Crate winsafe

Windows API and GUI in safe, idiomatic Rust.

Crate • GitHub • Docs (stable) • Docs (master branch) • Examples

WinSafe has:

low-level Win32 API constants, functions and structs;

high-level structs to build native Win32 GUI applications.

Usage

Add the dependency in your Cargo.toml:

[dependencies]
winsafe = { version = "0.0.20", features = [] }

Then you must enable the Cargo features you want to be included – these modules are named after native Windows DLL and library names, mostly.

The following Cargo features are available so far:

Feature	Description

comctl	ComCtl32.dll, for Common Controls

dshow	DirectShow

dwm	Desktop Window Manager

dxgi	DirectX Graphics Infrastructure

gdi	Gdi32.dll, the Windows GDI

gui	The WinSafe high-level GUI abstractions

kernel	Kernel32.dll, Advapi32.dll and Ktmw32.dll – all others will include it

mf	Media Foundation

ole	OLE and basic COM support

oleaut	OLE Automation

shell	Shell32.dll, Shlwapi.dll, and Userenv.dll, the COM-based Windows Shell

taskschd	Task Scheduler

user	User32.dll and ComDlg32.dll, the basic Windows GUI support

uxtheme	UxTheme.dll, extended window theming

version	Version.dll, to manipulate *.exe version info

If you’re looking for a comprehensive Win32 coverage, take a look at winapi or windows crates, which are unsafe, but have everything.

The GUI API

WinSafe features idiomatic bindings for the Win32 API, but on top of that, it features a set of high-level GUI structs, which scaffolds the boilerplate needed to build native Win32 GUI applications, event-oriented. Unless you’re doing something really specific, these high-level wrappers are highly recommended – you’ll usually start with the WindowMain.

One of the greatest strenghts of the GUI API is supporting the use of resource files, which can be created with a WYSIWYG resource editor.

GUI structs can be found in module gui.

Native function calls

The best way to understand the idea behind WinSafe bindings is comparing them to the correspondent C code.

For example, take the following C code:

HWND hwnd = GetDesktopWindow();
SetFocus(hwnd);

This is equivalent to:

use winsafe::{prelude::*, HWND};

let hwnd = HWND::GetDesktopWindow();
hwnd.SetFocus();

Note how GetDesktopWindow is a static method of HWND, and SetFocus is an instance method called directly upon hwnd. All native handles (HWND, HDC, HINSTANCE, etc.) are structs, thus:

native Win32 functions that return a handle are static methods in WinSafe;

native Win32 functions whose first parameter is a handle are instance methods.

Now this C code:

PostQuitMessage(0);

Is equivalent to:

use winsafe::PostQuitMessage;

PostQuitMessage(0);

Since PostQuitMessage is a free function, it’s simply at the root of the crate.

Also note that some functions which require a cleanup routine – like BeginPaint, for example – will return the resource wrapped in a guard, which will perform the cleanup automatically. You’ll never have to manually call EndPaint.

Sending messages are a special case, see the msg module.

Native constants

All native Win32 constants can be found in the co module. They’re all typed, what means that different constant types cannot be mixed (unless you explicitly say so).

Technically, each constant type is simply a newtype with a couple implementations, including those allowing bitflag operations. Also, all constant values can be converted to its underlying integer type.

The name of the constant type is often its prefix. For example, constants of MessageBox function, like MB_OKCANCEL, belong to a type called MB.

For example, take the following C code:

let hwnd = GetDesktopWindow();
MessageBox(hwnd, "Hello, world", "My hello", MB_OKCANCEL | MB_ICONINFORMATION);

This is equivalent to:

use winsafe::{prelude::*, co::MB, HWND};

let hwnd = HWND::GetDesktopWindow();
hwnd.MessageBox("Hello, world", "Title", MB::OKCANCEL | MB::ICONINFORMATION)?;

The method MessageBox, like most functions that can return errors, will return SysResult, which can contain an ERROR constant.

Native structs

WinSafe implements native Win32 structs in a very restricted way. First off, fields which control the size of the struct – often named cbSize – are private and automatically set when the struct is instantiated.

Pointer fields are also private, and they can be set and retrieved only through getter and setter methods. In particular, when setting a string pointer field, you need to pass a reference to a WString buffer, which will keep the actual string contents.

For example, the following C code:

WNDCLASSEX wcx = {0};
wcx.cbSize = sizeof(WNDCLASSEX);
wcx.lpszClassName = "MY_WINDOW";

if (RegisterClassEx(&wcx) == 0) {
    DWORD err = GetLastError();
    // handle error...
}

Is equivalent to:

use winsafe::{RegisterClassEx, WNDCLASSEX, WString};

let mut wcx = WNDCLASSEX::default();

let mut buf = WString::from_str("MY_WINDOW");
wcx.set_lpszClassName(Some(&mut buf));

if let Err(err) = RegisterClassEx(&wcx) {
    // handle error...
}

Note how you don’t need to call GetLastError to retrieve the error code: it’s returned by the method itself in the SysResult.

Text encoding

Windows natively uses Unicode UTF-16.

WinSafe uses Unicode UTF-16 internally but exposes idiomatic UTF-8, performing conversions automatically when needed, so you don’t have to worry about OsString or any low-level conversion.

However, there are cases where a string conversion is still needed, like when dealing with native Win32 structs. In such cases, you can use the WString struct, which is also capable of working as a buffer to receive text from Win32 calls.

Errors and result aliases

WinSafe declares a few Result aliases which are returned by its functions and methods:

Alias	Error	Used for

SysResult	ERROR	Standard system errors.

HrResult	HRESULT	COM errors.

AnyResult	Box<dyn Error + Send + Sync>	Holding different error types. All other Result aliases can be converted into it.

Utilities

Beyond the GUI API, WinSafe features a few high-level abstractions to deal with some particularly complex Win32 topics. Unless you need something specific, prefer using these over the raw, native calls:

Utility	Used for

Encoding	String encodings.

File	File read/write and other operations.

FileMapped	Memory-mapped file operations.

path	File path operations.

WString	Managing native wide strings.

Modules

cokernel

Native constants.

guardkernel

RAII implementation for various resources, which automatically perform cleanup routines when the object goes out of scope.

guigui

High-level GUI abstractions for user windows and native controls. They can be created programmatically or by loading resources from a .res file. These files can be created with a WYSIWYG resource editor.

msguser

Parameters of window messages.

pathkernel

File path utilities.

preludekernel

The WinSafe prelude.

vtole

Virtual tables of COM interfaces.

Macros

seq_ids

Generates sequential u16 constants starting from the given value.

Structs

ACCELuser

ACCEL struct.

ACLkernel

ACL struct.

ALTTABINFOuser

ALTTABINFO struct.

AM_MEDIA_TYPEdshow

AM_MEDIA_TYPE struct.

ATOMuser

ATOM returned by RegisterClassEx.

AccelMenuCtrlDatauser

Variant parameter for:

BITMAPgdi

BITMAP struct.

BITMAPFILEHEADERgdi

BITMAPFILEHEADER struct.

BITMAPINFOgdi

BITMAPINFO struct.

BITMAPINFOHEADERgdi

BITMAPINFOHEADER struct.

BLENDFUNCTIONuser

BLENDFUNCTION struct.

BSTRoleaut

A string data type used with COM automation.

BUTTON_IMAGELISTcomctl

BUTTON_IMAGELIST struct.

BUTTON_SPLITINFOcomctl

BUTTON_SPLITINFO struct.

BY_HANDLE_FILE_INFORMATIONkernel

BY_HANDLE_FILE_INFORMATION struct.

CHOOSECOLORuser

CHOOSECOLOR struct.

CLAIM_SECURITY_ATTRIBUTES_INFORMATIONkernel

CLAIM_SECURITY_ATTRIBUTES_INFORMATION struct.

CLAIM_SECURITY_ATTRIBUTE_FQBN_VALUEkernel

CLAIM_SECURITY_ATTRIBUTE_FQBN_VALUE struct.

CLAIM_SECURITY_ATTRIBUTE_OCTET_STRING_VALUEkernel

CLAIM_SECURITY_ATTRIBUTE_OCTET_STRING_VALUE struct.

CLAIM_SECURITY_ATTRIBUTE_V1kernel

CLAIM_SECURITY_ATTRIBUTE_V1 struct.

COAUTHIDENTITYole

COAUTHIDENTITY struct.

COAUTHINFOole

COAUTHINFO struct.

COLORREFuser

COLORREF struct.

COLORSCHEMEcomctl

COLORSCHEME struct.

COMBOBOXINFOuser

COMBOBOXINFO struct.

COMDLG_FILTERSPECshell

COMDLG_FILTERSPEC struct.

COMPAREITEMSTRUCTuser

COMPAREITEMSTRUCT struct.

CONSOLE_READCONSOLE_CONTROLkernel

CONSOLE_READCONSOLE_CONTROL struct.

COSERVERINFOole

COSERVERINFO struct.

CREATESTRUCTuser

CREATESTRUCT struct.

CURSORINFOuser

CURSORINFO struct.

DATETIMEPICKERINFOcomctl

DATETIMEPICKERINFO struct.

DELETEITEMSTRUCTuser

DELETEITEMSTRUCT struct.

DEVMODEuser

DEVMODE struct.

DEV_BROADCAST_DEVICEINTERFACEkernel

DEV_BROADCAST_DEVICEINTERFACE struct.

DEV_BROADCAST_HANDLEkernel

DEV_BROADCAST_HANDLE struct.

DEV_BROADCAST_HDRkernel

DEV_BROADCAST_HDR struct.

DEV_BROADCAST_OEMkernel

DEV_BROADCAST_OEM struct.

DEV_BROADCAST_PORTkernel

DEV_BROADCAST_PORT struct.

DEV_BROADCAST_VOLUMEkernel

DEV_BROADCAST_VOLUME struct.

DISK_SPACE_INFORMATIONkernel

DISK_SPACE_INFORMATION struct.

DISPLAY_DEVICEuser

DISPLAY_DEVICE struct.

DLGITEMTEMPLATEuser

DLGITEMTEMPLATE struct.

DLGTEMPLATEuser

DLGTEMPLATE struct.

DRAWITEMSTRUCTuser

DRAWITEMSTRUCT struct.

DRAWTEXTPARAMSuser

DRAWTEXTPARAMS struct.

DVINFOdshow

DVINFO struct.

DVTARGETDEVICEole

DVTARGETDEVICE struct.

DXGI_ADAPTER_DESCdxgi

DXGI_ADAPTER_DESC struct.

DXGI_ADAPTER_DESC1dxgi

DXGI_ADAPTER_DESC1 struct.

DXGI_FRAME_STATISTICSdxgi

DXGI_FRAME_STATISTICS struct.

DXGI_GAMMA_CONTROLdxgi

DXGI_GAMMA_CONTROL struct.

DXGI_GAMMA_CONTROL_CAPABILITIESdxgi

DXGI_GAMMA_CONTROL_CAPABILITIES struct.

DXGI_MAPPED_RECTdxgi

DXGI_MAPPED_RECT struct.

DXGI_MODE_DESCdxgi

DXGI_MODE_DESC struct.

DXGI_OUTPUT_DESCdxgi

DXGI_OUTPUT_DESC struct.

DXGI_RATIONALdxgi

DXGI_RATIONAL struct.

DXGI_RGBdxgi

DXGI_RGB struct.

DXGI_SAMPLE_DESCdxgi

DXGI_SAMPLE_DESC struct.

DXGI_SHARED_RESOURCEdxgi

DXGI_SHARED_RESOURCE struct.

DXGI_SURFACE_DESCdxgi

DXGI_SURFACE_DESC struct.

DXGI_SWAP_CHAIN_DESCdxgi

DXGI_SWAP_CHAIN_DESC struct.

EDITBALLOONTIPcomctl

EDITBALLOONTIP struct.

FILETIMEkernel

FILETIME struct.

FILTER_INFOdshow

FILTER_INFO struct.

FORMATETCole

FORMATETC struct.

Filekernel

Manages an HFILE handle, which provides file read/write and other operations. It is closed automatically when the object goes out of scope.

FileMappedkernel

Manages an HFILEMAP handle, which provides memory-mapped file operations, including read/write through slices. It is closed automatically when the object goes out of scope.

GUIDkernel

GUID struct.

GUITHREADINFOuser

GUITHREADINFO struct.

HACCELuser

Handle to an accelerator table.

HACCESSTOKENkernel

Handle to an access token. Originally just a HANDLE.

HARDWAREINPUTuser

HARDWAREINPUT struct.

HBITMAPuser

Handle to a bitmap.

HBRUSHuser

Handle to a brush.

HCURSORuser

Handle to a cursor.

HDCuser

Handle to a device context.

HDESKuser

Handle to a desktop.

HDHITTESTINFOcomctl

HDHITTESTINFO struct.

HDITEMcomctl

HDITEM struct.

HDLAYOUTcomctl

HDLAYOUT struct.

HDROPshell

Handle to an internal drop structure.

HDWPuser

Handle to a deferred window position.

HEAPLIST32kernel

HEAPLIST32 struct.

HELPINFOuser

HELPINFO struct.

HEVENTkernel

Handle to a named or unnamed event object. Originally just a HANDLE.

HEVENTLOGkernel

Handle to an event log. Originally just a HANDLE.

HFILEkernel

Handle to a file. Originally just a HANDLE.

HFILEMAPkernel

Handle to a file mapping. Originally just a HANDLE.

HFILEMAPVIEWkernel

Address of a mapped view. Originally just an LPVOID.

HFINDFILEkernel

Handle to a file search. Originally just a HANDLE.

HFONTgdi

Handle to a font.

HGLOBALkernel

Handle to a global memory block. Originally just a HANDLE.

HHEAPkernel

Handle to a heap object. Originally just a HANDLE.

HHOOKuser

Handle to a hook.

HICONuser

Handle to an icon.

HIMAGELISTcomctl

Handle to an image list.

HINSTANCEkernel

Handle to an instance, same as HMODULE.

HKEYkernel

Handle to a registry key.

HLOCALkernel

Handle to a local memory block.

HMENUuser

Handle to a menu.

HMONITORuser

Handle to a display monitor.

HPALETTEgdi

Handle to a palette.

HPENgdi

Handle to a pen GDI object.

HPIPEkernel

Handle to an anonymous pipe. Originally just a HANDLE.

HPROCESSkernel

Handle to a process. Originally just a HANDLE.

HPROCESSLISTkernel

Handle to a process list snapshot. Originally just a HANDLE.

HRGNuser

Handle to a region GDI object.

HRSRCkernel

Handle to a resource. Originally just a HANDLE.

HRSRCMEMkernel

Handle to a resource memory block. Originally just an HGLOBAL.

HSCkernel

Handle to a Service Control Manager. Originally SC_HANDLE.

HSERVICEkernel

Handle to a service. Originally SC_HANDLE.

HSERVICESTATUSkernel

Handle to a service status. Originally SERVICE_STATUS_HANDLE.

HSTDkernel

Handle to a standard device. Originally just a HANDLE.

HTHEMEuxtheme

Handle to a theme.

HTHREADkernel

Handle to a thread. Originally just a HANDLE.

HTRANSACTIONkernel

Handle to a transaction. Originally just a HANDLE.

HTREEITEMcomctl

Handle to a tree view item.

HUPDATERSRCkernel

Handle to an updateable resource. Originally just a HANDLE.

HVERSIONINFOversion

Handle to a version info block.

HWNDuser

Handle to a window.

IActiontaskschd

IAction COM interface over IActionVT.

IActionCollectiontaskschd

IActionCollection COM interface over IActionCollectionVT.

IAdviseSinkole

IAdviseSink COM interface over IAdviseSinkVT.

IBaseFilterdshow

IBaseFilter COM interface over IBaseFilterVT.

IBindCtxole

IBindCtx COM interface over IBindCtxVT.

IBootTriggertaskschd

IBootTrigger COM interface over IBootTriggerVT.

ICONINFOuser

ICONINFO struct.

ICONINFOEXuser

ICONINFOEX struct.

IComHandlerActiontaskschd

IComHandlerAction COM interface over IComHandlerActionVT.

IDXGIAdapterdxgi

IDXGIAdapter COM interface over IDXGIAdapterVT.

IDXGIAdapter1dxgi

IDXGIAdapter1 COM interface over IDXGIAdapter1VT.

IDXGIDevicedxgi

IDXGIDevice COM interface over IDXGIDeviceVT.

IDXGIDeviceSubObjectdxgi

IDXGIDeviceSubObject COM interface over IDXGIDeviceSubObjectVT.

IDXGIFactorydxgi

IDXGIFactory COM interface over IDXGIFactoryVT.

IDXGIFactory1dxgi

IDXGIFactory1 COM interface over IDXGIFactory1VT.

IDXGIKeyedMutexdxgi

IDXGIKeyedMutex COM interface over IDXGIKeyedMutexVT.

IDXGIObjectdxgi

IDXGIObject COM interface over IDXGIObjectVT.

IDXGIOutputdxgi

IDXGIOutput COM interface over IDXGIOutputVT.

IDXGIResourcedxgi

IDXGIResource COM interface over IDXGIResourceVT.

IDXGISurfacedxgi

IDXGISurface COM interface over IDXGISurfaceVT.

IDXGISwapChaindxgi

IDXGISwapChain COM interface over IDXGISwapChainVT.

IDailyTriggertaskschd

IDailyTrigger COM interface over IDailyTriggerVT.

IDataObjectole

IDataObject COM interface over IDataObjectVT.

IDispatcholeaut

IDispatch COM interface over IDispatchVT.

IDropTargetole

IDropTarget COM interface over IDropTargetVT.

IEmailActiontaskschd

IEmailAction COM interface over IEmailActionVT.

IEnumFiltersdshow

IEnumFilters COM interface over IEnumFiltersVT.

IEnumMediaTypesdshow

IEnumMediaTypes COM interface over IEnumMediaTypesVT.

IEnumPinsdshow

IEnumPins COM interface over IEnumPinsVT.

IEnumShellItemsshell

IEnumShellItems COM interface over IEnumShellItemsVT.

IEventTriggertaskschd

IEventTrigger COM interface over IEventTriggerVT.

IExecActiontaskschd

IExecAction COM interface over IExecActionVT.

IFileDialogshell

IFileDialog COM interface over IFileDialogVT.

IFileDialogEventsshell

IFileDialogEvents COM interface over IFileDialogEventsVT.

IFileOpenDialogshell

IFileOpenDialog COM interface over IFileOpenDialogVT.

IFileSaveDialogshell

IFileSaveDialog COM interface over IFileSaveDialogVT.

IFileSinkFilterdshow

IFileSinkFilter COM interface over IFileSinkFilterVT.

IFilterGraphdshow

IFilterGraph COM interface over IFilterGraphVT.

IFilterGraph2dshow

IFilterGraph2 COM interface over IFilterGraph2VT.

IGraphBuilderdshow

IGraphBuilder COM interface over IGraphBuilderVT.

IIdleTriggertaskschd

IIdleTrigger COM interface over IIdleTriggerVT.

ILogonTriggertaskschd

ILogonTrigger COM interface over ILogonTriggerVT.

IMAGELISTDRAWPARAMScomctl and gdi

IMAGELISTDRAWPARAMS struct.

IMFAsyncCallbackmf

IMFAsyncCallback COM interface over IMFAsyncCallbackVT.

IMFAsyncResultmf

IMFAsyncResult COM interface over IMFAsyncResultVT.

IMFAttributesmf

IMFAttributes COM interface over IMFAttributesVT.

IMFClockmf

IMFClock COM interface over IMFClockVT.

IMFGetServicemf

IMFGetService COM interface over IMFGetServiceVT.

IMFMediaEventmf

IMFMediaEvent COM interface over IMFMediaEventVT.

IMFMediaEventGeneratormf

IMFMediaEventGenerator COM interface over IMFMediaEventGeneratorVT.

IMFMediaSessionmf

IMFMediaSession COM interface over IMFMediaSessionVT.

IMFMediaSourcemf

IMFMediaSource COM interface over IMFMediaSourceVT.

IMFPresentationDescriptormf

IMFPresentationDescriptor COM interface over IMFPresentationDescriptorVT.

IMFSourceResolvermf

IMFSourceResolver COM interface over IMFSourceResolverVT.

IMFTopologymf

IMFTopology COM interface over IMFTopologyVT.

IMFTopologyNodemf

IMFTopologyNode COM interface over IMFTopologyNodeVT.

IMFVideoDisplayControlmf

IMFVideoDisplayControl COM interface over IMFVideoDisplayControlVT.

IMediaControldshow

IMediaControl COM interface over IMediaControlVT.

IMediaFilterdshow

IMediaFilter COM interface over IMediaFilterVT.

IMediaSeekingdshow

IMediaSeeking COM interface over IMediaSeekingVT. Inherits from IUnknown.

IModalWindowshell

IModalWindow COM interface over IModalWindowVT.

IMonikerole

IMoniker COM interface over IMonikerVT.

INITCOMMONCONTROLSEXcomctl

INITCOMMONCONTROLSEX struct

INPUTuser

INPUT struct.

IPersistole

IPersist COM interface over IPersistVT.

IPersistFileole

IPersistFile COM interface over IPersistFileVT.

IPersistStreamole

IPersistStream COM interface over IPersistStreamVT.

IPictureole

IPicture COM interface over IPictureVT.

IPindshow

IPin COM interface over IPinVT.

IPropertyStoreoleaut

IPropertyStore COM interface over IPropertyStoreVT.

IRegisteredTasktaskschd

IRegisteredTask COM interface over IRegisteredTaskVT.

IRegistrationInfotaskschd

IRegistrationInfo COM interface over IRegistrationInfoVT.

ISequentialStreamole

ISequentialStream COM interface over ISequentialStreamVT.

IShellItemshell

IShellItem COM interface over IShellItemVT.

IShellItem2shell

IShellItem2 COM interface over IShellItem2VT.

IShellItemArrayshell

IShellItemArray COM interface over IShellItemArrayVT.

IShellLinkshell

IShellLink COM interface over IShellLinkVT.

IStorageole

IStorage COM interface over IStorageVT.

IStreamole

IStream COM interface over IStreamVT.

ITaskDefinitiontaskschd

ITaskDefinition COM interface over ITaskDefinitionVT.

ITaskFoldertaskschd

ITaskFolder COM interface over ITaskFolderVT.

ITaskServicetaskschd

ITaskService COM interface over ITaskServiceVT.

ITaskbarListshell

ITaskbarList COM interface over ITaskbarListVT.

ITaskbarList2shell

ITaskbarList2 COM interface over ITaskbarList2VT.

ITaskbarList3shell

ITaskbarList3 COM interface over ITaskbarList3VT.

ITaskbarList4shell

ITaskbarList4 COM interface over ITaskbarList4VT.

ITriggertaskschd

ITrigger COM interface over ITriggerVT.

ITriggerCollectiontaskschd

ITriggerCollection COM interface over ITriggerCollectionVT.

ITypeInfooleaut

ITypeInfo COM interface over ITypeInfoVT.

IUnknownole

IUnknown COM interface over IUnknownVT. It’s the base to all COM interfaces.

KEYBDINPUTuser

KEYBDINPUT struct.

LANGIDkernel

LANGID language identifier.

LASTINPUTINFOuser

LASTINPUTINFO struct.

LCIDkernel

LCID locale identifier.

LITEMcomctl

LITEM struct.

LOGBRUSHgdi

LOGBRUSH struct.

LOGFONTgdi

LOGFONT struct.

LOGPALETTEgdi

LOGPALETTE struct.

LOGPENgdi

LOGPEN struct.

LUIDkernel

LUID identifier.

LUID_AND_ATTRIBUTESkernel

LUID_AND_ATTRIBUTES struct.

LVBKIMAGEcomctl

LVBKIMAGE struct.

LVCOLUMNcomctl

LVCOLUMN struct.

LVFINDINFOcomctl

LVFINDINFO struct.

LVFOOTERINFOcomctl

LVFOOTERINFO struct.

LVFOOTERITEMcomctl

LVFOOTERITEM struct.

LVGROUPcomctl

LVGROUP struct.

LVGROUPMETRICScomctl

LVGROUPMETRICS struct.

LVHITTESTINFOcomctl

LVHITTESTINFO struct.

LVINSERTGROUPSORTEDcomctl

LVINSERTGROUPSORTED struct.

LVINSERTMARKcomctl

LVINSERTMARK struct.

LVITEMcomctl

LVITEM struct.

LVITEMINDEXcomctl

LVITEMINDEX struct.

LVSETINFOTIPcomctl

LVSETINFOTIP struct.

LVTILEINFOcomctl

LVTILEINFO struct.

LVTILEVIEWINFOcomctl

LVTILEVIEWINFO struct.

MARGINSuxtheme

MARGINS struct.

MCGRIDINFOcomctl

MCGRIDINFO struct.

MCHITTESTINFOcomctl

MCHITTESTINFO struct.

MEMORYSTATUSEXkernel

MEMORYSTATUSEX struct.

MENUBARINFOuser

MENUBARINFO struct.

MENUINFOuser

MENUINFO struct.

MENUITEMINFOuser

MENUITEMINFO struct.

MFCLOCK_PROPERTIESmf

MFCLOCK_PROPERTIES struct.

MFVideoNormalizedRectmf

MFVideoNormalizedRect struct.

MINMAXINFOuser

MINMAXINFO struct.

MODULEENTRY32kernel

MODULEENTRY32 struct.

MONITORINFOEXuser

MONITORINFOEX struct.

MONTHDAYSTATEcomctl

MONTHDAYSTATE struct.

MOUSEINPUTuser

MOUSEINPUT struct.

MSGuser

MSG struct.

MULTI_QIole

MULTI_QI struct.

NCCALCSIZE_PARAMSuser

NCCALCSIZE_PARAMS struct.

NMBCDROPDOWNcomctl

NMBCDROPDOWN struct.

NMBCHOTITEMcomctl

NMBCHOTITEM struct.

NMCHARcomctl

NMCHAR struct.

NMCUSTOMDRAWcomctl

NMCUSTOMDRAW struct.

NMDATETIMECHANGEcomctl

NMDATETIMECHANGE struct.

NMDATETIMEFORMATcomctl

NMDATETIMEFORMAT struct.

NMDATETIMEFORMATQUERYcomctl

NMDATETIMEFORMATQUERY struct.

NMDATETIMESTRINGcomctl

NMDATETIMESTRING struct.

NMDATETIMEWMKEYDOWNcomctl

NMDATETIMEWMKEYDOWN struct.

NMDAYSTATEcomctl

NMDAYSTATE struct.

NMHDDISPINFOcomctl

NMHDDISPINFO struct.

NMHDFILTERBTNCLICKcomctl

NMHDFILTERBTNCLICK struct.

NMHDRcomctl

NMHDR struct.

NMHEADERcomctl

NMHEADER struct.

NMIPADDRESScomctl

NMIPADDRESS struct.

NMITEMACTIVATEcomctl

NMITEMACTIVATE struct.

NMLINKcomctl

NMLINK struct.

NMLISTVIEWcomctl

NMLISTVIEW struct.

NMLVCACHEHINTcomctl

NMLVCACHEHINT struct.

NMLVCUSTOMDRAWcomctl

NMLVCUSTOMDRAW struct.

NMLVDISPINFOcomctl

NMLVDISPINFO struct.

NMLVEMPTYMARKUPcomctl

NMLVEMPTYMARKUP struct.

NMLVFINDITEMcomctl

NMLVFINDITEM struct.

NMLVGETINFOTIPcomctl

NMLVGETINFOTIP struct.

NMLVKEYDOWNcomctl

NMLVKEYDOWN struct.

NMLVLINKcomctl

NMLVLINK struct.

NMLVODSTATECHANGEcomctl

NMLVODSTATECHANGE struct.

NMLVSCROLLcomctl

NMLVSCROLL struct.

NMMOUSEcomctl

NMMOUSE struct.

NMOBJECTNOTIFYcomctl

NMOBJECTNOTIFY struct.

NMSELCHANGEcomctl

NMSELCHANGE struct.

NMTCKEYDOWNcomctl

NMTCKEYDOWN struct.

NMTRBTHUMBPOSCHANGINGcomctl

NMTRBTHUMBPOSCHANGING struct.

NMTREEVIEWcomctl

NMTREEVIEW struct.

NMTVASYNCDRAWcomctl and gdi

NMTVASYNCDRAW struct.

NMTVCUSTOMDRAWcomctl

NMTVCUSTOMDRAW stuct.

NMTVITEMCHANGEcomctl

NMTVITEMCHANGE struct.

NMUPDOWNcomctl

NMUPDOWN struct.

NMVIEWCHANGEcomctl

NMVIEWCHANGE struct.

NONCLIENTMETRICSgdi

NONCLIENTMETRICS struct.

NOTIFYICONDATAshell

NOTIFYICONDATA struct.

OSVERSIONINFOEXkernel

OSVERSIONINFOEX struct.

OVERLAPPEDkernel

OVERLAPPED struct.

PAINTSTRUCTuser

PAINTSTRUCT struct.

PALETTEENTRYgdi

PALETTEENTRY struct.

PBRANGEcomctl

PBRANGE struct.

PIN_INFOdshow

PIN_INFO struct.

POINTuser

POINT struct.

POWERBROADCAST_SETTINGkernel

POWERBROADCAST_SETTING struct.

PROCESSENTRY32kernel

PROCESSENTRY32 struct.

PROCESSOR_NUMBERkernel

PROCESSOR_NUMBER struct.

PROCESS_HEAP_ENTRYkernel

PROCESS_HEAP_ENTRY struct.

PROCESS_HEAP_ENTRY_Blockkernel

PROCESS_HEAP_ENTRY Block.

PROCESS_HEAP_ENTRY_Regionkernel

PROCESS_HEAP_ENTRY Region.

PROCESS_INFORMATIONkernel

PROCESS_INFORMATION struct.

PROPERTYKEYoleaut

PROPERTYKEY struct.

PROPVARIANToleaut

PROPVARIANT struct.

RECTuser

RECT struct.

RGBQUADgdi

RGBQUAD struct.

SCROLLINFOuser

SCROLLINFO struct.

SECURITY_ATTRIBUTESkernel

SECURITY_ATTRIBUTES struct.

SECURITY_DESCRIPTORkernel

SECURITY_DESCRIPTOR struct.

SERVICE_STATUSkernel

SERVICE_STATUS struct.

SERVICE_TIMECHANGE_INFOkernel

SERVICE_TIMECHANGE_INFO struct.

SHELLEXECUTEINFOshell

SHELLEXECUTEINFO struct.

SHFILEINFOshell

SHFILEINFO struct.

SHFILEOPSTRUCTshell

SHFILEOPSTRUCT struct.

SHSTOCKICONINFOshell

SHSTOCKICONINFO struct.

SIDkernel

SID struct.

SID_AND_ATTRIBUTESkernel

SID_AND_ATTRIBUTES struct.

SID_AND_ATTRIBUTES_HASHkernel

SID_AND_ATTRIBUTES_HASH struct.

SID_IDENTIFIER_AUTHORITYkernel

SID_IDENTIFIER_AUTHORITY struct.

SIZEuser

SIZE struct.

SNBole

SNB struct.

STARTUPINFOkernel

STARTUPINFO struct.

STYLESTRUCTuser

STYLESTRUCT struct.

SYSTEMTIMEkernel

SYSTEMTIME struct.

SYSTEM_INFOkernel

SYSTEM_INFO struct.

TASKDIALOGCONFIGcomctl

TASKDIALOG_BUTTON struct.

TBADDBITMAPcomctl

TBADDBITMAP struct.

TBBUTTONcomctl

TBBUTTON struct.

TBBUTTONINFOcomctl

TBBUTTONINFO struct.

TBINSERTMARKcomctl

TBINSERTMARK struct.

TBMETRICScomctl

TBMETRICS struct.

TBREPLACEBITMAPcomctl

TBREPLACEBITMAP struct.

TBSAVEPARAMScomctl

TBSAVEPARAMS struct.

TCHITTESTINFOcomctl

TCHITTESTINFO struct.

TCITEMcomctl

TCITEM struct.

TEXTMETRICgdi

TEXTMETRIC struct.

THREADENTRY32kernel

THREADENTRY32 struct.

TIME_ZONE_INFORMATIONkernel

TIME_ZONE_INFORMATION struct.

TITLEBARINFOEXuser

TITLEBARINFOEX struct.

TOKEN_ACCESS_INFORMATIONkernel

TOKEN_ACCESS_INFORMATION struct.

TOKEN_APPCONTAINER_INFORMATIONkernel

TOKEN_APPCONTAINER_INFORMATION struct.

TOKEN_DEFAULT_DACLkernel

TOKEN_DEFAULT_DACL struct.

TOKEN_ELEVATIONkernel

TOKEN_ELEVATION struct.

TOKEN_GROUPSkernel

TOKEN_GROUPS struct.

TOKEN_GROUPS_AND_PRIVILEGESkernel

TOKEN_GROUPS_AND_PRIVILEGES struct.

TOKEN_LINKED_TOKENkernel

TOKEN_LINKED_TOKEN struct.

TOKEN_MANDATORY_LABELkernel

TOKEN_MANDATORY_LABEL struct.

TOKEN_MANDATORY_POLICYkernel

TOKEN_MANDATORY_POLICY struct.

TOKEN_ORIGINkernel

TOKEN_ORIGIN struct.

TOKEN_OWNERkernel

TOKEN_OWNER struct.

TOKEN_PRIMARY_GROUPkernel

TOKEN_PRIMARY_GROUP struct.

TOKEN_PRIVILEGESkernel

TOKEN_PRIVILEGES struct.

TOKEN_SOURCEkernel

TOKEN_SOURCE struct.

TOKEN_STATISTICSkernel

TOKEN_STATISTICS struct.

TOKEN_USERkernel

TOKEN_USER struct.

TRACKMOUSEEVENTuser

TRACKMOUSEEVENT struct.

TVHITTESTINFOcomctl

TVHITTESTINFO struct.

TVINSERTSTRUCTcomctl

TVINSERTSTRUCT struct.

TVITEMcomctl

TVITEM struct.

TVITEMEXcomctl

TVITEMEX struct.

TVSORTCBcomctl

TVSORTCB struct.

UDACCELcomctl

UDACCEL struct.

VALENTkernel

VALENT struct.

VARIANToleaut

VARIANT struct.

VS_FIXEDFILEINFOversion

VS_FIXEDFILEINFO struct.

WIN32_FIND_DATAkernel

WIN32_FIND_DATA struct.

WINDOWINFOuser

WINDOWINFO struct.

WINDOWPLACEMENTuser

WINDOWPLACEMENT struct.

WINDOWPOSuser

WINDOWPOS struct.

WNDCLASSEXuser

WNDCLASSEX struct.

WStringkernel

Stores a [u16] buffer for a null-terminated Unicode UTF-16 wide string natively used by Windows.

WTSSESSION_NOTIFICATIONkernel

WTSSESSION_NOTIFICATION struct.

Enums

AccelMenuCtrluser

Variant parameter for:

AtomStruser

Variant parameter for:

BmpIconuser

Variant parameter for:

BmpIconCurMetacomctl

Variant parameter for:

BmpIdbRescomctl

Variant parameter for:

BmpInstIdcomctl

Variant parameter for:

BmpPtrStruser

Variant parameter for:

ClaimSecurityAttrkernel

Variable parameter for:

ClrDefNonecomctl

Variant parameter for:

CurObjgdi

Variant parameter for:

DisabPrivkernel

Variable parameter for:

DispfNupuser

Variant parameter for:

Encodingkernel

String encodings.

FileAccesskernel

Access types for File::open and FileMapped::open.

GmidxEnumuser

Variant parameter for:

HwKbMouseuser

Variant parameter for:

HwndFocususer

Variant parameter for:

HwndHmenuuser

Variant parameter for:

HwndPlaceuser

Variant parameter for:

HwndPointIduser

Variant parameter for:

IcoMonshell

Variable parameter for:

IconIdcomctl

Variant parameter for:

IconIdTdcomctl

Variant parameter for:

IconRescomctl

Variant parameter for:

IdIdcStruser

Variant parameter for:

IdIdiStruser

Variant parameter for:

IdMenuuser

Variant parameter used in menu methods:

IdObmStrgdi

Variant parameter for:

IdOcrStrgdi

Variant parameter for:

IdOicStrgdi

Variant parameter for:

IdPosuser

Variant parameter for:

IdStrkernel

A resource identifier.

IdxCbNonecomctl

Variant type for:

IdxStrcomctl

Variant parameter for:

MenuItemuser

Variant parameter for:

MenuItemInfouser

Variant parameter for:

NccspRectuser

Variant parameter for:

PowerSettingkernel

Variant parameter for:

PowerSettingAwayModekernel

Variant parameter for:

PowerSettingLidkernel

Variant parameter for:

PtIdxcomctl

Variant parameter for:

PtsRcuser

Variant parameter for:

RegistryValuekernel

Registry value types.

ResStrscomctl

Variant parameter for:

RtStrkernel

A predefined resource identifier.

SuccessTimeoutuser

Variant parameter for:

SvcCtlkernel

Notification content for HSERVICESTATUS::RegisterServiceCtrlHandlerEx callback, describing co::SERVICE_CONTROL.

SvcCtlDeviceEventkernel

Notification content for SvcCtl.

SvcCtlPowerEventkernel

Notification content for SvcCtl.

TokenInfokernel

Variant parameter for:

TreeitemTvicomctl

Variant parameter for:

Functions

AdjustWindowRectExuser

AdjustWindowRectEx function.

AdjustWindowRectExForDpiuser

AdjustWindowRectExForDpi function.

AllocateAndInitializeSidkernel

AllocateAndInitializeSid function.

AllowSetForegroundWindowuser

AllowSetForegroundWindow function

AnyPopupuser

AnyPopup function.

AttachThreadInputuser

AttachThreadInput function.

BlockInputuser

BlockInput function.

BroadcastSystemMessage⚠user

BroadcastSystemMessage function.

CLSIDFromProgIDole

CLSIDFromProgID function.

CLSIDFromProgIDExole

CLSIDFromProgIDEx function.

CLSIDFromStringole

CLSIDFromString function.

ChangeDisplaySettingsuser

ChangeDisplaySettings function.

ChangeDisplaySettingsExuser

ChangeDisplaySettingsEx function.

ChooseColoruser

ChooseColor function.

ClipCursoruser

ClipCursor function.

CoCreateGuidole

CoCreateGuid function.

CoCreateInstanceole

CoCreateInstance function.

CoCreateInstanceExole

CoCreateInstanceEx function.

CoInitializeExole

CoInitializeEx function, which initializes the COM library. When succeeding, returns an informational error code.

CoLockObjectExternalole

CoLockObjectExternal function.

CoTaskMemAllocole

CoTaskMemAlloc function.

CoTaskMemReallocole

CoTaskMemRealloc function.

CommDlgExtendedErroruser

CommDlgExtendedError function.

CommandLineToArgvshell

CommandLineToArgv function.

ConvertSidToStringSidkernel

ConvertSidToStringSid function.

ConvertStringSidToSidkernel

ConvertStringSidToSid function.

CopyFilekernel

CopyFile function.

CopySidkernel

CopySid function.

CreateClassMonikerole

CreateClassMoniker function.

CreateDXGIFactorydxgi

CreateDXGIFactory function.

CreateDXGIFactory1dxgi

CreateDXGIFactory1 function.

CreateDirectorykernel

CreateDirectory function.

CreateFileMonikerole

CreateFileMoniker function.

CreateItemMonikerole

CreateItemMoniker function.

CreateObjrefMonikerole

CreateObjrefMoniker function.

CreatePointerMonikerole

CreatePointerMoniker function.

CreateWellKnownSidkernel

CreateWellKnownSid function.

DecryptFilekernel

DecryptFile function.

DeleteFilekernel

DeleteFile function.

DispatchMessage⚠user

DispatchMessage function.

DwmEnableMMCSSdwm

DwmEnableMMCSS function.

DwmFlushdwm

DwmFlush function.

DwmGetColorizationColordwm

DwmGetColorizationColor function.

DwmIsCompositionEnableddwm

DwmIsCompositionEnabled function.

EmptyClipboarduser

EmptyClipboard function.

EncryptFilekernel

EncryptFile function.

EncryptionDisablekernel

EncryptionDisable function.

EndMenuuser

EndMenu function.

EnumDisplayDevicesuser

EnumDisplayDevices function.

EnumDisplaySettingsuser

EnumDisplaySettings function.

EnumDisplaySettingsExuser

EnumDisplaySettingsEx function.

EnumThreadWindowsuser

EnumThreadWindows function.

EnumWindowsuser

EnumWindows function.

EqualDomainSidkernel

EqualDomainSid function.

EqualPrefixSidkernel

EqualPrefixSid function.

EqualSidkernel

EqualSid function.

ExitProcesskernel

ExitProcess function.

ExitThreadkernel

ExitThread function.

ExitWindowsExuser

ExitWindowsEx function.

ExpandEnvironmentStringskernel

ExpandEnvironmentStrings function.

FileTimeToSystemTimekernel

FileTimeToSystemTime function.

FlushProcessWriteBufferskernel

FlushProcessWriteBuffers function.

FormatMessage⚠kernel

FormatMessage function.

GdiFlushgdi

GdiFlush function.

GdiGetBatchLimitgdi

GdiGetBatchLimit function.

GdiSetBatchLimitgdi

GdiSetBatchLimit function.

GetAllUsersProfileDirectoryshell

GetAllUsersProfileDirectory function.

GetAsyncKeyStateuser

GetAsyncKeyState function.

GetBinaryTypekernel

GetBinaryType function.

GetCaretBlinkTimeuser

GetCaretBlinkTime function.

GetCaretPosuser

GetCaretPos function.

GetClipCursoruser

GetClipCursor function.

GetClipboardData⚠user

GetClipboardData function.

GetClipboardSequenceNumberuser

GetClipboardSequenceNumber function.

GetCommandLinekernel

GetCommandLine function.

GetComputerNamekernel

GetComputerName function.

GetCurrentDirectorykernel

GetCurrentDirectory function.

GetCurrentProcessIdkernel

GetCurrentProcessId function.

GetCurrentThreadIdkernel

GetCurrentThreadId function.

GetCursorInfouser

GetCursorInfo function.

GetCursorPosuser

GetCursorPos function.

GetDefaultUserProfileDirectoryshell

GetDefaultUserProfileDirectory function.

GetDialogBaseUnitsuser

GetDialogBaseUnits function.

GetDiskFreeSpaceExkernel

GetDiskFreeSpaceEx function.

GetDiskSpaceInformationkernel

GetDiskSpaceInformation function.

GetDoubleClickTimeuser

GetDoubleClickTime function.

GetDriveTypekernel

GetDriveType function.

GetEnvironmentStringskernel

GetEnvironmentStrings function.

GetFileAttributeskernel

GetFileAttributes function.

GetFirmwareTypekernel

GetFirmwareType function.

GetGUIThreadInfouser

GetGUIThreadInfo function.

GetLargePageMinimumkernel

GetLargePageMinimum function.

GetLastErrorkernel

GetLastError function.

GetLastInputInfouser

GetLastInputInfo function.

GetLengthSidkernel

GetLengthSid function.

GetLocalTimekernel

GetLocalTime function.

GetLogicalDriveStringskernel

GetLogicalDriveStrings function.

GetLogicalDriveskernel

GetLogicalDrives function.

GetMenuCheckMarkDimensionsuser

GetMenuCheckMarkDimensions function.

GetMessageuser

GetMessage function.

GetMessagePosuser

GetMessagePos function.

GetNativeSystemInfokernel

GetNativeSystemInfo function.

GetPrivateProfileSectionkernel

GetPrivateProfileSection function.

GetPrivateProfileSectionNameskernel

GetPrivateProfileSectionNames function.

GetPrivateProfileStringkernel

GetPrivateProfileString function.

GetProfilesDirectoryshell

GetProfilesDirectory function.

GetQueueStatususer

GetQueueStatus function.

GetSidLengthRequiredkernel

GetSidLengthRequired function.

GetStartupInfokernel

GetStartupInfo function.

GetSysColoruser

GetSysColor function.

GetSystemDirectorykernel

GetSystemDirectory function.

GetSystemFileCacheSizekernel

GetSystemFileCacheSize function.

GetSystemInfokernel

GetSystemInfo function.

GetSystemMetricsuser

GetSystemMetrics function.

GetSystemMetricsForDpiuser

GetSystemMetricsForDpi function.

GetSystemTimekernel

GetSystemTime function.

GetSystemTimeAsFileTimekernel

GetSystemTimeAsFileTime function.

GetSystemTimePreciseAsFileTimekernel

GetSystemTimePreciseAsFileTime function.

GetSystemTimeskernel

GetSystemTimes function.

GetTempFileNamekernel

GetTempFileName function.

GetTempPathkernel

GetTempPath function.

GetThreadDpiHostingBehavioruser

GetThreadDpiHostingBehavior function.

GetTickCount64kernel

GetTickCount64 function.

GetUserNamekernel

GetUserName function.

GetVolumeInformationkernel

GetVolumeInformation function.

GetVolumePathNamekernel

GetVolumePathName function.

GetWindowsAccountDomainSidkernel

GetWindowsAccountDomainSid function.

GlobalMemoryStatusExkernel

GlobalMemoryStatusEx function.

HIBYTEkernel

HIBYTE macro.

HIDWORDkernel

Returns the high-order u32 of an u64.

HIWORDkernel

HIWORD macro.

InSendMessageuser

InSendMessage function.

InSendMessageExuser and 64-bit

InSendMessageEx function.

InflateRectuser

InflateRect function.

InitCommonControlscomctl

InitCommonControls function.

InitCommonControlsExcomctl

InitCommonControlsEx function.

InitMUILanguagecomctl

InitMUILanguage function.

InitializeSecurityDescriptorkernel

InitializeSecurityDescriptor function.

InitiateSystemShutdownkernel

InitiateSystemShutdown function.

InitiateSystemShutdownExkernel

InitiateSystemShutdownEx function.

IntersectRectuser

IntersectRect function.

IsAppThemeduxtheme

IsAppThemed function.

IsCompositionActiveuxtheme

IsCompositionActive function.

IsDebuggerPresentkernel

IsDebuggerPresent function.

IsGUIThreaduser

IsGUIThread function.

IsNativeVhdBootkernel

IsNativeVhdBoot function.

IsRectEmptyuser

IsRectEmpty function.

IsThemeActiveuxtheme

IsThemeActive function.

IsThemeDialogTextureEnableduxtheme and 64-bit

IsThemeDialogTextureEnabled function.

IsValidSecurityDescriptorkernel

IsValidSecurityDescriptor function.

IsValidSidkernel

IsValidSid function.

IsWellKnownSidkernel

IsWellKnownSid function.

IsWindows7OrGreaterkernel

IsWindows7OrGreater function.

IsWindows8OrGreaterkernel

IsWindows8OrGreater function.

IsWindows8Point1OrGreaterkernel

IsWindows8Point1OrGreater function.

IsWindows10OrGreaterkernel

IsWindows10OrGreater function.

IsWindowsServerkernel

IsWindowsServer function.

IsWindowsVersionOrGreaterkernel

IsWindowsVersionOrGreater function.

IsWindowsVistaOrGreaterkernel

IsWindowsVistaOrGreater function.

IsWow64Messageuser

IsWow64Message function.

LOBYTEkernel

LOBYTE macro.

LODWORDkernel

Returns the low-order u32 of an u64.

LOWORDkernel

LOWORD macro.

LockSetForegroundWindowuser

LockSetForegroundWindow function.

LookupAccountNamekernel

LookupAccountName function.

LookupAccountSidkernel

LookupAccountSid function.

LookupPrivilegeNamekernel

LookupPrivilegeName function.

LookupPrivilegeValuekernel

LookupPrivilegeValue function.

MAKEDWORDkernel

Function analog to MAKELONG, MAKEWPARAM, and MAKELPARAM macros.

MAKEQWORDkernel

Similar to MAKEDWORD, but for u64.

MAKEWORDkernel

MAKEWORD macro.

MFCreateAsyncResultmf

MFCreateAsyncResult function.

MFCreateMediaSessionmf

MFCreateMediaSession function.

MFCreateSourceResolvermf

MFCreateSourceResolver function.

MFCreateTopologymf

MFCreateTopology function.

MFCreateTopologyNodemf

MFCreateTopologyNode function.

MFStartupmf

MFStartup function.

MoveFilekernel

MoveFile function.

MulDivkernel

MulDiv function.

MultiByteToWideCharkernel

MultiByteToWideChar function.

OffsetRectuser

OffsetRect function.

OleLoadPictureoleaut

OleLoadPicture function.

OleLoadPicturePatholeaut

OleLoadPicturePath function.

OutputDebugStringkernel

OutputDebugString function.

PSGetNameFromPropertyKeyoleaut

PSGetNameFromPropertyKey function.

PathCombineshell

PathCombine function.

PathCommonPrefixshell

PathCommonPrefix function.

PathSkipRootshell

PathSkipRoot function.

PathStripPathshell

PathStripPath function.

PathUndecorateshell

PathUndecorate function.

PathUnquoteSpacesshell

PathUnquoteSpaces function.

PeekMessageuser

PeekMessage function.

PostQuitMessageuser

PostQuitMessage function.

PostThreadMessage⚠user

PostThreadMessage function.

PtInRectuser

PtInRect function.

QueryPerformanceCounterkernel

QueryPerformanceCounter function.

QueryPerformanceFrequencykernel

QueryPerformanceFrequency function.

QueryUnbiasedInterruptTimekernel

QueryUnbiasedInterruptTime function.

QueryUnbiasedInterruptTimePrecisekernel

QueryUnbiasedInterruptTimePrecise function.

RegisterClassEx⚠user

RegisterClassEx function.

RegisterWindowMessageuser

RegisterWindowMessage function.

ReplaceFilekernel

ReplaceFile function.

SHAddToRecentDocs⚠shell

SHAddToRecentDocs function.

SHCreateItemFromParsingNameshell

SHCreateItemFromParsingName function.

SHCreateMemStreamshell

SHCreateMemStream function.

SHFileOperationshell

SHFileOperation function.

SHGetFileInfoshell

SHGetFileInfo function.

SHGetKnownFolderPathshell

SHGetKnownFolderPath function.

SHGetStockIconInfoshell

SHGetStockIconInfo function.

SendInputuser

SendInput function.

SetCaretBlinkTimeuser

SetCaretBlinkTime function.

SetCaretPosuser

SetCaretPos function.

SetClipboardData⚠user

SetClipboardData function.

SetCurrentDirectorykernel

SetCurrentDirectory function.

SetCursorPosuser

SetCursorPos function.

SetDoubleClickTimeuser

SetDoubleClickTime function.

SetFileAttributeskernel

SetFileAttributes function.

SetLastErrorkernel

SetLastError function.

SetProcessDPIAwareuser

SetProcessDPIAware function.

SetThreadDpiHostingBehavioruser

SetThreadDpiHostingBehavior function.

SetThreadStackGuaranteekernel

SetThreadStackGuarantee function.

ShellExecuteExshell

ShellExecuteEx function.

Shell_NotifyIconshell

Shell_NotifyIcon function.

ShowCursoruser

ShowCursor function.

Sleepkernel

Sleep function.

SoundSentryuser

SoundSentry function.

StringFromCLSIDole

StringFromCLSID function.

SubtractRectuser

SubtractRect function.

SwapMouseButtonuser

SwapMouseButton function.

SwitchToThreadkernel

SwitchToThread function.

SystemParametersInfo⚠user

SystemParametersInfo function.

SystemTimeToFileTimekernel

SystemTimeToFileTime function.

SystemTimeToTzSpecificLocalTimekernel

SystemTimeToTzSpecificLocalTime function.

SystemTimeToVariantTimeoleaut

SystemTimeToVariantTime function.

TaskDialogIndirectcomctl

TaskDialogIndirect function.

TrackMouseEventuser

TrackMouseEvent function.

TranslateMessageuser

TranslateMessage function.

UnionRectuser

UnionRect function.

UnregisterClassuser

UnregisterClass function.

VariantTimeToSystemTimeoleaut

VariantTimeToSystemTime function.

VerSetConditionMaskkernel

VerSetConditionMask function.

VerifyVersionInfokernel

VerifyVersionInfo function.

WaitMessageuser

WaitMessage function.

WideCharToMultiBytekernel

WideCharToMultiByte function.

WritePrivateProfileStringkernel

WritePrivateProfileString function.

Type Aliases

AnyResultkernel

A Result alias which returns a Box<dyn Error + Send + Sync> on failure.

CCHOOKPROCuser

Type alias to CCHOOKPROC callback function.

DLGPROCuser

Type alias to DLGPROC callback function.

EDITWORDBREAKPROCuser

Type alias to EDITWORDBREAKPROC callback function.

HOOKPROCuser

Type alias to HOOKPROC callback function.

HrResultole

A Result alias for COM error codes, which returns an HRESULT on failure.

PFNLVCOMPAREcomctl

Type alias to PFNLVCOMPARE callback function.

PFNLVGROUPCOMPAREcomctl

Type alias to PFNLVGROUPCOMPARE callback function.

PFNTVCOMPAREcomctl

Type alias to PFNTVCOMPARE callback function.

PFTASKDIALOGCALLBACKcomctl

Type alias to PFTASKDIALOGCALLBACK calback function.

SUBCLASSPROCcomctl

Type alias to SUBCLASSPROC callback function.

SysResultkernel

A Result alias for native system error codes, which returns an ERROR on failure.

TIMERPROCuser

Type alias to TIMERPROC callback function.

WNDPROCuser

Type alias to WNDPROC callback function.