winsafe::co

Struct APPCOMMAND

pub struct APPCOMMAND(/* private fields */);

Available on crate feature user only.

wm::AppCommand commands (u16).

Implementations

impl APPCOMMAND

pub const unsafe fn as_mut(&mut self) -> &mut u16

Returns a mutable reference to the underlying raw value.

Safety

Be sure the integer being set is meaningful for the actual type.

pub const unsafe fn from_raw(v: u16) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u16

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl APPCOMMAND

pub const BROWSER_BACKWARD: Self

pub const BROWSER_FORWARD: Self

pub const BROWSER_REFRESH: Self

pub const BROWSER_STOP: Self

pub const BROWSER_SEARCH: Self

pub const BROWSER_FAVORITES: Self

pub const BROWSER_HOME: Self

pub const VOLUME_MUTE: Self

pub const VOLUME_DOWN: Self

pub const VOLUME_UP: Self

pub const MEDIA_NEXTTRACK: Self

pub const MEDIA_PREVIOUSTRACK: Self

pub const MEDIA_STOP: Self

pub const MEDIA_PLAY_PAUSE: Self

pub const LAUNCH_MAIL: Self

pub const LAUNCH_MEDIA_SELECT: Self

pub const LAUNCH_APP1: Self

pub const LAUNCH_APP2: Self

pub const BASS_DOWN: Self

pub const BASS_BOOST: Self

pub const BASS_UP: Self

pub const TREBLE_DOWN: Self

pub const TREBLE_UP: Self

pub const MICROPHONE_VOLUME_MUTE: Self

pub const MICROPHONE_VOLUME_DOWN: Self

pub const MICROPHONE_VOLUME_UP: Self

pub const HELP: Self

pub const FIND: Self

pub const NEW: Self

pub const OPEN: Self

pub const CLOSE: Self

pub const SAVE: Self

pub const PRINT: Self

pub const UNDO: Self

pub const REDO: Self

pub const COPY: Self

pub const CUT: Self

pub const PASTE: Self

pub const REPLY_TO_MAIL: Self

pub const FORWARD_MAIL: Self

pub const SEND_MAIL: Self

pub const SPELL_CHECK: Self

pub const DICTATE_OR_COMMAND_CONTROL_TOGGLE: Self

pub const MIC_ON_OFF_TOGGLE: Self

pub const CORRECTION_LIST: Self

pub const MEDIA_PLAY: Self

pub const MEDIA_PAUSE: Self

pub const MEDIA_RECORD: Self

pub const MEDIA_FAST_FORWARD: Self

pub const MEDIA_REWIND: Self

pub const MEDIA_CHANNEL_UP: Self

pub const MEDIA_CHANNEL_DOWN: Self

pub const DELETE: Self

pub const DWM_FLIP3D: Self