winsafe::co

Struct CDERR

pub struct CDERR(/* private fields */);

Available on crate feature user only.

Common dialog box error codes.

Also includes PDERR, CFERR, FNERR and FRERR prefixes.

Implements the standard Error trait.

Does not implement SystemError because FormatMessage function does not offer support for it, so there is no way to obtain a textual description of the error codes.

Implementations

impl CDERR

pub const unsafe fn as_mut(&mut self) -> &mut u32

Returns a mutable reference to the underlying raw value.

Safety

Be sure the integer being set is meaningful for the actual type.

pub const unsafe fn from_raw(v: u32) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u32

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl CDERR

pub const NoValue: Self

None of the actual values (zero).

pub const DIALOGFAILURE: Self

pub const FINDRESFAILURE: Self

pub const INITIALIZATION: Self

pub const LOADRESFAILURE: Self

pub const LOADSTRFAILURE: Self

pub const LOCKRESFAILURE: Self

pub const MEMALLOCFAILURE: Self

pub const MEMLOCKFAILURE: Self

pub const NOHINSTANCE: Self

pub const NOHOOK: Self

pub const NOTEMPLATE: Self

pub const REGISTERMSGFAIL: Self

pub const STRUCTSIZE: Self

pub const PD_CREATEICFAILURE: Self

pub const PD_DEFAULTDIFFERENT: Self

pub const PD_DNDMMISMATCH: Self

pub const PD_GETDEVMODEFAIL: Self

pub const PD_INITFAILURE: Self

pub const PD_LOADDRVFAILURE: Self

pub const PD_NODEFAULTPRN: Self

pub const PD_NODEVICES: Self

pub const PD_PARSEFAILURE: Self

pub const PD_PRINTERNOTFOUND: Self

pub const PD_RETDEFFAILURE: Self

pub const PD_SETUPFAILURE: Self

pub const CF_MAXLESSTHANMIN: Self

pub const CF_NOFONTS: Self

pub const FN_BUFFERTOOSMALL: Self

pub const FN_INVALIDFILENAME: Self

pub const FN_SUBCLASSFAILURE: Self

pub const FR_BUFFERLENGTHZERO: Self