winsafe::co

Struct DMPAPER

pub struct DMPAPER(/* private fields */);

Available on crate feature user only.

DEVMODE dmPaperSize (i16).

Implementations

impl DMPAPER

pub const unsafe fn as_mut(&mut self) -> &mut i16

Returns a mutable reference to the underlying raw value.

Safety

Be sure the integer being set is meaningful for the actual type.

pub const unsafe fn from_raw(v: i16) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> i16

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl DMPAPER

pub const NoValue: Self

None of the actual values (zero).

pub const LETTER: Self

Letter 8 1/2 x 11 in.

pub const LETTERSMALL: Self

Letter Small 8 1/2 x 11 in.

pub const TABLOID: Self

Tabloid 11 x 17 in.

pub const LEDGER: Self

Ledger 17 x 11 in.

pub const LEGAL: Self

Legal 8 1/2 x 14 in.

pub const STATEMENT: Self

Statement 5 1/2 x 8 1/2 in.

pub const EXECUTIVE: Self

Executive 7 1/4 x 10 1/2 in.

pub const A3: Self

A3 297 x 420 mm.

pub const A4: Self

A4 210 x 297 mm.

pub const A4SMALL: Self

A4 Small 210 x 297 mm.

pub const A5: Self

A5 148 x 210 mm.

pub const B4: Self

B4 (JIS) 250 x 354.

pub const B5: Self

B5 (JIS) 182 x 257 mm.

pub const FOLIO: Self

Folio 8 1/2 x 13 in.

pub const QUARTO: Self

Quarto 215 x 275 mm.

pub const P10X14: Self

10x14 in.

pub const P11X17: Self

11x17 in.

pub const NOTE: Self

Note 8 1/2 x 11 in.

pub const ENV_9: Self

Envelope #9 3 7/8 x 8 7/8.

pub const ENV_10: Self

Envelope #10 4 1/8 x 9 1/2.

pub const ENV_11: Self

Envelope #11 4 1/2 x 10 3/8.

pub const ENV_12: Self

Envelope #12 4 \276 x 11.

pub const ENV_14: Self

Envelope #14 5 x 11 1/2.

pub const CSHEET: Self

C size sheet.

pub const DSHEET: Self

D size sheet.

pub const ESHEET: Self

E size sheet.

pub const ENV_DL: Self

Envelope DL 110 x 220mm.

pub const ENV_C5: Self

Envelope C5 162 x 229 mm.

pub const ENV_C3: Self

Envelope C3 324 x 458 mm.

pub const ENV_C4: Self

Envelope C4 229 x 324 mm.

pub const ENV_C6: Self

Envelope C6 114 x 162 mm.

pub const ENV_C65: Self

Envelope C65 114 x 229 mm.

pub const ENV_B4: Self

Envelope B4 250 x 353 mm.

pub const ENV_B5: Self

Envelope B5 176 x 250 mm.

pub const ENV_B6: Self

Envelope B6 176 x 125 mm.

pub const ENV_ITALY: Self

Envelope 110 x 230 mm.

pub const ENV_MONARCH: Self

Envelope Monarch 3.875 x 7.5 in.

pub const ENV_PERSONAL: Self

6 3/4 Envelope 3 5/8 x 6 1/2 in.

pub const FANFOLD_US: Self

US Std Fanfold 14 7/8 x 11 in.

pub const FANFOLD_STD_GERMAN: Self

German Std Fanfold 8 1/2 x 12 in.

pub const FANFOLD_LGL_GERMAN: Self

German Legal Fanfold 8 1/2 x 13 in.

pub const ISO_B4: Self

B4 (ISO) 250 x 353 mm.

pub const JAPANESE_POSTCARD: Self

Japanese Postcard 100 x 148 mm.

pub const P9X11: Self

9 x 11 in.

pub const P10X11: Self

10 x 11 in.

pub const P15X11: Self

15 x 11 in.

pub const ENV_INVITE: Self

Envelope Invite 220 x 220 mm.

pub const LETTER_EXTRA: Self

Letter Extra 9 275 x 12 in.

pub const LEGAL_EXTRA: Self

Legal Extra 9 275 x 15 in.

pub const TABLOID_EXTRA: Self

Tabloid Extra 11.69 x 18 in.

pub const A4_EXTRA: Self

A4 Extra 9.27 x 12.69 in.

pub const LETTER_TRANSVERSE: Self

Letter Transverse 8 275 x 11 in.

pub const A4_TRANSVERSE: Self

A4 Transverse 210 x 297 mm.

pub const LETTER_EXTRA_TRANSVERSE: Self

Letter Extra Transverse 9\275 x 12 in.

pub const A_PLUS: Self

SuperA/SuperA/A4 227 x 356 mm.

pub const B_PLUS: Self

SuperB/SuperB/A3 305 x 487 mm.

pub const LETTER_PLUS: Self

Letter Plus 8.5 x 12.69 in.

pub const A4_PLUS: Self

A4 Plus 210 x 330 mm.

pub const A5_TRANSVERSE: Self

A5 Transverse 148 x 210 mm.

pub const B5_TRANSVERSE: Self

B5 (JIS) Transverse 182 x 257 mm.

pub const A3_EXTRA: Self

A3 Extra 322 x 445 mm.

pub const A5_EXTRA: Self

A5 Extra 174 x 235 mm.

pub const B5_EXTRA: Self

B5 (ISO) Extra 201 x 276 mm.

pub const A2: Self

A2 420 x 594 mm.

pub const A3_TRANSVERSE: Self

A3 Transverse 297 x 420 mm.

pub const A3_EXTRA_TRANSVERSE: Self

A3 Extra Transverse 322 x 445 mm.

pub const DBL_JAPANESE_POSTCARD: Self

Japanese Double Postcard 200 x 148 mm.

pub const A6: Self

A6 105 x 148 mm.

pub const JENV_KAKU2: Self

Japanese Envelope Kaku #2.

pub const JENV_KAKU3: Self

Japanese Envelope Kaku #3.

pub const JENV_CHOU3: Self

Japanese Envelope Chou #3.

pub const JENV_CHOU4: Self

Japanese Envelope Chou #4.

pub const LETTER_ROTATED: Self

Letter Rotated 11 x 8 1/2 11 in.

pub const A3_ROTATED: Self

A3 Rotated 420 x 297 mm.

pub const A4_ROTATED: Self

A4 Rotated 297 x 210 mm.

pub const A5_ROTATED: Self

A5 Rotated 210 x 148 mm.

pub const B4_JIS_ROTATED: Self

B4 (JIS) Rotated 364 x 257 mm.

pub const B5_JIS_ROTATED: Self

B5 (JIS) Rotated 257 x 182 mm.

pub const JAPANESE_POSTCARD_ROTATED: Self

Japanese Postcard Rotated 148 x 100 mm.

pub const DBL_JAPANESE_POSTCARD_ROTATED: Self

Double Japanese Postcard Rotated 148 x 200 mm.

pub const A6_ROTATED: Self

A6 Rotated 148 x 105 mm.

pub const JENV_KAKU2_ROTATED: Self

Japanese Envelope Kaku #2 Rotated.

pub const JENV_KAKU3_ROTATED: Self

Japanese Envelope Kaku #3 Rotated.

pub const JENV_CHOU3_ROTATED: Self

Japanese Envelope Chou #3 Rotated.

pub const JENV_CHOU4_ROTATED: Self

Japanese Envelope Chou #4 Rotated.

pub const B6_JIS: Self

B6 (JIS) 128 x 182 mm.

pub const B6_JIS_ROTATED: Self

B6 (JIS) Rotated 182 x 128 mm.

pub const P12X11: Self

12 x 11 in.

pub const JENV_YOU4: Self

Japanese Envelope You #4.

pub const JENV_YOU4_ROTATED: Self

Japanese Envelope You #4 Rotated.

pub const P16K: Self

PRC 16K 146 x 215 mm.

pub const P32K: Self

PRC 32K 97 x 151 mm.

pub const P32KBIG: Self

PRC 32K (Big) 97 x 151 mm.

pub const PENV_1: Self

PRC Envelope #1 102 x 165 mm.

pub const PENV_2: Self

PRC Envelope #2 102 x 176 mm.

pub const PENV_3: Self

PRC Envelope #3 125 x 176 mm.

pub const PENV_4: Self

PRC Envelope #4 110 x 208 mm.

pub const PENV_5: Self

PRC Envelope #5 110 x 220 mm.

pub const PENV_6: Self

PRC Envelope #6 120 x 230 mm.

pub const PENV_7: Self

PRC Envelope #7 160 x 230 mm.

pub const PENV_8: Self

PRC Envelope #8 120 x 309 mm.

pub const PENV_9: Self

PRC Envelope #9 229 x 324 mm.

pub const PENV_10: Self

PRC Envelope #10 324 x 458 mm.

pub const P16K_ROTATED: Self

PRC 16K Rotated.

pub const P32K_ROTATED: Self

PRC 32K Rotated.

pub const P32KBIG_ROTATED: Self

PRC 32K(Big) Rotated.

pub const PENV_1_ROTATED: Self

PRC Envelope #1 Rotated 165 x 102 mm.

pub const PENV_2_ROTATED: Self

PRC Envelope #2 Rotated 176 x 102 mm.

pub const PENV_3_ROTATED: Self

PRC Envelope #3 Rotated 176 x 125 mm.

pub const PENV_4_ROTATED: Self

PRC Envelope #4 Rotated 208 x 110 mm.

pub const PENV_5_ROTATED: Self

PRC Envelope #5 Rotated 220 x 110 mm.

pub const PENV_6_ROTATED: Self

PRC Envelope #6 Rotated 230 x 120 mm.

pub const PENV_7_ROTATED: Self

PRC Envelope #7 Rotated 230 x 160 mm.

pub const PENV_8_ROTATED: Self

PRC Envelope #8 Rotated 309 x 120 mm.

pub const PENV_9_ROTATED: Self

PRC Envelope #9 Rotated 324 x 229 mm.

pub const PENV_10_ROTATED: Self

PRC Envelope #10 Rotated 458 x 324 mm.

pub const USER: Self

Other papers start here.