winsafe::co

Struct DXGI_FORMAT

pub struct DXGI_FORMAT(/* private fields */);

Available on crate feature dxgi only.

DXGI_FORMAT enumeration (u32).

Implementations

impl DXGI_FORMAT

pub const unsafe fn as_mut(&mut self) -> &mut u32

Returns a mutable reference to the underlying raw value.

Safety

Be sure the integer being set is meaningful for the actual type.

pub const unsafe fn from_raw(v: u32) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u32

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl DXGI_FORMAT

pub const UNKNOWN: Self

pub const R32G32B32A32_TYPELESS: Self

pub const R32G32B32A32_FLOAT: Self

pub const R32G32B32A32_UINT: Self

pub const R32G32B32A32_SINT: Self

pub const R32G32B32_TYPELESS: Self

pub const R32G32B32_FLOAT: Self

pub const R32G32B32_UINT: Self

pub const R32G32B32_SINT: Self

pub const R16G16B16A16_TYPELESS: Self

pub const R16G16B16A16_FLOAT: Self

pub const R16G16B16A16_UNORM: Self

pub const R16G16B16A16_UINT: Self

pub const R16G16B16A16_SNORM: Self

pub const R16G16B16A16_SINT: Self

pub const R32G32_TYPELESS: Self

pub const R32G32_FLOAT: Self

pub const R32G32_UINT: Self

pub const R32G32_SINT: Self

pub const R32G8X24_TYPELESS: Self

pub const D32_FLOAT_S8X24_UINT: Self

pub const R32_FLOAT_X8X24_TYPELESS: Self

pub const X32_TYPELESS_G8X24_UINT: Self

pub const R10G10B10A2_TYPELESS: Self

pub const R10G10B10A2_UNORM: Self

pub const R10G10B10A2_UINT: Self

pub const R11G11B10_FLOAT: Self

pub const R8G8B8A8_TYPELESS: Self

pub const R8G8B8A8_UNORM: Self

pub const R8G8B8A8_UNORM_SRGB: Self

pub const R8G8B8A8_UINT: Self

pub const R8G8B8A8_SNORM: Self

pub const R8G8B8A8_SINT: Self

pub const R16G16_TYPELESS: Self

pub const R16G16_FLOAT: Self

pub const R16G16_UNORM: Self

pub const R16G16_UINT: Self

pub const R16G16_SNORM: Self

pub const R16G16_SINT: Self

pub const R32_TYPELESS: Self

pub const D32_FLOAT: Self

pub const R32_FLOAT: Self

pub const R32_UINT: Self

pub const R32_SINT: Self

pub const R24G8_TYPELESS: Self

pub const D24_UNORM_S8_UINT: Self

pub const R24_UNORM_X8_TYPELESS: Self

pub const X24_TYPELESS_G8_UINT: Self

pub const R8G8_TYPELESS: Self

pub const R8G8_UNORM: Self

pub const R8G8_UINT: Self

pub const R8G8_SNORM: Self

pub const R8G8_SINT: Self

pub const R16_TYPELESS: Self

pub const R16_FLOAT: Self

pub const D16_UNORM: Self

pub const R16_UNORM: Self

pub const R16_UINT: Self

pub const R16_SNORM: Self

pub const R16_SINT: Self

pub const R8_TYPELESS: Self

pub const R8_UNORM: Self

pub const R8_UINT: Self

pub const R8_SNORM: Self

pub const R8_SINT: Self

pub const A8_UNORM: Self

pub const R1_UNORM: Self

pub const R9G9B9E5_SHAREDEXP: Self

pub const R8G8_B8G8_UNORM: Self

pub const G8R8_G8B8_UNORM: Self

pub const BC1_TYPELESS: Self

pub const BC1_UNORM: Self

pub const BC1_UNORM_SRGB: Self

pub const BC2_TYPELESS: Self

pub const BC2_UNORM: Self

pub const BC2_UNORM_SRGB: Self

pub const BC3_TYPELESS: Self

pub const BC3_UNORM: Self

pub const BC3_UNORM_SRGB: Self

pub const BC4_TYPELESS: Self

pub const BC4_UNORM: Self

pub const BC4_SNORM: Self

pub const BC5_TYPELESS: Self

pub const BC5_UNORM: Self

pub const BC5_SNORM: Self

pub const B5G6R5_UNORM: Self

pub const B5G5R5A1_UNORM: Self

pub const B8G8R8A8_UNORM: Self

pub const B8G8R8X8_UNORM: Self

pub const R10G10B10_XR_BIAS_A2_UNORM: Self

pub const B8G8R8A8_TYPELESS: Self

pub const B8G8R8A8_UNORM_SRGB: Self

pub const B8G8R8X8_TYPELESS: Self

pub const B8G8R8X8_UNORM_SRGB: Self

pub const BC6H_TYPELESS: Self

pub const BC6H_UF16: Self

pub const BC6H_SF16: Self

pub const BC7_TYPELESS: Self

pub const BC7_UNORM: Self

pub const BC7_UNORM_SRGB: Self

pub const AYUV: Self

pub const Y410: Self

pub const Y416: Self

pub const NV12: Self

pub const P010: Self

pub const P016: Self

pub const F420_OPAQUE: Self

pub const YUY2: Self

pub const Y210: Self

pub const Y216: Self

pub const NV11: Self

pub const AI44: Self

pub const IA44: Self

pub const P8: Self

pub const A8P8: Self

pub const B4G4R4A4_UNORM: Self

pub const P208: Self

pub const V208: Self

pub const V408: Self