Struct winsafe::co::EM

pub struct EM(/* private fields */);

Available on crate features kernel and user only.

Edit control messages (u32).

This is a window message, convertible to WM.

Implementations

impl EM

pub const SETCUEBANNER: Self = _

Available on crate feature comctl only.

pub const GETCUEBANNER: Self = _

Available on crate feature comctl only.

pub const SHOWBALLOONTIP: Self = _

Available on crate feature comctl only.

pub const HIDEBALLOONTIP: Self = _

Available on crate feature comctl only.

pub const SETHILITE: Self = _

Available on crate feature comctl only.

pub const GETHILITE: Self = _

Available on crate feature comctl only.

pub const NOSETFOCUS: Self = _

Available on crate feature comctl only.

pub const TAKEFOCUS: Self = _

Available on crate feature comctl only.

pub const SETEXTENDEDSTYLE: Self = _

Available on crate feature comctl only.

pub const GETEXTENDEDSTYLE: Self = _

Available on crate feature comctl only.

pub const SETENDOFLINE: Self = _

Available on crate feature comctl only.

pub const GETENDOFLINE: Self = _

Available on crate feature comctl only.

pub const ENABLESEARCHWEB: Self = _

Available on crate feature comctl only.

pub const SEARCHWEB: Self = _

Available on crate feature comctl only.

pub const SETCARETINDEX: Self = _

Available on crate feature comctl only.

pub const GETCARETINDEX: Self = _

Available on crate feature comctl only.

pub const GETZOOM: Self = _

Available on crate feature comctl only.

pub const SETZOOM: Self = _

Available on crate feature comctl only.

pub const FILELINEFROMCHAR: Self = _

Available on crate feature comctl only.

pub const FILELINEINDEX: Self = _

Available on crate feature comctl only.

pub const FILELINELENGTH: Self = _

Available on crate feature comctl only.

pub const GETFILELINE: Self = _

Available on crate feature comctl only.

pub const GETFILELINECOUNT: Self = _

Available on crate feature comctl only.

impl EM

pub const unsafe fn from_raw(v: u32) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u32

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl EM

pub const GETSEL: Self = _

pub const SETSEL: Self = _

pub const GETRECT: Self = _

pub const SETRECT: Self = _

pub const SETRECTNP: Self = _

pub const SCROLL: Self = _

pub const LINESCROLL: Self = _

pub const SCROLLCARET: Self = _

pub const GETMODIFY: Self = _

pub const SETMODIFY: Self = _

pub const GETLINECOUNT: Self = _

pub const LINEINDEX: Self = _

pub const SETHANDLE: Self = _

pub const GETHANDLE: Self = _

pub const GETTHUMB: Self = _

pub const LINELENGTH: Self = _

pub const REPLACESEL: Self = _

pub const GETLINE: Self = _

pub const LIMITTEXT: Self = _

pub const CANUNDO: Self = _

pub const UNDO: Self = _

pub const FMTLINES: Self = _

pub const LINEFROMCHAR: Self = _

pub const SETTABSTOPS: Self = _

pub const SETPASSWORDCHAR: Self = _

pub const EMPTYUNDOBUFFER: Self = _

pub const GETFIRSTVISIBLELINE: Self = _

pub const SETREADONLY: Self = _

pub const SETWORDBREAKPROC: Self = _

pub const GETWORDBREAKPROC: Self = _

pub const GETPASSWORDCHAR: Self = _

pub const SETMARGINS: Self = _

pub const GETMARGINS: Self = _

pub const SETLIMITTEXT: Self = _

pub const GETLIMITTEXT: Self = _

pub const POSFROMCHAR: Self = _

pub const CHARFROMPOS: Self = _

pub const SETIMESTATUS: Self = _

pub const GETIMESTATUS: Self = _

pub const ENABLEFEATURE: Self = _