Struct winsafe::co::FACILITY

pub struct FACILITY(/* private fields */);

Available on crate features kernel and ole only.

HRESULT facility (u32).

Implementations

impl FACILITY

pub const unsafe fn from_raw(v: u32) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u32

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl FACILITY

pub const NULL: Self = _

pub const RPC: Self = _

pub const DISPATCH: Self = _

pub const STORAGE: Self = _

pub const ITF: Self = _

pub const WIN32: Self = _

pub const WINDOWS: Self = _

pub const SSPI: Self = _

pub const SECURITY: Self = _

pub const CONTROL: Self = _

pub const CERT: Self = _

pub const INTERNET: Self = _

pub const MEDIASERVER: Self = _

pub const MSMQ: Self = _

pub const SETUPAPI: Self = _

pub const SCARD: Self = _

pub const COMPLUS: Self = _

pub const AAF: Self = _

pub const URT: Self = _

pub const ACS: Self = _

pub const DPLAY: Self = _

pub const UMI: Self = _

pub const SXS: Self = _

pub const WINDOWS_CE: Self = _

pub const HTTP: Self = _

pub const USERMODE_COMMONLOG: Self = _

pub const WER: Self = _

pub const USERMODE_FILTER_MANAGER: Self = _

pub const BACKGROUNDCOPY: Self = _

pub const CONFIGURATION: Self = _

pub const WIA: Self = _

pub const STATE_MANAGEMENT: Self = _

pub const METADIRECTORY: Self = _

pub const WINDOWSUPDATE: Self = _

pub const DIRECTORYSERVICE: Self = _

pub const GRAPHICS: Self = _

pub const SHELL: Self = _

pub const NAP: Self = _

pub const TPM_SERVICES: Self = _

pub const TPM_SOFTWARE: Self = _

pub const UI: Self = _

pub const XAML: Self = _

pub const ACTION_QUEUE: Self = _

pub const PLA: Self = _

pub const WINDOWS_SETUP: Self = _

pub const FVE: Self = _

pub const FWP: Self = _

pub const WINRM: Self = _

pub const NDIS: Self = _

pub const USERMODE_HYPERVISOR: Self = _

pub const CMI: Self = _

pub const USERMODE_VIRTUALIZATION: Self = _

pub const USERMODE_VOLMGR: Self = _

pub const BCD: Self = _

pub const USERMODE_VHD: Self = _

pub const USERMODE_HNS: Self = _

pub const SDIAG: Self = _

pub const WEBSERVICES: Self = _

pub const WINPE: Self = _

pub const WPN: Self = _

pub const WINDOWS_STORE: Self = _

pub const INPUT: Self = _

pub const EAP: Self = _

pub const WINDOWS_DEFENDER: Self = _

pub const OPC: Self = _

pub const XPS: Self = _

pub const MBN: Self = _

pub const POWERSHELL: Self = _

pub const RAS: Self = _

pub const P2P_INT: Self = _

pub const P2P: Self = _

pub const DAF: Self = _

pub const BLUETOOTH_ATT: Self = _

pub const AUDIO: Self = _

pub const STATEREPOSITORY: Self = _

pub const VISUALCPP: Self = _

pub const SCRIPT: Self = _

pub const PARSE: Self = _

pub const BLB: Self = _

pub const BLB_CLI: Self = _

pub const WSBAPP: Self = _

pub const BLBUI: Self = _

pub const USN: Self = _

pub const USERMODE_VOLSNAP: Self = _

pub const TIERING: Self = _

pub const WSB_ONLINE: Self = _

pub const ONLINE_ID: Self = _

pub const DEVICE_UPDATE_AGENT: Self = _

pub const DRVSERVICING: Self = _

pub const DLS: Self = _

pub const DELIVERY_OPTIMIZATION: Self = _

pub const USERMODE_SPACES: Self = _

pub const USER_MODE_SECURITY_CORE: Self = _

pub const USERMODE_LICENSING: Self = _

pub const SOS: Self = _

pub const DEBUGGERS: Self = _

pub const SPP: Self = _

pub const RESTORE: Self = _

pub const DMSERVER: Self = _

pub const DEPLOYMENT_SERVICES_SERVER: Self = _

pub const DEPLOYMENT_SERVICES_IMAGING: Self = _

pub const DEPLOYMENT_SERVICES_MANAGEMENT: Self = _

pub const DEPLOYMENT_SERVICES_UTIL: Self = _

pub const DEPLOYMENT_SERVICES_BINLSVC: Self = _

pub const DEPLOYMENT_SERVICES_PXE: Self = _

pub const DEPLOYMENT_SERVICES_TFTP: Self = _

pub const DEPLOYMENT_SERVICES_TRANSPORT_MANAGEMENT: Self = _

pub const DEPLOYMENT_SERVICES_DRIVER_PROVISIONING: Self = _

pub const DEPLOYMENT_SERVICES_MULTICAST_SERVER: Self = _

pub const DEPLOYMENT_SERVICES_MULTICAST_CLIENT: Self = _

pub const DEPLOYMENT_SERVICES_CONTENT_PROVIDER: Self = _

pub const LINGUISTIC_SERVICES: Self = _

pub const AUDIOSTREAMING: Self = _

pub const ACCELERATOR: Self = _

pub const WMAAECMA: Self = _

pub const DIRECTMUSIC: Self = _

pub const DIRECT3D10: Self = _

pub const DXGI: Self = _

pub const DXGI_DDI: Self = _

pub const DIRECT3D11: Self = _

pub const DIRECT3D11_DEBUG: Self = _

pub const DIRECT3D12: Self = _

pub const DIRECT3D12_DEBUG: Self = _

pub const LEAP: Self = _

pub const AUDCLNT: Self = _

pub const WINCODEC_DWRITE_DWM: Self = _

pub const WINML: Self = _

pub const DIRECT2D: Self = _

pub const DEFRAG: Self = _

pub const USERMODE_SDBUS: Self = _

pub const JSCRIPT: Self = _

pub const PIDGENX: Self = _

pub const EAS: Self = _

pub const WEB: Self = _

pub const WEB_SOCKET: Self = _

pub const MOBILE: Self = _

pub const SQLITE: Self = _

pub const UTC: Self = _

pub const WEP: Self = _

pub const SYNCENGINE: Self = _

pub const XBOX: Self = _

pub const GAME: Self = _

pub const PIX: Self = _