Struct winsafe::co::FOS

#[repr(C)]
pub struct FOS(_);

_FILEOPENDIALOGOPTIONS enumeration (u32).

Implementations

impl FOS

pub fn has(&self, other: FOS) -> bool

Tells whether other bitflag style is present. Equivalent to (val & other) != 0.

This method is common to all constant types.

impl FOS

pub const OVERWRITEPROMPT: Self

When saving a file, prompt before overwriting an existing file of the same name. This is a default value for the Save dialog.

pub const STRICTFILETYPES: Self

In the Save dialog, only allow the user to choose a file that has one of the file name extensions specified through IFileDialog::SetFileTypes.

pub const NOCHANGEDIR: Self

Don’t change the current working directory.

pub const PICKFOLDERS: Self

Present an Open dialog that offers a choice of folders rather than files.

pub const FORCEFILESYSTEM: Self

Ensures that returned items are file system items (SFGAO::FILESYSTEM). Note that this does not apply to items returned by IFileDialog::GetCurrentSelection.

pub const ALLNONSTORAGEITEMS: Self

Enables the user to choose any item in the Shell namespace, not just those with SFGAO::STREAM or SFAGO::FILESYSTEM attributes. This flag cannot be combined with FOS::FORCEFILESYSTEM.

pub const NOVALIDATE: Self

Do not check for situations that would prevent an application from opening the selected file, such as sharing violations or access denied errors.

pub const ALLOWMULTISELECT: Self

Enables the user to select multiple items in the open dialog. Note that when this flag is set, the IFileOpenDialog interface must be used to retrieve those items.

pub const PATHMUSTEXIST: Self

The item returned must be in an existing folder. This is a default value.

pub const FILEMUSTEXIST: Self

The item returned must exist. This is a default value for the Open dialog.

pub const CREATEPROMPT: Self

Prompt for creation if the item returned in the save dialog does not exist. Note that this does not actually create the item.

pub const SHAREAWARE: Self

In the case of a sharing violation when an application is opening a file, call the application back through OnShareViolation for guidance. This flag is overridden by FOS::NOVALIDATE.

pub const NOREADONLYRETURN: Self

Do not return read-only items. This is a default value for the Save dialog.

pub const NOTESTFILECREATE: Self

Do not test whether creation of the item as specified in the Save dialog will be successful. If this flag is not set, the calling application must handle errors, such as denial of access, discovered when the item is created.

pub const HIDEMRUPLACES: Self

Hide the list of places from which the user has recently opened or saved items. This value is not supported as of Windows 7.

pub const HIDEPINNEDPLACES: Self

Hide items shown by default in the view’s navigation pane. This flag is often used in conjunction with the IFileDialog::AddPlace method, to hide standard locations and replace them with custom locations.

Windows 7 and later. Hide all of the standard namespace locations (such as Favorites, Libraries, Computer, and Network) shown in the navigation pane.

Windows Vista. Hide the contents of the Favorite Links tree in the navigation pane. Note that the category itself is still displayed, but shown as empty.

pub const NODEREFERENCELINKS: Self

Shortcuts should not be treated as their target items. This allows an application to open a .lnk file rather than what that file is a shortcut to.

pub const OKBUTTONNEEDSINTERACTION: Self

(This constant has no official documentation.)

pub const DONTADDTORECENT: Self

Do not add the item being opened or saved to the recent documents list (SHAddToRecentDocs).

pub const FORCESHOWHIDDEN: Self

Include hidden and system items.

pub const DEFAULTNOMINIMODE: Self

Indicates to the Save As dialog box that it should open in expanded mode. Expanded mode is the mode that is set and unset by clicking the button in the lower-left corner of the Save As dialog box that switches between Browse Folders and Hide Folders when clicked. This value is not supported as of Windows 7.

pub const FORCEPREVIEWPANEON: Self

Indicates to the Open dialog box that the preview pane should always be displayed.

pub const SUPPORTSTREAMABLEITEMS: Self

Indicates that the caller is opening a file as a stream (BHID_Stream), so there is no need to download that file.