winsafe::co

Struct HRESULT

pub struct HRESULT(/* private fields */);

Available on crate feature ole only.

A native COM error code (u32), which can be considered a superset of ERROR.

The Result alias for this type is HrResult.

Implements the standard Error trait.

Implements the Debug and Display traits to show the error code along with the error description, taken from SystemError trait. For example, the code below:

use winsafe::{self as w, prelude::*, co};

println!("{}", co::HRESULT::E_INVALIDARG);

Will print:

[0x80070057 2147942487] The parameter is incorrect.

The numbers inside the brackets are the system error code, in hexadecimal and decimal formats for convenience.

An ERROR can be seamlessly converted into an HRESULT. This operation is equivalent to the HRESULT_FROM_WIN32 macro:

use winsafe::{self as w, prelude::*, co};

let err = co::ERROR::FILE_NOT_FOUND;

let hr = err.to_hresult();

Implementations

impl HRESULT

pub const VFW_E_INVALIDMEDIATYPE: Self

Available on crate feature dshow only.

pub const VFW_E_INVALIDSUBTYPE: Self

Available on crate feature dshow only.

pub const VFW_E_NEED_OWNER: Self

Available on crate feature dshow only.

pub const VFW_E_ENUM_OUT_OF_SYNC: Self

Available on crate feature dshow only.

pub const VFW_E_ALREADY_CONNECTED: Self

Available on crate feature dshow only.

pub const VFW_E_FILTER_ACTIVE: Self

Available on crate feature dshow only.

pub const VFW_E_NO_TYPES: Self

Available on crate feature dshow only.

pub const VFW_E_NO_ACCEPTABLE_TYPES: Self

Available on crate feature dshow only.

pub const VFW_E_INVALID_DIRECTION: Self

Available on crate feature dshow only.

pub const VFW_E_NOT_CONNECTED: Self

Available on crate feature dshow only.

pub const VFW_E_NO_ALLOCATOR: Self

Available on crate feature dshow only.

pub const VFW_E_RUNTIME_ERROR: Self

Available on crate feature dshow only.

pub const VFW_E_BUFFER_NOTSET: Self

Available on crate feature dshow only.

pub const VFW_E_BUFFER_OVERFLOW: Self

Available on crate feature dshow only.

pub const VFW_E_BADALIGN: Self

Available on crate feature dshow only.

pub const VFW_E_ALREADY_COMMITTED: Self

Available on crate feature dshow only.

pub const VFW_E_BUFFERS_OUTSTANDING: Self

Available on crate feature dshow only.

pub const VFW_E_NOT_COMMITTED: Self

Available on crate feature dshow only.

pub const VFW_E_SIZENOTSET: Self

Available on crate feature dshow only.

pub const VFW_E_NO_CLOCK: Self

Available on crate feature dshow only.

pub const VFW_E_NO_SINK: Self

Available on crate feature dshow only.

pub const VFW_E_NO_INTERFACE: Self

Available on crate feature dshow only.

pub const VFW_E_NOT_FOUND: Self

Available on crate feature dshow only.

pub const VFW_E_CANNOT_CONNECT: Self

Available on crate feature dshow only.

pub const VFW_E_CANNOT_RENDER: Self

Available on crate feature dshow only.

pub const VFW_E_CHANGING_FORMAT: Self

Available on crate feature dshow only.

pub const VFW_E_NO_COLOR_KEY_SET: Self

Available on crate feature dshow only.

pub const VFW_E_NOT_OVERLAY_CONNECTION: Self

Available on crate feature dshow only.

pub const VFW_E_NOT_SAMPLE_CONNECTION: Self

Available on crate feature dshow only.

pub const VFW_E_PALETTE_SET: Self

Available on crate feature dshow only.

pub const VFW_E_COLOR_KEY_SET: Self

Available on crate feature dshow only.

pub const VFW_E_NO_COLOR_KEY_FOUND: Self

Available on crate feature dshow only.

pub const VFW_E_NO_PALETTE_AVAILABLE: Self

Available on crate feature dshow only.

pub const VFW_E_NO_DISPLAY_PALETTE: Self

Available on crate feature dshow only.

pub const VFW_E_TOO_MANY_COLORS: Self

Available on crate feature dshow only.

pub const VFW_E_STATE_CHANGED: Self

Available on crate feature dshow only.

pub const VFW_E_NOT_STOPPED: Self

Available on crate feature dshow only.

pub const VFW_E_NOT_PAUSED: Self

Available on crate feature dshow only.

pub const VFW_E_NOT_RUNNING: Self

Available on crate feature dshow only.

pub const VFW_E_WRONG_STATE: Self

Available on crate feature dshow only.

pub const VFW_E_START_TIME_AFTER_END: Self

Available on crate feature dshow only.

pub const VFW_E_INVALID_RECT: Self

Available on crate feature dshow only.

pub const VFW_E_TYPE_NOT_ACCEPTED: Self

Available on crate feature dshow only.

pub const VFW_E_SAMPLE_REJECTED: Self

Available on crate feature dshow only.

pub const VFW_E_SAMPLE_REJECTED_EOS: Self

Available on crate feature dshow only.

pub const VFW_E_DUPLICATE_NAME: Self

Available on crate feature dshow only.

pub const VFW_S_DUPLICATE_NAME: Self

Available on crate feature dshow only.

pub const VFW_E_TIMEOUT: Self

Available on crate feature dshow only.

pub const VFW_E_INVALID_FILE_FORMAT: Self

Available on crate feature dshow only.

pub const VFW_E_ENUM_OUT_OF_RANGE: Self

Available on crate feature dshow only.

pub const VFW_E_CIRCULAR_GRAPH: Self

Available on crate feature dshow only.

pub const VFW_E_NOT_ALLOWED_TO_SAVE: Self

Available on crate feature dshow only.

pub const VFW_E_TIME_ALREADY_PASSED: Self

Available on crate feature dshow only.

pub const VFW_E_ALREADY_CANCELLED: Self

Available on crate feature dshow only.

pub const VFW_E_CORRUPT_GRAPH_FILE: Self

Available on crate feature dshow only.

pub const VFW_E_ADVISE_ALREADY_SET: Self

Available on crate feature dshow only.

pub const VFW_S_STATE_INTERMEDIATE: Self

Available on crate feature dshow only.

pub const VFW_E_NO_MODEX_AVAILABLE: Self

Available on crate feature dshow only.

pub const VFW_E_NO_ADVISE_SET: Self

Available on crate feature dshow only.

pub const VFW_E_NO_FULLSCREEN: Self

Available on crate feature dshow only.

pub const VFW_E_IN_FULLSCREEN_MODE: Self

Available on crate feature dshow only.

pub const VFW_E_UNKNOWN_FILE_TYPE: Self

Available on crate feature dshow only.

pub const VFW_E_CANNOT_LOAD_SOURCE_FILTER: Self

Available on crate feature dshow only.

pub const VFW_S_PARTIAL_RENDER: Self

Available on crate feature dshow only.

pub const VFW_E_FILE_TOO_SHORT: Self

Available on crate feature dshow only.

pub const VFW_E_INVALID_FILE_VERSION: Self

Available on crate feature dshow only.

pub const VFW_S_SOME_DATA_IGNORED: Self

Available on crate feature dshow only.

pub const VFW_S_CONNECTIONS_DEFERRED: Self

Available on crate feature dshow only.

pub const VFW_E_INVALID_CLSID: Self

Available on crate feature dshow only.

pub const VFW_E_INVALID_MEDIA_TYPE: Self

Available on crate feature dshow only.

pub const VFW_E_BAD_KEY: Self

Available on crate feature dshow only.

pub const VFW_S_NO_MORE_ITEMS: Self

Available on crate feature dshow only.

pub const VFW_E_SAMPLE_TIME_NOT_SET: Self

Available on crate feature dshow only.

pub const VFW_S_RESOURCE_NOT_NEEDED: Self

Available on crate feature dshow only.

pub const VFW_E_MEDIA_TIME_NOT_SET: Self

Available on crate feature dshow only.

pub const VFW_E_NO_TIME_FORMAT_SET: Self

Available on crate feature dshow only.

pub const VFW_E_MONO_AUDIO_HW: Self

Available on crate feature dshow only.

pub const VFW_S_MEDIA_TYPE_IGNORED: Self

Available on crate feature dshow only.

pub const VFW_E_NO_DECOMPRESSOR: Self

Available on crate feature dshow only.

pub const VFW_E_NO_AUDIO_HARDWARE: Self

Available on crate feature dshow only.

pub const VFW_S_VIDEO_NOT_RENDERED: Self

Available on crate feature dshow only.

pub const VFW_S_AUDIO_NOT_RENDERED: Self

Available on crate feature dshow only.

pub const VFW_E_RPZA: Self

Available on crate feature dshow only.

pub const VFW_S_RPZA: Self

Available on crate feature dshow only.

pub const VFW_E_PROCESSOR_NOT_SUITABLE: Self

Available on crate feature dshow only.

pub const VFW_E_UNSUPPORTED_AUDIO: Self

Available on crate feature dshow only.

pub const VFW_E_UNSUPPORTED_VIDEO: Self

Available on crate feature dshow only.

pub const VFW_E_MPEG_NOT_CONSTRAINED: Self

Available on crate feature dshow only.

pub const VFW_E_NOT_IN_GRAPH: Self

Available on crate feature dshow only.

pub const VFW_S_ESTIMATED: Self

Available on crate feature dshow only.

pub const VFW_E_NO_TIME_FORMAT: Self

Available on crate feature dshow only.

pub const VFW_E_READ_ONLY: Self

Available on crate feature dshow only.

pub const VFW_S_RESERVED: Self

Available on crate feature dshow only.

pub const VFW_E_BUFFER_UNDERFLOW: Self

Available on crate feature dshow only.

pub const VFW_E_UNSUPPORTED_STREAM: Self

Available on crate feature dshow only.

pub const VFW_E_NO_TRANSPORT: Self

Available on crate feature dshow only.

pub const VFW_S_STREAM_OFF: Self

Available on crate feature dshow only.

pub const VFW_S_CANT_CUE: Self

Available on crate feature dshow only.

pub const VFW_E_BAD_VIDEOCD: Self

Available on crate feature dshow only.

pub const VFW_S_NO_STOP_TIME: Self

Available on crate feature dshow only.

pub const VFW_E_OUT_OF_VIDEO_MEMORY: Self

Available on crate feature dshow only.

pub const VFW_E_VP_NEGOTIATION_FAILED: Self

Available on crate feature dshow only.

pub const VFW_E_DDRAW_CAPS_NOT_SUITABLE: Self

Available on crate feature dshow only.

pub const VFW_E_NO_VP_HARDWARE: Self

Available on crate feature dshow only.

pub const VFW_E_NO_CAPTURE_HARDWARE: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_OPERATION_INHIBITED: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_INVALIDDOMAIN: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_NO_BUTTON: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_GRAPHNOTREADY: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_RENDERFAIL: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_DECNOTENOUGH: Self

Available on crate feature dshow only.

pub const VFW_E_DDRAW_VERSION_NOT_SUITABLE: Self

Available on crate feature dshow only.

pub const VFW_E_COPYPROT_FAILED: Self

Available on crate feature dshow only.

pub const VFW_S_NOPREVIEWPIN: Self

Available on crate feature dshow only.

pub const VFW_E_TIME_EXPIRED: Self

Available on crate feature dshow only.

pub const VFW_S_DVD_NON_ONE_SEQUENTIAL: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_WRONG_SPEED: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_MENU_DOES_NOT_EXIST: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_CMD_CANCELLED: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_STATE_WRONG_VERSION: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_STATE_CORRUPT: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_STATE_WRONG_DISC: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_INCOMPATIBLE_REGION: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_NO_ATTRIBUTES: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_NO_GOUP_PGC: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_LOW_PARENTAL_LEVEL: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_NOT_IN_KARAOKE_MODE: Self

Available on crate feature dshow only.

pub const VFW_S_DVD_CHANNEL_CONTENTS_NOT_AVAILABLE: Self

Available on crate feature dshow only.

pub const VFW_S_DVD_NOT_ACCURATE: Self

Available on crate feature dshow only.

pub const VFW_E_FRAME_STEP_UNSUPPORTED: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_STREAM_DISABLED: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_TITLE_UNKNOWN: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_INVALID_DISC: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_NO_RESUME_INFORMATION: Self

Available on crate feature dshow only.

pub const VFW_E_PIN_ALREADY_BLOCKED_ON_THIS_THREAD: Self

Available on crate feature dshow only.

pub const VFW_E_PIN_ALREADY_BLOCKED: Self

Available on crate feature dshow only.

pub const VFW_E_CERTIFICATION_FAILURE: Self

Available on crate feature dshow only.

pub const VFW_E_VMR_NOT_IN_MIXER_MODE: Self

Available on crate feature dshow only.

pub const VFW_E_VMR_NO_AP_SUPPLIED: Self

Available on crate feature dshow only.

pub const VFW_E_VMR_NO_DEINTERLACE_HW: Self

Available on crate feature dshow only.

pub const VFW_E_VMR_NO_PROCAMP_HW: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_VMR9_INCOMPATIBLEDEC: Self

Available on crate feature dshow only.

pub const VFW_E_NO_COPP_HW: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_NONBLOCKING: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_TOO_MANY_RENDERERS_IN_FILTER_GRAPH: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_NON_EVR_RENDERER_IN_FILTER_GRAPH: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_RESOLUTION_ERROR: Self

Available on crate feature dshow only.

pub const E_PROP_SET_UNSUPPORTED: Self

Available on crate feature dshow only.

pub const E_PROP_ID_UNSUPPORTED: Self

Available on crate feature dshow only.

pub const VFW_E_CODECAPI_LINEAR_RANGE: Self

Available on crate feature dshow only.

pub const VFW_E_CODECAPI_ENUMERATED: Self

Available on crate feature dshow only.

pub const VFW_E_CODECAPI_NO_DEFAULT: Self

Available on crate feature dshow only.

pub const VFW_E_CODECAPI_NO_CURRENT_VALUE: Self

Available on crate feature dshow only.

pub const VFW_E_DVD_CHAPTER_DOES_NOT_EXIST: Self

Available on crate feature dshow only.

pub const VFW_S_DVD_RENDER_STATUS: Self

Available on crate feature dshow only.

impl HRESULT

pub const DWM_E_COMPOSITIONDISABLED: Self

Available on crate feature dwm only.

pub const DWM_E_REMOTING_NOT_SUPPORTED: Self

Available on crate feature dwm only.

pub const DWM_E_NO_REDIRECTION_SURFACE_AVAILABLE: Self

Available on crate feature dwm only.

pub const DWM_E_NOT_QUEUING_PRESENTS: Self

Available on crate feature dwm only.

pub const DWM_E_ADAPTER_NOT_FOUND: Self

Available on crate feature dwm only.

pub const DWM_S_GDI_REDIRECTION_SURFACE: Self

Available on crate feature dwm only.

pub const DWM_E_TEXTURE_TOO_LARGE: Self

Available on crate feature dwm only.

pub const DWM_S_GDI_REDIRECTION_SURFACE_BLT_VIA_GDI: Self

Available on crate feature dwm only.

impl HRESULT

pub const DXGI_STATUS_OCCLUDED: Self

Available on crate feature dxgi only.

pub const DXGI_STATUS_CLIPPED: Self

Available on crate feature dxgi only.

pub const DXGI_STATUS_NO_REDIRECTION: Self

Available on crate feature dxgi only.

pub const DXGI_STATUS_NO_DESKTOP_ACCESS: Self

Available on crate feature dxgi only.

pub const DXGI_STATUS_GRAPHICS_VIDPN_SOURCE_IN_USE: Self

Available on crate feature dxgi only.

pub const DXGI_STATUS_MODE_CHANGED: Self

Available on crate feature dxgi only.

pub const DXGI_STATUS_MODE_CHANGE_IN_PROGRESS: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_INVALID_CALL: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_NOT_FOUND: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_MORE_DATA: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_UNSUPPORTED: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_DEVICE_REMOVED: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_DEVICE_HUNG: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_DEVICE_RESET: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_WAS_STILL_DRAWING: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_FRAME_STATISTICS_DISJOINT: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_GRAPHICS_VIDPN_SOURCE_IN_USE: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_DRIVER_INTERNAL_ERROR: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_NONEXCLUSIVE: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_NOT_CURRENTLY_AVAILABLE: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_REMOTE_CLIENT_DISCONNECTED: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_REMOTE_OUTOFMEMORY: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_ACCESS_LOST: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_WAIT_TIMEOUT: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_SESSION_DISCONNECTED: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_RESTRICT_TO_OUTPUT_STALE: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_CANNOT_PROTECT_CONTENT: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_ACCESS_DENIED: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_NAME_ALREADY_EXISTS: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_SDK_COMPONENT_MISSING: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_NOT_CURRENT: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_HW_PROTECTION_OUTOFMEMORY: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_DYNAMIC_CODE_POLICY_VIOLATION: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_NON_COMPOSITED_UI: Self

Available on crate feature dxgi only.

pub const DXGI_STATUS_UNOCCLUDED: Self

Available on crate feature dxgi only.

pub const DXGI_STATUS_DDA_WAS_STILL_DRAWING: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_MODE_CHANGE_IN_PROGRESS: Self

Available on crate feature dxgi only.

pub const DXGI_STATUS_PRESENT_REQUIRED: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_CACHE_CORRUPT: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_CACHE_FULL: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_CACHE_HASH_COLLISION: Self

Available on crate feature dxgi only.

pub const DXGI_ERROR_ALREADY_EXISTS: Self

Available on crate feature dxgi only.

pub const DXGI_DDI_ERR_WASSTILLDRAWING: Self

Available on crate feature dxgi only.

pub const DXGI_DDI_ERR_UNSUPPORTED: Self

Available on crate feature dxgi only.

pub const DXGI_DDI_ERR_NONEXCLUSIVE: Self

Available on crate feature dxgi only.

impl HRESULT

pub const MF_E_PLATFORM_NOT_INITIALIZED: Self

Available on crate feature mf only.

pub const MF_E_BUFFERTOOSMALL: Self

Available on crate feature mf only.

pub const MF_E_INVALIDREQUEST: Self

Available on crate feature mf only.

pub const MF_E_INVALIDSTREAMNUMBER: Self

Available on crate feature mf only.

pub const MF_E_INVALIDMEDIATYPE: Self

Available on crate feature mf only.

pub const MF_E_NOTACCEPTING: Self

Available on crate feature mf only.

pub const MF_E_NOT_INITIALIZED: Self

Available on crate feature mf only.

pub const MF_E_UNSUPPORTED_REPRESENTATION: Self

Available on crate feature mf only.

pub const MF_E_NO_MORE_TYPES: Self

Available on crate feature mf only.

pub const MF_E_UNSUPPORTED_SERVICE: Self

Available on crate feature mf only.

pub const MF_E_UNEXPECTED: Self

Available on crate feature mf only.

pub const MF_E_INVALIDNAME: Self

Available on crate feature mf only.

pub const MF_E_INVALIDTYPE: Self

Available on crate feature mf only.

pub const MF_E_INVALID_FILE_FORMAT: Self

Available on crate feature mf only.

pub const MF_E_INVALIDINDEX: Self

Available on crate feature mf only.

pub const MF_E_INVALID_TIMESTAMP: Self

Available on crate feature mf only.

pub const MF_E_UNSUPPORTED_SCHEME: Self

Available on crate feature mf only.

pub const MF_E_UNSUPPORTED_BYTESTREAM_TYPE: Self

Available on crate feature mf only.

pub const MF_E_UNSUPPORTED_TIME_FORMAT: Self

Available on crate feature mf only.

pub const MF_E_NO_SAMPLE_TIMESTAMP: Self

Available on crate feature mf only.

pub const MF_E_NO_SAMPLE_DURATION: Self

Available on crate feature mf only.

pub const MF_E_INVALID_STREAM_DATA: Self

Available on crate feature mf only.

pub const MF_E_RT_UNAVAILABLE: Self

Available on crate feature mf only.

pub const MF_E_UNSUPPORTED_RATE: Self

Available on crate feature mf only.

pub const MF_E_THINNING_UNSUPPORTED: Self

Available on crate feature mf only.

pub const MF_E_REVERSE_UNSUPPORTED: Self

Available on crate feature mf only.

pub const MF_E_UNSUPPORTED_RATE_TRANSITION: Self

Available on crate feature mf only.

pub const MF_E_RATE_CHANGE_PREEMPTED: Self

Available on crate feature mf only.

pub const MF_E_NOT_FOUND: Self

Available on crate feature mf only.

pub const MF_E_NOT_AVAILABLE: Self

Available on crate feature mf only.

pub const MF_E_NO_CLOCK: Self

Available on crate feature mf only.

pub const MF_S_MULTIPLE_BEGIN: Self

Available on crate feature mf only.

pub const MF_E_MULTIPLE_BEGIN: Self

Available on crate feature mf only.

pub const MF_E_MULTIPLE_SUBSCRIBERS: Self

Available on crate feature mf only.

pub const MF_E_TIMER_ORPHANED: Self

Available on crate feature mf only.

pub const MF_E_STATE_TRANSITION_PENDING: Self

Available on crate feature mf only.

pub const MF_E_UNSUPPORTED_STATE_TRANSITION: Self

Available on crate feature mf only.

pub const MF_E_UNRECOVERABLE_ERROR_OCCURRED: Self

Available on crate feature mf only.

pub const MF_E_SAMPLE_HAS_TOO_MANY_BUFFERS: Self

Available on crate feature mf only.

pub const MF_E_SAMPLE_NOT_WRITABLE: Self

Available on crate feature mf only.

pub const MF_E_INVALID_KEY: Self

Available on crate feature mf only.

pub const MF_E_BAD_STARTUP_VERSION: Self

Available on crate feature mf only.

pub const MF_E_UNSUPPORTED_CAPTION: Self

Available on crate feature mf only.

pub const MF_E_INVALID_POSITION: Self

Available on crate feature mf only.

pub const MF_E_ATTRIBUTENOTFOUND: Self

Available on crate feature mf only.

pub const MF_E_PROPERTY_TYPE_NOT_ALLOWED: Self

Available on crate feature mf only.

pub const MF_E_PROPERTY_TYPE_NOT_SUPPORTED: Self

Available on crate feature mf only.

pub const MF_E_PROPERTY_EMPTY: Self

Available on crate feature mf only.

pub const MF_E_PROPERTY_NOT_EMPTY: Self

Available on crate feature mf only.

pub const MF_E_PROPERTY_VECTOR_NOT_ALLOWED: Self

Available on crate feature mf only.

pub const MF_E_PROPERTY_VECTOR_REQUIRED: Self

Available on crate feature mf only.

pub const MF_E_OPERATION_CANCELLED: Self

Available on crate feature mf only.

pub const MF_E_BYTESTREAM_NOT_SEEKABLE: Self

Available on crate feature mf only.

pub const MF_E_DISABLED_IN_SAFEMODE: Self

Available on crate feature mf only.

pub const MF_E_CANNOT_PARSE_BYTESTREAM: Self

Available on crate feature mf only.

pub const MF_E_SOURCERESOLVER_MUTUALLY_EXCLUSIVE_FLAGS: Self

Available on crate feature mf only.

pub const MF_E_MEDIAPROC_WRONGSTATE: Self

Available on crate feature mf only.

pub const MF_E_RT_THROUGHPUT_NOT_AVAILABLE: Self

Available on crate feature mf only.

pub const MF_E_RT_TOO_MANY_CLASSES: Self

Available on crate feature mf only.

pub const MF_E_RT_WOULDBLOCK: Self

Available on crate feature mf only.

pub const MF_E_NO_BITPUMP: Self

Available on crate feature mf only.

pub const MF_E_RT_OUTOFMEMORY: Self

Available on crate feature mf only.

pub const MF_E_RT_WORKQUEUE_CLASS_NOT_SPECIFIED: Self

Available on crate feature mf only.

pub const MF_E_INSUFFICIENT_BUFFER: Self

Available on crate feature mf only.

pub const MF_E_CANNOT_CREATE_SINK: Self

Available on crate feature mf only.

pub const MF_E_BYTESTREAM_UNKNOWN_LENGTH: Self

Available on crate feature mf only.

pub const MF_E_SESSION_PAUSEWHILESTOPPED: Self

Available on crate feature mf only.

pub const MF_S_ACTIVATE_REPLACED: Self

Available on crate feature mf only.

pub const MF_E_FORMAT_CHANGE_NOT_SUPPORTED: Self

Available on crate feature mf only.

pub const MF_E_INVALID_WORKQUEUE: Self

Available on crate feature mf only.

pub const MF_E_DRM_UNSUPPORTED: Self

Available on crate feature mf only.

pub const MF_E_UNAUTHORIZED: Self

Available on crate feature mf only.

pub const MF_E_OUT_OF_RANGE: Self

Available on crate feature mf only.

pub const MF_E_INVALID_CODEC_MERIT: Self

Available on crate feature mf only.

pub const MF_E_HW_MFT_FAILED_START_STREAMING: Self

Available on crate feature mf only.

pub const MF_E_OPERATION_IN_PROGRESS: Self

Available on crate feature mf only.

pub const MF_E_HARDWARE_DRM_UNSUPPORTED: Self

Available on crate feature mf only.

pub const MF_E_DURATION_TOO_LONG: Self

Available on crate feature mf only.

pub const MF_E_OPERATION_UNSUPPORTED_AT_D3D_FEATURE_LEVEL: Self

Available on crate feature mf only.

pub const MF_E_UNSUPPORTED_MEDIATYPE_AT_D3D_FEATURE_LEVEL: Self

Available on crate feature mf only.

pub const MF_S_ASF_PARSEINPROGRESS: Self

Available on crate feature mf only.

pub const MF_E_ASF_PARSINGINCOMPLETE: Self

Available on crate feature mf only.

pub const MF_E_ASF_MISSINGDATA: Self

Available on crate feature mf only.

pub const MF_E_ASF_INVALIDDATA: Self

Available on crate feature mf only.

pub const MF_E_ASF_OPAQUEPACKET: Self

Available on crate feature mf only.

pub const MF_E_ASF_NOINDEX: Self

Available on crate feature mf only.

pub const MF_E_ASF_OUTOFRANGE: Self

Available on crate feature mf only.

pub const MF_E_ASF_INDEXNOTLOADED: Self

Available on crate feature mf only.

pub const MF_E_ASF_TOO_MANY_PAYLOADS: Self

Available on crate feature mf only.

pub const MF_E_ASF_UNSUPPORTED_STREAM_TYPE: Self

Available on crate feature mf only.

pub const MF_E_ASF_DROPPED_PACKET: Self

Available on crate feature mf only.

pub const MF_E_NO_EVENTS_AVAILABLE: Self

Available on crate feature mf only.

pub const MF_E_INVALID_STATE_TRANSITION: Self

Available on crate feature mf only.

pub const MF_E_END_OF_STREAM: Self

Available on crate feature mf only.

pub const MF_E_SHUTDOWN: Self

Available on crate feature mf only.

pub const MF_E_MP3_NOTFOUND: Self

Available on crate feature mf only.

pub const MF_E_MP3_OUTOFDATA: Self

Available on crate feature mf only.

pub const MF_E_MP3_NOTMP3: Self

Available on crate feature mf only.

pub const MF_E_MP3_NOTSUPPORTED: Self

Available on crate feature mf only.

pub const MF_E_NO_DURATION: Self

Available on crate feature mf only.

pub const MF_E_INVALID_FORMAT: Self

Available on crate feature mf only.

pub const MF_E_PROPERTY_NOT_FOUND: Self

Available on crate feature mf only.

pub const MF_E_PROPERTY_READ_ONLY: Self

Available on crate feature mf only.

pub const MF_E_PROPERTY_NOT_ALLOWED: Self

Available on crate feature mf only.

pub const MF_E_MEDIA_SOURCE_NOT_STARTED: Self

Available on crate feature mf only.

pub const MF_E_UNSUPPORTED_FORMAT: Self

Available on crate feature mf only.

pub const MF_E_MP3_BAD_CRC: Self

Available on crate feature mf only.

pub const MF_E_NOT_PROTECTED: Self

Available on crate feature mf only.

pub const MF_E_MEDIA_SOURCE_WRONGSTATE: Self

Available on crate feature mf only.

pub const MF_E_MEDIA_SOURCE_NO_STREAMS_SELECTED: Self

Available on crate feature mf only.

pub const MF_E_CANNOT_FIND_KEYFRAME_SAMPLE: Self

Available on crate feature mf only.

pub const MF_E_UNSUPPORTED_CHARACTERISTICS: Self

Available on crate feature mf only.

pub const MF_E_NO_AUDIO_RECORDING_DEVICE: Self

Available on crate feature mf only.

pub const MF_E_AUDIO_RECORDING_DEVICE_IN_USE: Self

Available on crate feature mf only.

pub const MF_E_AUDIO_RECORDING_DEVICE_INVALIDATED: Self

Available on crate feature mf only.

pub const MF_E_VIDEO_RECORDING_DEVICE_INVALIDATED: Self

Available on crate feature mf only.

pub const MF_E_VIDEO_RECORDING_DEVICE_PREEMPTED: Self

Available on crate feature mf only.

pub const MF_E_NETWORK_RESOURCE_FAILURE: Self

Available on crate feature mf only.

pub const MF_E_NET_WRITE: Self

Available on crate feature mf only.

pub const MF_E_NET_READ: Self

Available on crate feature mf only.

pub const MF_E_NET_REQUIRE_NETWORK: Self

Available on crate feature mf only.

pub const MF_E_NET_REQUIRE_ASYNC: Self

Available on crate feature mf only.

pub const MF_E_NET_BWLEVEL_NOT_SUPPORTED: Self

Available on crate feature mf only.

pub const MF_E_NET_STREAMGROUPS_NOT_SUPPORTED: Self

Available on crate feature mf only.

pub const MF_E_NET_MANUALSS_NOT_SUPPORTED: Self

Available on crate feature mf only.

pub const MF_E_NET_INVALID_PRESENTATION_DESCRIPTOR: Self

Available on crate feature mf only.

pub const MF_E_NET_CACHESTREAM_NOT_FOUND: Self

Available on crate feature mf only.

pub const MF_I_MANUAL_PROXY: Self

Available on crate feature mf only.

pub const MF_E_NET_REQUIRE_INPUT: Self

Available on crate feature mf only.

pub const MF_E_NET_REDIRECT: Self

Available on crate feature mf only.

pub const MF_E_NET_REDIRECT_TO_PROXY: Self

Available on crate feature mf only.

pub const MF_E_NET_TOO_MANY_REDIRECTS: Self

Available on crate feature mf only.

pub const MF_E_NET_TIMEOUT: Self

Available on crate feature mf only.

pub const MF_E_NET_CLIENT_CLOSE: Self

Available on crate feature mf only.

pub const MF_E_NET_BAD_CONTROL_DATA: Self

Available on crate feature mf only.

pub const MF_E_NET_INCOMPATIBLE_SERVER: Self

Available on crate feature mf only.

pub const MF_E_NET_UNSAFE_URL: Self

Available on crate feature mf only.

pub const MF_E_NET_CACHE_NO_DATA: Self

Available on crate feature mf only.

pub const MF_E_NET_EOL: Self

Available on crate feature mf only.

pub const MF_E_NET_BAD_REQUEST: Self

Available on crate feature mf only.

pub const MF_E_NET_INTERNAL_SERVER_ERROR: Self

Available on crate feature mf only.

pub const MF_E_NET_SESSION_NOT_FOUND: Self

Available on crate feature mf only.

pub const MF_E_NET_NOCONNECTION: Self

Available on crate feature mf only.

pub const MF_E_NET_CONNECTION_FAILURE: Self

Available on crate feature mf only.

pub const MF_E_NET_INCOMPATIBLE_PUSHSERVER: Self

Available on crate feature mf only.

pub const MF_E_NET_SERVER_ACCESSDENIED: Self

Available on crate feature mf only.

pub const MF_E_NET_PROXY_ACCESSDENIED: Self

Available on crate feature mf only.

pub const MF_E_NET_CANNOTCONNECT: Self

Available on crate feature mf only.

pub const MF_E_NET_INVALID_PUSH_TEMPLATE: Self

Available on crate feature mf only.

pub const MF_E_NET_INVALID_PUSH_PUBLISHING_POINT: Self

Available on crate feature mf only.

pub const MF_E_NET_BUSY: Self

Available on crate feature mf only.

pub const MF_E_NET_RESOURCE_GONE: Self

Available on crate feature mf only.

pub const MF_E_NET_ERROR_FROM_PROXY: Self

Available on crate feature mf only.

pub const MF_E_NET_PROXY_TIMEOUT: Self

Available on crate feature mf only.

pub const MF_E_NET_SERVER_UNAVAILABLE: Self

Available on crate feature mf only.

pub const MF_E_NET_TOO_MUCH_DATA: Self

Available on crate feature mf only.

pub const MF_E_NET_SESSION_INVALID: Self

Available on crate feature mf only.

pub const MF_E_OFFLINE_MODE: Self

Available on crate feature mf only.

pub const MF_E_NET_UDP_BLOCKED: Self

Available on crate feature mf only.

pub const MF_E_NET_UNSUPPORTED_CONFIGURATION: Self

Available on crate feature mf only.

pub const MF_E_NET_PROTOCOL_DISABLED: Self

Available on crate feature mf only.

pub const MF_E_NET_COMPANION_DRIVER_DISCONNECT: Self

Available on crate feature mf only.

pub const MF_E_ALREADY_INITIALIZED: Self

Available on crate feature mf only.

pub const MF_E_BANDWIDTH_OVERRUN: Self

Available on crate feature mf only.

pub const MF_E_LATE_SAMPLE: Self

Available on crate feature mf only.

pub const MF_E_FLUSH_NEEDED: Self

Available on crate feature mf only.

pub const MF_E_INVALID_PROFILE: Self

Available on crate feature mf only.

pub const MF_E_INDEX_NOT_COMMITTED: Self

Available on crate feature mf only.

pub const MF_E_NO_INDEX: Self

Available on crate feature mf only.

pub const MF_E_CANNOT_INDEX_IN_PLACE: Self

Available on crate feature mf only.

pub const MF_E_MISSING_ASF_LEAKYBUCKET: Self

Available on crate feature mf only.

pub const MF_E_INVALID_ASF_STREAMID: Self

Available on crate feature mf only.

pub const MF_E_STREAMSINK_REMOVED: Self

Available on crate feature mf only.

pub const MF_E_STREAMSINKS_OUT_OF_SYNC: Self

Available on crate feature mf only.

pub const MF_E_STREAMSINKS_FIXED: Self

Available on crate feature mf only.

pub const MF_E_STREAMSINK_EXISTS: Self

Available on crate feature mf only.

pub const MF_E_SAMPLEALLOCATOR_CANCELED: Self

Available on crate feature mf only.

pub const MF_E_SAMPLEALLOCATOR_EMPTY: Self

Available on crate feature mf only.

pub const MF_E_SINK_ALREADYSTOPPED: Self

Available on crate feature mf only.

pub const MF_E_ASF_FILESINK_BITRATE_UNKNOWN: Self

Available on crate feature mf only.

pub const MF_E_SINK_NO_STREAMS: Self

Available on crate feature mf only.

pub const MF_S_SINK_NOT_FINALIZED: Self

Available on crate feature mf only.

pub const MF_E_METADATA_TOO_LONG: Self

Available on crate feature mf only.

pub const MF_E_SINK_NO_SAMPLES_PROCESSED: Self

Available on crate feature mf only.

pub const MF_E_SINK_HEADERS_NOT_FOUND: Self

Available on crate feature mf only.

pub const MF_E_VIDEO_REN_NO_PROCAMP_HW: Self

Available on crate feature mf only.

pub const MF_E_VIDEO_REN_NO_DEINTERLACE_HW: Self

Available on crate feature mf only.

pub const MF_E_VIDEO_REN_COPYPROT_FAILED: Self

Available on crate feature mf only.

pub const MF_E_VIDEO_REN_SURFACE_NOT_SHARED: Self

Available on crate feature mf only.

pub const MF_E_VIDEO_DEVICE_LOCKED: Self

Available on crate feature mf only.

pub const MF_E_NEW_VIDEO_DEVICE: Self

Available on crate feature mf only.

pub const MF_E_NO_VIDEO_SAMPLE_AVAILABLE: Self

Available on crate feature mf only.

pub const MF_E_NO_AUDIO_PLAYBACK_DEVICE: Self

Available on crate feature mf only.

pub const MF_E_AUDIO_PLAYBACK_DEVICE_IN_USE: Self

Available on crate feature mf only.

pub const MF_E_AUDIO_PLAYBACK_DEVICE_INVALIDATED: Self

Available on crate feature mf only.

pub const MF_E_AUDIO_SERVICE_NOT_RUNNING: Self

Available on crate feature mf only.

pub const MF_E_AUDIO_BUFFER_SIZE_ERROR: Self

Available on crate feature mf only.

pub const MF_E_AUDIO_CLIENT_WRAPPER_SPOOF_ERROR: Self

Available on crate feature mf only.

pub const MF_E_TOPO_INVALID_OPTIONAL_NODE: Self

Available on crate feature mf only.

pub const MF_E_TOPO_CANNOT_FIND_DECRYPTOR: Self

Available on crate feature mf only.

pub const MF_E_TOPO_CODEC_NOT_FOUND: Self

Available on crate feature mf only.

pub const MF_E_TOPO_CANNOT_CONNECT: Self

Available on crate feature mf only.

pub const MF_E_TOPO_UNSUPPORTED: Self

Available on crate feature mf only.

pub const MF_E_TOPO_INVALID_TIME_ATTRIBUTES: Self

Available on crate feature mf only.

pub const MF_E_TOPO_LOOPS_IN_TOPOLOGY: Self

Available on crate feature mf only.

pub const MF_E_TOPO_MISSING_PRESENTATION_DESCRIPTOR: Self

Available on crate feature mf only.

pub const MF_E_TOPO_MISSING_STREAM_DESCRIPTOR: Self

Available on crate feature mf only.

pub const MF_E_TOPO_STREAM_DESCRIPTOR_NOT_SELECTED: Self

Available on crate feature mf only.

pub const MF_E_TOPO_MISSING_SOURCE: Self

Available on crate feature mf only.

pub const MF_E_TOPO_SINK_ACTIVATES_UNSUPPORTED: Self

Available on crate feature mf only.

pub const MF_E_SEQUENCER_UNKNOWN_SEGMENT_ID: Self

Available on crate feature mf only.

pub const MF_S_SEQUENCER_CONTEXT_CANCELED: Self

Available on crate feature mf only.

pub const MF_E_NO_SOURCE_IN_CACHE: Self

Available on crate feature mf only.

pub const MF_S_SEQUENCER_SEGMENT_AT_END_OF_STREAM: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_TYPE_NOT_SET: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_STREAM_CHANGE: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_INPUT_REMAINING: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_PROFILE_MISSING: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_PROFILE_INVALID_OR_CORRUPT: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_PROFILE_TRUNCATED: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_PROPERTY_PID_NOT_RECOGNIZED: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_PROPERTY_VARIANT_TYPE_WRONG: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_PROPERTY_NOT_WRITEABLE: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_PROPERTY_ARRAY_VALUE_WRONG_NUM_DIM: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_PROPERTY_VALUE_SIZE_WRONG: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_PROPERTY_VALUE_OUT_OF_RANGE: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_PROPERTY_VALUE_INCOMPATIBLE: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_NOT_POSSIBLE_FOR_CURRENT_OUTPUT_MEDIATYPE: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_NOT_POSSIBLE_FOR_CURRENT_INPUT_MEDIATYPE: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_NOT_POSSIBLE_FOR_CURRENT_MEDIATYPE_COMBINATION: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_CONFLICTS_WITH_OTHER_CURRENTLY_ENABLED_FEATURES: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_NEED_MORE_INPUT: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_NOT_POSSIBLE_FOR_CURRENT_SPKR_CONFIG: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_CANNOT_CHANGE_MEDIATYPE_WHILE_PROCESSING: Self

Available on crate feature mf only.

pub const MF_S_TRANSFORM_DO_NOT_PROPAGATE_EVENT: Self

Available on crate feature mf only.

pub const MF_E_UNSUPPORTED_D3D_TYPE: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_ASYNC_LOCKED: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_CANNOT_INITIALIZE_ACM_DRIVER: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_STREAM_INVALID_RESOLUTION: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_ASYNC_MFT_NOT_SUPPORTED: Self

Available on crate feature mf only.

pub const MF_E_TRANSFORM_EXATTRIBUTE_NOT_SUPPORTED: Self

Available on crate feature mf only.

pub const MF_E_LICENSE_INCORRECT_RIGHTS: Self

Available on crate feature mf only.

pub const MF_E_LICENSE_OUTOFDATE: Self

Available on crate feature mf only.

pub const MF_E_LICENSE_REQUIRED: Self

Available on crate feature mf only.

pub const MF_E_DRM_HARDWARE_INCONSISTENT: Self

Available on crate feature mf only.

pub const MF_E_NO_CONTENT_PROTECTION_MANAGER: Self

Available on crate feature mf only.

pub const MF_E_LICENSE_RESTORE_NO_RIGHTS: Self

Available on crate feature mf only.

pub const MF_E_BACKUP_RESTRICTED_LICENSE: Self

Available on crate feature mf only.

pub const MF_E_LICENSE_RESTORE_NEEDS_INDIVIDUALIZATION: Self

Available on crate feature mf only.

pub const MF_S_PROTECTION_NOT_REQUIRED: Self

Available on crate feature mf only.

pub const MF_E_COMPONENT_REVOKED: Self

Available on crate feature mf only.

pub const MF_E_TRUST_DISABLED: Self

Available on crate feature mf only.

pub const MF_E_WMDRMOTA_NO_ACTION: Self

Available on crate feature mf only.

pub const MF_E_WMDRMOTA_ACTION_ALREADY_SET: Self

Available on crate feature mf only.

pub const MF_E_WMDRMOTA_DRM_HEADER_NOT_AVAILABLE: Self

Available on crate feature mf only.

pub const MF_E_WMDRMOTA_DRM_ENCRYPTION_SCHEME_NOT_SUPPORTED: Self

Available on crate feature mf only.

pub const MF_E_WMDRMOTA_ACTION_MISMATCH: Self

Available on crate feature mf only.

pub const MF_E_WMDRMOTA_INVALID_POLICY: Self

Available on crate feature mf only.

pub const MF_E_POLICY_UNSUPPORTED: Self

Available on crate feature mf only.

pub const MF_E_OPL_NOT_SUPPORTED: Self

Available on crate feature mf only.

pub const MF_E_TOPOLOGY_VERIFICATION_FAILED: Self

Available on crate feature mf only.

pub const MF_E_SIGNATURE_VERIFICATION_FAILED: Self

Available on crate feature mf only.

pub const MF_E_DEBUGGING_NOT_ALLOWED: Self

Available on crate feature mf only.

pub const MF_E_CODE_EXPIRED: Self

Available on crate feature mf only.

pub const MF_E_GRL_VERSION_TOO_LOW: Self

Available on crate feature mf only.

pub const MF_E_GRL_RENEWAL_NOT_FOUND: Self

Available on crate feature mf only.

pub const MF_E_GRL_EXTENSIBLE_ENTRY_NOT_FOUND: Self

Available on crate feature mf only.

pub const MF_E_KERNEL_UNTRUSTED: Self

Available on crate feature mf only.

pub const MF_E_PEAUTH_UNTRUSTED: Self

Available on crate feature mf only.

pub const MF_E_NON_PE_PROCESS: Self

Available on crate feature mf only.

pub const MF_E_REBOOT_REQUIRED: Self

Available on crate feature mf only.

pub const MF_S_WAIT_FOR_POLICY_SET: Self

Available on crate feature mf only.

pub const MF_S_VIDEO_DISABLED_WITH_UNKNOWN_SOFTWARE_OUTPUT: Self

Available on crate feature mf only.

pub const MF_E_GRL_INVALID_FORMAT: Self

Available on crate feature mf only.

pub const MF_E_GRL_UNRECOGNIZED_FORMAT: Self

Available on crate feature mf only.

pub const MF_E_ALL_PROCESS_RESTART_REQUIRED: Self

Available on crate feature mf only.

pub const MF_E_PROCESS_RESTART_REQUIRED: Self

Available on crate feature mf only.

pub const MF_E_USERMODE_UNTRUSTED: Self

Available on crate feature mf only.

pub const MF_E_PEAUTH_SESSION_NOT_STARTED: Self

Available on crate feature mf only.

pub const MF_E_PEAUTH_PUBLICKEY_REVOKED: Self

Available on crate feature mf only.

pub const MF_E_GRL_ABSENT: Self

Available on crate feature mf only.

pub const MF_S_PE_TRUSTED: Self

Available on crate feature mf only.

pub const MF_E_PE_UNTRUSTED: Self

Available on crate feature mf only.

pub const MF_E_PEAUTH_NOT_STARTED: Self

Available on crate feature mf only.

pub const MF_E_INCOMPATIBLE_SAMPLE_PROTECTION: Self

Available on crate feature mf only.

pub const MF_E_PE_SESSIONS_MAXED: Self

Available on crate feature mf only.

pub const MF_E_HIGH_SECURITY_LEVEL_CONTENT_NOT_ALLOWED: Self

Available on crate feature mf only.

pub const MF_E_TEST_SIGNED_COMPONENTS_NOT_ALLOWED: Self

Available on crate feature mf only.

pub const MF_E_ITA_UNSUPPORTED_ACTION: Self

Available on crate feature mf only.

pub const MF_E_ITA_ERROR_PARSING_SAP_PARAMETERS: Self

Available on crate feature mf only.

pub const MF_E_POLICY_MGR_ACTION_OUTOFBOUNDS: Self

Available on crate feature mf only.

pub const MF_E_BAD_OPL_STRUCTURE_FORMAT: Self

Available on crate feature mf only.

pub const MF_E_ITA_UNRECOGNIZED_ANALOG_VIDEO_PROTECTION_GUID: Self

Available on crate feature mf only.

pub const MF_E_NO_PMP_HOST: Self

Available on crate feature mf only.

pub const MF_E_ITA_OPL_DATA_NOT_INITIALIZED: Self

Available on crate feature mf only.

pub const MF_E_ITA_UNRECOGNIZED_ANALOG_VIDEO_OUTPUT: Self

Available on crate feature mf only.

pub const MF_E_ITA_UNRECOGNIZED_DIGITAL_VIDEO_OUTPUT: Self

Available on crate feature mf only.

pub const MF_E_RESOLUTION_REQUIRES_PMP_CREATION_CALLBACK: Self

Available on crate feature mf only.

pub const MF_E_INVALID_AKE_CHANNEL_PARAMETERS: Self

Available on crate feature mf only.

pub const MF_E_CONTENT_PROTECTION_SYSTEM_NOT_ENABLED: Self

Available on crate feature mf only.

pub const MF_E_UNSUPPORTED_CONTENT_PROTECTION_SYSTEM: Self

Available on crate feature mf only.

pub const MF_E_DRM_MIGRATION_NOT_SUPPORTED: Self

Available on crate feature mf only.

pub const MF_E_HDCP_AUTHENTICATION_FAILURE: Self

Available on crate feature mf only.

pub const MF_E_HDCP_LINK_FAILURE: Self

Available on crate feature mf only.

pub const MF_E_CLOCK_INVALID_CONTINUITY_KEY: Self

Available on crate feature mf only.

pub const MF_E_CLOCK_NO_TIME_SOURCE: Self

Available on crate feature mf only.

pub const MF_E_CLOCK_STATE_ALREADY_SET: Self

Available on crate feature mf only.

pub const MF_E_CLOCK_NOT_SIMPLE: Self

Available on crate feature mf only.

pub const MF_S_CLOCK_STOPPED: Self

Available on crate feature mf only.

pub const MF_E_CLOCK_AUDIO_DEVICE_POSITION_UNEXPECTED: Self

Available on crate feature mf only.

pub const MF_E_CLOCK_AUDIO_RENDER_POSITION_UNEXPECTED: Self

Available on crate feature mf only.

pub const MF_E_CLOCK_AUDIO_RENDER_TIME_UNEXPECTED: Self

Available on crate feature mf only.

pub const MF_E_NO_MORE_DROP_MODES: Self

Available on crate feature mf only.

pub const MF_E_NO_MORE_QUALITY_LEVELS: Self

Available on crate feature mf only.

pub const MF_E_DROPTIME_NOT_SUPPORTED: Self

Available on crate feature mf only.

pub const MF_E_QUALITYKNOB_WAIT_LONGER: Self

Available on crate feature mf only.

pub const MF_E_QM_INVALIDSTATE: Self

Available on crate feature mf only.

pub const MF_E_TRANSCODE_NO_CONTAINERTYPE: Self

Available on crate feature mf only.

pub const MF_E_TRANSCODE_PROFILE_NO_MATCHING_STREAMS: Self

Available on crate feature mf only.

pub const MF_E_TRANSCODE_NO_MATCHING_ENCODER: Self

Available on crate feature mf only.

pub const MF_E_TRANSCODE_INVALID_PROFILE: Self

Available on crate feature mf only.

pub const MF_E_ALLOCATOR_NOT_INITIALIZED: Self

Available on crate feature mf only.

pub const MF_E_ALLOCATOR_NOT_COMMITED: Self

Available on crate feature mf only.

pub const MF_E_ALLOCATOR_ALREADY_COMMITED: Self

Available on crate feature mf only.

pub const MF_E_STREAM_ERROR: Self

Available on crate feature mf only.

pub const MF_E_INVALID_STREAM_STATE: Self

Available on crate feature mf only.

pub const MF_E_HW_STREAM_NOT_CONNECTED: Self

Available on crate feature mf only.

pub const MF_E_NO_CAPTURE_DEVICES_AVAILABLE: Self

Available on crate feature mf only.

pub const MF_E_CAPTURE_SINK_OUTPUT_NOT_SET: Self

Available on crate feature mf only.

pub const MF_E_CAPTURE_SINK_MIRROR_ERROR: Self

Available on crate feature mf only.

pub const MF_E_CAPTURE_SINK_ROTATE_ERROR: Self

Available on crate feature mf only.

pub const MF_E_CAPTURE_ENGINE_INVALID_OP: Self

Available on crate feature mf only.

pub const MF_E_CAPTURE_ENGINE_ALL_EFFECTS_REMOVED: Self

Available on crate feature mf only.

pub const MF_E_CAPTURE_SOURCE_NO_INDEPENDENT_PHOTO_STREAM_PRESENT: Self

Available on crate feature mf only.

pub const MF_E_CAPTURE_SOURCE_NO_VIDEO_STREAM_PRESENT: Self

Available on crate feature mf only.

pub const MF_E_CAPTURE_SOURCE_NO_AUDIO_STREAM_PRESENT: Self

Available on crate feature mf only.

pub const MF_E_CAPTURE_SOURCE_DEVICE_EXTENDEDPROP_OP_IN_PROGRESS: Self

Available on crate feature mf only.

pub const MF_E_CAPTURE_PROPERTY_SET_DURING_PHOTO: Self

Available on crate feature mf only.

pub const MF_E_CAPTURE_NO_SAMPLES_IN_QUEUE: Self

Available on crate feature mf only.

pub const MF_E_HW_ACCELERATED_THUMBNAIL_NOT_SUPPORTED: Self

Available on crate feature mf only.

pub const MF_E_UNSUPPORTED_CAPTURE_DEVICE_PRESENT: Self

Available on crate feature mf only.

pub const MF_E_TIMELINECONTROLLER_UNSUPPORTED_SOURCE_TYPE: Self

Available on crate feature mf only.

pub const MF_E_TIMELINECONTROLLER_NOT_ALLOWED: Self

Available on crate feature mf only.

pub const MF_E_TIMELINECONTROLLER_CANNOT_ATTACH: Self

Available on crate feature mf only.

pub const MF_E_MEDIA_EXTENSION_APPSERVICE_CONNECTION_FAILED: Self

Available on crate feature mf only.

pub const MF_E_MEDIA_EXTENSION_APPSERVICE_REQUEST_FAILED: Self

Available on crate feature mf only.

pub const MF_E_MEDIA_EXTENSION_PACKAGE_INTEGRITY_CHECK_FAILED: Self

Available on crate feature mf only.

pub const MF_E_MEDIA_EXTENSION_PACKAGE_LICENSE_INVALID: Self

Available on crate feature mf only.

impl HRESULT

pub const unsafe fn as_mut(&mut self) -> &mut u32

Returns a mutable reference to the underlying raw value.

Safety

Be sure the integer being set is meaningful for the actual type.

pub const unsafe fn from_raw(v: u32) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u32

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl HRESULT

pub const fn code(self) -> u16

HRESULT_CODE macro.

pub const fn facility(self) -> FACILITY

HRESULT_FACILITY macro.

pub const fn severity(self) -> SEVERITY

HRESULT_SEVERITY macro.

pub const fn FAILED(self) -> bool

FAILED macro.

pub const fn SUCCEEDED(self) -> bool

SUCCEEDED macro.

impl HRESULT

pub const S_OK: Self

Operation successful.

pub const S_FALSE: Self

pub const E_UNEXPECTED: Self

pub const E_NOTIMPL: Self

pub const E_OUTOFMEMORY: Self

pub const E_INVALIDARG: Self

pub const E_NOINTERFACE: Self

pub const E_POINTER: Self

pub const E_HANDLE: Self

pub const E_ABORT: Self

pub const E_FAIL: Self

pub const E_ACCESSDENIED: Self

pub const E_PENDING: Self

pub const E_BOUNDS: Self

pub const E_CHANGED_STATE: Self

pub const E_ILLEGAL_STATE_CHANGE: Self

pub const E_ILLEGAL_METHOD_CALL: Self

pub const E_STRING_NOT_NULL_TERMINATED: Self

pub const E_ILLEGAL_DELEGATE_ASSIGNMENT: Self

pub const E_ASYNC_OPERATION_NOT_STARTED: Self

pub const E_APPLICATION_EXITING: Self

pub const E_APPLICATION_VIEW_EXITING: Self

pub const CO_E_INIT_TLS: Self

pub const CO_E_INIT_SHARED_ALLOCATOR: Self

pub const CO_E_INIT_MEMORY_ALLOCATOR: Self

pub const CO_E_INIT_CLASS_CACHE: Self

pub const CO_E_INIT_RPC_CHANNEL: Self

pub const CO_E_INIT_TLS_SET_CHANNEL_CONTROL: Self

pub const CO_E_INIT_TLS_CHANNEL_CONTROL: Self

pub const CO_E_INIT_UNACCEPTED_USER_ALLOCATOR: Self

pub const CO_E_INIT_SCM_MUTEX_EXISTS: Self

pub const CO_E_INIT_SCM_FILE_MAPPING_EXISTS: Self

pub const CO_E_INIT_SCM_MAP_VIEW_OF_FILE: Self

pub const CO_E_INIT_SCM_EXEC_FAILURE: Self

pub const CO_E_INIT_ONLY_SINGLE_THREADED: Self

pub const CO_E_CANT_REMOTE: Self

pub const CO_E_BAD_SERVER_NAME: Self

pub const CO_E_WRONG_SERVER_IDENTITY: Self

pub const CO_E_OLE1DDE_DISABLED: Self

pub const CO_E_RUNAS_SYNTAX: Self

pub const CO_E_CREATEPROCESS_FAILURE: Self

pub const CO_E_RUNAS_CREATEPROCESS_FAILURE: Self

pub const CO_E_RUNAS_LOGON_FAILURE: Self

pub const CO_E_LAUNCH_PERMSSION_DENIED: Self

pub const CO_E_START_SERVICE_FAILURE: Self

pub const CO_E_REMOTE_COMMUNICATION_FAILURE: Self

pub const CO_E_SERVER_START_TIMEOUT: Self

pub const CO_E_CLSREG_INCONSISTENT: Self

pub const CO_E_IIDREG_INCONSISTENT: Self

pub const CO_E_NOT_SUPPORTED: Self

pub const CO_E_RELOAD_DLL: Self

pub const CO_E_MSI_ERROR: Self

pub const CO_E_ATTEMPT_TO_CREATE_OUTSIDE_CLIENT_CONTEXT: Self

pub const CO_E_SERVER_PAUSED: Self

pub const CO_E_SERVER_NOT_PAUSED: Self

pub const CO_E_CLASS_DISABLED: Self

pub const CO_E_CLRNOTAVAILABLE: Self

pub const CO_E_ASYNC_WORK_REJECTED: Self

pub const CO_E_SERVER_INIT_TIMEOUT: Self

pub const CO_E_NO_SECCTX_IN_ACTIVATE: Self

pub const CO_E_TRACKER_CONFIG: Self

pub const CO_E_THREADPOOL_CONFIG: Self

pub const CO_E_SXS_CONFIG: Self

pub const CO_E_MALFORMED_SPN: Self

pub const CO_E_UNREVOKED_REGISTRATION_ON_APARTMENT_SHUTDOWN: Self

pub const CO_E_PREMATURE_STUB_RUNDOWN: Self

pub const OLE_E_OLEVERB: Self

pub const OLE_E_ADVF: Self

pub const OLE_E_ENUM_NOMORE: Self

pub const OLE_E_ADVISENOTSUPPORTED: Self

pub const OLE_E_NOCONNECTION: Self

pub const OLE_E_NOTRUNNING: Self

pub const OLE_E_NOCACHE: Self

pub const OLE_E_BLANK: Self

pub const OLE_E_CLASSDIFF: Self

pub const OLE_E_CANT_GETMONIKER: Self

pub const OLE_E_CANT_BINDTOSOURCE: Self

pub const OLE_E_STATIC: Self

pub const OLE_E_PROMPTSAVECANCELLED: Self

pub const OLE_E_INVALIDRECT: Self

pub const OLE_E_WRONGCOMPOBJ: Self

pub const OLE_E_INVALIDHWND: Self

pub const OLE_E_NOT_INPLACEACTIVE: Self

pub const OLE_E_CANTCONVERT: Self

pub const OLE_E_NOSTORAGE: Self

pub const DV_E_FORMATETC: Self

pub const DV_E_DVTARGETDEVICE: Self

pub const DV_E_STGMEDIUM: Self

pub const DV_E_STATDATA: Self

pub const DV_E_LINDEX: Self

pub const DV_E_TYMED: Self

pub const DV_E_CLIPFORMAT: Self

pub const DV_E_DVASPECT: Self

pub const DV_E_DVTARGETDEVICE_SIZE: Self

pub const DV_E_NOIVIEWOBJECT: Self

pub const CLASS_E_NOAGGREGATION: Self

pub const CLASS_E_CLASSNOTAVAILABLE: Self

pub const CLASS_E_NOTLICENSED: Self

pub const CO_E_NOTINITIALIZED: Self

pub const CO_E_ALREADYINITIALIZED: Self

pub const CO_E_CANTDETERMINECLASS: Self

pub const CO_E_CLASSSTRING: Self

pub const CO_E_IIDSTRING: Self

pub const CO_E_APPNOTFOUND: Self

pub const CO_E_APPSINGLEUSE: Self

pub const CO_E_ERRORINAPP: Self

pub const CO_E_DLLNOTFOUND: Self

pub const CO_E_ERRORINDLL: Self

pub const CO_E_WRONGOSFORAPP: Self

pub const CO_E_OBJNOTREG: Self

pub const CO_E_OBJISREG: Self

pub const CO_E_OBJNOTCONNECTED: Self

pub const CO_E_APPDIDNTREG: Self

pub const CO_E_RELEASED: Self

pub const CO_E_ACTIVATIONFAILED: Self

pub const CO_E_ACTIVATIONFAILED_EVENTLOGGED: Self

pub const CO_E_ACTIVATIONFAILED_CATALOGERROR: Self

pub const CO_E_ACTIVATIONFAILED_TIMEOUT: Self

pub const CO_E_INITIALIZATIONFAILED: Self

pub const CO_E_THREADINGMODEL_CHANGED: Self

pub const CO_E_NOIISINTRINSICS: Self

pub const CO_E_NOCOOKIES: Self

pub const CO_E_DBERROR: Self

pub const CO_E_NOTPOOLED: Self

pub const CO_E_NOTCONSTRUCTED: Self

pub const CO_E_NOSYNCHRONIZATION: Self

pub const CO_E_ISOLEVELMISMATCH: Self

pub const CO_E_CALL_OUT_OF_TX_SCOPE_NOT_ALLOWED: Self

pub const CO_E_EXIT_TRANSACTION_SCOPE_NOT_CALLED: Self

pub const OLE_S_USEREG: Self

pub const OLE_S_STATIC: Self

pub const OLE_S_MAC_CLIPFORMAT: Self

pub const CO_E_CLASS_CREATE_FAILED: Self

pub const CO_E_SCM_ERROR: Self

pub const CO_E_SCM_RPC_FAILURE: Self

pub const CO_E_BAD_PATH: Self

pub const CO_E_SERVER_EXEC_FAILURE: Self

pub const CO_E_OBJSRV_RPC_FAILURE: Self

pub const CO_E_SERVER_STOPPING: Self

pub const CO_S_NOTALLINTERFACES: Self

pub const CO_S_MACHINENAMENOTFOUND: Self

pub const CO_E_MISSING_DISPLAYNAME: Self

pub const CO_E_RUNAS_VALUE_MUST_BE_AAA: Self

pub const CO_E_ELEVATION_DISABLED: Self

pub const DISP_E_UNKNOWNINTERFACE: Self

pub const DISP_E_MEMBERNOTFOUND: Self

pub const DISP_E_PARAMNOTFOUND: Self

pub const DISP_E_TYPEMISMATCH: Self

pub const DISP_E_UNKNOWNNAME: Self

pub const DISP_E_NONAMEDARGS: Self

pub const DISP_E_BADVARTYPE: Self

pub const DISP_E_EXCEPTION: Self

pub const DISP_E_OVERFLOW: Self

pub const DISP_E_BADINDEX: Self

pub const DISP_E_UNKNOWNLCID: Self

pub const DISP_E_ARRAYISLOCKED: Self

pub const DISP_E_BADPARAMCOUNT: Self

pub const DISP_E_PARAMNOTOPTIONAL: Self

pub const DISP_E_BADCALLEE: Self

pub const DISP_E_NOTACOLLECTION: Self

pub const DISP_E_DIVBYZERO: Self

pub const DISP_E_BUFFERTOOSMALL: Self

pub const RPC_E_CALL_REJECTED: Self

pub const RPC_E_CALL_CANCELED: Self

pub const RPC_E_CANTPOST_INSENDCALL: Self

pub const RPC_E_CANTCALLOUT_INASYNCCALL: Self

pub const RPC_E_CANTCALLOUT_INEXTERNALCALL: Self

pub const RPC_E_CONNECTION_TERMINATED: Self

pub const RPC_E_SERVER_DIED: Self

pub const RPC_E_CLIENT_DIED: Self

pub const RPC_E_INVALID_DATAPACKET: Self

pub const RPC_E_CANTTRANSMIT_CALL: Self

pub const RPC_E_CLIENT_CANTMARSHAL_DATA: Self

pub const RPC_E_CLIENT_CANTUNMARSHAL_DATA: Self

pub const RPC_E_SERVER_CANTMARSHAL_DATA: Self

pub const RPC_E_SERVER_CANTUNMARSHAL_DATA: Self

pub const RPC_E_INVALID_DATA: Self

pub const RPC_E_INVALID_PARAMETER: Self

pub const RPC_E_CANTCALLOUT_AGAIN: Self

pub const RPC_E_SERVER_DIED_DNE: Self

pub const RPC_E_SYS_CALL_FAILED: Self

pub const RPC_E_OUT_OF_RESOURCES: Self

pub const RPC_E_ATTEMPTED_MULTITHREAD: Self

pub const RPC_E_NOT_REGISTERED: Self

pub const RPC_E_FAULT: Self

pub const RPC_E_SERVERFAULT: Self

pub const RPC_E_CHANGED_MODE: Self

pub const RPC_E_INVALIDMETHOD: Self

pub const RPC_E_DISCONNECTED: Self

pub const RPC_E_RETRY: Self

pub const RPC_E_SERVERCALL_RETRYLATER: Self

pub const RPC_E_SERVERCALL_REJECTED: Self

pub const RPC_E_INVALID_CALLDATA: Self

pub const RPC_E_CANTCALLOUT_ININPUTSYNCCALL: Self

pub const RPC_E_WRONG_THREAD: Self

pub const RPC_E_THREAD_NOT_INIT: Self

pub const RPC_E_VERSION_MISMATCH: Self

pub const RPC_E_INVALID_HEADER: Self

pub const RPC_E_INVALID_EXTENSION: Self

pub const RPC_E_INVALID_IPID: Self

pub const RPC_E_INVALID_OBJECT: Self

pub const RPC_S_CALLPENDING: Self

pub const RPC_S_WAITONTIMER: Self

pub const RPC_E_CALL_COMPLETE: Self

pub const RPC_E_UNSECURE_CALL: Self

pub const RPC_E_TOO_LATE: Self

pub const RPC_E_NO_GOOD_SECURITY_PACKAGES: Self

pub const RPC_E_ACCESS_DENIED: Self

pub const RPC_E_REMOTE_DISABLED: Self

pub const RPC_E_INVALID_OBJREF: Self

pub const RPC_E_NO_CONTEXT: Self

pub const RPC_E_TIMEOUT: Self

pub const RPC_E_NO_SYNC: Self

pub const RPC_E_FULLSIC_REQUIRED: Self

pub const RPC_E_INVALID_STD_NAME: Self

pub const CO_E_FAILEDTOIMPERSONATE: Self

pub const CO_E_FAILEDTOGETSECCTX: Self

pub const CO_E_FAILEDTOOPENTHREADTOKEN: Self

pub const CO_E_FAILEDTOGETTOKENINFO: Self

pub const CO_E_TRUSTEEDOESNTMATCHCLIENT: Self

pub const CO_E_FAILEDTOQUERYCLIENTBLANKET: Self

pub const CO_E_FAILEDTOSETDACL: Self

pub const CO_E_ACCESSCHECKFAILED: Self

pub const CO_E_NETACCESSAPIFAILED: Self

pub const CO_E_WRONGTRUSTEENAMESYNTAX: Self

pub const CO_E_INVALIDSID: Self

pub const CO_E_CONVERSIONFAILED: Self

pub const CO_E_NOMATCHINGSIDFOUND: Self

pub const CO_E_LOOKUPACCSIDFAILED: Self

pub const CO_E_NOMATCHINGNAMEFOUND: Self

pub const CO_E_LOOKUPACCNAMEFAILED: Self

pub const CO_E_SETSERLHNDLFAILED: Self

pub const CO_E_FAILEDTOGETWINDIR: Self

pub const CO_E_PATHTOOLONG: Self

pub const CO_E_FAILEDTOGENUUID: Self

pub const CO_E_FAILEDTOCREATEFILE: Self

pub const CO_E_FAILEDTOCLOSEHANDLE: Self

pub const CO_E_EXCEEDSYSACLLIMIT: Self

pub const CO_E_ACESINWRONGORDER: Self

pub const CO_E_INCOMPATIBLESTREAMVERSION: Self

pub const CO_E_FAILEDTOOPENPROCESSTOKEN: Self

pub const CO_E_DECODEFAILED: Self

pub const CO_E_ACNOTINITIALIZED: Self

pub const CO_E_CANCEL_DISABLED: Self

pub const RPC_E_UNEXPECTED: Self

impl HRESULT

pub const CTL_E_INVALIDPICTURE: Self

Available on crate feature oleaut only.

pub const INPLACE_E_NOTUNDOABLE: Self

Available on crate feature oleaut only.

pub const INPLACE_E_NOTOOLSPACE: Self

Available on crate feature oleaut only.

pub const INPLACE_S_TRUNCATED: Self

Available on crate feature oleaut only.