winsafe::co

Struct HTTP_QUERY

pub struct HTTP_QUERY(/* private fields */);

Available on crate feature wininet only.

HINTERNETREQUEST::HttpQueryInfo info (u32).

Implementations

impl HTTP_QUERY

pub const unsafe fn as_mut(&mut self) -> &mut u32

Returns a mutable reference to the underlying raw value.

Safety

Be sure the integer being set is meaningful for the actual type.

pub const unsafe fn from_raw(v: u32) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u32

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl HTTP_QUERY

pub const MIME_VERSION: Self

pub const CONTENT_TYPE: Self

pub const CONTENT_TRANSFER_ENCODING: Self

pub const CONTENT_ID: Self

pub const CONTENT_DESCRIPTION: Self

pub const CONTENT_LENGTH: Self

pub const CONTENT_LANGUAGE: Self

pub const ALLOW: Self

pub const PUBLIC: Self

pub const DATE: Self

pub const EXPIRES: Self

pub const LAST_MODIFIED: Self

pub const MESSAGE_ID: Self

pub const URI: Self

pub const DERIVED_FROM: Self

pub const COST: Self

pub const LINK: Self

pub const PRAGMA: Self

pub const VERSION: Self

pub const STATUS_CODE: Self

pub const STATUS_TEXT: Self

pub const RAW_HEADERS: Self

pub const RAW_HEADERS_CRLF: Self

pub const CONNECTION: Self

pub const ACCEPT: Self

pub const ACCEPT_CHARSET: Self

pub const ACCEPT_ENCODING: Self

pub const ACCEPT_LANGUAGE: Self

pub const AUTHORIZATION: Self

pub const CONTENT_ENCODING: Self

pub const FORWARDED: Self

pub const FROM: Self

pub const IF_MODIFIED_SINCE: Self

pub const LOCATION: Self

pub const ORIG_URI: Self

pub const REFERER: Self

pub const RETRY_AFTER: Self

pub const SERVER: Self

pub const TITLE: Self

pub const USER_AGENT: Self

pub const WWW_AUTHENTICATE: Self

pub const PROXY_AUTHENTICATE: Self

pub const ACCEPT_RANGES: Self

pub const SET_COOKIE: Self

pub const COOKIE: Self

pub const REQUEST_METHOD: Self

pub const REFRESH: Self

pub const CONTENT_DISPOSITION: Self

pub const AGE: Self

pub const CACHE_CONTROL: Self

pub const CONTENT_BASE: Self

pub const CONTENT_LOCATION: Self

pub const CONTENT_MD5: Self

pub const CONTENT_RANGE: Self

pub const ETAG: Self

pub const HOST: Self

pub const IF_MATCH: Self

pub const IF_NONE_MATCH: Self

pub const IF_RANGE: Self

pub const IF_UNMODIFIED_SINCE: Self

pub const MAX_FORWARDS: Self

pub const PROXY_AUTHORIZATION: Self

pub const RANGE: Self

pub const TRANSFER_ENCODING: Self

pub const UPGRADE: Self

pub const VARY: Self

pub const VIA: Self

pub const WARNING: Self

pub const EXPECT: Self

pub const PROXY_CONNECTION: Self

pub const UNLESS_MODIFIED_SINCE: Self

pub const ECHO_REQUEST: Self

pub const ECHO_REPLY: Self

pub const ECHO_HEADERS: Self

pub const ECHO_HEADERS_CRLF: Self

pub const PROXY_SUPPORT: Self

pub const AUTHENTICATION_INFO: Self

pub const PASSPORT_URLS: Self

pub const PASSPORT_CONFIG: Self

pub const X_CONTENT_TYPE_OPTIONS: Self

pub const P3P: Self

pub const X_P2P_PEERDIST: Self

pub const TRANSLATE: Self

pub const X_UA_COMPATIBLE: Self

pub const DEFAULT_STYLE: Self

pub const X_FRAME_OPTIONS: Self

pub const X_XSS_PROTECTION: Self

pub const SET_COOKIE2: Self

pub const DO_NOT_TRACK: Self

pub const KEEP_ALIVE: Self

pub const HTTP2_SETTINGS: Self

pub const STRICT_TRANSPORT_SECURITY: Self

pub const TOKEN_BINDING: Self

pub const INCLUDE_REFERRED_TOKEN_BINDING_ID: Self

pub const INCLUDE_REFERER_TOKEN_BINDING_ID: Self

pub const PUBLIC_KEY_PINS: Self

pub const PUBLIC_KEY_PINS_REPORT_ONLY: Self