winsafe::co

Struct KNOWNFOLDERID

pub struct KNOWNFOLDERID(/* private fields */);

Available on crate feature shell only.

Known system folders (GUID).

Implementations

impl KNOWNFOLDERID

pub const unsafe fn from_raw(guid_str: &str) -> Self

Creates a new GUID from a representative hex string, which can be copied straight from standard GUID declarations.

Safety

Be sure the given value is meaningful for the actual type.

impl KNOWNFOLDERID

pub const NetworkFolder: KNOWNFOLDERID

pub const ComputerFolder: KNOWNFOLDERID

pub const InternetFolder: KNOWNFOLDERID

pub const ControlPanelFolder: KNOWNFOLDERID

pub const PrintersFolder: KNOWNFOLDERID

pub const SyncManagerFolder: KNOWNFOLDERID

pub const SyncSetupFolder: KNOWNFOLDERID

pub const ConflictFolder: KNOWNFOLDERID

pub const SyncResultsFolder: KNOWNFOLDERID

pub const RecycleBinFolder: KNOWNFOLDERID

pub const ConnectionsFolder: KNOWNFOLDERID

pub const Fonts: KNOWNFOLDERID

pub const Desktop: KNOWNFOLDERID

pub const Startup: KNOWNFOLDERID

pub const Programs: KNOWNFOLDERID

pub const StartMenu: KNOWNFOLDERID

pub const Recent: KNOWNFOLDERID

pub const SendTo: KNOWNFOLDERID

pub const Documents: KNOWNFOLDERID

pub const Favorites: KNOWNFOLDERID

pub const NetHood: KNOWNFOLDERID

pub const PrintHood: KNOWNFOLDERID

pub const Templates: KNOWNFOLDERID

pub const CommonStartup: KNOWNFOLDERID

pub const CommonPrograms: KNOWNFOLDERID

pub const CommonStartMenu: KNOWNFOLDERID

pub const PublicDesktop: KNOWNFOLDERID

pub const ProgramData: KNOWNFOLDERID

pub const CommonTemplates: KNOWNFOLDERID

pub const PublicDocuments: KNOWNFOLDERID

pub const RoamingAppData: KNOWNFOLDERID

pub const LocalAppData: KNOWNFOLDERID

pub const LocalAppDataLow: KNOWNFOLDERID

pub const InternetCache: KNOWNFOLDERID

pub const Cookies: KNOWNFOLDERID

pub const History: KNOWNFOLDERID

pub const System: KNOWNFOLDERID

pub const SystemX86: KNOWNFOLDERID

pub const Windows: KNOWNFOLDERID

pub const Profile: KNOWNFOLDERID

pub const Pictures: KNOWNFOLDERID

pub const ProgramFilesX86: KNOWNFOLDERID

pub const ProgramFilesCommonX86: KNOWNFOLDERID

pub const ProgramFilesX64: KNOWNFOLDERID

pub const ProgramFilesCommonX64: KNOWNFOLDERID

pub const ProgramFiles: KNOWNFOLDERID

pub const ProgramFilesCommon: KNOWNFOLDERID

pub const UserProgramFiles: KNOWNFOLDERID

pub const UserProgramFilesCommon: KNOWNFOLDERID

pub const AdminTools: KNOWNFOLDERID

pub const CommonAdminTools: KNOWNFOLDERID

pub const Music: KNOWNFOLDERID

pub const Videos: KNOWNFOLDERID

pub const Ringtones: KNOWNFOLDERID

pub const PublicPictures: KNOWNFOLDERID

pub const PublicMusic: KNOWNFOLDERID

pub const PublicVideos: KNOWNFOLDERID

pub const PublicRingtones: KNOWNFOLDERID

pub const ResourceDir: KNOWNFOLDERID

pub const LocalizedResourcesDir: KNOWNFOLDERID

pub const CommonOEMLinks: KNOWNFOLDERID

pub const CDBurning: KNOWNFOLDERID

pub const UserProfiles: KNOWNFOLDERID

pub const Playlists: KNOWNFOLDERID

pub const SamplePlaylists: KNOWNFOLDERID

pub const SampleMusic: KNOWNFOLDERID

pub const SamplePictures: KNOWNFOLDERID

pub const SampleVideos: KNOWNFOLDERID

pub const PhotoAlbums: KNOWNFOLDERID

pub const Public: KNOWNFOLDERID

pub const ChangeRemovePrograms: KNOWNFOLDERID

pub const AppUpdates: KNOWNFOLDERID

pub const AddNewPrograms: KNOWNFOLDERID

pub const Downloads: KNOWNFOLDERID

pub const PublicDownloads: KNOWNFOLDERID

pub const SavedSearches: KNOWNFOLDERID

pub const QuickLaunch: KNOWNFOLDERID

pub const Contacts: KNOWNFOLDERID

pub const SidebarParts: KNOWNFOLDERID

pub const SidebarDefaultParts: KNOWNFOLDERID

pub const PublicGameTasks: KNOWNFOLDERID

pub const GameTasks: KNOWNFOLDERID

pub const SavedGames: KNOWNFOLDERID

pub const Games: KNOWNFOLDERID

pub const SEARCH_MAPI: KNOWNFOLDERID

pub const SEARCH_CSC: KNOWNFOLDERID

pub const Links: KNOWNFOLDERID

pub const UsersFiles: KNOWNFOLDERID

pub const UsersLibraries: KNOWNFOLDERID

pub const SearchHome: KNOWNFOLDERID

pub const OriginalImages: KNOWNFOLDERID

pub const DocumentsLibrary: KNOWNFOLDERID

pub const MusicLibrary: KNOWNFOLDERID

pub const PicturesLibrary: KNOWNFOLDERID

pub const VideosLibrary: KNOWNFOLDERID

pub const RecordedTVLibrary: KNOWNFOLDERID

pub const HomeGroup: KNOWNFOLDERID

pub const HomeGroupCurrentUser: KNOWNFOLDERID

pub const DeviceMetadataStore: KNOWNFOLDERID

pub const Libraries: KNOWNFOLDERID

pub const PublicLibraries: KNOWNFOLDERID

pub const UserPinned: KNOWNFOLDERID

pub const ImplicitAppShortcuts: KNOWNFOLDERID

pub const AccountPictures: KNOWNFOLDERID

pub const PublicUserTiles: KNOWNFOLDERID

pub const AppsFolder: KNOWNFOLDERID

pub const StartMenuAllPrograms: KNOWNFOLDERID

pub const CommonStartMenuPlaces: KNOWNFOLDERID

pub const ApplicationShortcuts: KNOWNFOLDERID

pub const RoamingTiles: KNOWNFOLDERID

pub const RoamedTileImages: KNOWNFOLDERID

pub const Screenshots: KNOWNFOLDERID

pub const CameraRoll: KNOWNFOLDERID

pub const SkyDrive: KNOWNFOLDERID

pub const OneDrive: KNOWNFOLDERID

pub const SkyDriveDocuments: KNOWNFOLDERID

pub const SkyDrivePictures: KNOWNFOLDERID

pub const SkyDriveMusic: KNOWNFOLDERID

pub const SkyDriveCameraRoll: KNOWNFOLDERID

pub const SearchHistory: KNOWNFOLDERID

pub const SearchTemplates: KNOWNFOLDERID

pub const CameraRollLibrary: KNOWNFOLDERID

pub const SavedPictures: KNOWNFOLDERID

pub const SavedPicturesLibrary: KNOWNFOLDERID

pub const RetailDemo: KNOWNFOLDERID

pub const Device: KNOWNFOLDERID

pub const DevelopmentFiles: KNOWNFOLDERID

pub const Objects3D: KNOWNFOLDERID

pub const AppCaptures: KNOWNFOLDERID

pub const LocalDocuments: KNOWNFOLDERID

pub const LocalPictures: KNOWNFOLDERID

pub const LocalVideos: KNOWNFOLDERID

pub const LocalMusic: KNOWNFOLDERID

pub const LocalDownloads: KNOWNFOLDERID

pub const RecordedCalls: KNOWNFOLDERID

pub const AllAppMods: KNOWNFOLDERID

pub const CurrentAppMods: KNOWNFOLDERID

pub const AppDataDesktop: KNOWNFOLDERID

pub const AppDataDocuments: KNOWNFOLDERID

pub const AppDataFavorites: KNOWNFOLDERID

pub const AppDataProgramData: KNOWNFOLDERID