Struct winsafe::co::LANG

pub struct LANG(/* private fields */);

Available on crate feature kernel only.

Language identifier (u16).

Implementations

impl LANG

pub const unsafe fn from_raw(v: u16) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u16

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl LANG

pub const NEUTRAL: Self = _

pub const INVARIANT: Self = _

pub const AFRIKAANS: Self = _

pub const ALBANIAN: Self = _

pub const ALSATIAN: Self = _

pub const AMHARIC: Self = _

pub const ARABIC: Self = _

pub const ARMENIAN: Self = _

pub const ASSAMESE: Self = _

pub const AZERI: Self = _

pub const AZERBAIJANI: Self = _

pub const BANGLA: Self = _

pub const BASHKIR: Self = _

pub const BASQUE: Self = _

pub const BELARUSIAN: Self = _

pub const BENGALI: Self = _

pub const BRETON: Self = _

pub const BOSNIAN: Self = _

pub const BOSNIAN_NEUTRAL: Self = _

pub const BULGARIAN: Self = _

pub const CATALAN: Self = _

pub const CENTRAL_KURDISH: Self = _

pub const CHEROKEE: Self = _

pub const CHINESE: Self = _

pub const CHINESE_SIMPLIFIED: Self = _

pub const CHINESE_TRADITIONAL: Self = _

pub const CORSICAN: Self = _

pub const CROATIAN: Self = _

pub const CZECH: Self = _

pub const DANISH: Self = _

pub const DARI: Self = _

pub const DIVEHI: Self = _

pub const DUTCH: Self = _

pub const ENGLISH: Self = _

pub const ESTONIAN: Self = _

pub const FAEROESE: Self = _

pub const FARSI: Self = _

pub const FILIPINO: Self = _

pub const FINNISH: Self = _

pub const FRENCH: Self = _

pub const FRISIAN: Self = _

pub const FULAH: Self = _

pub const GALICIAN: Self = _

pub const GEORGIAN: Self = _

pub const GERMAN: Self = _

pub const GREEK: Self = _

pub const GREENLANDIC: Self = _

pub const GUJARATI: Self = _

pub const HAUSA: Self = _

pub const HAWAIIAN: Self = _

pub const HEBREW: Self = _

pub const HINDI: Self = _

pub const HUNGARIAN: Self = _

pub const ICELANDIC: Self = _

pub const IGBO: Self = _

pub const INDONESIAN: Self = _

pub const INUKTITUT: Self = _

pub const IRISH: Self = _

pub const ITALIAN: Self = _

pub const JAPANESE: Self = _

pub const KANNADA: Self = _

pub const KASHMIRI: Self = _

pub const KAZAK: Self = _

pub const KHMER: Self = _

pub const KICHE: Self = _

pub const KINYARWANDA: Self = _

pub const KONKANI: Self = _

pub const KOREAN: Self = _

pub const KYRGYZ: Self = _

pub const LAO: Self = _

pub const LATVIAN: Self = _

pub const LITHUANIAN: Self = _

pub const LOWER_SORBIAN: Self = _

pub const LUXEMBOURGISH: Self = _

pub const MACEDONIAN: Self = _

pub const MALAY: Self = _

pub const MALAYALAM: Self = _

pub const MALTESE: Self = _

pub const MANIPURI: Self = _

pub const MAORI: Self = _

pub const MAPUDUNGUN: Self = _

pub const MARATHI: Self = _

pub const MOHAWK: Self = _

pub const MONGOLIAN: Self = _

pub const NEPALI: Self = _

pub const NORWEGIAN: Self = _

pub const OCCITAN: Self = _

pub const ODIA: Self = _

pub const ORIYA: Self = _

pub const PASHTO: Self = _

pub const PERSIAN: Self = _

pub const POLISH: Self = _

pub const PORTUGUESE: Self = _

pub const PULAR: Self = _

pub const PUNJABI: Self = _

pub const QUECHUA: Self = _

pub const ROMANIAN: Self = _

pub const ROMANSH: Self = _

pub const RUSSIAN: Self = _

pub const SAKHA: Self = _

pub const SAMI: Self = _

pub const SANSKRIT: Self = _

pub const SCOTTISH_GAELIC: Self = _

pub const SERBIAN: Self = _

pub const SERBIAN_NEUTRAL: Self = _

pub const SINDHI: Self = _

pub const SINHALESE: Self = _

pub const SLOVAK: Self = _

pub const SLOVENIAN: Self = _

pub const SOTHO: Self = _

pub const SPANISH: Self = _

pub const SWAHILI: Self = _

pub const SWEDISH: Self = _

pub const SYRIAC: Self = _

pub const TAJIK: Self = _

pub const TAMAZIGHT: Self = _

pub const TAMIL: Self = _

pub const TATAR: Self = _

pub const TELUGU: Self = _

pub const THAI: Self = _

pub const TIBETAN: Self = _

pub const TIGRIGNA: Self = _

pub const TIGRINYA: Self = _

pub const TSWANA: Self = _

pub const TURKISH: Self = _

pub const TURKMEN: Self = _

pub const UIGHUR: Self = _

pub const UKRAINIAN: Self = _

pub const UPPER_SORBIAN: Self = _

pub const URDU: Self = _

pub const UZBEK: Self = _

pub const VALENCIAN: Self = _

pub const VIETNAMESE: Self = _

pub const WELSH: Self = _

pub const WOLOF: Self = _

pub const XHOSA: Self = _

pub const YAKUT: Self = _

pub const YI: Self = _

pub const YORUBA: Self = _

pub const ZULU: Self = _