Struct winsafe::co::LVM

pub struct LVM(/* private fields */);

Available on crate features kernel and comctl only.

List view control messages (u32).

This is a window message, convertible to WM.

Implementations

impl LVM

pub const unsafe fn from_raw(v: u32) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u32

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl LVM

pub const SETUNICODEFORMAT: Self = _

pub const GETUNICODEFORMAT: Self = _

pub const GETBKCOLOR: Self = _

pub const SETBKCOLOR: Self = _

pub const GETIMAGELIST: Self = _

pub const SETIMAGELIST: Self = _

pub const GETITEMCOUNT: Self = _

pub const DELETEITEM: Self = _

pub const DELETEALLITEMS: Self = _

pub const GETCALLBACKMASK: Self = _

pub const SETCALLBACKMASK: Self = _

pub const GETNEXTITEM: Self = _

pub const GETITEMRECT: Self = _

pub const SETITEMPOSITION: Self = _

pub const GETITEMPOSITION: Self = _

pub const HITTEST: Self = _

pub const ENSUREVISIBLE: Self = _

pub const SCROLL: Self = _

pub const REDRAWITEMS: Self = _

pub const ARRANGE: Self = _

pub const GETEDITCONTROL: Self = _

pub const DELETECOLUMN: Self = _

pub const GETCOLUMNWIDTH: Self = _

pub const SETCOLUMNWIDTH: Self = _

pub const GETHEADER: Self = _

pub const CREATEDRAGIMAGE: Self = _

pub const GETVIEWRECT: Self = _

pub const GETTEXTCOLOR: Self = _

pub const SETTEXTCOLOR: Self = _

pub const GETTEXTBKCOLOR: Self = _

pub const SETTEXTBKCOLOR: Self = _

pub const GETTOPINDEX: Self = _

pub const GETCOUNTPERPAGE: Self = _

pub const GETORIGIN: Self = _

pub const UPDATE: Self = _

pub const SETITEMSTATE: Self = _

pub const GETITEMSTATE: Self = _

pub const SETITEMCOUNT: Self = _

pub const SORTITEMS: Self = _

pub const SETITEMPOSITION32: Self = _

pub const GETSELECTEDCOUNT: Self = _

pub const GETITEMSPACING: Self = _

pub const SETICONSPACING: Self = _

pub const SETEXTENDEDLISTVIEWSTYLE: Self = _

pub const GETEXTENDEDLISTVIEWSTYLE: Self = _

pub const GETSUBITEMRECT: Self = _

pub const SUBITEMHITTEST: Self = _

pub const SETCOLUMNORDERARRAY: Self = _

pub const GETCOLUMNORDERARRAY: Self = _

pub const SETHOTITEM: Self = _

pub const GETHOTITEM: Self = _

pub const SETHOTCURSOR: Self = _

pub const GETHOTCURSOR: Self = _

pub const APPROXIMATEVIEWRECT: Self = _

pub const SETWORKAREAS: Self = _

pub const GETSELECTIONMARK: Self = _

pub const SETSELECTIONMARK: Self = _

pub const GETWORKAREAS: Self = _

pub const SETHOVERTIME: Self = _

pub const GETHOVERTIME: Self = _

pub const GETNUMBEROFWORKAREAS: Self = _

pub const SETTOOLTIPS: Self = _

pub const GETITEM: Self = _

pub const SETITEM: Self = _

pub const INSERTITEM: Self = _

pub const GETTOOLTIPS: Self = _

pub const SORTITEMSEX: Self = _

pub const FINDITEM: Self = _

pub const GETSTRINGWIDTH: Self = _

pub const GETGROUPSTATE: Self = _

pub const GETFOCUSEDGROUP: Self = _

pub const GETCOLUMN: Self = _

pub const SETCOLUMN: Self = _

pub const INSERTCOLUMN: Self = _

pub const GETGROUPRECT: Self = _

pub const GETITEMTEXT: Self = _

pub const SETITEMTEXT: Self = _

pub const GETISEARCHSTRING: Self = _

pub const EDITLABEL: Self = _

pub const SETBKIMAGE: Self = _

pub const GETBKIMAGE: Self = _

pub const SETSELECTEDCOLUMN: Self = _

pub const SETVIEW: Self = _

pub const GETVIEW: Self = _

pub const INSERTGROUP: Self = _

pub const SETGROUPINFO: Self = _

pub const GETGROUPINFO: Self = _

pub const REMOVEGROUP: Self = _

pub const MOVEGROUP: Self = _

pub const GETGROUPCOUNT: Self = _

pub const GETGROUPINFOBYINDEX: Self = _

pub const MOVEITEMTOGROUP: Self = _

pub const SETGROUPMETRICS: Self = _

pub const GETGROUPMETRICS: Self = _

pub const ENABLEGROUPVIEW: Self = _

pub const SORTGROUPS: Self = _

pub const INSERTGROUPSORTED: Self = _

pub const REMOVEALLGROUPS: Self = _

pub const HASGROUP: Self = _

pub const SETTILEVIEWINFO: Self = _

pub const GETTILEVIEWINFO: Self = _

pub const SETTILEINFO: Self = _

pub const GETTILEINFO: Self = _

pub const SETINSERTMARK: Self = _

pub const GETINSERTMARK: Self = _

pub const INSERTMARKHITTEST: Self = _

pub const GETINSERTMARKRECT: Self = _

pub const SETINSERTMARKCOLOR: Self = _

pub const GETINSERTMARKCOLOR: Self = _

pub const SETINFOTIP: Self = _

pub const GETSELECTEDCOLUMN: Self = _

pub const ISGROUPVIEWENABLED: Self = _

pub const GETOUTLINECOLOR: Self = _

pub const SETOUTLINECOLOR: Self = _

pub const CANCELEDITLABEL: Self = _

pub const MAPINDEXTOID: Self = _

pub const MAPIDTOINDEX: Self = _

pub const ISITEMVISIBLE: Self = _

pub const GETEMPTYTEXT: Self = _

pub const GETFOOTERRECT: Self = _

pub const GETFOOTERINFO: Self = _

pub const GETFOOTERITEMRECT: Self = _

pub const GETFOOTERITEM: Self = _

pub const GETITEMINDEXRECT: Self = _

pub const SETITEMINDEXSTATE: Self = _

pub const GETNEXTITEMINDEX: Self = _