Struct winsafe::co::MB

#[repr(transparent)]
pub struct MB(_);

HWND::MessageBox flags (u32).

Implementations

impl MB

pub const ABORTRETRYIGNORE: Self

The message box contains three push buttons: Abort Retry and Ignore.

pub const CANCELTRYCONTINUE: Self

The message box contains three push buttons: Cancel Try Again, Continue. Use this message box type instead of MB::ABORTRETRYIGNORE.

pub const HELP: Self

Adds a Help button to the message box. When the user clicks the Help button or presses F1 the system sends a wm::Help message to the owner.

pub const OK: Self

The message box contains one push button: OK. This is the default.

pub const OKCANCEL: Self

The message box contains two push buttons: OK and Cancel.

pub const RETRYCANCEL: Self

The message box contains two push buttons: Retry and Cancel.

pub const YESNO: Self

The message box contains two push buttons: Yes and No.

pub const YESNOCANCEL: Self

The message box contains three push buttons: Yes No and Cancel.

pub const ICONEXCLAMATION: Self

An exclamation-point icon appears in the message box.

pub const ICONWARNING: Self

An exclamation-point icon appears in the message box.

pub const ICONINFORMATION: Self

An icon consisting of a lowercase letter i in a circle appears in the message box.

pub const ICONASTERISK: Self

An icon consisting of a lowercase letter i in a circle appears in the message box.

pub const ICONQUESTION: Self

A question-mark icon appears in the message box. The question-mark message icon is no longer recommended because it does not clearly represent a specific type of message and because the phrasing of a message as a question could apply to any message type. In addition, users can confuse the message symbol question mark with Help information. Therefore do not use this question mark message symbol in your message boxes. The system continues to support its inclusion only for backward compatibility.

pub const ICONSTOP: Self

A stop-sign icon appears in the message box.

pub const ICONERROR: Self

A stop-sign icon appears in the message box.

pub const ICONHAND: Self

A stop-sign icon appears in the message box.

pub const DEFBUTTON1: Self

The first button is the default button. MB::DEFBUTTON1 is the default unless MB::DEFBUTTON2, MB::DEFBUTTON3 or MB::DEFBUTTON4 is specified.

pub const DEFBUTTON2: Self

The second button is the default button.

pub const DEFBUTTON3: Self

The third button is the default button.

pub const DEFBUTTON4: Self

The fourth button is the default button.

pub const APPLMODAL: Self

The user must respond to the message box before continuing work in the window identified by the hWnd parameter. However the user can move to the windows of other threads and work in those windows.

Depending on the hierarchy of windows in the application the user may be able to move to other windows within the thread. All child windows of the parent of the message box are automatically disabled but pop-up windows are not.

MB::APPLMODAL is the default if neither MB::SYSTEMMODAL nor MB::TASKMODAL is specified.

pub const SYSTEMMODAL: Self

Same as MB::APPLMODAL except that the message box has the WS_EX::TOPMOST style. Use system-modal message boxes to notify the user of serious, potentially damaging errors that require immediate attention (for example running out of memory). This flag has no effect on the user’s ability to interact with windows other than those associated with hWnd.

pub const TASKMODAL: Self

Same as MB::APPLMODAL except that all the top-level windows belonging to the current thread are disabled if the hWnd parameter is NULL. Use this flag when the calling application or library does not have a window handle available but still needs to prevent input to other windows in the calling thread without suspending other threads.

pub const DEFAULT_DESKTOP_ONLY: Self

Same as desktop of the interactive window station. For more information, see Window Stations.

If the current input desktop is not the default desktop, HWND::MessageBox does not return until the user switches to the default desktop.

pub const RIGHT: Self

The text is right-justified.

pub const RTLREADING: Self

Displays message and caption text using right-to-left reading order on Hebrew and Arabic systems.

pub const SETFOREGROUND: Self

The message box becomes the foreground window. Internally the system calls the HWND::SetForegroundWindow function for the message box.

pub const TOPMOST: Self

The message box is created with the WS_EX::TOPMOST window style.

pub const SERVICE_NOTIFICATION: Self

The caller is a service notifying the user of an event. The function displays a message box on the current active desktop even if there is no user logged on to the computer.

Terminal Services: If the calling thread has an impersonation token the function directs the message box to the session specified in the impersonation token.

If this flag is set the hWnd parameter must be NULL. This is so that the message box can appear on a desktop other than the desktop corresponding to the hWnd.

For information on security considerations in regard to using this flag, see Interactive Services. In particular be aware that this flag can produce interactive content on a locked desktop and should therefore be used for only a very limited set of scenarios such as resource exhaustion.