Struct winsafe::co::ME

pub struct ME(/* private fields */);

Available on crate features kernel and mf only.

IMFMediaEvent::GetType return value (u32).

Implementations

impl ME

pub const unsafe fn from_raw(v: u32) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u32

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl ME

pub const Unknown: Self = _

pub const Error: Self = _

pub const ExtendedType: Self = _

pub const NonFatalError: Self = _

pub const GenericV1Anchor: Self = _

pub const SessionUnknown: Self = _

pub const SessionTopologySet: Self = _

pub const SessionTopologiesCleared: Self = _

pub const SessionStarted: Self = _

pub const SessionPaused: Self = _

pub const SessionStopped: Self = _

pub const SessionClosed: Self = _

pub const SessionEnded: Self = _

pub const SessionRateChanged: Self = _

pub const SessionScrubSampleComplete: Self = _

pub const SessionCapabilitiesChanged: Self = _

pub const SessionTopologyStatus: Self = _

pub const SessionNotifyPresentationTime: Self = _

pub const NewPresentation: Self = _

pub const LicenseAcquisitionStart: Self = _

pub const LicenseAcquisitionCompleted: Self = _

pub const IndividualizationStart: Self = _

pub const IndividualizationCompleted: Self = _

pub const EnablerProgress: Self = _

pub const EnablerCompleted: Self = _

pub const PolicyError: Self = _

pub const PolicyReport: Self = _

pub const BufferingStarted: Self = _

pub const BufferingStopped: Self = _

pub const ConnectStart: Self = _

pub const ConnectEnd: Self = _

pub const ReconnectStart: Self = _

pub const ReconnectEnd: Self = _

pub const RendererEvent: Self = _

pub const SessionStreamSinkFormatChanged: Self = _

pub const SessionV1Anchor: Self = _

pub const SourceUnknown: Self = _

pub const SourceStarted: Self = _

pub const StreamStarted: Self = _

pub const SourceSeeked: Self = _

pub const StreamSeeked: Self = _

pub const NewStream: Self = _

pub const UpdatedStream: Self = _

pub const SourceStopped: Self = _

pub const StreamStopped: Self = _

pub const SourcePaused: Self = _

pub const StreamPaused: Self = _

pub const EndOfPresentation: Self = _

pub const EndOfStream: Self = _

pub const MediaSample: Self = _

pub const StreamTick: Self = _

pub const StreamThinMode: Self = _

pub const StreamFormatChanged: Self = _

pub const SourceRateChanged: Self = _

pub const EndOfPresentationSegment: Self = _

pub const SourceCharacteristicsChanged: Self = _

pub const SourceRateChangeRequested: Self = _

pub const SourceMetadataChanged: Self = _

pub const SequencerSourceTopologyUpdated: Self = _

pub const SourceV1Anchor: Self = _

pub const SinkUnknown: Self = _

pub const StreamSinkStarted: Self = _

pub const StreamSinkStopped: Self = _

pub const StreamSinkPaused: Self = _

pub const StreamSinkRateChanged: Self = _

pub const StreamSinkRequestSample: Self = _

pub const StreamSinkMarker: Self = _

pub const StreamSinkPrerolled: Self = _

pub const StreamSinkScrubSampleComplete: Self = _

pub const StreamSinkFormatChanged: Self = _

pub const StreamSinkDeviceChanged: Self = _

pub const QualityNotify: Self = _

pub const SinkInvalidated: Self = _

pub const AudioSessionNameChanged: Self = _

pub const AudioSessionVolumeChanged: Self = _

pub const AudioSessionDeviceRemoved: Self = _

pub const AudioSessionServerShutdown: Self = _

pub const AudioSessionGroupingParamChanged: Self = _

pub const AudioSessionIconChanged: Self = _

pub const AudioSessionFormatChanged: Self = _

pub const AudioSessionDisconnected: Self = _

pub const AudioSessionExclusiveModeOverride: Self = _

pub const SinkV1Anchor: Self = _

pub const CaptureAudioSessionVolumeChanged: Self = _

pub const CaptureAudioSessionDeviceRemoved: Self = _

pub const CaptureAudioSessionFormatChanged: Self = _

pub const CaptureAudioSessionDisconnected: Self = _

pub const CaptureAudioSessionExclusiveModeOverride: Self = _

pub const CaptureAudioSessionServerShutdown: Self = _

pub const SinkV2Anchor: Self = _

pub const TrustUnknown: Self = _

pub const PolicyChanged: Self = _

pub const ContentProtectionMessage: Self = _

pub const PolicySet: Self = _

pub const TrustV1Anchor: Self = _

pub const WMDRMLicenseBackupCompleted: Self = _

pub const WMDRMLicenseBackupProgress: Self = _

pub const WMDRMLicenseRestoreCompleted: Self = _

pub const WMDRMLicenseRestoreProgress: Self = _

pub const WMDRMLicenseAcquisitionCompleted: Self = _

pub const WMDRMIndividualizationCompleted: Self = _

pub const WMDRMIndividualizationProgress: Self = _

pub const WMDRMProximityCompleted: Self = _

pub const WMDRMLicenseStoreCleaned: Self = _

pub const WMDRMRevocationDownloadCompleted: Self = _

pub const WMDRMV1Anchor: Self = _

pub const TransformUnknown: Self = _

pub const TransformNeedInput: Self = _

pub const TransformHaveOutput: Self = _

pub const TransformDrainComplete: Self = _

pub const TransformMarker: Self = _

pub const TransformInputStreamStateChanged: Self = _

pub const ByteStreamCharacteristicsChanged: Self = _

pub const VideoCaptureDeviceRemoved: Self = _

pub const VideoCaptureDevicePreempted: Self = _

pub const StreamSinkFormatInvalidated: Self = _

pub const EncodingParameters: Self = _

pub const ContentProtectionMetadata: Self = _

pub const DeviceThermalStateChanged: Self = _