Struct winsafe::co::SE_PRIV

pub struct SE_PRIV(_);

Available on crate feature kernel only.

NT defined privileges for LookupPrivilegeValue (&'static str).

Originally has SE prefix.

Implementations

impl SE_PRIV

pub const unsafe fn from_raw(v: &'static str) -> Self

Constructs a new object by wrapping the given &'static str value.

Safety

Be sure the given value is meaningful for the actual type.

pub const CREATE_TOKEN_NAME: Self = _

pub const ASSIGNPRIMARYTOKEN_NAME: Self = _

pub const LOCK_MEMORY_NAME: Self = _

pub const INCREASE_QUOTA_NAME: Self = _

pub const UNSOLICITED_INPUT_NAME: Self = _

pub const MACHINE_ACCOUNT_NAME: Self = _

pub const TCB_NAME: Self = _

pub const SECURITY_NAME: Self = _

pub const TAKE_OWNERSHIP_NAME: Self = _

pub const LOAD_DRIVER_NAME: Self = _

pub const SYSTEM_PROFILE_NAME: Self = _

pub const SYSTEMTIME_NAME: Self = _

pub const PROF_SINGLE_PROCESS_NAME: Self = _

pub const INC_BASE_PRIORITY_NAME: Self = _

pub const CREATE_PAGEFILE_NAME: Self = _

pub const CREATE_PERMANENT_NAME: Self = _

pub const BACKUP_NAME: Self = _

pub const RESTORE_NAME: Self = _

pub const SHUTDOWN_NAME: Self = _

pub const DEBUG_NAME: Self = _

pub const AUDIT_NAME: Self = _

pub const SYSTEM_ENVIRONMENT_NAME: Self = _

pub const CHANGE_NOTIFY_NAME: Self = _

pub const REMOTE_SHUTDOWN_NAME: Self = _

pub const UNDOCK_NAME: Self = _

pub const SYNC_AGENT_NAME: Self = _

pub const ENABLE_DELEGATION_NAME: Self = _

pub const MANAGE_VOLUME_NAME: Self = _

pub const IMPERSONATE_NAME: Self = _

pub const CREATE_GLOBAL_NAME: Self = _

pub const TRUSTED_CREDMAN_ACCESS_NAME: Self = _

pub const RELABEL_NAME: Self = _

pub const INC_WORKING_SET_NAME: Self = _

pub const TIME_ZONE_NAME: Self = _

pub const CREATE_SYMBOLIC_LINK_NAME: Self = _

pub const DELEGATE_SESSION_USER_IMPERSONATE_NAME: Self = _