Struct winsafe::co::SIID

pub struct SIID(/* private fields */);

Available on crate features kernel and shell only.

SHSTOCKICONID enumeration, SHGetStockIconInfo siid (u32).

Implementations

impl SIID

pub const unsafe fn from_raw(v: u32) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u32

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl SIID

pub const DOCNOASSOC: Self = _

pub const DOCASSOC: Self = _

pub const APPLICATION: Self = _

pub const FOLDER: Self = _

pub const FOLDEROPEN: Self = _

pub const DRIVE525: Self = _

pub const DRIVE35: Self = _

pub const DRIVEREMOVE: Self = _

pub const DRIVEFIXED: Self = _

pub const DRIVENET: Self = _

pub const DRIVENETDISABLED: Self = _

pub const DRIVECD: Self = _

pub const DRIVERAM: Self = _

pub const WORLD: Self = _

pub const SERVER: Self = _

pub const PRINTER: Self = _

pub const MYNETWORK: Self = _

pub const FIND: Self = _

pub const HELP: Self = _

pub const SHARE: Self = _

pub const LINK: Self = _

pub const SLOWFILE: Self = _

pub const RECYCLER: Self = _

pub const RECYCLERFULL: Self = _

pub const MEDIACDAUDIO: Self = _

pub const LOCK: Self = _

pub const AUTOLIST: Self = _

pub const PRINTERNET: Self = _

pub const SERVERSHARE: Self = _

pub const PRINTERFAX: Self = _

pub const PRINTERFAXNET: Self = _

pub const PRINTERFILE: Self = _

pub const STACK: Self = _

pub const MEDIASVCD: Self = _

pub const STUFFEDFOLDER: Self = _

pub const DRIVEUNKNOWN: Self = _

pub const DRIVEDVD: Self = _

pub const MEDIADVD: Self = _

pub const MEDIADVDRAM: Self = _

pub const MEDIADVDRW: Self = _

pub const MEDIADVDR: Self = _

pub const MEDIADVDROM: Self = _

pub const MEDIACDAUDIOPLUS: Self = _

pub const MEDIACDRW: Self = _

pub const MEDIACDR: Self = _

pub const MEDIACDBURN: Self = _

pub const MEDIABLANKCD: Self = _

pub const MEDIACDROM: Self = _

pub const AUDIOFILES: Self = _

pub const IMAGEFILES: Self = _

pub const VIDEOFILES: Self = _

pub const MIXEDFILES: Self = _

pub const FOLDERBACK: Self = _

pub const FOLDERFRONT: Self = _

pub const SHIELD: Self = _

pub const WARNING: Self = _

pub const INFO: Self = _

pub const ERROR: Self = _

pub const KEY: Self = _

pub const SOFTWARE: Self = _

pub const RENAME: Self = _

pub const DELETE: Self = _

pub const MEDIAAUDIODVD: Self = _

pub const MEDIAMOVIEDVD: Self = _

pub const MEDIAENHANCEDCD: Self = _

pub const MEDIAENHANCEDDVD: Self = _

pub const MEDIAHDDVD: Self = _

pub const MEDIABLURAY: Self = _

pub const MEDIAVCD: Self = _

pub const MEDIADVDPLUSR: Self = _

pub const MEDIADVDPLUSRW: Self = _

pub const DESKTOPPC: Self = _

pub const MOBILEPC: Self = _

pub const USERS: Self = _

pub const MEDIASMARTMEDIA: Self = _

pub const MEDIACOMPACTFLASH: Self = _

pub const DEVICECELLPHONE: Self = _

pub const DEVICECAMERA: Self = _

pub const DEVICEVIDEOCAMERA: Self = _

pub const DEVICEAUDIOPLAYER: Self = _

pub const NETWORKCONNECT: Self = _

pub const INTERNET: Self = _

pub const ZIPFILE: Self = _

pub const SETTINGS: Self = _

pub const DRIVEHDDVD: Self = _

pub const DRIVEBD: Self = _

pub const MEDIAHDDVDROM: Self = _

pub const MEDIAHDDVDR: Self = _

pub const MEDIAHDDVDRAM: Self = _

pub const MEDIABDROM: Self = _

pub const MEDIABDR: Self = _

pub const MEDIABDRE: Self = _

pub const CLUSTEREDDRIVE: Self = _

pub const MAX_ICONS: Self = _