winsafe::co

Struct SM

pub struct SM(/* private fields */);

Available on crate feature user only.

GetSystemMetrics index (i32).

Implementations

impl SM

pub const unsafe fn as_mut(&mut self) -> &mut i32

Returns a mutable reference to the underlying raw value.

Safety

Be sure the integer being set is meaningful for the actual type.

pub const unsafe fn from_raw(v: i32) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> i32

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl SM

pub const CXSCREEN: Self

pub const CYSCREEN: Self

pub const CXVSCROLL: Self

pub const CYHSCROLL: Self

pub const CYCAPTION: Self

pub const CXBORDER: Self

pub const CYBORDER: Self

pub const CXDLGFRAME: Self

pub const CYDLGFRAME: Self

pub const CYVTHUMB: Self

pub const CXHTHUMB: Self

pub const CXICON: Self

pub const CYICON: Self

pub const CXCURSOR: Self

pub const CYCURSOR: Self

pub const CYMENU: Self

pub const CXFULLSCREEN: Self

pub const CYFULLSCREEN: Self

pub const CYKANJIWINDOW: Self

pub const MOUSEPRESENT: Self

pub const CYVSCROLL: Self

pub const CXHSCROLL: Self

pub const DEBUG: Self

pub const SWAPBUTTON: Self

pub const RESERVED1: Self

pub const RESERVED2: Self

pub const RESERVED3: Self

pub const RESERVED4: Self

pub const CXMIN: Self

The minimum width of a window in pixels.

pub const CYMIN: Self

The minimum height of a window in pixels.

pub const CXSIZE: Self

The width of a button in a window caption or title bar in pixels.

pub const CYSIZE: Self

The height of a button in a window caption or title bar in pixels.

pub const CXFRAME: Self

pub const CYFRAME: Self

pub const CXMINTRACK: Self

The minimum tracking width of a window in pixels. The user cannot drag the window frame to a size smaller than these dimensions. A window can override this value by processing the wm::GetMinMaxInfo message.

pub const CYMINTRACK: Self

The minimum tracking height of a window in pixels. The user cannot drag the window frame to a size smaller than these dimensions. A window can override this value by processing the wm::GetMinMaxInfo message.

pub const CXDOUBLECLK: Self

pub const CYDOUBLECLK: Self

pub const CXICONSPACING: Self

pub const CYICONSPACING: Self

pub const MENUDROPALIGNMENT: Self

pub const PENWINDOWS: Self

pub const DBCSENABLED: Self

pub const CMOUSEBUTTONS: Self

pub const CXFIXEDFRAME: Self

pub const CYFIXEDFRAME: Self

pub const CXSIZEFRAME: Self

pub const CYSIZEFRAME: Self

pub const SECURE: Self

pub const CXEDGE: Self

pub const CYEDGE: Self

pub const CXMINSPACING: Self

pub const CYMINSPACING: Self

pub const CXSMICON: Self

pub const CYSMICON: Self

pub const CYSMCAPTION: Self

pub const CXSMSIZE: Self

pub const CYSMSIZE: Self

pub const CXMENUSIZE: Self

pub const CYMENUSIZE: Self

pub const ARRANGE: Self

pub const CXMINIMIZED: Self

pub const CYMINIMIZED: Self

pub const CXMAXTRACK: Self

pub const CYMAXTRACK: Self

pub const CXMAXIMIZED: Self

pub const CYMAXIMIZED: Self

pub const NETWORK: Self

pub const CLEANBOOT: Self

pub const CXDRAG: Self

pub const CYDRAG: Self

pub const SHOWSOUNDS: Self

pub const CXMENUCHECK: Self

pub const CYMENUCHECK: Self

pub const SLOWMACHINE: Self

pub const MIDEASTENABLED: Self

pub const MOUSEWHEELPRESENT: Self

pub const XVIRTUALSCREEN: Self

pub const YVIRTUALSCREEN: Self

pub const CXVIRTUALSCREEN: Self

pub const CYVIRTUALSCREEN: Self

pub const CMONITORS: Self

pub const SAMEDISPLAYFORMAT: Self

pub const IMMENABLED: Self

pub const CXFOCUSBORDER: Self

pub const CYFOCUSBORDER: Self

pub const TABLETPC: Self

pub const MEDIACENTER: Self

pub const STARTER: Self

pub const SERVERR2: Self

pub const MOUSEHORIZONTALWHEELPRESENT: Self

pub const CXPADDEDBORDER: Self

pub const DIGITIZER: Self

pub const MAXIMUMTOUCHES: Self

pub const CMETRICS: Self

pub const REMOTESESSION: Self

pub const SHUTTINGDOWN: Self

pub const REMOTECONTROL: Self

pub const CARETBLINKINGENABLED: Self

pub const CONVERTIBLESLATEMODE: Self

pub const SYSTEMDOCKED: Self