Struct winsafe::co::SPI

pub struct SPI(/* private fields */);

Available on crate features kernel and user only.

SystemParametersInfo action (u32).

Implementations

impl SPI

pub const unsafe fn from_raw(v: u32) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u32

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl SPI

pub const GETBEEP: Self = _

pub const SETBEEP: Self = _

pub const GETMOUSE: Self = _

pub const SETMOUSE: Self = _

pub const GETBORDER: Self = _

pub const SETBORDER: Self = _

pub const GETKEYBOARDSPEED: Self = _

pub const SETKEYBOARDSPEED: Self = _

pub const LANGDRIVER: Self = _

pub const ICONHORIZONTALSPACING: Self = _

pub const GETSCREENSAVETIMEOUT: Self = _

pub const SETSCREENSAVETIMEOUT: Self = _

pub const GETSCREENSAVEACTIVE: Self = _

pub const SETSCREENSAVEACTIVE: Self = _

pub const GETGRIDGRANULARITY: Self = _

pub const SETGRIDGRANULARITY: Self = _

pub const SETDESKWALLPAPER: Self = _

pub const SETDESKPATTERN: Self = _

pub const GETKEYBOARDDELAY: Self = _

pub const SETKEYBOARDDELAY: Self = _

pub const ICONVERTICALSPACING: Self = _

pub const GETICONTITLEWRAP: Self = _

pub const SETICONTITLEWRAP: Self = _

pub const GETMENUDROPALIGNMENT: Self = _

pub const SETMENUDROPALIGNMENT: Self = _

pub const SETDOUBLECLKWIDTH: Self = _

pub const SETDOUBLECLKHEIGHT: Self = _

pub const GETICONTITLELOGFONT: Self = _

pub const SETDOUBLECLICKTIME: Self = _

pub const SETMOUSEBUTTONSWAP: Self = _

pub const SETICONTITLELOGFONT: Self = _

pub const GETFASTTASKSWITCH: Self = _

pub const SETFASTTASKSWITCH: Self = _

pub const SETDRAGFULLWINDOWS: Self = _

pub const GETDRAGFULLWINDOWS: Self = _

pub const GETNONCLIENTMETRICS: Self = _

pub const SETNONCLIENTMETRICS: Self = _

pub const GETMINIMIZEDMETRICS: Self = _

pub const SETMINIMIZEDMETRICS: Self = _

pub const GETICONMETRICS: Self = _

pub const SETICONMETRICS: Self = _

pub const SETWORKAREA: Self = _

pub const GETWORKAREA: Self = _

pub const SETPENWINDOWS: Self = _

pub const GETHIGHCONTRAST: Self = _

pub const SETHIGHCONTRAST: Self = _

pub const GETKEYBOARDPREF: Self = _

pub const SETKEYBOARDPREF: Self = _

pub const GETSCREENREADER: Self = _

pub const SETSCREENREADER: Self = _

pub const GETANIMATION: Self = _

pub const SETANIMATION: Self = _

pub const GETFONTSMOOTHING: Self = _

pub const SETFONTSMOOTHING: Self = _

pub const SETDRAGWIDTH: Self = _

pub const SETDRAGHEIGHT: Self = _

pub const SETHANDHELD: Self = _

pub const GETLOWPOWERTIMEOUT: Self = _

pub const GETPOWEROFFTIMEOUT: Self = _

pub const SETLOWPOWERTIMEOUT: Self = _

pub const SETPOWEROFFTIMEOUT: Self = _

pub const GETLOWPOWERACTIVE: Self = _

pub const GETPOWEROFFACTIVE: Self = _

pub const SETLOWPOWERACTIVE: Self = _

pub const SETPOWEROFFACTIVE: Self = _

pub const SETCURSORS: Self = _

pub const SETICONS: Self = _

pub const GETDEFAULTINPUTLANG: Self = _

pub const SETDEFAULTINPUTLANG: Self = _

pub const SETLANGTOGGLE: Self = _

pub const GETWINDOWSEXTENSION: Self = _

pub const SETMOUSETRAILS: Self = _

pub const GETMOUSETRAILS: Self = _

pub const SETSCREENSAVERRUNNING: Self = _

pub const SCREENSAVERRUNNING: Self = _

pub const GETFILTERKEYS: Self = _

pub const SETFILTERKEYS: Self = _

pub const GETTOGGLEKEYS: Self = _

pub const SETTOGGLEKEYS: Self = _

pub const GETMOUSEKEYS: Self = _

pub const SETMOUSEKEYS: Self = _

pub const GETSHOWSOUNDS: Self = _

pub const SETSHOWSOUNDS: Self = _

pub const GETSTICKYKEYS: Self = _

pub const SETSTICKYKEYS: Self = _

pub const GETACCESSTIMEOUT: Self = _

pub const SETACCESSTIMEOUT: Self = _

pub const GETSERIALKEYS: Self = _

pub const SETSERIALKEYS: Self = _

pub const GETSOUNDSENTRY: Self = _

pub const SETSOUNDSENTRY: Self = _

pub const GETSNAPTODEFBUTTON: Self = _

pub const SETSNAPTODEFBUTTON: Self = _

pub const GETMOUSEHOVERWIDTH: Self = _

pub const SETMOUSEHOVERWIDTH: Self = _

pub const GETMOUSEHOVERHEIGHT: Self = _

pub const SETMOUSEHOVERHEIGHT: Self = _

pub const GETMOUSEHOVERTIME: Self = _

pub const SETMOUSEHOVERTIME: Self = _

pub const GETWHEELSCROLLLINES: Self = _

pub const SETWHEELSCROLLLINES: Self = _

pub const GETMENUSHOWDELAY: Self = _

pub const SETMENUSHOWDELAY: Self = _

pub const GETWHEELSCROLLCHARS: Self = _

pub const SETWHEELSCROLLCHARS: Self = _

pub const GETSHOWIMEUI: Self = _

pub const SETSHOWIMEUI: Self = _

pub const GETMOUSESPEED: Self = _

pub const SETMOUSESPEED: Self = _

pub const GETSCREENSAVERRUNNING: Self = _

pub const GETDESKWALLPAPER: Self = _

pub const GETAUDIODESCRIPTION: Self = _

pub const SETAUDIODESCRIPTION: Self = _

pub const GETSCREENSAVESECURE: Self = _

pub const SETSCREENSAVESECURE: Self = _

pub const GETHUNGAPPTIMEOUT: Self = _

pub const SETHUNGAPPTIMEOUT: Self = _

pub const GETWAITTOKILLTIMEOUT: Self = _

pub const SETWAITTOKILLTIMEOUT: Self = _

pub const GETWAITTOKILLSERVICETIMEOUT: Self = _

pub const SETWAITTOKILLSERVICETIMEOUT: Self = _

pub const GETMOUSEDOCKTHRESHOLD: Self = _

pub const SETMOUSEDOCKTHRESHOLD: Self = _

pub const GETPENDOCKTHRESHOLD: Self = _

pub const SETPENDOCKTHRESHOLD: Self = _

pub const GETWINARRANGING: Self = _

pub const SETWINARRANGING: Self = _

pub const GETMOUSEDRAGOUTTHRESHOLD: Self = _

pub const SETMOUSEDRAGOUTTHRESHOLD: Self = _

pub const GETPENDRAGOUTTHRESHOLD: Self = _

pub const SETPENDRAGOUTTHRESHOLD: Self = _

pub const GETMOUSESIDEMOVETHRESHOLD: Self = _

pub const SETMOUSESIDEMOVETHRESHOLD: Self = _

pub const GETPENSIDEMOVETHRESHOLD: Self = _

pub const SETPENSIDEMOVETHRESHOLD: Self = _

pub const GETDRAGFROMMAXIMIZE: Self = _

pub const SETDRAGFROMMAXIMIZE: Self = _

pub const GETSNAPSIZING: Self = _

pub const SETSNAPSIZING: Self = _

pub const GETDOCKMOVING: Self = _

pub const SETDOCKMOVING: Self = _