Struct winsafe::co::SUBLANG

pub struct SUBLANG(/* private fields */);

Available on crate feature kernel only.

Sublanguage identifier (u16).

Implementations

impl SUBLANG

pub const unsafe fn from_raw(v: u16) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u16

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl SUBLANG

pub const NEUTRAL: Self = _

pub const DEFAULT: Self = _

pub const SYS_DEFAULT: Self = _

pub const CUSTOM_DEFAULT: Self = _

pub const CUSTOM_UNSPECIFIED: Self = _

pub const UI_CUSTOM_DEFAULT: Self = _

pub const AFRIKAANS_SOUTH_AFRICA: Self = _

pub const ALBANIAN_ALBANIA: Self = _

pub const ALSATIAN_FRANCE: Self = _

pub const AMHARIC_ETHIOPIA: Self = _

pub const ARABIC_SAUDI_ARABIA: Self = _

pub const ARABIC_IRAQ: Self = _

pub const ARABIC_EGYPT: Self = _

pub const ARABIC_LIBYA: Self = _

pub const ARABIC_ALGERIA: Self = _

pub const ARABIC_MOROCCO: Self = _

pub const ARABIC_TUNISIA: Self = _

pub const ARABIC_OMAN: Self = _

pub const ARABIC_YEMEN: Self = _

pub const ARABIC_SYRIA: Self = _

pub const ARABIC_JORDAN: Self = _

pub const ARABIC_LEBANON: Self = _

pub const ARABIC_KUWAIT: Self = _

pub const ARABIC_UAE: Self = _

pub const ARABIC_BAHRAIN: Self = _

pub const ARABIC_QATAR: Self = _

pub const ARMENIAN_ARMENIA: Self = _

pub const ASSAMESE_INDIA: Self = _

pub const AZERI_LATIN: Self = _

pub const AZERI_CYRILLIC: Self = _

pub const AZERBAIJANI_AZERBAIJAN_LATIN: Self = _

pub const AZERBAIJANI_AZERBAIJAN_CYRILLIC: Self = _

pub const BANGLA_INDIA: Self = _

pub const BANGLA_BANGLADESH: Self = _

pub const BASHKIR_RUSSIA: Self = _

pub const BASQUE_BASQUE: Self = _

pub const BELARUSIAN_BELARUS: Self = _

pub const BENGALI_INDIA: Self = _

pub const BENGALI_BANGLADESH: Self = _

pub const BOSNIAN_BOSNIA_HERZEGOVINA_LATIN: Self = _

pub const BOSNIAN_BOSNIA_HERZEGOVINA_CYRILLIC: Self = _

pub const BRETON_FRANCE: Self = _

pub const BULGARIAN_BULGARIA: Self = _

pub const CATALAN_CATALAN: Self = _

pub const CENTRAL_KURDISH_IRAQ: Self = _

pub const CHEROKEE_CHEROKEE: Self = _

pub const CHINESE_TRADITIONAL: Self = _

pub const CHINESE_SIMPLIFIED: Self = _

pub const CHINESE_HONGKONG: Self = _

pub const CHINESE_SINGAPORE: Self = _

pub const CHINESE_MACAU: Self = _

pub const CORSICAN_FRANCE: Self = _

pub const CZECH_CZECH_REPUBLIC: Self = _

pub const CROATIAN_CROATIA: Self = _

pub const CROATIAN_BOSNIA_HERZEGOVINA_LATIN: Self = _

pub const DANISH_DENMARK: Self = _

pub const DARI_AFGHANISTAN: Self = _

pub const DIVEHI_MALDIVES: Self = _

pub const DUTCH: Self = _

pub const DUTCH_BELGIAN: Self = _

pub const ENGLISH_US: Self = _

pub const ENGLISH_UK: Self = _

pub const ENGLISH_AUS: Self = _

pub const ENGLISH_CAN: Self = _

pub const ENGLISH_NZ: Self = _

pub const ENGLISH_EIRE: Self = _

pub const ENGLISH_SOUTH_AFRICA: Self = _

pub const ENGLISH_JAMAICA: Self = _

pub const ENGLISH_CARIBBEAN: Self = _

pub const ENGLISH_BELIZE: Self = _

pub const ENGLISH_TRINIDAD: Self = _

pub const ENGLISH_ZIMBABWE: Self = _

pub const ENGLISH_PHILIPPINES: Self = _

pub const ENGLISH_INDIA: Self = _

pub const ENGLISH_MALAYSIA: Self = _

pub const ENGLISH_SINGAPORE: Self = _

pub const ESTONIAN_ESTONIA: Self = _

pub const FAEROESE_FAROE_ISLANDS: Self = _

pub const FILIPINO_PHILIPPINES: Self = _

pub const FINNISH_FINLAND: Self = _

pub const FRENCH: Self = _

pub const FRENCH_BELGIAN: Self = _

pub const FRENCH_CANADIAN: Self = _

pub const FRENCH_SWISS: Self = _

pub const FRENCH_LUXEMBOURG: Self = _

pub const FRENCH_MONACO: Self = _

pub const FRISIAN_NETHERLANDS: Self = _

pub const FULAH_SENEGAL: Self = _

pub const GALICIAN_GALICIAN: Self = _

pub const GEORGIAN_GEORGIA: Self = _

pub const GERMAN: Self = _

pub const GERMAN_SWISS: Self = _

pub const GERMAN_AUSTRIAN: Self = _

pub const GERMAN_LUXEMBOURG: Self = _

pub const GERMAN_LIECHTENSTEIN: Self = _

pub const GREEK_GREECE: Self = _

pub const GREENLANDIC_GREENLAND: Self = _

pub const GUJARATI_INDIA: Self = _

pub const HAUSA_NIGERIA_LATIN: Self = _

pub const HAWAIIAN_US: Self = _

pub const HEBREW_ISRAEL: Self = _

pub const HINDI_INDIA: Self = _

pub const HUNGARIAN_HUNGARY: Self = _

pub const ICELANDIC_ICELAND: Self = _

pub const IGBO_NIGERIA: Self = _

pub const INDONESIAN_INDONESIA: Self = _

pub const INUKTITUT_CANADA: Self = _

pub const INUKTITUT_CANADA_LATIN: Self = _

pub const IRISH_IRELAND: Self = _

pub const ITALIAN: Self = _

pub const ITALIAN_SWISS: Self = _

pub const JAPANESE_JAPAN: Self = _

pub const KANNADA_INDIA: Self = _

pub const KASHMIRI_SASIA: Self = _

pub const KASHMIRI_INDIA: Self = _

pub const KAZAK_KAZAKHSTAN: Self = _

pub const KHMER_CAMBODIA: Self = _

pub const KICHE_GUATEMALA: Self = _

pub const KINYARWANDA_RWANDA: Self = _

pub const KONKANI_INDIA: Self = _

pub const KOREAN: Self = _

pub const KYRGYZ_KYRGYZSTAN: Self = _

pub const LAO_LAO: Self = _

pub const LATVIAN_LATVIA: Self = _

pub const LITHUANIAN: Self = _

pub const LOWER_SORBIAN_GERMANY: Self = _

pub const LUXEMBOURGISH_LUXEMBOURG: Self = _

pub const MACEDONIAN_MACEDONIA: Self = _

pub const MALAY_MALAYSIA: Self = _

pub const MALAY_BRUNEI_DARUSSALAM: Self = _

pub const MALAYALAM_INDIA: Self = _

pub const MALTESE_MALTA: Self = _

pub const MAORI_NEW_ZEALAND: Self = _

pub const MAPUDUNGUN_CHILE: Self = _

pub const MARATHI_INDIA: Self = _

pub const MOHAWK_MOHAWK: Self = _

pub const MONGOLIAN_CYRILLIC_MONGOLIA: Self = _

pub const MONGOLIAN_PRC: Self = _

pub const NEPALI_INDIA: Self = _

pub const NEPALI_NEPAL: Self = _

pub const NORWEGIAN_BOKMAL: Self = _

pub const NORWEGIAN_NYNORSK: Self = _

pub const OCCITAN_FRANCE: Self = _

pub const ODIA_INDIA: Self = _

pub const ORIYA_INDIA: Self = _

pub const PASHTO_AFGHANISTAN: Self = _

pub const PERSIAN_IRAN: Self = _

pub const POLISH_POLAND: Self = _

pub const PORTUGUESE: Self = _

pub const PORTUGUESE_BRAZILIAN: Self = _

pub const PULAR_SENEGAL: Self = _

pub const PUNJABI_INDIA: Self = _

pub const PUNJABI_PAKISTAN: Self = _

pub const QUECHUA_BOLIVIA: Self = _

pub const QUECHUA_ECUADOR: Self = _

pub const QUECHUA_PERU: Self = _

pub const ROMANIAN_ROMANIA: Self = _

pub const ROMANSH_SWITZERLAND: Self = _

pub const RUSSIAN_RUSSIA: Self = _

pub const SAKHA_RUSSIA: Self = _

pub const SAMI_NORTHERN_NORWAY: Self = _

pub const SAMI_NORTHERN_SWEDEN: Self = _

pub const SAMI_NORTHERN_FINLAND: Self = _

pub const SAMI_LULE_NORWAY: Self = _

pub const SAMI_LULE_SWEDEN: Self = _

pub const SAMI_SOUTHERN_NORWAY: Self = _

pub const SAMI_SOUTHERN_SWEDEN: Self = _

pub const SAMI_SKOLT_FINLAND: Self = _

pub const SAMI_INARI_FINLAND: Self = _

pub const SANSKRIT_INDIA: Self = _

pub const SCOTTISH_GAELIC: Self = _

pub const SERBIAN_BOSNIA_HERZEGOVINA_LATIN: Self = _

pub const SERBIAN_BOSNIA_HERZEGOVINA_CYRILLIC: Self = _

pub const SERBIAN_MONTENEGRO_LATIN: Self = _

pub const SERBIAN_MONTENEGRO_CYRILLIC: Self = _

pub const SERBIAN_SERBIA_LATIN: Self = _

pub const SERBIAN_SERBIA_CYRILLIC: Self = _

pub const SERBIAN_CROATIA: Self = _

pub const SERBIAN_LATIN: Self = _

pub const SERBIAN_CYRILLIC: Self = _

pub const SINDHI_INDIA: Self = _

pub const SINDHI_PAKISTAN: Self = _

pub const SINDHI_AFGHANISTAN: Self = _

pub const SINHALESE_SRI_LANKA: Self = _

pub const SOTHO_NORTHERN_SOUTH_AFRICA: Self = _

pub const SLOVAK_SLOVAKIA: Self = _

pub const SLOVENIAN_SLOVENIA: Self = _

pub const SPANISH: Self = _

pub const SPANISH_MEXICAN: Self = _

pub const SPANISH_MODERN: Self = _

pub const SPANISH_GUATEMALA: Self = _

pub const SPANISH_COSTA_RICA: Self = _

pub const SPANISH_PANAMA: Self = _

pub const SPANISH_DOMINICAN_REPUBLIC: Self = _

pub const SPANISH_VENEZUELA: Self = _

pub const SPANISH_COLOMBIA: Self = _

pub const SPANISH_PERU: Self = _

pub const SPANISH_ARGENTINA: Self = _

pub const SPANISH_ECUADOR: Self = _

pub const SPANISH_CHILE: Self = _

pub const SPANISH_URUGUAY: Self = _

pub const SPANISH_PARAGUAY: Self = _

pub const SPANISH_BOLIVIA: Self = _

pub const SPANISH_EL_SALVADOR: Self = _

pub const SPANISH_HONDURAS: Self = _

pub const SPANISH_NICARAGUA: Self = _

pub const SPANISH_PUERTO_RICO: Self = _

pub const SPANISH_US: Self = _

pub const SWAHILI_KENYA: Self = _

pub const SWEDISH: Self = _

pub const SWEDISH_FINLAND: Self = _

pub const SYRIAC_SYRIA: Self = _

pub const TAJIK_TAJIKISTAN: Self = _

pub const TAMAZIGHT_ALGERIA_LATIN: Self = _

pub const TAMAZIGHT_MOROCCO_TIFINAGH: Self = _

pub const TAMIL_INDIA: Self = _

pub const TAMIL_SRI_LANKA: Self = _

pub const TATAR_RUSSIA: Self = _

pub const TELUGU_INDIA: Self = _

pub const THAI_THAILAND: Self = _

pub const TIBETAN_PRC: Self = _

pub const TIGRIGNA_ERITREA: Self = _

pub const TIGRINYA_ERITREA: Self = _

pub const TIGRINYA_ETHIOPIA: Self = _

pub const TSWANA_BOTSWANA: Self = _

pub const TSWANA_SOUTH_AFRICA: Self = _

pub const TURKISH_TURKEY: Self = _

pub const TURKMEN_TURKMENISTAN: Self = _

pub const UIGHUR_PRC: Self = _

pub const UKRAINIAN_UKRAINE: Self = _

pub const UPPER_SORBIAN_GERMANY: Self = _

pub const URDU_PAKISTAN: Self = _

pub const URDU_INDIA: Self = _

pub const UZBEK_LATIN: Self = _

pub const UZBEK_CYRILLIC: Self = _

pub const VALENCIAN_VALENCIA: Self = _

pub const VIETNAMESE_VIETNAM: Self = _

pub const WELSH_UNITED_KINGDOM: Self = _

pub const WOLOF_SENEGAL: Self = _

pub const XHOSA_SOUTH_AFRICA: Self = _

pub const YAKUT_RUSSIA: Self = _

pub const YI_PRC: Self = _

pub const YORUBA_NIGERIA: Self = _

pub const ZULU_SOUTH_AFRICA: Self = _