Struct winsafe::co::TMT

pub struct TMT(/* private fields */);

Available on crate features kernel and uxtheme only.

Theme property identifiers (i32).

Implementations

impl TMT

pub const unsafe fn from_raw(v: i32) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> i32

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl TMT

pub const DIBDATA: Self = _

pub const GLYPHDIBDATA: Self = _

pub const ENUM: Self = _

pub const STRING: Self = _

pub const INT: Self = _

pub const BOOL: Self = _

pub const COLOR: Self = _

pub const MARGINS: Self = _

pub const FILENAME: Self = _

pub const SIZE: Self = _

pub const POSITION: Self = _

pub const RECT: Self = _

pub const FONT: Self = _

pub const INTLIST: Self = _

pub const HBITMAP: Self = _

pub const DISKSTREAM: Self = _

pub const STREAM: Self = _

pub const BITMAPREF: Self = _

pub const FLOAT: Self = _

pub const FLOATLIST: Self = _

pub const COLORSCHEMES: Self = _

pub const SIZES: Self = _

pub const CHARSET: Self = _

pub const NAME: Self = _

pub const DISPLAYNAME: Self = _

pub const TOOLTIP: Self = _

pub const COMPANY: Self = _

pub const AUTHOR: Self = _

pub const COPYRIGHT: Self = _

pub const URL: Self = _

pub const VERSION: Self = _

pub const DESCRIPTION: Self = _

pub const FIRST_RCSTRING_NAME: Self = _

pub const LAST_RCSTRING_NAME: Self = _

pub const CAPTIONFONT: Self = _

pub const SMALLCAPTIONFONT: Self = _

pub const MENUFONT: Self = _

pub const STATUSFONT: Self = _

pub const MSGBOXFONT: Self = _

pub const ICONTITLEFONT: Self = _

pub const HEADING1FONT: Self = _

pub const HEADING2FONT: Self = _

pub const BODYFONT: Self = _

pub const FIRSTFONT: Self = _

pub const LASTFONT: Self = _

pub const FLATMENUS: Self = _

pub const FIRSTBOOL: Self = _

pub const LASTBOOL: Self = _

pub const SIZINGBORDERWIDTH: Self = _

pub const SCROLLBARWIDTH: Self = _

pub const SCROLLBARHEIGHT: Self = _

pub const CAPTIONBARWIDTH: Self = _

pub const CAPTIONBARHEIGHT: Self = _

pub const SMCAPTIONBARWIDTH: Self = _

pub const SMCAPTIONBARHEIGHT: Self = _

pub const MENUBARWIDTH: Self = _

pub const MENUBARHEIGHT: Self = _

pub const PADDEDBORDERWIDTH: Self = _

pub const FIRSTSIZE: Self = _

pub const LASTSIZE: Self = _

pub const MINCOLORDEPTH: Self = _

pub const FIRSTINT: Self = _

pub const LASTINT: Self = _

pub const CSSNAME: Self = _

pub const XMLNAME: Self = _

pub const LASTUPDATED: Self = _

pub const ALIAS: Self = _

pub const FIRSTSTRING: Self = _

pub const LASTSTRING: Self = _

pub const SCROLLBAR: Self = _

pub const BACKGROUND: Self = _

pub const ACTIVECAPTION: Self = _

pub const INACTIVECAPTION: Self = _

pub const MENU: Self = _

pub const WINDOW: Self = _

pub const WINDOWFRAME: Self = _

pub const MENUTEXT: Self = _

pub const WINDOWTEXT: Self = _

pub const CAPTIONTEXT: Self = _

pub const ACTIVEBORDER: Self = _

pub const INACTIVEBORDER: Self = _

pub const APPWORKSPACE: Self = _

pub const HIGHLIGHT: Self = _

pub const HIGHLIGHTTEXT: Self = _

pub const BTNFACE: Self = _

pub const BTNSHADOW: Self = _

pub const GRAYTEXT: Self = _

pub const BTNTEXT: Self = _

pub const INACTIVECAPTIONTEXT: Self = _

pub const BTNHIGHLIGHT: Self = _

pub const DKSHADOW3D: Self = _

pub const LIGHT3D: Self = _

pub const INFOTEXT: Self = _

pub const INFOBK: Self = _

pub const BUTTONALTERNATEFACE: Self = _

pub const HOTTRACKING: Self = _

pub const GRADIENTACTIVECAPTION: Self = _

pub const GRADIENTINACTIVECAPTION: Self = _

pub const MENUHILIGHT: Self = _

pub const MENUBAR: Self = _

pub const FIRSTCOLOR: Self = _

pub const LASTCOLOR: Self = _

pub const FROMHUE1: Self = _

pub const FROMHUE2: Self = _

pub const FROMHUE3: Self = _

pub const FROMHUE4: Self = _

pub const FROMHUE5: Self = _

pub const TOHUE1: Self = _

pub const TOHUE2: Self = _

pub const TOHUE3: Self = _

pub const TOHUE4: Self = _

pub const TOHUE5: Self = _

pub const FROMCOLOR1: Self = _

pub const FROMCOLOR2: Self = _

pub const FROMCOLOR3: Self = _

pub const FROMCOLOR4: Self = _

pub const FROMCOLOR5: Self = _

pub const TOCOLOR1: Self = _

pub const TOCOLOR2: Self = _

pub const TOCOLOR3: Self = _

pub const TOCOLOR4: Self = _

pub const TOCOLOR5: Self = _

pub const TRANSPARENT: Self = _

pub const AUTOSIZE: Self = _

pub const BORDERONLY: Self = _

pub const COMPOSITED: Self = _

pub const BGFILL: Self = _

pub const GLYPHTRANSPARENT: Self = _

pub const GLYPHONLY: Self = _

pub const ALWAYSSHOWSIZINGBAR: Self = _

pub const MIRRORIMAGE: Self = _

pub const UNIFORMSIZING: Self = _

pub const INTEGRALSIZING: Self = _

pub const SOURCEGROW: Self = _

pub const SOURCESHRINK: Self = _

pub const DRAWBORDERS: Self = _

pub const NOETCHEDEFFECT: Self = _

pub const TEXTAPPLYOVERLAY: Self = _

pub const TEXTGLOW: Self = _

pub const TEXTITALIC: Self = _

pub const COMPOSITEDOPAQUE: Self = _

pub const LOCALIZEDMIRRORIMAGE: Self = _

pub const IMAGECOUNT: Self = _

pub const ALPHALEVEL: Self = _

pub const BORDERSIZE: Self = _

pub const ROUNDCORNERWIDTH: Self = _

pub const ROUNDCORNERHEIGHT: Self = _

pub const GRADIENTRATIO1: Self = _

pub const GRADIENTRATIO2: Self = _

pub const GRADIENTRATIO3: Self = _

pub const GRADIENTRATIO4: Self = _

pub const GRADIENTRATIO5: Self = _

pub const PROGRESSCHUNKSIZE: Self = _

pub const PROGRESSSPACESIZE: Self = _

pub const SATURATION: Self = _

pub const TEXTBORDERSIZE: Self = _

pub const ALPHATHRESHOLD: Self = _

pub const WIDTH: Self = _

pub const HEIGHT: Self = _

pub const GLYPHINDEX: Self = _

pub const TRUESIZESTRETCHMARK: Self = _

pub const MINDPI1: Self = _

pub const MINDPI2: Self = _

pub const MINDPI3: Self = _

pub const MINDPI4: Self = _

pub const MINDPI5: Self = _

pub const TEXTGLOWSIZE: Self = _

pub const FRAMESPERSECOND: Self = _

pub const PIXELSPERFRAME: Self = _

pub const ANIMATIONDELAY: Self = _

pub const GLOWINTENSITY: Self = _

pub const OPACITY: Self = _

pub const COLORIZATIONCOLOR: Self = _

pub const COLORIZATIONOPACITY: Self = _

pub const MINDPI6: Self = _

pub const MINDPI7: Self = _

pub const GLYPHFONT: Self = _

pub const IMAGEFILE: Self = _

pub const IMAGEFILE1: Self = _

pub const IMAGEFILE2: Self = _

pub const IMAGEFILE3: Self = _

pub const IMAGEFILE4: Self = _

pub const IMAGEFILE5: Self = _

pub const GLYPHIMAGEFILE: Self = _

pub const IMAGEFILE6: Self = _

pub const IMAGEFILE7: Self = _

pub const TEXT: Self = _

pub const CLASSICVALUE: Self = _

pub const OFFSET: Self = _

pub const TEXTSHADOWOFFSET: Self = _

pub const MINSIZE: Self = _

pub const MINSIZE1: Self = _

pub const MINSIZE2: Self = _

pub const MINSIZE3: Self = _

pub const MINSIZE4: Self = _

pub const MINSIZE5: Self = _

pub const NORMALSIZE: Self = _

pub const MINSIZE6: Self = _

pub const MINSIZE7: Self = _

pub const SIZINGMARGINS: Self = _

pub const CONTENTMARGINS: Self = _

pub const CAPTIONMARGINS: Self = _

pub const BORDERCOLOR: Self = _

pub const FILLCOLOR: Self = _

pub const TEXTCOLOR: Self = _

pub const EDGELIGHTCOLOR: Self = _

pub const EDGEHIGHLIGHTCOLOR: Self = _

pub const EDGESHADOWCOLOR: Self = _

pub const EDGEDKSHADOWCOLOR: Self = _

pub const EDGEFILLCOLOR: Self = _

pub const TRANSPARENTCOLOR: Self = _

pub const GRADIENTCOLOR1: Self = _

pub const GRADIENTCOLOR2: Self = _

pub const GRADIENTCOLOR3: Self = _

pub const GRADIENTCOLOR4: Self = _

pub const GRADIENTCOLOR5: Self = _

pub const SHADOWCOLOR: Self = _

pub const GLOWCOLOR: Self = _

pub const TEXTBORDERCOLOR: Self = _

pub const TEXTSHADOWCOLOR: Self = _

pub const GLYPHTEXTCOLOR: Self = _

pub const GLYPHTRANSPARENTCOLOR: Self = _

pub const FILLCOLORHINT: Self = _

pub const BORDERCOLORHINT: Self = _

pub const ACCENTCOLORHINT: Self = _

pub const TEXTCOLORHINT: Self = _

pub const HEADING1TEXTCOLOR: Self = _

pub const HEADING2TEXTCOLOR: Self = _

pub const BODYTEXTCOLOR: Self = _

pub const BGTYPE: Self = _

pub const BORDERTYPE: Self = _

pub const FILLTYPE: Self = _

pub const SIZINGTYPE: Self = _

pub const HALIGN: Self = _

pub const CONTENTALIGNMENT: Self = _

pub const VALIGN: Self = _

pub const OFFSETTYPE: Self = _

pub const ICONEFFECT: Self = _

pub const TEXTSHADOWTYPE: Self = _

pub const IMAGELAYOUT: Self = _

pub const GLYPHTYPE: Self = _

pub const IMAGESELECTTYPE: Self = _

pub const GLYPHFONTSIZINGTYPE: Self = _

pub const TRUESIZESCALINGTYPE: Self = _

pub const USERPICTURE: Self = _

pub const DEFAULTPANESIZE: Self = _

pub const BLENDCOLOR: Self = _

pub const CUSTOMSPLITRECT: Self = _

pub const ANIMATIONBUTTONRECT: Self = _

pub const ANIMATIONDURATION: Self = _

pub const TRANSITIONDURATIONS: Self = _

pub const SCALEDBACKGROUND: Self = _

pub const ATLASIMAGE: Self = _

pub const ATLASINPUTIMAGE: Self = _

pub const ATLASRECT: Self = _