winsafe::co

Struct VK

pub struct VK(/* private fields */);

Available on crate feature user only.

Virtual key codes (u16).

Implementations

impl VK

pub const unsafe fn as_mut(&mut self) -> &mut u16

Returns a mutable reference to the underlying raw value.

Safety

Be sure the integer being set is meaningful for the actual type.

pub const unsafe fn from_raw(v: u16) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u16

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl VK

pub const NoValue: Self

None of the actual values (zero).

pub const LBUTTON: Self

Left mouse button.

pub const RBUTTON: Self

Right mouse button.

pub const CANCEL: Self

Control-break processing.

pub const MBUTTON: Self

Middle mouse button (three-button mouse).

pub const XBUTTON1: Self

X1 mouse button.

pub const XBUTTON2: Self

X2 mouse button.

pub const BACK: Self

BACKSPACE key.

pub const TAB: Self

TAB key.

pub const CLEAR: Self

CLEAR key.

pub const RETURN: Self

ENTER key.

pub const SHIFT: Self

SHIFT key.

pub const CONTROL: Self

CTRL key.

pub const MENU: Self

ALT key.

pub const PAUSE: Self

PAUSE key.

pub const CAPITAL: Self

CAPS LOCK key.

pub const KANA: Self

IME Kana mode.

pub const HANGUL: Self

IME Hangul mode.

pub const IME_ON: Self

IME On.

pub const JUNJA: Self

IME Junja mode.

pub const FINAL: Self

IME final mode.

pub const HANJA: Self

IME Hanja mode.

pub const KANJI: Self

IME Kanji mode.

pub const IME_OFF: Self

IME off.

pub const ESCAPE: Self

ESC key.

pub const CONVERT: Self

IME convert.

pub const NONCONVERT: Self

IME nonconvert.

pub const ACCEPT: Self

IME accept.

pub const MODECHANGE: Self

IME mode change request.

pub const SPACE: Self

SPACEBAR key.

pub const PRIOR: Self

PAGE UP key.

pub const NEXT: Self

PAGE DOWN key.

pub const END: Self

END key.

pub const HOME: Self

HOME key.

pub const LEFT: Self

LEFT ARROW key.

pub const UP: Self

UP ARROW key.

pub const RIGHT: Self

RIGHT ARROW key.

pub const DOWN: Self

DOWN ARROW key.

pub const SELECT: Self

SELECT key.

pub const PRINT: Self

PRINT key.

pub const EXECUTE: Self

EXECUTE key.

pub const SNAPSHOT: Self

PRINT SCREEN key.

pub const INSERT: Self

INS key.

pub const DELETE: Self

DEL key.

pub const HELP: Self

HELP key.

pub const CHAR_0: Self

Number 0 key.

pub const CHAR_1: Self

Number 1 key.

pub const CHAR_2: Self

Number 2 key.

pub const CHAR_3: Self

Number 3 key.

pub const CHAR_4: Self

Number 4 key.

pub const CHAR_5: Self

Number 5 key.

pub const CHAR_6: Self

Number 6 key.

pub const CHAR_7: Self

Number 7 key.

pub const CHAR_8: Self

Number 8 key.

pub const CHAR_9: Self

Number 9 key.

pub const CHAR_A: Self

Character A key.

pub const CHAR_B: Self

Character B key.

pub const CHAR_C: Self

Character C key.

pub const CHAR_D: Self

Character D key.

pub const CHAR_E: Self

Character E key.

pub const CHAR_F: Self

Character F key.

pub const CHAR_G: Self

Character G key.

pub const CHAR_H: Self

Character H key.

pub const CHAR_I: Self

Character I key.

pub const CHAR_J: Self

Character J key.

pub const CHAR_K: Self

Character K key.

pub const CHAR_L: Self

Character L key.

pub const CHAR_M: Self

Character M key.

pub const CHAR_N: Self

Character N key.

pub const CHAR_O: Self

Character O key.

pub const CHAR_P: Self

Character P key.

pub const CHAR_Q: Self

Character Q key.

pub const CHAR_R: Self

Character R key.

pub const CHAR_S: Self

Character S key.

pub const CHAR_T: Self

Character T key.

pub const CHAR_U: Self

Character U key.

pub const CHAR_V: Self

Character V key.

pub const CHAR_W: Self

Character W key.

pub const CHAR_X: Self

Character X key.

pub const CHAR_Y: Self

Character Y key.

pub const CHAR_Z: Self

Character Z key.

pub const LWIN: Self

Left Windows key (Natural keyboard).

pub const RWIN: Self

Right Windows key (Natural keyboard).

pub const APPS: Self

Applications key context menu (Natural keyboard).

pub const SLEEP: Self

Computer Sleep key.

pub const NUMPAD0: Self

Numeric keypad 0 key.

pub const NUMPAD1: Self

Numeric keypad 1 key.

pub const NUMPAD2: Self

Numeric keypad 2 key.

pub const NUMPAD3: Self

Numeric keypad 3 key.

pub const NUMPAD4: Self

Numeric keypad 4 key.

pub const NUMPAD5: Self

Numeric keypad 5 key.

pub const NUMPAD6: Self

Numeric keypad 6 key.

pub const NUMPAD7: Self

Numeric keypad 7 key.

pub const NUMPAD8: Self

Numeric keypad 8 key.

pub const NUMPAD9: Self

Numeric keypad 9 key.

pub const MULTIPLY: Self

Numeric keypad multiply key.

pub const ADD: Self

Numeric keypad add key.

pub const SEPARATOR: Self

Numeric keypad separator key.

pub const SUBTRACT: Self

Numeric keypad subtract key.

pub const DECIMAL: Self

Numeric keypad decimal key.

pub const DIVIDE: Self

Numeric keypad divide key.

pub const F1: Self

pub const F2: Self

pub const F3: Self

pub const F4: Self

pub const F5: Self

pub const F6: Self

pub const F7: Self

pub const F8: Self

pub const F9: Self

pub const F10: Self

pub const F11: Self

pub const F12: Self

pub const F13: Self

pub const F14: Self

pub const F15: Self

pub const F16: Self

pub const F17: Self

pub const F18: Self

pub const F19: Self

pub const F20: Self

pub const F21: Self

pub const F22: Self

pub const F23: Self

pub const F24: Self

pub const NUMLOCK: Self

NUM LOCK key.

pub const SCROLL: Self

SCROLL LOCK key.

pub const OEM_NEC_EQUAL: Self

pub const OEM_FJ_JISHO: Self

pub const OEM_FJ_MASSHOU: Self

pub const OEM_FJ_TOUROKU: Self

pub const OEM_FJ_LOYA: Self

pub const OEM_FJ_ROYA: Self

pub const LSHIFT: Self

Left SHIFT key.

pub const RSHIFT: Self

Right SHIFT key.

pub const LCONTROL: Self

Left CONTROL key.

pub const RCONTROL: Self

Right CONTROL key.

pub const LMENU: Self

Left MENU key.

pub const RMENU: Self

Right MENU key.

pub const BROWSER_BACK: Self

pub const BROWSER_FORWARD: Self

pub const BROWSER_REFRESH: Self

pub const BROWSER_STOP: Self

pub const BROWSER_SEARCH: Self

pub const BROWSER_FAVORITES: Self

pub const BROWSER_HOME: Self

pub const VOLUME_MUTE: Self

pub const VOLUME_DOWN: Self

pub const VOLUME_UP: Self

pub const MEDIA_NEXT_TRACK: Self

pub const MEDIA_PREV_TRACK: Self

pub const MEDIA_STOP: Self

pub const MEDIA_PLAY_PAUSE: Self

pub const LAUNCH_MAIL: Self

pub const LAUNCH_MEDIA_SELECT: Self

pub const LAUNCH_APP1: Self

pub const LAUNCH_APP2: Self

pub const OEM_1: Self

pub const OEM_PLUS: Self

pub const OEM_COMMA: Self

pub const OEM_MINUS: Self

pub const OEM_PERIOD: Self

pub const OEM_2: Self

pub const OEM_3: Self

pub const OEM_4: Self

pub const OEM_5: Self

pub const OEM_6: Self

pub const OEM_7: Self

pub const OEM_8: Self

pub const OEM_AX: Self

pub const OEM_102: Self

pub const ICO_HELP: Self

pub const ICO_00: Self

pub const PROCESSKEY: Self

pub const ICO_CLEAR: Self

pub const PACKET: Self

pub const OEM_RESET: Self

pub const OEM_JUMP: Self

pub const OEM_PA1: Self

pub const OEM_PA2: Self

pub const OEM_PA3: Self

pub const OEM_WSCTRL: Self

pub const OEM_CUSEL: Self

pub const OEM_ATTN: Self

pub const OEM_FINISH: Self

pub const OEM_COPY: Self

pub const OEM_AUTO: Self

pub const OEM_ENLW: Self

pub const OEM_BACKTAB: Self

pub const ATTN: Self

pub const CRSEL: Self

pub const EXSEL: Self

pub const EREOF: Self

pub const PLAY: Self

pub const ZOOM: Self

pub const NONAME: Self

pub const PA1: Self

pub const OEM_CLEAR: Self