winsafe::co

Struct VT

pub struct VT(/* private fields */);

Available on crate feature oleaut only.

VARENUM enumeration (u16).

This is a bitflag constant.

Implementations

impl VT

pub const unsafe fn as_mut(&mut self) -> &mut u16

Returns a mutable reference to the underlying raw value.

Safety

Be sure the integer being set is meaningful for the actual type.

pub const unsafe fn from_raw(v: u16) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u16

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl VT

pub const fn has(&self, other: Self) -> bool

Returns true if the other bitflag value is present.

Equivalent to (val & other) != 0.

impl VT

pub const EMPTY: Self

Nothing.

pub const NULL: Self

SQL style NULL.

pub const I2: Self

2 byte signed int.

pub const I4: Self

4 byte signed int.

pub const R4: Self

4 byte real.

pub const R8: Self

8 byte real.

pub const CY: Self

Currency.

pub const DATE: Self

Date.

pub const BSTR: Self

OLE Automation string.

pub const DISPATCH: Self

IDispatch pointer.

pub const ERROR: Self

SCODE.

pub const BOOL: Self

True = -1, False = 0.

pub const VARIANT: Self

VARIANT pointer.

pub const UNKNOWN: Self

IUnknown pointer.

pub const DECIMAL: Self

16 byte fixed point.

pub const I1: Self

Signed char.

pub const UI1: Self

Unsigned char.

pub const UI2: Self

Unsigned short.

pub const UI4: Self

ULONG.

pub const I8: Self

Signed 64-bit int.

pub const UI8: Self

Unsigned 64-bit int.

pub const INT: Self

Signed machine int.

pub const UINT: Self

Unsigned machine int.

pub const VOID: Self

C style void.

pub const HRESULT: Self

Standard return type.

pub const PTR: Self

Pointer type.

pub const SAFEARRAY: Self

Use VT::ARRAY in VARIANT.

pub const CARRAY: Self

C style array.

pub const USERDEFINED: Self

User defined type.

pub const LPSTR: Self

Null terminated string.

pub const LPWSTR: Self

Wide null terminated string.

pub const RECORD: Self

User defined type.

pub const INT_PTR: Self

Signed machine register size width.

pub const UINT_PTR: Self

Unsigned machine register size width.

pub const FILETIME: Self

FILETIME.

pub const BLOB: Self

Length of prefixed bytes.

pub const STREAM: Self

Name of the stream follows.

pub const STORAGE: Self

Name of the storage follows.

pub const STREAMED_OBJECT: Self

Stream contains an object.

pub const STORED_OBJECT: Self

Storage contains an object.

pub const BLOB_OBJECT: Self

Blob contains an object.

pub const CF: Self

Clipboard format.

pub const CLSID: Self

A class ID.

pub const VERSIONED_STREAM: Self

Stream with a GUID version.

pub const BSTR_BLOB: Self

Reserved for system use.

pub const VECTOR: Self

Simple counted array.

pub const ARRAY: Self

SAFEARRAY pointer.

pub const BYREF: Self

Void pointer for local use.

pub const RESERVED: Self

pub const ILLEGAL: Self

pub const ILLEGALMASKED: Self

pub const TYPEMASK: Self