winsafe::co

Struct WELL_KNOWN_SID_TYPE

pub struct WELL_KNOWN_SID_TYPE(/* private fields */);

Available on crate feature advapi only.

WELL_KNOWN_SID_TYPE enumeration (u32).

Implementations

impl WELL_KNOWN_SID_TYPE

pub const unsafe fn as_mut(&mut self) -> &mut u32

Returns a mutable reference to the underlying raw value.

Safety

Be sure the integer being set is meaningful for the actual type.

pub const unsafe fn from_raw(v: u32) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u32

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl WELL_KNOWN_SID_TYPE

pub const Null: Self

pub const World: Self

pub const Local: Self

pub const CreatorOwner: Self

pub const CreatorGroup: Self

pub const CreatorOwnerServer: Self

pub const CreatorGroupServer: Self

pub const NtAuthority: Self

pub const Dialup: Self

pub const Network: Self

pub const Batch: Self

pub const Interactive: Self

pub const Service: Self

pub const Anonymous: Self

pub const Proxy: Self

pub const EnterpriseControllers: Self

pub const SelfSid: Self

pub const AuthenticatedUser: Self

pub const RestrictedCode: Self

pub const TerminalServer: Self

pub const RemoteLogonId: Self

pub const LogonIds: Self

pub const LocalSystem: Self

pub const LocalService: Self

pub const NetworkService: Self

pub const BuiltinDomain: Self

pub const BuiltinAdministrators: Self

pub const BuiltinUsers: Self

pub const BuiltinGuests: Self

pub const BuiltinPowerUsers: Self

pub const BuiltinAccountOperators: Self

pub const BuiltinSystemOperators: Self

pub const BuiltinPrintOperators: Self

pub const BuiltinBackupOperators: Self

pub const BuiltinReplicator: Self

pub const BuiltinPreWindows2000CompatibleAccess: Self

pub const BuiltinRemoteDesktopUsers: Self

pub const BuiltinNetworkConfigurationOperators: Self

pub const AccountAdministrator: Self

pub const AccountGuest: Self

pub const AccountKrbtgt: Self

pub const AccountDomainAdmins: Self

pub const AccountDomainUsers: Self

pub const AccountDomainGuests: Self

pub const AccountComputers: Self

pub const AccountControllers: Self

pub const AccountCertAdmins: Self

pub const AccountSchemaAdmins: Self

pub const AccountEnterpriseAdmins: Self

pub const AccountPolicyAdmins: Self

pub const AccountRasAndIasServers: Self

pub const NTLMAuthentication: Self

pub const DigestAuthentication: Self

pub const SChannelAuthentication: Self

pub const ThisOrganization: Self

pub const OtherOrganization: Self

pub const BuiltinIncomingForestTrustBuilders: Self

pub const BuiltinPerfMonitoringUsers: Self

pub const BuiltinPerfLoggingUsers: Self

pub const BuiltinAuthorizationAccess: Self

pub const BuiltinTerminalServerLicenseServers: Self

pub const BuiltinDCOMUsers: Self

pub const BuiltinIUsers: Self

pub const IUser: Self

pub const BuiltinCryptoOperators: Self

pub const UntrustedLabel: Self

pub const LowLabel: Self

pub const MediumLabel: Self

pub const HighLabel: Self

pub const SystemLabel: Self

pub const WriteRestrictedCode: Self

pub const CreatorOwnerRights: Self

pub const CacheablePrincipalsGroup: Self

pub const NonCacheablePrincipalsGroup: Self

pub const EnterpriseReadonlyControllers: Self

pub const AccountReadonlyControllers: Self

pub const BuiltinEventLogReadersGroup: Self

pub const NewEnterpriseReadonlyControllers: Self

pub const BuiltinCertSvcDComAccessGroup: Self

pub const MediumPlusLabel: Self

pub const LocalLogon: Self

pub const ConsoleLogon: Self

pub const ThisOrganizationCertificate: Self

pub const ApplicationPackageAuthority: Self

pub const BuiltinAnyPackage: Self

pub const CapabilityInternetClient: Self

pub const CapabilityInternetClientServer: Self

pub const CapabilityPrivateNetworkClientServer: Self

pub const CapabilityPicturesLibrary: Self

pub const CapabilityVideosLibrary: Self

pub const CapabilityMusicLibrary: Self

pub const CapabilityDocumentsLibrary: Self

pub const CapabilitySharedUserCertificates: Self

pub const CapabilityEnterpriseAuthentication: Self

pub const CapabilityRemovableStorage: Self

pub const BuiltinRDSRemoteAccessServers: Self

pub const BuiltinRDSEndpointServers: Self

pub const BuiltinRDSManagementServers: Self

pub const UserModeDrivers: Self

pub const BuiltinHyperVAdmins: Self

pub const AccountCloneableControllers: Self

pub const BuiltinAccessControlAssistanceOperators: Self

pub const BuiltinRemoteManagementUsers: Self

pub const AuthenticationAuthorityAsserted: Self

pub const AuthenticationServiceAsserted: Self

pub const LocalAccount: Self

pub const LocalAccountAndAdministrator: Self

pub const AccountProtectedUsers: Self

pub const CapabilityAppointments: Self

pub const CapabilityContacts: Self

pub const AccountDefaultSystemManaged: Self

pub const BuiltinDefaultSystemManagedGroup: Self

pub const BuiltinStorageReplicaAdmins: Self

pub const AccountKeyAdmins: Self

pub const AccountEnterpriseKeyAdmins: Self

pub const AuthenticationKeyTrust: Self

pub const AuthenticationKeyPropertyMFA: Self

pub const AuthenticationKeyPropertyAttestation: Self

pub const AuthenticationFreshKeyAuth: Self

pub const BuiltinDeviceOwners: Self