Struct winsafe::co::WM

pub struct WM(/* private fields */);

Available on crate features kernel and user only.

Window message codes (u32).

Note: Control-specific messages have their own types, which are convertible to WM.

Implementations

impl WM

pub const unsafe fn from_raw(v: u32) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u32

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl WM

pub const NULL: Self = _

pub const CREATE: Self = _

pub const DESTROY: Self = _

pub const MOVE: Self = _

pub const SIZE: Self = _

pub const ACTIVATE: Self = _

pub const SETFOCUS: Self = _

pub const KILLFOCUS: Self = _

pub const ENABLE: Self = _

pub const SETREDRAW: Self = _

pub const SETTEXT: Self = _

pub const GETTEXT: Self = _

pub const GETTEXTLENGTH: Self = _

pub const PAINT: Self = _

pub const CLOSE: Self = _

pub const QUERYENDSESSION: Self = _

pub const QUERYOPEN: Self = _

pub const ENDSESSION: Self = _

pub const QUIT: Self = _

pub const ERASEBKGND: Self = _

pub const SYSCOLORCHANGE: Self = _

pub const SHOWWINDOW: Self = _

pub const WININICHANGE: Self = _

pub const DEVMODECHANGE: Self = _

pub const ACTIVATEAPP: Self = _

pub const FONTCHANGE: Self = _

pub const TIMECHANGE: Self = _

pub const CANCELMODE: Self = _

pub const SETCURSOR: Self = _

pub const MOUSEACTIVATE: Self = _

pub const CHILDACTIVATE: Self = _

pub const QUEUESYNC: Self = _

pub const GETMINMAXINFO: Self = _

pub const PAINTICON: Self = _

pub const ICONERASEBKGND: Self = _

pub const NEXTDLGCTL: Self = _

pub const SPOOLERSTATUS: Self = _

pub const DRAWITEM: Self = _

pub const MEASUREITEM: Self = _

pub const DELETEITEM: Self = _

pub const VKEYTOITEM: Self = _

pub const CHARTOITEM: Self = _

pub const SETFONT: Self = _

pub const GETFONT: Self = _

pub const SETHOTKEY: Self = _

pub const GETHOTKEY: Self = _

pub const QUERYDRAGICON: Self = _

pub const COMPAREITEM: Self = _

pub const GETOBJECT: Self = _

pub const COPYDATA: Self = _

pub const COMPACTING: Self = _

pub const COMMNOTIFY: Self = _

pub const WINDOWPOSCHANGING: Self = _

pub const WINDOWPOSCHANGED: Self = _

pub const POWER: Self = _

pub const NOTIFY: Self = _

pub const INPUTLANGCHANGEREQUEST: Self = _

pub const INPUTLANGCHANGE: Self = _

pub const TCARD: Self = _

pub const HELP: Self = _

pub const USERCHANGED: Self = _

pub const NOTIFYFORMAT: Self = _

pub const CONTEXTMENU: Self = _

pub const STYLECHANGING: Self = _

pub const STYLECHANGED: Self = _

pub const DISPLAYCHANGE: Self = _

pub const GETICON: Self = _

pub const SETICON: Self = _

pub const NCCREATE: Self = _

pub const NCDESTROY: Self = _

pub const NCCALCSIZE: Self = _

pub const NCHITTEST: Self = _

pub const NCPAINT: Self = _

pub const NCACTIVATE: Self = _

pub const GETDLGCODE: Self = _

pub const SYNCPAINT: Self = _

pub const NCMOUSEMOVE: Self = _

pub const NCLBUTTONDOWN: Self = _

pub const NCLBUTTONUP: Self = _

pub const NCLBUTTONDBLCLK: Self = _

pub const NCRBUTTONDOWN: Self = _

pub const NCRBUTTONUP: Self = _

pub const NCRBUTTONDBLCLK: Self = _

pub const NCMBUTTONDOWN: Self = _

pub const NCMBUTTONUP: Self = _

pub const NCMBUTTONDBLCLK: Self = _

pub const NCXBUTTONDOWN: Self = _

pub const NCXBUTTONUP: Self = _

pub const NCXBUTTONDBLCLK: Self = _

pub const INPUT_DEVICE_CHANGE: Self = _

pub const INPUT: Self = _

pub const KEYFIRST: Self = _

pub const KEYDOWN: Self = _

pub const KEYUP: Self = _

pub const CHAR: Self = _

pub const DEADCHAR: Self = _

pub const SYSKEYDOWN: Self = _

pub const SYSKEYUP: Self = _

pub const SYSCHAR: Self = _

pub const SYSDEADCHAR: Self = _

pub const UNICHAR: Self = _

pub const KEYLAST: Self = _

pub const IME_STARTCOMPOSITION: Self = _

pub const IME_ENDCOMPOSITION: Self = _

pub const IME_COMPOSITION: Self = _

pub const IME_KEYLAST: Self = _

pub const INITDIALOG: Self = _

pub const COMMAND: Self = _

pub const SYSCOMMAND: Self = _

pub const TIMER: Self = _

pub const HSCROLL: Self = _

pub const VSCROLL: Self = _

pub const INITMENU: Self = _

pub const INITMENUPOPUP: Self = _

pub const GESTURE: Self = _

pub const GESTURENOTIFY: Self = _

pub const MENUSELECT: Self = _

pub const MENUCHAR: Self = _

pub const ENTERIDLE: Self = _

pub const MENURBUTTONUP: Self = _

pub const MENUDRAG: Self = _

pub const MENUGETOBJECT: Self = _

pub const UNINITMENUPOPUP: Self = _

pub const MENUCOMMAND: Self = _

pub const CHANGEUISTATE: Self = _

pub const UPDATEUISTATE: Self = _

pub const QUERYUISTATE: Self = _

pub const CTLCOLORMSGBOX: Self = _

pub const CTLCOLOREDIT: Self = _

pub const CTLCOLORLISTBOX: Self = _

pub const CTLCOLORBTN: Self = _

pub const CTLCOLORDLG: Self = _

pub const CTLCOLORSCROLLBAR: Self = _

pub const CTLCOLORSTATIC: Self = _

pub const MN_GETHMENU: Self = _

Originally has no WM prefix.

pub const MOUSEFIRST: Self = _

pub const MOUSEMOVE: Self = _

pub const LBUTTONDOWN: Self = _

pub const LBUTTONUP: Self = _

pub const LBUTTONDBLCLK: Self = _

pub const RBUTTONDOWN: Self = _

pub const RBUTTONUP: Self = _

pub const RBUTTONDBLCLK: Self = _

pub const MBUTTONDOWN: Self = _

pub const MBUTTONUP: Self = _

pub const MBUTTONDBLCLK: Self = _

pub const MOUSEHWHEEL: Self = _

pub const XBUTTONDOWN: Self = _

pub const XBUTTONUP: Self = _

pub const XBUTTONDBLCLK: Self = _

pub const MOUSELAST: Self = _

pub const PARENTNOTIFY: Self = _

pub const ENTERMENULOOP: Self = _

pub const EXITMENULOOP: Self = _

pub const NEXTMENU: Self = _

pub const SIZING: Self = _

pub const CAPTURECHANGED: Self = _

pub const MOVING: Self = _

pub const POWERBROADCAST: Self = _

pub const DEVICECHANGE: Self = _

pub const MDICREATE: Self = _

pub const MDIDESTROY: Self = _

pub const MDIACTIVATE: Self = _

pub const MDIRESTORE: Self = _

pub const MDINEXT: Self = _

pub const MDIMAXIMIZE: Self = _

pub const MDITILE: Self = _

pub const MDICASCADE: Self = _

pub const MDIICONARRANGE: Self = _

pub const MDIGETACTIVE: Self = _

pub const MDISETMENU: Self = _

pub const ENTERSIZEMOVE: Self = _

pub const EXITSIZEMOVE: Self = _

pub const DROPFILES: Self = _

pub const MDIREFRESHMENU: Self = _

pub const POINTERDEVICECHANGE: Self = _

pub const POINTERDEVICEINRANGE: Self = _

pub const POINTERDEVICEOUTOFRANGE: Self = _

pub const TOUCH: Self = _

pub const NCPOINTERUPDATE: Self = _

pub const NCPOINTERDOWN: Self = _

pub const NCPOINTERUP: Self = _

pub const POINTERUPDATE: Self = _

pub const POINTERDOWN: Self = _

pub const POINTERUP: Self = _

pub const POINTERENTER: Self = _

pub const POINTERLEAVE: Self = _

pub const POINTERACTIVATE: Self = _

pub const POINTERCAPTURECHANGED: Self = _

pub const TOUCHHITTESTING: Self = _

pub const POINTERWHEEL: Self = _

pub const POINTERHWHEEL: Self = _

pub const DM_POINTERHITTEST: Self = _

Originally has no WM prefix.

pub const POINTERROUTEDTO: Self = _

pub const POINTERROUTEDAWAY: Self = _

pub const POINTERROUTEDRELEASED: Self = _

pub const IME_SETCONTEXT: Self = _

pub const IME_NOTIFY: Self = _

pub const IME_CONTROL: Self = _

pub const IME_COMPOSITIONFULL: Self = _

pub const IME_SELECT: Self = _

pub const IME_CHAR: Self = _

pub const IME_REQUEST: Self = _

pub const IME_KEYDOWN: Self = _

pub const IME_KEYUP: Self = _

pub const MOUSEHOVER: Self = _

pub const MOUSELEAVE: Self = _

pub const NCMOUSEHOVER: Self = _

pub const NCMOUSELEAVE: Self = _

pub const WTSSESSION_CHANGE: Self = _

pub const TABLET_FIRST: Self = _

pub const TABLET_LAST: Self = _

pub const DPICHANGED: Self = _

pub const DPICHANGED_BEFOREPARENT: Self = _

pub const DPICHANGED_AFTERPARENT: Self = _

pub const GETDPISCALEDSIZE: Self = _

pub const CUT: Self = _

pub const COPY: Self = _

pub const PASTE: Self = _

pub const CLEAR: Self = _

pub const UNDO: Self = _

pub const RENDERFORMAT: Self = _

pub const RENDERALLFORMATS: Self = _

pub const DESTROYCLIPBOARD: Self = _

pub const DRAWCLIPBOARD: Self = _

pub const PAINTCLIPBOARD: Self = _

pub const VSCROLLCLIPBOARD: Self = _

pub const SIZECLIPBOARD: Self = _

pub const ASKCBFORMATNAME: Self = _

pub const CHANGECBCHAIN: Self = _

pub const HSCROLLCLIPBOARD: Self = _

pub const QUERYNEWPALETTE: Self = _

pub const PALETTEISCHANGING: Self = _

pub const PALETTECHANGED: Self = _

pub const HOTKEY: Self = _

pub const PRINT: Self = _

pub const PRINTCLIENT: Self = _

pub const APPCOMMAND: Self = _

pub const THEMECHANGED: Self = _

pub const CLIPBOARDUPDATE: Self = _

pub const DWMCOMPOSITIONCHANGED: Self = _

pub const DWMNCRENDERINGCHANGED: Self = _

pub const DWMCOLORIZATIONCOLORCHANGED: Self = _

pub const DWMWINDOWMAXIMIZEDCHANGE: Self = _

pub const DWMSENDICONICTHUMBNAIL: Self = _

pub const DWMSENDICONICLIVEPREVIEWBITMAP: Self = _

pub const GETTITLEBARINFOEX: Self = _

pub const HANDHELDFIRST: Self = _

pub const HANDHELDLAST: Self = _

pub const AFXFIRST: Self = _

pub const AFXLAST: Self = _

pub const PENWINFIRST: Self = _

pub const PENWINLAST: Self = _

pub const APP: Self = _

pub const USER: Self = _