winsafe

Enum DwmAttr

pub enum DwmAttr {

    NcRenderingEnabled(bool),
    NcRenderingPolicy(DWMNCRENDERINGPOLICY),
    TransitionsForceDisabled(bool),
    AllowNcPaint(bool),
    CaptionButtonBounds(RECT),
    NonClientRtlLayout(bool),
    ForceIconicRepresentation(bool),
    Flip3dPolicy(DWMFLIP3DWINDOWPOLICY),
    ExtendedFrameBounds(RECT),
    HasIconicBitmap(bool),
    DisallowPeek(bool),
    ExcludedFromPeek(bool),
    Cloak(bool),
    Cloaked(DWM_CLOAKED),
    FreezeRepresentation(bool),
    PassiveUpdateMode(bool),
    UseHostBackdropBrush(bool),
    UseImmersiveDarkMode(bool),
    WindowCornerPreference(DWMWCP),
    BorderColor(COLORREF),
    CaptionColor(COLORREF),
    TextColor(COLORREF),
    VisibleFrameBorderThickness(u32),
    SystemBackdropType(DWMSBT),

}

Available on crate feature dwm only.

Variant parameter for:

HWND::DwmGetWindowAttribute

HWND::DwmSetWindowAttribute

Variants

NcRenderingEnabled(bool)

NcRenderingPolicy(DWMNCRENDERINGPOLICY)

TransitionsForceDisabled(bool)

AllowNcPaint(bool)

CaptionButtonBounds(RECT)

NonClientRtlLayout(bool)

ForceIconicRepresentation(bool)

Flip3dPolicy(DWMFLIP3DWINDOWPOLICY)

ExtendedFrameBounds(RECT)

HasIconicBitmap(bool)

DisallowPeek(bool)

ExcludedFromPeek(bool)

Cloak(bool)

Cloaked(DWM_CLOAKED)

FreezeRepresentation(bool)

PassiveUpdateMode(bool)

UseHostBackdropBrush(bool)

Since Windows 11 Build 22000.

UseImmersiveDarkMode(bool)

Since Windows 11 Build 22000.

WindowCornerPreference(DWMWCP)

Since Windows 11 Build 22000.

BorderColor(COLORREF)

Since Windows 11 Build 22000.

CaptionColor(COLORREF)

Since Windows 11 Build 22000.

TextColor(COLORREF)

Since Windows 11 Build 22000.

VisibleFrameBorderThickness(u32)

Since Windows 11 Build 22000.

SystemBackdropType(DWMSBT)

Since Windows 11 Build 22621.

Implementations

impl DwmAttr

pub const fn flag(&self) -> DWMWA

Returns the correspondent co::DWMWA flag.

pub const fn unwrap_bool(&self) -> bool

If the enum contains a bool, returns it; otherwise panics.

This is a syntactic sugar method to be used when you are sure of the content, its general use is discouraged.

Panics

Panics if the value is not bool.

pub const fn unwrap_rect(&self) -> RECT

If the enum contains a RECT, returns it; otherwise panics.

This is a syntactic sugar method to be used when you are sure of the content, its general use is discouraged.

Panics

Panics if the value is not RECT.

pub const fn unwrap_colorref(&self) -> COLORREF

If the enum contains a COLORREF, returns it; otherwise panics.

This is a syntactic sugar method to be used when you are sure of the content, its general use is discouraged.

Panics

Panics if the value is not COLORREF.

pub const fn unwrap_u32(&self) -> u32

If the enum contains an u32, returns it; otherwise panics.

This is a syntactic sugar method to be used when you are sure of the content, its general use is discouraged.

Panics

Panics if the value is not u32.

pub const fn unwrap_ncrenderingpolicy(&self) -> DWMNCRENDERINGPOLICY

If the enum contains a co::DWMNCRENDERINGPOLICY (DwmAttr::NcRenderingPolicy), returns it; otherwise panics.

This is a syntactic sugar method to be used when you are sure of the content, its general use is discouraged.

Panics

Panics if the value is not co::DWMNCRENDERINGPOLICY.

pub const fn unwrap_flip3dpolicy(&self) -> DWMFLIP3DWINDOWPOLICY

If the enum contains a co::DWMFLIP3DWINDOWPOLICY (DwmAttr::Flip3dPolicy), returns it; otherwise panics.

This is a syntactic sugar method to be used when you are sure of the content, its general use is discouraged.

Panics

Panics if the value is not co::DWMFLIP3DWINDOWPOLICY.

pub const fn unwrap_cloaked(&self) -> DWM_CLOAKED

If the enum contains a co::DWM_CLOAKED (DwmAttr::Cloaked), returns it; otherwise panics.

This is a syntactic sugar method to be used when you are sure of the content, its general use is discouraged.

Panics

Panics if the value is not co::DWM_CLOAKED.

pub const fn unwrap_wcp(&self) -> DWMWCP

If the enum contains a co::DWMWCP (DwmAttr::WindowCornerPreference), returns it; otherwise panics.

This is a syntactic sugar method to be used when you are sure of the content, its general use is discouraged.

Panics

Panics if the value is not co::DWMWCP.

pub const fn unwrap_sbt(&self) -> DWMSBT

If the enum contains a co::DWMSBT (DwmAttr::SystemBackdropType), returns it; otherwise panics.

This is a syntactic sugar method to be used when you are sure of the content, its general use is discouraged.

Panics

Panics if the value is not co::DWMSBT.