winsafe

Enum HttpInfo

pub enum HttpInfo {
    Number(u32),
    Number64(u64),
    Str(String),
    Time(SYSTEMTIME),
}

Available on crate feature wininet only.

Variant parameter for:

HINTERNETREQUEST::HttpQueryInfo

Variants

Number(u32)

If co::HTTP_QUERY_FLAG::NUMBER is present.

Number64(u64)

If co::HTTP_QUERY_FLAG::NUMBER64 is present.

Str(String)

Default value format.

Time(SYSTEMTIME)

If co::HTTP_QUERY_FLAG::SYSTEMTIME is present.

Implementations

impl HttpInfo

pub const fn unwrap_number(&self) -> u32

If the value is HttpInfo::Number, returns it; otherwise panics.

This is a syntactic sugar method to be used when you are sure of the HttpInfo content, its general use is discouraged.

Panics

Panics if the value is not HttpInfo::Number.

pub const fn unwrap_number64(&self) -> u64

If the value is HttpInfo::Number64, returns it; otherwise panics.

This is a syntactic sugar method to be used when you are sure of the HttpInfo content, its general use is discouraged.

Panics

Panics if the value is not HttpInfo::Number64.

pub fn unwrap_str(&self) -> String

If the value is HttpInfo::Str, returns a clone it; otherwise panics.

This is a syntactic sugar method to be used when you are sure of the HttpInfo content, its general use is discouraged.

Panics

Panics if the value is not HttpInfo::Str.

pub const fn unwrap_time(&self) -> SYSTEMTIME

If the value is HttpInfo::Time, returns a copy of it; otherwise panics.

This is a syntactic sugar method to be used when you are sure of the HttpInfo content, its general use is discouraged.

Panics

Panics if the value is not HttpInfo::Time.