winsafe

Enum Snd

pub enum Snd<'a> {
    AliasAsync {
        alias: SND_ALIAS,
        stop: bool,
        sentry: bool,
        loops: bool,
    },
    AliasSync {
        alias: SND_ALIAS,
        stop: bool,
        sentry: bool,
    },
    FileAsync {
        path: &'a str,
        default: bool,
        stop: bool,
        sentry: bool,
        loops: bool,
    },
    FileSync {
        path: &'a str,
        default: bool,
        stop: bool,
        sentry: bool,
    },
    MemSync {
        src: &'a [u8],
        default: bool,
        stop: bool,
        sentry: bool,
    },
    ResAsync {
        id: IdStr,
        hinst: &'a HINSTANCE,
        default: bool,
        stop: bool,
        sentry: bool,
        loops: bool,
    },
    ResSync {
        id: IdStr,
        hinst: &'a HINSTANCE,
        default: bool,
        stop: bool,
        sentry: bool,
    },
    Stop,
}

Available on crate feature winmm only.

Variable parameter for:

PlaySound

Variants

AliasAsync

Plays the system sound of the given co::SND_ALIAS asynchronously.

Fields

alias: SND_ALIAS

System sound alias.

stop: bool

If true, stops any sound currently playing asynchronously.

If set to false and a sound is currently playing, PlaySound will return an error.

sentry: bool

Calls SoundSentry to give a visual cue as an accessibility feature.

loops: bool

Plays the sound repeatedly until you call PlaySound with Snd::Stop.

AliasSync

Plays the system sound of the given co::SND_ALIAS asynchronously.

Fields

alias: SND_ALIAS

System sound alias.

stop: bool

If true, stops any sound currently playing asynchronously.

If set to false and a sound is currently playing, PlaySound will return an error.

sentry: bool

Calls SoundSentry to give a visual cue as an accessibility feature.

FileAsync

Loads a file to be played asynchronously.

Fields

path: &'a str

File path to be loaded.

default: bool

If true, a default sound will be played if path is not found.

stop: bool

If true, stops any sound currently playing asynchronously.

If set to false and a sound is currently playing, PlaySound will return an error.

sentry: bool

Calls SoundSentry to give a visual cue as an accessibility feature.

loops: bool

Plays the sound repeatedly until you call PlaySound with Snd::Stop.

FileSync

Loads a file to be played synchronously.

Fields

path: &'a str

File path to be loaded.

default: bool

If true, a default sound will be played if path is not found.

stop: bool

If true, stops any sound currently playing asynchronously.

If set to false and a sound is currently playing, PlaySound will return an error.

sentry: bool

Calls SoundSentry to give a visual cue as an accessibility feature.

MemSync

Reads the sound from memory and plays it synchronously.

Since PlaySound doesn’t provide a mechanism to notify its exit, there is no safe way to play from memory asynchronously.

Fields

src: &'a [u8]

Source to be read from.

default: bool

If true, a default sound will be played if src loading fails.

stop: bool

If true, stops any sound currently playing.

If set to false and a sound is currently playing, PlaySound will return an error.

sentry: bool

Calls SoundSentry to give a visual cue as an accessibility feature.

ResAsync

Loads a sound from the resource and plays it asynchronously.

Fields

id: IdStr

Resource identifier.

hinst: &'a HINSTANCE

Instance handle.

default: bool

If true, a default sound will be played if src loading fails.

stop: bool

If true, stops any sound currently playing.

If set to false and a sound is currently playing, PlaySound will return an error.

sentry: bool

Calls SoundSentry to give a visual cue as an accessibility feature.

loops: bool

Plays the sound repeatedly until you call PlaySound with Snd::Stop.

ResSync

Loads a sound from the resource and plays it synchronously.

Fields

id: IdStr

Resource identifier.

hinst: &'a HINSTANCE

Instance handle.

default: bool

If true, a default sound will be played if src loading fails.

stop: bool

If true, stops any sound currently playing.

If set to false and a sound is currently playing, PlaySound will return an error.

sentry: bool

Calls SoundSentry to give a visual cue as an accessibility feature.

Stop

Stops any sound currently playing asynchronously.