Function winsafe::MFStartup

pub fn MFStartup(flags: MFSTARTUP) -> HrResult<MFShutdownGuard>

Available on crate feature mf only.

MFStartup function.

In the original C implementation, you must call MFShutdown as a cleanup operation.

Here, the cleanup is performed automatically, because MFStartup returns a MFShutdownGuard, which automatically calls MFShutdown when the guard goes out of scope. You must, however, keep the guard alive, otherwise the cleanup will be performed right away.

Examples

use winsafe::{self as w, prelude::*, co};

let _mf_guard = w::MFStartup( // keep guard alive
    co::MFSTARTUP::NOSOCKET,
)?;

// program runs...

// MFShutdown() automatically called