winsafe::guard

Struct ClosePrinterGuard

pub struct ClosePrinterGuard { /* private fields */ }

Available on crate feature winspool only.

RAII implementation for HPRINTER which automatically calls ClosePrinter when the object goes out of scope.

Implementations

impl ClosePrinterGuard

pub const unsafe fn new(handle: HPRINTER) -> Self

Constructs the guard by taking ownership of the handle.

This method can be used as an escape hatch to interoperate with other libraries.

Safety

Be sure the handle must be freed with the specified function at the end of scope.

pub fn leak(&mut self) -> HPRINTER

Ejects the underlying handle, leaving a Handle::INVALID in its place.

Since the internal handle will be invalidated, the destructor will not run. Itโs your responsability to run it, otherwise youโll cause a resource leak.