winsafe::guard

Struct CoTaskMemFreeGuard

pub struct CoTaskMemFreeGuard { /* private fields */ }

Available on crate feature ole only.

RAII implementation which automatically calls CoTaskMemFree when the object goes out of scope.

Implementations

impl CoTaskMemFreeGuard

pub const unsafe fn new(pmem: *mut c_void, sz: usize) -> Self

Constructs the guard.

Safety

Be sure the pointer must be freed with CoTaskMemFree at the end of the scope, and the size is correct.

pub fn leak(&mut self) -> (*mut c_void, usize)

Ejects the underlying memory pointer and size, leaving null and zero in their places.

Since the internal memory pointer will be invalidated, the destructor will not run. It’s your responsibility to run it, otherwise you’ll cause a memory leak.

pub const fn as_ptr(&self) -> *const c_void

Returns a pointer to the allocated memory block, or null if not allocated.

pub const fn as_mut_ptr(&mut self) -> *mut c_void

Returns a mutable pointer to the allocated memory block, or null if not allocated.

pub const fn as_slice(&self) -> &[u8]

Returns a slice over the allocated memory block.

pub const fn as_mut_slice(&mut self) -> &mut [u8]

Returns a mutable slice over the allocated memory block.

pub const unsafe fn as_slice_aligned<T>(&self) -> &[T]

Returns a slice over the allocated memory block, aligned to the given type.

Safety

Make sure the alignment is correct.

pub const unsafe fn as_mut_slice_aligned<T>(&mut self) -> &mut [T]

Returns a mutable slice over the allocated memory block, aligned to the given type.

Safety

Make sure the alignment is correct.

pub const fn len(&self) -> usize

Returns the size of the allocated memory block.