winsafe::guard

Struct DestroyMenuGuard

pub struct DestroyMenuGuard { /* private fields */ }

Available on crate feature user only.

RAII implementation for HMENU which automatically calls DestroyMenu when the object goes out of scope.

Implementations

impl DestroyMenuGuard

pub const unsafe fn new(handle: HMENU) -> Self

Constructs the guard by taking ownership of the handle.

This method can be used as an escape hatch to interoperate with other libraries.

Safety

Be sure the handle must be freed with the specified function at the end of scope.

pub fn leak(&mut self) -> HMENU

Ejects the underlying handle, leaving a Handle::INVALID in its place.

Since the internal handle will be invalidated, the destructor will not run. It’s your responsability to run it, otherwise you’ll cause a resource leak.