winsafe::guard

Struct FreeLibraryGuard

pub struct FreeLibraryGuard { /* private fields */ }

Available on crate feature kernel only.

RAII implementation for HINSTANCE which automatically calls FreeLibrary when the object goes out of scope.

Implementations

impl FreeLibraryGuard

pub const unsafe fn new(handle: HINSTANCE) -> Self

Constructs the guard by taking ownership of the handle.

This method can be used as an escape hatch to interoperate with other libraries.

Safety

Be sure the handle must be freed with the specified function at the end of scope.

pub fn leak(&mut self) -> HINSTANCE

Ejects the underlying handle, leaving a Handle::INVALID in its place.

Since the internal handle will be invalidated, the destructor will not run. It’s your responsability to run it, otherwise you’ll cause a resource leak.