winsafe::guard
Struct HeapDestroyGuard

pub struct HeapDestroyGuard { /* private fields */ }

Available on crate feature kernel only.

RAII implementation for HHEAP which automatically calls HeapDestroy when the object goes out of scope.

Implementations

impl HeapDestroyGuard

pub const unsafe fn new(handle: HHEAP) -> Self

Constructs the guard by taking ownership of the handle.

This method can be used as an escape hatch to interoperate with other libraries.

Safety

Be sure the handle must be freed with the specified function at the end of scope.

pub fn leak(&mut self) -> HHEAP

Ejects the underlying handle, leaving a Handle::INVALID in its place.

Since the internal handle will be invalidated, the destructor will not run. It’s your responsability to run it, otherwise you’ll cause a resource leak.