winsafe::guard

Struct RegCloseKeyGuard

pub struct RegCloseKeyGuard { /* private fields */ }

Available on crate feature advapi only.

RAII implementation for HKEY which automatically calls RegCloseKey when the object goes out of scope.

Implementations

impl RegCloseKeyGuard

pub const unsafe fn new(hkey: HKEY) -> Self

Constructs the guard by taking ownership of the handle.

Safety

Be sure the handle must be freed with RegCloseKey at the end of scope.

pub fn leak(&mut self) -> HKEY

Ejects the underlying handle, leaving Handle::INVALID in its place.

Since the internal handle will be invalidated, the destructor will not run. It’s your responsibility to run it, otherwise you’ll cause a resource leak.