Struct winsafe::guard::SidGuard

pub struct SidGuard { /* private fields */ }

Available on crate features kernel and advapi only.

RAII implementation for SID, returned by CopySid, CreateWellKnownSid, GetWindowsAccountDomainSid and LookupAccountName, which automatically frees the underlying memory block when the object goes out of scope.

Implementations

impl SidGuard

pub const unsafe fn new(ptr: GlobalFreeGuard) -> Self

Constructs a new guard by taking ownership of the data.

Safety

Be sure the data is an allocated SID structure.