winsafe::gui::collections

Struct ListBoxItems

pub struct ListBoxItems<'a> { /* private fields */ }

Available on crate feature gui only.

Exposes item methods of a ListBox control.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl<'a> ListBoxItems<'a>

pub fn add(&self, items: &[impl AsRef<str>]) -> SysResult<()>

Adds new texts by sending lb::AddString messages.

Examples

use winsafe::{self as w, prelude::*, gui};

let my_list: gui::ListBox; // initialized somewhere

my_list.items().add(&["John", "Mary"])?;

pub fn count(&self) -> SysResult<u32>

Retrieves the number of items by sending an lb::GetCount message.

pub fn delete(&self, index: u32) -> SysResult<()>

Deletes the item at the given index by sending an lb::DeleteString message.

pub fn delete_all(&self)

Deletes all items by sending an lb::ResetContent message.

pub fn ensure_visible(&self, index: u32) -> SysResult<()>

Ensures that the specified item in a list box is visible by sending an lb::SetTopIndex message.

pub fn iter( &self, ) -> SysResult<impl DoubleEndedIterator<Item = SysResult<String>> + 'a>

Returns an iterator over the texts.

Examples

use winsafe::{self as w, prelude::*, gui};

let my_list: gui::ListBox; // initialized somewhere

for text in my_list.items().iter()? {
    println!("Text {}", text?);
}

pub fn iter_selected( &self, ) -> SysResult<impl DoubleEndedIterator<Item = SysResult<(u32, String)>> + 'a>

Returns an iterator over the currently selected texts.

This method works for both single and multiple-selection lists.

Examples

use winsafe::{self as w, prelude::*, gui};

let my_list: gui::ListBox; // initialized somewhere

for idx_text in my_list.items().iter_selected()? {
    let (idx, text) = idx_text?;
    println!("Text {idx}: {text}");
}

pub fn selected_count(&self) -> SysResult<u32>

Retrieves the number of selected items by sending an lb::GetSelCount message.

pub fn text(&self, index: u32) -> SysResult<String>

Retrieves the text at the given position, if any, by sending a lb::GetText message.