winsafe::gui::collections

Struct StatusBarParts

pub struct StatusBarParts<'a> { /* private fields */ }

Available on crate feature gui only.

Exposes the part methods of a StatusBar control.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl<'a> StatusBarParts<'a>

pub fn count(&self) -> u32

Retrieves the number of parts by sending an sb::GetParts message.

pub const fn get(&self, index: u32) -> StatusBarPart<'a>

Retrieves the part at the given zero-based position.

Note: This method is cheap – even if index is beyond the range of existing parts, an object will still be returned. However, operations upon this object will produce no effect.

pub fn iter(&self) -> impl DoubleEndedIterator<Item = StatusBarPart<'a>> + 'a

Returns an iterator over all parts.

pub fn last(&self) -> Option<StatusBarPart<'a>>

Returns the last part, if any.

pub fn set_texts(&self, texts: &[Option<impl AsRef<str>>])

Sets the texts of multiple parts at once.

Panics

Panics if texts is empty, or if the number of texts is greater than the number of columns.

Examples

use winsafe::{self as w, prelude::*, gui};

let my_sb: gui::StatusBar; // initialized somewhere

my_sb.parts().set_texts(&[
    Some("First"),
    None, // 2nd part won't have its text changed
    Some("Third"),
    Some("Fourth"),
]);