Module winsafe::gui::events

Available on crate feature gui only.

Events exposed by windows and controls, which allow the handling of native Windows messages.

Structs

ButtonEvents

Exposes button control notifications.

ComboBoxEvents

Exposes combo box control notifications.

DateTimePickerEvents

Exposes date and time picker control notifications.

EditEvents

Exposes edit control notifications.

LabelEvents

Exposes label control notifications.

ListBoxEvents

Exposes list box control notifications.

ListViewEvents

Exposes list view control notifications.

MonthCalendarEvents

Exposes month calendar control notifications.

RadioGroupEvents

Exposes button control notifications for a RadioGroup.

StatusBarEvents

Exposes status bar control notifications.

TabEvents

Exposes tab control notifications.

TrackbarEvents

Exposes trackbar control notifications.

TreeViewEvents

Exposes tree view control notifications.

UpDownEvents

Exposes up-down control notifications.

WindowEvents

Exposes window messages.

WindowEventsAll

Exposes window messages, plus timer and native control notifications.