Trait winsafe::gui::events::prelude::EventsViewAll

pub trait EventsViewAll: SealedEventsWmNfy + EventsView {
    fn wm_notify<F>(&self, id_from: i32, code: NM, func: F)
    where
        F: Fn(Notify<'_>) -> ErrResult<isize> + 'static,
    { ... }
}

Exposes the methods of WindowEventsAll.

Provided methods

fn wm_notify<F>(&self, id_from: i32, code: NM, func: F) where
F: Fn(Notify<'_>) -> ErrResult<isize> + 'static,

WM_NOTIFY message, for specific ID and notification code.

A notification must be narrowed by the notification code and the control ID, so the closure will be fired for that specific control at the specific event.

Note: Instead of using this event, you should always prefer the specific notifications, which will give you the correct notification struct. This generic method should be used when you have a custom, non-standard window notification.