Struct winsafe::gui::events::ComboBoxEvents

pub struct ComboBoxEvents(/* private fields */);

Available on crate feature gui only.

Exposes combo box control notifications.

These event methods are just proxies to the WindowEvents of the parent window, who is the real responsible for the child event handling.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl ComboBoxEvents

pub fn cbn_close_up<F>(&self, func: F)where F: Fn() -> AnyResult<()> + 'static,

CBN_CLOSEUP command notification.

pub fn cbn_dbl_clk<F>(&self, func: F)where F: Fn() -> AnyResult<()> + 'static,

CBN_DBLCLK command notification.

pub fn cbn_drop_down<F>(&self, func: F)where F: Fn() -> AnyResult<()> + 'static,

CBN_DROPDOWN command notification.

pub fn cbn_edit_change<F>(&self, func: F)where F: Fn() -> AnyResult<()> + 'static,

CBN_EDITCHANGE command notification.

pub fn cbn_edit_update<F>(&self, func: F)where F: Fn() -> AnyResult<()> + 'static,

CBN_EDITUPDATE command notification.

pub fn cbn_err_space<F>(&self, func: F)where F: Fn() -> AnyResult<()> + 'static,

CBN_ERRSPACE command notification.

pub fn cbn_kill_focus<F>(&self, func: F)where F: Fn() -> AnyResult<()> + 'static,

CBN_KILLFOCUS command notification.

pub fn cbn_sel_change<F>(&self, func: F)where F: Fn() -> AnyResult<()> + 'static,

CBN_SELCHANGE command notification.

Examples

use winsafe::{self as w, prelude::*, gui};

let wnd: gui::WindowMain; // initialized somewhere
let cmb: gui::ComboBox;

let cmb2 = cmb.clone(); // to pass into the closure

cmb.on().cbn_sel_change(
    move || -> w::AnyResult<()> {
        if let Some(sel_text) = cmb2.items().selected_text() {
            println!("New selected text: {}", sel_text);
        }
        Ok(())
    },
);

pub fn cbn_sel_end_cancel<F>(&self, func: F)where F: Fn() -> AnyResult<()> + 'static,

CBN_SELENDCANCEL command notification.

pub fn cbn_sel_end_ok<F>(&self, func: F)where F: Fn() -> AnyResult<()> + 'static,

CBN_SELENDOK command notification.

pub fn cbn_set_focus<F>(&self, func: F)where F: Fn() -> AnyResult<()> + 'static,

CBN_SETFOCUS command notification.