Struct winsafe::gui::events::DateTimePickerEvents

pub struct DateTimePickerEvents(/* private fields */);

Available on crate feature gui only.

Exposes date and time picker control notifications.

These event methods are just proxies to the WindowEvents of the parent window, who is the real responsible for the child event handling.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl DateTimePickerEvents

pub fn dtn_close_up<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

DTN_CLOSEUP notification.

pub fn dtn_date_time_change<F>(&self, func: F)
where F: Fn(&NMDATETIMECHANGE) -> AnyResult<()> + 'static,

DTN_DATETIMECHANGE notification.

Examples

use winsafe::{self as w, prelude::*, gui};

let wnd: gui::WindowMain; // initialized somewhere
let dtp: gui::DateTimePicker;

dtp.on().dtn_date_time_change(
    move |p: &w::NMDATETIMECHANGE| -> w::AnyResult<()> {
        println!("{}-{}-{}", p.st.wYear, p.st.wMonth, p.st.wDay);
        Ok(())
    },
);

pub fn dtn_drop_down<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

DTN_DROPDOWN notification.

pub fn dtn_format<F>(&self, func: F)
where F: Fn(&mut NMDATETIMEFORMAT<'_>) -> AnyResult<()> + 'static,

DTN_FORMAT notification.

pub fn dtn_format_query<F>(&self, func: F)
where F: Fn(&mut NMDATETIMEFORMATQUERY<'_>) -> AnyResult<()> + 'static,

DTN_FORMATQUERY notification.

pub fn dtn_user_string<F>(&self, func: F)
where F: Fn(&mut NMDATETIMESTRING<'_>) -> AnyResult<()> + 'static,

DTN_USERSTRING notification.

pub fn dtn_wm_key_down<F>(&self, func: F)
where F: Fn(&NMDATETIMEWMKEYDOWN<'_>) -> AnyResult<()> + 'static,

DTN_WMKEYDOWN notification.

pub fn nm_kill_focus<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

NM_KILLFOCUS notification.

pub fn nm_set_focus<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

NM_SETFOCUS notification.