Struct winsafe::gui::events::EditEvents

pub struct EditEvents(/* private fields */);

Available on crate feature gui only.

Exposes edit control notifications.

These event methods are just proxies to the WindowEvents of the parent window, who is the real responsible for the child event handling.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl EditEvents

pub fn en_align_ltr_ec<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

EN_ALIGN_LTR_EC command notification.

pub fn en_align_rtl_ec<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

EN_ALIGN_LTR_EC command notification.

pub fn en_change<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

EN_CHANGE command notification.

Examples

use winsafe::{self as w, prelude::*, gui};

let wnd: gui::WindowMain; // initialized somewhere
let txt: gui::Edit;

let txt2 = txt.clone(); // to pass into the closure

txt.on().en_change(
    move || -> w::AnyResult<()> {
        println!("Text: {}", txt2.text());
        Ok(())
    },
);

pub fn en_err_space<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

EN_ERRSPACE command notification.

pub fn en_h_scroll<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

EN_HSCROLL command notification.

pub fn en_kill_focus<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

EN_KILLFOCUS command notification.

pub fn en_max_text<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

EN_MAXTEXT command notification.

pub fn en_set_focus<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

EN_SETFOCUS command notification.

pub fn en_update<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

EN_UPDATE command notification.

pub fn en_v_scroll<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

EN_VSCROLL command notification.