winsafe::gui::events

Struct HeaderEvents

pub struct HeaderEvents(/* private fields */);

Available on crate feature gui only.

Exposes header control notifications.

These event methods are just proxies to the WindowEvents of the parent window, who is the real responsible for the child event handling.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl HeaderEvents

pub fn hdn_begin_drag<F>(&self, func: F) -> &Self
where F: Fn(&NMHEADER<'_>) -> AnyResult<bool> + 'static,

HDN_BEGINDRAG notification.

pub fn hdn_begin_filter_edit<F>(&self, func: F) -> &Self
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_BEGINFILTEREDIT notification.

pub fn hdn_begin_track<F>(&self, func: F) -> &Self
where F: Fn(&NMHEADER<'_>) -> AnyResult<bool> + 'static,

HDN_BEGINTRACK notification.

pub fn hdn_divider_dbl_click<F>(&self, func: F) -> &Self
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_DIVIDERDBLCLICK notification.

pub fn hdn_drop_down<F>(&self, func: F) -> &Self
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_DROPDOWN notification.

pub fn hdn_end_drag<F>(&self, func: F) -> &Self
where F: Fn(&NMHEADER<'_>) -> AnyResult<bool> + 'static,

HDN_ENDDRAG notification.

pub fn hdn_end_filter_edit<F>(&self, func: F) -> &Self
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_ENDFILTEREDIT notification.

pub fn hdn_end_track<F>(&self, func: F) -> &Self
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_ENDTRACK notification.

pub fn hdn_filter_btn_click<F>(&self, func: F) -> &Self
where F: Fn(&NMHDFILTERBTNCLICK) -> AnyResult<bool> + 'static,

HDN_FILTERBTNCLICK notification.

pub fn hdn_filter_change<F>(&self, func: F) -> &Self
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_FILTERCHANGE notification.

pub fn hdn_get_disp_info<F>(&self, func: F) -> &Self
where F: Fn(&mut NMHDDISPINFO<'_>) -> AnyResult<isize> + 'static,

HDN_GETDISPINFO notification.

pub fn hdn_item_changed<F>(&self, func: F) -> &Self
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_ITEMCHANGED notification.

pub fn hdn_item_changing<F>(&self, func: F) -> &Self
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_ITEMCHANGING notification.

pub fn hdn_item_click<F>(&self, func: F) -> &Self
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_ITEMCLICK notification.

pub fn hdn_item_dbl_click<F>(&self, func: F) -> &Self
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_ITEMDBLCLICK notification.

pub fn hdn_item_key_down<F>(&self, func: F) -> &Self
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_ITEMKEYDOWN notification.

pub fn hdn_item_state_icon_click<F>(&self, func: F) -> &Self
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_ITEMSTATEICONCLICK notification.

pub fn hdn_overflow_click<F>(&self, func: F) -> &Self
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_OVERFLOWCLICK notification.

pub fn hdn_track<F>(&self, func: F) -> &Self
where F: Fn(&NMHDFILTERBTNCLICK) -> AnyResult<bool> + 'static,

HDN_TRACK notification.

pub fn nm_custom_draw<F>(&self, func: F) -> &Self
where F: Fn(&mut NMCUSTOMDRAW) -> AnyResult<CDRF> + 'static,

NM_CUSTOMDRAW notification.

pub fn nm_r_click<F>(&self, func: F) -> &Self
where F: Fn() -> AnyResult<i32> + 'static,

NM_RCLICK notification.

pub fn nm_released_capture<F>(&self, func: F) -> &Self
where F: Fn() -> AnyResult<()> + 'static,

NM_RELEASEDCAPTURE notification.