winsafe::gui::events

Struct LabelEvents

pub struct LabelEvents(/* private fields */);

Available on crate feature gui only.

Exposes label control notifications.

These event methods are just proxies to the WindowEvents of the parent window, who is the real responsible for the child event handling.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl LabelEvents

pub fn stn_clicked<F>(&self, func: F) -> &Self
where F: Fn() -> AnyResult<()> + 'static,

STN_CLICKED notification.

pub fn stn_dbl_clk<F>(&self, func: F) -> &Self
where F: Fn() -> AnyResult<()> + 'static,

STN_DBLCLK notification.

pub fn stn_disable<F>(&self, func: F) -> &Self
where F: Fn() -> AnyResult<()> + 'static,

STN_DISABLE notification.

pub fn stn_enable<F>(&self, func: F) -> &Self
where F: Fn() -> AnyResult<()> + 'static,

STN_ENABLE notification.