Struct winsafe::gui::events::TabEvents

pub struct TabEvents(/* private fields */);

Available on crate feature gui only.

Exposes tab control notifications.

These event methods are just proxies to the WindowEvents of the parent window, who is the real responsible for the child event handling.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl TabEvents

pub fn nm_click<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

NM_KILLFOCUS notification.

pub fn nm_dbl_clk<F>(&self, func: F)
where F: Fn() -> AnyResult<i32> + 'static,

NM_DBLCLK notification.

pub fn nm_r_click<F>(&self, func: F)
where F: Fn() -> AnyResult<i32> + 'static,

NM_RCLICK notification.

pub fn nm_r_dbl_clk<F>(&self, func: F)
where F: Fn() -> AnyResult<i32> + 'static,

NM_RDBLCLK notification.

pub fn nm_released_capture<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

NM_RELEASEDCAPTURE notification.

pub fn tcn_focus_change<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

TCN_FOCUSCHANGE notification.

pub fn tcn_get_object<F>(&self, func: F)
where F: Fn(&NMOBJECTNOTIFY<'_>) -> AnyResult<()> + 'static,

TCN_GETOBJECT notification.

pub fn tcn_key_down<F>(&self, func: F)
where F: Fn(&NMTCKEYDOWN) -> AnyResult<()> + 'static,

TCN_KEYDOWN notification.

pub fn tcn_sel_change<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

TCN_SELCHANGE notification.

Examples

use winsafe::{self as w, prelude::*, gui};

let wnd: gui::WindowMain; // initialized somewhere
let tab: gui::Tab;

tab.on().tcn_sel_change(
    move || -> w::AnyResult<()> {
        println!("Changed tab.");
        Ok(())
    },
);

pub fn tcn_sel_changing<F>(&self, func: F)
where F: Fn() -> AnyResult<bool> + 'static,

TCN_SELCHANGING notification.