Struct winsafe::gui::events::TreeViewEvents

pub struct TreeViewEvents(_);

Exposes tree view control notifications.

These event methods are just proxies to the WindowEvents of the parent window, who is the real responsible for the child event handling.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl TreeViewEvents

pub fn tvn_delete_item<F>(&self, func: F) where
F: Fn(&NMTREEVIEW<'_, '_>) -> ErrResult<()> + 'static,

TVN_DELETEITEM notification.

Notifies a tree-view control’s parent window that an item is being deleted.

pub fn tvn_item_changed<F>(&self, func: F) where
F: Fn(&NMTREEVIEW<'_, '_>) -> ErrResult<()> + 'static,

TVN_ITEMCHANGED notification.

Notifies a tree-view control’s parent window that item attributes have changed.

pub fn tvn_item_changing<F>(&self, func: F) where
F: Fn(&NMTREEVIEW<'_, '_>) -> ErrResult<bool> + 'static,

TVN_ITEMCHANGING notification.

Notifies a tree-view control’s parent window that item attributes are about to change.

pub fn tvn_item_expanded<F>(&self, func: F) where
F: Fn(&NMTREEVIEW<'_, '_>) -> ErrResult<()> + 'static,

TVN_ITEMEXPANDED notification.

Notifies a tree-view control’s parent window that a parent item’s list of child items has expanded or collapsed.

pub fn tvn_item_expanding<F>(&self, func: F) where
F: Fn(&NMTREEVIEW<'_, '_>) -> ErrResult<bool> + 'static,

TVN_ITEMEXPANDING notification.

Notifies a tree-view control’s parent window that a parent item’s list of child items is about to expand or collapse.

pub fn tvn_sel_changed<F>(&self, func: F) where
F: Fn(&NMTREEVIEW<'_, '_>) -> ErrResult<()> + 'static,

TVN_SELCHANGED notification.

Notifies a tree-view control’s parent window that the selection has changed from one item to another.

pub fn tvn_sel_changing<F>(&self, func: F) where
F: Fn(&NMTREEVIEW<'_, '_>) -> ErrResult<bool> + 'static,

TVN_SELCHANGING notification.

Notifies a tree-view control’s parent window that the selection is about to change from one item to another.

pub fn nm_click<F>(&self, func: F) where
F: Fn() -> ErrResult<i32> + 'static,

NM_CLICK notification.

Notifies the parent window of a tree-view control that the user has clicked the left mouse button within the control.

pub fn nm_custom_draw<F>(&self, func: F) where
F: Fn(&NMTVCUSTOMDRAW) -> ErrResult<CDRF> + 'static,

NM_CUSTOMDRAW notification.

Sent by a tree-view control to notify its parent window about drawing operations.

pub fn nm_dbl_clk<F>(&self, func: F) where
F: Fn() -> ErrResult<i32> + 'static,

NM_DBLCLK notification.

Notifies the parent window of a tree-view control that the user has double-clicked the left mouse button within the control.

pub fn nm_kill_focus<F>(&self, func: F) where
F: Fn() -> ErrResult<()> + 'static,

NM_KILLFOCUS notification.

Notifies a tree-view control’s parent window that the control has lost the input focus.

pub fn nm_r_click<F>(&self, func: F) where
F: Fn() -> ErrResult<i32> + 'static,

NM_RCLICK notification.

Notifies the parent window of a tree-view control that the user has clicked the right mouse button within the control.

pub fn nm_r_dbl_clk<F>(&self, func: F) where
F: Fn() -> ErrResult<i32> + 'static,

NM_RDBLCLK notification.

Notifies the parent of a tree-view control that the user has double-clicked the right mouse button within the control.

pub fn nm_return<F>(&self, func: F) where
F: Fn() -> ErrResult<()> + 'static,

NM_RETURN notification.

Notifies a tree-view control’s parent window that the control has the input focus and that the user has pressed the ENTER key.

pub fn nm_set_cursor<F>(&self, func: F) where
F: Fn(&NMMOUSE) -> ErrResult<i32> + 'static,

NM_MOUSE notification.

Notifies a tree-view control’s parent window that the control is setting the cursor in response to a wm::SetCursor message.

pub fn nm_set_focus<F>(&self, func: F) where
F: Fn() -> ErrResult<()> + 'static,

NM_SETFOCUS notification.

Notifies a tree-view control’s parent window that the control has received the input focus.