Struct winsafe::gui::events::WindowEventsAll

pub struct WindowEventsAll { /* private fields */ }

Available on crate feature gui only.

Exposes window messages, plus native control notifications.

You cannot directly instantiate this object, it is created internally by the window.

Implementations

impl WindowEventsAll

pub fn wm_timer<F>(&self, timer_id: u32, func: F)where
F: Fn() -> AnyResult<()> + 'static,

WM_TIMER message, narrowed to a specific timer ID.

pub fn wm_command<F>(&self, code: impl Into<CMD>, ctrl_id: u16, func: F)where
F: Fn() -> AnyResult<()> + 'static,

WM_COMMAND message, for specific code and control ID.

A command notification must be narrowed by the command code and the control ID, so the closure will be fired for that specific control at that specific event.

Note: Instead of using this event, you should always prefer the specific command notifications, which will give you the correct message parameters. This generic method should be used only when you have a custom, non-standard window notification.

use winsafe::prelude::*;
use winsafe::{co, gui, msg, AnyResult};

let wnd: gui::WindowMain; // initialized somewhere

const ID_BTN: u16 = 1000;

wnd.on().wm_command(co::BN::CLICKED, ID_BTN,
    move || -> AnyResult<()> {
        println!("Button clicked!");
        Ok(())
    },
);

pub fn wm_command_accel_menu<F>(&self, ctrl_id: u16, func: F)where
F: Fn() -> AnyResult<()> + 'static,

WM_COMMAND message, handling both CMD::Accelerator and CMD::Menu, for a specific command ID.

Ideal to be used with menu commands whose IDs are shared with accelerators.

pub fn wm_notify<F>(&self, id_from: u16, code: impl Into<NM>, func: F)where
F: Fn(Notify<'_>) -> AnyResult<Option<isize>> + 'static,

WM_NOTIFY message, for specific ID and notification code.

Note: Instead of using this event, you should always prefer the specific notifications, which will give you the correct notification struct. This generic method should be used only when you have a custom, non-standard window notification.