Struct winsafe::gui::spec::ComboBoxItems

pub struct ComboBoxItems<'a> { /* private fields */ }

This is supported on crate feature gui only.

Exposes item methods of a ComboBox control.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl<'a> ComboBoxItems<'a>

pub fn add(&self, items: &[impl AsRef<str>]) -> WinResult<()>

Adds new texts by sending cb::AddString messages.

Examples

use winsafe::prelude::*;
use winsafe::gui;

let my_combo: gui::ComboBox; // initialized somewhere

my_combo.items().add(&["John", "Mary"]);

pub fn count(&self) -> WinResult<u32>

Retrieves the number of items by sending a cb::GetCount message.

pub fn delete(&self, index: u32) -> WinResult<()>

Deletes the item at the given index by sending a cb::DeleteString message.

pub fn delete_all(&self)

Deletes all items by sending a cb::ResetContent message.

pub fn iter(&self) -> impl Iterator<Item = WinResult<String>> + 'a

Returns an iterator over the texts.

Examples

use winsafe::prelude::*;
use winsafe::gui;

let my_combo: gui::ComboBox; // initialized somewhere

for text in my_combo.items().iter() {
    let text = text?;
    println!("Text {}", text);
}

pub fn select(&self, index: Option<u32>)

Sets the currently selected index, or clears it, by sending a cb::SetCurSel message.

pub fn selected_index(&self) -> Option<u32>

Retrieves the index of the currently selected item, if any, by sending a cb::GetCurSel message.

pub fn selected_text(&self) -> WinResult<Option<String>>

Retrieves the currently selected text, if any, by calling selected_index and text methods.

pub fn text(&self, index: u32) -> WinResult<Option<String>>

Retrieves the text at the given position, if any, by sending a cb::GetLbText message.