Struct winsafe::gui::spec::ListViewColumns

pub struct ListViewColumns<'a, T: 'static> { /* private fields */ }

Available on crate feature gui only.

Exposes column methods of a ListView control.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl<'a, T> ListViewColumns<'a, T>

pub fn add(&self, texts_and_widths: &[(impl AsRef<str>, u32)])

Adds many columns at once by sending an lvm::InsertColumn message.

Widths will be adjusted to match current system DPI.

Examples

use winsafe::{self as w, prelude::*, gui};

let my_list: gui::ListView; // initialized somewhere

my_list.columns().add(&[
    ("Name", 300),
    ("Address", 500),
]);

pub fn count(&self) -> u32

Retrieves the number of columns by sending an hdm::GetItemCount message to the handle returned by lvm::GetHeader.

pub const fn get(&self, index: u32) -> ListViewColumn<'a, T>

Retrieves the column at the given zero-based position.

Note: This method is cheap – even if index is beyond the range of existing columns, an object will still be returned. However, operations upon this object will produce no effect.

pub fn iter(&self) -> impl Iterator<Item = ListViewColumn<'a, T>> + 'a

Returns an iterator over all columns.