Struct winsafe::gui::spec::ListViewItem

pub struct ListViewItem<'a> { /* private fields */ }

This is supported on crate feature gui only.

Represents a single item of a ListView control.

Note: Each object keeps the zero-based index of an item. If new items are added/removed from the list view control, the object may then point to a different item.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl<'a> ListViewItem<'a>

pub fn delete(&self) -> WinResult<()>

Deletes the item by sending an lvm::DeleteItem message.

pub fn ensure_visible(&self) -> WinResult<()>

Scrolls the list by sending an lvm::EnsureVisible message so that the item is visible in the list.

pub fn focus(&self) -> WinResult<()>

Sets the item as the focused one sending an lvm:SetItemState message.

pub fn icon_index(&self) -> WinResult<Option<u32>>

Retrieves the icon index of the item by sending an lvm::GetItem message.

pub const fn index(&self) -> u32

Returns the zero-based index of the item.

pub fn is_focused(&self) -> bool

Tells if the item is the focused one by sending an lvm::GetItemState message.

pub fn is_selected(&self) -> bool

Tells if the item is selected by sending an lvm::GetItemState message.

pub fn is_visible(&self) -> bool

Tells if the item is currently visible by sending an lvm::IsItemVisible message.

pub fn lparam(&self) -> WinResult<isize>

Retrieves the user-defined value by sending an lvm::GetItem message.

pub fn map_index_to_id(&self) -> Option<u32>

Retrieves the unique ID for the item index by sending an lvm::MapIndexToId message.

If the item index has became invalid, returns None.

pub fn rect(&self, portion: LVIR) -> WinResult<RECT>

Retrieves the bound rectangle of item by sending an lvm::GetItemRect message.

pub fn select(&self, set: bool) -> WinResult<()>

Sets or removes the selection from the item by sending an lvm::SetItemState message.

pub fn set_icon_index(&self, icon_index: Option<u32>) -> WinResult<()>

Sets the icon index of the item by sending an lvm::SetItem message.

pub fn set_lparam(&self, lparam: isize) -> WinResult<()>

Sets the user-defined value by sending an lvm::SetItem message.

pub fn set_text(&self, column_index: u32, text: &str) -> WinResult<()>

Sets the text of the item under a column by sending an lvm::SetItemText message.

pub fn text(&self, column_index: u32) -> String

Retrieves the text of an item under a column by sending an lvm::GetItemText message.