Struct winsafe::gui::spec::TreeViewItem

pub struct TreeViewItem<'a> { /* private fields */ }

Available on crate feature gui only.

A single item of a TreeView control.

Each object keeps an unique HTREEITEM handle.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl<'a> TreeViewItem<'a>

pub fn add_child(&self, text: &str, icon_index: Option<u32>) -> Self

Adds a new child item by sending a tvm::InsertItem message, and returns the newly added item.

pub fn delete(&self)

Deletes the item by sending a tvm::DeleteItem message.

pub fn edit_label(&self) -> HWND

Begins in-place editing of the item’s text by sending a tvm::EditLabel message.

Returns a handle to the edit control.

pub fn ensure_visible(&self) -> bool

Ensures that a tree-view item is visible, expanding the parent item or scrolling the tree-view control, if necessary, by sending a tvm::EnsureVisible message.

Returns whether a scroll occurred and no items were expanded.

pub fn expand(&self, expand: bool)

Expands or collapse the item by sending a tvm::Expand message.

pub const fn htreeitem(&self) -> &HTREEITEM

Returns the underlying handle of the item.

pub fn is_expanded(&self) -> bool

Tells if the item is expanded by sending a tvm::GetItemState message.

pub fn is_root(&self) -> bool

Tells if the item is a root by sending a tvm::GetNextItem message.

pub fn iter_children(&self) -> impl Iterator<Item = TreeViewItem<'a>> + 'a

Returns an iterator over the child items.

pub fn iter_next_siblings(&self) -> impl Iterator<Item = TreeViewItem<'a>> + 'a

Returns an iterator over the next sibling items.

pub fn iter_prev_siblings(&self) -> impl Iterator<Item = TreeViewItem<'a>> + 'a

Returns an iterator over the previous sibling items.

pub fn parent(&self) -> Option<Self>

Retrieves the parent of the item by sending a tvm::GetNextItem message.

pub fn set_text(&self, text: &str)

Sets the text of the item by sending a tvm::SetItem message.

pub fn text(&self) -> String

Retrieves the text of the item by sending a tvm::GetItem message.