Struct winsafe::gui::Button

pub struct Button(_);

Available on crate feature gui only.

Native button control.

Implementations

impl Button

pub fn new(parent: &impl GuiParent, opts: ButtonOpts) -> Button

Instantiates a new Button object, to be created on the parent window with HWND::CreateWindowEx.

pub fn new_dlg(
parent: &impl GuiParent,
ctrl_id: u16,
resize_behavior: (Horz, Vert)
) -> Button

Instantiates a new Button object, to be loaded from a dialog resource with HWND::GetDlgItem.

pub fn trigger_click(&self)

Fires the click event for the button by sending a bm::Click message.