Struct winsafe::gui::ButtonOpts

pub struct ButtonOpts {
    pub text: String,
    pub position: POINT,
    pub width: u32,
    pub height: u32,
    pub button_style: BS,
    pub window_style: WS,
    pub window_ex_style: WS_EX,
    pub ctrl_id: u16,
    pub horz_resize: Horz,
    pub vert_resize: Vert,
}

This is supported on crate feature gui only.

Options to create a Button programmatically with Button::new.

Fields
text: String

Text of the control to be created.

Defaults to empty string.

position: POINT

Control position within parent client area, to be created.

If the parent window is a dialog, the values are in Dialog Template Units; otherwise in pixels, which will be multiplied to match current system DPI.

Defaults to 0 x 0.

width: u32

Control width, to be created.

If the parent window is a dialog, the value is in Dialog Template Units; otherwise in pixels, which will be multiplied to match current system DPI.

Defaults to 80.

height: u32

Control height, to be created.

If the parent window is a dialog, the value is in Dialog Template Units; otherwise in pixels, which will be multiplied to match current system DPI.

Defaults to 23.

button_style: BS

Button styles to be created.

Defaults to BS::PUSHBUTTON.

Suggestions:

replace with BS::DEFPUSHBUTTON for the default button of the window;

add BS::NOTIFY to receive notifications other than the simple click.

window_style: WS

Window styles to be created.

Defaults to WS::CHILD | WS::VISIBLE | WS::TABSTOP | WS::GROUP.

window_ex_style: WS_EX

Extended window styles to be created.

Defaults to WS_EX::LEFT.

ctrl_id: u16

The control ID.

Defaults to an auto-generated ID.

horz_resize: Horz

Horizontal behavior when the parent is resized.

Defaults to Horz::None.

vert_resize: Vert

Vertical behavior when the parent is resized.

Defaults to Vert::None.