winsafe::gui
Struct CheckBox

pub struct CheckBox(/* private fields */);

Available on crate feature gui only.

Native check box control, actually a variation of the ordinary Button: just a button with a specific style.

Implementations

impl CheckBox

pub fn new(parent: &impl GuiParent, opts: CheckBoxOpts) -> Self

Instantiates a new CheckBox object, to be created on the parent window with HWND::CreateWindowEx.

Panics

Panics if the parent window was already created – that is, you cannot dynamically create a CheckBox in an event closure.

pub fn new_dlg( parent: &impl GuiParent, ctrl_id: u16, resize_behavior: (Horz, Vert), ) -> Self

Instantiates a new CheckBox object, to be loaded from a dialog resource with HWND::GetDlgItem.

Panics

Panics if the parent dialog was already created – that is, you cannot dynamically create a CheckBox in an event closure.

pub fn check_state(&self) -> CheckState

Retrieves the current check state by sending a bm::GetCheck message.

pub fn emulate_click(&self)

Emulates the click event for the check box by sending a bm::Click message.

pub fn is_checked(&self) -> bool

Calls check_state and compares the result with CheckState::Checked.

pub fn set_check_state(&self, state: CheckState)

Sets the current check state by sending a bm::SetCheck message.

pub fn set_check_state_and_trigger(&self, state: CheckState)

Sets the current check state by sending a bm::SetCheck message, then sends a wm::Command message to the parent, so it can handle the event.

pub fn set_text_and_resize(&self, text: &str)

Calls set_text and resizes the control to exactly fit the new text.