winsafe::gui
Struct ComboBox

pub struct ComboBox(/* private fields */);

Available on crate feature gui only.

Native combo box control.

Implementations

impl ComboBox

pub fn new(parent: &impl GuiParent, opts: ComboBoxOpts) -> Self

Instantiates a new ComboBox object, to be created on the parent window with HWND::CreateWindowEx.

Panics

Panics if the parent window was already created – that is, you cannot dynamically create a ComboBox in an event closure.

Examples

use winsafe::{self as w, prelude::*, gui};

let wnd: gui::WindowMain; // initialized somewhere

let cmb = gui::ComboBox::new(
    &wnd,
    gui::ComboBoxOpts {
        position: (10, 10),
        width: 140,
        items: vec![
            "Avocado".to_owned(),
            "Banana".to_owned(),
            "Grape".to_owned(),
            "Orange".to_owned(),
        ],
        selected_item: Some(0),
        ..Default::default()
    },
);

pub fn new_dlg( parent: &impl GuiParent, ctrl_id: u16, resize_behavior: (Horz, Vert), ) -> Self

Instantiates a new ComboBox object, to be loaded from a dialog resource with HWND::GetDlgItem.

Panics

Panics if the parent dialog was already created – that is, you cannot dynamically create a ComboBox in an event closure.

pub const fn items(&self) -> ComboBoxItems<'_>

Item methods.