Struct winsafe::gui::DateTimePicker

pub struct DateTimePicker(/* private fields */);

Available on crate feature gui only.

Native date and time picker control.

Implementations

impl DateTimePicker

pub fn new(parent: &impl GuiParent, opts: DateTimePickerOpts) -> Self

Instantiates a new DateTimePicker object, to be created on the parent window with HWND::CreateWindowEx.

Panics

Panics if the parent window was already created – that is, you cannot dynamically create a DateTimePicker in an event closure.

pub fn new_dlg( parent: &impl GuiParent, ctrl_id: u16, resize_behavior: (Horz, Vert) ) -> Self

Instantiates a new DateTimePicker object, to be loaded from a dialog resource with HWND::GetDlgItem.

Panics

Panics if the parent dialog was already created – that is, you cannot dynamically create a DateTimePicker in an event closure.

pub fn date(&self) -> SYSTEMTIME

Retrieves the currently selected date by sending a dtm::GetSystemTime message.

pub fn set_date(&self, st: &SYSTEMTIME)

Sets the currently selected date by sending a dtm::SetSystemTime message.