winsafe::gui

Struct DateTimePickerOpts

pub struct DateTimePickerOpts {
    pub position: (i32, i32),
    pub width: i32,
    pub control_style: DTS,
    pub window_style: WS,
    pub window_ex_style: WS_EX,
    pub ctrl_id: u16,
    pub resize_behavior: (Horz, Vert),
    pub date: SYSTEMTIME,
}

Available on crate feature gui only.

Options to create a DateTimePicker programmatically with DateTimePicker::new.

Fields
position: (i32, i32)

Left and top position coordinates of control within parent’s client area, to be created.

Defaults to gui::dpi(0, 0).

width: i32

Control width to be created.

Defaults to ideal width retrieved with dtm::GetIdealSize message, usually around gui::dpi_x(250).

control_style: DTS

Date and time picker styles to be created.

Defaults to DTS::LONGDATEFORMAT.

window_style: WS

Window styles to be created.

Defaults to WS::CHILD | WS::GROUP | WS::TABSTOP | WS::VISIBLE.

window_ex_style: WS_EX

Extended window styles to be created.

Defaults to WS_EX::LEFT.

ctrl_id: u16

The control ID.

Defaults to an auto-generated ID.

resize_behavior: (Horz, Vert)

Horizontal and vertical behavior of the control when the parent window is resized.

Note: A DateTimePicker cannot be resized vertically, so it will panic if you use Vert::Resize.

Defaults to (gui::Horz::None, gui::Vert::None).

date: SYSTEMTIME

Initial date to be displayed.

Defaults to now.