Struct winsafe::gui::Edit

pub struct Edit(/* private fields */);

Available on crate feature gui only.

Native edit control.

Implementations

impl Edit

pub fn new(parent: &impl GuiParent, opts: EditOpts) -> Self

Instantiates a new Edit object, to be created on the parent window with HWND::CreateWindowEx.

Panics

Panics if the parent window was already created – that is, you cannot dynamically create an Edit in an event closure.

Examples

use winsafe::{self as w, prelude::*, gui};

let wnd: gui::WindowMain; // initialized somewhere

let txt = gui::Edit::new(
    &wnd,
    gui::EditOpts {
        position: (10, 10),
        width: 120,
        ..Default::default()
    },
);

pub fn new_dlg( parent: &impl GuiParent, ctrl_id: u16, resize_behavior: (Horz, Vert) ) -> Self

Instantiates a new Edit object, to be loaded from a dialog resource with HWND::GetDlgItem.

Panics

Panics if the parent dialog was already created – that is, you cannot dynamically create an Edit in an event closure.

pub fn hide_balloon_tip(&self)

Hides any balloon tip by sending an em::HideBalloonTip message.

pub fn iter_lines<'a>(&'a self) -> impl Iterator<Item = String> + 'a

Returns an iterator over the lines in the Edit.

Examples

use winsafe::{self as w, prelude::*, gui};

let my_edit: gui::Edit; // initialized somewhere

for line in my_edit.iter_lines() {
    println!("{}", line);
}

pub fn limit_text(&self, max_chars: Option<u32>)

Limits the number of characters that can be type by sending an em::SetLimitText message.

pub fn line_count(&self) -> u32

Returns the number of lines by sending an em::GetLineCount message.

pub fn set_font(&self, font: &HFONT)

Sets the font to the Edit by sending an wm::SetFont message.

Note that the font must remain alive while being used in the control.

pub fn set_selection(&self, start: i32, end: i32)

Sets the selection range of the text by sending an em::SetSel message.

Examples

Selecting all text in the control:

use winsafe::{self as w, prelude::*, gui};

let my_edit: gui::Edit; // initialized somewhere

my_edit.set_selection(0, -1);

Clearing the selection:

use winsafe::gui;

let my_edit: gui::Edit; // initialized somewhere

my_edit.set_selection(-1, -1);

pub fn show_ballon_tip(&self, title: &str, text: &str, icon: TTI)

Displays a balloon tip by sending an em::ShowBalloonTip message.