winsafe::gui
Struct Header

pub struct Header(/* private fields */);

Available on crate feature gui only.

Native header control.

Implementations

impl Header

pub fn new(parent: &impl GuiParent, opts: HeaderOpts) -> Self

Instantiates a new Header object, to be created on the parent window with HWND::CreateWindowEx.

Panics

Panics if the parent window was already created – that is, you cannot dynamically create a Header in an event closure.

pub fn new_dlg( parent: &impl GuiParent, ctrl_id: u16, resize_behavior: (Horz, Vert), ) -> Self

Instantiates a new Header object, to be loaded from a dialog resource with HWND::GetDlgItem.

Panics

Panics if the parent dialog was already created – that is, you cannot dynamically create a Header in an event closure.

pub fn from_list_view<T: 'static>(list_view: &ListView<T>) -> Self

Instantiates a new Header object to be loaded from a ListView control. This will give you access to the inner Header control of that ListView.

pub fn image_list(&self, kind: HDSIL) -> Option<&HIMAGELIST>

Retrieves a reference to one of the associated image lists by sending an hdm::GetImageList message.

The image list is owned by the control.

pub const fn items(&self) -> HeaderItems<'_>

Exposes the item methods.

pub fn set_image_list( &self, kind: HDSIL, himagelist: ImageListDestroyGuard, ) -> Option<ImageListDestroyGuard>

Sets the one of the associated image lists by sending an hdm::SetImageList message.

The image list will be owned by the control. Returns the previous one, if any.