Struct winsafe::gui::Label

pub struct Label(/* private fields */);

Available on crate feature gui only.

Native label control.

Implementations

impl Label

pub fn new(parent: &impl GuiParent, opts: LabelOpts) -> Self

Instantiates a new Label object, to be created on the parent window with HWND::CreateWindowEx.

Panics

Panics if the parent window was already created – that is, you cannot dynamically create a Label in an event closure.

pub fn new_dlg( parent: &impl GuiParent, ctrl_id: u16, resize_behavior: (Horz, Vert), ) -> Self

Instantiates a new CheckBox object, to be loaded from a dialog resource with HWND::GetDlgItem.

Panics

Panics if the parent dialog was already created – that is, you cannot dynamically create a Label in an event closure.

pub fn set_text_and_resize(&self, text: &str)

Calls set_text and resizes the control to exactly fit the new text.

Examples

use winsafe::{self as w, prelude::*, gui};

let my_label: gui::Label; // initialized somewhere

my_label.set_text_and_resize("This my text");