Struct winsafe::gui::LabelOpts

pub struct LabelOpts {
    pub text: String,
    pub position: POINT,
    pub size: SIZE,
    pub label_style: SS,
    pub window_style: WS,
    pub window_ex_style: WS_EX,
    pub ctrl_id: u16,
    pub horz_resize: Horz,
    pub vert_resize: Vert,
}

Available on crate feature gui only.

Options to create a Label programmatically with label::new.

Fields
text: String

Text of the control to be created.

Defaults to empty string.

position: POINT

Control position within parent client area, to be created.

If the parent window is a dialog, the values are in Dialog Template Units; otherwise in pixels, which will be multiplied to match current system DPI.

Defaults to 0 x 0.

size: SIZE

Control size, to be created.

If the parent window is a dialog, the values are in Dialog Template Units; otherwise in pixels, which will be multiplied to match current system DPI.

Defaults to the size needed to fit the text.

label_style: SS

label styles to be created.

Defaults to SS::LEFT | SS:NOTIFY.

window_style: WS

Window styles to be created.

Defaults to WS::CHILD | WS::VISIBLE.

window_ex_style: WS_EX

Extended window styles to be created.

Defaults to WS_EX::LEFT.

ctrl_id: u16

The control ID.

Defaults to an auto-generated ID.

horz_resize: Horz

Horizontal behavior when the parent is resized.

Defaults to Horz::None.

vert_resize: Vert

Vertical behavior when the parent is resized.

Defaults to Vert::None.