Struct winsafe::gui::ListView

pub struct ListView<T: 'static = ()>(/* private fields */);

Available on crate feature gui only.

Native list view control. Not to be confused with the simpler list box control.

The generic parameter specifies the type of the object that will be embedded on each item – if you don’t want to store anything, just use () as the type. Internally, this storage is implemented with pointers in the item’s LPARAM fields.

You can have access to the internal header of the list view by creating a Header object.

Implementations

impl<T> ListView<T>

pub fn new(parent: &impl GuiParent, opts: ListViewOpts) -> Self

Instantiates a new ListView object, to be created on the parent window with HWND::CreateWindowEx.

Panics

Panics if the parent window was already created – that is, you cannot dynamically create a ListView in an event closure.

pub fn new_dlg( parent: &impl GuiParent, ctrl_id: u16, resize_behavior: (Horz, Vert), context_menu_id: Option<u16> ) -> Self

Instantiates a new ListView object, to be loaded from a dialog resource with HWND::GetDlgItem.

Panics

Panics if the parent dialog was already created – that is, you cannot dynamically create a ListView in an event closure.

Panics if the context menu, when specified, does not exist.

pub const fn columns(&self) -> ListViewColumns<'_, T>

Exposes the column methods.

pub fn context_menu(&self) -> Option<&HMENU>

Returns the context menu attached to this list view, if any.

The context menu is attached when the list view is created, either by calling ListView::new or ListView::new_dlg.

pub fn image_list(&self, kind: LVSIL) -> Option<HIMAGELIST>

Retrieves one of the associated image lists by sending an lvm::GetImageList message.

pub const fn items(&self) -> ListViewItems<'_, T>

Exposes the item methods.

pub fn current_view(&self) -> LV_VIEW

Retrieves the current view by sending an lvm::GetView message.

pub fn set_current_view(&self, view: LV_VIEW)

Sets the current view by sending an lvm::SetView message.

pub fn set_extended_style(&self, set: bool, ex_style: LVS_EX)

Sets or unsets the given extended list view styles by sending an lvm::SetExtendedListViewStyle message.

pub fn set_image_list( &self, kind: LVSIL, himagelist: &HIMAGELIST ) -> Option<HIMAGELIST>

Sets the one of the associated image lists by sending an lvm::SetImageList message.

Returns the previous image list, if any.

pub fn set_redraw(&self, can_redraw: bool)

Allows or disallows the redrawing of the control by sending a wm::SetRedraw message.