Struct winsafe::gui::ListViewColumns

pub struct ListViewColumns { /* fields omitted */ }

Exposes column methods of a ListView control.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl ListViewColumns

pub fn add(&self, texts_and_widths: &[(&str, u32)]) -> WinResult<()>

Adds many columns at once by sending an LVM_INSERTCOLUMN message.

Widths will be adjusted to match current system DPI.

Examples

use winsafe::gui;

let my_list: gui::ListView; // initialized somewhere

my_list.columns().add(&[
    ("Name", 300),
    ("Address", 500),
]).unwrap();

pub fn all_texts(&self, column_index: u32) -> Vec<String>

Retrieves the texts of all items at the given column.

pub fn count(&self) -> WinResult<u32>

Retrieves the number of columns by sending an HDM_GETITEMCOUNT message to the handle returned by LVM_GETHEADER.

pub fn info(&self, column_index: u32, lvc: &mut LVCOLUMN<'_>) -> WinResult<()>

Retrieves information about the column by sending an LVM_GETCOLUMN message.

pub fn selected_texts(&self, column_index: u32) -> Vec<String>

Retrieves the texts of the selected items at the given column.

pub fn set_info(&self, column_index: u32, lvc: &LVCOLUMN<'_>) -> WinResult<()>

Sets information of the column by sending an LVM_SETCOLUMN message.

pub fn set_title(&self, column_index: u32, text: &str) -> WinResult<()>

Sets the title of the column by calling set_info.

pub fn set_width(&self, column_index: u32, width: u32) -> WinResult<()>

Sets the width of the column by sending an LVM_SETCOLUMNWIDTH message.

Width will be adjusted to match current system DPI.

pub fn set_width_to_fill(&self, column_index: u32) -> WinResult<()>

Sets the width of the column by sending an LVM_SETCOLUMNWIDTH message. The width will be calculated to fill the remaining space.

pub fn title(&self, column_index: u32, buf: &mut WString) -> WinResult<String>

Retrieves the title of the column by calling info.

The passed buffer will be automatically allocated.

This method can be more performant than title_str because the buffer can be reused, avoiding multiple allocations. However, it has the inconvenient of the manual conversion from WString to String.

pub fn title_str(&self, column_index: u32) -> WinResult<String>

A more convenient text, which directly returns a String instead of requiring an external buffer.

pub fn width(&self, column_index: u32) -> WinResult<u32>

Retrieves the width of the column by sending an LVM_GETCOLUMNWIDTH message.