winsafe::gui

Struct ListViewItem

pub struct ListViewItem<'a, T: 'static = ()> { /* private fields */ }

Available on crate feature gui only.

A single item of a ListView control.

Note: Each object keeps the zero-based index of an item. If new items are added/removed from the list view control, the object may then point to a different item.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl<'a, T> ListViewItem<'a, T>

pub fn data(&self) -> Rc<RefCell<T>>

Returns a Rc/RefCell with the stored data by sending an lvm::GetItem message.

Panics

Panics if the ListView doesnโt have an actual type, that is, if it was declared as ListView<()>.

Panics if the item index is invalid.

Examples

use winsafe::{self as w, prelude::*, gui};

struct Person {
    name: String,
    age: u32,
}

let my_list: gui::ListView<Person>; // initialized somewhere

// Create item with data.
let person = Person { name: "".to_owned(), age: 50 };
my_list.items().add(&["foo"], None, person)?;

// Retrieve and modify item data.
let rc_person = my_list.items().get(0).data();
rc_person.try_borrow_mut()?.age = 60;

pub fn delete(&self) -> SysResult<()>

Deletes the item by sending an lvm::DeleteItem message.

pub fn ensure_visible(&self) -> SysResult<Self>

Scrolls the list by sending an lvm::EnsureVisible message so that the item is visible in the list.

Returns the same item, so further operations can be chained.

pub fn focus(&self) -> SysResult<Self>

Sets the item as the focused one sending an lvm:SetItemState message.

Returns the same item, so further operations can be chained.

pub fn icon_index(&self) -> SysResult<Option<u32>>

Retrieves the icon index of the item by sending an lvm::GetItem message.

pub const fn index(&self) -> u32

Returns the zero-based index of the item.

pub fn is_focused(&self) -> bool

Tells if the item is the focused one by sending an lvm::GetItemState message.

pub fn is_selected(&self) -> bool

Tells if the item is selected by sending an lvm::GetItemState message.

pub fn is_visible(&self) -> bool

Tells if the item is currently visible by sending an lvm::IsItemVisible message.

pub fn rect(&self, portion: LVIR) -> SysResult<RECT>

Retrieves the bounding rectangle of the item by sending an lvm::GetItemRect message.

pub fn select(&self, set: bool) -> SysResult<Self>

Sets or removes the selection from the item by sending an lvm::SetItemState message.

Returns the same item, so further operations can be chained.

pub fn set_icon_index(&self, icon_index: Option<u32>) -> SysResult<Self>

Sets the icon index of the item by sending an lvm::SetItem message.

Returns the same item, so further operations can be chained.

pub fn set_text(&self, column_index: u32, text: &str) -> SysResult<Self>

Sets the text of the item under a column by sending an lvm::SetItemText message.

Returns the same item, so further operations can be chained.

pub fn text(&self, column_index: u32) -> String

Retrieves the text of an item under a column by sending an lvm::GetItemText message.

pub fn uid(&self) -> Option<u32>

Retrieves the unique ID for the item index by sending an lvm::MapIndexToId message.

If the item index has became invalid, returns None.