Struct winsafe::gui::ListViewItems

pub struct ListViewItems { /* fields omitted */ }

Exposes item methods of a ListView control.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl ListViewItems

pub fn add<S: AsRef<str>>(
&self,
texts: &[S],
icon_index: Option<u32>
) -> WinResult<u32>

Appends a new item by sending an LVM_INSERTITEM message, and returns its index.

The texts are relative to each column.

Examples

use winsafe::gui;

let my_list: gui::ListView; // initialized somewhere

my_list.items().add(
    &[
        "First column text",
        "Second column text",
    ],
    None, // no icon; requires set_image_list() before
).unwrap();

Panics

Panics if texts is empty, or if the number of texts is greater than the number of columns.

pub fn count(&self) -> u32

Retrieves the total number of items by sending an LVM_GETITEMCOUNT message.

pub fn delete(&self, item_indexes: &[u32]) -> WinResult<()>

Deletes the items at the given indexes by sending an LVM_DELETEITEM message.

The indexes are iterated backwards, so the last item will be deleted first.

pub fn delete_all(&self) -> WinResult<()>

Deletes all items by sending an LVM_DELETEALLITEMS message.

pub fn delete_selected(&self) -> WinResult<()>

Deletes the selected items by sending LVM_DELETEITEM messages.

pub fn ensure_visible(&self, item_index: u32) -> WinResult<()>

Scrolls the list by sending an LVM_ENSUREVISIBLE message so that an item is visible in the list.

pub fn find(&self, text: &str) -> Option<u32>

Searches for an item with the given text, case-insensitive, by sending an LVM_FINDITEM message.

pub fn focused(&self) -> Option<u32>

Retrieves the index of the focused item by sending an LVM_GETNEXTITEM message.

pub fn hit_test(&self, info: &mut LVHITTESTINFO) -> Option<u32>

Retrieves the item at the specified position by sending an LVM_HITTEST message

pub fn is_focused(&self, item_index: u32) -> bool

Tells if the item is the focused one by sending an LVM_GETITEMSTATE message.

pub fn is_selected(&self, item_index: u32) -> bool

Tells if the item is selected by sending an LVM_GETITEMSTATE message.

pub fn is_visible(&self, item_index: u32) -> bool

Tells if the item is currently visible by sending an LVM_ISITEMVISIBLE message.

pub fn map_id_to_index(&self, item_id: u32) -> Option<u32>

Retrieves the actual index of the unique ID by sending an LVM_MAPIDTOINDEX message.

pub fn map_index_to_id(&self, item_index: u32) -> Option<u32>

Retrieves an unique ID for the given index by sending an LVM_MAPINDEXTOID message.

pub fn rect(&self, item_index: u32, portion: LVIR) -> WinResult<RECT>

Retrieves the bound rectangle of item by sending an LVM_GETITEMRECT message.

pub fn selected(&self) -> Vec<u32>

Retrieves the indexes of the selected items by sending LVM_GETNEXTITEM messages.

pub fn selected_count(&self) -> u32

Retrieves the number of selected items by sending an LVM_GETSELECTEDCOUNT message.

pub fn set_focused(&self, item_index: u32) -> WinResult<()>

Sets the focused item by sending an LVM_SETITEMSTATE message.

pub fn set_selected(&self, set: bool, item_indexes: &[u32]) -> WinResult<()>

Sets or remove the selection from the given item indexes by sending LVM_SETITEMSTATE messages.

pub fn set_selected_all(&self, set: bool) -> WinResult<()>

Sets or remove the selection for all items by sending an LVM_SETITEMSTATE message.

pub fn set_text(
&self,
item_index: u32,
column_index: u32,
text: &str
) -> WinResult<()>

Sets the text of an item under a column by sending an LVM_SETITEMTEXT message.

pub fn text(&self, item_index: u32, column_index: u32, buf: &mut WString)

Retrieves the text of an item under a column by sending an LVM_GETITEMTEXT message.

The passed buffer will be automatically allocated.

This method can be more performant than text_str because the buffer can be reused, avoiding multiple allocations. However, it has the inconvenient of the manual conversion from WString to String.

Examples

use winsafe::{gui, WString};

let my_list: gui::ListView; // initialized somewhere

let mut buf = WString::default();
my_list.items().text(0, 2, &mut buf); // 1st item, 3rd column

println!("Text: {}", buf.to_string());

pub fn text_str(&self, item_index: u32, column_index: u32) -> String

A more convenient text, which directly returns a String instead of requiring an external buffer.

Examples

use winsafe::gui;

let my_list: gui::ListView; // initialized somewhere

println!("Text: {}", my_list.items().text(0, 2)); // 1st item, 3rd column