Struct winsafe::gui::ListViewOpts

pub struct ListViewOpts {
    pub position: POINT,
    pub size: SIZE,
    pub list_view_style: LVS,
    pub list_view_ex_style: LVS_EX,
    pub window_style: WS,
    pub window_ex_style: WS_EX,
    pub ctrl_id: u16,
    pub horz_resize: Horz,
    pub vert_resize: Vert,
    pub context_menu: Option<HMENU>,
    pub columns: Vec<(String, u32)>,
}

Available on crate feature gui only.

Options to create a ListView programmatically with ListView::new.

Fields
position: POINT

Control position within parent client area, to be created.

If the parent window is a dialog, the values are in Dialog Template Units; otherwise in pixels, which will be multiplied to match current system DPI.

Defaults to 0 x 0.

size: SIZE

Control size, to be created.

If the parent window is a dialog, the values are in Dialog Template Units; otherwise in pixels, which will be multiplied to match current system DPI.

Defaults to 50 x 50.

list_view_style: LVS

List view styles to be created.

Defaults to LVS::REPORT | LVS::NOSORTHEADER | LVS::SHOWSELALWAYS | LVS::SHAREIMAGELISTS.

list_view_ex_style: LVS_EX

Extended list view styles to be created.

Defaults to LVS_EX::NoValue.

window_style: WS

Window styles to be created.

Defaults to WS::CHILD | WS::VISIBLE | WS::TABSTOP | WS::GROUP.

window_ex_style: WS_EX

Extended window styles to be created.

Defaults to WS_EX::LEFT | WS_EX::CLIENTEDGE.

ctrl_id: u16

The control ID.

Defaults to an auto-generated ID.

horz_resize: Horz

Horizontal behavior when the parent is resized.

Defaults to Horz::None.

vert_resize: Vert

Vertical behavior when the parent is resized.

Defaults to Vert::None.

context_menu: Option<HMENU>

Context popup menu.

This menu is shared: it must be destroyed manually after the control is destroyed. But note that menus loaded from resources don’t need to be destroyed.

Defaults to None.

columns: Vec<(String, u32)>

Text and width of columns to be added right away. The columns only show in report mode.

Defaults to none.