winsafe::gui

Struct ProgressBar

pub struct ProgressBar(/* private fields */);

Available on crate feature gui only.

Native progress bar control.

Implementations

impl ProgressBar

pub fn new(parent: &impl GuiParent, opts: ProgressBarOpts) -> Self

Instantiates a new ProgressBar object, to be created on the parent window with HWND::CreateWindowEx.

Panics

Panics if the parent window was already created – that is, you cannot dynamically create a ProgressBar in an event closure.

pub fn new_dlg( parent: &impl GuiParent, ctrl_id: u16, resize_behavior: (Horz, Vert), ) -> Self

Instantiates a new ProgressBar object, to be loaded from a dialog resource with HWND::GetDlgItem.

Panics

Panics if the parent dialog was already created – that is, you cannot dynamically create a ProgressBar in an event closure.

pub fn position(&self) -> u32

Retrieves the current position by sending a pbm::GetPos message.

pub fn range(&self) -> (u32, u32)

Retrieves the current minimum and maximum values by sending a pbm::GetRange message. Default values are 0 and 100.

pub fn set_marquee(&self, marquee: bool)

Sets or unsets the marquee mode by sending a pbm::SetMarquee message combined with a SetWindowLongPtr call for a style change.

pub fn set_position(&self, position: u32) -> u32

Sets the current position by sending a pbm::SetPos message, returning the previous position.

pub fn set_range(&self, min: u32, max: u32)

Sets the minimum and maximum values by sending a pbm::SetRange32 message. Default values are 0 and 100.

pub fn set_state(&self, state: PBST) -> PBST

Sets the current state by sending a pbm::SetState message, retuning the previous state.

pub fn state(&self) -> PBST

Retrieves the current state by sending a pbm::GetState message.