winsafe::gui

Struct PropSheet

pub struct PropSheet(/* private fields */);

Available on crate feature gui only.

Native property sheet control.

Examples

The example below creates a property sheet with one PropSheetPage page.

use winsafe::{self as w, co, gui, prelude::*, seq_ids};

seq_ids! {
    DLG_PAGE_1 = 100; // must match the resource dialog
}

fn main() {
    let ps = gui::PropSheet::new(gui::PropSheetOpts {
        title: "Main title",
        pages: &[
            ("First page", FirstPage::new().into()),
        ],
        ..Default::default()
    });

    // In a real application, the parent window is passed to show().
    if let Err(e) = ps.show(None::<&gui::WindowMain>) {
        eprintln!("{}", e);
    }
}

#[derive(Clone)]
struct FirstPage {
    wnd: gui::PropSheetPage,
}
impl Into<gui::PropSheetPage> for FirstPage {
    fn into(self) -> gui::PropSheetPage {
        self.wnd.clone() // so we can pass our custom struct to PropSheet::new()
    }
}
impl FirstPage {
    fn new() -> Self {
        let wnd = gui::PropSheetPage::new_dlg(DLG_PAGE_1);
        let new_self = Self { wnd };
        new_self.events();
        new_self
    }
    fn events(&self) {
        let self2 = self.clone();
        self.wnd.on().psn_apply(move || {
            // When the user clicks OK, each page will receive this notification.
            println!("OK from the first page");
            Ok(co::PSNRET::NOERROR)
        });
    }
}

Implementations

impl PropSheet

pub fn new(opts: PropSheetOpts<'_>) -> Self

Instantiates a new PropSheet object, to be created with PropertySheet.

Panics

Panics if PropSheetOpts has no pages.

pub fn show(&self, parent: Option<&impl GuiParent>) -> AnyResult<()>

Physically creates the property sheet window. This method will block until the window is closed.

Panics

Panics if the window is already created.