Struct winsafe::gui::RadioButton

pub struct RadioButton { /* private fields */ }

Available on crate feature gui only.

Native radio button control.

You cannot directly instantiate this object, you must use RadioGroup.

Implementations

impl RadioButton

pub fn emulate_click(&self)

Emulates the click event for the radio button by sending a bm::Click message.

pub fn is_selected(&self) -> bool

Tells if this radio button is the currently selected one by sending a bm::GetCheck message.

pub fn select(&self, selected: bool)

Sets the this radio button as the currently selected one by sending a bm::SetCheck message.

pub fn select_and_trigger(&self, selected: bool) -> SysResult<()>

Sets the this radio button as the currently selected one by sending a bm::SetCheck message, then sends a wm::Command message to the parent, so it can handle the event.

pub fn set_text_and_resize(&self, text: &str)

Calls set_text and resizes the control to exactly fit the new text.