Struct winsafe::gui::RadioGroup

pub struct RadioGroup(/* private fields */);

Available on crate feature gui only.

A group of native RadioButton controls.

Implementations

impl RadioGroup

pub fn new(parent: &impl GuiParent, opts: &[RadioButtonOpts]) -> Self

Instantiates a new RadioGroup object, each RadioButton to be created on the parent window with HWND::CreateWindowEx.

Panics

Panics if opts is empty.

Panics if the parent window was already created – that is, you cannot dynamically create a RadioGroup in an event closure.

pub fn new_dlg(parent: &impl GuiParent, ctrls: &[(u16, Horz, Vert)]) -> Self

Instantiates a new RadioGroup object, to be loaded from a dialog resource with HWND::GetDlgItem.

Panics

Panics if ctrls is empty.

Panics if the parent dialog was already created – that is, you cannot dynamically create a RadioGroup in an event closure.

pub fn iter(&self) -> Iter<'_, RadioButton>

Returns an iterator over the internal RadioButton slice.

Example

Changing the text of all radio buttons to "One":

use winsafe::{self as w, prelude::*, gui};

let radio_group: gui::RadioGroup; // initialized somewhere

for single_radio in radio_group.iter() {
    single_radio.hwnd().SetWindowText("One");
}

pub fn checked(&self) -> Option<&RadioButton>

Returns the currently checked RadioButton of this group, if any.

pub fn checked_index(&self) -> Option<usize>

Returns the index of the currently selected RadioButton of this group, if any.

pub fn count(&self) -> usize

Returns the number of RadioButton controls in this group.