Struct winsafe::gui::StatusBarParts

pub struct StatusBarParts { /* fields omitted */ }

Exposes the part methods of a StatusBar control.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl StatusBarParts

pub fn count(&self) -> u8

Retrieves the number of parts by sending an SB_GETPARTS message.

pub fn set_text(&self, part_index: u8, text: &str) -> WinResult<()>

Sets the text of a part by sending an SB_SETTEXT message.

pub fn text(&self, part_index: u8, buf: &mut WString)

Retrieves the text of the item by sending a SB_GETTEXT message.

The passed buffer will be automatically allocated.

This method can be more performant than text_str because the buffer can be reused, avoiding multiple allocations. However, it has the inconvenient of the manual conversion from WString to String.

Examples

use winsafe::{gui, WString};

let my_sb: gui::StatusBar; // initialized somewhere

let mut buf = WString::default();
my_sb.parts().text(0, &mut buf).unwrap();

println!("Text: {}", buf.to_string());

pub fn text_str(&self, part_index: u8) -> String

A more convenient text, which directly returns a String instead of requiring an external buffer.

Examples

use winsafe::gui;

let my_sb: gui::StatusBar; // initialized somewhere

println!("Text: {}", my_sb.parts().text(0).unwrap());