winsafe::gui

Struct Tab

pub struct Tab(/* private fields */);

Available on crate feature gui only.

Native tab control.

Implementations

impl Tab

pub fn on_subclass(&self) -> &WindowEvents

Exposes the subclass events. If at least one event exists, the control will be subclassed.

Note: Subclassing may impact performance, use with care.

Panics

Panics if the control or the parent window are already created. Events must be set before control and parent window creation.

pub fn on(&self) -> &TabEvents

Exposes the specific control events.

Panics

Panics if the control is already created. Events must be set before control creation.

impl Tab

pub fn new(parent: &(impl GuiParent + 'static), opts: TabOpts) -> Self

Instantiates a new Tab object, to be created on the parent window with HWND::CreateWindowEx.

Panics

Panics if the parent window was already created – that is, you cannot dynamically create a Tab in an event closure.

pub fn new_dlg( parent: &(impl GuiParent + 'static), ctrl_id: u16, resize_behavior: (Horz, Vert), items: Vec<(String, Box<dyn AsRef<WindowControl>>)>, ) -> Self

Instantiates a new Tab object, to be loaded from a dialog resource with HWND::GetDlgItem.

Panics

Panics if the parent dialog was already created – that is, you cannot dynamically create a Tab in an event closure.

pub fn image_list(&self) -> Option<&HIMAGELIST>

Retrieves a reference to the associated image list by sending a tcm::GetImageList message.

The image list is owned by the control.

pub const fn items(&self) -> TabItems<'_>

Item methods.

pub fn set_extended_style(&self, set: bool, ex_style: TCS_EX)

Sets or unsets the given extended list view styles by sending a tcm::SetExtendedStyle message.

pub fn set_image_list( &self, himagelist: ImageListDestroyGuard, ) -> Option<ImageListDestroyGuard>

Sets the associated image list by sending a tcm::SetImageList message.

The image list will be owned by the control. Returns the previous one, if any.