Struct winsafe::gui::UpDownOpts

pub struct UpDownOpts {
    pub position: (i32, i32),
    pub height: u32,
    pub up_down_style: UDS,
    pub window_style: WS,
    pub window_ex_style: WS_EX,
    pub ctrl_id: u16,
    pub range: (i32, i32),
}

Available on crate feature gui only.

Options to create an UpDown programmatically with UpDown::new.

Fields
position: (i32, i32)

Left and top position coordinates of control within parent’s client area, to be created.

If the parent window is a dialog, the values are in Dialog Template Units; otherwise in pixels, which will be multiplied to match current system DPI.

Note that the UDS::AUTOBUDDY style automatically positions the UpDown; thus, with this style, position is meaningless.

Defaults to (0, 0).

height: u32

Control height to be created.

If the parent window is a dialog, the value is in Dialog Template Units; otherwise in pixels, which will be multiplied to match current system DPI.

Note that the UDS::AUTOBUDDY style automatically resizes the UpDown; thus, with this style, height is meaningless.

Defaults to 40.

up_down_style: UDS

Up-down styles to be created.

Note that the UDS::AUTOBUDDY style will take the control created immediately before the UpDown as the buddy one, attaching the UpDown to it. This control should be an Edit with ES::NUMBER style.

Defaults to UDS::AUTOBUDDY | UDS::SETBUDDYINT | UDS::ALIGNRIGHT | UDS::ARROWKEYS | UDS::HOTTRACK.

window_style: WS

Window styles to be created.

Defaults to WS::CHILDWINDOW | WS::VISIBLE.

window_ex_style: WS_EX

Extended window styles to be created.

Defaults to WS_EX::LEFT.

ctrl_id: u16

The control ID.

Defaults to an auto-generated ID.

range: (i32, i32)

The minimum and maximum position values.

Defaults to (0, 100).