Struct winsafe::gui::WindowControl

pub struct WindowControl { /* private fields */ }

This is supported on crate feature gui only.

An user child window, which can handle events. Can be programmatically created or load a dialog resource from a .res file.

Implementations

impl WindowControl

pub fn new(parent: &impl GuiParent, opts: WindowControlOpts) -> WindowControl

Instantiates a new WindowControl object, to be created with HWND::CreateWindowEx.

pub fn new_dlg(
parent: &impl GuiParent,
dialog_id: u16,
position: POINT,
resize_behavior: (Horz, Vert),
ctrl_id: Option<u16>
) -> WindowControl

Instantiates a new WindowControl object, to be loaded from a dialog resource with HWND::GetDlgItem.

If the parent window is a dialog, position is in Dialog Template Units; otherwise in pixels, which will be multiplied to match current system DPI.