Struct winsafe::gui::WindowMain

pub struct WindowMain(_);

Available on crate feature gui only.

An user main window, which can handle events. Usually, this is the first window of your application, launched directly from the main function. Can be programmatically created or load a dialog resource from a .res file.

Implementations

impl WindowMain

pub fn new(opts: WindowMainOpts) -> WindowMain

Instantiates a new WindowMain object, to be created internally with HWND::CreateWindowEx.

pub fn new_dlg(
dialog_id: u16,
icon_id: Option<u16>,
accel_table_id: Option<u16>
) -> WindowMain

Instantiates a new WindowMain object, to be loaded from a dialog resource with HWND::GetDlgItem.

pub fn run_main(&self, cmd_show: Option<SW>) -> AnyResult<i32>

Physically creates the window, then runs the main application loop. This method will block until the window is closed.

The cmd_show parameter defaults to co::SW::SHOW.

Panics

Panics if the window is already created.