winsafe::gui

Struct WindowModal

pub struct WindowModal(/* private fields */);

Available on crate feature gui only.

An user modal window, which can handle events. Can be programmatically created or load a dialog resource from a .res file.

Implementations

impl WindowModal

pub fn new(opts: WindowModalOpts<'_>) -> Self

Instantiates a new WindowModal object, to be created internally with HWND::CreateWindowEx.

pub fn new_dlg(dlg_id: u16) -> Self

Instantiates a new WindowModal object, to be loaded from a dialog resource with HINSTANCE::DialogBoxParam.

pub fn show_modal(&self, parent: &impl GuiParent) -> AnyResult<()>

Physically creates the window, then runs the modal loop. This method will block until the window is closed.

Note that, if the user clicks the โXโ to close the modal, the default behavior is to call EndDialog(0). To override this behavior, handle the modalโs wm_close yourself.

Panics

Panics if the window is already created.

Panics if the creation process fails.

pub fn on(&self) -> &impl GuiEventsParent

Exposes methods to handle window messages.

Panics

Panics if the window is already created. Events must be set before window creation.

pub fn close(&self)

Closes the window by posting a WM_CLOSE message. This is the safest way to close any popup window, because youโll able to process the wm_close event, just like if the user clicked the window โXโ button.