Trait winsafe::gui::Parent

pub trait Parent: Sealed {
    fn as_any(&self) -> &dyn Any;
}

Trait to any window which can host child controls.

Note: This is a sealed trait which cannot be implemented outside the library.

Required methods

fn as_any(&self) -> &dyn Any

Returns a reference to the Any trait, allowing downcasting.