winsafe

Macro bind_ig

macro_rules! bind_ig {
    ($arg:ident, $fun:expr) => { ... };
}

Available on crate feature gui only.

Binds, to an event closure, a function ignoring the event parameter.

This is just syntactic sugar to use a member function as an event closure.

Example

use winsafe::{self as w, prelude::*, gui, bind_ig};

#[derive(Clone)]
struct MyWindow {
    wnd: gui::WindowMain,
}

impl MyWindow {
    fn new() -> Self {
        let wnd = gui::WindowMain::new(gui::WindowMainOpts::default());
        let new_self = Self { wnd };
        new_self.events();
        new_self
    }

    fn events(&self) {
        self.wnd.on().wm_create(bind_ig!(self, Self::on_create));
    }

    fn on_create(&self) -> w::AnyResult<i32> {
        Ok(0)
    }
}

Related macros

bind

bind_p