winsafe

Module msg

Available on crate feature user only.

Parameters of window messages.

Wm is the basic message, with WPARAM and LPARAM fields. All messages can be sent, so they all implement the MsgSend trait. Some messages can also be received – that means you can handle them if you’re implementing a custom window from scratch –, and these implement the MsgSendRecv trait.

Sending messages

We want to delete the 3rd element of a ListView control. This can be done by sending it an LVM_DELETEITEM message via HWND::SendMessage. The message itself is a struct, which is initialized with the specific message parameters.

The message struct also defines the data type returned by SendMessage. In the example below, LVM_DELETEITEM returns SysResult<()>.

use winsafe::{self as w, prelude::*, msg};

let hlistview: w::HWND; // initialized somewhere

hlistview.SendMessage(
    msg::LvmDeleteItem {
        index: 2,
    },
).expect("Failed to delete item 2.");

Messages can be found in the msg module.

Custom messages

In order to create a custom message, you must create a struct with the data it contains (if any) and implement the MsgSend and MsgSendRecv traits:

use winsafe::{self as w, prelude::*, co, msg};

/// The integer value of our message ID.
pub const MAKE_TOAST: co::WM = unsafe { co::WM::from_raw(co::WM::USER.raw() + 20) };

/// Our message with its parameter.
struct MakeToast {
    how_many: u32,
}

impl MsgSend for MakeToast {
    type RetType = ();

    fn convert_ret(&self, _: isize) -> Self::RetType {
        ()
    }

    fn as_generic_wm(&mut self) -> msg::Wm {
        msg::Wm {
            msg_id: MAKE_TOAST,
            wparam: self.how_many as _,
            lparam: 0,
        }
    }
}

impl MsgSendRecv for MakeToast {
    fn from_generic_wm(p: msg::Wm) -> Self {
        Self {
            how_many: p.wparam as _,
        }
    }
}

Structs

BcmGetIdealSizecomctl

BCM_GETIDEALSIZE message parameters.

BcmGetImageListcomctl

BCM_GETIMAGELIST message parameters.

BcmGetNotecomctl

BCM_GETNOTE message parameters.

BcmGetNoteLengthcomctl

BCM_GETNOTELENGTH message, which has no parameters.

BcmGetSplitInfocomctl

BCM_GETSPLITINFO message parameters.

BcmGetTextMargincomctl

BCM_GETTEXTMARGIN message parameters.

BcmSetDropDownStatecomctl

BCM_SETDROPDOWNSTATE message parameters.

BcmSetImageListcomctl

BCM_SETIMAGELIST message parameters.

BcmSetNotecomctl

BCM_SETNOTE message parameters.

BcmSetShieldcomctl

BCM_SETSHIELD message parameters.

BcmSetSplitInfocomctl

BCM_SETSPLITINFO message parameters.

BcmSetTextMargincomctl

BCM_SETTEXTMARGIN message parameters.

BmClick

BM_CLICK message, which has no parameters.

BmGetCheck

BM_GETCHECK message parameters.

BmGetImage

BM_GETIMAGE message parameters.

BmGetState

BM_GETSTATE message, which has no parameters.

BmSetCheck

BM_SETCHECK message parameters.

BmSetDontClick

BM_SETDONTCLICK message parameters.

BmSetImage

BM_SETIMAGE message parameters.

BmSetState

BM_SETSTATE message parameters.

BmSetStyle

BM_SETSTYLE message parameters.

CbAddString

CB_ADDSTRING message parameters.

CbDeleteString

CB_DELETESTRING message parameters.

CbDir

CB_DIR message parameters.

CbFindString

CB_FINDSTRING message parameters.

CbFindStringExact

CB_FINDSTRINGEXACT message parameters.

CbGetComboBoxInfo

CB_GETCOMBOBOXINFO message parameters.

CbGetCount

CB_GETCOUNT message, which has no parameters.

CbGetCueBannercomctl

CB_GETCUEBANNER message parameters.

CbGetCurSel

CB_GETCURSEL message, which has no parameters.

CbGetDroppedControlRect

CB_GETDROPPEDCONTROLRECT message parameters.

CbGetDroppedState

CB_GETDROPPEDSTATE message, which has no parameters.

CbGetDroppedWidth

CB_GETDROPPEDWIDTH message, which has no parameters.

CbGetEditSel

CB_GETEDITSEL message, which has no parameters.

CbGetExtendedUi

CB_GETEXTENDEDUI message, which has no parameters.

CbGetHorizontalExtent

CB_GETHORIZONTALEXTENT message, which has no parameters.

CbGetItemData

CB_GETITEMDATA message parameters.

CbGetItemHeight

CB_GETITEMHEIGHT message parameters.

CbGetLbText

CB_GETLBTEXT message parameters.

CbGetLbTextLen

CB_GETLBTEXTLEN message parameters.

CbGetLocale

CB_GETLOCALE message, which has no parameters.

CbGetMinVisiblecomctl

CB_GETMINVISIBLE message, which has no parameters.

CbGetTopIndex

CB_GETTOPINDEX message, which has no parameters.

CbInitStorage

CB_INITSTORAGE message parameters.

CbInsertString

CB_INSERTSTRING message parameters.

CbLimitText

CB_LIMITTEXT message parameters.

CbResetContent

CB_RESETCONTENT message, which has no parameters.

CbSelectString

CB_SELECTSTRING message parameters.

CbSetCueBannercomctl

CB_SETCUEBANNER message parameters.

CbSetCurSel

CB_SETCURSEL message parameters.

CbSetDroppedWidth

CB_SETDROPPEDWIDTH message parameters.

CbSetEditSel

CB_SETEDITSEL message parameters.

CbSetExtendedUi

CB_SETEXTENDEDUI message parameters.

CbSetHorizontalExtent

CB_SETHORIZONTALEXTENT message parameters.

CbSetItemData

CB_SETITEMDATA message parameters.

CbSetItemHeight

CB_SETITEMHEIGHT message parameters.

CbSetLocale

CB_SETLOCALE message parameters.

CbSetMinVisiblecomctl

CB_SETMINVISIBLE message parameters.

CbSetTopIndex

CB_SETTOPINDEX message parameters.

CbShowDropDown

CB_SHOWDROPDOWN message parameters.

DtmCloseMonthCalcomctl

DTM_CLOSEMONTHCAL message, which has no parameters.

DtmGetDateTimePickerInfocomctl

DTM_GETDATETIMEPICKERINFO message parameters.

DtmGetIdealSizecomctl

DTM_GETIDEALSIZE message parameters.

DtmGetMcColorcomctl

DTM_GETMCCOLOR message parameters.

DtmGetMcFontcomctl and gdi

DTM_GETMCFONT message, which has no parameters.

DtmGetMcStylecomctl

DTM_GETMCSTYLE message, which has no parameters.

DtmGetMonthCalcomctl

DTM_GETMONTHCAL message, which has no parameters.

DtmGetRangecomctl

DTM_GETRANGE message parameters.

DtmGetSystemTimecomctl

DTM_GETSYSTEMTIME message parameters.

DtmSetFormatcomctl

DTM_SETFORMAT message parameters.

DtmSetMcColorcomctl

DTM_SETMCCOLOR message parameters.

DtmSetMcFontcomctl and gdi

DTM_SETMCFONT message parameters.

DtmSetMcStylecomctl

DTM_SETMCSTYLE message parameters.

DtmSetRangecomctl

DTM_SETRANGE message parameters.

DtmSetSystemTimecomctl

DTM_SETSYSTEMTIME message parameters.

EmCanUndo

EN_CANUNDO message, which has no parameters.

EmCharFromPos

EM_CHARFROMPOS message parameters.

EmEmptyUndoBuffer

EM_EMPTYUNDOBUFFER message, which has no parameters.

EmFmtLines

EM_FMTLINES message parameters.

EmGetCueBannercomctl

EM_GETCUEBANNER message parameters.

EmGetFirstVisibleLine

EM_GETFIRSTVISIBLELINE message, which has no parameters.

EmGetHandle

EM_GETHANDLE message, which has no parameters.

EmGetImeStatus

EM_GETIMESTATUS message, which has no parameters.

EmGetLimitText

EM_GETLIMITTEXT message, which has no parameters.

EmGetLine

EM_GETLINE message parameters.

EmGetLineCount

EM_GETLINECOUNT message, which has no parameters.

EmGetMargins

EM_GETMARGINS message, which has no parameters.

EmGetModify

EM_GETMODIFY message, which has no parameters.

EmGetPasswordChar

EM_GETPASSWORDCHAR message, which has no parameters.

EmGetRect

EM_GETRECT message parameters.

EmGetSel

EM_GETSEL message parameters.

EmGetThumb

EM_GETTHUMB message, which has no parameters.

EmGetWordBreakProc

EM_GETWORDBREAKPROC message, which has no parameters.

EmHideBalloonTipcomctl

EM_HIDEBALLOONTIP message, which has no parameters.

EmLimitText

EM_LIMITTEXT message parameters.

EmLineFromChar

EM_LINEFROMCHAR message parameters.

EmLineIndex

EM_LINEINDEX message parameters.

EmLineLength

EM_LINELENGTH message parameters.

EmLineScroll

EM_LINESCROLL message parameters.

EmPosFromChar

EM_POSFROMCHAR message parameters.

EmReplaceSel

EM_REPLACESEL message parameters.

EmScroll

EM_SCROLL message parameters.

EmScrollCaret

EM_SCROLLCARET message, which has no parameters.

EmSetCueBannercomctl

EM_SETCUEBANNER message parameters..

EmSetHandle

EM_SETHANDLE message parameters.

EmSetImeStatus

EM_SETIMESTATUS message parameters.

EmSetLimitText

EM_SETLIMITTEXT message parameters.

EmSetMargins

EM_SETMARGINS message parameters.

EmSetModify

EM_SETMODIFY message parameters.

EmSetPasswordChar

EM_SETPASSWORDCHAR message parameters.

EmSetReadOnly

EM_SETREADONLY message parameters.

EmSetRect

EM_SETRECT message parameters.

EmSetRectNp

EM_SETRECTNP message parameters.

EmSetSel

EM_SETSEL message parameters.

EmSetTabStops

EM_SETTABSTOPS message parameters.

EmSetWordBreakProc

EM_SETWORDBREAKPROC message parameters.

EmShowBalloonTipcomctl

EM_SHOWBALLOONTIP message parameters.

EmUndo

EM_UNDO message, which has no parameters.

HdmClearFiltercomctl

HDM_CLEARFILTER message parameters.

HdmCreateDragImagecomctl

HDM_CREATEDRAGIMAGE message parameters.

HdmDeleteItemcomctl

HDM_DELETEITEM message parameters.

HdmEditFiltercomctl

HDM_EDITFILTER message parameters.

HdmGetBitmapMargincomctl

HDM_GETBITMAPMARGIN message, which has no parameters.

HdmGetFocusedItemcomctl

HDM_GETFOCUSEDITEM message, which has no parameters.

HdmGetImageListcomctl

HDM_GETIMAGELIST message parameters.

HdmGetItemcomctl

HDM_GETITEM message parameters.

HdmGetItemCountcomctl

HDM_GETITEMCOUNT message, which has no parameters.

HdmGetItemDropDownRectcomctl

HDM_GETITEMDROPDOWNRECT message parameters.

HdmGetItemRectcomctl

HDM_GETITEMRECT message parameters.

HdmGetOrderArraycomctl

HDM_GETORDERARRAY message parameters.

HdmGetOverflowRectcomctl

HDM_GETOVERFLOWRECT message parameters.

HdmGetUnicodeFormatcomctl

HDM_GETUNICODEFORMAT message, which has no parameters.

HdmHitTestcomctl

HDM_HITTEST message parameters.

HdmInsertItemcomctl

HDM_INSERTITEM message parameters.

HdmLayoutcomctl

HDM_LAYOUT message parameters.

HdmOrderToIndexcomctl

HDM_ORDERTOINDEX message parameters.

HdmSetBitmapMargincomctl

HDM_SETBITMAPMARGIN message parameters.

HdmSetFilterChangeTimeoutcomctl

HDM_SETFILTERCHANGETIMEOUT message parameters.

HdmSetFocusedItemcomctl

HDM_SETFOCUSEDITEM message parameters.

HdmSetHotDividercomctl

HDM_SETHOTDIVIDER message parameters.

HdmSetImageListcomctl

HDM_SETIMAGELIST message parameters.

HdmSetItemcomctl

HDM_SETITEM message parameters.

HdmSetOrderArraycomctl

HDM_SETORDERARRAY message parameters.

HdmSetUnicodeFormatcomctl

HDM_SETUNICODEFORMAT message parameters.

LbAddFile

LB_ADDFILE message parameters.

LbAddString

LB_ADDSTRING message parameters.

LbDeleteString

LB_DELETESTRING message parameters.

LbDir

LB_DIR message parameters.

LbFindString

LB_FINDSTRING message parameters.

LbFindStringExact

LB_FINDSTRINGEXACT message parameters.

LbGetAnchorIndex

LB_GETANCHORINDEX message, which has no parameters.

LbGetCaretIndex

LB_GETCARETINDEX message, which has no parameters.

LbGetCount

LB_GETCOUNT message, which has no parameters.

LbGetCurSel

LB_GETCURSEL message, which has no parameters.

LbGetHorizontalExtent

LB_GETHORIZONTALEXTENT message, which has no parameters.

LbGetItemData

LB_GETITEMDATA message parameters.

LbGetItemHeight

LB_GETITEMHEIGHT message parameters.

LbGetItemRect

LB_GETITEMRECT message parameters.

LbGetListBoxInfo

LB_GETLISTBOXINFO message, which has no parameters.

LbGetLocale

LB_GETLOCALE message, which has no parameters.

LbGetSel

LB_GETSEL message parameters.

LbGetSelCount

LB_GETSELCOUNT message, which has no parameters.

LbGetSelItems

LB_GETSELITEMS message parameters.

LbGetText

LB_GETTEXT message parameters.

LbGetTextLen

LB_GETTEXTLEN message parameters.

LbGetTopIndex

LB_GETTOPINDEX message parameters.

LbInitStorage

LB_INITSTORAGE message parameters.

LbInsertString

LB_INSERTSTRING message parameters.

LbItemFromPoint

LB_ITEMFROMPOINT message parameters.

LbResetContent

LB_RESETCONTENT message, which has no parameters.

LbSelItemRange

LB_SELITEMRANGE message parameters.

LbSelItemRangeEx

LB_SELITEMRANGEEX message parameters.

LbSelectString

LB_SELECTSTRING message parameters.

LbSetAnchorIndex

LB_SETANCHORINDEX message parameters.

LbSetCaretIndex

LB_SETCARETINDEX message parameters.

LbSetColumnWidth

LB_SETCOLUMNWIDTH message parameters.

LbSetCount

LB_SETCOUNT message parameters.

LbSetCurSel

LB_SETCURSEL message parameters.

LbSetHorizontalExtent

LB_SETHORIZONTALEXTENT message parameters.

LbSetItemData

LB_SETITEMDATA message parameters.

LbSetItemHeight

LB_SETITEMHEIGHT message parameters.

LbSetLocale

LB_SETLOCALE message parameters.

LbSetSel

LB_SETSEL message parameters.

LbSetTabStops

LB_SETTABSTOPS message parameters.

LbSetTopIndex

LB_SETTOPINDEX message parameters.

LvmApproximateViewRectcomctl

LVM_APPROXIMATEVIEWRECT message parameters.

LvmArrangecomctl

LVM_ARRANGE message parameters.

LvmCancelEditLabelcomctl

LVM_CANCELEDITLABEL message, which has no parameters.

LvmCreateDragImagecomctl

LVM_CREATEDRAGIMAGE message parameters.

LvmDeleteAllItemscomctl

LVM_DELETEALLITEMS message, which has no parameters.

LvmDeleteColumncomctl

LVM_DELETECOLUMN message parameters.

LvmDeleteItemcomctl

LVM_DELETEITEM message parameters.

LvmEditLabelcomctl

LVM_EDITLABEL message parameters.

LvmEnableGroupViewcomctl

LVM_ENABLEGROUPVIEW message parameters.

LvmEnsureVisiblecomctl

LVM_ENSUREVISIBLE message parameters.

LvmFindItemcomctl

LVM_FINDITEM message parameters.

LvmGetBkColorcomctl

LVM_GETBKCOLOR message, which has no parameters.

LvmGetBkImagecomctl

LVM_GETBKIMAGE message parameters.

LvmGetCallbackMaskcomctl

LVM_GETCALLBACKMASK message, which has no parameters.

LvmGetColumncomctl

LVM_GETCOLUMN message parameters.

LvmGetColumnOrderArraycomctl

LVM_GETCOLUMNORDERARRAY message parameters.

LvmGetColumnWidthcomctl

LVM_GETCOLUMNWIDTH message parameters.

LvmGetCountPerPagecomctl

LVM_GETCOUNTPERPAGE message, which has no parameters.

LvmGetEditControlcomctl

LVM_EDITCONTROL message, which has no parameters.

LvmGetEmptyTextcomctl

LVM_GETEMPTYTEXT message parameters.

LvmGetExtendedListViewStylecomctl

LVM_GETEXTENDEDLISTVIEWSTYLE message, which has no parameters.

LvmGetFocusedGroupcomctl

LVM_GETFOCUSEDGROUP message, which has no parameters.

LvmGetFooterInfocomctl

LVM_GETFOOTERINFO message parameters.

LvmGetFooterItemcomctl

LVM_GETFOOTERITEM message parameters.

LvmGetFooterItemRectcomctl

LVM_GETFOOTERITEMRECT message parameters.

LvmGetFooterRectcomctl

LVM_GETFOOTERRECT message parameters.

LvmGetGroupCountcomctl

LVM_GROUPCOUNT message, which has no parameters.

LvmGetGroupInfocomctl

LVM_GROUPINFO message parameters.

LvmGetGroupInfoByIndexcomctl

LVM_GETGROUPINFOBYINDEX message parameters.

LvmGetGroupMetricscomctl

LVM_GETGROUPMETRICS message parameters.

LvmGetGroupRectcomctl

LVM_GETGROUPRECT message parameters.

LvmGetGroupStatecomctl

LVM_GETGROUPSTATE message parameters.

LvmGetHeadercomctl

LVM_GETHEADER message, which has no parameters.

LvmGetHotCursorcomctl

LVM_GETHOTCURSOR message, which has no parameters.

LvmGetHotItemcomctl

LVM_GETHOTITEM message, which has no parameters.

LvmGetHoverTimecomctl

LVM_GETHOVERTIME message, which has no parameters.

LvmGetISearchStringcomctl

LVM_GETISEARCHSTRING message parameters.

LvmGetImageListcomctl

LVM_GETIMAGELIST message parameters.

LvmGetInsertMarkcomctl

LVM_GETINSERTMARK message parameters.

LvmGetInsertMarkColorcomctl

LVM_GETINSERTMARKCOLOR message, which has no parameters.

LvmGetInsertMarkRectcomctl

LVM_GETINSERTMARKRECT message parameters.

LvmGetItemcomctl

LVM_GETITEM message parameters.

LvmGetItemCountcomctl

LVM_GETITEMCOUNT message, which has no parameters.

LvmGetItemIndexRectcomctl

LVM_GETITEMINDEXRECT message parameters.

LvmGetItemPositioncomctl

LVM_GETITEMPOSITION message parameters.

LvmGetItemRectcomctl

LVM_GETITEMRECT message parameters.

LvmGetItemSpacingcomctl

LVM_GETITEMSPACING message parameters.

LvmGetItemStatecomctl

LVM_GETITEMSTATE message parameters.

LvmGetItemTextcomctl

LVM_GETITEMTEXT message parameters.

LvmGetNextItemcomctl

LVM_GETNEXTITEM message parameters.

LvmGetNextItemIndexcomctl

LVM_GETNEXTITEMINDEX message parameters.

LvmGetNumberOfWorkAreascomctl

LVM_GETNUMBEROFWORKAREAS message parameters.

LvmGetOrigincomctl

LVM_GETORIGIN message parameters.

LvmGetOutlineColorcomctl

LVM_GETOUTLINECOLOR message, which has no parameters.

LvmGetSelectedColumncomctl

LVM_GETSELECTEDCOLUMN message, which has no parameters.

LvmGetSelectedCountcomctl

LVM_GETSELECTEDCOUNT message, which has no parameters.

LvmGetSelectionMarkcomctl

LVM_GETSELECTIONMARK message, which has no parameters.

LvmGetStringWidthcomctl

LVM_GETSTRINGWIDTH message parameters.

LvmGetSubItemRectcomctl

LVM_GETSUBITEMRECT message parameters.

LvmGetTextBkColorcomctl

LVM_GETTEXTBKCOLOR message, which has no parameters.

LvmGetTextColorcomctl

LVM_GETTEXTCOLOR message, which has no parameters.

LvmGetTileInfocomctl

LVM_GETTILEINFO message parameters.

LvmGetTileViewInfocomctl

LVM_GETTILEVIEWINFO message parameters.

LvmGetTooltipscomctl

LVM_GETTOOLTIPS message, which has no parameters.

LvmGetTopIndexcomctl

LVM_GETTOPINDEX message, which has no parameters.

LvmGetUnicodeFormatcomctl

LVM_GETUNICODEFORMAT message, which has no parameters.

LvmGetViewcomctl

LVM_GETVIEW message, which has no parameters.

LvmGetViewRectcomctl

LVM_GETVIEWRECT message parameters.

LvmGetWorkAreascomctl

LVM_GETWORKAREAS message parameters.

LvmHasGroupcomctl

LVM_HASGROUP message parameters.

LvmHitTestcomctl

LVM_HITTEST message parameters.

LvmInsertColumncomctl

LVM_INSERTCOLUMN message parameters.

LvmInsertGroupcomctl

LVM_INSERTGROUP message parameters.

LvmInsertGroupSortedcomctl

LVM_INSERTGROUPSORTED message parameters.

LvmInsertItemcomctl

LVM_INSERTITEM message parameters.

LvmInsertMarkHitTestcomctl

LVM_INSERTMARKHITTEST message parameters.

LvmIsGroupViewEnabledcomctl

LVM_ISGROUPVIEWENABLED message, which has no parameters.

LvmIsItemVisiblecomctl

LVM_ISITEMVISIBLE message parameters.

LvmMapIdToIndexcomctl

LVM_MAPIDTOINDEX message parameters.

LvmMapIndexToIdcomctl

LVM_MAPINDEXTOID message parameters.

LvmRedrawItemscomctl

LVM_REDRAWITEMS message parameters.

LvmRemoveAllGroupscomctl

LVM_REMOVEALLGROUPS message, which has no parameters.

LvmRemoveGroupcomctl

LVM_REMOVEGROUP message parameters.

LvmScrollcomctl

LVM_SCROLL message parameters.

LvmSetBkColorcomctl

LVM_SETBKCOLOR message parameters.

LvmSetBkImagecomctl

LVM_SETBKIMAGE message parameters.

LvmSetCallbackMaskcomctl

LVM_SETCALLBACKMASK message parameters.

LvmSetColumncomctl

LVM_SETCOLUMN message parameters.

LvmSetColumnOrderArraycomctl

LVM_SETCOLUMNORDERARRAY message parameters.

LvmSetColumnWidthcomctl

LVM_SETCOLUMNWIDTH message parameters.

LvmSetExtendedListViewStylecomctl

LVM_SETEXTENDEDLISTVIEWSTYLE message parameters.

LvmSetGroupInfocomctl

LVM_SETGROUPINFO message parameters.

LvmSetGroupMetricscomctl

LVM_SETGROUPMETRICS message parameters.

LvmSetHotCursorcomctl

LVM_SETHOTCURSOR message parameters.

LvmSetHotItemcomctl

LVM_SETHOTITEM message parameters.

LvmSetHoverTimecomctl

LVM_SETHOVERTIME message parameters.

LvmSetIconSpacingcomctl

LVM_SETICONSPACING message parameters.

LvmSetImageListcomctl

LVM_SETIMAGELIST message parameters.

LvmSetInfoTipcomctl

LVM_SETINFOTIP message parameters.

LvmSetInsertMarkcomctl

LVM_SETINSERTMARK message parameters.

LvmSetInsertMarkColorcomctl

LVM_SETINSERTMARKCOLOR message parameters.

LvmSetItemcomctl

LVM_SETITEM message parameters.

LvmSetItemCountcomctl

LVM_SETITEMCOUNT message parameters.

LvmSetItemIndexStatecomctl

LVM_SETITEMINDEXSTATE message parameters.

LvmSetItemPositioncomctl

LVM_SETITEMPOSITION message parameters.

LvmSetItemPosition32comctl

LVM_SETITEMPOSITION32 message parameters.

LvmSetItemStatecomctl

LVM_SETITEMSTATE message parameters.

LvmSetItemTextcomctl

LVM_SETITEMTEXT message parameters.

LvmSetOutlineColorcomctl

LVM_SETOUTLINECOLOR message parameters.

LvmSetSelectedColumncomctl

LVM_SETSELECTEDCOLUMN message parameters.

LvmSetSelectionMarkcomctl

LVM_SETSELECTIONMARK message parameters.

LvmSetTextBkColorcomctl

LVM_SETTEXTBKCOLOR message parameters.

LvmSetTextColorcomctl

LVM_SETTEXTCOLOR message parameters.

LvmSetTileInfocomctl

LVM_SETTILEINFO message parameters.

LvmSetTileViewInfocomctl

LVM_SETTILEVIEWINFO message parameters.

LvmSetTooltipscomctl

LVM_SETTOOLTIPS message parameters.

LvmSetUnicodeFormatcomctl

LVM_SETUNICODEFORMAT message parameters.

LvmSetViewcomctl

LVM_SETVIEW message parameters.

LvmSetWorkAreascomctl

LVM_SETWORKAREAS message parameters.

LvmSortGroupscomctl

LVM_SORTGROUPS message parameters.

LvmSortItemscomctl

LVM_SORTITEMS message parameters.

LvmSortItemsExcomctl

LVM_SORTITEMSEX message parameters.

LvmSubItemHitTestcomctl

LVM_SUBITEMHITTEST message parameters.

LvmUpdatecomctl

LVM_UPDATE message parameters.

McmGetCalIdcomctl

MCM_GETCALID message, which has no parameters.

McmGetCalendarBordercomctl

MCM_GETCALENDARBORDER message, which has no parameters.

McmGetCalendarCountcomctl

MCM_GETCALENDARCOUNT message, which has no parameters.

McmGetCalendarGridInfocomctl

MCM_GETCALENDARGRIDINFO message parameters.

McmGetColorcomctl

MCM_GETCOLOR message parameters.

McmGetCurSelcomctl

MCM_GETCURSEL message parameters.

McmGetCurrentViewcomctl

MCM_GETCURRENTVIEW message, which has no parameters.

McmGetFirstDayOfWeekcomctl

MCM_GETFIRSTDAYOFWEEK message, which has no parameters.

McmGetMaxSelCountcomctl

MCM_GETMAXSELCOUNT message, which has no parameters.

McmGetMaxTodayWidthcomctl

MCM_GETMAXTODAYWIDTH message, which has no parameters.

McmGetMinReqRectcomctl

MCM_GETMINREQRECT message parameters.

McmGetMonthDeltacomctl

MCM_GETMONTHDELTA message, which has no parameters.

McmGetMonthRangecomctl

MCM_GETMONTHRANGE message parameters.

McmGetRangecomctl

MCM_GETRANGE message parameters.

McmGetSelRangecomctl

MCM_GETSELRANGE message parameters.

McmGetTodaycomctl

MCM_GETTODAY message parameters.

McmGetUnicodeFormatcomctl

MCM_GETUNICODEFORMAT message, which has no parameters.

McmHitTestcomctl

MCM_HITTEST message parameters.

McmSetCalIdcomctl

MCM_SETCALID message parameters.

McmSetCalendarBordercomctl

MCM_SETCALENDARBORDER message parameters.

McmSetColorcomctl

MCM_SETCOLOR message parameters.

McmSetCurSelcomctl

MCM_SETCURSEL message parameters.

McmSetCurrentViewcomctl

MCM_SETCURRENTVIEW message parameters.

McmSetDayStatecomctl

MCM_SETDAYSTATE message parameters.

McmSetFirstDayOfWeekcomctl

MCM_SETFIRSTDAYOFWEEK message parameters.

McmSetMaxSelCountcomctl

MCM_SETMAXSELCOUNT message parameters.

McmSetMonthDeltacomctl

MCM_SETMONTHDELTA message parameters.

McmSetRangecomctl

MCM_SETRANGE message parameters.

McmSetSelRangecomctl

MCM_SETSELRANGE message parameters.

McmSetTodaycomctl

MCM_SETTODAY message parameters.

McmSetUnicodeFormatcomctl

MCM_SETUNICODEFORMAT message parameters.

McmSizeRectToMincomctl

MCM_SIZERECTTOMIN message parameters.

PbmDeltaPoscomctl

PBM_DELTAPOS message parameters.

PbmGetBarColorcomctl

PBM_GETBARCOLOR message, which has no parameters.

PbmGetBkColorcomctl

PBM_GETBKCOLOR message, which has no parameters.

PbmGetPoscomctl

PBM_GETPOS message, which has no parameters.

PbmGetRangecomctl

PBM_GETRANGE message parameters.

PbmGetStatecomctl

PBM_GETSTATE message, which has no parameters.

PbmGetStepcomctl

PBM_GETSTEP message, which has no parameters.

PbmSetBarColorcomctl

PBM_SETBARCOLOR message parameters.

PbmSetBkColorcomctl

PBM_SETBKCOLOR message parameters.

PbmSetMarqueecomctl

PBM_SETMARQUEE message parameters.

PbmSetPoscomctl

PBM_SETPOS message parameters.

PbmSetRangecomctl

PBM_SETRANGE message parameters.

PbmSetRange32comctl

PBM_SETRANGE32 message parameters.

PbmSetStatecomctl

PBM_SETSTATE message parameters.

PbmSetStepcomctl

PBM_SETSTEP message parameters.

PbmStepItcomctl

PBM_STEPIT message, which has no parameters.

PsmAddPagecomctl

PSM_ADDPAGE message parameters.

PsmApplycomctl

PSM_APPLY message, which has no parameters.

PsmCancelToClosecomctl

PSM_CANCELTOCLOSE message, which has no parameters.

PsmChangedcomctl

PSM_CHANGED message, which has no parameters.

PsmEnableWizButtonscomctl

PSM_ENABLEWIZBUTTONS message parameters.

PsmGetCurrentPageHwndcomctl

PSM_GETCURRENTPAGEHWND message, which has no parameters.

PsmGetResultcomctl

PSM_GETRESULT message, which has no parameters.

PsmGetTabControlcomctl

PSM_GETTABCONTROL message, which has no parameters.

PsmHwndToIndexcomctl

PSM_HWNDTOINDEX message parameters.

PsmIdToIndexcomctl

PSM_IDTOINDEX message parameters.

PsmIndexToHwndcomctl

PSM_INDEXTOHWND message parameters.

PsmIndexToIdcomctl

PSM_INDEXTOID message parameters.

PsmIndexToPagecomctl

PSM_INDEXTOPAGE message parameters.

PsmInsertPagecomctl

PSM_INSERTPAGE message parameters.

PsmIsDialogMessagecomctl

PSM_ISDIALOGMESSAGE message parameters.

PsmPageToIndexcomctl

PSM_PAGETOINDEX message parameters.

PsmPressButtoncomctl

PSM_PRESSBUTTON message parameters.

PsmQuerySiblingscomctl

PSM_QUERYSIBLINGS message parameters.

PsmRebootSystemcomctl

PSM_REBOOTSYSTEM message, which has no parameters.

PsmRecalcPageSizescomctl

PSM_RECALCPAGESIZES message, which has no parameters.

PsmRemovePagecomctl

PSM_REMOVEPAGE message parameters.

PsmRestartWindowscomctl

PSM_RESTARTWINDOWS message, which has no parameters.

PsmSetButtonTextcomctl

PSM_SETBUTTONTEXT message parameters.

PsmSetCurSelcomctl

PSM_SETCURSEL message parameters.

PsmSetCurSelIdcomctl

PSM_SETCURSELID message parameters.

PsmSetFinishTextcomctl

PSM_SETFINISHTEXT message parameters.

PsmSetHeaderSubtitlecomctl

PSM_SETHEADERSUBTITLE message parameters.

PsmSetHeaderTitlecomctl

PSM_SETHEADERTITLE message parameters.

PsmSetNextTextcomctl

PSM_SETNEXTTEXT message parameters.

PsmSetTitlecomctl

PSM_SETTITLE message parameters.

PsmSetWizButtonscomctl

PSM_SETWIZBUTTONS message parameters.

PsmShowWizButtonscomctl

PSM_SHOWWIZBUTTONS message parameters.

PsmUnchangedcomctl

PSM_UNCHANGED message parameters.

SbGetBorderscomctl

SB_GETBORDERS message parameters.

SbGetIconcomctl

SB_GETICON message parameters.

SbGetPartscomctl

SB_GETPARTS message parameters.

SbGetRectcomctl

SB_GETRECT message parameters.

SbGetTextcomctl

SB_GETTEXT message parameters.

SbGetTextLengthcomctl

SB_GETTEXTLENGTH message parameters.

SbGetTipTextcomctl

SB_GETTIPTEXT message parameters.

SbGetUnicodeFormatcomctl

SB_GETUNICODEFORMAT message, which has no parameters.

SbIsSimplecomctl

SB_ISSIMPLE message, which has no parameters.

SbSetBkColorcomctl

SB_SETBKCOLOR message parameters.

SbSetIconcomctl

SB_SETICON message parameters.

SbSetMinHeightcomctl

SB_SETMINHEIGHT message parameters.

SbSetPartscomctl

SB_SETPARTS message parameters.

SbSetTextcomctl

SB_SETTEXT message parameters.

SbSetTipTextcomctl

SB_SETTIPTEXT message parameters.

SbSetUnicodeFormatcomctl

SB_SETUNICODEFORMAT message parameters.

SbSimplecomctl

SB_SIMPLE message parameters.

StmGetIconcomctl

STM_GETICON message, which has no parameters.

StmGetImagecomctl

STM_GETIMAGE message parameters.

StmSetIconcomctl

STM_SETICON message parameters.

StmSetImagecomctl

STM_SETIMAGE message parameters.

TbAddBitmapcomctl

TB_ADDBITMAP message parameters.

TbAddButtonscomctl

TB_ADDBUTTONS message parameters.

TbAddStringcomctl

TB_ADDSTRING message parameters.

TbAutoSizecomctl

TB_AUTOSIZE message, which has no parameters.

TbButtonCountcomctl

TB_BUTTONCOUNT message, which has no parameters.

TbButtonStructSizecomctl

TB_BUTTONSTRUCTSIZE message parameters.

TbChangeBitmapcomctl

TB_CHANGEBITMAP message parameters.

TbCheckButtoncomctl

TB_CHECKBUTTON message parameters.

TbCommandToIndexcomctl

TB_COMMANDTOINDEX message parameters.

TbCustomizecomctl

TB_CUSTOMIZE message, which has no parameters.

TbDeleteButtoncomctl

TB_DELETEBUTTON message parameters.

TbEnableButtoncomctl

TB_ENABLEBUTTON message parameters.

TbGetAnchorHighlightcomctl

TB_GETANCHORHIGHLIGHT message, which has no parameters.

TbGetBitmapcomctl

TB_GETBITMAP message parameters.

TbGetBitmapFlagscomctl

TB_GETBITMAPFLAGS message, which has no parameters.

TbGetButtoncomctl

TB_GETBUTTON message parameters.

TbGetButtonInfocomctl

TB_GETBUTTONINFO message parameters.

TbGetButtonSizecomctl

TB_GETBUTTONSIZE message, which has no parameters.

TbGetButtonTextcomctl

TB_GETBUTTONTEXT message parameters.

TbGetColorSchemecomctl

TB_GETCOLORSCHEME message parameters.

TbGetDisabledImageListcomctl

TB_GETDISABLEDIMAGELIST message, which has no parameters.

TbGetExtendedStylecomctl

TB_GETEXTENDEDSTYLE message, which has no parameters.

TbGetHotImageListcomctl

TB_GETHOTIMAGELIST message, which has no parameters.

TbGetHotItemcomctl

TB_GETHOTITEM message, which has no parameters.

TbGetIdealSizecomctl

TB_GETIDEALSIZE message parameters.

TbGetImageListcomctl

TB_GETIMAGELIST message, which has no parameters.

TbGetImageListCountcomctl

TB_GETIMAGELISTCOUNT message, which has no parameters.

TbGetInsertMarkcomctl

TB_GETINSERTMARK message parameters.

TbGetInsertMarkColorcomctl

TB_GETINSERTMARKCOLOR message, which has no parameters.

TbGetItemDropdownRectcomctl

TB_GETITEMDROPDOWNRECT message parameters.

TbGetItemRectcomctl

TB_GETITEMRECT message parameters.

TbGetMaxSizecomctl

TB_GETMAXSIZE message parameters.

TbGetMetricscomctl

TB_GETMETRICS message parameters.

TbGetObjectcomctl

TB_GETOBJECT message parameters.

TbGetPaddingcomctl

TB_GETPADDING message, which has no parameters.

TbGetPressedImageListcomctl

TB_GETPRESSEDIMAGELIST message, which has no parameters.

TbGetRectcomctl

TB_GETRECT message parameters.

TbGetRowscomctl

TB_GETROWS message, which has no parameters.

TbGetStatecomctl

TB_GETSTATE message parameters.

TbGetStringcomctl

TB_GETSTRING message parameters.

TbGetStylecomctl

TB_GETSTYLE message, which has no parameters.

TbGetTextRowscomctl

TB_GETTEXTROWS message, which has no parameters.

TbGetTooltipscomctl

TB_GETTOOLTIPS message, which has no parameters.

TbGetUnicodeFormatcomctl

TB_GETUNICODEFORMAT message, which has no parameters.

TbHideButtoncomctl

TB_HIDEBUTTON message parameters.

TbHitTestcomctl

TB_HITTEST message parameters.

TbIndeterminatecomctl

TB_INDETERMINATE message parameters.

TbInsertButtoncomctl

TB_INSERTBUTTON message parameters.

TbInsertMarkHitTestcomctl

TB_INSERTMARKHITTEST message parameters.

TbIsButtonCheckedcomctl

TB_ISBUTTONCHECKED message parameters.

TbIsButtonEnabledcomctl

TB_ISBUTTONENABLED message parameters.

TbIsButtonHiddencomctl

TB_ISBUTTONHIDDEN message parameters.

TbIsButtonHighlightedcomctl

TB_ISBUTTONHIGHLIGHTED message parameters.

TbIsButtonIndeterminatecomctl

TB_ISBUTTONINDETERMINATE message parameters.

TbIsButtonPressedcomctl

TB_ISBUTTONPRESSED message parameters.

TbLoadImagescomctl

TB_LOADIMAGES message parameters.

TbMapAcceleratorcomctl

TB_MAPACCELERATOR message parameters.

TbMarkButtoncomctl

TB_MARKBUTTON message parameters.

TbMoveButtoncomctl

TB_MOVEBUTTON message parameters.

TbPressButtoncomctl

TB_PRESSBUTTON message parameters.

TbReplaceBitmapcomctl

TB_REPLACEBITMAP message parameters.

TbSaveRestoreadvapi and comctl

TB_SAVERESTORE message parameters.

TbSetAnchorHighlightcomctl

TB_SETANCHORHIGHLIGHT message parameters.

TbSetBitmapSizecomctl

TB_SETBITMAPSIZE message parameters.

TbSetButtonInfocomctl

TB_SETBUTTONINFO message parameters.

TbSetButtonSizecomctl

TB_SETBUTTONSIZE message parameters.

TbSetButtonWidthcomctl

TB_SETBUTTONWIDTH message parameters.

TbSetCmdIdcomctl

TB_SETCMDID message parameters.

TbSetColorSchemecomctl

TB_SETCOLORSCHEME message parameters.

TbSetDisabledImageListcomctl

TB_SETDISABLEDIMAGELIST message, which has no parameters.

TbSetDrawTextFlagscomctl

TB_SETDRAWTEXTFLAGS message parameters.

TbSetExtendedStylecomctl

TB_SETEXTENDEDSTYLE message parameters.

TbSetHotImageListcomctl

TB_SETHOTIMAGELIST message, which has no parameters.

TbSetHotItemcomctl

TB_SETHOTITEM message parameters.

TbSetHotItem2comctl

TB_SETHOTITEM2 message parameters.

TbSetImageListcomctl

TB_SETIMAGELIST message, which has no parameters.

TbSetIndentcomctl

TB_SETINDENT message parameters.

TbSetInsertMarkcomctl

TB_SETINSERTMARK message parameters.

TbSetInsertMarkColorcomctl

TB_SETINSERTMARKCOLOR message parameters.

TbSetListGapcomctl

TB_SETLISTGAP message parameters.

TbSetMaxTextRowscomctl

TB_SETMAXTEXTROWS message parameters.

TbSetMetricscomctl

TB_SETMETRICS message parameters.

TbSetPaddingcomctl

TB_SETPADDING message parameters.

TbSetParentcomctl

TB_SETPARENT message parameters.

TbSetPressedImageListcomctl

TB_SETPRESSEDIMAGELIST message, which has no parameters.

TbSetRowscomctl

TB_SETROWS message parameters.

TbSetStatecomctl

TB_SETSTATE message parameters.

TbSetStylecomctl

TB_SETSTYLE message parameters.

TbSetTooltipscomctl

TB_SETTOOLTIPS message parameters.

TbSetUnicodeFormatcomctl

TB_SETUNICODEFORMAT message parameters.

TbSetWindowThemecomctl

TB_SETWINDOWTHEME message parameters.

TbmClearSelcomctl

TBM_CLEARSEL message parameters.

TbmClearTicscomctl

TBM_CLEARTICS message parameters.

TbmGetBuddycomctl

TBM_GETBUDDY message parameters.

TbmGetChannelRectcomctl

TBM_GETCHANNELRECT message parameters.

TbmGetLineSizecomctl

TBM_GETLINESIZE message, which has no parameters.

TbmGetNumTicscomctl

TBM_GETNUMTICS message, which has no parameters.

TbmGetPageSizecomctl

TBM_GETPAGESIZE message, which has no parameters.

TbmGetPoscomctl

TBM_GETPOS message, which has no parameters.

TbmGetRangeMaxcomctl

TBM_GETRANGEMAX message, which has no parameters.

TbmGetRangeMincomctl

TBM_GETRANGEMIN message, which has no parameters.

TbmGetSelEndcomctl

TBM_GETSELEND message, which has no parameters.

TbmGetSelStartcomctl

TBM_GETSELSTART message, which has no parameters.

TbmGetThumbLengthcomctl

TBM_GETTHUMBLENGTH message, which has no parameters.

TbmGetThumbRectcomctl

TBM_GETTHUMBRECT message parameters.

TbmGetTiccomctl

TBM_GETTIC message parameters.

TbmGetTicPoscomctl

TBM_GETTICPOS message parameters.

TbmGetTooltipscomctl

TBM_GETTOOLTIPS message, which has no parameters.

TbmGetUnicodeFormatcomctl

TBM_GETUNICODEFORMAT message, which has no parameters.

TbmSetBuddycomctl

TBM_SETBUDDY message parameters.

TbmSetLineSizecomctl

TBM_SETLINESIZE message parameters.

TbmSetPageSizecomctl

TBM_SETPAGSIZE message parameters.

TbmSetPoscomctl

TBM_SETPOS message parameters.

TbmSetPosNotifycomctl

TBM_SETPOSNOTIFY message parameters.

TbmSetRangecomctl

TBM_SETRANGE message parameters.

TbmSetRangeMaxcomctl

TBM_SETRANGEMAX message parameters.

TbmSetRangeMincomctl

TBM_SETRANGEMIN message parameters.

TbmSetSelcomctl

TBM_SETSEL message parameters.

TbmSetSelEndcomctl

TBM_SETSELEND message parameters.

TbmSetSelStartcomctl

TBM_SETSELSTART message parameters.

TbmSetThumbLengthcomctl

TBM_SETTHUMBLENGTH message parameters.

TbmSetTiccomctl

TBM_SETTIC message parameters.

TbmSetTicFreqcomctl

TBM_SETTICFREQ message parameters.

TbmSetTipSidecomctl

TBM_SETTIPSIDE message parameters.

TbmSetTooltipscomctl

TBM_SETTOOLTIPS message parameters.

TbmSetUnicodeFormatcomctl

TBM_SETUNICODEFORMAT message parameters.

TcmAdjustRectcomctl

TCM_ADJUSTRECT message parameters.

TcmDeleteAllItemscomctl

TCM_DELETEALLITEMS message, which has no parameters.

TcmDeleteItemcomctl

TCM_DELETEITEM message parameters.

TcmDeselectAllcomctl

TCM_DESELECTALL message parameters.

TcmGetCurFocuscomctl

TCM_GETCURFOCUS message, which has no parameters.

TcmGetCurSelcomctl

TCM_GETCURSEL message, which has no parameters.

TcmGetExtendedStylecomctl

TCM_GETEXTENDEDSTYLE message, which has no parameters.

TcmGetImageListcomctl

TCM_GETIMAGELIST message, which has no parameters.

TcmGetItemcomctl

TCM_GETITEM message parameters.

TcmGetItemCountcomctl

TCM_GETITEMCOUNT message, which has no parameters.

TcmGetItemRectcomctl

TCM_GETITEMRECT message parameters.

TcmGetRowCountcomctl

TCM_GETROWCOUNT message, which has no parameters.

TcmGetTooltipscomctl

TCM_GETTOOLTIPS message, which has no parameters.

TcmGetUnicodeFormatcomctl

TCM_GETUNICODEFORMAT message, which has no parameters.

TcmHighlightItemcomctl

TCM_HIGHLIGHTITEM message parameters.

TcmHitTestcomctl

TCM_HITTEST message parameters.

TcmInsertItemcomctl

TCM_INSERTITEM message parameters.

TcmRemoveImagecomctl

TCM_REMOVEIMAGE message parameters.

TcmSetCurFocuscomctl

TCM_SETCURFOCUS message parameters.

TcmSetCurSelcomctl

TCM_SETCURSEL message parameters.

TcmSetExtendedStylecomctl

TCM_SETEXTENDEDSTYLE message parameters.

TcmSetImageListcomctl

TCM_SETIMAGELIST message parameters.

TcmSetItemcomctl

TCM_SETITEM message parameters.

TcmSetItemExtracomctl

TCM_SETITEMEXTRA message parameters.

TcmSetItemSizecomctl

TCM_SETITEMSIZE message parameters.

TcmSetMinTabWidthcomctl

TCM_SETMINTABWIDTH message parameters.

TcmSetPaddingcomctl

TCM_SETPADDING message parameters.

TcmSetTooltipscomctl

TCM_SETTOOLTIPS message parameters.

TcmSetUnicodeFormatcomctl

TCM_SETUNICODEFORMAT message parameters.

TvmCreateDragImagecomctl

TVM_CREATEDRAGIMAGE message parameters.

TvmDeleteItemcomctl

TVM_DELETEITEM message parameters.

TvmEditLabelcomctl

TVM_EDITLABEL message parameters.

TvmEndEditLabelNowcomctl

TVM_ENDEDITLABELNOW message parameters.

TvmEnsureVisiblecomctl

TVM_ENSUREVISIBLE message parameters.

TvmExpandcomctl

TVM_EXPAND message parameters.

TvmGetBkColorcomctl

TVM_GETBKCOLOR message, which has no parameters.

TvmGetCountcomctl

TVM_GETCOUNT message, which has no parameters.

TvmGetEditControlcomctl

TVM_GETEDITCONTROL message, which has no parameters.

TvmGetExtendedStylecomctl

TVM_GETEXTENDEDSTYLE message, which has no parameters.

TvmGetISearchStringcomctl

TVM_GETISEARCHSTRING message parameters.

TvmGetImageListcomctl

TVM_GETIMAGELIST message parameters.

TvmGetIndentcomctl

TVM_GETINDENT message, which has no parameters.

TvmGetInsertMarkColorcomctl

TVM_GETINSERTMARKCOLOR message, which has no parameters.

TvmGetItemcomctl

TVM_GETITEM message parameters.

TvmGetItemHeightcomctl

TVM_GETITEMHEIGHT message, which has no parameters.

TvmGetItemRectcomctl

TVM_GETITEMRECT message parameters.

TvmGetItemStatecomctl

TVM_GETITEMSTATE message parameters.

TvmGetLineColorcomctl

TVM_GETLINECOLOR message, which has no parameters.

TvmGetNextItemcomctl

TVM_GETNEXTITEM message parameters.

TvmGetScrollTimecomctl

TVM_GETSCROLLTIME message, which has no parameters.

TvmGetTextColorcomctl

TVM_GETTEXTCOLOR message, which has no parameters.

TvmGetTooltipscomctl

TVM_GETTOOLTIPS message, which has no parameters.

TvmGetUnicodeFormatcomctl

HDM_GETUNICODEFORMAT message, which has no parameters.

TvmGetVisibleCountcomctl

TVM_GETVISIBLECOUNT message, which has no parameters.

TvmHitTestcomctl

TVM_HITTEST message parameters.

TvmInsertItemcomctl

TVM_INSERTITEM message parameters.

TvmMapAccIdToHtreeitemcomctl

TVM_MAPACCIDTOHTREEITEM message parameters.

TvmMapHtreeitemToAccIdcomctl

TVM_MAPHTREEITEMTOACCID message parameters.

TvmSelectItemcomctl

TVM_SELECTITEM message parameters.

TvmSetAutoScrollInfocomctl

TVM_SETAUTOSCROLLINFO message parameters.

TvmSetBkColorcomctl

TVM_SETBKCOLOR message parameters.

TvmSetBordercomctl

TVM_SETBORDER message parameters.

TvmSetExtendedStylecomctl

TVM_SETEXTENDEDSTYLE message parameters.

TvmSetHotcomctl

TVM_SETHOT message parameters.

TvmSetImageListcomctl

TVM_SETIMAGELIST message parameters.

TvmSetIndentcomctl

TVM_SETINDENT message parameters.

TvmSetInsertMarkcomctl

TVM_SETINSERTMARK message parameters.

TvmSetInsertMarkColorcomctl

TVM_SETINSERTMARKCOLOR message parameters.

TvmSetItemcomctl

TVM_SETITEM message parameters.

TvmSetItemHeightcomctl

TVM_SETITEMHEIGHT message parameters.

TvmSetLineColorcomctl

TVM_SETLINECOLOR message parameters.

TvmSetScrollTimecomctl

TVM_SETSCROLLTIME message parameters.

TvmSetTextColorcomctl

TVM_SETTEXTCOLOR message parameters.

TvmSetTooltipscomctl

TVM_SETTOOLTIPS message parameters.

TvmSetUnicodeFormatcomctl

TVM_SETUNICODEFORMAT message parameters.

TvmShowInfoTipcomctl

TVM_SHOWINFOTIP message parameters.

TvmSortChildrencomctl

TVM_SORTCHILDREN message parameters.

TvmSortChildrenCbcomctl

TVM_SORTCHILDRENCB message parameters.

UdmGetAccelcomctl

UDM_GETACCEL message parameters.

UdmGetBasecomctl

UDM_GETBASE message, which has no parameters.

UdmGetBuddycomctl

UDM_GETBUDDY message, which has no parameters.

UdmGetPoscomctl

UDM_GETPOS message, which has no parameters.

UdmGetPos32comctl

UDM_GETPOS32 message parameters.

UdmGetRangecomctl

UDM_GETRANGE message, which has no parameters.

UdmGetRange32comctl

UDM_GETRANGE32 message parameters.

UdmGetUnicodeFormatcomctl

UDM_GETUNICODEFORMAT message, which has no parameters.

UdmSetAccelcomctl

UDM_SETACCEL message parameters.

UdmSetBasecomctl

UDM_SETBASE message parameters.

UdmSetBuddycomctl

UDM_SETBUDDY message parameters.

UdmSetPoscomctl

UDM_SETPOS message parameters.

UdmSetPos32comctl

UDM_SETPOS32 message parameters.

UdmSetRangecomctl

UDM_SETRANGE message parameters.

UdmSetRange32comctl

UDM_SETRANGE32 message parameters.

UdmSetUnicodeFormatcomctl

UDM_SETUNICODEFORMAT message parameters.

Wm

Generic window message parameters: WPARAM and LPARAM.

WmActivate

WM_ACTIVATE message parameters.

WmActivateApp

WM_ACTIVATEAPP message parameters.

WmAppCommand

WM_APPCOMMAND message parameters.

WmCancelMode

WM_CANCELMODE message, which has no parameters.

WmCaptureChanged

WM_CAPTURECHANGED message parameters.

WmChar

WM_CHAR message parameters.

WmChildActivate

WM_CHILDACTIVATE message, which has no parameters.

WmClose

WM_CLOSE message, which has no parameters.

WmCommand

WM_COMMAND message parameters.

WmContextMenu

WM_CONTEXTMENU message parameters.

WmCreate

WM_CREATE message parameters.

WmCtlColorBtngdi

WM_CTLCOLORBTN message parameters.

WmCtlColorDlggdi

WM_CTLCOLORDLG message parameters.

WmCtlColorEditgdi

WM_CTLCOLOREDIT message parameters.

WmCtlColorListBoxgdi

WM_CTLCOLORLISTBOX message parameters.

WmCtlColorScrollBargdi

WM_CTLCOLORSCROLLBAR message parameters.

WmCtlColorStaticgdi

WM_CTLCOLORSTATIC message parameters.

WmDeadChargdi

WM_DEADCHAR message parameters.

WmDeleteItem

WM_DELETEITEM message parameters.

WmDestroy

WM_DESTROY message, which has no parameters.

WmDeviceChange

WM_DEVICECHANGE message parameters.

WmDisplayChangegdi

WM_DISPLAYCHANGE message parameters.

WmDropFilesshell

WM_DROPFILES message parameters.

WmEnable

WM_ENABLE message parameters.

WmEndSession

WM_ENDSESSION message parameters.

WmEnterIdle

WM_ENTERIDLE message parameters.

WmEnterMenuLoop

WM_ENTERMENULOOP message parameters.

WmEnterSizeMove

WM_ENTERSIZEMOVE message, which has no parameters.

WmEraseBkgnd

WM_ERASEBKGND message parameters.

WmExitMenuLoop

WM_EXITMENULOOP message parameters.

WmExitSizeMove

WM_EXITSIZEMOVE message, which has no parameters.

WmGetDlgCode

WM_GETDLGCODE message parameters.

WmGetFontgdi

WM_GETFONT message, which has no parameters.

WmGetHMenu

WM_GETHMENU message, which has no parameters. Originally has MN prefix.

WmGetHotKey

WM_GETHOTKEY message, which has no parameters.

WmGetMinMaxInfo

WM_GETMINMAXINFO message parameters.

WmGetText

WM_GETTEXT message parameters.

WmGetTextLength

WM_GETTEXTLENGTH message, which has no parameters.

WmGetTitleBarInfoEx

WM_GETTITLEBARINFOEX message parameters.

WmHScroll

WM_HSCROLL message parameters.

WmHelp

WM_HELP message parameters.

WmInitDialog

WM_INITDIALOG message parameters.

WmInitMenuPopup

WM_INITMENUPOPUP message parameters.

WmKeyDown

WM_KEYDOWN message parameters.

WmKeyUp

WM_KEYUP message parameters.

WmKillFocus

WM_KILLFOCUS message parameters.

WmLButtonDblClk

WM_LBUTTONDBLCLK message parameters.

WmLButtonDown

WM_LBUTTONDOWN message parameters.

WmLButtonUp

WM_LBUTTONUP message parameters.

WmMButtonDblClk

WM_MBUTTONDBLCLK message parameters.

WmMButtonDown

WM_MBUTTONDOWN message parameters.

WmMButtonUp

WM_MBUTTONUP message parameters.

WmMenuCommand

WM_MENUCOMMAND message parameters.

WmMenuDrag

WM_MENUDRAG message parameters.

WmMenuRButtonUp

WM_MENURBUTTONUP message parameters.

WmMouseHWheel

WM_MOUSEHWHEEL message parameters.

WmMouseHover

WM_MOUSEHOVER message parameters.

WmMouseLeave

WM_MOUSELEAVE message, which has no parameters.

WmMouseMove

WM_MOUSEMOVE message parameters.

WmMouseWheel

WM_MOUSEWHEEL message parameters.

WmMove

WM_MOVE message parameters.

WmMoving

WM_MOVING message parameters.

WmNcCalcSize

WM_NCCALCSIZE message parameters.

WmNcCreate

WM_NCCREATE message parameters.

WmNcDestroy

WM_NCDESTROY message, which has no parameters.

WmNcHitTest

WM_NCHITTEST message parameters.

WmNcPaintgdi

WM_NCPAINT message parameters.

WmNextDlgCtl

WM_NEXTDLGCTL message parameters.

WmNotifycomctl

WM_NOTIFY message parameters.

WmNull

WM_NULL message, which has no parameters.

WmPaintgdi

WM_PAINT message, which has no parameters.

WmParentNotify

WM_PARENTNOTIFY message parameters.

WmPowerBroadcast

WM_POWERBROADCAST message parameters.

WmQueryOpen

WM_QUERYOPEN message, which has no parameters.

WmRButtonDblClk

WM_RBUTTONDBLCLK message parameters.

WmRButtonDown

WM_RBUTTONDOWN message parameters.

WmRButtonUp

WM_RBUTTONUP message parameters.

WmSetCursor

WM_SETCURSOR message parameters.

WmSetFocus

WM_SETFOCUS message parameters.

WmSetFontgdi

WM_SETFONT message parameters.

WmSetHotKey

WM_SETHOTKEY message parameters.

WmSetIcon

WM_SETICON message parameters.

WmSetRedrawgdi

WM_SETREDRAW message parameters.

WmSetText

WM_SETTEXT message parameters.

WmShowWindow

WM_SHOWWINDOW message parameters.

WmSize

WM_SIZE message parameters.

WmSizing

WM_SIZING message parameters.

WmStyleChanged

WM_STYLECHANGED message parameters.

WmStyleChanging

WM_STYLECHANGING message parameters.

WmSyncPaintgdi

WM_SYNCPAINT message, which has no parameters.

WmSysChar

WM_SYSCHAR message parameters.

WmSysCommand

WM_SYSCOMMAND message parameters.

WmSysDeadChar

WM_SYSDEADCHAR message parameters.

WmSysKeyDown

WM_SYSKEYDOWN message parameters.

WmSysKeyUp

WM_SYSKEYUP message parameters.

WmThemeChanged

WM_THEMECHANGED message, which has no parameters.

WmTimer

WM_TIMER message parameters.

WmUndo

WM_UNDO message, which has no parameters.

WmUninitMenuPopup

WM_UNINITMENUPOPUP message parameters.

WmVScroll

WM_VSCROLL message parameters.

WmWindowPosChanged

WM_WINDOWPOSCHANGED message parameters.

WmWindowPosChanging

WM_WINDOWPOSCHANGING message parameters.

WmWtsSessionChange

WM_WTSSESSION_CHANGE message parameters.

WmXButtonDblClk

WM_XBUTTONDBLCLK message parameters.

WmXButtonDown

WM_XBUTTONDOWN message parameters.

WmXButtonUp

WM_XBUTTONUP message parameters.