Trait winsafe::prelude::AsAny

pub trait AsAny {
    fn as_any(&self) -> &dyn Any;
}

Used to convert a reference to the Any trait.

Required methods

fn as_any(&self) -> &dyn Any

Converts a reference to the Any trait. This is useful when storing a collection of polymorphic controls, because Any allows downcasting.

Examples

use std::sync::Arc;
use winsafe::prelude::*;
use winsafe::gui;

let parent: gui::WindowMain; // initialized somewhere

let ctrls: Vec<Arc<dyn gui::NativeControl>> = vec![
    Arc::new( gui::Edit::new(&parent, gui::EditOpts::default()) ),
    Arc::new( gui::Button::new(&parent, gui::ButtonOpts::default()) ),
];

let edit = ctrls[0].as_any().downcast_ref::<gui::Edit>()?;
edit.set_text("Foo")?;