Trait winsafe::prelude::ComctlHimagelist

pub trait ComctlHimagelist: Handle {

    fn Add(
        self, 
        hbmp_image: HBITMAP, 
        hbmp_mask: Option<HBITMAP>
    ) -> WinResult<u32> { ... }

    fn AddIcon(self, hicon: HICON) -> WinResult<u32> { ... }

    fn AddMasked(
        self, 
        hbmp_image: HBITMAP, 
        color_mask: COLORREF
    ) -> WinResult<u32> { ... }

    fn BeginDrag(self, track: u32, hotspot: POINT) -> WinResult<()> { ... }

    fn Create(
        image_sz: SIZE, 
        flags: ILC, 
        initial_size: i32, 
        grow_size: i32
    ) -> WinResult<HIMAGELIST> { ... }

    fn Destroy(self) -> WinResult<()> { ... }

    fn DragMove(self, x: i32, y: i32) -> WinResult<()> { ... }

    fn DragShowNolock(show: bool) -> WinResult<()> { ... }

    fn EndDrag() { ... }

    fn GetIconSize(self) -> WinResult<SIZE> { ... }

    fn GetImageCount(self) -> u32 { ... }

    fn Remove(self, index: Option<u32>) -> WinResult<()> { ... }

    fn ReplaceIcon(self, index: Option<u32>, hicon_new: HICON) -> WinResult<u32> { ... }

    fn SetImageCount(self, new_count: u32) -> WinResult<()> { ... }

}

This is supported on crate feature comctl only.

HIMAGELIST methods from comctl feature.

Provided methods

fn Add(self, hbmp_image: HBITMAP, hbmp_mask: Option<HBITMAP>) -> WinResult<u32>

ImageList_Add method.

Note: A copy of the bitmap is made, and this copy is then stored. You’re still responsible for freeing the original bitmap.

fn AddIcon(self, hicon: HICON) -> WinResult<u32>

ImageList_AddIcon method.

The icon can be destroyed right after this call, its handle its duplicated inside the image list.

fn AddMasked(self, hbmp_image: HBITMAP, color_mask: COLORREF) -> WinResult<u32>

ImageList_AddMasked method.

fn BeginDrag(self, track: u32, hotspot: POINT) -> WinResult<()>

ImageList_BeginDrag method.

Note: Must be paired with an HIMAGELIST::EndDrag call.

fn Create(
image_sz: SIZE,
flags: ILC,
initial_size: i32,
grow_size: i32
) -> WinResult<HIMAGELIST>

ImageList_Create static method.

Note: Must be paired with an HIMAGELIST::Destroy call.

Examples

use winsafe::prelude::*;
use winsafe::{co, HIMAGELIST, SIZE};

let himgl = HIMAGELIST::Create(
    SIZE::new(16, 16), co::ILC::COLOR32, 1, 1)?;

himgl.Destroy()?;

fn Destroy(self) -> WinResult<()>

ImageList_Destroy method.

fn DragMove(self, x: i32, y: i32) -> WinResult<()>

ImageList_DragMove method.

fn DragShowNolock(show: bool) -> WinResult<()>

ImageList_DragShowNolock static method.

fn EndDrag()

ImageList_EndDrag static method.

fn GetIconSize(self) -> WinResult<SIZE>

ImageList_GetIconSize method.

fn GetImageCount(self) -> u32

ImageList_GetImageCount method.

fn Remove(self, index: Option<u32>) -> WinResult<()>

ImageList_Remove method.

fn ReplaceIcon(self, index: Option<u32>, hicon_new: HICON) -> WinResult<u32>

ImageList_ReplaceIcon method.

The icon can be destroyed right after this call, its handle its duplicated inside the image list.

fn SetImageCount(self, new_count: u32) -> WinResult<()>

ImageList_SetImageCount methods.