Trait winsafe::prelude::ComctlHwnd

pub trait ComctlHwnd: Handle {
    fn DefSubclassProc<M>(self, msg: M) -> M::RetType
    where
        M: MsgSend,
    { ... }

    fn RemoveWindowSubclass(
        self, 
        subclass_func: SUBCLASSPROC, 
        subclass_id: usize
    ) -> WinResult<()> { ... }

    fn SetWindowSubclass(
        self, 
        subclass_proc: SUBCLASSPROC, 
        subclass_id: usize, 
        ref_data: usize
    ) -> WinResult<()> { ... }
}

This is supported on crate feature comctl only.

HWND methods from comctl feature.

Provided methods

fn DefSubclassProc<M>(self, msg: M) -> M::RetType where
M: MsgSend,

DefSubclassProc method.

The return type is variable, being defined by the RetType associated type of the MsgSend trait. That means each message can define its own return type.

fn RemoveWindowSubclass(
self,
subclass_func: SUBCLASSPROC,
subclass_id: usize
) -> WinResult<()>

RemoveWindowSubclass method.

fn SetWindowSubclass(
self,
subclass_proc: SUBCLASSPROC,
subclass_id: usize,
ref_data: usize
) -> WinResult<()>

SetWindowSubclass method.