Trait winsafe::prelude::ComctlShellHimagelist

pub trait ComctlShellHimagelist: ComctlHimagelist {
    fn add_icon_from_shell(self, file_extensions: &[&str]) -> WinResult<()> { ... }
}

This is supported on crate features comctl and shell only.

HIMAGELIST methods from comctl+shell features.

Provided methods

fn add_icon_from_shell(self, file_extensions: &[&str]) -> WinResult<()>

Calls SHGetFileInfo to retrieve one or more shell file icons, then passes them to AddIcon.

Examples

use winsafe::prelude::*;
use winsafe::{co, HIMAGELIST, SIZE};

let himgl = HIMAGELIST::Create(
    SIZE::new(16, 16), co::ILC::COLOR32, 1, 1)?;

himgl.add_icon_from_shell(&["mp3", "wav"])?;

himgl.Destroy()?;