winsafe::prelude

Trait GuiControl

pub trait GuiControl: GuiWindow {
    // Required method
    fn ctrl_id(&self) -> u16;

    // Provided method
    fn focus(&self) -> SysResult<()> { ... }
}

Available on crate feature gui only.

This trait is enabled with the gui feature, and is implemented by all child controls.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Required Methods

fn ctrl_id(&self) -> u16

Returns the control ID, which is defined at control creation.

The control ID should be unique within a parent.

Provided Methods

fn focus(&self) -> SysResult<()>

If the control belongs to an ordinary window, simply calls HWND:SetFocus.

If the control belongs to a dialog window, sends a wm::NextDlgCtl message. This is preferable to the HWND::SetFocus because it takes care of border highlighting, like the native Button control needs.