Trait winsafe::prelude::GuiEvents

pub trait GuiEvents {

    fn wm<F>(&self, ident: WM, func: F)
    where
        F: Fn(WndMsg) -> AnyResult<Option<isize>> + 'static;

    fn wm_activate<F>(&self, func: F)
    where
        F: Fn(Activate) -> AnyResult<()> + 'static,
    { ... }
    fn wm_activate_app<F>(&self, func: F)
    where
        F: Fn(ActivateApp) -> AnyResult<()> + 'static,
    { ... }
    fn wm_app_command<F>(&self, func: F)
    where
        F: Fn(AppCommand) -> AnyResult<()> + 'static,
    { ... }
    fn wm_cancel_mode<F>(&self, func: F)
    where
        F: Fn() -> AnyResult<()> + 'static,
    { ... }
    fn wm_capture_changed<F>(&self, func: F)
    where
        F: Fn(CaptureChanged) -> AnyResult<()> + 'static,
    { ... }
    fn wm_char<F>(&self, func: F)
    where
        F: Fn(Char) -> AnyResult<()> + 'static,
    { ... }
    fn wm_child_activate<F>(&self, func: F)
    where
        F: Fn() -> AnyResult<()> + 'static,
    { ... }
    fn wm_close<F>(&self, func: F)
    where
        F: Fn() -> AnyResult<()> + 'static,
    { ... }
    fn wm_context_menu<F>(&self, func: F)
    where
        F: Fn() -> AnyResult<()> + 'static,
    { ... }
    fn wm_create<F>(&self, func: F)
    where
        F: Fn(Create<'_, '_, '_>) -> AnyResult<i32> + 'static,
    { ... }
    fn wm_ctl_color_btn<F>(&self, func: F)
    where
        F: Fn(CtlColorBtn) -> AnyResult<HBRUSH> + 'static,
    { ... }
    fn wm_ctl_color_dlg<F>(&self, func: F)
    where
        F: Fn(CtlColorDlg) -> AnyResult<HBRUSH> + 'static,
    { ... }
    fn wm_ctl_color_edit<F>(&self, func: F)
    where
        F: Fn(CtlColorEdit) -> AnyResult<HBRUSH> + 'static,
    { ... }
    fn wm_ctl_color_list_box<F>(&self, func: F)
    where
        F: Fn(CtlColorListBox) -> AnyResult<HBRUSH> + 'static,
    { ... }
    fn wm_ctl_color_scroll_bar<F>(&self, func: F)
    where
        F: Fn(CtlColorScrollBar) -> AnyResult<HBRUSH> + 'static,
    { ... }
    fn wm_ctl_color_static<F>(&self, func: F)
    where
        F: Fn(CtlColorStatic) -> AnyResult<HBRUSH> + 'static,
    { ... }
    fn wm_dead_char<F>(&self, func: F)
    where
        F: Fn(DeadChar) -> AnyResult<()> + 'static,
    { ... }
    fn wm_delete_item<F>(&self, func: F)
    where
        F: Fn(DeleteItem<'_>) -> AnyResult<bool> + 'static,
    { ... }
    fn wm_destroy<F>(&self, func: F)
    where
        F: Fn() -> AnyResult<()> + 'static,
    { ... }
    fn wm_display_change<F>(&self, func: F)
    where
        F: Fn(DisplayChange) -> AnyResult<()> + 'static,
    { ... }
    fn wm_drop_files<F>(&self, func: F)
    where
        F: Fn(DropFiles) -> AnyResult<()> + 'static,
    { ... }
    fn wm_enable<F>(&self, func: F)
    where
        F: Fn(Enable) -> AnyResult<()> + 'static,
    { ... }
    fn wm_end_session<F>(&self, func: F)
    where
        F: Fn(EndSession) -> AnyResult<()> + 'static,
    { ... }
    fn wm_enter_idle<F>(&self, func: F)
    where
        F: Fn(EnterIdle) -> AnyResult<()> + 'static,
    { ... }
    fn wm_enter_menu_loop<F>(&self, func: F)
    where
        F: Fn(EnterMenuLoop) -> AnyResult<()> + 'static,
    { ... }
    fn wm_enter_size_move<F>(&self, func: F)
    where
        F: Fn() -> AnyResult<()> + 'static,
    { ... }
    fn wm_erase_bkgnd<F>(&self, func: F)
    where
        F: Fn(EraseBkgnd) -> AnyResult<i32> + 'static,
    { ... }
    fn wm_exit_menu_loop<F>(&self, func: F)
    where
        F: Fn(ExitMenuLoop) -> AnyResult<()> + 'static,
    { ... }
    fn wm_exit_size_move<F>(&self, func: F)
    where
        F: Fn() -> AnyResult<()> + 'static,
    { ... }
    fn wm_get_dlg_code<F>(&self, func: F)
    where
        F: Fn(GetDlgCode<'_>) -> AnyResult<DLGC> + 'static,
    { ... }
    fn wm_get_font<F>(&self, func: F)
    where
        F: Fn() -> AnyResult<Option<HFONT>> + 'static,
    { ... }
    fn wm_get_hmenu<F>(&self, func: F)
    where
        F: Fn() -> AnyResult<Option<HMENU>> + 'static,
    { ... }
    fn wm_get_min_max_info<F>(&self, func: F)
    where
        F: Fn(GetMinMaxInfo<'_>) -> AnyResult<()> + 'static,
    { ... }
    fn wm_get_text<F>(&self, func: F)
    where
        F: Fn(GetText<'_>) -> AnyResult<u32> + 'static,
    { ... }
    fn wm_get_text_length<F>(&self, func: F)
    where
        F: Fn() -> AnyResult<u32> + 'static,
    { ... }
    fn wm_get_title_bar_info_ex<F>(&self, func: F)
    where
        F: Fn(GetTitleBarInfoEx<'_>) -> AnyResult<()> + 'static,
    { ... }
    fn wm_help<F>(&self, func: F)
    where
        F: Fn(Help<'_>) -> AnyResult<()> + 'static,
    { ... }
    fn wm_h_scroll<F>(&self, func: F)
    where
        F: Fn(HScroll) -> AnyResult<()> + 'static,
    { ... }
    fn wm_init_dialog<F>(&self, func: F)
    where
        F: Fn(InitDialog) -> AnyResult<bool> + 'static,
    { ... }
    fn wm_init_menu_popup<F>(&self, func: F)
    where
        F: Fn(InitMenuPopup) -> AnyResult<()> + 'static,
    { ... }
    fn wm_key_down<F>(&self, func: F)
    where
        F: Fn(KeyDown) -> AnyResult<()> + 'static,
    { ... }
    fn wm_key_up<F>(&self, func: F)
    where
        F: Fn(KeyUp) -> AnyResult<()> + 'static,
    { ... }
    fn wm_kill_focus<F>(&self, func: F)
    where
        F: Fn(KillFocus) -> AnyResult<()> + 'static,
    { ... }
    fn wm_l_button_dbl_clk<F>(&self, func: F)
    where
        F: Fn(LButtonDblClk) -> AnyResult<()> + 'static,
    { ... }
    fn wm_l_button_down<F>(&self, func: F)
    where
        F: Fn(LButtonDown) -> AnyResult<()> + 'static,
    { ... }
    fn wm_l_button_up<F>(&self, func: F)
    where
        F: Fn(LButtonUp) -> AnyResult<()> + 'static,
    { ... }
    fn wm_m_button_dbl_clk<F>(&self, func: F)
    where
        F: Fn(MButtonDblClk) -> AnyResult<()> + 'static,
    { ... }
    fn wm_m_button_down<F>(&self, func: F)
    where
        F: Fn(MButtonDown) -> AnyResult<()> + 'static,
    { ... }
    fn wm_m_button_up<F>(&self, func: F)
    where
        F: Fn(MButtonUp) -> AnyResult<()> + 'static,
    { ... }
    fn wm_menu_command<F>(&self, func: F)
    where
        F: Fn(MenuCommand) -> AnyResult<()> + 'static,
    { ... }
    fn wm_menu_drag<F>(&self, func: F)
    where
        F: Fn(MenuDrag) -> AnyResult<MND> + 'static,
    { ... }
    fn wm_menu_r_button_up<F>(&self, func: F)
    where
        F: Fn(MenuRButtonUp) -> AnyResult<()> + 'static,
    { ... }
    fn wm_mouse_hover<F>(&self, func: F)
    where
        F: Fn(MouseHover) -> AnyResult<()> + 'static,
    { ... }
    fn wm_mouse_leave<F>(&self, func: F)
    where
        F: Fn() -> AnyResult<()> + 'static,
    { ... }
    fn wm_mouse_move<F>(&self, func: F)
    where
        F: Fn(MouseMove) -> AnyResult<()> + 'static,
    { ... }
    fn wm_move<F>(&self, func: F)
    where
        F: Fn(Move) -> AnyResult<()> + 'static,
    { ... }
    fn wm_moving<F>(&self, func: F)
    where
        F: Fn(Moving<'_>) -> AnyResult<()> + 'static,
    { ... }
    fn wm_nc_calc_size<F>(&self, func: F)
    where
        F: Fn(NcCalcSize<'_, '_>) -> AnyResult<WVR> + 'static,
    { ... }
    fn wm_nc_create<F>(&self, func: F)
    where
        F: Fn(NcCreate<'_, '_, '_>) -> AnyResult<bool> + 'static,
    { ... }
    fn wm_nc_destroy<F>(&self, func: F)
    where
        F: Fn() -> AnyResult<()> + 'static,
    { ... }
    fn wm_nc_hit_test<F>(&self, func: F)
    where
        F: Fn(NcHitTest) -> AnyResult<HT> + 'static,
    { ... }
    fn wm_nc_paint<F>(&self, func: F)
    where
        F: Fn(NcPaint) -> AnyResult<()> + 'static,
    { ... }
    fn wm_next_dlg_ctl<F>(&self, func: F)
    where
        F: Fn(NextDlgCtl) -> AnyResult<()> + 'static,
    { ... }
    fn wm_null<F>(&self, func: F)
    where
        F: Fn() -> AnyResult<()> + 'static,
    { ... }
    fn wm_paint<F>(&self, func: F)
    where
        F: Fn() -> AnyResult<()> + 'static,
    { ... }
    fn wm_parent_notify<F>(&self, func: F)
    where
        F: Fn(ParentNotify) -> AnyResult<()> + 'static,
    { ... }
    fn wm_query_open<F>(&self, func: F)
    where
        F: Fn() -> AnyResult<bool> + 'static,
    { ... }
    fn wm_r_button_dbl_clk<F>(&self, func: F)
    where
        F: Fn(RButtonDblClk) -> AnyResult<()> + 'static,
    { ... }
    fn wm_r_button_down<F>(&self, func: F)
    where
        F: Fn(RButtonDown) -> AnyResult<()> + 'static,
    { ... }
    fn wm_r_button_up<F>(&self, func: F)
    where
        F: Fn(RButtonUp) -> AnyResult<()> + 'static,
    { ... }
    fn wm_set_cursor<F>(&self, func: F)
    where
        F: Fn(SetCursor) -> AnyResult<bool> + 'static,
    { ... }
    fn wm_set_focus<F>(&self, func: F)
    where
        F: Fn(SetFocus) -> AnyResult<()> + 'static,
    { ... }
    fn wm_set_font<F>(&self, func: F)
    where
        F: Fn(SetFont) -> AnyResult<()> + 'static,
    { ... }
    fn wm_set_icon<F>(&self, func: F)
    where
        F: Fn(SetIcon) -> AnyResult<Option<HICON>> + 'static,
    { ... }
    fn wm_set_redraw<F>(&self, func: F)
    where
        F: Fn(SetRedraw) -> AnyResult<()> + 'static,
    { ... }
    fn wm_set_text<F>(&self, func: F)
    where
        F: Fn(SetText) -> AnyResult<bool> + 'static,
    { ... }
    fn wm_show_window<F>(&self, func: F)
    where
        F: Fn(ShowWindow) -> AnyResult<()> + 'static,
    { ... }
    fn wm_size<F>(&self, func: F)
    where
        F: Fn(Size) -> AnyResult<()> + 'static,
    { ... }
    fn wm_sizing<F>(&self, func: F)
    where
        F: Fn(Sizing<'_>) -> AnyResult<()> + 'static,
    { ... }
    fn wm_style_changed<F>(&self, func: F)
    where
        F: Fn(StyleChanged<'_>) -> AnyResult<()> + 'static,
    { ... }
    fn wm_style_changing<F>(&self, func: F)
    where
        F: Fn(StyleChanging<'_>) -> AnyResult<()> + 'static,
    { ... }
    fn wm_sync_paint<F>(&self, func: F)
    where
        F: Fn() -> AnyResult<()> + 'static,
    { ... }
    fn wm_sys_char<F>(&self, func: F)
    where
        F: Fn(SysChar) -> AnyResult<()> + 'static,
    { ... }
    fn wm_sys_command<F>(&self, func: F)
    where
        F: Fn(SysCommand) -> AnyResult<()> + 'static,
    { ... }
    fn wm_sys_dead_char<F>(&self, func: F)
    where
        F: Fn(SysDeadChar) -> AnyResult<()> + 'static,
    { ... }
    fn wm_sys_key_down<F>(&self, func: F)
    where
        F: Fn(SysKeyDown) -> AnyResult<()> + 'static,
    { ... }
    fn wm_sys_key_up<F>(&self, func: F)
    where
        F: Fn(SysKeyUp) -> AnyResult<()> + 'static,
    { ... }
    fn wm_theme_changed<F>(&self, func: F)
    where
        F: Fn() -> AnyResult<()> + 'static,
    { ... }
    fn wm_uninit_menu_popup<F>(&self, func: F)
    where
        F: Fn(UninitMenuPopup) -> AnyResult<()> + 'static,
    { ... }
    fn wm_undo<F>(&self, func: F)
    where
        F: Fn() -> AnyResult<bool> + 'static,
    { ... }
    fn wm_v_scroll<F>(&self, func: F)
    where
        F: Fn(VScroll) -> AnyResult<()> + 'static,
    { ... }
    fn wm_window_pos_changed<F>(&self, func: F)
    where
        F: Fn(WindowPosChanged<'_>) -> AnyResult<()> + 'static,
    { ... }
    fn wm_window_pos_changing<F>(&self, func: F)
    where
        F: Fn(WindowPosChanging<'_>) -> AnyResult<()> + 'static,
    { ... }
    fn wm_x_button_dbl_clk<F>(&self, func: F)
    where
        F: Fn(XButtonDblClk) -> AnyResult<()> + 'static,
    { ... }
    fn wm_x_button_down<F>(&self, func: F)
    where
        F: Fn(XButtonDown) -> AnyResult<()> + 'static,
    { ... }
    fn wm_x_button_up<F>(&self, func: F)
    where
        F: Fn(XButtonUp) -> AnyResult<()> + 'static,
    { ... }

}

Available on crate feature gui only.

Exposes the basic window message methods of WindowEvents.

Required Methods

fn wm<F>(&self, ident: WM, func: F)where
F: Fn(WndMsg) -> AnyResult<Option<isize>> + 'static,

Event to any window message.

Note: Instead of using this event, you should always prefer the specific events, which will give you the correct message parameters. This generic method should be used only when you have a custom, non-standard window message – which should be pretty rare.

Examples

Handling a custom, user-defined message:

use winsafe::prelude::*;
use winsafe::{co, gui, msg, AnyResult};

let wnd: gui::WindowMain; // initialized somewhere

let CUSTOM_MSG = co::WM::from(0x1234);

wnd.on().wm(CUSTOM_MSG, {
    let wnd = wnd.clone(); // to pass into the closure
    move |p: msg::WndMsg| -> AnyResult<Option<isize>> {
        println!("HWND: {}, msg ID: {}", wnd.hwnd(), p.msg_id);
        Ok(Some(0))
    }
});

Provided Methods

fn wm_activate<F>(&self, func: F)where
F: Fn(Activate) -> AnyResult<()> + 'static,

WM_ACTIVATE message.

Default handling

If you handle this event, you’ll overwrite the default handling in:

non-dialog WindowMain.

fn wm_activate_app<F>(&self, func: F)where
F: Fn(ActivateApp) -> AnyResult<()> + 'static,

WM_ACTIVATEAPP message.

fn wm_app_command<F>(&self, func: F)where
F: Fn(AppCommand) -> AnyResult<()> + 'static,

WM_APPCOMMAND message.

fn wm_cancel_mode<F>(&self, func: F)where
F: Fn() -> AnyResult<()> + 'static,

WM_CANCELMODE message.

fn wm_capture_changed<F>(&self, func: F)where
F: Fn(CaptureChanged) -> AnyResult<()> + 'static,

WM_CAPTURECHANGED message.

fn wm_char<F>(&self, func: F)where
F: Fn(Char) -> AnyResult<()> + 'static,

WM_CHAR message.

fn wm_child_activate<F>(&self, func: F)where
F: Fn() -> AnyResult<()> + 'static,

WM_CHILDACTIVATE message.

fn wm_close<F>(&self, func: F)where
F: Fn() -> AnyResult<()> + 'static,

WM_CLOSE message.

Default handling

If you handle this event, you’ll overwrite the default handling in:

dialog WindowMain;

dialog WindowModal;

non-dialog WindowModal.

fn wm_context_menu<F>(&self, func: F)where
F: Fn() -> AnyResult<()> + 'static,

WM_CONTEXTMENU message.

fn wm_create<F>(&self, func: F)where
F: Fn(Create<'_, '_, '_>) -> AnyResult<i32> + 'static,

WM_CREATE message, sent only to non-dialog windows. Dialog windows receive WM_INITDIALOG instead.

Examples

use winsafe::prelude::*;
use winsafe::{gui, msg, AnyResult};

let wnd: gui::WindowMain; // initialized somewhere

wnd.on().wm_create({
    let wnd = wnd.clone(); // to pass into the closure
    move |p: msg::wm::Create| -> AnyResult<i32> {
        println!("HWND: {}, client area: {}x{}",
            wnd.hwnd(),
            p.createstruct.cx,
            p.createstruct.cy,
        );
        Ok(0)
    }
});

fn wm_ctl_color_btn<F>(&self, func: F)where
F: Fn(CtlColorBtn) -> AnyResult<HBRUSH> + 'static,

WM_CTLCOLORBTN message.

fn wm_ctl_color_dlg<F>(&self, func: F)where
F: Fn(CtlColorDlg) -> AnyResult<HBRUSH> + 'static,

WM_CTLCOLORDLG message.

fn wm_ctl_color_edit<F>(&self, func: F)where
F: Fn(CtlColorEdit) -> AnyResult<HBRUSH> + 'static,

WM_CTLCOLOREDIT message.

fn wm_ctl_color_list_box<F>(&self, func: F)where
F: Fn(CtlColorListBox) -> AnyResult<HBRUSH> + 'static,

WM_CTLCOLORLISTBOX message.

fn wm_ctl_color_scroll_bar<F>(&self, func: F)where
F: Fn(CtlColorScrollBar) -> AnyResult<HBRUSH> + 'static,

WM_CTLCOLORSCROLLBAR message.

fn wm_ctl_color_static<F>(&self, func: F)where
F: Fn(CtlColorStatic) -> AnyResult<HBRUSH> + 'static,

WM_CTLCOLORSTATIC message.

fn wm_dead_char<F>(&self, func: F)where
F: Fn(DeadChar) -> AnyResult<()> + 'static,

WM_DEADCHAR message.

fn wm_delete_item<F>(&self, func: F)where
F: Fn(DeleteItem<'_>) -> AnyResult<bool> + 'static,

WM_DELETEITEM message.

fn wm_destroy<F>(&self, func: F)where
F: Fn() -> AnyResult<()> + 'static,

WM_DESTROY message.

Examples

use winsafe::prelude::*;
use winsafe::{gui, AnyResult};

let wnd: gui::WindowMain; // initialized somewhere

wnd.on().wm_destroy(move || -> AnyResult<()> {
    println!("Window is gone, goodbye!");
    Ok(())
});

fn wm_display_change<F>(&self, func: F)where
F: Fn(DisplayChange) -> AnyResult<()> + 'static,

WM_DISPLAYCHANGE message.

fn wm_drop_files<F>(&self, func: F)where
F: Fn(DropFiles) -> AnyResult<()> + 'static,

WM_DROPFILES message.

Examples

use winsafe::prelude::*;
use winsafe::{gui, msg, AnyResult};

let wnd: gui::WindowMain; // initialized somewhere

wnd.on().wm_drop_files(move |p: msg::wm::DropFiles| -> AnyResult<()> {
    for dropped_file in p.hdrop.iter()? {
        let dropped_file = dropped_file?;
        println!("Dropped: {}", dropped_file);
    }
    Ok(())
});

fn wm_enable<F>(&self, func: F)where
F: Fn(Enable) -> AnyResult<()> + 'static,

WM_ENABLE message.

fn wm_end_session<F>(&self, func: F)where
F: Fn(EndSession) -> AnyResult<()> + 'static,

WM_ENDSESSION message.

fn wm_enter_idle<F>(&self, func: F)where
F: Fn(EnterIdle) -> AnyResult<()> + 'static,

WM_ENTERIDLE message.

fn wm_enter_menu_loop<F>(&self, func: F)where
F: Fn(EnterMenuLoop) -> AnyResult<()> + 'static,

WM_ENTERMENULOOP message.

fn wm_enter_size_move<F>(&self, func: F)where
F: Fn() -> AnyResult<()> + 'static,

WM_ENTERSIZEMOVE message.

fn wm_erase_bkgnd<F>(&self, func: F)where
F: Fn(EraseBkgnd) -> AnyResult<i32> + 'static,

WM_ERASEBKGND message.

fn wm_exit_menu_loop<F>(&self, func: F)where
F: Fn(ExitMenuLoop) -> AnyResult<()> + 'static,

WM_EXITMENULOOP message.

fn wm_exit_size_move<F>(&self, func: F)where
F: Fn() -> AnyResult<()> + 'static,

WM_EXITSIZEMOVE message.

fn wm_get_dlg_code<F>(&self, func: F)where
F: Fn(GetDlgCode<'_>) -> AnyResult<DLGC> + 'static,

WM_GETDLGCODE message.

fn wm_get_font<F>(&self, func: F)where
F: Fn() -> AnyResult<Option<HFONT>> + 'static,

WM_GETFONT message.

fn wm_get_hmenu<F>(&self, func: F)where
F: Fn() -> AnyResult<Option<HMENU>> + 'static,

WM_GETHMENU message. Originally has MN prefix.

fn wm_get_min_max_info<F>(&self, func: F)where
F: Fn(GetMinMaxInfo<'_>) -> AnyResult<()> + 'static,

WM_GETMINMAXINFO message.

fn wm_get_text<F>(&self, func: F)where
F: Fn(GetText<'_>) -> AnyResult<u32> + 'static,

WM_GETTEXT message.

fn wm_get_text_length<F>(&self, func: F)where
F: Fn() -> AnyResult<u32> + 'static,

WM_GETTEXTLENGTH message.

fn wm_get_title_bar_info_ex<F>(&self, func: F)where
F: Fn(GetTitleBarInfoEx<'_>) -> AnyResult<()> + 'static,

WM_GETTITLEBARINFOEX message.

fn wm_help<F>(&self, func: F)where
F: Fn(Help<'_>) -> AnyResult<()> + 'static,

WM_HELP message.

fn wm_h_scroll<F>(&self, func: F)where
F: Fn(HScroll) -> AnyResult<()> + 'static,

WM_HSCROLL message.

fn wm_init_dialog<F>(&self, func: F)where
F: Fn(InitDialog) -> AnyResult<bool> + 'static,

WM_INITDIALOG message, sent only to dialog windows. Non-dialog windows receive WM_CREATE instead.

Examples

use winsafe::prelude::*;
use winsafe::{gui, msg, AnyResult};

let wnd: gui::WindowMain; // initialized somewhere

let wnd = wnd.clone(); // to pass into the closure
wnd.on().wm_init_dialog(move |p: msg::wm::InitDialog| -> AnyResult<bool> {
    println!("Focused HWND: {}", p.hwnd_focus);
    Ok(true)
});

fn wm_init_menu_popup<F>(&self, func: F)where
F: Fn(InitMenuPopup) -> AnyResult<()> + 'static,

WM_INITMENUPOPUP message.

fn wm_key_down<F>(&self, func: F)where
F: Fn(KeyDown) -> AnyResult<()> + 'static,

WM_KEYDOWN message.

fn wm_key_up<F>(&self, func: F)where
F: Fn(KeyUp) -> AnyResult<()> + 'static,

WM_KEYUP message.

fn wm_kill_focus<F>(&self, func: F)where
F: Fn(KillFocus) -> AnyResult<()> + 'static,

WM_KILLFOCUS message.

fn wm_l_button_dbl_clk<F>(&self, func: F)where
F: Fn(LButtonDblClk) -> AnyResult<()> + 'static,

WM_LBUTTONDBLCLK message.

Examples

use winsafe::prelude::*;
use winsafe::{gui, msg, AnyResult};

let wnd: gui::WindowMain; // initialized somewhere

let wnd = wnd.clone(); // to pass into the closure
wnd.on().wm_l_button_dbl_clk(move |p: msg::wm::LButtonDblClk| -> AnyResult<()> {
    println!("Point: {}x{}", p.coords.x, p.coords.y);
    Ok(())
});

fn wm_l_button_down<F>(&self, func: F)where
F: Fn(LButtonDown) -> AnyResult<()> + 'static,

WM_LBUTTONDOWN message.

Examples

use winsafe::prelude::*;
use winsafe::{gui, msg, AnyResult};

let wnd: gui::WindowMain; // initialized somewhere

let wnd = wnd.clone(); // to pass into the closure
wnd.on().wm_l_button_down(move |p: msg::wm::LButtonDown| -> AnyResult<()> {
    println!("Point: {}x{}", p.coords.x, p.coords.y);
    Ok(())
});

fn wm_l_button_up<F>(&self, func: F)where
F: Fn(LButtonUp) -> AnyResult<()> + 'static,

WM_LBUTTONUP message.

fn wm_m_button_dbl_clk<F>(&self, func: F)where
F: Fn(MButtonDblClk) -> AnyResult<()> + 'static,

WM_MBUTTONDBLCLK message.

fn wm_m_button_down<F>(&self, func: F)where
F: Fn(MButtonDown) -> AnyResult<()> + 'static,

WM_MBUTTONDOWN message.

fn wm_m_button_up<F>(&self, func: F)where
F: Fn(MButtonUp) -> AnyResult<()> + 'static,

WM_MBUTTONUP message.

fn wm_menu_command<F>(&self, func: F)where
F: Fn(MenuCommand) -> AnyResult<()> + 'static,

WM_MENUCOMMAND message.

fn wm_menu_drag<F>(&self, func: F)where
F: Fn(MenuDrag) -> AnyResult<MND> + 'static,

WM_MENUDRAG message.

fn wm_menu_r_button_up<F>(&self, func: F)where
F: Fn(MenuRButtonUp) -> AnyResult<()> + 'static,

WM_MENURBUTTONUP message.

fn wm_mouse_hover<F>(&self, func: F)where
F: Fn(MouseHover) -> AnyResult<()> + 'static,

WM_MOUSEHOVER message.

fn wm_mouse_leave<F>(&self, func: F)where
F: Fn() -> AnyResult<()> + 'static,

WM_MOUSELEAVE message.

fn wm_mouse_move<F>(&self, func: F)where
F: Fn(MouseMove) -> AnyResult<()> + 'static,

WM_MOUSEMOVE message.

fn wm_move<F>(&self, func: F)where
F: Fn(Move) -> AnyResult<()> + 'static,

WM_MOVE message.

fn wm_moving<F>(&self, func: F)where
F: Fn(Moving<'_>) -> AnyResult<()> + 'static,

WM_MOVING message.

fn wm_nc_calc_size<F>(&self, func: F)where
F: Fn(NcCalcSize<'_, '_>) -> AnyResult<WVR> + 'static,

WM_NCCALCSIZE message.

fn wm_nc_create<F>(&self, func: F)where
F: Fn(NcCreate<'_, '_, '_>) -> AnyResult<bool> + 'static,

WM_NCCREATE message.

fn wm_nc_destroy<F>(&self, func: F)where
F: Fn() -> AnyResult<()> + 'static,

WM_NCDESTROY message.

Default handling

If you handle this event, you’ll overwrite the default handling in:

non-dialog WindowMain;

dialog WindowMain.

fn wm_nc_hit_test<F>(&self, func: F)where
F: Fn(NcHitTest) -> AnyResult<HT> + 'static,

WM_NCHITTEST message.

fn wm_nc_paint<F>(&self, func: F)where
F: Fn(NcPaint) -> AnyResult<()> + 'static,

WM_NCPAINT message.

Default handling

If you handle this event, you’ll overwrite the default handling in:

non-dialog WindowControl;

dialog WindowControl.

fn wm_next_dlg_ctl<F>(&self, func: F)where
F: Fn(NextDlgCtl) -> AnyResult<()> + 'static,

WM_NEXTDLGCTL message.

fn wm_null<F>(&self, func: F)where
F: Fn() -> AnyResult<()> + 'static,

WM_NULL message.

fn wm_paint<F>(&self, func: F)where
F: Fn() -> AnyResult<()> + 'static,

WM_PAINT message.

Examples

use winsafe::prelude::*;
use winsafe::{gui, AnyResult};

let wnd: gui::WindowMain; // initialized somewhere

wnd.on().wm_paint({
    let wnd = wnd.clone(); // to pass into the closure
    move || -> AnyResult<()> {
        let hdc = wnd.hwnd().BeginPaint()?;

        // hdc painting...

        Ok(())
    }
});

fn wm_parent_notify<F>(&self, func: F)where
F: Fn(ParentNotify) -> AnyResult<()> + 'static,

WM_PARENTNOTIFY message.

fn wm_query_open<F>(&self, func: F)where
F: Fn() -> AnyResult<bool> + 'static,

WM_QUERYOPEN message.

fn wm_r_button_dbl_clk<F>(&self, func: F)where
F: Fn(RButtonDblClk) -> AnyResult<()> + 'static,

WM_RBUTTONDBLCLK message.

fn wm_r_button_down<F>(&self, func: F)where
F: Fn(RButtonDown) -> AnyResult<()> + 'static,

WM_RBUTTONDOWN message.

fn wm_r_button_up<F>(&self, func: F)where
F: Fn(RButtonUp) -> AnyResult<()> + 'static,

WM_RBUTTONUP

fn wm_set_cursor<F>(&self, func: F)where
F: Fn(SetCursor) -> AnyResult<bool> + 'static,

WM_SETCURSOR message.

fn wm_set_focus<F>(&self, func: F)where
F: Fn(SetFocus) -> AnyResult<()> + 'static,

WM_SETFOCUS message.

Default handling

If you handle this event, you’ll overwrite the default handling in:

non-dialog WindowMain;

non-dialog WindowModal.

fn wm_set_font<F>(&self, func: F)where
F: Fn(SetFont) -> AnyResult<()> + 'static,

WM_SETFONT message.

fn wm_set_icon<F>(&self, func: F)where
F: Fn(SetIcon) -> AnyResult<Option<HICON>> + 'static,

WM_SETICON message.

fn wm_set_redraw<F>(&self, func: F)where
F: Fn(SetRedraw) -> AnyResult<()> + 'static,

WM_SETREDRAW message.

fn wm_set_text<F>(&self, func: F)where
F: Fn(SetText) -> AnyResult<bool> + 'static,

WM_SETTEXT message.

fn wm_show_window<F>(&self, func: F)where
F: Fn(ShowWindow) -> AnyResult<()> + 'static,

WM_SHOWWINDOW message.

fn wm_size<F>(&self, func: F)where
F: Fn(Size) -> AnyResult<()> + 'static,

WM_SIZE message.

Examples

use winsafe::prelude::*;
use winsafe::{gui, msg, AnyResult};

let wnd: gui::WindowMain; // initialized somewhere

wnd.on().wm_size({
    let wnd = wnd.clone(); // to pass into the closure
    move |p: msg::wm::Size| -> AnyResult<()> {
        println!("HWND: {}, client area: {}x{}",
            wnd.hwnd(),
            p.client_area.cx,
            p.client_area.cy,
        );
        Ok(())
    }
});

fn wm_sizing<F>(&self, func: F)where
F: Fn(Sizing<'_>) -> AnyResult<()> + 'static,

WM_SIZING message.

fn wm_style_changed<F>(&self, func: F)where
F: Fn(StyleChanged<'_>) -> AnyResult<()> + 'static,

WM_STYLECHANGED message.

fn wm_style_changing<F>(&self, func: F)where
F: Fn(StyleChanging<'_>) -> AnyResult<()> + 'static,

WM_STYLECHANGING message.

fn wm_sync_paint<F>(&self, func: F)where
F: Fn() -> AnyResult<()> + 'static,

WM_SYNCPAINT message.

fn wm_sys_char<F>(&self, func: F)where
F: Fn(SysChar) -> AnyResult<()> + 'static,

WM_SYSCHAR message.

fn wm_sys_command<F>(&self, func: F)where
F: Fn(SysCommand) -> AnyResult<()> + 'static,

WM_SYSCOMMAND message.

fn wm_sys_dead_char<F>(&self, func: F)where
F: Fn(SysDeadChar) -> AnyResult<()> + 'static,

WM_SYSDEADCHAR message.

fn wm_sys_key_down<F>(&self, func: F)where
F: Fn(SysKeyDown) -> AnyResult<()> + 'static,

WM_SYSKEYDOWN message.

fn wm_sys_key_up<F>(&self, func: F)where
F: Fn(SysKeyUp) -> AnyResult<()> + 'static,

WM_SYSKEYUP message.

fn wm_theme_changed<F>(&self, func: F)where
F: Fn() -> AnyResult<()> + 'static,

WM_THEMECHANGED message.

fn wm_uninit_menu_popup<F>(&self, func: F)where
F: Fn(UninitMenuPopup) -> AnyResult<()> + 'static,

WM_UNINITMENUPOPUP message.

fn wm_undo<F>(&self, func: F)where
F: Fn() -> AnyResult<bool> + 'static,

WM_UNDO message.

fn wm_v_scroll<F>(&self, func: F)where
F: Fn(VScroll) -> AnyResult<()> + 'static,

WM_VSCROLL message.

fn wm_window_pos_changed<F>(&self, func: F)where
F: Fn(WindowPosChanged<'_>) -> AnyResult<()> + 'static,

WM_WINDOWPOSCHANGED message.

fn wm_window_pos_changing<F>(&self, func: F)where
F: Fn(WindowPosChanging<'_>) -> AnyResult<()> + 'static,

WM_WINDOWPOSCHANGING message.

fn wm_x_button_dbl_clk<F>(&self, func: F)where
F: Fn(XButtonDblClk) -> AnyResult<()> + 'static,

WM_XBUTTONDBLCLK message.

fn wm_x_button_down<F>(&self, func: F)where
F: Fn(XButtonDown) -> AnyResult<()> + 'static,

WM_XBUTTONDOWN message.

fn wm_x_button_up<F>(&self, func: F)where
F: Fn(XButtonUp) -> AnyResult<()> + 'static,

WM_XBUTTONUP message.