winsafe::prelude

Trait GuiEventsButton

pub trait GuiEventsButton: GuiEvents {
    // Provided methods
    fn bcn_drop_down<F>(&self, func: F)
       where F: Fn(&NMBCDROPDOWN) -> AnyResult<()> + 'static { ... }
    fn bcn_hot_item_change<F>(&self, func: F)
       where F: Fn(&NMBCHOTITEM) -> AnyResult<()> + 'static { ... }
    fn bn_clicked<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn bn_dbl_clk<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn bn_kill_focus<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn bn_set_focus<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn nm_custom_draw<F>(&self, func: F)
       where F: Fn(&NMCUSTOMDRAW) -> AnyResult<CDRF> + 'static { ... }
}

Available on crate feature gui only.

This trait is enabled with the gui feature, and exposes button control notifications.

These event methods are just proxies to the GuiEventsParent of the parent window, who is the real responsible for the child event handling.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn bcn_drop_down<F>(&self, func: F)
where F: Fn(&NMBCDROPDOWN) -> AnyResult<()> + 'static,

BCN_DROPDOWN notification.

fn bcn_hot_item_change<F>(&self, func: F)
where F: Fn(&NMBCHOTITEM) -> AnyResult<()> + 'static,

BCN_HOTITEMCHANGE notification.

fn bn_clicked<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

BN_CLICKED command notification.

Examples

use winsafe::{self as w, prelude::*, gui};

let wnd: gui::WindowMain; // initialized somewhere
let btn: gui::Button;

btn.on().bn_clicked(
    move || -> w::AnyResult<()> {
        println!("Clicked.");
        Ok(())
    },
);

fn bn_dbl_clk<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

BN_DBLCLK command notification.

fn bn_kill_focus<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

BN_KILLFOCUS command notification.

fn bn_set_focus<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

BN_SETFOCUS command notification.

fn nm_custom_draw<F>(&self, func: F)
where F: Fn(&NMCUSTOMDRAW) -> AnyResult<CDRF> + 'static,

NM_CUSTOMDRAW notification.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.