winsafe::prelude

Trait GuiEventsComboBox

pub trait GuiEventsComboBox: GuiEvents {
    // Provided methods
    fn cbn_close_up<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn cbn_dbl_clk<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn cbn_drop_down<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn cbn_edit_change<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn cbn_edit_update<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn cbn_err_space<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn cbn_kill_focus<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn cbn_sel_change<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn cbn_sel_end_cancel<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn cbn_sel_end_ok<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn cbn_set_focus<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
}

Available on crate feature gui only.

This trait is enabled with the gui feature, and exposes combo box control notifications.

These event methods are just proxies to the GuiEventsParent of the parent window, who is the real responsible for the child event handling.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn cbn_close_up<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

CBN_CLOSEUP command notification.

fn cbn_dbl_clk<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

CBN_DBLCLK command notification.

fn cbn_drop_down<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

CBN_DROPDOWN command notification.

fn cbn_edit_change<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

CBN_EDITCHANGE command notification.

fn cbn_edit_update<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

CBN_EDITUPDATE command notification.

fn cbn_err_space<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

CBN_ERRSPACE command notification.

fn cbn_kill_focus<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

CBN_KILLFOCUS command notification.

fn cbn_sel_change<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

CBN_SELCHANGE command notification.

Examples

use winsafe::{self as w, prelude::*, gui};

let wnd: gui::WindowMain; // initialized somewhere
let cmb: gui::ComboBox;

let cmb2 = cmb.clone(); // to pass into the closure

cmb.on().cbn_sel_change(
    move || -> w::AnyResult<()> {
        if let Some(sel_text) = cmb2.items().selected_text()? {
            println!("New selected text: {}", sel_text);
        }
        Ok(())
    },
);

fn cbn_sel_end_cancel<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

CBN_SELENDCANCEL command notification.

fn cbn_sel_end_ok<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

CBN_SELENDOK command notification.

fn cbn_set_focus<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

CBN_SETFOCUS command notification.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.