winsafe::prelude

Trait GuiEventsDateTimePicker

pub trait GuiEventsDateTimePicker: GuiEvents {
    // Provided methods
    fn dtn_close_up<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn dtn_date_time_change<F>(&self, func: F)
       where F: Fn(&NMDATETIMECHANGE) -> AnyResult<()> + 'static { ... }
    fn dtn_drop_down<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn dtn_format<F>(&self, func: F)
       where F: Fn(&mut NMDATETIMEFORMAT<'_>) -> AnyResult<()> + 'static { ... }
    fn dtn_format_query<F>(&self, func: F)
       where F: Fn(&mut NMDATETIMEFORMATQUERY<'_>) -> AnyResult<()> + 'static { ... }
    fn dtn_user_string<F>(&self, func: F)
       where F: Fn(&mut NMDATETIMESTRING<'_>) -> AnyResult<()> + 'static { ... }
    fn dtn_wm_key_down<F>(&self, func: F)
       where F: Fn(&NMDATETIMEWMKEYDOWN<'_>) -> AnyResult<()> + 'static { ... }
    fn nm_kill_focus<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn nm_set_focus<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
}

Available on crate feature gui only.

This trait is enabled with the gui feature, and exposes date and time picker control notifications.

These event methods are just proxies to the GuiEventsParent of the parent window, who is the real responsible for the child event handling.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn dtn_close_up<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

DTN_CLOSEUP notification.

fn dtn_date_time_change<F>(&self, func: F)
where F: Fn(&NMDATETIMECHANGE) -> AnyResult<()> + 'static,

DTN_DATETIMECHANGE notification.

Examples

use winsafe::{self as w, prelude::*, gui};

let wnd: gui::WindowMain; // initialized somewhere
let dtp: gui::DateTimePicker;

dtp.on().dtn_date_time_change(
    move |p: &w::NMDATETIMECHANGE| -> w::AnyResult<()> {
        println!("{}-{}-{}", p.st.wYear, p.st.wMonth, p.st.wDay);
        Ok(())
    },
);

fn dtn_drop_down<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

DTN_DROPDOWN notification.

fn dtn_format<F>(&self, func: F)
where F: Fn(&mut NMDATETIMEFORMAT<'_>) -> AnyResult<()> + 'static,

DTN_FORMAT notification.

fn dtn_format_query<F>(&self, func: F)
where F: Fn(&mut NMDATETIMEFORMATQUERY<'_>) -> AnyResult<()> + 'static,

DTN_FORMATQUERY notification.

fn dtn_user_string<F>(&self, func: F)
where F: Fn(&mut NMDATETIMESTRING<'_>) -> AnyResult<()> + 'static,

DTN_USERSTRING notification.

fn dtn_wm_key_down<F>(&self, func: F)
where F: Fn(&NMDATETIMEWMKEYDOWN<'_>) -> AnyResult<()> + 'static,

DTN_WMKEYDOWN notification.

fn nm_kill_focus<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

NM_KILLFOCUS notification.

fn nm_set_focus<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

NM_SETFOCUS notification.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.