winsafe::prelude

Trait GuiEventsHeader

pub trait GuiEventsHeader: GuiEvents {

    // Provided methods
    fn hdn_begin_drag<F>(&self, func: F)
       where F: Fn(&NMHEADER<'_>) -> AnyResult<bool> + 'static { ... }
    fn hdn_begin_filter_edit<F>(&self, func: F)
       where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static { ... }
    fn hdn_begin_track<F>(&self, func: F)
       where F: Fn(&NMHEADER<'_>) -> AnyResult<bool> + 'static { ... }
    fn hdn_divider_dbl_click<F>(&self, func: F)
       where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static { ... }
    fn hdn_drop_down<F>(&self, func: F)
       where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static { ... }
    fn hdn_end_drag<F>(&self, func: F)
       where F: Fn(&NMHEADER<'_>) -> AnyResult<bool> + 'static { ... }
    fn hdn_end_filter_edit<F>(&self, func: F)
       where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static { ... }
    fn hdn_end_track<F>(&self, func: F)
       where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static { ... }
    fn hdn_filter_btn_click<F>(&self, func: F)
       where F: Fn(&NMHDFILTERBTNCLICK) -> AnyResult<bool> + 'static { ... }
    fn hdn_filter_change<F>(&self, func: F)
       where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static { ... }
    fn hdn_get_disp_info<F>(&self, func: F)
       where F: Fn(&mut NMHDDISPINFO<'_>) -> AnyResult<isize> + 'static { ... }
    fn hdn_item_changed<F>(&self, func: F)
       where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static { ... }
    fn hdn_item_changing<F>(&self, func: F)
       where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static { ... }
    fn hdn_item_click<F>(&self, func: F)
       where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static { ... }
    fn hdn_item_dbl_click<F>(&self, func: F)
       where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static { ... }
    fn hdn_item_key_down<F>(&self, func: F)
       where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static { ... }
    fn hdn_item_state_icon_click<F>(&self, func: F)
       where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static { ... }
    fn hdn_overflow_click<F>(&self, func: F)
       where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static { ... }
    fn hdn_track<F>(&self, func: F)
       where F: Fn(&NMHDFILTERBTNCLICK) -> AnyResult<bool> + 'static { ... }
    fn nm_custom_draw<F>(&self, func: F)
       where F: Fn(&mut NMCUSTOMDRAW) -> AnyResult<CDRF> + 'static { ... }
    fn nm_r_click<F>(&self, func: F)
       where F: Fn() -> AnyResult<i32> + 'static { ... }
    fn nm_released_capture<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }

}

Available on crate feature gui only.

This trait is enabled with the gui feature, and exposes header control notifications.

These event methods are just proxies to the GuiEventsParent of the parent window, who is the real responsible for the child event handling.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn hdn_begin_drag<F>(&self, func: F)
where F: Fn(&NMHEADER<'_>) -> AnyResult<bool> + 'static,

HDN_BEGINDRAG notification.

fn hdn_begin_filter_edit<F>(&self, func: F)
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_BEGINFILTEREDIT notification.

fn hdn_begin_track<F>(&self, func: F)
where F: Fn(&NMHEADER<'_>) -> AnyResult<bool> + 'static,

HDN_BEGINTRACK notification.

fn hdn_divider_dbl_click<F>(&self, func: F)
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_DIVIDERDBLCLICK notification.

fn hdn_drop_down<F>(&self, func: F)
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_DROPDOWN notification.

fn hdn_end_drag<F>(&self, func: F)
where F: Fn(&NMHEADER<'_>) -> AnyResult<bool> + 'static,

HDN_ENDDRAG notification.

fn hdn_end_filter_edit<F>(&self, func: F)
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_ENDFILTEREDIT notification.

fn hdn_end_track<F>(&self, func: F)
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_ENDTRACK notification.

fn hdn_filter_btn_click<F>(&self, func: F)
where F: Fn(&NMHDFILTERBTNCLICK) -> AnyResult<bool> + 'static,

HDN_FILTERBTNCLICK notification.

fn hdn_filter_change<F>(&self, func: F)
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_FILTERCHANGE notification.

fn hdn_get_disp_info<F>(&self, func: F)
where F: Fn(&mut NMHDDISPINFO<'_>) -> AnyResult<isize> + 'static,

HDN_GETDISPINFO notification.

fn hdn_item_changed<F>(&self, func: F)
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_ITEMCHANGED notification.

fn hdn_item_changing<F>(&self, func: F)
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_ITEMCHANGING notification.

fn hdn_item_click<F>(&self, func: F)
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_ITEMCLICK notification.

fn hdn_item_dbl_click<F>(&self, func: F)
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_ITEMDBLCLICK notification.

fn hdn_item_key_down<F>(&self, func: F)
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_ITEMKEYDOWN notification.

fn hdn_item_state_icon_click<F>(&self, func: F)
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_ITEMSTATEICONCLICK notification.

fn hdn_overflow_click<F>(&self, func: F)
where F: Fn(&NMHEADER<'_>) -> AnyResult<()> + 'static,

HDN_OVERFLOWCLICK notification.

fn hdn_track<F>(&self, func: F)
where F: Fn(&NMHDFILTERBTNCLICK) -> AnyResult<bool> + 'static,

HDN_TRACK notification.

fn nm_custom_draw<F>(&self, func: F)
where F: Fn(&mut NMCUSTOMDRAW) -> AnyResult<CDRF> + 'static,

NM_CUSTOMDRAW notification.

fn nm_r_click<F>(&self, func: F)
where F: Fn() -> AnyResult<i32> + 'static,

NM_RCLICK notification.

fn nm_released_capture<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

NM_RELEASEDCAPTURE notification.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.