winsafe::prelude

Trait GuiEventsLabel

pub trait GuiEventsLabel: GuiEvents {
    // Provided methods
    fn stn_clicked<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn stn_dbl_clk<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn stn_disable<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn stn_enable<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
}

Available on crate feature gui only.

This trait is enabled with the gui feature, and exposes label control notifications.

These event methods are just proxies to the GuiEventsParent of the parent window, who is the real responsible for the child event handling.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn stn_clicked<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

STN_CLICKED notification.

fn stn_dbl_clk<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

STN_DBLCLK notification.

fn stn_disable<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

STN_DISABLE notification.

fn stn_enable<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

STN_ENABLE notification.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.