winsafe::prelude

Trait GuiEventsListBox

pub trait GuiEventsListBox: GuiEvents {
    // Provided methods
    fn lbn_dbl_clk<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn lbn_err_space<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn lbn_kill_focus<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn lbn_sel_cancel<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn lbn_sel_change<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn lbn_set_focus<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
}

Available on crate feature gui only.

This trait is enabled with the gui feature, and exposes list box control notifications.

These event methods are just proxies to the GuiEventsParent of the parent window, who is the real responsible for the child event handling.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn lbn_dbl_clk<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

LBN_DBLCLK command notification.

fn lbn_err_space<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

LBN_ERRSPACE command notification.

fn lbn_kill_focus<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

LBN_KILLFOCUS command notification.

fn lbn_sel_cancel<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

LBN_SELCANCEL command notification.

fn lbn_sel_change<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

LBN_SELCHANGE command notification.

fn lbn_set_focus<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

LBN_SETFOCUS command notification.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.