winsafe::prelude

Trait GuiEventsListView

pub trait GuiEventsListView: GuiEvents {

    // Provided methods
    fn lvn_begin_drag<F>(&self, func: F)
       where F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static { ... }
    fn lvn_begin_label_edit<F>(&self, func: F)
       where F: Fn(&NMLVDISPINFO<'_>) -> AnyResult<bool> + 'static { ... }
    fn lvn_begin_r_drag<F>(&self, func: F)
       where F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static { ... }
    fn lvn_begin_scroll<F>(&self, func: F)
       where F: Fn(&NMLVSCROLL) -> AnyResult<()> + 'static { ... }
    fn lvn_column_click<F>(&self, func: F)
       where F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static { ... }
    fn lvn_column_drop_down<F>(&self, func: F)
       where F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static { ... }
    fn lvn_column_overflow_click<F>(&self, func: F)
       where F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static { ... }
    fn lvn_delete_all_items<F>(&self, func: F)
       where F: Fn() -> AnyResult<bool> + 'static { ... }
    fn lvn_delete_item<F>(&self, func: F)
       where F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static { ... }
    fn lvn_end_label_edit<F>(&self, func: F)
       where F: Fn(&NMLVDISPINFO<'_>) -> AnyResult<bool> + 'static { ... }
    fn lvn_end_scroll<F>(&self, func: F)
       where F: Fn(&NMLVSCROLL) -> AnyResult<()> + 'static { ... }
    fn lvn_get_disp_info<F>(&self, func: F)
       where F: Fn(&mut NMLVDISPINFO<'_>) -> AnyResult<()> + 'static { ... }
    fn lvn_get_empty_markup<F>(&self, func: F)
       where F: Fn(&mut NMLVEMPTYMARKUP) -> AnyResult<bool> + 'static { ... }
    fn lvn_get_info_tip<F>(&self, func: F)
       where F: Fn(&NMLVGETINFOTIP<'_>) -> AnyResult<()> + 'static { ... }
    fn lvn_hot_track<F>(&self, func: F)
       where F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static { ... }
    fn lvn_incremental_search<F>(&self, func: F)
       where F: Fn(&NMLVFINDITEM<'_>) -> AnyResult<()> + 'static { ... }
    fn lvn_insert_item<F>(&self, func: F)
       where F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static { ... }
    fn lvn_item_activate<F>(&self, func: F)
       where F: Fn(&NMITEMACTIVATE) -> AnyResult<()> + 'static { ... }
    fn lvn_item_changed<F>(&self, func: F)
       where F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static { ... }
    fn lvn_item_changing<F>(&self, func: F)
       where F: Fn(&NMLISTVIEW) -> AnyResult<bool> + 'static { ... }
    fn lvn_key_down<F>(&self, func: F)
       where F: Fn(&NMLVKEYDOWN) -> AnyResult<()> + 'static { ... }
    fn lvn_link_click<F>(&self, func: F)
       where F: Fn(&NMLVLINK) -> AnyResult<()> + 'static { ... }
    fn lvn_marquee_begin<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn lvn_od_cache_hint<F>(&self, func: F)
       where F: Fn(&NMLVCACHEHINT) -> AnyResult<()> + 'static { ... }
    fn lvn_od_find_item<F>(&self, func: F)
       where F: Fn(&mut NMLVFINDITEM<'_>) -> AnyResult<Option<u32>> + 'static { ... }
    fn lvn_od_state_changed<F>(&self, func: F)
       where F: Fn(&NMLVODSTATECHANGE) -> AnyResult<()> + 'static { ... }
    fn lvn_set_disp_info<F>(&self, func: F)
       where F: Fn(&NMLVDISPINFO<'_>) -> AnyResult<()> + 'static { ... }
    fn nm_click<F>(&self, func: F)
       where F: Fn(&NMITEMACTIVATE) -> AnyResult<()> + 'static { ... }
    fn nm_custom_draw<F>(&self, func: F)
       where F: Fn(&mut NMLVCUSTOMDRAW) -> AnyResult<CDRF> + 'static { ... }
    fn nm_dbl_clk<F>(&self, func: F)
       where F: Fn(&NMITEMACTIVATE) -> AnyResult<()> + 'static { ... }
    fn nm_hover<F>(&self, func: F)
       where F: Fn() -> AnyResult<i32> + 'static { ... }
    fn nm_kill_focus<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn nm_r_click<F>(&self, func: F)
       where F: Fn(&NMITEMACTIVATE) -> AnyResult<i32> + 'static { ... }
    fn nm_r_dbl_clk<F>(&self, func: F)
       where F: Fn(&NMITEMACTIVATE) -> AnyResult<()> + 'static { ... }
    fn nm_released_capture<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn nm_return<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn nm_set_focus<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }

}

Available on crate feature gui only.

This trait is enabled with the gui feature, and exposes list view control notifications.

These event methods are just proxies to the GuiEventsParent of the parent window, who is the real responsible for the child event handling.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn lvn_begin_drag<F>(&self, func: F)
where F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static,

LVN_BEGINDRAG notification.

fn lvn_begin_label_edit<F>(&self, func: F)
where F: Fn(&NMLVDISPINFO<'_>) -> AnyResult<bool> + 'static,

LVN_BEGINLABELEDIT notification.

fn lvn_begin_r_drag<F>(&self, func: F)
where F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static,

LVN_BEGINRDRAG notification.

fn lvn_begin_scroll<F>(&self, func: F)
where F: Fn(&NMLVSCROLL) -> AnyResult<()> + 'static,

LVN_BEGINSCROLL notification.

fn lvn_column_click<F>(&self, func: F)
where F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static,

LVN_COLUMNCLICK notification.

fn lvn_column_drop_down<F>(&self, func: F)
where F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static,

LVN_COLUMNDROPDOWN notification.

fn lvn_column_overflow_click<F>(&self, func: F)
where F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static,

LVN_COLUMNOVERFLOWCLICK notification.

fn lvn_delete_all_items<F>(&self, func: F)
where F: Fn() -> AnyResult<bool> + 'static,

LVN_DELETEALLITEMS notification.

Examples

use winsafe::{self as w, prelude::*, gui};

let wnd: gui::WindowMain; // initialized somewhere
let list: gui::ListView;

list.on().lvn_delete_all_items(
    move || -> w::AnyResult<bool> {
        println!("All items deleted.");
        Ok(true) // supress subsequent WM_DELETEITEM notifications
    },
);

fn lvn_delete_item<F>(&self, func: F)
where F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static,

LVN_DELETEITEM notification.

Examples

use winsafe::{self as w, prelude::*, gui};

let wnd: gui::WindowMain; // initialized somewhere
let list: gui::ListView;

list.on().lvn_delete_item(
    move |p: &w::NMLISTVIEW| -> w::AnyResult<()> {
        println!("Item: {}", p.iItem);
        Ok(())
    },
);

fn lvn_end_label_edit<F>(&self, func: F)
where F: Fn(&NMLVDISPINFO<'_>) -> AnyResult<bool> + 'static,

LVN_ENDLABELEDIT notification.

fn lvn_end_scroll<F>(&self, func: F)
where F: Fn(&NMLVSCROLL) -> AnyResult<()> + 'static,

LVN_ENDSCROLL notification.

fn lvn_get_disp_info<F>(&self, func: F)
where F: Fn(&mut NMLVDISPINFO<'_>) -> AnyResult<()> + 'static,

LVN_GETDISPINFO notification.

fn lvn_get_empty_markup<F>(&self, func: F)
where F: Fn(&mut NMLVEMPTYMARKUP) -> AnyResult<bool> + 'static,

LVN_GETEMPTYMARKUP notification.

fn lvn_get_info_tip<F>(&self, func: F)
where F: Fn(&NMLVGETINFOTIP<'_>) -> AnyResult<()> + 'static,

LVN_GETINFOTIP notification.

fn lvn_hot_track<F>(&self, func: F)
where F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static,

LVN_HOTTRACK notification.

fn lvn_incremental_search<F>(&self, func: F)
where F: Fn(&NMLVFINDITEM<'_>) -> AnyResult<()> + 'static,

LVN_INCREMENTALSEARCH notification.

fn lvn_insert_item<F>(&self, func: F)
where F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static,

LVN_INSERTITEM notification.

fn lvn_item_activate<F>(&self, func: F)
where F: Fn(&NMITEMACTIVATE) -> AnyResult<()> + 'static,

LVN_ITEMACTIVATE notification.

fn lvn_item_changed<F>(&self, func: F)
where F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static,

LVN_ITEMCHANGED notification.

Examples

use winsafe::{self as w, prelude::*, gui};

let wnd: gui::WindowMain; // initialized somewhere
let list: gui::ListView;

list.on().lvn_item_changed(
    move |p: &w::NMLISTVIEW| -> w::AnyResult<()> {
        println!("Item: {}", p.iItem);
        Ok(())
    },
);

fn lvn_item_changing<F>(&self, func: F)
where F: Fn(&NMLISTVIEW) -> AnyResult<bool> + 'static,

LVN_ITEMCHANGING notification.

fn lvn_key_down<F>(&self, func: F)
where F: Fn(&NMLVKEYDOWN) -> AnyResult<()> + 'static,

LVN_KEYDOWN notification.

Examples

use winsafe::{self as w, prelude::*, co, gui};

let wnd: gui::WindowMain; // initialized somewhere
let list: gui::ListView;

list.on().lvn_key_down(
    move |p: &w::NMLVKEYDOWN| -> w::AnyResult<()> {
        if p.wVKey == co::VK::DELETE {
            println!("DEL key was pressed.");
        }
        Ok(())
    },
);

fn lvn_link_click<F>(&self, func: F)
where F: Fn(&NMLVLINK) -> AnyResult<()> + 'static,

LVN_LINKCLICK notification.

fn lvn_marquee_begin<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

LVN_MARQUEEBEGIN notification.

fn lvn_od_cache_hint<F>(&self, func: F)
where F: Fn(&NMLVCACHEHINT) -> AnyResult<()> + 'static,

LVN_ODCACHEHINT notification.

fn lvn_od_find_item<F>(&self, func: F)
where F: Fn(&mut NMLVFINDITEM<'_>) -> AnyResult<Option<u32>> + 'static,

LVN_ODFINDITEM notification.

fn lvn_od_state_changed<F>(&self, func: F)
where F: Fn(&NMLVODSTATECHANGE) -> AnyResult<()> + 'static,

LVN_ODSTATECHANGED notification.

fn lvn_set_disp_info<F>(&self, func: F)
where F: Fn(&NMLVDISPINFO<'_>) -> AnyResult<()> + 'static,

LVN_SETDISPINFO notification.

fn nm_click<F>(&self, func: F)
where F: Fn(&NMITEMACTIVATE) -> AnyResult<()> + 'static,

NM_CLICK notification.

fn nm_custom_draw<F>(&self, func: F)
where F: Fn(&mut NMLVCUSTOMDRAW) -> AnyResult<CDRF> + 'static,

NM_CUSTOMDRAW notification.

fn nm_dbl_clk<F>(&self, func: F)
where F: Fn(&NMITEMACTIVATE) -> AnyResult<()> + 'static,

NM_DBLCLK notification.

fn nm_hover<F>(&self, func: F)
where F: Fn() -> AnyResult<i32> + 'static,

NM_HOVER notification.

fn nm_kill_focus<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

NM_KILLFOCUS notification.

fn nm_r_click<F>(&self, func: F)
where F: Fn(&NMITEMACTIVATE) -> AnyResult<i32> + 'static,

NM_RCLICK notification.

fn nm_r_dbl_clk<F>(&self, func: F)
where F: Fn(&NMITEMACTIVATE) -> AnyResult<()> + 'static,

NM_RDBLCLK notification.

fn nm_released_capture<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

NM_RELEASEDCAPTURE notification.

fn nm_return<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

NM_RETURN notification.

fn nm_set_focus<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

NM_SETFOCUS notification.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.