winsafe::prelude

Trait GuiEventsParent

pub trait GuiEventsParent: GuiEventsWindow {
    // Required methods
    fn wm_command<F>(
        &self,
        ctrl_id: impl Into<u16>,
        code: impl Into<CMD>,
        func: F,
    )
       where F: Fn() -> AnyResult<()> + 'static;
    fn wm_notify<F>(
        &self,
        id_from: impl Into<u16>,
        code: impl Into<NmhdrCode>,
        func: F,
    )
       where F: Fn(Notify<'_>) -> AnyResult<isize> + 'static;

    // Provided methods
    fn wm_command_acc_menu<F>(&self, ctrl_id: impl Into<u16> + Copy, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn wm_create<F>(&self, func: F)
       where F: Fn(Create<'_, '_, '_>) -> AnyResult<i32> + 'static { ... }
    fn wm_init_dialog<F>(&self, func: F)
       where F: Fn(InitDialog) -> AnyResult<bool> + 'static { ... }
    fn wm_nc_create<F>(&self, func: F)
       where F: Fn(NcCreate<'_, '_, '_>) -> AnyResult<bool> + 'static { ... }
}

Available on crate feature gui only.

This trait is enabled with the gui feature, and provides events for window notifications.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Required Methods

fn wm_command<F>(&self, ctrl_id: impl Into<u16>, code: impl Into<CMD>, func: F)
where F: Fn() -> AnyResult<()> + 'static,

WM_COMMAND message, for specific code and control ID.

A command notification must be narrowed by the command code and the control ID, so the closure will be fired for that specific control at that specific event.

Instead of using this event, you should always prefer the specific command notifications, which will give you the correct message parameters. This generic method should be used only when you have a custom, non-standard window notification.

use winsafe::{self as w, prelude::*, co, gui};

let wnd: gui::WindowMain; // initialized somewhere

const CTRL_ID: u16 = 1010;

wnd.on().wm_command(
    CTRL_ID,
    co::BN::CLICKED,
    move || -> w::AnyResult<()> {
        println!("Button clicked!");
        Ok(())
    },
);

fn wm_notify<F>( &self, id_from: impl Into<u16>, code: impl Into<NmhdrCode>, func: F, )
where F: Fn(Notify<'_>) -> AnyResult<isize> + 'static,

WM_NOTIFY message, for specific ID and notification code.

Instead of using this event, you should always prefer the specific notifications, which will give you the correct notification struct. This generic method should be used only when you have a custom, non-standard window notification.

use winsafe::{self as w, prelude::*, co, gui};

let wnd: gui::WindowMain; // initialized somewhere

const CTRL_ID: u16 = 1010;

wnd.on().wm_notify(
    CTRL_ID,
    co::NM::DBLCLK,
    move |_| -> w::AnyResult<isize> {
        println!("Status bar double clicked!");
        Ok(0)
    },
);

Provided Methods

fn wm_command_acc_menu<F>(&self, ctrl_id: impl Into<u16> + Copy, func: F)
where F: Fn() -> AnyResult<()> + 'static,

WM_COMMAND message, handling both CMD::Accelerator and CMD::Menu, for a specific command ID.

Ideal to be used with menu commands whose IDs are shared with accelerators, like menu items.

Examples

use winsafe::{self as w, prelude::*, co, gui};

let wnd: gui::WindowMain; // initialized somewhere

const CTRL_ID: u16 = 1010;

wnd.on().wm_command_acc_menu(
    CTRL_ID,
    move || -> w::AnyResult<()> {
        println!("Hello!");
        Ok(())
    },
);

fn wm_create<F>(&self, func: F)
where F: Fn(Create<'_, '_, '_>) -> AnyResult<i32> + 'static,

WM_CREATE message, sent only to non-dialog windows. Dialog windows must handle wm_init_dialog instead.

Examples

use winsafe::{self as w, prelude::*, gui, msg};

let wnd: gui::WindowMain; // initialized somewhere

wnd.on().wm_create(
    move |p: msg::wm::Create| -> w::AnyResult<i32> {
        println!("Client area: {}x{}",
            p.createstruct.cx,
            p.createstruct.cy,
        );
        Ok(0)
    },
);

fn wm_init_dialog<F>(&self, func: F)
where F: Fn(InitDialog) -> AnyResult<bool> + 'static,

WM_INITDIALOG message, sent only to dialog windows. Non-dialog windows must handle wm_create instead.

Return true to set the focus to the first control in the dialog.

Examples

use winsafe::{self as w, prelude::*, gui, msg};

let wnd: gui::WindowMain; // initialized somewhere

wnd.on().wm_init_dialog(
    move |p: msg::wm::InitDialog| -> w::AnyResult<bool> {
        println!("Focused HWND: {}", p.hwnd_focus);
        Ok(true)
    },
);

fn wm_nc_create<F>(&self, func: F)
where F: Fn(NcCreate<'_, '_, '_>) -> AnyResult<bool> + 'static,

WM_NCCREATE message.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.