winsafe::prelude

Trait GuiEventsPropSheetPage

pub trait GuiEventsPropSheetPage: GuiEventsParent {
    // Provided methods
    fn psn_apply<F>(&self, func: F)
       where F: Fn() -> AnyResult<PSNRET> + 'static { ... }
    fn psn_get_object<F>(&self, func: F)
       where F: Fn(&NMOBJECTNOTIFY<'_>) -> AnyResult<()> + 'static { ... }
    fn psn_help<F>(&self, func: F)
       where F: Fn(&PSHNOTIFY) -> AnyResult<()> + 'static { ... }
    fn psn_kill_active<F>(&self, func: F)
       where F: Fn(&PSHNOTIFY) -> AnyResult<bool> + 'static { ... }
    fn psn_query_cancel<F>(&self, func: F)
       where F: Fn(&PSHNOTIFY) -> AnyResult<bool> + 'static { ... }
    fn psn_query_initial_focus<F>(&self, func: F)
       where F: Fn(&PSHNOTIFY) -> AnyResult<Option<HWND>> + 'static { ... }
    fn psn_reset<F>(&self, func: F)
       where F: Fn(&PSHNOTIFY) -> AnyResult<()> + 'static { ... }
    fn psn_set_active<F>(&self, func: F)
       where F: Fn(&PSHNOTIFY) -> AnyResult<i32> + 'static { ... }
    fn psn_translate_accelerator<F>(&self, func: F)
       where F: Fn() -> AnyResult<PSNRET> + 'static { ... }
    fn psn_wiz_back<F>(&self, func: F)
       where F: Fn(&PSHNOTIFY) -> AnyResult<i32> + 'static { ... }
    fn psn_wiz_finish<F>(&self, func: F)
       where F: Fn(&PSHNOTIFY) -> AnyResult<bool> + 'static { ... }
    fn psn_wiz_next<F>(&self, func: F)
       where F: Fn(&PSHNOTIFY) -> AnyResult<i32> + 'static { ... }
}

Available on crate feature gui only.

This trait is enabled with the gui feature, and exposes property sheet page control notifications.

These event methods are just proxies to the GuiEventsParent of the parent window, who is the real responsible for the child event handling.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn psn_apply<F>(&self, func: F)
where F: Fn() -> AnyResult<PSNRET> + 'static,

PSN_APPLY notification.

fn psn_get_object<F>(&self, func: F)
where F: Fn(&NMOBJECTNOTIFY<'_>) -> AnyResult<()> + 'static,

PSN_GETOBJECT notification.

fn psn_help<F>(&self, func: F)
where F: Fn(&PSHNOTIFY) -> AnyResult<()> + 'static,

PSN_HELP notification.

fn psn_kill_active<F>(&self, func: F)
where F: Fn(&PSHNOTIFY) -> AnyResult<bool> + 'static,

PSN_KILLACTIVE notification.

fn psn_query_cancel<F>(&self, func: F)
where F: Fn(&PSHNOTIFY) -> AnyResult<bool> + 'static,

PSN_QUERYCANCEL notification.

fn psn_query_initial_focus<F>(&self, func: F)
where F: Fn(&PSHNOTIFY) -> AnyResult<Option<HWND>> + 'static,

PSN_QUERYINITIALFOCUS notification.

fn psn_reset<F>(&self, func: F)
where F: Fn(&PSHNOTIFY) -> AnyResult<()> + 'static,

PSN_RESET notification.

fn psn_set_active<F>(&self, func: F)
where F: Fn(&PSHNOTIFY) -> AnyResult<i32> + 'static,

PSN_SETACTIVE notification.

fn psn_translate_accelerator<F>(&self, func: F)
where F: Fn() -> AnyResult<PSNRET> + 'static,

PSN_TRANSLATEACCELERATOR notification.

fn psn_wiz_back<F>(&self, func: F)
where F: Fn(&PSHNOTIFY) -> AnyResult<i32> + 'static,

PSN_WIZBACK notification.

fn psn_wiz_finish<F>(&self, func: F)
where F: Fn(&PSHNOTIFY) -> AnyResult<bool> + 'static,

PSN_WIZFINISH notification.

fn psn_wiz_next<F>(&self, func: F)
where F: Fn(&PSHNOTIFY) -> AnyResult<i32> + 'static,

PSN_WIZNEXT notification.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.