winsafe::prelude

Trait GuiEventsRadioGroup

pub trait GuiEventsRadioGroup: GuiEvents {
    // Provided methods
    fn bn_clicked<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn bn_dbl_clk<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn bn_kill_focus<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn bn_set_focus<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
}

Available on crate feature gui only.

This trait is enabled with the gui feature, and exposes button control notifications for a RadioGroup.

These event methods are just proxies to the GuiEventsParent of the parent window, who is the real responsible for the child event handling.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn bn_clicked<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

BN_CLICKED command notification for all radio buttons in the group.

Sent when the user clicks a button.

Examples

use winsafe::{self as w, prelude::*, gui};

let wnd: gui::WindowMain; // initialized somewhere
let radios: gui::RadioGroup;

radios.on().bn_clicked({
    let radios = radios.clone();
    move || -> w::AnyResult<()> {
        println!("Selected {}",
            radios.selected().unwrap()
                .hwnd().GetWindowText()?,
        );
        Ok(())
    }
});

fn bn_dbl_clk<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

BN_DBLCLK command notification for all radio buttons in the group.

Sent when the user double-clicks a button. This notification code is sent automatically for BS::USERBUTTON, BS::RADIOBUTTON, and BS::OWNERDRAW buttons. Other button types send only if they have the BS::NOTIFY style.

fn bn_kill_focus<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

BN_KILLFOCUS command notification for all radio buttons in the group.

Sent when a button loses the keyboard focus. The button must have the BS::NOTIFY style to send this notification code.

fn bn_set_focus<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

BN_SETFOCUS command notification for all radio buttons in the group.

Sent when a button receives the keyboard focus. The button must have the BS::NOTIFY style to send this notification code.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.