winsafe::prelude

Trait GuiEventsTrackbar

pub trait GuiEventsTrackbar: GuiEvents {
    // Provided methods
    fn trbn_thumb_pos_changing<F>(&self, func: F)
       where F: Fn(&NMTRBTHUMBPOSCHANGING) -> AnyResult<()> + 'static { ... }
    fn wm_h_scroll<F>(&self, func: F)
       where F: Fn(HScroll) -> AnyResult<()> + 'static { ... }
    fn wm_v_scroll<F>(&self, func: F)
       where F: Fn(VScroll) -> AnyResult<()> + 'static { ... }
    fn nm_custom_draw<F>(&self, func: F)
       where F: Fn(&NMCUSTOMDRAW) -> AnyResult<CDRF> + 'static { ... }
    fn nm_released_capture<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
}

Available on crate feature gui only.

This trait is enabled with the gui feature, and exposes trackbar control notifications.

These event methods are just proxies to the GuiEventsParent of the parent window, who is the real responsible for the child event handling.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn trbn_thumb_pos_changing<F>(&self, func: F)
where F: Fn(&NMTRBTHUMBPOSCHANGING) -> AnyResult<()> + 'static,

TRBN_THUMBPOSCHANGING notification.

fn wm_h_scroll<F>(&self, func: F)
where F: Fn(HScroll) -> AnyResult<()> + 'static,

WM_HSCROLL notification.

fn wm_v_scroll<F>(&self, func: F)
where F: Fn(VScroll) -> AnyResult<()> + 'static,

WM_VSCROLL notification.

fn nm_custom_draw<F>(&self, func: F)
where F: Fn(&NMCUSTOMDRAW) -> AnyResult<CDRF> + 'static,

NM_CUSTOMDRAW notification.

fn nm_released_capture<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

NM_RELEASEDCAPTURE notification.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.