winsafe::prelude

Trait GuiEventsTreeView

pub trait GuiEventsTreeView: GuiEvents {

    // Provided methods
    fn tvn_delete_item<F>(&self, func: F)
       where F: Fn(&NMTREEVIEW<'_, '_>) -> AnyResult<()> + 'static { ... }
    fn tvn_item_changed<F>(&self, func: F)
       where F: Fn(&NMTREEVIEW<'_, '_>) -> AnyResult<()> + 'static { ... }
    fn tvn_item_changing<F>(&self, func: F)
       where F: Fn(&NMTREEVIEW<'_, '_>) -> AnyResult<bool> + 'static { ... }
    fn tvn_item_expanded<F>(&self, func: F)
       where F: Fn(&NMTREEVIEW<'_, '_>) -> AnyResult<()> + 'static { ... }
    fn tvn_item_expanding<F>(&self, func: F)
       where F: Fn(&NMTREEVIEW<'_, '_>) -> AnyResult<bool> + 'static { ... }
    fn tvn_sel_changed<F>(&self, func: F)
       where F: Fn(&NMTREEVIEW<'_, '_>) -> AnyResult<()> + 'static { ... }
    fn tvn_sel_changing<F>(&self, func: F)
       where F: Fn(&NMTREEVIEW<'_, '_>) -> AnyResult<bool> + 'static { ... }
    fn nm_click<F>(&self, func: F)
       where F: Fn() -> AnyResult<i32> + 'static { ... }
    fn nm_custom_draw<F>(&self, func: F)
       where F: Fn(&mut NMTVCUSTOMDRAW) -> AnyResult<CDRF> + 'static { ... }
    fn nm_dbl_clk<F>(&self, func: F)
       where F: Fn() -> AnyResult<i32> + 'static { ... }
    fn nm_kill_focus<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn nm_r_click<F>(&self, func: F)
       where F: Fn() -> AnyResult<i32> + 'static { ... }
    fn nm_r_dbl_clk<F>(&self, func: F)
       where F: Fn() -> AnyResult<i32> + 'static { ... }
    fn nm_return<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn nm_set_cursor<F>(&self, func: F)
       where F: Fn(&NMMOUSE) -> AnyResult<i32> + 'static { ... }
    fn nm_set_focus<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }

}

Available on crate feature gui only.

This trait is enabled with the gui feature, and exposes tree view control notifications.

These event methods are just proxies to the GuiEventsParent of the parent window, who is the real responsible for the child event handling.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn tvn_delete_item<F>(&self, func: F)
where F: Fn(&NMTREEVIEW<'_, '_>) -> AnyResult<()> + 'static,

TVN_DELETEITEM notification.

fn tvn_item_changed<F>(&self, func: F)
where F: Fn(&NMTREEVIEW<'_, '_>) -> AnyResult<()> + 'static,

TVN_ITEMCHANGED notification.

fn tvn_item_changing<F>(&self, func: F)
where F: Fn(&NMTREEVIEW<'_, '_>) -> AnyResult<bool> + 'static,

TVN_ITEMCHANGING notification.

fn tvn_item_expanded<F>(&self, func: F)
where F: Fn(&NMTREEVIEW<'_, '_>) -> AnyResult<()> + 'static,

TVN_ITEMEXPANDED notification.

fn tvn_item_expanding<F>(&self, func: F)
where F: Fn(&NMTREEVIEW<'_, '_>) -> AnyResult<bool> + 'static,

TVN_ITEMEXPANDING notification.

fn tvn_sel_changed<F>(&self, func: F)
where F: Fn(&NMTREEVIEW<'_, '_>) -> AnyResult<()> + 'static,

TVN_SELCHANGED notification.

Examples

use winsafe::{self as w, prelude::*, gui, co};

let wnd: gui::WindowMain; // initialized somewhere
let tree: gui::TreeView;

tree.on().tvn_sel_changed(
    move |p: &w::NMTREEVIEW| -> w::AnyResult<()> {
        println!(
            "Old item: {} - new item: {} - action: {}",
            p.itemOld.pszText().unwrap(),
            p.itemNew.pszText().unwrap(),
            unsafe { co::TVC::from_raw(p.action) },
        );
        Ok(())
    },
);

fn tvn_sel_changing<F>(&self, func: F)
where F: Fn(&NMTREEVIEW<'_, '_>) -> AnyResult<bool> + 'static,

TVN_SELCHANGING notification.

fn nm_click<F>(&self, func: F)
where F: Fn() -> AnyResult<i32> + 'static,

NM_CLICK notification.

fn nm_custom_draw<F>(&self, func: F)
where F: Fn(&mut NMTVCUSTOMDRAW) -> AnyResult<CDRF> + 'static,

NM_CUSTOMDRAW notification.

fn nm_dbl_clk<F>(&self, func: F)
where F: Fn() -> AnyResult<i32> + 'static,

NM_DBLCLK notification.

fn nm_kill_focus<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

NM_KILLFOCUS notification.

fn nm_r_click<F>(&self, func: F)
where F: Fn() -> AnyResult<i32> + 'static,

NM_RCLICK notification.

fn nm_r_dbl_clk<F>(&self, func: F)
where F: Fn() -> AnyResult<i32> + 'static,

NM_RDBLCLK notification.

fn nm_return<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

NM_RETURN notification.

fn nm_set_cursor<F>(&self, func: F)
where F: Fn(&NMMOUSE) -> AnyResult<i32> + 'static,

NM_MOUSE notification.

fn nm_set_focus<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

NM_SETFOCUS notification.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.