winsafe::prelude

Trait GuiEventsUpDown

pub trait GuiEventsUpDown: GuiEvents {
    // Provided methods
    fn nm_released_capture<F>(&self, func: F)
       where F: Fn() -> AnyResult<()> + 'static { ... }
    fn udn_delta_pos<F>(&self, func: F)
       where F: Fn(&NMUPDOWN) -> AnyResult<i32> + 'static { ... }
}

Available on crate feature gui only.

This trait is enabled with the gui feature, and exposes up-down control notifications.

These event methods are just proxies to the GuiEventsParent of the parent window, who is the real responsible for the child event handling.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn nm_released_capture<F>(&self, func: F)
where F: Fn() -> AnyResult<()> + 'static,

NM_RELEASEDCAPTURE notification.

fn udn_delta_pos<F>(&self, func: F)
where F: Fn(&NMUPDOWN) -> AnyResult<i32> + 'static,

UDN_DELTAPOS notification.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.