Trait winsafe::prelude::GuiParent

pub trait GuiParent: GuiWindow {
    // Required methods
    fn on(&self) -> &WindowEventsAll;
    unsafe fn as_base(&self) -> *mut c_void;
}

Available on crate features kernel and gui only.

Any window which can host child controls.

Required Methods

fn on(&self) -> &WindowEventsAll

Exposes methods to handle the basic window messages, plus timer and native control notifications.

Panics

Panics if the window is already created. Events must be set before window creation.

unsafe fn as_base(&self) -> *mut c_void

Returns a pointer to the inner base window structure, declared internally in the library.

Safety

Do not use this method – it’s for internal use of the library only.