winsafe::prelude
Trait GuiParentPopup

pub trait GuiParentPopup: GuiParent {
    // Provided method
    fn close(&self) { ... }
}

Available on crate features kernel and gui only.

A closeable popup parent window.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn close(&self)

Closes the window by posting a WM_CLOSE message. This is the safest way to close any popup window, because you’re able to process the wm_close event, just like if the user clicked the window “X” button.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.