Trait winsafe::prelude::KernelHfindfile

pub trait KernelHfindfile: Handle {
    fn iter<'a>(
        path_and_pattern: &'a str
    ) -> Box<dyn Iterator<Item = WinResult<String>> + 'a> { ... }

    fn FindClose(self) -> WinResult<()> { ... }

    fn FindFirstFile(
        file_name: &str, 
        wfd: &mut WIN32_FIND_DATA
    ) -> WinResult<(HFINDFILE, bool)> { ... }

    fn FindNextFile(self, wfd: &mut WIN32_FIND_DATA) -> WinResult<bool> { ... }
}

This is supported on crate feature kernel only.

HFILEMAPVIEW methods from kernel feature.

Provided methods

fn iter<'a>(
path_and_pattern: &'a str
) -> Box<dyn Iterator<Item = WinResult<String>> + 'a>

Returns an iterator over the found items, by calling HFINDFILE::FindFirstFile, then subsequent HFINDFILE::FindNextFile, and finally freeing the resource by calling HFINDFILE::FindClose.

Examples

Enumerating all TXT files in a directory:

use winsafe::prelude::*;
use winsafe::HFINDFILE;

for file_path in HFINDFILE::iter("C:\\Temp\\*.txt") {
    let file_path = file_path?;
    println!("File: {}", file_path);
}

Collecting the strings into a Vec:

use winsafe::prelude::*;
use winsafe::{HFINDFILE, WinResult};

let file_paths = HFINDFILE::iter("")
    .collect::<WinResult<Vec<_>>>()?;

fn FindClose(self) -> WinResult<()>

FindClose method.

fn FindFirstFile(
file_name: &str,
wfd: &mut WIN32_FIND_DATA
) -> WinResult<(HFINDFILE, bool)>

FindFirstFile static method.

Note: Must be paired with an HFINDFILE::FindClose call.

This method is rather tricky, consider using HFINDFILE::iter.

fn FindNextFile(self, wfd: &mut WIN32_FIND_DATA) -> WinResult<bool>

FindNextFile method.

This method is rather tricky, consider using HFINDFILE::iter.