Trait winsafe::prelude::Main

pub trait Main: Parent {
    fn run_main(&self, cmd_show: Option<SW>) -> ErrResult<i32>;
}

The main window of an application.

This is a sealed trait, which cannot be implemented outside winsafe.

Required methods

fn run_main(&self, cmd_show: Option<SW>) -> ErrResult<i32>

Physically creates the window, then runs the main application loop. This method will block until the window is closed.

The cmd_show parameter defaults to co::SW::SHOW.

Panics

Panics if the window is already created.