winsafe::prelude

Trait MsgSendRecv

pub trait MsgSendRecv: MsgSend {
    // Required method
    unsafe fn from_generic_wm(parm: Wm) -> Self;
}

Available on crate feature user only.

This trait is enabled with the gui feature, and is implemented by the parameters of messages which can be sent and handled (received).

Allows the conversion from and to the generic Wm parameters, and also defines the return type of the message.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Required Methods

unsafe fn from_generic_wm(parm: Wm) -> Self

Unmarshaling method which converts the generic Wm parameters struct into the specific message struct.

Safety

Message parameters often involve pointers and require casts, make sure the conversions are correct.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".