Trait winsafe::prelude::NativeControl

pub trait NativeControl: Child {
    fn on_subclass(&self) -> &WindowEvents;
}

Any native child control.

Required methods

fn on_subclass(&self) -> &WindowEvents

Exposes the subclass events. If at least one event exists, the control will be subclassed.

Note: Subclassing may impact performance, use with care.

Panics

Panics if the control or the parent window are already created. Events must be set before control and parent window creation.