Trait winsafe::prelude::ShellHdrop

pub trait ShellHdrop: Handle {
    fn iter<'a>(
        &'a self
    ) -> WinResult<Box<dyn Iterator<Item = WinResult<String>> + 'a>> { ... }

    fn DragFinish(self) { ... }

    fn DragQueryFile(
        self, 
        ifile: Option<u32>, 
        buf: Option<&mut WString>
    ) -> WinResult<u32> { ... }

    fn DragQueryPoint(self) -> (POINT, bool) { ... }
}

This is supported on crate feature shell only.

HDROP methods from shell feature.

Provided methods

fn iter<'a>(
&'a self
) -> WinResult<Box<dyn Iterator<Item = WinResult<String>> + 'a>>

Returns an iterator over the dropped files by calling HDROP::DragQueryFile consecutively, then frees the handle by calling HDROP::DragFinish.

Examples

Iterating over the strings:

use winsafe::prelude::*;
use winsafe::HDROP;

let hdrop: HDROP; // initialized somewhere

for file_path in hdrop.iter()? {
    let file_path = file_path?;
    println!("File: {}", file_path);
}

Collecting the strings into a Vec:

use winsafe::prelude::*;
use winsafe::{HDROP, WinResult};

let hdrop: HDROP; // initialized somewhere

let file_paths = hdrop.iter()?
    .collect::<WinResult<Vec<_>>>()?;

fn DragFinish(self)

DragFinish method.

Prefer using HDROP::iter, which calls DragFinish automatically.

fn DragQueryFile(
self,
ifile: Option<u32>,
buf: Option<&mut WString>
) -> WinResult<u32>

DragQueryFile method.

This method is rather tricky, consider using HDROP::iter.

fn DragQueryPoint(self) -> (POINT, bool)

DragQueryPoint method.

Returns the coordinates and whether the drop occurred in the client area of the window.