winsafe::prelude

Trait SystemError

pub trait SystemError: Into<u32> {
    // Provided method
    fn FormatMessage(self) -> String { ... }
}

Available on crate feature kernel only.

This trait is enabled with the kernel feature, and is implemented by all system error types which can be formatted with FormatMessage, exhibiting a description string provided by the OS.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn FormatMessage(self) -> String

Returns the textual description of the system error, by calling FormatMessage. function.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.