Trait winsafe::prelude::UserHmenu

pub trait UserHmenu: Handle {

    fn AppendMenu(
        self, 
        flags: MF, 
        new_item: IdMenu, 
        content: BmpPtrStr
    ) -> WinResult<()> { ... }

    fn AppendMenuEnum(self, items: &[MenuEnum<'_>]) -> WinResult<()> { ... }

    fn CheckMenuItem(self, id_or_pos: IdPos, check: bool) -> WinResult<MF> { ... }

    fn CheckMenuRadioItem(
        self, 
        first: IdPos, 
        last: IdPos, 
        check: IdPos
    ) -> WinResult<()> { ... }

    fn CreateMenu() -> WinResult<HMENU> { ... }

    fn CreatePopupMenu() -> WinResult<HMENU> { ... }

    fn DeleteMenu(self, id_or_pos: IdPos) -> WinResult<()> { ... }

    fn DestroyMenu(self) -> WinResult<()> { ... }

    fn EnableMenuItem(self, id_or_pos: IdPos, enable: bool) -> WinResult<MF> { ... }

    fn GetMenuDefaultItem(self, by_pos: bool, flags: GMDI) -> WinResult<IdPos> { ... }

    fn GetMenuInfo(self, mi: &mut MENUINFO) -> WinResult<()> { ... }

    fn GetMenuItemCount(self) -> WinResult<u32> { ... }

    fn GetMenuItemID(self, nPos: i32) -> Option<i32> { ... }

    fn GetMenuItemInfo(
        self, 
        id_or_pos: IdPos, 
        mii: &mut MENUITEMINFO
    ) -> WinResult<()> { ... }

    fn GetMenuState(self, id_or_pos: IdPos) -> WinResult<MF> { ... }

    fn GetMenuString(self, id_or_pos: IdPos) -> WinResult<String> { ... }

    fn GetSubMenu(self, pos: u32) -> Option<HMENU> { ... }

    fn InsertMenuItem(
        self, 
        id_or_pos: IdPos, 
        mii: &MENUITEMINFO
    ) -> WinResult<()> { ... }

    fn IsMenu(self) -> bool { ... }

    fn RemoveMenu(self, id_or_pos: IdPos) -> WinResult<()> { ... }

    fn SetMenuDefaultItem(self, id_or_pos: IdPos) -> WinResult<()> { ... }

    fn SetMenuInfo(self, mi: &MENUINFO) -> WinResult<()> { ... }

    fn SetMenuItemBitmaps(
        self, 
        id_or_pos: IdPos, 
        hbmp_unchecked: Option<HBITMAP>, 
        hbmp_checked: Option<HBITMAP>
    ) -> WinResult<()> { ... }

    fn SetMenuItemInfo(
        self, 
        id_or_pos: IdPos, 
        mii: &MENUITEMINFO
    ) -> WinResult<()> { ... }

    fn TrackPopupMenu(
        self, 
        flags: TPM, 
        location: POINT, 
        hwnd: HWND
    ) -> WinResult<Option<i32>> { ... }

    fn TrackPopupMenuAtPoint(
        self, 
        pos: POINT, 
        hwnd_parent: HWND, 
        hwnd_coords_relative_to: HWND
    ) -> WinResult<()> { ... }

}

HMENU methods from user feature.

Provided methods

fn AppendMenu(
self,
flags: MF,
new_item: IdMenu,
content: BmpPtrStr
) -> WinResult<()>

AppendMenu method.

This method is rather tricky, consider using HMENU::AppendMenuEnum.

fn AppendMenuEnum(self, items: &[MenuEnum<'_>]) -> WinResult<()>

A more convenient HMENU::AppendMenu.

Examples

Adding multiple entries at once, with their command IDs:

use winsafe::prelude::*;
use winsafe::{HMENU, MenuEnum, seq_ids};

seq_ids! {
    ID_FILE_OPEN = 2001;
    ID_FILE_SAVE
    ID_FILE_EXIT
}

let hmenu: HMENU; // initialized somewhere

hmenu.AppendMenuEnum(&[
    MenuEnum::Entry(ID_FILE_OPEN, "&Open"),
    MenuEnum::Entry(ID_FILE_OPEN, "&Save"),
    MenuEnum::Separator,
    MenuEnum::Entry(ID_FILE_EXIT, "E&xit"),
])?;

fn CheckMenuItem(self, id_or_pos: IdPos, check: bool) -> WinResult<MF>

CheckMenuItem method.

fn CheckMenuRadioItem(
self,
first: IdPos,
last: IdPos,
check: IdPos
) -> WinResult<()>

CheckMenuRadioItem method.

Panics

Panics if first, last and check don’t use the same enum field.

fn CreateMenu() -> WinResult<HMENU>

CreateMenu static method.

Note: If not attached to a window, must be paired with an HMENU::DestroyMenu call.

fn CreatePopupMenu() -> WinResult<HMENU>

CreatePopupMenu static method.

Note: If not attached to a window, must be paired with an HMENU::DestroyMenu call.

fn DeleteMenu(self, id_or_pos: IdPos) -> WinResult<()>

DeleteMenu method.

fn DestroyMenu(self) -> WinResult<()>

DestroyMenu method.

fn EnableMenuItem(self, id_or_pos: IdPos, enable: bool) -> WinResult<MF>

EnableMenuItem method.

Examples

Disabling a menu item:

use winsafe::prelude::*;
use winsafe::{HMENU, IdPos, seq_ids};

seq_ids! {
    ID_FILE_OPEN = 101;
}

let hmenu: HMENU; // initialized somewhere

hmenu.EnableMenuItem(
    IdPos::Id(ID_FILE_OPEN),
    false,
)?;

Disabling multiple menu items at once:

use winsafe::prelude::*;
use winsafe::{HMENU, IdPos, seq_ids};

seq_ids! {
    ID_FILE_OPEN = 201;
    ID_FILE_SAVE
}

let hmenu: HMENU; // initialized somewhere

[
    (ID_FILE_OPEN, "Open\tCtrl+O"),
    (ID_FILE_SAVE, "&Save"),
].iter()
    .map(|(id, txt)| hmenu.EnableMenuItem(
        IdPos::Id(*id),
        false,
    ))
    .collect::<Result<Vec<_>, _>>()?;

fn GetMenuDefaultItem(self, by_pos: bool, flags: GMDI) -> WinResult<IdPos>

GetMenuDefaultItem method.

fn GetMenuInfo(self, mi: &mut MENUINFO) -> WinResult<()>

GetMenuInfo method.

fn GetMenuItemCount(self) -> WinResult<u32>

GetMenuItemCount method.

fn GetMenuItemID(self, nPos: i32) -> Option<i32>

GetMenuItemID method.

fn GetMenuItemInfo(
self,
id_or_pos: IdPos,
mii: &mut MENUITEMINFO
) -> WinResult<()>

GetMenuItemInfo method.

fn GetMenuState(self, id_or_pos: IdPos) -> WinResult<MF>

GetMenuState method.

fn GetMenuString(self, id_or_pos: IdPos) -> WinResult<String>

GetMenuString method.

fn GetSubMenu(self, pos: u32) -> Option<HMENU>

GetSubMenu method.

fn InsertMenuItem(self, id_or_pos: IdPos, mii: &MENUITEMINFO) -> WinResult<()>

InsertMenuItem method.

fn IsMenu(self) -> bool

IsMenu method.

fn RemoveMenu(self, id_or_pos: IdPos) -> WinResult<()>

RemoveMenu method.

fn SetMenuDefaultItem(self, id_or_pos: IdPos) -> WinResult<()>

SetMenuDefaultItem method.

fn SetMenuInfo(self, mi: &MENUINFO) -> WinResult<()>

SetMenuInfo method.

fn SetMenuItemBitmaps(
self,
id_or_pos: IdPos,
hbmp_unchecked: Option<HBITMAP>,
hbmp_checked: Option<HBITMAP>
) -> WinResult<()>

SetMenuItemBitmaps method.

fn SetMenuItemInfo(self, id_or_pos: IdPos, mii: &MENUITEMINFO) -> WinResult<()>

SetMenuItemInfo method.

fn TrackPopupMenu(
self,
flags: TPM,
location: POINT,
hwnd: HWND
) -> WinResult<Option<i32>>

TrackPopupMenu method.

Note: If you just want to display a popup menu, consider the simpler HMENU::TrackPopupMenuAtPoint.

fn TrackPopupMenuAtPoint(
self,
pos: POINT,
hwnd_parent: HWND,
hwnd_coords_relative_to: HWND
) -> WinResult<()>

Shows the popup menu anchored at the given coordinates using TrackPopupMenu, and performs other needed operations.

This method will block until the menu disappears.