winsafe::prelude

Trait VariableSized

pub trait VariableSized { }

Available on crate feature kernel only.

A struct natively defined with its last element being an 1-element array. The struct is dynamically allocated to accomodate a variable number of elements, whose amount is stored in a specific field. The elements are accessed by reaching the past-struct memory room.

Structs of this kind are managed through a guard.