winsafe::prelude

Trait comctl_gdi_Himagelist

pub trait comctl_gdi_Himagelist: Handle {
    // Provided method
    fn DrawIndirect(&self, imldp: &IMAGELISTDRAWPARAMS) -> HrResult<()> { ... }
}

Available on crate features comctl and gdi and kernel only.

This trait is enabled with comctl and gdi features, and provides methods for HIMAGELIST.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn DrawIndirect(&self, imldp: &IMAGELISTDRAWPARAMS) -> HrResult<()>

ImageList_DrawIndirect function.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.