Trait winsafe::prelude::comctl_ole_Hwnd

pub trait comctl_ole_Hwnd: Handle {
    fn InitializeFlatSB(self) -> HrResult<()> { ... }
    fn TaskDialog(
        self,
        hinstance: Option<HINSTANCE>,
        window_title: Option<&str>,
        main_instruction: Option<&str>,
        content: Option<&str>,
        common_buttons: TDCBF,
        icon: IdTdiconStr
    ) -> HrResult<DLGID> { ... }
    fn UninitializeFlatSB(self) -> HrResult<()> { ... }
}

Available on crate features comctl and ole only.

This trait is enabled with comctl and ole features, and provides methods for HWND.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn InitializeFlatSB(self) -> HrResult<()>

InitializeFlatSB method.

fn TaskDialog(
self,
hinstance: Option<HINSTANCE>,
window_title: Option<&str>,
main_instruction: Option<&str>,
content: Option<&str>,
common_buttons: TDCBF,
icon: IdTdiconStr
) -> HrResult<DLGID>

TaskDialog method.

Unless you need something specific, consider using the task_dlg high-level abstractions.

If you need more customization, see the TaskDialogIndirect function.

Examples

An information message with just an OK button:

use winsafe::prelude::*;
use winsafe::{co, HWND, IdTdiconStr};

let hwnd: HWND; // initialized somewhere

hwnd.TaskDialog(
    None,
    Some("My app name"),
    Some("Operation successful"),
    Some("The operation completed successfully."),
    co::TDCBF::OK,
    IdTdiconStr::Tdicon(co::TD_ICON::INFORMATION),
)?;

Prompt the user to click OK or Cancel upon a question:

use winsafe::prelude::*;
use winsafe::{co, HWND, IdTdiconStr};

let hwnd: HWND; // initialized somewhere

let answer = hwnd.TaskDialog(
    None,
    Some("My app name"),
    Some("File modified"),
    Some("The file has been modified.\nProceed closing the application?"),
    co::TDCBF::OK | co::TDCBF::CANCEL,
    IdTdiconStr::Tdicon(co::TD_ICON::WARNING),
)?;

if answer == co::DLGID::OK {
    println!("User clicked OK.");
}

fn UninitializeFlatSB(self) -> HrResult<()>

UninitializeFlatSB method.