winsafe::prelude
Trait gdi_Hbrush

pub trait gdi_Hbrush: Handle {
    // Provided methods
    fn from_sys_color(color: COLOR) -> HBRUSH { ... }
    fn CreateBrushIndirect(
        lb: &LOGBRUSH,
    ) -> SysResult<DeleteObjectGuard<HBRUSH>> { ... }
    fn CreateHatchBrush(
        hatch: HS,
        color: COLORREF,
    ) -> SysResult<DeleteObjectGuard<HBRUSH>> { ... }
    fn CreatePatternBrush(
        hbmp: &HBITMAP,
    ) -> SysResult<DeleteObjectGuard<HBRUSH>> { ... }
    fn CreateSolidBrush(color: COLORREF) -> SysResult<DeleteObjectGuard<HBRUSH>> { ... }
    fn GetObject(&self, pv: &mut LOGBRUSH) -> SysResult<()> { ... }
    fn GetStockObject(sb: STOCK_BRUSH) -> SysResult<HBRUSH> { ... }
    fn GetSysColorBrush(index: COLOR) -> SysResult<HBRUSH> { ... }
    fn UnrealizeObject(&self) -> SysResult<()> { ... }
}

Available on crate features kernel and gdi only.

This trait is enabled with the gdi feature, and provides methods for HBRUSH.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn from_sys_color(color: COLOR) -> HBRUSH

Creates a brush with the given system color.

Note: This should be used only to initialize the WNDCLASSEX’s hbrBackground field. Any other use will yield an invalid handle.

fn CreateBrushIndirect(lb: &LOGBRUSH) -> SysResult<DeleteObjectGuard<HBRUSH>>

CreateBrushIndirect function.

fn CreateHatchBrush( hatch: HS, color: COLORREF, ) -> SysResult<DeleteObjectGuard<HBRUSH>>

CreateHatchBrush function.

fn CreatePatternBrush(hbmp: &HBITMAP) -> SysResult<DeleteObjectGuard<HBRUSH>>

CreatePatternBrush function.

fn CreateSolidBrush(color: COLORREF) -> SysResult<DeleteObjectGuard<HBRUSH>>

CreateSolidBrush function.

fn GetObject(&self, pv: &mut LOGBRUSH) -> SysResult<()>

GetObject function.

Examples

use winsafe::{self as w, prelude::*};

let hbr: w::HBRUSH; // initialized somewhere

let mut brush = w::LOGBRUSH::default();
hbr.GetObject(&mut brush)?;

fn GetStockObject(sb: STOCK_BRUSH) -> SysResult<HBRUSH>

GetStockObject function.

fn GetSysColorBrush(index: COLOR) -> SysResult<HBRUSH>

GetSysColorBrush function.

fn UnrealizeObject(&self) -> SysResult<()>

UnrealizeObject function.

Object Safety

This trait is not object safe.