winsafe::prelude

Trait gdi_Hrgn

pub trait gdi_Hrgn: Handle {
    // Provided methods
    fn CreateRectRgn(bounds: RECT) -> SysResult<DeleteObjectGuard<HRGN>> { ... }
    fn CreateRectRgnIndirect(rc: RECT) -> SysResult<DeleteObjectGuard<HRGN>> { ... }
    fn CreateRoundRectRgn(
        bounds: RECT,
        size: SIZE,
    ) -> SysResult<DeleteObjectGuard<HRGN>> { ... }
    fn CombineRgn(
        &self,
        src1: &HRGN,
        src2: &HRGN,
        mode: RGN,
    ) -> SysResult<REGION> { ... }
    fn EqualRgn(&self, other: &HRGN) -> bool { ... }
    fn OffsetClipRgn(&self, x: i32, y: i32) -> SysResult<REGION> { ... }
    fn OffsetRgn(&self, x: i32, y: i32) -> SysResult<REGION> { ... }
    fn PtInRegion(&self, x: i32, y: i32) -> bool { ... }
    fn RectInRegion(&self, rc: RECT) -> bool { ... }
}

Available on crate features kernel and gdi only.

This trait is enabled with the gdi feature, and provides methods for HRGN.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn CreateRectRgn(bounds: RECT) -> SysResult<DeleteObjectGuard<HRGN>>

CreateRectRgn function.

fn CreateRectRgnIndirect(rc: RECT) -> SysResult<DeleteObjectGuard<HRGN>>

CreateRectRgnIndirect function.

fn CreateRoundRectRgn( bounds: RECT, size: SIZE, ) -> SysResult<DeleteObjectGuard<HRGN>>

CreateRoundRectRgn function.

fn CombineRgn(&self, src1: &HRGN, src2: &HRGN, mode: RGN) -> SysResult<REGION>

CombineRgn function.

Examples

Creating a clipping region with a square hole in it:

use winsafe::{self as w, prelude::*, co};

let hdc: w::HDC; // initialized somewhere

let rc_hole = w::RECT { left: 0, top: 0, right: 100, bottom: 100 };
let hrgn_hole = w::HRGN::CreateRectRgnIndirect(rc_hole)?;

let hrgn_clip = w::HRGN::CreateRectRgnIndirect(
    w::InflateRect(rc_hole, 10, 10)?,
)?;
hrgn_clip.CombineRgn(&hrgn_clip, &hrgn_hole, co::RGN::DIFF)?;

hdc.SelectClipRgn(&hrgn_clip)?;

fn EqualRgn(&self, other: &HRGN) -> bool

EqualRgn function.

fn OffsetClipRgn(&self, x: i32, y: i32) -> SysResult<REGION>

OffsetClipRgn function.

fn OffsetRgn(&self, x: i32, y: i32) -> SysResult<REGION>

OffsetRgn function.

fn PtInRegion(&self, x: i32, y: i32) -> bool

PtInRegion function.

fn RectInRegion(&self, rc: RECT) -> bool

RectInRegion function.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.