Trait winsafe::prelude::gdi_mf_IMFVideoDisplayControl

pub trait gdi_mf_IMFVideoDisplayControl: mf_IMFVideoDisplayControl {
    // Provided method
    fn GetCurrentImage(&self) -> HrResult<(BITMAPINFOHEADER, Vec<u8>, i64)> { ... }
}

Available on crate features gdi and mf and kernel only.

This trait is enabled with gdi and mf features, and provides methods for IMFVideoDisplayControl.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn GetCurrentImage(&self) -> HrResult<(BITMAPINFOHEADER, Vec<u8>, i64)>

GetCurrentImage method.

Returns bitmap description, DIB bytes and time stamp.

Object Safety

This trait is not object safe.