Trait winsafe::prelude::kernel_Haccesstoken

pub trait kernel_Haccesstoken: Handle {
    // Provided methods
    fn AdjustTokenPrivileges(&self, new_state: DisabPriv<'_>) -> SysResult<()> { ... }
    fn CheckTokenCapability(
        &self,
        capability_sid_to_check: &SID
    ) -> SysResult<bool> { ... }
    fn CheckTokenMembership(&self, sid_to_check: &SID) -> SysResult<bool> { ... }
    fn DuplicateToken(
        &self,
        level: SECURITY_IMPERSONATION
    ) -> SysResult<CloseHandleGuard<HACCESSTOKEN>> { ... }
    fn GetCurrentProcessToken() -> HACCESSTOKEN { ... }
    fn GetCurrentThreadEffectiveToken() -> HACCESSTOKEN { ... }
    unsafe fn GetTokenInformation<T>(
        &self,
        information_class: TOKEN_INFORMATION_CLASS,
        information: &mut T
    ) -> SysResult<()> { ... }
    fn ImpersonateLoggedOnUser(&self) -> SysResult<()> { ... }
    fn IsTokenRestricted(&self) -> SysResult<bool> { ... }
}

Available on crate feature kernel only.

This trait is enabled with the kernel feature, and provides methods for HACCESSTOKEN.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn AdjustTokenPrivileges(&self, new_state: DisabPriv<'_>) -> SysResult<()>

AdjustTokenPrivileges function.

Examples

use winsafe::{self as w, prelude::*, co};

let htoken = w::HPROCESS::GetCurrentProcess()
    .OpenProcessToken(co::TOKEN::ADJUST_PRIVILEGES | co::TOKEN::QUERY)?;

let luid = w::LookupPrivilegeValue(None, co::SE_PRIV::SHUTDOWN_NAME)?;

let privs = w::TOKEN_PRIVILEGES::new(&[
    w::LUID_AND_ATTRIBUTES::new(luid, co::SE_PRIV_ATTR::ENABLED),
]);

htoken.AdjustTokenPrivileges(w::DisabPriv::Privs(&privs))?;

fn CheckTokenCapability(&self, capability_sid_to_check: &SID) -> SysResult<bool>

CheckTokenCapability function.

fn CheckTokenMembership(&self, sid_to_check: &SID) -> SysResult<bool>

CheckTokenMembership function.

fn DuplicateToken( &self, level: SECURITY_IMPERSONATION ) -> SysResult<CloseHandleGuard<HACCESSTOKEN>>

DuplicateToken function.

fn GetCurrentProcessToken() -> HACCESSTOKEN

GetCurrentProcessToken function.

fn GetCurrentThreadEffectiveToken() -> HACCESSTOKEN

GetCurrentThreadEffectiveToken function.

unsafe fn GetTokenInformation<T>( &self, information_class: TOKEN_INFORMATION_CLASS, information: &mut T ) -> SysResult<()>

GetTokenInformation function.

Safety

Make sure the information type is the correct one, matching that in information_class.

Examples

Checking if the current process has elevated privileges:

use winsafe::{self as w, prelude::*, co};

let htoken = w::HPROCESS::GetCurrentProcess()
    .OpenProcessToken(co::TOKEN::QUERY)?;

let mut elevation = w::TOKEN_ELEVATION::default();
unsafe {
    htoken.GetTokenInformation(
        co::TOKEN_INFORMATION_CLASS::Elevation,
        &mut elevation,
     )?;
}
println!("Is elevated: {}", elevation.TokenIsElevated());

fn ImpersonateLoggedOnUser(&self) -> SysResult<()>

ImpersonateLoggedOnUser function.

fn IsTokenRestricted(&self) -> SysResult<bool>

IsTokenRestricted function.

Object Safety

This trait is not object safe.