Trait winsafe::prelude::kernel_Hfindfile

pub trait kernel_Hfindfile: Handle {
    fn iter<'a>(
        path_and_pattern: &'a str
    ) -> Box<dyn Iterator<Item = SysResult<String>> + 'a> { ... }
    fn FindClose(self) -> SysResult<()> { ... }
    fn FindFirstFile(
        file_name: &str,
        wfd: &mut WIN32_FIND_DATA
    ) -> SysResult<(HFINDFILE, bool)> { ... }
    fn FindNextFile(self, wfd: &mut WIN32_FIND_DATA) -> SysResult<bool> { ... }
}

Available on crate feature kernel only.

This trait is enabled with the kernel feature, and provides methods for HFINDFILE.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn iter<'a>(
path_and_pattern: &'a str
) -> Box<dyn Iterator<Item = SysResult<String>> + 'a>

Returns an iterator over the found items, by calling HFINDFILE::FindFirstFile, then subsequent HFINDFILE::FindNextFile, and finally freeing the resource by calling HFINDFILE::FindClose.

Examples

Enumerating all TXT files in a directory:

use winsafe::prelude::*;
use winsafe::HFINDFILE;

for file_path in HFINDFILE::iter("C:\\Temp\\*.txt") {
    let file_path = file_path?;
    println!("File: {}", file_path);
}

Collecting the strings into a Vec:

use winsafe::prelude::*;
use winsafe::{HFINDFILE, SysResult};

let file_paths = HFINDFILE::iter("")
    .collect::<SysResult<Vec<_>>>()?;

fn FindClose(self) -> SysResult<()>

FindClose method.

fn FindFirstFile(
file_name: &str,
wfd: &mut WIN32_FIND_DATA
) -> SysResult<(HFINDFILE, bool)>

FindFirstFile static method.

Note: Must be paired with an HFINDFILE::FindClose call.

This method is rather tricky, consider using HFINDFILE::iter.

fn FindNextFile(self, wfd: &mut WIN32_FIND_DATA) -> SysResult<bool>

FindNextFile method.

This method is rather tricky, consider using HFINDFILE::iter.