winsafe::prelude
Trait mf_IMFAttributes

pub trait mf_IMFAttributes: ole_IUnknown {

    // Provided methods
    fn Compare(
        &self,
        theirs: &impl mf_IMFAttributes,
        match_type: MF_ATTRIBUTES_MATCH,
    ) -> HrResult<bool> { ... }
    fn CompareItem(
        &self,
        guid_key: &GUID,
        value: &PROPVARIANT,
    ) -> HrResult<bool> { ... }
    fn CopyAllItems(&self, dest: &impl mf_IMFAttributes) -> HrResult<()> { ... }
    fn DeleteAllItems(&self) -> HrResult<()> { ... }
    fn DeleteItem(&self, guid_key: &GUID) -> HrResult<()> { ... }
    fn GetAllocatedBlob(&self, guid_key: &GUID) -> HrResult<Vec<u8>> { ... }
    fn GetAllocatedString(&self, guid_key: &GUID) -> HrResult<String> { ... }
    fn GetBlob(&self, guid_key: &GUID) -> HrResult<Vec<u8>> { ... }
    fn GetBlobSize(&self, guid_key: &GUID) -> HrResult<u32> { ... }
    fn GetCount(&self) -> HrResult<u32> { ... }
    fn GetDouble(&self, guid_key: &GUID) -> HrResult<f64> { ... }
    fn GetGUID(&self, guid_key: &GUID) -> HrResult<GUID> { ... }
    fn GetItem(&self, guid_key: &GUID) -> HrResult<PROPVARIANT> { ... }
    fn GetItemByIndex(&self, index: u32) -> HrResult<(GUID, PROPVARIANT)> { ... }
    fn GetItemType(&self, guid_key: &GUID) -> HrResult<MF_ATTRIBUTE> { ... }
    fn GetString(&self, guid_key: &GUID) -> HrResult<String> { ... }
    fn GetStringLength(&self, guid_key: &GUID) -> HrResult<u32> { ... }
    fn GetUINT32(&self, guid_key: &GUID) -> HrResult<u32> { ... }
    fn GetUINT64(&self, guid_key: &GUID) -> HrResult<u64> { ... }
    fn GetUnknown<T>(&self, guid_key: &GUID) -> HrResult<T>
       where T: ole_IUnknown { ... }
    fn LockStore(&self) -> HrResult<()> { ... }
    fn SetBlob(&self, guid_key: &GUID, buf: &[u8]) -> HrResult<()> { ... }
    fn SetDouble(&self, guid_key: &GUID, value: f64) -> HrResult<()> { ... }
    fn SetGUID(&self, guid_key: &GUID, value: &GUID) -> HrResult<()> { ... }
    fn SetItem(&self, guid_key: &GUID, value: &PROPVARIANT) -> HrResult<()> { ... }
    fn SetString(&self, guid_key: &GUID, value: &str) -> HrResult<()> { ... }
    fn SetUINT32(&self, guid_key: &GUID, value: u32) -> HrResult<()> { ... }
    fn SetUINT64(&self, guid_key: &GUID, value: u64) -> HrResult<()> { ... }
    fn SetUnknown(
        &self,
        guid_key: &GUID,
        value: &impl ole_IUnknown,
    ) -> HrResult<()> { ... }
    fn UnlockStore(&self) -> HrResult<()> { ... }

}

Available on crate features kernel and mf only.

This trait is enabled with the mf feature, and provides methods for IMFAttributes.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn Compare( &self, theirs: &impl mf_IMFAttributes, match_type: MF_ATTRIBUTES_MATCH, ) -> HrResult<bool>

IMFAttributes::Compare method.

fn CompareItem(&self, guid_key: &GUID, value: &PROPVARIANT) -> HrResult<bool>

IMFAttributes::CompareItem method.

fn CopyAllItems(&self, dest: &impl mf_IMFAttributes) -> HrResult<()>

IMFAttributes::CopyAllItems method.

fn DeleteAllItems(&self) -> HrResult<()>

IMFAttributes::DeleteAllItems method.

fn DeleteItem(&self, guid_key: &GUID) -> HrResult<()>

IMFAttributes::DeleteItem method.

fn GetAllocatedBlob(&self, guid_key: &GUID) -> HrResult<Vec<u8>>

IMFAttributes::GetAllocatedBlob method.

Note that this method allocates the buffer twice, whereas IMFAttributes::GetBlob allocates only once, thus being more efficient.

fn GetAllocatedString(&self, guid_key: &GUID) -> HrResult<String>

IMFAttributes::GetAllocatedString method.

Note that this method allocates the buffer twice, whereas IMFAttributes::GetString allocates only once, thus being more efficient.

fn GetBlob(&self, guid_key: &GUID) -> HrResult<Vec<u8>>

IMFAttributes::GetBlob method.

Calls IMFAttributes::GetBlobSize to alloc the buffer.

fn GetBlobSize(&self, guid_key: &GUID) -> HrResult<u32>

IMFAttributes::GetBlobSize method.

fn GetCount(&self) -> HrResult<u32>

IMFAttributes::GetCount method.

fn GetDouble(&self, guid_key: &GUID) -> HrResult<f64>

IMFAttributes::GetDouble method.

fn GetGUID(&self, guid_key: &GUID) -> HrResult<GUID>

IMFAttributes::GetGUID method.

fn GetItem(&self, guid_key: &GUID) -> HrResult<PROPVARIANT>

IMFAttributes::GetItem method.

fn GetItemByIndex(&self, index: u32) -> HrResult<(GUID, PROPVARIANT)>

IMFAttributes::GetItemByIndex method.

fn GetItemType(&self, guid_key: &GUID) -> HrResult<MF_ATTRIBUTE>

IMFAttributes::GetItemType method.

fn GetString(&self, guid_key: &GUID) -> HrResult<String>

IMFAttributes::GetString method.

Calls IMFAttributes::GetStringLength to alloc the buffer.

fn GetStringLength(&self, guid_key: &GUID) -> HrResult<u32>

IMFAttributes::GetStringLength method.

fn GetUINT32(&self, guid_key: &GUID) -> HrResult<u32>

IMFAttributes::GetUINT32 method.

fn GetUINT64(&self, guid_key: &GUID) -> HrResult<u64>

IMFAttributes::GetUINT64 method.

fn GetUnknown<T>(&self, guid_key: &GUID) -> HrResult<T>
where T: ole_IUnknown,

IMFAttributes::GetUnknown method.

fn LockStore(&self) -> HrResult<()>

IMFAttributes::LockStore method.

fn SetBlob(&self, guid_key: &GUID, buf: &[u8]) -> HrResult<()>

IMFAttributes::SetBlob method.

fn SetDouble(&self, guid_key: &GUID, value: f64) -> HrResult<()>

IMFAttributes::SetDouble method.

fn SetGUID(&self, guid_key: &GUID, value: &GUID) -> HrResult<()>

IMFAttributes::SetGUID method.

fn SetItem(&self, guid_key: &GUID, value: &PROPVARIANT) -> HrResult<()>

IMFAttributes::SetItem method.

fn SetString(&self, guid_key: &GUID, value: &str) -> HrResult<()>

IMFAttributes::SetString method.

fn SetUINT32(&self, guid_key: &GUID, value: u32) -> HrResult<()>

IMFAttributes::SetUINT32 method.

fn SetUINT64(&self, guid_key: &GUID, value: u64) -> HrResult<()>

IMFAttributes::SetUINT64 method.

fn SetUnknown(&self, guid_key: &GUID, value: &impl ole_IUnknown) -> HrResult<()>

IMFAttributes::SetUnknown method.

fn UnlockStore(&self) -> HrResult<()>

IMFAttributes::UnlockStore method.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.