Trait winsafe::prelude::ole_IMoniker

pub trait ole_IMoniker: ole_IPersistStream {

    // Provided methods
    fn BindToObject<T>(
        &self,
        bind_ctx: &impl ole_IBindCtx,
        moniker_to_left: Option<&impl ole_IMoniker>
    ) -> HrResult<T>
       where T: ole_IUnknown { ... }
    fn BindToStorage<T>(
        &self,
        bind_ctx: &impl ole_IBindCtx,
        moniker_to_left: Option<&impl ole_IMoniker>
    ) -> HrResult<T>
       where T: ole_IUnknown { ... }
    fn CommonPrefixWith(&self, other: &impl ole_IMoniker) -> HrResult<IMoniker> { ... }
    fn ComposeWith(
        &self,
        moniker_to_right: &impl ole_IMoniker,
        only_if_not_generic: bool
    ) -> HrResult<IMoniker> { ... }
    fn Enum(&self, forward: bool) -> HrResult<IMoniker> { ... }
    fn GetDisplayName(
        &self,
        bind_ctx: &impl ole_IBindCtx,
        moniker_to_left: Option<&impl ole_IMoniker>
    ) -> HrResult<String> { ... }
    fn GetTimeOfLastChange(
        &self,
        bind_ctx: &impl ole_IBindCtx,
        moniker_to_left: Option<&impl ole_IMoniker>
    ) -> HrResult<FILETIME> { ... }
    fn Hash(&self) -> HrResult<u32> { ... }
    fn Inverse(&self) -> HrResult<IMoniker> { ... }
    fn IsEqual(&self, other_moniker: &impl ole_IMoniker) -> HrResult<bool> { ... }
    fn IsRunning(
        &self,
        bind_ctx: &impl ole_IBindCtx,
        moniker_to_left: Option<&impl ole_IMoniker>,
        moniker_newly_running: Option<&impl ole_IMoniker>
    ) -> HrResult<bool> { ... }
    fn IsSystemMoniker(&self) -> HrResult<(bool, MKSYS)> { ... }
    fn ParseDisplayName(
        &self,
        bind_ctx: &impl ole_IBindCtx,
        moniker_to_left: &impl ole_IMoniker,
        display_name: &str
    ) -> HrResult<(u32, IMoniker)> { ... }
    fn Reduce(
        &self,
        bind_ctx: &impl ole_IBindCtx,
        reduce_how_far: MKRREDUCE
    ) -> HrResult<(IMoniker, IMoniker)> { ... }
    fn RelativePathTo(
        &self,
        other_moniker: &impl ole_IMoniker
    ) -> HrResult<IMoniker> { ... }

}

Available on crate features kernel and ole only.

This trait is enabled with the ole feature, and provides methods for IMoniker.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn BindToObject<T>( &self, bind_ctx: &impl ole_IBindCtx, moniker_to_left: Option<&impl ole_IMoniker> ) -> HrResult<T>
where T: ole_IUnknown,

IMoniker::BindToObject method.

fn BindToStorage<T>( &self, bind_ctx: &impl ole_IBindCtx, moniker_to_left: Option<&impl ole_IMoniker> ) -> HrResult<T>
where T: ole_IUnknown,

IMoniker::BindToStorage method.

fn CommonPrefixWith(&self, other: &impl ole_IMoniker) -> HrResult<IMoniker>

IMoniker::CommonPrefixWith method.

fn ComposeWith( &self, moniker_to_right: &impl ole_IMoniker, only_if_not_generic: bool ) -> HrResult<IMoniker>

IMoniker::ComposeWith method.

fn Enum(&self, forward: bool) -> HrResult<IMoniker>

IMoniker::Enum method.

fn GetDisplayName( &self, bind_ctx: &impl ole_IBindCtx, moniker_to_left: Option<&impl ole_IMoniker> ) -> HrResult<String>

IMoniker::GetDisplayName method.

fn GetTimeOfLastChange( &self, bind_ctx: &impl ole_IBindCtx, moniker_to_left: Option<&impl ole_IMoniker> ) -> HrResult<FILETIME>

IMoniker::GetTimeOfLastChange method.

fn Hash(&self) -> HrResult<u32>

IMoniker::Hash method.

fn Inverse(&self) -> HrResult<IMoniker>

IMoniker::Inverse method.

fn IsEqual(&self, other_moniker: &impl ole_IMoniker) -> HrResult<bool>

IMoniker::IsEqual method.

fn IsRunning( &self, bind_ctx: &impl ole_IBindCtx, moniker_to_left: Option<&impl ole_IMoniker>, moniker_newly_running: Option<&impl ole_IMoniker> ) -> HrResult<bool>

IMoniker::IsRunning method.

fn IsSystemMoniker(&self) -> HrResult<(bool, MKSYS)>

`IMoniker::IsSystemMoniker method.

fn ParseDisplayName( &self, bind_ctx: &impl ole_IBindCtx, moniker_to_left: &impl ole_IMoniker, display_name: &str ) -> HrResult<(u32, IMoniker)>

IMoniker::ParseDisplayName method.

fn Reduce( &self, bind_ctx: &impl ole_IBindCtx, reduce_how_far: MKRREDUCE ) -> HrResult<(IMoniker, IMoniker)>

IMoniker::Reduce method.

Returns the moniker to the left, and the reduced moniker, respectively.

fn RelativePathTo( &self, other_moniker: &impl ole_IMoniker ) -> HrResult<IMoniker>

IMoniker::RelativePathTo method.

Object Safety

This trait is not object safe.