winsafe::prelude

Trait oleaut_IDispatch

pub trait oleaut_IDispatch: ole_IUnknown {
    // Provided methods
    fn GetIDsOfNames(
        &self,
        names: &[impl AsRef<str>],
        lcid: LCID,
    ) -> HrResult<Vec<i32>> { ... }
    fn GetTypeInfoCount(&self) -> HrResult<u32> { ... }
    fn GetTypeInfo(&self, info_type: u32, lcid: LCID) -> HrResult<ITypeInfo> { ... }
    fn Invoke(
        &self,
        disp_id_member: i32,
        lcid: LCID,
        flags: DISPATCH,
        disp_params: &mut DISPPARAMS<'_, '_>,
    ) -> AnyResult<VARIANT> { ... }
    fn invoke_get(
        &self,
        property_name: &str,
        params: &[&Variant],
    ) -> AnyResult<Variant> { ... }
    fn invoke_method(
        &self,
        method_name: &str,
        params: &[&Variant],
    ) -> AnyResult<Variant> { ... }
    fn invoke_put(
        &self,
        property_name: &str,
        param: &Variant,
    ) -> AnyResult<Variant> { ... }
}

Available on crate feature oleaut only.

This trait is enabled with the oleaut feature, and provides methods for IDispatch.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn GetIDsOfNames( &self, names: &[impl AsRef<str>], lcid: LCID, ) -> HrResult<Vec<i32>>

IDispatch::GetIDsOfNames method.

fn GetTypeInfoCount(&self) -> HrResult<u32>

IDispatch::GetTypeInfoCount method.

fn GetTypeInfo(&self, info_type: u32, lcid: LCID) -> HrResult<ITypeInfo>

IDispatch::GetTypeInfo method.

fn Invoke( &self, disp_id_member: i32, lcid: LCID, flags: DISPATCH, disp_params: &mut DISPPARAMS<'_, '_>, ) -> AnyResult<VARIANT>

IDispatch::Invoke method.

If the remote call raises an exception, the returned error will be an EXCEPINFO.

This is a low-level method, prefer using:

IDispatch::invoke_get;

IDispatch::invoke_method; or

IDispatch::invoke_put.

fn invoke_get( &self, property_name: &str, params: &[&Variant], ) -> AnyResult<Variant>

Calls IDispatch::GetIDsOfNames and IDispatch::Invoke with co::DISPATCH::PROPERTYGET.

If the remote call raises an exception, the returned error will be an EXCEPINFO.

Examples

use winsafe::{self as w, prelude::*, co};

let _com_guard = w::CoInitializeEx(
    co::COINIT::APARTMENTTHREADED | co::COINIT::DISABLE_OLE1DDE)?;

let excel = w::CoCreateInstance::<w::IDispatch>(
    &w::CLSIDFromProgID("Excel.Application")?,
    None::<&w::IUnknown>,
    co::CLSCTX::LOCAL_SERVER,
)?;

let books = excel.invoke_get("Workbooks", &[])?.unwrap_dispatch();

let workbook = books
    .invoke_method("Open", &[&w::Variant::from_str("C:\\Temp\\bar.xlsx")])?
    .unwrap_dispatch();

let worksheets = workbook.invoke_get("Worksheets", &[])?.unwrap_dispatch();

let sheet1 = worksheets
    .invoke_get("Item", &[&w::Variant::from_str("Sheet1")])?
    .unwrap_dispatch();

fn invoke_method( &self, method_name: &str, params: &[&Variant], ) -> AnyResult<Variant>

Calls IDispatch::GetIDsOfNames and IDispatch::Invoke with co::DISPATCH::METHOD.

If the remote call raises an exception, the returned error will be an EXCEPINFO.

Examples

use winsafe::{self as w, prelude::*, co};

let _com_guard = w::CoInitializeEx(
    co::COINIT::APARTMENTTHREADED | co::COINIT::DISABLE_OLE1DDE)?;

let excel = w::CoCreateInstance::<w::IDispatch>(
    &w::CLSIDFromProgID("Excel.Application")?,
    None::<&w::IUnknown>,
    co::CLSCTX::LOCAL_SERVER,
)?;

let books = excel.invoke_get("Workbooks", &[])?
    .unwrap_dispatch();

let file = books.invoke_method(
    "Open",
    &[&w::Variant::from_str("C:\\Temp\\foo.xlsx")],
)?
.unwrap_dispatch();

file.invoke_method(
    "SaveAs",
    &[&w::Variant::from_str("C:\\Temp\\bar.xlsx")],
)?;

file.invoke_method("Close", &[])?;

fn invoke_put(&self, property_name: &str, param: &Variant) -> AnyResult<Variant>

Calls IDispatch::GetIDsOfNames and IDispatch::Invoke with co::DISPATCH::PROPERTYPUT.

If the remote call raises an exception, the returned error will be an EXCEPINFO.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.