Trait winsafe::prelude::oleaut_IPicture

pub trait oleaut_IPicture: ole_IUnknown {
    fn from_file(
        path: &str,
        transparent_color: Option<COLORREF>
    ) -> HrResult<IPicture> { ... }
    fn from_slice(src: &[u8], keep_original_format: bool) -> HrResult<IPicture> { ... }
    fn from_stream(
        stream: &IStream,
        size: Option<u32>,
        keep_original_format: bool
    ) -> HrResult<IPicture> { ... }
    fn get_CurDC(&self) -> HrResult<HDC> { ... }
    fn get_Height(&self) -> HrResult<i32> { ... }
    fn get_Type(&self) -> HrResult<PICTYPE> { ... }
    fn get_Width(&self) -> HrResult<i32> { ... }
    fn PictureChanged(&self) -> HrResult<()> { ... }
    fn put_KeepOriginalFormat(&self, keep: bool) -> HrResult<()> { ... }
    fn Render(
        &self,
        hdc: HDC,
        dest_pt: POINT,
        dest_sz: SIZE,
        src_offset: Option<POINT>,
        src_extent: SIZE,
        metafile_bounds: Option<&RECT>
    ) -> HrResult<()> { ... }
    fn SelectPicture(&self, hdc: HDC) -> HrResult<(HDC, HBITMAP)> { ... }
}

Available on crate feature oleaut only.

This trait is enabled with the oleaut feature, and provides methods for IPicture.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn from_file(
path: &str,
transparent_color: Option<COLORREF>
) -> HrResult<IPicture>

Calls OleLoadPicturePath to load a picture from a file.

The picture must be in BMP (bitmap), JPEG, WMF (metafile), ICO (icon), or GIF format.

fn from_slice(src: &[u8], keep_original_format: bool) -> HrResult<IPicture>

Calls IStream::from_slice and IPicture::from_stream to load a picture straight from a slice.

fn from_stream(
stream: &IStream,
size: Option<u32>,
keep_original_format: bool
) -> HrResult<IPicture>

Calls OleLoadPicture to load a picture from a stream.

fn get_CurDC(&self) -> HrResult<HDC>

IPicture::get_CurDC method.

fn get_Height(&self) -> HrResult<i32>

IPicture::get_Height method.

Note: Returns a value in HIMETRIC units. To convert it to pixels, prefer using the simpler IPicture::size_px, or use HDC::HiMetricToPixel to perform the conversion manually.

Examples

use winsafe::prelude::*;
use winsafe::{HWND, IPicture};

let pic: IPicture; // initialized somewhere

let hdc = HWND::NULL.GetDC()?;

let (_, height) = hdc.HiMetricToPixel(0, pic.get_Height()?);
println!("Height: {} px", height);

HWND::NULL.ReleaseDC(hdc)?;

fn get_Type(&self) -> HrResult<PICTYPE>

IPicture::get_Type method.

fn get_Width(&self) -> HrResult<i32>

IPicture::get_Width method.

Note: Returns a value in HIMETRIC units. To convert it to pixels, prefer using the simpler IPicture::size_px, or use HDC::HiMetricToPixel to perform the conversion manually.

Examples

use winsafe::prelude::*;
use winsafe::{HWND, IPicture};

let pic: IPicture; // initialized somewhere

let hdc = HWND::NULL.GetDC()?;

let (width, _) = hdc.HiMetricToPixel(pic.get_Width()?, 0);
println!("Width: {} px", width);

HWND::NULL.ReleaseDC(hdc)?;

fn PictureChanged(&self) -> HrResult<()>

IPicture::PictureChanged method.

fn put_KeepOriginalFormat(&self, keep: bool) -> HrResult<()>

IPicture::put_KeepOriginalFormat method.

fn Render(
&self,
hdc: HDC,
dest_pt: POINT,
dest_sz: SIZE,
src_offset: Option<POINT>,
src_extent: SIZE,
metafile_bounds: Option<&RECT>
) -> HrResult<()>

IPicture::Render method.

fn SelectPicture(&self, hdc: HDC) -> HrResult<(HDC, HBITMAP)>

IPicture::SelectPicture method.