winsafe::prelude

Trait oleaut_ITypeInfo

pub trait oleaut_ITypeInfo: ole_IUnknown {
    // Provided methods
    fn AddressOfMember(
        &self,
        member_id: i32,
        inv_kind: INVOKEKIND,
    ) -> HrResult<*mut c_void> { ... }
    fn CreateInstance<T>(
        &self,
        iunk_outer: Option<&impl ole_IUnknown>,
    ) -> HrResult<T>
       where T: ole_IUnknown { ... }
    fn GetDllEntry(
        &self,
        member_id: i32,
        inv_kind: INVOKEKIND,
    ) -> HrResult<(String, String, u16)> { ... }
    fn GetDocumentation(
        &self,
        member_id: i32,
    ) -> HrResult<(String, String, u32, String)> { ... }
    fn GetIDsOfNames(&self, names: &[impl AsRef<str>]) -> HrResult<Vec<i32>> { ... }
}

Available on crate feature oleaut only.

This trait is enabled with the oleaut feature, and provides methods for ITypeInfo.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn AddressOfMember( &self, member_id: i32, inv_kind: INVOKEKIND, ) -> HrResult<*mut c_void>

ITypeInfo::AddressOfMember method.

fn CreateInstance<T>( &self, iunk_outer: Option<&impl ole_IUnknown>, ) -> HrResult<T>
where T: ole_IUnknown,

ITypeInfo::CreateInstance method.

fn GetDllEntry( &self, member_id: i32, inv_kind: INVOKEKIND, ) -> HrResult<(String, String, u16)>

ITypeInfo::GetDllEntry method.

Returns:

DLL name;

entry point name;

ordinal.

fn GetDocumentation( &self, member_id: i32, ) -> HrResult<(String, String, u32, String)>

ITypeInfo::GetDocumentation method.

Returns:

item name;

documentation;

help localization context;

fully qualified name of the file containing the DLL used for help file.

fn GetIDsOfNames(&self, names: &[impl AsRef<str>]) -> HrResult<Vec<i32>>

ITypeInfo::GetIDsOfNames method.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.