winsafe::prelude
Trait shell_IShellLink

pub trait shell_IShellLink: ole_IUnknown {

    // Provided methods
    fn GetArguments(&self) -> HrResult<String> { ... }
    fn GetDescription(&self) -> HrResult<String> { ... }
    fn GetIconLocation(&self) -> HrResult<(String, i32)> { ... }
    fn GetPath(
        &self,
        fd: Option<&mut WIN32_FIND_DATA>,
        flags: SLGP,
    ) -> HrResult<String> { ... }
    fn GetShowCmd(&self) -> HrResult<SW> { ... }
    fn GetWorkingDirectory(&self) -> HrResult<String> { ... }
    fn Resolve(&self, hwnd: &HWND, flags: SLR) -> HrResult<()> { ... }
    fn SetArguments(&self, args: &str) -> HrResult<()> { ... }
    fn SetDescription(&self, args: &str) -> HrResult<()> { ... }
    fn SetIconLocation(&self, path: &str, index: i32) -> HrResult<()> { ... }
    fn SetPath(&self, file: &str) -> HrResult<()> { ... }
    fn SetRelativePath(&self, file: &str) -> HrResult<()> { ... }
    fn SetShowCmd(&self, show_cmd: SW) -> HrResult<()> { ... }
    fn SetWorkingDirectory(&self, dir: &str) -> HrResult<()> { ... }

}

Available on crate features kernel and shell only.

This trait is enabled with the shell feature, and provides methods for IShellLink.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn GetArguments(&self) -> HrResult<String>

IShellLink::GetArguments method.

fn GetDescription(&self) -> HrResult<String>

IShellLink::GetDescription method.

fn GetIconLocation(&self) -> HrResult<(String, i32)>

IShellLink::GetIconLocation method.

Returns the path of the icon and its index within the file.

fn GetPath( &self, fd: Option<&mut WIN32_FIND_DATA>, flags: SLGP, ) -> HrResult<String>

IShellLink::GetPath method.

fn GetShowCmd(&self) -> HrResult<SW>

IShellLink::GetShowCmd method.

fn GetWorkingDirectory(&self) -> HrResult<String>

IShellLink::GetWorkingDirectory method.

fn Resolve(&self, hwnd: &HWND, flags: SLR) -> HrResult<()>

IShellLink::Resolve method.

fn SetArguments(&self, args: &str) -> HrResult<()>

IShellLink::SetArguments method.

fn SetDescription(&self, args: &str) -> HrResult<()>

IShellLink::SetDescription method.

fn SetIconLocation(&self, path: &str, index: i32) -> HrResult<()>

IShellLink::SetIconLocation method.

fn SetPath(&self, file: &str) -> HrResult<()>

IShellLink::SetPath method.

fn SetRelativePath(&self, file: &str) -> HrResult<()>

IShellLink::SetRelativePath method.

fn SetShowCmd(&self, show_cmd: SW) -> HrResult<()>

IShellLink::SetShowCmd method.

fn SetWorkingDirectory(&self, dir: &str) -> HrResult<()>

IShellLink::SetWorkingDirectory method.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.