Trait winsafe::prelude::user_Haccel

pub trait user_Haccel: Handle {
    fn CreateAcceleratorTable(accel: &mut [ACCEL]) -> SysResult<HACCEL> { ... }
    fn DestroyAcceleratorTable(&self) -> bool { ... }
}

Available on crate feature user only.

This trait is enabled with the user feature, and provides methods for HACCEL.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn CreateAcceleratorTable(accel: &mut [ACCEL]) -> SysResult<HACCEL>

CreateAcceleratorTable static method.

Note: Must be paired with an HACCEL::DestroyAcceleratorTable call.

fn DestroyAcceleratorTable(&self) -> bool

DestroyAcceleratorTable method.

After calling this method, the handle will be invalidated and further operations will fail with ERROR::INVALID_HANDLE error code.