winsafe::prelude

Trait user_Hclipboard

pub trait user_Hclipboard: Handle {
    // Provided methods
    fn EmptyClipboard(&self) -> SysResult<()> { ... }
    fn GetClipboardData(&self, format: CF) -> SysResult<Vec<u8>> { ... }
    fn GetClipboardSequenceNumber(&self) -> u32 { ... }
    fn SetClipboardData(&self, format: CF, data: &[u8]) -> SysResult<()> { ... }
}

Available on crate features kernel and user only.

This trait is enabled with the user feature, and provides methods for HCLIPBOARD.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn EmptyClipboard(&self) -> SysResult<()>

EmptyClipboard function.

fn GetClipboardData(&self, format: CF) -> SysResult<Vec<u8>>

GetClipboardData function.

Calls HGLOBAL::GlobalSize and HGLOBAL::GlobalLock internally to retrieve a copy of the raw clipboard data.

Note that you should not trust the clipboard format – the binary data can be anything, despite what the format says. Be careful when parsing the binary into your desired format.

Examples

use winsafe::{self as w, prelude::*, co};

let hclip = w::HWND::NULL.OpenClipboard()?;
let data = hclip.GetClipboardData(co::CF::TEXT)?;

fn GetClipboardSequenceNumber(&self) -> u32

GetClipboardSequenceNumber function.

fn SetClipboardData(&self, format: CF, data: &[u8]) -> SysResult<()>

SetClipboardData function.

Calls HGLOBAL::GlobalAlloc and HGLOBAL::GlobalLock internally before copying the data into the clipboard.

Examples

use winsafe::{self as w, prelude::*, co};

let hclip = w::HWND::NULL.OpenClipboard()?;

let str_nullt = "foo"
    .as_bytes()
    .iter()
    .map(|ch| *ch)
    .chain(std::iter::once(0)) // null-terminated
    .collect::<Vec<_>>();

hclip.SetClipboardData(co::CF::TEXT, &str_nullt)?;

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.