Trait winsafe::prelude::user_Hdwp

pub trait user_Hdwp: Handle {
    // Provided methods
    fn BeginDeferWindowPos(
        num_windows: u32
    ) -> SysResult<EndDeferWindowPosGuard> { ... }
    fn DeferWindowPos(
        &mut self,
        hwnd: &HWND,
        hwnd_insert_after: HwndPlace,
        top_left: POINT,
        sz: SIZE,
        flags: SWP
    ) -> SysResult<()> { ... }
}

Available on crate features kernel and user only.

This trait is enabled with the user feature, and provides methods for HDWP.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn BeginDeferWindowPos(num_windows: u32) -> SysResult<EndDeferWindowPosGuard>

BeginDeferWindowPos function.

In the original C implementation, BeginDeferWindowPos returns a handle which must be passed to EndDeferWindowPos as a cleanup operation.

Here, the cleanup is performed automatically, because BeginDeferWindowPos returns an EndDeferWindowPosGuard, which automatically calls EndDeferWindowPos when the guard goes out of scope.

fn DeferWindowPos( &mut self, hwnd: &HWND, hwnd_insert_after: HwndPlace, top_left: POINT, sz: SIZE, flags: SWP ) -> SysResult<()>

DeferWindowPos function.

Originally this method returns the handle to the reallocated memory object; here the original handle is automatically updated.

Object Safety

This trait is not object safe.