winsafe::prelude
Trait user_Hhook

pub trait user_Hhook: Handle {
    // Provided methods
    fn CallNextHookEx(&self, code: WH, wparam: usize, lparam: isize) -> isize { ... }
    fn SetWindowsHookEx(
        hook_id: WH,
        proc: HOOKPROC,
        module: Option<&HINSTANCE>,
        thread_id: Option<u32>,
    ) -> SysResult<HHOOK> { ... }
    fn UnhookWindowsHookEx(&mut self) -> SysResult<()> { ... }
}

Available on crate features kernel and user only.

This trait is enabled with the user feature, and provides methods for HHOOK.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn CallNextHookEx(&self, code: WH, wparam: usize, lparam: isize) -> isize

CallNextHookEx function.

fn SetWindowsHookEx( hook_id: WH, proc: HOOKPROC, module: Option<&HINSTANCE>, thread_id: Option<u32>, ) -> SysResult<HHOOK>

SetWindowsHookEx function.

fn UnhookWindowsHookEx(&mut self) -> SysResult<()>

UnhookWindowsHookEx function.

After calling this method, the handle will be invalidated and further operations will fail with ERROR::INVALID_HANDLE error code.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.