winsafe::prelude

Trait version_Hversioninfo

pub trait version_Hversioninfo: Handle {
    // Provided methods
    fn GetFileVersionInfo(file_name: &str) -> SysResult<VersionInfoGuard> { ... }
    fn GetFileVersionInfoSize(file_name: &str) -> SysResult<u32> { ... }
    fn langs_and_cps(&self) -> SysResult<&[(LANGID, CP)]> { ... }
    unsafe fn VerQueryValue<T>(
        &self,
        sub_block: &str,
    ) -> SysResult<(*const T, u32)> { ... }
    fn str_val(
        &self,
        lang_and_cp: (LANGID, CP),
        name: &str,
    ) -> SysResult<String> { ... }
    fn version_info(&self) -> SysResult<&VS_FIXEDFILEINFO> { ... }
}

Available on crate features kernel and version only.

This trait is enabled with the version feature, and provides methods for HVERSIONINFO.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn GetFileVersionInfo(file_name: &str) -> SysResult<VersionInfoGuard>

GetFileVersionInfo function.

The returned buffer will be automatically allocated with HVERSIONINFO::GetFileVersionInfoSize.

fn GetFileVersionInfoSize(file_name: &str) -> SysResult<u32>

GetFileVersionInfoSize function.

You don’t need to call this function directly, because HVERSIONINFO::GetFileVersionInfo already calls it.

fn langs_and_cps(&self) -> SysResult<&[(LANGID, CP)]>

Calls HVERSIONINFO::VerQueryValue to retrieve a reference to a slice with all languages and code pages.

Examples

Listing all pairs of language and code page:

use winsafe::{self as w, prelude::*};

let exe_name = w::HINSTANCE::NULL.GetModuleFileName()?;
let hversion = w::HVERSIONINFO::GetFileVersionInfo(&exe_name)?;

for (lang, cp) in hversion.langs_and_cps()?.iter() {
    println!("{} {}", lang, cp);
}

unsafe fn VerQueryValue<T>(&self, sub_block: &str) -> SysResult<(*const T, u32)>

VerQueryValue function.

Safety

The returned pointer and size vary according to lpSubBlock. If you set it wrong, you’re likely to cause a buffer overrun.

This is a low-level method, prefer using:

langs_and_cps;

str_val;

version_info.

Examples

Reading version information from resource:

use winsafe::{self as w, prelude::*};

let exe_name = w::HINSTANCE::NULL.GetModuleFileName()?;
let hversion = w::HVERSIONINFO::GetFileVersionInfo(&exe_name)?;

let (pvsf, sz_data) = unsafe {
    hversion.VerQueryValue::<w::VS_FIXEDFILEINFO>("\\")?
};

let ver = unsafe { &*pvsf }.dwFileVersion();
println!("Version {}.{}.{}.{}",
    ver[0], ver[1], ver[2], ver[3]);

fn str_val(&self, lang_and_cp: (LANGID, CP), name: &str) -> SysResult<String>

Calls HVERSIONINFO::VerQueryValue to retrieve a string value.

Common value names are:

Comments

CompanyName

FileDescription

FileVersion

InternalName

LegalCopyright

LegalTrademarks

OriginalFilename

ProductName

ProductVersion

PrivateBuild

SpecialBuild

Examples

Reading product name and legal copyright from resource:

use winsafe::{self as w, prelude::*};

let exe_name = w::HINSTANCE::NULL.GetModuleFileName()?;
let hversion = w::HVERSIONINFO::GetFileVersionInfo(&exe_name)?;

let lang_cp = hversion.langs_and_cps()?[0]; // first language and code page

println!(
    "{}\n{}",
    hversion.str_val(lang_cp, "ProductName")?,
    hversion.str_val(lang_cp, "LegalCopyright")?,
);

fn version_info(&self) -> SysResult<&VS_FIXEDFILEINFO>

Calls HVERSIONINFO::VerQueryValue to retrieve a reference to the fixed version block, if any.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.