winsafe

Struct BSTR

pub struct BSTR(/* private fields */);

Available on crate feature oleaut only.

A string data type used with COM automation.

Automatically calls SysFreeString when the object goes out of scope.

Implementations

impl BSTR

pub fn SysAllocString(s: &str) -> HrResult<Self>

SysAllocString function.

pub fn SysReAllocString(&mut self, s: &str) -> HrResult<()>

SysReAllocString function.

The underlying pointer is automatically updated.

pub fn SysStringLen(&self) -> u32

SysStringLen function.

pub const unsafe fn from_ptr(p: *mut u16) -> Self

Creates a new BSTR by wrapping a pointer.

Safety

Be sure the pointer has the correct type and isn’t owned by anyone else, otherwise you may cause memory access violations.

pub const fn as_ptr(&self) -> *mut u16

Returns the underlying LPWSTR pointer to the null-terminated wide string.

pub const fn as_mut_ptr(&mut self) -> *mut *mut u16

Returns a pointer to the underlying LPWSTR pointer to the null-terminated wide string.

pub fn as_slice(&self) -> &[u16]

Returns the underlying LPWSTR memory block as a null-terminated u16 slice.

pub const fn leak(&mut self) -> *mut u16

Ejects the underlying LPWSTR pointer leaving a null pointer in its place, so that SysFreeString won’t be called.

Be sure to free the pointer, otherwise, as the name of this method implies, you will cause a memory leak.