Struct winsafe::ComPtr

#[repr(transparent)]
pub struct ComPtr(_);

Available on crate feature ole only.

A pointer to pointer to a COM virtual table.

Implementations

impl ComPtr

pub const unsafe fn into_ptr<T>(self) -> *mut *mut T

Converts the underlying raw pointer converted to a specific virtual table pointer.

Safety

Be sure the pointer actually points to the given virtual table type.

This method is used internally by the library, and not intended to be used externally.

pub const unsafe fn null() -> Self

Creates a null pointer to a COM virtual table.

Safety

A null pointer to a COM virtual table has no practical use, and calling methods upon it is undefined behavior.

This method is used internally by the library, and not intended to be used externally.

pub fn as_opt(&self) -> Option<&Self>

Returns None if the COM pointer is null, otherwise returns Some(&self).