winsafe

Struct FileMapped

pub struct FileMapped { /* private fields */ }

Available on crate feature kernel only.

Manages an HFILEMAP handle, which provides memory-mapped file operations, including read/write through slices. It is closed automatically when the object goes out of scope.

Examples

Usually the fastest way to read a file as string is by memory-mapping it, like:

use winsafe::{self as w, prelude::*};

let f = w::FileMapped::open(
    "C:\\Temp\\foo.txt",
    w::FileAccess::ExistingReadOnly,
)?;
let raw_bytes = f.as_slice();
let text = w::WString::parse(raw_bytes)?.to_string();

Implementations

impl FileMapped

pub fn open(file_path: &str, access: FileAccess) -> SysResult<Self>

Opens a file with the desired access, then map its contents in memory.

pub fn as_mut_slice(&mut self) -> &mut [u8]

Returns a mutable slice to the mapped memory.

pub fn as_slice(&self) -> &[u8]

Returns a slice to the mapped memory.

pub fn hfile(&self) -> &HFILE

Returns the underlying file handle.

pub const fn size(&self) -> u64

Returns the size of the file.

This value is cached.

pub fn times(&self) -> SysResult<(SYSTEMTIME, SYSTEMTIME, SYSTEMTIME)>

Returns, in current time zone, 3 times of the file, respectively:

creation time;

last access time;

last write time.