winsafe

Struct HACCESSTOKEN

pub struct HACCESSTOKEN(/* private fields */);

Available on crate feature advapi only.

Handle to an access token. Originally just a HANDLE.

Implementations

impl HACCESSTOKEN

pub const unsafe fn from_ptr(p: *mut c_void) -> Self

Constructs a new handle object by wrapping a pointer.

This method can be used as an escape hatch to interoperate with other libraries.

Safety

Be sure the pointer has the correct type and isn’t owned by anyone else, otherwise you may cause memory access violations.

pub const unsafe fn raw_copy(&self) -> Self

Returns a raw copy of the underlying handle pointer.

Safety

As the name implies, raw_copy returns a raw copy of the handle, so closing one of the copies won’t close the others. This means a handle can be used after it has been closed, what can lead to errors and undefined behavior. Even worse: sometimes Windows reuses handle values, so you can call a method on a completely different handle type, what can be catastrophic.

However, in some cases the Windows API demands a copy of the handle – raw_copy is an escape hatch to fill this gap.

pub const unsafe fn as_mut(&mut self) -> &mut *mut c_void

Returns a mutable reference to the underlying raw pointer.

This method can be used as an escape hatch to interoperate with other libraries.

Safety

This method exposes the raw pointer used by raw Windows calls. It’s an opaque pointer to an internal Windows structure, and no dereferencings should be attempted.

pub const fn ptr(&self) -> *mut c_void

Returns the underlying raw pointer.

This method exposes the raw pointer used by raw Windows calls. It’s an opaque pointer to an internal Windows structure, and no dereferencings should be attempted.

This method can be used as an escape hatch to interoperate with other libraries.

impl HACCESSTOKEN

pub fn AdjustTokenPrivileges(&self, new_state: DisabPriv<'_>) -> SysResult<()>

AdjustTokenPrivileges function.

Examples

use winsafe::{self as w, prelude::*, co};

let htoken = w::HPROCESS::GetCurrentProcess()
    .OpenProcessToken(co::TOKEN::ADJUST_PRIVILEGES | co::TOKEN::QUERY)?;

let luid = w::LookupPrivilegeValue(None, co::SE_PRIV::SHUTDOWN_NAME)?;

let privs = w::TOKEN_PRIVILEGES::new(&[
    w::LUID_AND_ATTRIBUTES::new(luid, co::SE_PRIV_ATTR::ENABLED),
])?;

htoken.AdjustTokenPrivileges(w::DisabPriv::Privs(&privs))?;

pub fn CheckTokenCapability( &self, capability_sid_to_check: &SID, ) -> SysResult<bool>

CheckTokenCapability function.

pub fn CheckTokenMembership(&self, sid_to_check: &SID) -> SysResult<bool>

CheckTokenMembership function.

pub fn DuplicateToken( &self, level: SECURITY_IMPERSONATION, ) -> SysResult<CloseHandleGuard<HACCESSTOKEN>>

DuplicateToken function.

pub fn GetCurrentProcessToken() -> HACCESSTOKEN

GetCurrentProcessToken function.

pub fn GetCurrentThreadEffectiveToken() -> HACCESSTOKEN

GetCurrentThreadEffectiveToken function.

pub fn GetTokenInformation( &self, information_class: TOKEN_INFORMATION_CLASS, ) -> SysResult<TokenInfo<'_, '_, '_, '_, '_, '_>>

GetTokenInformation function.

The returned enum variant will correspond to the passed information_class.

Examples

Retrieving the Groups information:

use winsafe::{self as w, prelude::*, co};

let htoken = w::HPROCESS::GetCurrentProcess()
    .OpenProcessToken(co::TOKEN::QUERY)?;

let nfo = htoken.GetTokenInformation(co::TOKEN_INFORMATION_CLASS::Groups)?;
let w::TokenInfo::Groups(groups) = nfo else { unreachable!() };

for (idx, g) in groups.Groups().iter().enumerate() {
    println!("{}: {}", idx, g.Sid().unwrap());
}

pub fn ImpersonateLoggedOnUser(&self) -> SysResult<()>

ImpersonateLoggedOnUser function.

pub fn IsTokenRestricted(&self) -> SysResult<bool>

IsTokenRestricted function.