winsafe

Struct HDC

pub struct HDC(/* private fields */);

Available on crate feature user only.

Handle to a device context.

Implementations

impl HDC

pub fn AbortDoc(&self) -> SysResult<()>

Available on crate feature gdi only.

AbortDoc function.

pub fn AbortPath(&self) -> SysResult<()>

Available on crate feature gdi only.

AborthPath function.

pub fn AlphaBlend( &self, origin_dest: RECT, hdc_src: &HDC, origin_src: RECT, ftn: &BLENDFUNCTION, ) -> SysResult<()>

Available on crate feature gdi only.

AlphaBlend function.

pub fn AngleArc( &self, center: POINT, radius: u32, start_angle: f32, sweep_angle: f32, ) -> SysResult<()>

Available on crate feature gdi only.

AngleArc function.

pub fn Arc( &self, bound: RECT, radial_start: POINT, radial_end: POINT, ) -> SysResult<()>

Available on crate feature gdi only.

Arc function.

pub fn ArcTo( &self, bound: RECT, radial_start: POINT, radial_end: POINT, ) -> SysResult<()>

Available on crate feature gdi only.

ArcTo function.

pub fn BeginPath(&self) -> SysResult<()>

Available on crate feature gdi only.

BeginPath function.

pub fn BitBlt( &self, dest_pos: POINT, sz: SIZE, hdc_src: &HDC, src_src: POINT, rop: ROP, ) -> SysResult<()>

Available on crate feature gdi only.

BitBlt function.

pub fn CancelDC(&self) -> SysResult<()>

Available on crate feature gdi only.

CancelDC function.

pub fn ChoosePixelFormat(&self, pfd: &PIXELFORMATDESCRIPTOR) -> SysResult<i32>

Available on crate feature gdi only.

ChoosePixelFormat function.

pub fn Chord( &self, bounds: RECT, start_radial: POINT, end_radial: POINT, ) -> SysResult<()>

Available on crate feature gdi only.

Chord function.

pub fn CloseFigure(&self) -> SysResult<()>

Available on crate feature gdi only.

CloseFigure function.

pub fn CreateCompatibleBitmap( &self, cx: i32, cy: i32, ) -> SysResult<DeleteObjectGuard<HBITMAP>>

Available on crate feature gdi only.

CreateCompatibleBitmap function.

pub fn CreateCompatibleDC(&self) -> SysResult<DeleteDCGuard>

Available on crate feature gdi only.

CreateCompatibleDC function.

pub fn CreateHalftonePalette(&self) -> SysResult<DeleteObjectPaletteGuard>

Available on crate feature gdi only.

CreateHalftonePalette function.

pub fn DescribePixelFormat( &self, index: i32, ) -> SysResult<PIXELFORMATDESCRIPTOR>

Available on crate feature gdi only.

DescribePixelFormat function.

pub fn Ellipse(&self, bound: RECT) -> SysResult<()>

Available on crate feature gdi only.

Ellipse function.

pub fn EndPath(&self) -> SysResult<()>

Available on crate feature gdi only.

EndPath function.

pub fn ExcludeClipRect(&self, rc: RECT) -> SysResult<REGION>

Available on crate feature gdi only.

ExcludeClipRect function.

pub fn FillPath(&self) -> SysResult<()>

Available on crate feature gdi only.

FillPath function.

pub fn FillRect(&self, rc: RECT, hbr: &HBRUSH) -> SysResult<()>

Available on crate feature gdi only.

FillRect function.

pub fn FillRgn(&self, rgn: &HRGN, brush: &HBRUSH) -> SysResult<()>

Available on crate feature gdi only.

FillRgn function.

pub fn FlattenPath(&self) -> SysResult<()>

Available on crate feature gdi only.

FlattenPath function.

pub fn FrameRgn( &self, rgn: &HRGN, brush: &HBRUSH, w: i32, h: i32, ) -> SysResult<()>

Available on crate feature gdi only.

FrameRgn function.

pub fn GetBkColor(&self) -> SysResult<COLORREF>

Available on crate feature gdi only.

GetBkColor function.

pub fn GetBkMode(&self) -> SysResult<BKMODE>

Available on crate feature gdi only.

GetBkMode function.

pub fn GetCurrentObject(&self, kind: CUR_OBJ) -> SysResult<CurObj>

Available on crate feature gdi only.

GetCurrentObject function.

Examples

use winsafe::{self as w, prelude::*, co};

let hdc: w::HDC; // initialized somewhere

let obj = hdc.GetCurrentObject(co::CUR_OBJ::BRUSH)?;
let w::CurObj::Brush(hbrush) = obj else { unreachable!() };

println!("HBRUSH: {}", hbrush);
w::SysResult::Ok(())

pub fn GetCurrentPositionEx(&self) -> SysResult<POINT>

Available on crate feature gdi only.

GetCurrentPositionEx function.

pub fn GetDCBrushColor(&self) -> SysResult<COLORREF>

Available on crate feature gdi only.

GetDCBrushColor function.

pub fn GetDCPenColor(&self) -> SysResult<COLORREF>

Available on crate feature gdi only.

GetDCPenColor function.

pub fn GetDeviceCaps(&self, index: GDC) -> i32

Available on crate feature gdi only.

GetDeviceCaps function.

pub unsafe fn GetDIBits( &self, hbm: &HBITMAP, first_scan_line: u32, num_scan_lines: u32, bmp_data_buf: Option<&mut [u8]>, bmi: &mut BITMAPINFO, usage: DIB, ) -> SysResult<i32>

Available on crate feature gdi only.

GetDIBits function.

Safety

If bmpDataBuf is smaller than needed, you’ll have a buffer overflow.

Examples

Taking a screenshot and saving to file:

use winsafe::{self as w, prelude::*, co};

let cx_screen = w::GetSystemMetrics(co::SM::CXSCREEN);
let cy_screen = w::GetSystemMetrics(co::SM::CYSCREEN);

let hdc_screen = w::HWND::DESKTOP.GetDC()?;
let hbmp = hdc_screen.CreateCompatibleBitmap(cx_screen, cy_screen)?;
let hdc_mem = hdc_screen.CreateCompatibleDC()?;
let _hbmp_guard = hdc_mem.SelectObject(&*hbmp)?;

hdc_mem.BitBlt(
    w::POINT::new(),
    w::SIZE::with(cx_screen, cy_screen),
    &hdc_screen,
    w::POINT::new(),
    co::ROP::SRCCOPY,
)?;

let bmp_obj = hbmp.GetObject()?;

let mut bi = w::BITMAPINFO::default();
bi.bmiHeader.biWidth = cx_screen;
bi.bmiHeader.biHeight = cy_screen;
bi.bmiHeader.biPlanes = 1;
bi.bmiHeader.biBitCount = 32;
bi.bmiHeader.biCompression = co::BI::RGB;

let bmp_size = (bmp_obj.bmWidth * (bi.bmiHeader.biBitCount as i32) + 31)
    / 32 * 4 * bmp_obj.bmHeight;
let mut data_buf = vec![0u8; bmp_size as _];

unsafe {
    hdc_screen.GetDIBits(
        &hbmp,
        0,
        cy_screen as _,
        Some(&mut data_buf),
        &mut bi,
        co::DIB::RGB_COLORS,
    )?;
}

let mut bfh = w::BITMAPFILEHEADER::default();
bfh.bfOffBits = (std::mem::size_of::<w::BITMAPFILEHEADER>()
    + std::mem::size_of::<w::BITMAPINFOHEADER>()) as _;
bfh.bfSize = bfh.bfOffBits + (bmp_size as u32);

let fo = w::File::open("C:\\Temp\\foo.bmp", w::FileAccess::OpenOrCreateRW)?;
fo.write(bfh.serialize())?;
fo.write(bi.bmiHeader.serialize())?;
fo.write(&data_buf)?;

pub fn GetPixelFormat(&self) -> SysResult<i32>

Available on crate feature gdi only.

GetPixelFormat function.

pub fn GetStretchBltMode(&self) -> SysResult<STRETCH_MODE>

Available on crate feature gdi only.

GetStretchBltMode function.

pub fn GetTextColor(&self) -> SysResult<COLORREF>

Available on crate feature gdi only.

GetTextColor function.

pub fn GetTextExtentPoint32(&self, text: &str) -> SysResult<SIZE>

Available on crate feature gdi only.

GetTextExtentPoint32 function.

pub fn GetTextFace(&self) -> SysResult<String>

Available on crate feature gdi only.

GetTextFace function.

pub fn GetTextMetrics(&self) -> SysResult<TEXTMETRIC>

Available on crate feature gdi only.

GetTextMetrics function.

pub fn GetViewportExtEx(&self) -> SysResult<SIZE>

Available on crate feature gdi only.

GetViewportExtEx function.

pub fn GetViewportOrgEx(&self) -> SysResult<POINT>

Available on crate feature gdi only.

GetViewportOrgEx function.

pub fn GetWindowExtEx(&self) -> SysResult<SIZE>

Available on crate feature gdi only.

GetWindowExtEx function.

pub fn GetWindowOrgEx(&self) -> SysResult<POINT>

Available on crate feature gdi only.

GetWindowOrgEx function.

pub fn HiMetricToPixel(&self, x: i32, y: i32) -> (i32, i32)

Available on crate feature gdi only.

AtlHiMetricToPixel function.

Converts HIMETRIC units to pixels. The inverse operation is HDC::PixelToHiMetric.

pub fn IntersectClipRect(&self, rc: RECT) -> SysResult<()>

Available on crate feature gdi only.

IntersectClipRect function.

pub fn InvertRgn(&self, hrgn: &HRGN) -> SysResult<()>

Available on crate feature gdi only.

InvertRgn function.

pub fn LineTo(&self, x: i32, y: i32) -> SysResult<()>

Available on crate feature gdi only.

LineTo function.

pub fn MaskBlt( &self, dest_top_left: POINT, sz: SIZE, hdc_src: &HDC, src_top_left: POINT, hbm_mask: &HBITMAP, mask_offset: POINT, rop: ROP, ) -> SysResult<()>

Available on crate feature gdi only.

MaskBlt function.

pub fn MoveToEx(&self, x: i32, y: i32, pt: Option<&mut POINT>) -> SysResult<()>

Available on crate feature gdi only.

MoveToEx function.

pub fn PaintRgn(&self, hrgn: &HRGN) -> SysResult<()>

Available on crate feature gdi only.

PaintRgn function.

pub fn PatBlt(&self, top_left: POINT, sz: SIZE, rop: ROP) -> SysResult<()>

Available on crate feature gdi only.

PatBlt function.

pub fn PathToRegion(&self) -> SysResult<DeleteObjectGuard<HRGN>>

Available on crate feature gdi only.

PathToRegion function.

pub fn Pie( &self, bounds: RECT, radial_1: POINT, radial_2: POINT, ) -> SysResult<()>

Available on crate feature gdi only.

Pie function.

pub fn PixelToHiMetric(&self, x: i32, y: i32) -> (i32, i32)

Available on crate feature gdi only.

AtlPixelToHiMetric function.

Converts pixels to HIMETRIC units. The inverse operation is HDC::HiMetricToPixel.

pub fn PolyBezier(&self, pts: &[POINT]) -> SysResult<()>

Available on crate feature gdi only.

PolyBezier function.

pub fn PolyBezierTo(&self, pts: &[POINT]) -> SysResult<()>

Available on crate feature gdi only.

PolyBezierTo function.

pub fn PolyDraw(&self, pts: &[(POINT, PT)]) -> SysResult<()>

Available on crate feature gdi only.

PolyDraw function.

pub fn Polygon(&self, pts: &[POINT]) -> SysResult<()>

Available on crate feature gdi only.

Polygon function.

pub fn Polyline(&self, pts: &[POINT]) -> SysResult<()>

Available on crate feature gdi only.

Polyline function.

pub fn PolylineTo(&self, pts: &[POINT]) -> SysResult<()>

Available on crate feature gdi only.

PolylineTo function.

pub fn PolyPolygon(&self, polygons: &[&[POINT]]) -> SysResult<()>

Available on crate feature gdi only.

PolyPolygon function.

pub fn PolyPolyline(&self, polylines: &[&[POINT]]) -> SysResult<()>

Available on crate feature gdi only.

PolyPolyline function.

pub fn PtVisible(&self, x: i32, y: i32) -> SysResult<bool>

Available on crate feature gdi only.

PtVisible function.

pub fn RealizePalette(&self) -> SysResult<u32>

Available on crate feature gdi only.

RealizePalette function.

pub fn Rectangle(&self, bounds: RECT) -> SysResult<()>

Available on crate feature gdi only.

Rectangle function.

pub fn RestoreDC(&self, saved_dc: i32) -> SysResult<()>

Available on crate feature gdi only.

RestoreDC function.

pub fn RoundRect(&self, bounds: RECT, sz: SIZE) -> SysResult<()>

Available on crate feature gdi only.

RoundRect function.

pub fn SaveDC(&self) -> SysResult<i32>

Available on crate feature gdi only.

SaveDC function.

pub fn SelectClipPath(&self, mode: RGN) -> SysResult<()>

Available on crate feature gdi only.

SelectClipPath function.

pub fn SelectClipRgn(&self, rgn: &HRGN) -> SysResult<REGION>

Available on crate feature gdi only.

SelectClipRgn function.

pub fn SelectObject<G: GdiObject>( &self, hgdiobj: &G, ) -> SysResult<SelectObjectGuard<'_, G>>

Available on crate feature gdi only.

SelectObject function.

In the original C implementation, SelectObject returns a handle to the object being replaced. You must perform a cleanup operation, calling SelectObject again, passing the handle to the replaced object.

Here, the cleanup is performed automatically, because SelectObject returns a SelectObjectGuard, which stores the replaced handle and calls SelectObject automatically when the guard goes out of scope. You must, however, keep the guard alive, otherwise the cleanup will be performed right away.

Examples

use winsafe::{self as w, prelude::*, co};

let hdc: w::HDC; // initialized somewhere

let hpen = w::HPEN::CreatePen(
    co::PS::SOLID,
    1,
    w::COLORREF::from_rgb(0xff, 0x00, 0x88),
)?;

let _pen_guard = hdc.SelectObject(&*hpen); // keep guard alive

pub fn SelectPalette( &self, hpal: &HPALETTE, force_bkgd: bool, ) -> SysResult<HPALETTE>

Available on crate feature gdi only.

SelectPalette function.

pub fn SetArcDirection(&self, dir: AD) -> SysResult<AD>

Available on crate feature gdi only.

SetArcDirection function.

pub fn SetBkColor(&self, color: COLORREF) -> SysResult<COLORREF>

Available on crate feature gdi only.

SetBkColor function.

pub fn SetBkMode(&self, mode: BKMODE) -> SysResult<BKMODE>

Available on crate feature gdi only.

SetBkMode function.

pub fn SetBrushOrgEx(&self, new_origin: POINT) -> SysResult<POINT>

Available on crate feature gdi only.

SetBrushOrgEx function.

pub fn SetDCBrushColor(&self, color: COLORREF) -> SysResult<COLORREF>

Available on crate feature gdi only.

SetDCBrushColor function.

pub fn SetDCPenColor(&self, color: COLORREF) -> SysResult<COLORREF>

Available on crate feature gdi only.

SetDCPenColor function.

pub fn SetDIBits( &self, hbm: &HBITMAP, first_scan_line: u32, num_scan_lines: u32, dib_color_data: &[u8], bmi: &BITMAPINFO, color_use: DIB, ) -> SysResult<i32>

Available on crate feature gdi only.

SetDIBits function.

pub fn SetGraphicsMode(&self, mode: GM) -> SysResult<GM>

Available on crate feature gdi only.

SetGraphicsMode function.

pub fn SetStretchBltMode(&self, mode: STRETCH_MODE) -> SysResult<STRETCH_MODE>

Available on crate feature gdi only.

SetStretchBltMode function.

pub fn SetTextAlign(&self, align: TA) -> SysResult<TA>

Available on crate feature gdi only.

SetTextAlign function.

pub fn SetTextColor(&self, color: COLORREF) -> SysResult<COLORREF>

Available on crate feature gdi only.

SetTextColor function.

pub fn SetTextJustification(&self, extra: i32, count: i32) -> SysResult<()>

Available on crate feature gdi only.

SetTextJustification function.

pub fn SetViewportExtEx(&self, x: i32, y: i32) -> SysResult<SIZE>

Available on crate feature gdi only.

SetViewportExtEx function.

pub fn SetViewportOrgEx(&self, x: i32, y: i32) -> SysResult<POINT>

Available on crate feature gdi only.

SetViewportOrgEx function.

pub fn SetWindowExtEx(&self, x: i32, y: i32) -> SysResult<SIZE>

Available on crate feature gdi only.

SetWindowExtEx function.

pub fn SetWindowOrgEx(&self, x: i32, y: i32) -> SysResult<POINT>

Available on crate feature gdi only.

SetWindowOrgEx function.

pub fn StretchBlt( &self, pos_dest: POINT, sz_dest: SIZE, hdc_src: &HDC, pt_src: POINT, sz_src: SIZE, rop: ROP, ) -> SysResult<()>

Available on crate feature gdi only.

StretchBlt function.

pub fn StrokeAndFillPath(&self) -> SysResult<()>

Available on crate feature gdi only.

StrokeAndFillPath function.

pub fn StrokePath(&self) -> SysResult<()>

Available on crate feature gdi only.

StrokePath function.

pub fn TextOut(&self, x: i32, y: i32, text: &str) -> SysResult<()>

Available on crate feature gdi only.

TextOut function.

pub fn TransparentBlt( &self, dest_top_left: POINT, dest_sz: SIZE, hdc_src: HDC, src_top_left: POINT, src_sz: SIZE, color_transparent: COLORREF, ) -> SysResult<()>

Available on crate feature gdi only.

TransparentBlt function.

pub fn UpdateColors(&self) -> SysResult<()>

Available on crate feature gdi only.

UpdateColors function.

pub fn WidenPath(&self) -> SysResult<()>

Available on crate feature gdi only.

WidenPath function.

impl HDC

pub const unsafe fn from_ptr(p: *mut c_void) -> Self

Constructs a new handle object by wrapping a pointer.

This method can be used as an escape hatch to interoperate with other libraries.

Safety

Be sure the pointer has the correct type and isn’t owned by anyone else, otherwise you may cause memory access violations.

pub const unsafe fn raw_copy(&self) -> Self

Returns a raw copy of the underlying handle pointer.

Safety

As the name implies, raw_copy returns a raw copy of the handle, so closing one of the copies won’t close the others. This means a handle can be used after it has been closed, what can lead to errors and undefined behavior. Even worse: sometimes Windows reuses handle values, so you can call a method on a completely different handle type, what can be catastrophic.

However, in some cases the Windows API demands a copy of the handle – raw_copy is an escape hatch to fill this gap.

pub const unsafe fn as_mut(&mut self) -> &mut *mut c_void

Returns a mutable reference to the underlying raw pointer.

This method can be used as an escape hatch to interoperate with other libraries.

Safety

This method exposes the raw pointer used by raw Windows calls. It’s an opaque pointer to an internal Windows structure, and no dereferencings should be attempted.

pub const fn ptr(&self) -> *mut c_void

Returns the underlying raw pointer.

This method exposes the raw pointer used by raw Windows calls. It’s an opaque pointer to an internal Windows structure, and no dereferencings should be attempted.

This method can be used as an escape hatch to interoperate with other libraries.

impl HDC

pub fn DrawFocusRect(&self, rect: RECT) -> SysResult<()>

DrawFocusRect function.

pub fn DrawText(&self, text: &str, bounds: RECT, format: DT) -> SysResult<i32>

DrawText function.

pub fn DrawTextEx( &self, text: &str, bounds: RECT, format: DT, dtp: Option<&DRAWTEXTPARAMS>, ) -> SysResult<i32>

DrawTextExW function.

pub fn EnumDisplayMonitors<F>( &self, rc_clip: Option<RECT>, func: F, ) -> SysResult<()>
where F: FnMut(HMONITOR, HDC, &RECT) -> bool,

EnumDisplayMonitors function.

Examples

use winsafe::{self as w, prelude::*};

let hdc: w::HDC; // initialized somewhere

hdc.EnumDisplayMonitors(
    None,
    |hmon: w::HMONITOR, hdc: w::HDC, rc: &w::RECT| -> bool {
        println!("HMONITOR: {}, ", hmon);
        true
    },
)?;

pub fn ExcludeUpdateRgn(&self, hwnd: &HWND) -> SysResult<REGION>

ExcludeUpdateRgn function.

pub fn FrameRect(&self, rc: RECT, hbr: &HBRUSH) -> SysResult<()>

FrameRect function.

pub fn InvertRect(&self, rc: RECT) -> SysResult<()>

InvertRect function.

pub fn PaintDesktop(&self) -> SysResult<()>

PaintDesktop function.

pub fn WindowFromDC(&self) -> Option<HWND>

WindowFromDC function.