winsafe

Struct HMENU

pub struct HMENU(/* private fields */);

Available on crate feature user only.

Handle to a menu.

Implementations

impl HMENU

pub const unsafe fn from_ptr(p: *mut c_void) -> Self

Constructs a new handle object by wrapping a pointer.

This method can be used as an escape hatch to interoperate with other libraries.

Safety

Be sure the pointer has the correct type and isn’t owned by anyone else, otherwise you may cause memory access violations.

pub const unsafe fn raw_copy(&self) -> Self

Returns a raw copy of the underlying handle pointer.

Safety

As the name implies, raw_copy returns a raw copy of the handle, so closing one of the copies won’t close the others. This means a handle can be used after it has been closed, what can lead to errors and undefined behavior. Even worse: sometimes Windows reuses handle values, so you can call a method on a completely different handle type, what can be catastrophic.

However, in some cases the Windows API demands a copy of the handle – raw_copy is an escape hatch to fill this gap.

pub const unsafe fn as_mut(&mut self) -> &mut *mut c_void

Returns a mutable reference to the underlying raw pointer.

This method can be used as an escape hatch to interoperate with other libraries.

Safety

This method exposes the raw pointer used by raw Windows calls. It’s an opaque pointer to an internal Windows structure, and no dereferencings should be attempted.

pub const fn ptr(&self) -> *mut c_void

Returns the underlying raw pointer.

This method exposes the raw pointer used by raw Windows calls. It’s an opaque pointer to an internal Windows structure, and no dereferencings should be attempted.

This method can be used as an escape hatch to interoperate with other libraries.

impl HMENU

pub fn append_item(&self, items: &[MenuItem<'_>]) -> SysResult<()>

A more convenient HMENU::AppendMenu.

Examples

Adding multiple entries at once, with their command IDs:

use winsafe::{self as w, prelude::*, seq_ids};

seq_ids! {
    ID_FILE_OPEN = 2001;
    ID_FILE_SAVE
    ID_FILE_EXIT
}

let hmenu: w::HMENU; // initialized somewhere

hmenu.append_item(&[
    w::MenuItem::Entry { cmd_id: ID_FILE_OPEN, text: "&Open" },
    w::MenuItem::Entry { cmd_id: ID_FILE_OPEN, text: "&Save" },
    w::MenuItem::Separator,
    w::MenuItem::Entry { cmd_id: ID_FILE_EXIT, text: "E&xit" },
])?;

pub fn item_info(&self, id_or_pos: IdPos) -> SysResult<MenuItemInfo>

Simpler version of HMENU::GetMenuItemInfo, which returns a MenuItemInfo instead of the tricky MENUITEMINFO.

Examples

use winsafe::{self as w, prelude::*};

let hmenu: w::HMENU; // initialized somewhere

let item_info = hmenu.item_info(w::IdPos::Id(0))?;
match item_info {
    w::MenuItemInfo::Entry { cmd_id, text } =>
        println!("item {} {}", cmd_id, text),
    w::MenuItemInfo::Separator =>
        println!("separator"),
    w::MenuItemInfo::Submenu { submenu, text } =>
        println!("submenu {} {}", submenu, text),
}

pub fn iter_items( &self, ) -> SysResult<impl DoubleEndedIterator<Item = SysResult<MenuItemInfo>> + '_>

Returns an iterator over all menu items, including submenus and separators.

Examples

use winsafe::{self as w, prelude::*};

let hmenu: w::HMENU; // initialized somewhere

for item_info in hmenu.iter_items()? {
    let item_info = item_info?;
    match item_info {
        w::MenuItemInfo::Entry { cmd_id, text } =>
            println!("item {} {}", cmd_id, text),
        w::MenuItemInfo::Separator =>
            println!("separator"),
        w::MenuItemInfo::Submenu { submenu, text } =>
            println!("submenu {} {}", submenu, text),
    }
}

pub fn track_popup_menu_at_point( &self, pos: POINT, hwnd_parent: &HWND, hwnd_coords_relative_to: &HWND, ) -> SysResult<()>

Shows the popup menu anchored at the given coordinates using TrackPopupMenu, and performs other needed operations.

This method will block until the menu disappears.

pub fn AppendMenu( &self, flags: MF, new_item: IdMenu<'_>, content: BmpPtrStr, ) -> SysResult<()>

AppendMenu function.

This method is rather tricky, consider using HMENU::append_item.

pub fn CheckMenuItem(&self, id_or_pos: IdPos, check: bool) -> SysResult<MF>

CheckMenuItem function.

pub fn CheckMenuRadioItem( &self, first: IdPos, last: IdPos, check: IdPos, ) -> SysResult<()>

CheckMenuRadioItem function.

Panics

Panics if first, last and check don’t use the same enum field.

pub fn CreateMenu() -> SysResult<HMENU>

CreateMenu function.

Note: If not attached to a window, must be paired with an HMENU::DestroyMenu call.

pub fn CreatePopupMenu() -> SysResult<HMENU>

CreatePopupMenu function.

Note: When a menu is attached to a window, it’s automatically destroyed along with the window. However, if the menu is not attached to any window, you must call HMENU::DestroyMenu.

pub fn DeleteMenu(&self, id_or_pos: IdPos) -> SysResult<()>

DeleteMenu function.

pub fn DestroyMenu(&mut self) -> SysResult<()>

DestroyMenu function.

After calling this method, the handle will be invalidated and further operations will fail with ERROR::INVALID_HANDLE error code.

pub fn EnableMenuItem(&self, id_or_pos: IdPos, enable: bool) -> SysResult<MF>

EnableMenuItem function.

Examples

Disabling a menu item:

use winsafe::{self as w, prelude::*, seq_ids};

seq_ids! {
    ID_FILE_OPEN = 101;
}

let hmenu: w::HMENU; // initialized somewhere

hmenu.EnableMenuItem(
    w::IdPos::Id(ID_FILE_OPEN),
    false,
)?;

Disabling multiple menu items at once:

use winsafe::{self as w, prelude::*, seq_ids};

seq_ids! {
    ID_FILE_OPEN = 201;
    ID_FILE_SAVE
}

let hmenu: w::HMENU; // initialized somewhere

[ID_FILE_OPEN, ID_FILE_SAVE]
    .into_iter()
    .try_for_each(|id|
        hmenu.EnableMenuItem(
            w::IdPos::Id(id),
            false,
        ).map(|_| ())
    )?;

pub fn GetMenuDefaultItem(&self, by_pos: bool, flags: GMDI) -> SysResult<IdPos>

GetMenuDefaultItem function.

pub fn GetMenuInfo(&self, mi: &mut MENUINFO) -> SysResult<()>

GetMenuInfo function.

pub fn GetMenuItemCount(&self) -> SysResult<u32>

GetMenuItemCount function.

pub fn GetMenuItemID(&self, item_index: i32) -> Option<u16>

GetMenuItemID function.

If item_index corresponds to a submenu, returns None.

pub fn GetMenuItemInfo( &self, id_or_pos: IdPos, mii: &mut MENUITEMINFO, ) -> SysResult<()>

GetMenuItemInfo function.

This method is rather tricky, consider using HMENU::item_info.

pub fn GetMenuState(&self, id_or_pos: IdPos) -> SysResult<MF>

GetMenuState function.

pub fn GetMenuString(&self, id_or_pos: IdPos) -> SysResult<String>

GetMenuString function.

pub fn GetSubMenu(&self, pos: u32) -> Option<HMENU>

GetSubMenu function.

pub fn InsertMenuItem( &self, id_or_pos: IdPos, mii: &MENUITEMINFO, ) -> SysResult<()>

InsertMenuItem function.

pub fn IsMenu(&self) -> bool

IsMenu function.

pub fn RemoveMenu(&self, id_or_pos: IdPos) -> SysResult<()>

RemoveMenu function.

pub fn SetMenuDefaultItem(&self, id_or_pos: IdPos) -> SysResult<()>

SetMenuDefaultItem function.

pub fn SetMenuInfo(&self, mi: &MENUINFO) -> SysResult<()>

SetMenuInfo function.

pub fn SetMenuItemBitmaps( &self, id_or_pos: IdPos, hbmp_unchecked: Option<&HBITMAP>, hbmp_checked: Option<&HBITMAP>, ) -> SysResult<()>

SetMenuItemBitmaps function.

pub fn SetMenuItemInfo( &self, id_or_pos: IdPos, mii: &MENUITEMINFO, ) -> SysResult<()>

SetMenuItemInfo function.

pub fn TrackPopupMenu( &self, flags: TPM, location: POINT, hwnd: &HWND, ) -> SysResult<Option<i32>>

TrackPopupMenu function.

Note: If you just want to display a popup menu, consider the simpler HMENU::track_popup_menu_at_point.