winsafe

Struct HPROCESS

pub struct HPROCESS(/* private fields */);

Available on crate feature kernel only.

Handle to a process. Originally just a HANDLE.

Implementations

impl HPROCESS

pub fn OpenProcessToken( &self, desired_access: TOKEN, ) -> SysResult<CloseHandleGuard<HACCESSTOKEN>>

Available on crate feature advapi only.

OpenProcessToken function.

Examples

use winsafe::{self as w, prelude::*, co};

let htoken = w::HPROCESS::GetCurrentProcess()
    .OpenProcessToken(co::TOKEN::ADJUST_PRIVILEGES | co::TOKEN::QUERY)?;

impl HPROCESS

pub const unsafe fn from_ptr(p: *mut c_void) -> Self

Constructs a new handle object by wrapping a pointer.

This method can be used as an escape hatch to interoperate with other libraries.

Safety

Be sure the pointer has the correct type and isn’t owned by anyone else, otherwise you may cause memory access violations.

pub const unsafe fn raw_copy(&self) -> Self

Returns a raw copy of the underlying handle pointer.

Safety

As the name implies, raw_copy returns a raw copy of the handle, so closing one of the copies won’t close the others. This means a handle can be used after it has been closed, what can lead to errors and undefined behavior. Even worse: sometimes Windows reuses handle values, so you can call a method on a completely different handle type, what can be catastrophic.

However, in some cases the Windows API demands a copy of the handle – raw_copy is an escape hatch to fill this gap.

pub const unsafe fn as_mut(&mut self) -> &mut *mut c_void

Returns a mutable reference to the underlying raw pointer.

This method can be used as an escape hatch to interoperate with other libraries.

Safety

This method exposes the raw pointer used by raw Windows calls. It’s an opaque pointer to an internal Windows structure, and no dereferencings should be attempted.

pub const fn ptr(&self) -> *mut c_void

Returns the underlying raw pointer.

This method exposes the raw pointer used by raw Windows calls. It’s an opaque pointer to an internal Windows structure, and no dereferencings should be attempted.

This method can be used as an escape hatch to interoperate with other libraries.

impl HPROCESS

pub fn CheckRemoteDebuggerPresent(&self) -> SysResult<bool>

CheckRemoteDebuggerPresent function.

pub fn FlushInstructionCache( &self, base_address: *mut c_void, size: usize, ) -> SysResult<()>

FlushInstructionCache function.

pub fn GetCurrentProcess() -> HPROCESS

GetCurrentProcess function.

pub fn GetExitCodeProcess(&self) -> SysResult<u32>

GetExitCodeProcess function.

pub fn GetGuiResources(&self, flags: GR) -> SysResult<u32>

GetGuiResources function.

pub fn GetPriorityClass(&self) -> SysResult<PRIORITY_CLASS>

GetPriorityClass function.

pub fn GetProcessHandleCount(&self) -> SysResult<u32>

GetProcessHandleCount function.

pub fn GetProcessId(&self) -> SysResult<u32>

GetProcessId function.

pub fn GetProcessTimes( &self, ) -> SysResult<(FILETIME, FILETIME, FILETIME, FILETIME)>

GetProcessTimes function.

Returns, respectively:

creation time;

exit time;

kernel time;

user time.

Examples

use winsafe::{self as w, prelude::*, co};

let hprocess: w::HPROCESS; // initialized somewhere

let (creation, exit, kernel, user) = hprocess.GetProcessTimes()?;

pub fn IsProcessCritical(&self) -> SysResult<bool>

IsProcessCritical function.

pub fn IsWow64Process(&self) -> SysResult<bool>

IsWow64Process function.

pub fn OpenProcess( desired_access: PROCESS, inherit_handle: bool, process_id: u32, ) -> SysResult<CloseHandleGuard<HPROCESS>>

OpenProcess function.

This method will return ERROR::INVALID_PARAMETER if you try to open a system process.

pub fn QueryFullProcessImageName( &self, flags: PROCESS_NAME, ) -> SysResult<String>

QueryFullProcessImageName function.

pub fn QueryProcessAffinityUpdateMode(&self) -> SysResult<PROCESS_AFFINITY>

QueryProcessAffinityUpdateMode function.

pub fn QueryProcessCycleTime(&self) -> SysResult<u64>

QueryProcessCycleTime function.

pub fn ReadProcessMemory( &self, base_address: *mut c_void, buffer: &mut [u8], ) -> SysResult<usize>

ReadProcessMemory function.

Reads at most buffer.len() bytes. Returns how many bytes were actually read.

pub fn SetPriorityClass(&self, prority_class: PRIORITY_CLASS) -> SysResult<()>

SetPriorityClass function.

pub fn SetProcessAffinityUpdateMode( &self, flags: PROCESS_AFFINITY, ) -> SysResult<()>

SetProcessAffinityUpdateMode function.

pub fn SetProcessPriorityBoost( &self, disable_priority_boost: bool, ) -> SysResult<()>

SetProcessPriorityBoost function.

pub fn TerminateProcess(&self, exit_code: u32) -> SysResult<()>

TerminateProcess function.

pub fn VirtualQueryEx( &self, address: Option<*mut c_void>, ) -> SysResult<MEMORY_BASIC_INFORMATION>

VirtualQueryEx function.

pub fn WaitForSingleObject(&self, milliseconds: Option<u32>) -> SysResult<WAIT>

WaitForSingleObject function.

pub fn WriteProcessMemory( &self, base_address: *mut c_void, buffer: &[u8], ) -> SysResult<usize>

WriteProcessMemory function.

Returns how many bytes were actually written.

impl HPROCESS

pub fn EmptyWorkingSet(&self) -> SysResult<()>

Available on crate feature psapi only.

EmptyWorkingSet function.

pub fn EnumProcessModules(&self) -> SysResult<Vec<HINSTANCE>>

Available on crate feature psapi only.

EnumProcessModules function.

pub fn GetMappedFileName(&self, address: *mut c_void) -> SysResult<String>

Available on crate feature psapi only.

GetMappedFileName function.

pub fn GetModuleBaseName( &self, hmodule: Option<&HINSTANCE>, ) -> SysResult<String>

Available on crate feature psapi only.

GetModuleBaseName function.

pub fn GetModuleFileNameEx( &self, hmodule: Option<&HINSTANCE>, ) -> SysResult<String>

Available on crate feature psapi only.

GetModuleFileNameEx function.

pub fn GetModuleInformation( &self, hmodule: Option<&HINSTANCE>, ) -> SysResult<MODULEINFO>

Available on crate feature psapi only.

GetModuleInformation function.

pub fn GetProcessImageFileName(&self) -> SysResult<String>

Available on crate feature psapi only.

GetProcessImageFileName function.

pub fn GetProcessMemoryInfo(&self) -> SysResult<PROCESS_MEMORY_COUNTERS_EX>

Available on crate feature psapi only.

GetProcessMemoryInfo function.

impl HPROCESS

pub unsafe fn SetUserObjectInformation<T>( &self, index: UOI, pv_info: &mut T, ) -> SysResult<()>

Available on crate feature user only.

SetUserObjectInformation function.

Safety

The pv_info type varies according to index. If you set it wrong, you’re likely to cause a buffer overrun.

pub fn WaitForInputIdle(&self, milliseconds: u32) -> SysResult<SuccessTimeout>

Available on crate feature user only.

WaitForInputIdle function.