winsafe

Struct HPROCESSLIST

pub struct HPROCESSLIST(/* private fields */);

Available on crate feature kernel only.

Handle to a process list snapshot. Originally just a HANDLE.

Implementations

impl HPROCESSLIST

pub const unsafe fn from_ptr(p: *mut c_void) -> Self

Constructs a new handle object by wrapping a pointer.

This method can be used as an escape hatch to interoperate with other libraries.

Safety

Be sure the pointer has the correct type and isn’t owned by anyone else, otherwise you may cause memory access violations.

pub const unsafe fn raw_copy(&self) -> Self

Returns a raw copy of the underlying handle pointer.

Safety

As the name implies, raw_copy returns a raw copy of the handle, so closing one of the copies won’t close the others. This means a handle can be used after it has been closed, what can lead to errors and undefined behavior. Even worse: sometimes Windows reuses handle values, so you can call a method on a completely different handle type, what can be catastrophic.

However, in some cases the Windows API demands a copy of the handle – raw_copy is an escape hatch to fill this gap.

pub const unsafe fn as_mut(&mut self) -> &mut *mut c_void

Returns a mutable reference to the underlying raw pointer.

This method can be used as an escape hatch to interoperate with other libraries.

Safety

This method exposes the raw pointer used by raw Windows calls. It’s an opaque pointer to an internal Windows structure, and no dereferencings should be attempted.

pub const fn ptr(&self) -> *mut c_void

Returns the underlying raw pointer.

This method exposes the raw pointer used by raw Windows calls. It’s an opaque pointer to an internal Windows structure, and no dereferencings should be attempted.

This method can be used as an escape hatch to interoperate with other libraries.

impl HPROCESSLIST

pub fn iter_heaps( &mut self, ) -> impl Iterator<Item = SysResult<&HEAPLIST32>> + '_

Returns an iterator over the heaps of a process, with HEAPLIST32 structs. Calls HPROCESSLIST::Heap32ListFirst and then HPROCESSLIST::Heap32ListNext consecutively.

Examples

use winsafe::{self as w, prelude::*, co};

let mut hpl = w::HPROCESSLIST::
    CreateToolhelp32Snapshot(co::TH32CS::SNAPHEAPLIST, None)?;

for heap_entry in hpl.iter_heaps() {
    let heap_entry = heap_entry?;
    let is_default_heap = heap_entry.dwFlags == co::HF32::DEFAULT;
    println!("{} {}",
        heap_entry.th32HeapID, heap_entry.th32ProcessID);
}

pub fn iter_modules( &mut self, ) -> impl Iterator<Item = SysResult<&MODULEENTRY32>> + '_

Returns an iterator over the modules of a process, with MODULEENTRY32 structs. Calls HPROCESSLIST::Module32First and then HPROCESSLIST::Module32Next consecutively.

Examples

use winsafe::{self as w, prelude::*, co};

let mut hpl = w::HPROCESSLIST::
    CreateToolhelp32Snapshot(co::TH32CS::SNAPMODULE, None)?;

for mod_entry in hpl.iter_modules() {
    let mod_entry = mod_entry?;
    println!("{} {}",
        mod_entry.szModule(), mod_entry.th32ProcessID);
}

pub fn iter_processes( &mut self, ) -> impl Iterator<Item = SysResult<&PROCESSENTRY32>> + '_

Returns an iterator over the processes of a process, with PROCESSENTRY32 structs. Calls HPROCESSLIST::Process32First and then HPROCESSLIST::Process32Next consecutively.

Examples

use winsafe::{self as w, prelude::*, co};

let mut hpl = w::HPROCESSLIST::
    CreateToolhelp32Snapshot(co::TH32CS::SNAPPROCESS, None)?;

for proc_entry in hpl.iter_processes() {
    let proc_entry = proc_entry?;
    println!("{} {} {}",
        proc_entry.szExeFile(), proc_entry.th32ProcessID, proc_entry.cntThreads);
}

pub fn iter_threads( &mut self, ) -> impl Iterator<Item = SysResult<&THREADENTRY32>> + '_

Returns an iterator over the threads of a process, with THREADENTRY32 structs. Calls HPROCESSLIST::Thread32First and then HPROCESSLIST::Thread32Next consecutively.

Examples

use winsafe::{self as w, prelude::*, co};

let mut hpl = w::HPROCESSLIST::CreateToolhelp32Snapshot(
    co::TH32CS::SNAPTHREAD,
    Some(w::GetCurrentProcessId()),
)?;

for thread_entry in hpl.iter_threads() {
    let thread_entry = thread_entry?;
    println!("{} {}",
        thread_entry.th32ThreadID, thread_entry.th32OwnerProcessID);
}

pub fn CreateToolhelp32Snapshot( flags: TH32CS, th32_process_id: Option<u32>, ) -> SysResult<CloseHandleGuard<HPROCESSLIST>>

CreateToolhelp32Snapshot function.

pub fn Heap32ListFirst(&mut self, hl: &mut HEAPLIST32) -> SysResult<bool>

HeapList32First function.

After the listing ends, the handle will be invalidated and further operations will fail with ERROR::INVALID_HANDLE error code.

Prefer using HPROCESSLIST::iter_heaps, which is simpler.

pub fn Heap32ListNext(&mut self, hl: &mut HEAPLIST32) -> SysResult<bool>

HeapList32Next function.

After the listing ends, the handle will be invalidated and further operations will fail with ERROR::INVALID_HANDLE error code.

Prefer using HPROCESSLIST::iter_heaps, which is simpler.

pub fn Module32First(&mut self, me: &mut MODULEENTRY32) -> SysResult<bool>

Module32First function.

After the listing ends, the handle will be invalidated and further operations will fail with ERROR::INVALID_HANDLE error code.

Prefer using HPROCESSLIST::iter_modules, which is simpler.

pub fn Module32Next(&mut self, me: &mut MODULEENTRY32) -> SysResult<bool>

Module32Next function.

After the listing ends, the handle will be invalidated and further operations will fail with ERROR::INVALID_HANDLE error code.

Prefer using HPROCESSLIST::iter_modules, which is simpler.

pub fn Process32First(&mut self, pe: &mut PROCESSENTRY32) -> SysResult<bool>

Process32First function.

After the listing ends, the handle will be invalidated and further operations will fail with ERROR::INVALID_HANDLE error code.

Prefer using HPROCESSLIST::iter_processes, which is simpler.

pub fn Process32Next(&mut self, pe: &mut PROCESSENTRY32) -> SysResult<bool>

Process32Next function.

After the listing ends, the handle will be invalidated and further operations will fail with ERROR::INVALID_HANDLE error code.

Prefer using HPROCESSLIST::iter_processes, which is simpler.

pub fn Thread32First(&mut self, te: &mut THREADENTRY32) -> SysResult<bool>

Thread32First function.

After the listing ends, the handle will be invalidated and further operations will fail with ERROR::INVALID_HANDLE error code.

Prefer using HPROCESSLIST::iter_threads, which is simpler.

pub fn Thread32Next(&mut self, te: &mut THREADENTRY32) -> SysResult<bool>

Thread32First function.

After the listing ends, the handle will be invalidated and further operations will fail with ERROR::INVALID_HANDLE error code.

Prefer using HPROCESSLIST::iter_threads, which is simpler.