winsafe

Struct HWND

pub struct HWND(/* private fields */);

Available on crate feature user only.

Handle to a window.

Implementations

impl HWND

pub unsafe fn DefSubclassProc<M>(&self, msg: M) -> M::RetType
where M: MsgSend,

Available on crate feature comctl only.

DefSubclassProc function.

The return type is variable, being defined by the RetType associated type of the MsgSend trait. That means each message can define its own return type.

Safety

Messages manipulate pointers, copies and window states. Improper use may lead to undefined behavior.

pub fn InitializeFlatSB(&self) -> HrResult<()>

Available on crate feature comctl only.

InitializeFlatSB function.

pub fn RemoveWindowSubclass( &self, subclass_func: SUBCLASSPROC, subclass_id: usize, ) -> SysResult<()>

Available on crate feature comctl only.

RemoveWindowSubclass function.

pub unsafe fn SetWindowSubclass( &self, subclass_proc: SUBCLASSPROC, subclass_id: usize, ref_data: usize, ) -> SysResult<()>

Available on crate feature comctl only.

SetWindowSubclass function.

Safety

You must provide a subclass procedure.

pub fn TaskDialog( &self, window_title: Option<&str>, main_instruction: Option<&str>, content: Option<&str>, common_buttons: TDCBF, icon: IconRes<'_>, ) -> HrResult<DLGID>

Available on crate feature comctl only.

TaskDialog function.

If you need more customization, see the TaskDialogIndirect function.

Examples

An information message with just an OK button:

use winsafe::{self as w, prelude::*, co};

let hwnd: w::HWND; // initialized somewhere

hwnd.TaskDialog(
    Some("Operation successful"),
    None,
    Some("The operation completed successfully."),
    co::TDCBF::OK,
    w::IconRes::Info,
)?;

Prompt the user to click OK or Cancel upon a question:

use winsafe::{self as w, prelude::*, co};

let hwnd: w::HWND; // initialized somewhere

let answer = hwnd.TaskDialog(
    Some("My app name"),
    Some("File modified"),
    Some("The file has been modified.\nProceed closing the application?"),
    co::TDCBF::OK | co::TDCBF::CANCEL,
    w::IconRes::Warn,
)?;

if answer == co::DLGID::OK {
    println!("User clicked OK.");
}

pub fn UninitializeFlatSB(&self) -> HrResult<()>

Available on crate feature comctl only.

UninitializeFlatSB function.

impl HWND

pub fn DwmExtendFrameIntoClientArea( &self, margins_inset: &MARGINS, ) -> HrResult<()>

Available on crate feature dwm only.

DwmExtendFrameIntoClientArea function.

pub fn DwmGetWindowAttribute(&self, attribute: DWMWA) -> HrResult<DwmAttr>

Available on crate feature dwm only.

DwmGetWindowAttribute function.

pub fn DwmInvalidateIconicBitmaps(&self) -> HrResult<()>

Available on crate feature dwm only.

DwmInvalidateIconicBitmaps function.

pub fn DwmModifyPreviousDxFrameDuration( &self, num_refreshes: i32, relative_to_current: bool, ) -> HrResult<()>

Available on crate feature dwm only.

DwmModifyPreviousDxFrameDuration function.

pub fn DwmSetIconicLivePreviewBitmap( &self, hbmp: HBITMAP, pt_client: Option<POINT>, sit_flags: Option<DWM_SIT>, ) -> HrResult<()>

Available on crate feature dwm only.

DwmSetIconicLivePreviewBitmap function.

pub fn DwmSetIconicThumbnail( &self, hbmp: HBITMAP, sit_flags: Option<DWM_SIT>, ) -> HrResult<()>

Available on crate feature dwm only.

DwmSetIconicThumbnail function.

pub fn DwmSetWindowAttribute(&self, attribute: DwmAttr) -> HrResult<()>

Available on crate feature dwm only.

DwmSetWindowAttribute function.

impl HWND

pub fn RegisterDragDrop(&self, drop_target: &IDropTarget) -> HrResult<()>

Available on crate feature ole only.

RegisterDragDrop function.

Note that if you don’t call OleInitialize before this function, you’ll receive an ERROR::OUTOFMEMORY error.

pub fn RevokeDragDrop(&self) -> HrResult<()>

Available on crate feature ole only.

RevokeDragDrop function.

impl HWND

pub fn DragAcceptFiles(&self, accept: bool)

Available on crate feature shell only.

DragAcceptFiles function.

pub fn ShellAbout( &self, title_bar: &str, first_line: Option<&str>, other_stuff: Option<&str>, hicon: Option<&HICON>, ) -> SysResult<()>

Available on crate feature shell only.

ShellAbout function.

pub fn ShellExecute( &self, operation: &str, file: &str, parameters: Option<&str>, directory: Option<&str>, show_cmd: SW, ) -> SysResult<()>

Available on crate feature shell only.

ShellExecute function.

impl HWND

pub const unsafe fn from_ptr(p: *mut c_void) -> Self

Constructs a new handle object by wrapping a pointer.

This method can be used as an escape hatch to interoperate with other libraries.

Safety

Be sure the pointer has the correct type and isn’t owned by anyone else, otherwise you may cause memory access violations.

pub const unsafe fn raw_copy(&self) -> Self

Returns a raw copy of the underlying handle pointer.

Safety

As the name implies, raw_copy returns a raw copy of the handle, so closing one of the copies won’t close the others. This means a handle can be used after it has been closed, what can lead to errors and undefined behavior. Even worse: sometimes Windows reuses handle values, so you can call a method on a completely different handle type, what can be catastrophic.

However, in some cases the Windows API demands a copy of the handle – raw_copy is an escape hatch to fill this gap.

pub const unsafe fn as_mut(&mut self) -> &mut *mut c_void

Returns a mutable reference to the underlying raw pointer.

This method can be used as an escape hatch to interoperate with other libraries.

Safety

This method exposes the raw pointer used by raw Windows calls. It’s an opaque pointer to an internal Windows structure, and no dereferencings should be attempted.

pub const fn ptr(&self) -> *mut c_void

Returns the underlying raw pointer.

This method exposes the raw pointer used by raw Windows calls. It’s an opaque pointer to an internal Windows structure, and no dereferencings should be attempted.

This method can be used as an escape hatch to interoperate with other libraries.

impl HWND

pub const BROADCAST: HWND

Represents all top-level windows in HWND::PostMessage and HWND::SendMessage.

pub const DESKTOP: HWND

Represents the desktop window in HWND::GetDC.

pub fn hinstance(&self) -> HINSTANCE

Calls HWND::GetWindowLongPtr to retrieve the window HINSTANCE.

pub fn is_dialog(&self) -> bool

Calls HWND::GetClassLongPtr to retrieve the class atom and check whether the window was created from a dialog resource.

pub fn set_style(&self, style: impl Into<WS>)

Calls HWND::SetWindowLongPtr to set the window styles.

pub fn set_style_ex(&self, ex_style: impl Into<WS_EX>)

Calls HWND::SetWindowLongPtr to set the extended window styles.

pub fn style(&self) -> WS

Calls HWND::GetWindowLongPtr to retrieve the window styles.

pub fn style_ex(&self) -> WS_EX

Calls HWND::GetWindowLongPtr to retrieve the extended window styles.

pub fn ArrangeIconicWindows(&self) -> SysResult<u32>

ArrangeIconicWindows function.

pub fn BeginPaint(&self) -> SysResult<EndPaintGuard<'_>>

BeginPaint function.

In the original C implementation, BeginPaint returns a handle which must be passed to EndPaint, as a cleanup operation. Also, you must allocate and pass a PAINTSTRUCT object.

Here, the cleanup is performed automatically, because BeginPaint returns an EndPaintGuard, which stores the PAINTSTRUCT and automatically calls EndPaint when the guard goes out of scope. You must, however, keep the guard alive, otherwise the cleanup will be performed right away.

Examples

use winsafe::{self as w, prelude::*};

let hwnd: w::HWND; // initialized somewhere

let hdc = hwnd.BeginPaint()?;

// do your hdc painting...

// EndPaint() called automatically

If you don’t use the returned device context handle, you must still keep the guard alive:

use winsafe::{self as w, prelude::*};

let hwnd: w::HWND; // initialized somewhere

let _hdc = hwnd.BeginPaint()?; // keep guard alive

// do your hdc painting...

// EndPaint() called automatically

pub fn BringWindowToTop(&self) -> SysResult<()>

BringWindowToTop function.

pub fn ChildWindowFromPoint(&self, pt: POINT) -> Option<HWND>

ChildWindowFromPoint function.

pub fn ClientToScreen(&self, pt: POINT) -> SysResult<POINT>

ClientToScreen function.

If you need to convert a RECT, see the HWND::ClientToScreenRc function.

pub fn ClientToScreenRc(&self, rc: RECT) -> SysResult<RECT>

ClientToScreen method for a RECT.

pub fn CloseWindow(&self) -> SysResult<()>

CloseWindow function.

Note that this method will actually minimize the window, not destroy it.

pub unsafe fn CreateWindowEx( ex_style: WS_EX, class_name: AtomStr, title: Option<&str>, style: WS, pos: POINT, size: SIZE, hwnd_parent: Option<&HWND>, hmenu: IdMenu<'_>, hinstance: &HINSTANCE, lparam: Option<isize>, ) -> SysResult<HWND>

CreateWindowEx function.

Safety

This method will create raw dynamic windows and controls outside the library safety – it’s up to you to handle all the messages. You must use a properly registered class name and, if creating a custom window, provide its own window procedure.

The usable ID range for child controls is 8 to 57,343.

pub unsafe fn DefWindowProc<M>(&self, msg: M) -> M::RetType
where M: MsgSend,

DefWindowProc function.

The return type is variable, being defined by the RetType associated type of the MsgSend trait. That means each message can define its own return type.

Safety

Messages manipulate pointers, copies and window states. Improper use may lead to undefined behavior.

pub fn DestroyWindow(&self) -> SysResult<()>

DestroyWindow function.

Usually you don’t need to call this method directly, since it’s automatically called inside the internal message loop. The ordinary way to close a window is sending a wm::Close message.

pub fn DragDetect(&self, pt: POINT) -> bool

DragDetect function.

pub fn DrawCaption( &self, hdc: &HDC, rect: RECT, flags: Option<DC>, ) -> SysResult<()>

DrawCaption function.

pub fn DrawMenuBar(&self) -> SysResult<()>

DrawMenuBar function.

pub fn EnableScrollBar(&self, sb_flags: SBB, arrows: ESB) -> SysResult<()>

EnableScrollBar function.

pub fn EnableWindow(&self, enable: bool) -> bool

EnableWindow function.

pub fn EndDialog(&self, result: isize) -> SysResult<()>

EndDialog function.

pub fn EnumChildWindows<F>(&self, func: F)
where F: FnMut(HWND) -> bool,

EnumChildWindows function.

Examples

use winsafe::{self as w, prelude::*};

let hwnd: w::HWND; // initialized somewhere

hwnd.EnumChildWindows(|hchild: w::HWND| -> bool {
    println!("Child HWND: {}", hchild);
    true
});

pub fn FindWindow( class_name: Option<AtomStr>, title: Option<&str>, ) -> SysResult<Option<HWND>>

FindWindow function.

pub fn FindWindowEx( &self, hwnd_child_after: Option<&HWND>, class_name: AtomStr, title: Option<&str>, ) -> SysResult<Option<HWND>>

FindWindowEx function.

pub fn GetActiveWindow() -> Option<HWND>

GetActiveWindow function.

pub fn GetAltTabInfo( &self, item: Option<u32>, ati: &mut ALTTABINFO, sz_item_text: Option<u32>, ) -> SysResult<String>

GetAltTabInfo function.

If item is None, the item text is not retrieved.

The sz_item_text is the maximum number of expected chars for the item text. If None, defaults to 100.

pub fn GetAncestor(&self, flags: GA) -> Option<HWND>

GetAncestor function.

pub fn GetCapture() -> Option<HWND>

GetCapture function.

pub fn GetClassLongPtr(&self, index: GCLP) -> usize

GetClassLongPtr function.

If you just want to check whether the window is a dialog, prefer using HWND::is_dialog method.

pub fn GetClassName(&self) -> SysResult<String>

GetClassName function.

pub fn GetClientRect(&self) -> SysResult<RECT>

GetClientRect function.

pub fn GetDC(&self) -> SysResult<ReleaseDCGuard<'_>>

GetDC function.

Examples

Retrieving the device context of the desktop window:

use winsafe::{self as w, prelude::*};

let hdc_desktop = w::HWND::DESKTOP.GetDC()?;

pub fn GetDCEx( &self, hrgn_clip: &HRGN, flags: DCX, ) -> SysResult<ReleaseDCGuard<'_>>

GetDCEx function.

pub fn GetDesktopWindow() -> HWND

GetDesktopWindow function.

pub fn GetDialogDpiChangeBehavior(&self) -> SysResult<DDC>

GetDialogDpiChangeBehavior function.

pub fn GetDlgCtrlID(&self) -> SysResult<u16>

GetDlgCtrlID function.

pub fn GetDlgItem(&self, ctrl_id: u16) -> SysResult<HWND>

GetDlgItem function.

pub fn GetDpiForWindow(&self) -> u32

GetDpiForWindow function.

pub fn GetFocus() -> Option<HWND>

GetFocus function.

pub fn GetForegroundWindow() -> Option<HWND>

GetForegroundWindow function.

pub fn GetLastActivePopup(&self) -> Option<HWND>

GetLastActivePopup function.

pub fn GetMenu(&self) -> Option<HMENU>

GetMenu function.

pub fn GetMenuBarInfo( &self, obj_id: OBJID, item_id: u32, mbi: &mut MENUBARINFO, ) -> SysResult<()>

GetMenuBarInfo function.

pub fn GetMenuItemRect(&self, hmenu: &HMENU, item_pos: u32) -> SysResult<RECT>

GetMenuItemRect function.

pub fn GetNextDlgGroupItem( &self, hwnd_ctrl: &HWND, previous: bool, ) -> SysResult<HWND>

GetNextDlgGroupItem function.

pub fn GetNextDlgTabItem( &self, hwnd_ctrl: &HWND, previous: bool, ) -> SysResult<HWND>

GetNextDlgTabItem function.

pub fn GetParent(&self) -> SysResult<HWND>

GetParent function.

pub fn GetScrollInfo(&self, bar: SBB, si: &mut SCROLLINFO) -> SysResult<()>

GetScrollInfo function.

pub fn GetScrollPos(&self, bar: SBB) -> SysResult<i32>

GetScrollPos function.

pub fn GetShellWindow() -> Option<HWND>

GetShellWindow function.

pub fn GetSystemMenu(&self, revert: bool) -> Option<HMENU>

GetSystemMenu function.

pub fn GetTopWindow(&self) -> SysResult<Option<HWND>>

GetTopWindow function.

pub fn GetUpdateRect(&self, erase: bool) -> Option<RECT>

GetUpdateRect function.

pub fn GetUpdateRgn(&self, hrgn: &HRGN, erase: bool) -> SysResult<REGION>

GetUpdateRgn function.

pub fn GetWindow(&self, cmd: GW) -> SysResult<HWND>

GetWindow function.

pub fn GetWindowDC(&self) -> SysResult<ReleaseDCGuard<'_>>

GetWindowDC function.

pub fn GetWindowDisplayAffinity(&self) -> SysResult<WDA>

GetWindowDisplayAffinity function.

pub fn GetWindowDpiHostingBehavior(&self) -> DPI_HOSTING_BEHAVIOR

GetWindowDpiHostingBehavior function.

pub fn GetWindowInfo(&self, wi: &mut WINDOWINFO) -> SysResult<()>

GetWindowInfo function.

pub fn GetWindowLongPtr(&self, index: GWLP) -> isize

GetWindowLong function (x32) or GetWindowLongPtr function (x64).

If you just want to retrieve the window HINSTANCE, prefer using HWND::hinstance.

If you just want to retrieve the window styles, prefer using HWND::style and HWND::style_ex.

pub fn GetWindowModuleFileName(&self) -> String

GetWindowModuleFileName function.

pub fn GetWindowPlacement(&self, wp: &mut WINDOWPLACEMENT) -> SysResult<()>

GetWindowPlacement function.

pub fn GetWindowRect(&self) -> SysResult<RECT>

GetWindowRect function.

pub fn GetWindowRgn(&self, hrgn: &HRGN) -> SysResult<REGION>

GetWindowRgn function.

pub fn GetWindowRgnBox(&self) -> SysResult<(RECT, REGION)>

GetWindowRgnBox function.

pub fn GetWindowText(&self) -> SysResult<String>

GetWindowText function.

Calls GetWindowTextLength and performs all necessary allocations, returning an ordinary String.

use winsafe::{self as w, prelude::*};

let hwnd: w::HWND; // initialized somewhere

let text = hwnd.GetWindowText()?;
println!("Text: {}", text);

pub fn GetWindowTextLength(&self) -> SysResult<i32>

GetWindowTextLength function.

Does not count the terminating null.

You usually don’t need to call this method directly, since GetWindowText returns a String, performing all necessary allocations.

pub fn GetWindowThreadProcessId(&self) -> (u32, u32)

GetWindowThreadProcessId function.

Returns thread ID and process ID, respectively.

pub fn HideCaret(&self) -> SysResult<()>

HideCaret function.

pub fn HiliteMenuItem( &self, hmenu: &HMENU, id_or_pos: IdPos, hilite: bool, ) -> bool

HiliteMenuItem function.

pub fn InheritWindowMonitor(&self, hwnd_inherit: &HWND) -> SysResult<()>

InheritWindowMonitor function.

pub fn InvalidateRect(&self, rc: Option<&RECT>, erase: bool) -> SysResult<()>

InvalidateRect function.

Examples

Most of the time you’ll just want update the entire client area:

use winsafe::{self as w, prelude::*};

let hwnd: w::HWND; // initialized somewhere

hwnd.InvalidateRect(None, true)?;

pub fn InvalidateRgn(&self, hrgn: &HRGN, erase: bool)

InvalidateRgn function.

pub fn IsChild(&self, hwnd_possible_child: &HWND) -> bool

IsChild function.

pub fn IsDialogMessage(&self, msg: &mut MSG) -> bool

IsDialogMessage function.

pub fn IsIconic(&self) -> bool

IsIconic function.

pub fn IsWindow(&self) -> bool

IsWindow function.

pub fn IsWindowEnabled(&self) -> bool

IsWindowEnabled function.

pub fn IsWindowUnicode(&self) -> bool

IsWindowUnicode function.

pub fn IsWindowVisible(&self) -> bool

IsWindowVisible function.

pub fn IsZoomed(&self) -> bool

IsZoomed function.

pub fn KillTimer(&self, event_id: usize) -> SysResult<()>

KillTimer function.

This function ends the timer calls for the given timer ID. If you don’t call this function, the timer calls will continue until the window is destroyed – at this point, any remaining timers will be automatically cleared.

pub fn LockWindowUpdate(&self) -> SysResult<()>

LockWindowUpdate function.

Examples

use winsafe::{self as w, prelude::*};

let hwnd: w::HWND; // initialized somewhere

// Lock the window – only one window can be locked at a time.
hwnd.LockWindowUpdate()?;

// After all operations, unlock the currently locked window.
w::HWND::NULL.LockWindowUpdate()?;

pub fn LogicalToPhysicalPoint(&self, pt: *mut POINT) -> SysResult<()>

LogicalToPhysicalPoint function.

pub fn MapDialogRect(&self, rc: RECT) -> SysResult<RECT>

MapDialogRect function.

pub fn MapWindowPoints( &self, hdest: &HWND, points: PtsRc<'_>, ) -> SysResult<(i16, i16)>

MapWindowPoints function.

This method can convert either a series of POINT structs or a single RECT.

Examples

use winsafe::{self as w, prelude::*};

let hwnd: w::HWND; // initialized somewhere
let hwnd_dest: w::HWND;

let mut points = vec![w::POINT::default(), w::POINT::default()];

hwnd.MapWindowPoints(
    &hwnd_dest,
    w::PtsRc::Pts(&mut points),
)?;

pub fn MessageBox( &self, text: &str, caption: &str, flags: MB, ) -> SysResult<DLGID>

MessageBox function.

Consider using the more modern HWND::TaskDialog method.

Examples

A modal message box, which blocks its parent:

use winsafe::{self as w, prelude::*, co};

let hwnd: w::HWND; // initialized somewhere

hwnd.MessageBox("Hello, world", "title",
    co::MB::OKCANCEL | co::MB::ICONINFORMATION)?;

Usually the message box has a valid parent window, however, if for some reason you don’t have a window to serve as parent, you still can show a non-modal, parent-less message box by using the null window handle:

use winsafe::{self as w, prelude::*, co};

w::HWND::NULL
    .MessageBox("Hello, world", "Title", co::MB::ICONEXCLAMATION)?;

pub fn MonitorFromWindow(&self, flags: MONITOR) -> HMONITOR

MonitorFromWindow function.

pub fn MoveWindow(&self, pos: POINT, size: SIZE, repaint: bool) -> SysResult<()>

MoveWindow function.

pub fn OpenClipboard(&self) -> SysResult<CloseClipboardGuard<'_>>

OpenClipboard function.

In the original C implementation, you must call CloseClipboard as a cleanup operation.

Here, the cleanup is performed automatically, because OpenClipboard returns a CloseClipboardGuard, which automatically calls CloseClipboard when the guard goes out of scope. You must, however, keep the guard alive, otherwise the cleanup will be performed right away.

Examples

use winsafe::{self as w, prelude::*, co};

let hwnd: w::HWND; // initialized somewhere

let hclip = hwnd.OpenClipboard()?;
let data = hclip.GetClipboardData(co::CF::TEXT)?;

You can also open the clipboard without an HWND owner:

use winsafe::{self as w, prelude::*, co};

let hclip = w::HWND::NULL.OpenClipboard()?;
let data = hclip.GetClipboardData(co::CF::TEXT)?;

pub unsafe fn PostMessage<M>(&self, msg: M) -> SysResult<()>
where M: MsgSend + Send + Copy + 'static,

PostMessage function.

Note that this method is asychronous.

Safety

Messages manipulate pointers, copies and window states. Improper use may lead to undefined behavior.

pub fn RealChildWindowFromPoint( &self, pt_parent_client_coords: POINT, ) -> Option<HWND>

RealChildWindowFromPoint function.

pub fn RealGetWindowClass(&self) -> SysResult<String>

RealGetWindowClass function.

pub fn RedrawWindow( &self, rc_update: RECT, hrgn_update: &HRGN, flags: RDW, ) -> SysResult<()>

RedrawWindow function.

pub fn RegisterHotKey( &self, id: i32, modifiers: MOD, vkey_code: VK, ) -> SysResult<()>

RegisterHotKey function.

pub fn ScreenToClient(&self, pt: POINT) -> SysResult<POINT>

ScreenToClient function.

If you need to convert a RECT, see the HWND::ScreenToClientRc function.

pub fn ScreenToClientRc(&self, rc: RECT) -> SysResult<RECT>

ScreenToClient method for a RECT.

pub fn ScrollWindowEx( &self, dx: i32, dy: i32, client_area_portion: Option<&RECT>, clipping_rect: Option<&RECT>, hrgn_update: Option<&HRGN>, updated_boundaries: Option<&mut RECT>, flags: SCROLLW, ) -> SysResult<REGION>

ScrollWindowEx function.

pub fn SendCommand(&self, cmd: AccelMenuCtrl)

SendMessage function, specialized to send a wm::Command message.

Unlike the general SendMessage, sending a command is safe.

Examples

use winsafe::{self as w, prelude::*, msg};

let hwnd: w::HWND; // initialized somewhere

const ID_MENU_FILE_OPEN: u16 = 103;

hwnd.SendCommand(
    w::AccelMenuCtrl::Menu(ID_MENU_FILE_OPEN),
);

pub unsafe fn SendMessage<M: MsgSend>(&self, msg: M) -> M::RetType

SendMessage function.

The return type is variable, being defined by the RetType associated type of the MsgSend trait. That means each message can define its own return type.

Safety

Messages manipulate pointers, copies and window states. Improper use may lead to undefined behavior.

Examples

Sending a bm::GetImage button message, which demands an image type parameter. Note that this specific message can also return an error, which is handled with ?:

use winsafe::{self as w, prelude::*, co, msg};

let hwnd: w::HWND; // initialized somewhere

let bmp = unsafe {
    hwnd.SendMessage(
        msg::bm::GetImage {
            img_type: co::IMAGE_TYPE::BITMAP,
        },
    )
}?;

Sending an em::CharFromPos edit message, which receives point coordinates and returns two values:

use winsafe::{self as w, prelude::*, msg};

let hwnd: w::HWND; // initialized somewhere

let (char_pos, line_pos) = unsafe {
    hwnd.SendMessage(
        msg::em::CharFromPos {
            coords: w::POINT::with(12, 20),
        },
    )
};

pub unsafe fn SendMessageTimeout<M: MsgSend>( &self, msg: M, flags: SMTO, timeout_ms: u32, ) -> SysResult<M::RetType>

SendMessageTimeout function.

Safety

Messages manipulate pointers, copies and window states. Improper use may lead to undefined behavior.

pub fn SetActiveWindow(&self) -> SysResult<HWND>

SetActiveWindow function.

pub fn SetCapture(&self) -> ReleaseCaptureGuard<'_>

SetCapture function.

pub fn SetDialogDpiChangeBehavior( &self, mask: DDC, values: DDC, ) -> SysResult<()>

SetDialogDpiChangeBehavior function.

pub fn SetFocus(&self) -> Option<HWND>

SetFocus function.

pub fn SetForegroundWindow(&self) -> bool

SetForegroundWindow function.

pub fn SetLayeredWindowAttributes( &self, transparency_color_key: COLORREF, alpha: u8, flags: LWA, ) -> SysResult<()>

SetLayeredWindowAttributes function.

pub fn SetMenu(&self, hmenu: &HMENU) -> SysResult<()>

SetMenu function.

pub fn SetParent(&self, hwnd_new_parent: &HWND) -> SysResult<Option<HWND>>

SetParent function.

pub fn SetScrollInfo(&self, bar: SBB, si: &SCROLLINFO, redraw: bool) -> i32

SetScrollInfo function.

pub fn ShowScrollBar(&self, bar: SBB, show: bool) -> SysResult<()>

ShowScrollBar function.

pub fn SetScrollPos(&self, b: SBB, pos: i32, redraw: bool) -> SysResult<i32>

SetScrollPos function.

pub fn SetScrollRange( &self, bar: SBB, min_pos: i32, max_pos: i32, redraw: bool, ) -> SysResult<()>

SetScrollRange function.

pub fn SetTimer( &self, event_id: usize, elapse_ms: u32, timer_func: Option<TIMERPROC>, ) -> SysResult<usize>

This method returns the timer ID, to be passed to HWND::KillTimer.

The timer calls – either wm_timer message or callback function – will continuously be executed until you call KillTimer. If you don’t call KillTimer, the timer calls will continue until the window is destroyed – at this point, any remaining timers will be automatically cleared.

Why not closures?

A common C++ technique to use closures with SetTimer is allocating a closure on the heap and use its pointer as the timer ID. When the callback function is called, the pointer is dereferenced and the closure is then executed.

The problem with this approach is that the closure must be freed after KillTimer, which can be called from anywhere, including from the closure itself – that means you must keep the pointer outside the closure and free it somehow after the closure finishes.

Such approach is, obviously, incredibly unsafe, and only possible within Rust’s rigid ownership rules if we use some sort of garbage-collection, which will free the allocated closure some time after KillTimer is called and the closure itself finishes. Since that would incur in a performance penalty, the current implementation of SetTimer will only accept ordinary function pointers, not closures.

Handling the wm_timer message is simply more practical and efficient, so the use of a callback is discouraged here.

pub fn SetWindowDisplayAffinity(&self, affinity: WDA) -> SysResult<()>

SetWindowDisplayAffinity function.

pub unsafe fn SetWindowLongPtr(&self, index: GWLP, new_long: isize) -> isize

SetWindowLongPtr function.

If you just want to set the window styles, prefer using HWND::set_style and HWND::set_style_ex.

Safety

Changing these values may potentially cause undefined behavior to the window, and passed pointers must be handled correctly.

pub fn SetWindowPlacement(&self, wp: &WINDOWPLACEMENT) -> SysResult<()>

SetWindowPlacement function.

pub fn SetWindowPos( &self, hwnd_insert_after: HwndPlace, pos: POINT, size: SIZE, flags: SWP, ) -> SysResult<()>

SetWindowPos function.

Examples

use winsafe::{self as w, prelude::*, co};

let hwnd: w::HWND; // initialized somewhere

hwnd.SetWindowPos(
    w::HwndPlace::None,
    w::POINT::with(10, 10),
    w::SIZE::default(),
    co::SWP::NOZORDER | co::SWP::NOSIZE,
)?;

pub fn SetWindowRgn(&self, hrgn: &HRGN, redraw: bool) -> SysResult<()>

SetWindowRgn function.

pub fn SetWindowText(&self, text: &str) -> SysResult<()>

SetWindowText function.

pub fn ShowCaret(&self) -> SysResult<()>

ShowCaret function.

pub fn ShowOwnedPopups(&self, show: bool) -> SysResult<()>

ShowOwnedPopups function.

pub fn ShowWindow(&self, show_cmd: SW) -> bool

ShowWindow function.

pub fn ShowWindowAsync(&self, show_cmd: SW) -> SysResult<()>

ShowWindowAsync function.

pub fn ShutdownBlockReasonCreate(&self, reason: &str) -> SysResult<()>

ShutdownBlockReasonCreate function.

pub fn ShutdownBlockReasonDestroy(&self) -> SysResult<()>

ShutdownBlockReasonDestroy function.

pub fn ShutdownBlockReasonQuery(&self) -> SysResult<String>

ShutdownBlockReasonQuery function.

pub fn TileWindows( &self, how: MDITILE, rect: Option<RECT>, kids: &[&HWND], ) -> SysResult<u16>

TileWindows function.

pub fn TranslateAccelerator( &self, haccel_table: &HACCEL, msg: &mut MSG, ) -> SysResult<()>

TranslateAccelerator function.

pub fn UnregisterHotKey(&self, id: i32) -> SysResult<()>

UnregisterHotKey function.

pub fn UpdateLayeredWindow( &self, hdc_dest: Option<&HDC>, pt_dest: Option<&POINT>, size: Option<&SIZE>, hdc_src: Option<&HDC>, pt_src: Option<&POINT>, key: COLORREF, blend: &BLENDFUNCTION, flags: ULW, ) -> SysResult<()>

UpdateLayeredWindow function.

pub fn UpdateWindow(&self) -> SysResult<()>

UpdateWindow function.

pub fn ValidateRect(&self, rc: Option<RECT>) -> SysResult<()>

ValidateRect function.

pub fn ValidateRgn(&self, hrgn: &HRGN) -> SysResult<()>

ValidateRgn function.

pub fn WindowFromPhysicalPoint(pt: POINT) -> Option<HWND>

WindowFromPhysicalPoint function.

pub fn WindowFromPoint(pt: POINT) -> Option<HWND>

WindowFromPoint function.

pub fn WinHelp(&self, help_file: &str, cmd: HELPW, data: usize) -> SysResult<()>

WinHelp function.

impl HWND

pub fn OpenThemeData(&self, class_list: &str) -> Option<CloseThemeDataGuard>

Available on crate feature uxtheme only.

OpenThemeData function.

pub fn SetWindowTheme( &self, sub_app_name: &str, sub_id_list: Option<&str>, ) -> HrResult<()>

Available on crate feature uxtheme only.

SetWindowTheme function.

impl HWND

pub fn AddPrinterConnection2( &self, name: &str, connection_info: &PRINTER_CONNECTION_INFO_1<'_>, ) -> SysResult<()>

Available on crate feature winspool only.

AddPrinterConnection2 function.

pub fn AdvancedDocumentProperties( &self, hprinter: &HPRINTER, device_name: &str, mode_input: &DEVMODE, ) -> SysResult<DEVMODE>

Available on crate feature winspool only.

AdvancedDocumentProperties function.

pub fn PrinterProperties(&self, hprinter: &HPRINTER) -> SysResult<()>

Available on crate feature winspool only.

PrinterProperties function.