Struct winsafe::Ini

pub struct Ini {
    pub sections: Vec<IniSection>,
}

Available on crate feature kernel only.

High-level abstraction to load, manage and serialize sections and key/value pairs of a .ini file.

Examples

Printing all sections, keys and values:

use winsafe::{self as w, prelude::*};

let ini = w::Ini::parse_from_file("C:\\Temp\\foo.ini")?;

for section in ini.sections.iter() {
    println!("Section: {}", section.name);
    for entry in section.entries.iter() {
        println!("Key: {}; Value: {}", entry.key, entry.val);
    }
}

Reading a value:

use winsafe::{self as w, prelude::*};

let ini = w::Ini::parse_from_file("C:\\Temp\\foo.ini")?;

println!("{}", ini.value("the_section", "the_key").unwrap());

Fields
sections: Vec<IniSection>

All the sections of the file. They can be modified at will.

Implementations

impl Ini

pub fn parse_str(contents: &str) -> Self

Parses an Ini from a string.

pub fn parse_bytes(bytes: &[u8]) -> SysResult<Self>

Parses an Ini from raw bytes with WString::parse.

pub fn parse_from_file(ini_path: &str) -> SysResult<Self>

Parses an Ini directly from a file with WString::parse. The file will be mapped in memory during reading for maximum performance.

pub fn find_section(&self, name: &str) -> Option<&IniSection>

Returns a reference to the IniSection with the given name, if any.

The search is case-insensitive.

pub fn find_section_mut(&mut self, name: &str) -> Option<&mut IniSection>

Returns a mutable reference to the IniSection with the given name, if any.

The search is case-insensitive.

pub fn serialize_to_str(&self) -> String

Serializes the sections and entries to a string.

pub fn serialize_to_bytes(&self) -> Vec<u8>

Serializes the sections and entries to raw bytes with String::into_bytes.

pub fn serialize_to_file(&self, ini_path: &str) -> SysResult<()>

Serializes the data directly to a file with String::into_bytes.

pub fn set_value(&mut self, section: &str, key: &str, new_value: &str)

If the section/key exists, changes its value, otherwise creates it.

Examples

Changing or creating a value:

use winsafe::{self as w, prelude::*};

let ini_path = "C:\\Temp\\foo.ini";
let mut ini = w::Ini::parse_from_file(ini_path)?;

ini.set_value("the_section", "the_key", "new_value");

ini.serialize_to_file(ini_path)?;

pub fn value(&self, section: &str, key: &str) -> Option<&str>

Returns a reference to the specified value, if any.

The search is case-insensitive.

pub fn value_mut(&mut self, section: &str, key: &str) -> Option<&mut String>

Returns a mutable reference to the specified value, if any.

The search is case-insensitive.

Examples

Changing an existing value:

use winsafe::{self as w, prelude::*};

let ini_path = "C:\\Temp\\foo.ini";
let mut ini = w::Ini::parse_from_file(ini_path)?;

*ini.value_mut("the_section", "the_key").unwrap() = "new_value".to_owned();

ini.serialize_to_file(ini_path)?;