winsafe

Struct LVGROUP

pub struct LVGROUP<'a, 'b, 'c, 'd, 'e, 'f, 'g> {
    pub mask: LVGF,
    pub iGroupId: i32,
    pub stateMask: LVGS,
    pub state: LVGS,
    pub uAlign: LVGA_FH,
    pub iTitleImage: i32,
    pub iExtendedImage: i32,
    pub iFirstItem: i32,
    pub cItems: u32,
    /* private fields */
}

Available on crate feature comctl only.

LVGROUP struct.

Fields
mask: LVGFiGroupId: i32stateMask: LVGSstate: LVGSuAlign: LVGA_FHiTitleImage: i32iExtendedImage: i32iFirstItem: i32cItems: u32
Implementations

impl<'a, 'b, 'c, 'd, 'e, 'f, 'g> LVGROUP<'a, 'b, 'c, 'd, 'e, 'f, 'g>

pub const fn raw_pszHeader(&self) -> (*mut u16, i32)

Returns the raw pointer to the string field, and its declared size.

This method can be used as an escape hatch to interoperate with other libraries.

pub fn pszHeader(&self) -> Option<String>

Returns the string field.

pub fn set_pszHeader(&mut self, buf: Option<&'a mut WString>)

Sets the string field.

pub const fn raw_pszFooter(&self) -> (*mut u16, i32)

Returns the raw pointer to the string field, and its declared size.

This method can be used as an escape hatch to interoperate with other libraries.

pub fn pszFooter(&self) -> Option<String>

Returns the string field.

pub fn set_pszFooter(&mut self, buf: Option<&'b mut WString>)

Sets the string field.

pub const fn raw_pszSubtitle(&self) -> (*mut u16, i32)

Returns the raw pointer to the string field, and its declared size.

This method can be used as an escape hatch to interoperate with other libraries.

pub fn pszSubtitle(&self) -> Option<String>

Returns the string field.

pub fn set_pszSubtitle(&mut self, buf: Option<&'c mut WString>)

Sets the string field.

pub const fn raw_pszTask(&self) -> (*mut u16, i32)

Returns the raw pointer to the string field, and its declared size.

This method can be used as an escape hatch to interoperate with other libraries.

pub fn pszTask(&self) -> Option<String>

Returns the string field.

pub fn set_pszTask(&mut self, buf: Option<&'d mut WString>)

Sets the string field.

pub const fn raw_pszDescriptionTop(&self) -> (*mut u16, i32)

Returns the raw pointer to the string field, and its declared size.

This method can be used as an escape hatch to interoperate with other libraries.

pub fn pszDescriptionTop(&self) -> Option<String>

Returns the string field.

pub fn set_pszDescriptionTop(&mut self, buf: Option<&'e mut WString>)

Sets the string field.

pub const fn raw_pszDescriptionBottom(&self) -> (*mut u16, i32)

Returns the raw pointer to the string field, and its declared size.

This method can be used as an escape hatch to interoperate with other libraries.

pub fn pszDescriptionBottom(&self) -> Option<String>

Returns the string field.

pub fn set_pszDescriptionBottom(&mut self, buf: Option<&'f mut WString>)

Sets the string field.

pub const fn raw_pszSubsetTitle(&self) -> (*mut u16, i32)

Returns the raw pointer to the string field, and its declared size.

This method can be used as an escape hatch to interoperate with other libraries.

pub fn pszSubsetTitle(&self) -> Option<String>

Returns the string field.

pub fn set_pszSubsetTitle(&mut self, buf: Option<&'g mut WString>)

Sets the string field.