Struct winsafe::ResourceInfo

pub struct ResourceInfo { /* fields omitted */ }

Retrieves data from an embedded resource, which can be read from an executable file or a DLL.

Examples

Reading version information:

use winsafe::{HINSTANCE, ResourceInfo};

let exe_name = HINSTANCE::NULL.GetModuleFileName()?;
let ri = ResourceInfo::read_from(&exe_name)?;

if let Some(ffi) = ri.fixed_file_info() {
    let ver = ffi.dwFileVersion();
    println!("Version: {}.{}.{}.{}",
        ver[0], ver[1], ver[2], ver[3]);
}

Reading information strings. An embedded resource can have multiple string blocks, and each block is identified by a language/code page pair. Each block can have their own information strings:

use winsafe::{HINSTANCE, ResourceInfo};

let exe_name = HINSTANCE::NULL.GetModuleFileName()?;
let ri = ResourceInfo::read_from(&exe_name)?;

if let Some(langs_cps) = ri.langs_and_code_pages() {
    for (lang, code_page) in langs_cps.iter() {

        if let Some(product_name) = ri.product_name(*lang, *code_page) {
            println!("Product name: {}", product_name);
        }

        if let Some(copyright) = ri.legal_copyright(*lang, *code_page) {
            println!("Copyright: {}", copyright);
        }

    }
}

Implementations

impl ResourceInfo

pub fn read_from(exe_file: &str) -> WinResult<ResourceInfo>

Reads and stores the resource data from an executable file or a DLL.

pub fn fixed_file_info(&self) -> Option<&VS_FIXEDFILEINFO>

Returns the version information, if any.

pub fn langs_and_code_pages(&self) -> Option<&[(LANGID, CP)]>

Return the languages and code pages available in the resource.

pub fn comments(&self, lang: LANGID, code_page: CP) -> Option<String>

Retrieves the version information string, if any.

pub fn company_name(&self, lang: LANGID, code_page: CP) -> Option<String>

Retrieves the version information string, if any.

pub fn file_description(&self, lang: LANGID, code_page: CP) -> Option<String>

Retrieves the version information string, if any.

pub fn file_version(&self, lang: LANGID, code_page: CP) -> Option<String>

Retrieves the version information string, if any.

pub fn internal_name(&self, lang: LANGID, code_page: CP) -> Option<String>

Retrieves the version information string, if any.

pub fn legal_copyright(&self, lang: LANGID, code_page: CP) -> Option<String>

Retrieves the version information string, if any.

pub fn legal_trademarks(&self, lang: LANGID, code_page: CP) -> Option<String>

Retrieves the version information string, if any.

pub fn original_filename(&self, lang: LANGID, code_page: CP) -> Option<String>

Retrieves the version information string, if any.

pub fn product_name(&self, lang: LANGID, code_page: CP) -> Option<String>

Retrieves the version information string, if any.

pub fn product_version(&self, lang: LANGID, code_page: CP) -> Option<String>

Retrieves the version information string, if any.

pub fn private_build(&self, lang: LANGID, code_page: CP) -> Option<String>

Retrieves the version information string, if any.

pub fn special_build(&self, lang: LANGID, code_page: CP) -> Option<String>

Retrieves the version information string, if any.