Struct winsafe::SHELLEXECUTEINFO

#[repr(C)]
pub struct SHELLEXECUTEINFO<'a, 'b, 'c, 'd, 'e> {
    pub fMask: SEE_MASK,
    pub hwnd: HWND,
    pub nShow: SW,
    pub hInstApp: HINSTANCE,
    pub lpIDList: *mut c_void,
    pub hkeyClass: HKEY,
    pub hProcess: HPROCESS,
    /* private fields */
}

Available on crate feature shell only.

SHELLEXECUTEINFO struct.

Fields
fMask: SEE_MASKhwnd: HWNDnShow: SWhInstApp: HINSTANCElpIDList: *mut c_voidhkeyClass: HKEYhProcess: HPROCESS
Implementations

impl<'a, 'b, 'c, 'd, 'e> SHELLEXECUTEINFO<'a, 'b, 'c, 'd, 'e>

pub fn lpVerb(&self) -> Option<String>

Returns the string field, if any.

pub fn set_lpVerb(&mut self, buf: Option<&'a mut WString>)

Sets the string field.

pub fn lpFile(&self) -> Option<String>

Returns the string field, if any.

pub fn set_lpFile(&mut self, buf: Option<&'b mut WString>)

Sets the string field.

pub fn lpParameters(&self) -> Option<String>

Returns the string field, if any.

pub fn set_lpParameters(&mut self, buf: Option<&'c mut WString>)

Sets the string field.

pub fn lpDirectory(&self) -> Option<String>

Returns the string field, if any.

pub fn set_lpDirectory(&mut self, buf: Option<&'d mut WString>)

Sets the string field.

pub fn lpClass(&self) -> Option<String>

Returns the string field, if any.

pub fn set_lpClass(&mut self, buf: Option<&'e mut WString>)

Sets the string field.

pub const fn dwHotKey(&self) -> (VK, HOTKEYF)

Retrieves the dwHotKey field.

pub fn set_dwHotKey(&mut self, val: (VK, HOTKEYF))

Sets the dwHotKey field.

pub unsafe fn hIcon_hMonitor(&self) -> (HICON, HMONITOR)

Retrieves the hIcon/hMonitor union field.

Safety

Both hIcon and hMonitor fields share the same memory space so both are returned, and you must choose which one you want.

Examples

Retrieving hIcon:

use winsafe::{self as w, prelude::*};

let sei = w::SHELLEXECUTEINFO::default();

let (hicon, _) = unsafe { sei.hIcon_hMonitor() };

pub unsafe fn set_hIcon(&mut self, hicon: &HICON)

Sets the hIcon/hMonitor union field as hIcon.

Safety

Both hIcon and hMonitor fields share the same memory space, so when setting one, you replace the other.

pub unsafe fn set_hMonitor(&mut self, hmonitor: &HMONITOR)

Sets the hIcon/hMonitor union field as hMonitor.

Safety

Both hIcon and hMonitor fields share the same memory space, so when setting one, you replace the other.