winsafe

Struct TASKDIALOGCONFIG

pub struct TASKDIALOGCONFIG<'a, 'b, 'c, 'd, 'e, 'f, 'g, 'h, 'i, 'j, 'k, 'l, 'm, 'n, 'o, 'p> {

    pub hwnd_parent: Option<&'a HWND>,
    pub hinstance: Option<&'b HINSTANCE>,
    pub flags: TDF,
    pub common_buttons: TDCBF,
    pub window_title: Option<&'c str>,
    pub main_icon: IconIdTd<'d>,
    pub main_instruction: Option<&'e str>,
    pub content: Option<&'f str>,
    pub buttons: &'g [(u16, &'h str)],
    pub default_button_id: u16,
    pub radio_buttons: &'i [(u16, &'j str)],
    pub default_radio_button_id: u16,
    pub verification_text: Option<&'k str>,
    pub more_info: Option<&'l str>,
    pub more_info_btn_expanded: Option<&'m str>,
    pub more_info_btn_collapsed: Option<&'n str>,
    pub footer_icon: IconId<'o>,
    pub footer_text: Option<&'p str>,
    pub callback: Option<Box<dyn Fn(&HWND, Tdn) -> HRESULT>>,
    pub width: u32,

}

Available on crate feature comctl only.

TASKDIALOGCONFIG struct.

Used with TaskDialogIndirect function.

Not all flags constants are available, some of them are automatically set as you fill other parameters.

Fields
hwnd_parent: Option<&'a HWND>hinstance: Option<&'b HINSTANCE>flags: TDFcommon_buttons: TDCBF

Predefined buttons. Will be placed after the custom buttons.

window_title: Option<&'c str>

Window caption. If not specified, the .exe name is used.

main_icon: IconIdTd<'d>main_instruction: Option<&'e str>

Text shown before the main content, in a larger font.

content: Option<&'f str>

The main text of the dialog.

buttons: &'g [(u16, &'h str)]

Command ID and button text. Will be placed before the predefined common_buttons.

default_button_id: u16

Any ID from common_buttons or buttons.

radio_buttons: &'i [(u16, &'j str)]

Command ID and radio button text.

default_radio_button_id: u16

Any ID from radio_buttons.

verification_text: Option<&'k str>

Text of the label of the verification check box.

more_info: Option<&'l str>

Text of the collapsible section.

more_info_btn_expanded: Option<&'m str>

Text of the button that expands/collapses more_info, when the section is expanded.

more_info_btn_collapsed: Option<&'n str>

Text of the button that expands/collapses more_info, when the section is collapsed.

footer_icon: IconId<'o>footer_text: Option<&'p str>callback: Option<Box<dyn Fn(&HWND, Tdn) -> HRESULT>>width: u32