winsafe

Struct WinUsbFreeGuard

pub struct WinUsbFreeGuard { /* private fields */ }

Available on crate feature winusb only.

RAII implementation for HUSB which automatically calls WinUsb_Free when the object goes out of scope.

Implementations

impl WinUsbFreeGuard

pub const unsafe fn new(handle: HUSB) -> Self

Constructs the guard by taking ownership of the handle.

This method can be used as an escape hatch to interoperate with other libraries.

Safety

Be sure the handle must be freed with the specified function at the end of scope.

pub fn leak(&mut self) -> HUSB

Ejects the underlying handle, leaving a Handle::INVALID in its place.

Since the internal handle will be invalidated, the destructor will not run. It’s your responsability to run it, otherwise you’ll cause a resource leak.