Module winsafe::task_dlg

Available on crate feature comctl only.

Provides high-level abstractions to TaskDialogIndirect and HWND::TaskDialog functions.

Functions

error

Displays an error modal window with an OK button.

info

Displays an information modal window with an OK button.

ok_cancel

Displays a question modal window with OK and Cancel buttons. The text of the OK button can be customized.

warn

Displays a warning modal window with an OK button.

yes_no

Displays a question modal window with Yes and No buttons.

yes_no_cancel

Displays a question modal window with Yes, No and Cancel buttons.