Crate winsafe

Windows API and GUI in safe, idiomatic Rust.

Crate • GitHub • Docs (stable) • Docs (master branch) • Examples

WinSafe has:

high-level structs to build native Win32 GUI applications;

low-level Win32 API constants, functions and structs related to GUI.

If you’re looking for a comprehensive Win32 coverage, take a look at winapi or windows crates, which are unsafe, but have everything.

Usage

Add the dependency in your Cargo.toml:

[dependencies]
winsafe = { version = "0.0.9", features = [] }

Then you must enable the Cargo features you want to be included – these modules are named after native Windows DLL and library names, mostly.

The following Cargo features are available so far:

Feature	Description

advapi	Advapi32.dll, for Windows Registry

comctl	ComCtl32.dll, for Common Controls

comdlg	ComDlg32.dll, for the old Common Dialogs

dshow	DirectShow

gdi	Gdi32.dll, the Windows GDI

gui	The WinSafe high-level GUI structs

kernel	Kernel32.dll, required by all others

msimg	Msimg32.dll

ole	OLE and basic COM support

oleaut	OLE Automation

shell	Shell32.dll, the COM-based Windows Shell

shlwapi	Shlwapi.dll, for some Shell functions

user	User32.dll, the basic Windows UI support

uxtheme	UxTheme.dll, extended UI theming

version	Version.dll, to manipulate *.exe version info

Note that a Cargo feature may depend on other features, which will be enabled automatically.

The GUI API

WinSafe features idiomatic bindings for the Win32 API, but on top of that, it features a set of high-level GUI structs, which scaffolds the boilerplate needed to build native Win32 GUI applications, event-oriented. Unless you’re doing something really specific, these high-level wrappers are highly recommended – you’ll usually start with the WindowMain.

One of the greatest strenghts of the GUI API is supporting the use of resource files, which can be created with a WYSIWYG resource editor.

GUI structs can be found in module gui.

Native function calls

The best way to understand the idea behind WinSafe bindings is comparing them to the correspondent C code.

For example, take the following C code:

HWND hwnd = GetDesktopWindow();
SetFocus(hwnd);

This is equivalent to:

use winsafe::prelude::*;
use winsafe::HWND;

let hwnd = HWND::GetDesktopWindow();
hwnd.SetFocus();

Note how GetDesktopWindow is a static method of HWND, and SetFocus is an instance method called directly upon hwnd. All native handles (HWND, HDC, HINSTANCE, etc.) are structs, thus:

native Win32 functions that return a handle are static methods in WinSafe;

native Win32 functions whose first parameter is a handle are instance methods.

Now this C code:

PostQuitMessage(0);

Is equivalent to:

use winsafe::prelude::*;
use winsafe::PostQuitMessage;

PostQuitMessage(0);

Since PostQuitMessage is a free function, it’s simply at the root of the crate.

Native constants

All native Win32 constants can be found in the co module. They’re all typed, what means that different constant types cannot be mixed (unless you explicitly say so).

Technically, each constant type is simply a newtype with a couple implementations, including those allowing bitflag operations. Also, all constant values can be converted to its underlying integer type.

The name of the constant type is often its prefix. For example, constants of MessageBox function, like MB_OKCANCEL, belong to a type called MB.

For example, take the following C code:

let hwnd = GetDesktopWindow();
MessageBox(hwnd, "Hello, world", "My hello", MB_OKCANCEL | MB_ICONINFORMATION);

This is equivalent to:

use winsafe::prelude::*;
use winsafe::{co::MB, HWND};

let hwnd = HWND::GetDesktopWindow();
hwnd.MessageBox("Hello, world", "Title", MB::OKCANCEL | MB::ICONINFORMATION)?;

The method MessageBox, like most functions that can return errors, will return WinResult, which can contain an ERROR constant.

Native structs

WinSafe implements native Win32 structs in a very restricted way. First off, fields which control the size of the struct – often named cbSize – are private and automatically set when the struct is instantiated.

Pointer fields are also private, and they can be set and retrieved only through getter and setter methods. In particular, when setting a string pointer field, you need to pass a reference to a WString buffer, which will keep the actual string contents.

For example, the following C code:

WNDCLASSEX wcx = {0};
wcx.cbSize = sizeof(WNDCLASSEX);
wcx.lpszClassName = "MY_WINDOW";

if (RegisterClassEx(&wcx) == 0) {
    DWORD err = GetLastError();
    // handle error...
}

Is equivalent to:

use winsafe::prelude::*;
use winsafe::{RegisterClassEx, WNDCLASSEX, WString};

let mut wcx = WNDCLASSEX::default();

let mut buf = WString::from_str("MY_WINDOW");
wcx.set_lpszClassName(Some(&mut buf));

if let Err(err) = RegisterClassEx(&wcx) {
    // handle error...
}

Note how you don’t need to call GetLastError to retrieve the error code: it’s returned by the method itself in the WinResult.

Text encoding

Windows natively uses Unicode UTF-16.

WinSafe uses Unicode UTF-16 internally but exposes idiomatic UTF-8, performing conversions automatically when needed, so you don’t have to worry about OsString or any low-level conversion.

However, there are cases where a string conversion is still needed, like when dealing with native Win32 structs. In such cases, you can use the WString struct, which is also capable of working as a buffer to receive text from Win32 calls.

Modules

co

Native constants.

gui

High-level GUI abstractions for user windows and native controls. They can be created programmatically or by loading resources from a .res file. These files can be created with a WYSIWYG resource editor.

msg

Parameters of window messages.

path

File path utilities.

prelude

The WinSafe prelude.

vt

Virtual tables of COM interfaces.

Macros

seq_ids

Generates sequential u16 constants starting from the given value.

Structs

ACCELuser

ACCEL struct.

ACLkernel

ACL struct.

ALTTABINFOuser

ALTTABINFO struct.

AM_MEDIA_TYPEdshow

AM_MEDIA_TYPE struct.

ATOMuser

ATOM returned by RegisterClassEx.

AccelMenuCtrlDatauser

The data of the AccelMenuCtrl Ctrl option.

BHIDshell

IShellItem::BindToHandler bhid. Just a safe abstraction over a GUID.

BITMAPgdi

BITMAP struct.

BITMAPINFOgdi

BITMAPINFO struct.

BITMAPINFOHEADERgdi

BITMAPINFOHEADER struct.

BUTTON_IMAGELISTcomctl

BUTTON_IMAGELIST struct.

BUTTON_SPLITINFOcomctl

BUTTON_SPLITINFO struct.

BY_HANDLE_FILE_INFORMATIONkernel

BY_HANDLE_FILE_INFORMATION struct.

CHOOSECOLORcomdlg

CHOOSECOLOR struct.

CLSIDole

COM class ID. Just a safe abstraction over a GUID.

COLORREFuser

COLORREF struct.

COMBOBOXINFOuser

COMBOBOXINFO struct.

COMDLG_FILTERSPECshell

COMDLG_FILTERSPEC struct.

COMPAREITEMSTRUCTuser

COMPAREITEMSTRUCT struct.

CREATESTRUCTuser

CREATESTRUCT struct.

ComPtrole

A pointer to a COM virtual table.

DATETIMEPICKERINFOcomctl

DATETIMEPICKERINFO struct.

DELETEITEMSTRUCTuser

DELETEITEMSTRUCT struct.

DEVMODEuser

DEVMODE struct.

DRAWITEMSTRUCTuser

DRAWITEMSTRUCT struct.

DVINFOdshow

DVINFO struct.

EDITBALLOONTIPcomctl

EDITBALLOONTIP struct.

FILETIMEkernel

FILETIME struct.

Filekernel

Manages an HFILE handle, which is closed automatically when the object goes out of scope.

FileMappedkernel

Manages a memory-mapped file, which can be read/written through slices. It is closed automatically when the object goes out of scope.

GUIDole

GUID struct.

GUITHREADINFOuser

GUITHREADINFO struct.

HACCELuser

Handle to an accelerator table.

HACCESSTOKENkernel

Handle to an access token. Originally just a HANDLE.

HBITMAPuser

Handle to a bitmap.

HBRUSHuser

Handle to a brush.

HCURSORuser

Handle to a cursor.

HDCuser

Handle to a device context.

HDHITTESTINFOcomctl

HDHITTESTINFO struct.

HDITEMcomctl

HDITEM struct.

HDLAYOUTcomctl

HDLAYOUT struct.

HDROPshell

Handle to an internal drop structure.

HDWPuser32

Handle to a deferred window position.

HELPINFOuser

HELPINFO struct.

HEVENTkernel

Handle to an event. Originally just a HANDLE.

HFILEkernel

Handle to a file. Originally just a HANDLE.

HFILEMAPkernel

Handle to a file mapping. Originally just a HANDLE.

HFILEMAPVIEWkernel

Address of a mapped view. Originally just an LPVOID.

HFINDFILEkernel

Handle to a file search. Originally just a HANDLE.

HFONTgdi

Handle to a font.

HGLOBALkernel

Handle to a global memory block. Originally just a HANDLE.

HHOOKuser

Handle to a hook.

HICONuser

Handle to an icon.

HIMAGELISTcomctl

Handle to an image list.

HINSTANCEkernel

Handle to an instance, same as HMODULE.

HKEYadvapi

Handle to a registry key.

HLOCALkernel

Handle to a local memory block.

HMENUuser

Handle to a menu.

HMONITORuser

Handle to a display monitor.

HPENgdi

Handle to a pen GDI object.

HPIPEkernel

Handle to an anonymous pipe. Originally just a HANDLE.

HPROCESSkernel

Handle to a process. Originally just a HANDLE.

HPROCESSLISTkernel

Handle to a process list snapshot. Originally just a HANDLE.

HRGNuser

Handle to a region GDI object.

HRSRCkernel

Handle to a resource. Originally just a HANDLE.

HRSRCMEMkernel

Handle to a resource memory block. Originally just an HGLOBAL.

HTHEMEuxtheme

Handle to a theme.

HTHREADkernel

Handle to a thread. Originally just a HANDLE.

HTREEITEMcomctl

Handle to an tree view item.

HUPDATERSRCkernel

Handle to an updateable resource. Originally just a HANDLE.

HWNDuser

Handle to a window.

IBaseFilterdshow

IBaseFilter COM interface over IBaseFilterVT.

IBindCtxshlwapi

IBindCtx COM interface over IBindCtxVT.

IDispatcholeaut

IDispatch COM interface over IDispatchVT.

IEnumFiltersdshow

IEnumFilters COM interface over IEnumFiltersVT.

IFileDialogshell

IFileDialog COM interface over IFileDialogVT.

IFileOpenDialogshell

IFileOpenDialog COM interface over IFileOpenDialogVT.

IFileSaveDialogshell

IFileSaveDialog COM interface over IFileSaveDialogVT.

IFileSinkFilterdshow

IFileSinkFilter COM interface over IFileSinkFilterVT.

IFilterGraphdshow

IFilterGraph COM interface over IFilterGraphVT.

IGraphBuilderdshow

IGraphBuilder COM interface over IGraphBuilderVT.

IIDole

COM interface ID. Just a safe abstraction over a GUID.

IMAGELISTDRAWPARAMScomctl and gdi

IMAGELISTDRAWPARAMS struct.

IMFGetServicedshow

IMFGetService COM interface over IMFGetServiceVT.

IMFVideoDisplayControldshow

IMFVideoDisplayControl COM interface over IMFVideoDisplayControlVT.

IMediaControldshow

IMediaControl COM interface over IMediaControlVT.

IMediaFilterdshow

IMediaFilter COM interface over IMediaFilterVT.

IMediaSeekingdshow

IMediaSeeking COM interface over IMediaSeekingVT. Inherits from IUnknown.

IModalWindowshell

IModalWindow COM interface over IModalWindowVT.

INITCOMMONCONTROLSEXcomctl

INITCOMMONCONTROLSEX struct

IPersistshlwapi

IPersist COM interface over IPersistVT.

IPictureoleaut

IPicture COM interface over IPictureVT.

IPindshow

IPin COM interface over IPinVT.

ISequentialStreamshlwapi

ISequentialStream COM interface over ISequentialStreamVT.

IShellItemshell

IShellItem COM interface over IShellItemVT.

IShellItemArrayshell

IShellItemArray COM interface over IShellItemArrayVT.

IStreamshlwapi

IStream COM interface over IStreamVT.

ITaskbarListshell

ITaskbarList COM interface over ITaskbarListVT.

ITaskbarList2shell

ITaskbarList2 COM interface over ITaskbarList2VT.

ITaskbarList3shell

ITaskbarList3 COM interface over ITaskbarList3VT.

ITaskbarList4shell

ITaskbarList4 COM interface over ITaskbarList4VT.

ITypeInfooleaut

ITypeInfo COM interface over ITypeInfoVT.

IUnknownole

IUnknown COM interface over IUnknownVT. It’s the base to all COM interfaces.

Inikernel

Keeps sections and key/value pairs of a .ini file, also doing parsing and serialization of the data.

IniEntrykernel

A single key/value pair of an IniSection of an Ini.

IniSectionkernel

A single section of an Ini.

KNOWNFOLDERIDshell

Constants that represent known system folders.

LANGIDkernel

LANGID language identifier.

LCIDkernel

LCID locale identifier.

LITEMcomctl

LITEM struct.

LOGBRUSHgdi

LOGBRUSH struct.

LOGFONTgdi

LOGFONT struct.

LOGPENgdi

LOGPEN struct.

LVBKIMAGEcomctl

LVBKIMAGE struct.

LVCOLUMNcomctl

LVCOLUMN struct.

LVFINDINFOcomctl

LVFINDINFO struct.

LVFOOTERINFOcomctl

LVFOOTERINFO struct.

LVFOOTERITEMcomctl

LVFOOTERITEM struct.

LVGROUPcomctl

LVGROUP struct.

LVGROUPMETRICScomctl

LVGROUPMETRICS struct.

LVHITTESTINFOcomctl

LVHITTESTINFO struct.

LVINSERTGROUPSORTEDcomctl

LVINSERTGROUPSORTED struct.

LVINSERTMARKcomctl

LVINSERTMARK struct.

LVITEMcomctl

LVITEM struct.

LVITEMINDEXcomctl

LVITEMINDEX struct.

LVSETINFOTIPcomctl

LVSETINFOTIP struct.

LVTILEINFOcomctl

LVTILEINFO struct.

LVTILEVIEWINFOcomctl

LVTILEVIEWINFO struct.

MEDIATYPEdshow

AM_MEDIA_TYPE majortype. Just a safe abstraction over a GUID.

MEDIA_FORMATdshow

AM_MEDIA_TYPE formattype, originally with FORMAT prefix. Just a safe abstraction over a GUID.

MEMORYSTATUSEXkernel

MEMORYSTATUSEX struct.

MENUBARINFOuser

MENUBARINFO struct.

MENUINFOuser

MENUINFO struct.

MENUITEMINFOuser

MENUITEMINFO struct.

MFVideoNormalizedRectdshow

MFVideoNormalizedRect struct.

MINMAXINFOuser

MINMAXINFO struct.

MONITORINFOuser

MONITORINFO struct.

MSGuser

MSG struct.

NCCALCSIZE_PARAMSuser

NCCALCSIZE_PARAMS struct.

NMBCDROPDOWNcomctl

NMBCDROPDOWN struct.

NMBCHOTITEMcomctl

NMBCHOTITEM struct.

NMCHARcomctl

NMCHAR struct.

NMCUSTOMDRAWcomctl

NMCUSTOMDRAW struct.

NMDATETIMECHANGEcomctl

NMDATETIMECHANGE struct.

NMDATETIMEFORMATcomctl

NMDATETIMEFORMAT struct.

NMDATETIMEFORMATQUERYcomctl

NMDATETIMEFORMATQUERY struct.

NMDATETIMESTRINGcomctl

NMDATETIMESTRING struct.

NMDATETIMEWMKEYDOWNcomctl

NMDATETIMEWMKEYDOWN struct.

NMDAYSTATEcomctl

NMDAYSTATE struct.

NMHDRcomctl

NMHDR struct.

NMIPADDRESScomctl

NMIPADDRESS struct.

NMITEMACTIVATEcomctl

NMITEMACTIVATE struct.

NMLINKcomctl

NMLINK struct.

NMLISTVIEWcomctl

NMLISTVIEW struct.

NMLVCACHEHINTcomctl

NMLVCACHEHINT struct.

NMLVCUSTOMDRAWcomctl

NMLVCUSTOMDRAW struct.

NMLVDISPINFOcomctl

NMLVDISPINFO struct.

NMLVEMPTYMARKUPcomctl

NMLVEMPTYMARKUP struct.

NMLVFINDITEMcomctl

NMLVFINDITEM struct.

NMLVGETINFOTIPcomctl

NMLVGETINFOTIP struct.

NMLVKEYDOWNcomctl

NMLVKEYDOWN struct.

NMLVLINKcomctl

NMLVLINK struct.

NMLVODSTATECHANGEcomctl

NMLVODSTATECHANGE struct.

NMLVSCROLLcomctl

NMLVSCROLL struct.

NMMOUSEcomctl

NMMOUSE struct.

NMSELCHANGEcomctl

NMSELCHANGE struct.

NMTRBTHUMBPOSCHANGINGcomctl

NMTRBTHUMBPOSCHANGING struct.

NMTREEVIEWcomctl

NMTREEVIEW struct.

NMTVASYNCDRAWcomctl and gdi

NMTVASYNCDRAW struct.

NMTVCUSTOMDRAWcomctl

NMTVCUSTOMDRAW stuct.

NMTVITEMCHANGEcomctl

NMTVITEMCHANGE struct.

NMVIEWCHANGEcomctl

NMVIEWCHANGE struct.

NONCLIENTMETRICSgdi

NONCLIENTMETRICS struct.

NOTIFYICONDATAshell

NOTIFYICONDATA struct.

OSVERSIONINFOEXkernel

OSVERSIONINFOEX struct.

OVERLAPPEDkernel

OVERLAPPED struct.

PAINTSTRUCTuser

PAINTSTRUCT struct.

PBRANGEcomctl

PBRANGE struct.

POINTuser

POINT struct.

PROCESSENTRY32kernel

PROCESSENTRY32 struct.

PROCESS_INFORMATIONkernel

PROCESS_INFORMATION struct.

RECTuser

RECT struct.

RGBQUADgdi

RGBQUAD struct.

ResourceInfoversion

Retrieves data from an embedded resource, which can be read from an executable file or a DLL.

ResourceInfoBlockversion

An language block of ResourceInfo, composed of a language ID and a code page.

SCROLLINFOuser

SCROLLINFO struct.

SECURITY_ATTRIBUTESkernel

SECURITY_ATTRIBUTES struct.

SECURITY_DESCRIPTORkernel

SECURITY_DESCRIPTOR struct.

SHFILEINFOshell

SHFILEINFO struct.

SHFILEOPSTRUCTshell

SHFILEOPSTRUCT struct.

SIZEuser

SIZE struct.

STARTUPINFOkernel

STARTUPINFO struct.

STYLESTRUCTuser

STYLESTRUCT struct.

SYSTEMTIMEkernel

SYSTEMTIME struct.

SYSTEM_INFOkernel

SYSTEM_INFO struct.

TASKDIALOGCONFIGcomctl

TASKDIALOGCONFIG struct.

TASKDIALOG_BUTTONcomctl

TASKDIALOG_BUTTON struct.

TBADDBITMAPcomctl

TBADDBITMAP struct.

TBBUTTONcomctl

TBBUTTON struct.

TEXTMETRICgdi

TEXTMETRIC struct.

TIME_FORMATdshow

IMediaSeeking::SetTimeFormat format. Just a safe abstraction over a GUID.

TIME_ZONE_INFORMATIONkernel

TIME_ZONE_INFORMATION struct.

TITLEBARINFOEXuser

TITLEBARINFOEX struct.

TRACKMOUSEEVENTuser

TRACKMOUSEEVENT struct.

TVINSERTSTRUCTcomctl

TVINSERTSTRUCT struct.

TVITEMcomctl

TVITEM struct.

TVITEMEXcomctl

TVITEMEX struct.

VS_FIXEDFILEINFOversion

VS_FIXEDFILEINFO struct.

WIN32_FIND_DATAkernel

WIN32_FIND_DATA struct.

WINDOWINFOuser

WINDOWINFO struct.

WINDOWPLACEMENTuser

WINDOWPLACEMENT struct.

WINDOWPOSuser

WINDOWPOS struct.

WNDCLASSEXuser

WNDCLASSEX struct.

WStringkernel

Stores a Vec<u16> buffer for a null-terminated Unicode UTF-16 wide string natively used by Windows.

Enums

AccelMenuCtrluser

Variant parameters of a wm::Command message.

AtomStruser

Variant parameter used in window class functions:

BmpIconuser

Variant parameter for:

BmpIdbRescomctl

Variant parameter for:

BmpPtrStruser

Variant parameter for:

DispfNupuser

Variant parameter for:

Encodingkernel

String encodings that can be guessed by WString::guess_encoding.

FileAccesskernel

Access types for File::open and FileMapped::open.

HwndFocususer

Variant parameter for:

HwndHmenuuser

Variant parameter for:

HwndPlaceuser

Variant parameter for:

HwndPointIduser

Variant parameter for:

IconIdcomctl

Variant parameter for:

IconIdTdiconcomctl

Variant parameter for:

IdIdcStruser

Variant parameter for:

IdIdiStruser

Variant parameter for:

IdMenuuser

Variant parameter used in menu methods:

IdPosuser

Variant parameter for:

IdStrkernel

A resource identifier.

IdTdiconStrcomctl

Variant parameter for:

IndexStrcomctl

Variant parameter for:

MenuEnumuser

Variant parameter for:

NccspRectuser

Variant parameter for:

PtIdxcomctl

Variant parameter for:

RegistryValueadvapi

Registry value types.

ResStrscomctl

Variant parameter for:

RtStrkernel

A predefined resource identifier.

TreeitemTvicomctl

Variant parameter for:

Functions

AdjustWindowRectExuser

AdjustWindowRectEx function.

AllowSetForegroundWindowuser

AllowSetForegroundWindow function

AnyPopupuser

AnyPopup function.

AttachThreadInputuser

AttachThreadInput function.

ChangeDisplaySettingsuser

ChangeDisplaySettings function.

ChooseColorcomdlg

ChooseColor function.

ClipCursoruser

ClipCursor function.

CloseClipboarduser

CloseClipboard function.

CoCreateInstanceole

CoCreateInstance function.

CoInitializeExole

CoInitializeEx function. Returns some error codes as success status.

CoTaskMemFreeole

CoTaskMemFree function.

CoUninitializeole

CoUninitialize function.

CommDlgExtendedErrorcomdlg

CommDlgExtendedError function.

CommandLineToArgvshell

CommandLineToArgv function.

CopyFilekernel

CopyFile function.

DecryptFileadvapi

DecryptFile function.

DeleteFilekernel

DeleteFile function.

DispatchMessageuser

DispatchMessage function.

EmptyClipboarduser

EmptyClipboard function.

EncryptFileadvapi

EncryptFile function.

EncryptionDisableadvapi

EncryptionDisable function.

EndMenuuser

EndMenu function.

EnumDisplaySettingsExuser

EnumDisplaySettingsEx function

EnumWindowsuser

EnumWindows function.

ExpandEnvironmentStringskernel

ExpandEnvironmentStrings function.

FileTimeToSystemTimekernel

FileTimeToSystemTime function.

GetAsyncKeyStateuser

GetAsyncKeyState function.

GetBinaryTypekernel

GetBinaryType function.

GetClipCursoruser

GetClipCursor function.

GetCommandLinekernel

GetCommandLine function.

GetComputerNamekernel

GetComputerName function.

GetCurrentDirectorykernel

GetCurrentDirectory function.

GetCurrentProcessIdkernel

GetCurrentProcessId function.

GetCurrentThreadIdkernel

GetCurrentThreadId function.

GetCursorPosuser

GetCursorPos function.

GetDialogBaseUnitsuser

GetDialogBaseUnits function.

GetDoubleClickTimeuser

GetDoubleClickTime function.

GetEnvironmentStringskernel

GetEnvironmentStrings function.

GetFileAttributeskernel

GetFileAttributes function.

GetFileVersionInfoversion

GetFileVersionInfo function.

GetFileVersionInfoSizeversion

GetFileVersionInfoSize function.

GetFirmwareTypekernel

GetFirmwareType function.

GetGUIThreadInfouser

GetGUIThreadInfo function.

GetLargePageMinimumkernel

GetLargePageMinimum function.

GetLastErrorkernel

GetLastError function.

GetLogicalDriveStringskernel

GetLogicalDriveStrings function.

GetMenuCheckMarkDimensionsuser

GetMenuCheckMarkDimensions function.

GetMessageuser

GetMessage function.

GetMessagePosuser

GetMessagePos function.

GetNativeSystemInfokernel

GetNativeSystemInfo function.

GetQueueStatususer

GetQueueStatus function.

GetStartupInfokernel

GetStartupInfo function.

GetSysColoruser

GetSysColor function.

GetSystemDirectorykernel

GetSystemDirectory function.

GetSystemInfokernel

GetSystemInfo function.

GetSystemMetricsuser

GetSystemMetrics function.

GetSystemTimekernel

GetSystemTime function.

GetSystemTimeAsFileTimekernel

GetSystemTimeAsFileTime function.

GetSystemTimePreciseAsFileTimekernel

GetSystemTimePreciseAsFileTime function.

GetSystemTimeskernel

GetSystemTimes function.

GetTempPathkernel

GetTempPath function.

GetTickCount64kernel

GetTickCount64 function.

GetUserNameadvapi

GetUserName function.

GlobalMemoryStatusExkernel

GlobalMemoryStatusEx function.

HIBYTEkernel

HIBYTE function. Originally a macro.

HIDWORDkernel

Returns the high-order u32 of an u64.

HIWORDkernel

HIWORD function. Originally a macro.

InSendMessageuser

InSendMessage function.

InSendMessageExuser

InSendMessageEx function.

InitCommonControlscomctl

InitCommonControls function.

InitCommonControlsExcomctl

InitCommonControlsEx function.

InitMUILanguagecomctl

InitMUILanguage function.

IsAppThemeduxtheme

IsAppThemed static method.

IsCompositionActiveuxtheme

IsCompositionActive static method.

IsGUIThreaduser

IsGUIThread function.

IsNativeVhdBootkernel

IsNativeVhdBoot function.

IsThemeActiveuxtheme

IsThemeActive static method.

IsThemeDialogTextureEnableduxtheme

IsThemeDialogTextureEnabled static method.

IsWindows7OrGreaterkernel

IsWindows7OrGreater function.

IsWindows8OrGreaterkernel

IsWindows8OrGreater function.

IsWindows8Point1OrGreaterkernel

IsWindows8Point1OrGreater function.

IsWindows10OrGreaterkernel

IsWindows10OrGreater function.

IsWindowsServerkernel

IsWindowsServer function.

IsWindowsVersionOrGreaterkernel

IsWindowsVersionOrGreater function.

IsWindowsVistaOrGreaterkernel

IsWindowsVistaOrGreater function.

IsWow64Messageuser

IsWow64Message function.

LOBYTEkernel

LOBYTE function. Originally a macro.

LODWORDkernel

Returns the low-order u32 of an u64.

LOWORDkernel

LOWORD function. Originally a macro.

LockSetForegroundWindowuser

LockSetForegroundWindow function.

MAKEDWORDkernel

Function that implements MAKELONG, MAKEWPARAM, and MAKELPARAM macros.

MAKEQWORDkernel

Similar to MAKEDWORD, but for u64.

MAKEWORDkernel

MAKEWORD function. Originally a macro.

MoveFilekernel

MoveFile function.

MulDivkernel

MulDiv function.

MultiByteToWideCharkernel

MultiByteToWideChar function.

OutputDebugStringkernel

OutputDebugString function.

PeekMessageuser

PeekMessage function.

PostQuitMessageuser

PostQuitMessage function.

PostThreadMessageuser

PostThreadMessage function.

QueryPerformanceCounterkernel

QueryPerformanceCounter function.

QueryPerformanceFrequencykernel

QueryPerformanceFrequency function.

RegisterClassExuser

RegisterClassEx function.

ReleaseCaptureuser

ReleaseCapture function.

ReplaceFilekernel

ReplaceFile function.

SHAddToRecentDocs⚠shell

SHAddToRecentDocs function.

SHFileOperationshell

SHFileOperation function.

SHGetFileInfoshell

SHGetFileInfo function.

SHGetKnownFolderPathshell

SHGetKnownFolderPath function.

SetCaretBlinkTimeuser

SetCaretBlinkTime function.

SetCaretPosuser

SetCaretPos function.

SetClipboardDatauser

SetClipboardData function.

SetCurrentDirectorykernel

SetCurrentDirectory function.

SetCursorPosuser

SetCursorPos function.

SetLastErrorkernel

SetLastError function.

SetProcessDPIAwareuser

SetProcessDPIAware function.

Shell_NotifyIconshell

Shell_NotifyIcon function.

ShowCursoruser

ShowCursor function.

Sleepkernel

Sleep function.

SoundSentryuser

SoundSentry function.

SystemParametersInfo⚠kernel

SystemParametersInfo function.

SystemTimeToFileTimekernel

SystemTimeToFileTime function.

SystemTimeToTzSpecificLocalTimekernel

SystemTimeToTzSpecificLocalTime function.

TaskDialogIndirectcomctl and ole

TaskDialogIndirect function.

TrackMouseEventuser

TrackMouseEvent function.

TranslateMessageuser

TranslateMessage function.

UnregisterClassuser

UnregisterClass function.

VarQueryValue⚠version

VarQueryValue function.

VerSetConditionMaskkernel

VerSetConditionMask function.

VerifyVersionInfokernel

VerifyVersionInfo function.

WaitMessageuser

WaitMessage function.

WideCharToMultiBytekernel

WideCharToMultiByte function.

Type Definitions

CCHOOKPROCcomdlg

Type alias to CCHOOKPROC callback function.

DLGPROCuser

Type alias to DLGPROC callback function.

EDITWORDBREAKPROCuser

Type alias to EDITWORDBREAKPROC callback function.

ErrResultkernel

A specialized Result which returns a Box<dyn Error + Send + Sync> on failure.

HOOKPROCuser

Type alias to HOOKPROC callback function.

HrResultole

A specialized Result for Win32 COM operations, which returns an HRESULT on failure.

PFNLVCOMPAREcomctl

Type alias to PFNLVCOMPARE callback function.

PFNLVGROUPCOMPAREcomctl

Type alias to PFNLVGROUPCOMPARE callback function.

PFTASKDIALOGCALLBACKcomctl

Type alias to PFTASKDIALOGCALLBACK calback function.

SUBCLASSPROCcomctl

Type alias to SUBCLASSPROC callback function.

TIMERPROCuser

Type alias to TIMERPROC callback function.

WNDPROCuser

Type alias to WNDPROC callback function.

WinResultkernel

A specialized Result for Win32 operations, which returns an ERROR on failure.