Crate winsafe

Win32 GUI and related APIs in safe, idiomatic Rust.

WinSafe has:

high-level structs to build native Win32 GUI applications;

low-level Win32 API constants, functions and structs related to GUI.

If you’re looking for a comprehensive Win32 coverage, take a look at winapi or windows crates, which are unsafe, but have everything.

Links:

Crate – crates.io/crates/winsafe;

GitHub – github.com/rodrigocfd/winsafe;

examples – github.com/rodrigocfd/winsafe-examples.

Usage

Add the dependency in your Cargo.toml:

[dependencies]
winsafe = "0.0.5"

To enable the DirectShow COM module, use:

[dependencies]
winsafe = { version = "0.0.5", features = ["dshow"] }

To enable the Shell COM module, use:

[dependencies]
winsafe = { version = "0.0.5", features = ["shell"] }

Modules overview

The Win32 bindings are divided into a few modules:

root – the root of the crate has Win32 free functions and structs;

co – types and values of Win32 constants;

msg – window messages;

gui – high-level GUI wrappers.

And additionally:

dshow – Win32 DirectShow COM interfaces;

shell – Win32 Shell COM interfaces.

The GUI API

WinSafe features idiomatic bindings for the Win32 API, but on top of that, it features a set of high-level GUI structs, which scaffolds the boilerplate needed to build native Win32 GUI applications, event-oriented. Unless you’re doing something really specific, these high-level wrappers are highly recommended – you’ll usually start with the WindowMain.

One of the greatest strenghts of the GUI API is supporting the use of resource files, which can be created with a WYSIWYG resource editor.

GUI structs can be found in module gui.

Native function calls

The best way to understand the idea behind WinSafe bindings is comparing them to the correspondent C code.

For example, take the following C code:

HWND hwnd = GetDesktopWindow();
SetFocus(hwnd);

This is equivalent to:

use winsafe::HWND;

let hwnd = HWND::GetDesktopWindow();
hwnd.SetFocus();

Note how GetDesktopWindow is a static method of HWND, and SetFocus is an instance method called directly upon hwnd. All native handles (HWND, HDC, HINSTANCE, etc.) are structs, thus:

native Win32 functions that return a handle are static methods in WinSafe;

native Win32 functions whose first parameter is a handle are instance methods.

Now this C code:

PostQuitMessage(0);

Is equivalent to:

use winsafe::PostQuitMessage;

PostQuitMessage(0);

Since PostQuitMessage is a free function, it’s simply at the root of the crate.

Native constants

All native Win32 constants can be found in the co module. They’re all typed, what means that different constant types cannot be mixed (unless you explicitly say so).

Technically, each constant type is simply a newtype with a couple implementations, including those allowing bitflag operations. Also, all constant values can be converted to its underlying integer type.

The name of the constant type is often its prefix. For example, constants of MessageBox function, like MB_OKCANCEL, belong to a type called MB.

For example, take the following C code:

let hwnd = GetDesktopWindow();
MessageBox(hwnd, "Hello, world", "My hello", MB_OKCANCEL | MB_ICONINFORMATION);

This is equivalent to:

use winsafe::{co::MB, HWND};

let hwnd = HWND::GetDesktopWindow();
hwnd.MessageBox("Hello, world", "Title", MB::OKCANCEL | MB::ICONINFORMATION)?;

The method MessageBox, like all native functions that can return errors, will return WinResult, which can contain an ERROR constant.

Native structs

WinSafe implements native Win32 structs in a very restricted way. First off, fields which control the size of the struct – often named cbSize – are private and automatically set when the struct is instantiated.

Pointer fields are also private, and they can be set and retrieved only through getter and setter methods. In particular, when setting a string pointer field, you need to pass a reference to a WString buffer, which will keep the actual string contents.

For example, the following C code:

WNDCLASSEX wcx = {0};
wcx.cbSize = sizeof(WNDCLASSEX);
wcx.lpszClassName = "MY_WINDOW";

if (RegisterClassEx(&wcx) == 0) {
    DWORD err = GetLastError();
    // handle error...
}

Is equivalent to:

use winsafe::{RegisterClassEx, WNDCLASSEX, WString};

let buf = WString::from_str("MY_WINDOW");

let mut wcx = WNDCLASSEX::default();
wcx.set_lpszClassName(&buf);

if let Err(err) = RegisterClassEx(&wcx) {
    // handle error...
}

Note how you don’t need to call GetLastError to retrieve the error code: it’s returned by the method itself in the WinResult.

Text encoding

Windows natively uses Unicode UTF-16.

WinSafe uses Unicode UTF-16 internally but exposes idiomatic UTF-8, performing conversions automatically when needed, so you don’t have to worry about OsString or any low-level conversion.

However, there are cases where a string conversion is still needed, like when dealing with native Win32 structs. In such cases, you can use the WString struct, which is also capable of working as a buffer to receive text from Win32 calls.

Modules

co

Win32 constants and types of constants.

dshow

DirectShow COM interfaces, structs and constants.

gui

High-level GUI abstractions for user windows and native controls. They can be created programmatically or by loading resources from a .res file. These files can be created with a WYSIWYG resource editor.

msg

Parameters of window messages.

shell

Shell COM interfaces, structs and constants.

Structs

ACCEL

ACCEL struct.

ACL

ACL struct.

ALTTABINFO

ALTTABINFO struct.

ATOM

ATOM returned by RegisterClassEx.

AccelMenuCtrlData

The data of the AccelMenuCtrl Ctrl option.

BITMAP

BITMAP struct.

BITMAPINFO

BITMAPINFO struct.

BITMAPINFOHEADER

BITMAPINFOHEADER struct.

BUTTON_IMAGELIST

BUTTON_IMAGELIST struct.

BUTTON_SPLITINFO

BUTTON_SPLITINFO struct.

BY_HANDLE_FILE_INFORMATION

BY_HANDLE_FILE_INFORMATION struct.

CHOOSECOLOR

CHOOSECOLOR struct.

CLSID

COM class ID. Just a safe abstraction over a GUID.

COLORREF

COLORREF struct.

COMBOBOXINFO

COMBOBOXINFO struct.

COMPAREITEMSTRUCT

COMPAREITEMSTRUCT struct.

CREATESTRUCT

CREATESTRUCT struct.

DATETIMEPICKERINFO

DATETIMEPICKERINFO struct.

DELETEITEMSTRUCT

DELETEITEMSTRUCT struct.

DEVMODE

DEVMODE struct.

DRAWITEMSTRUCT

DRAWITEMSTRUCT struct.

FILETIME

FILETIME struct.

GUID

GUID struct.

GUITHREADINFO

GUITHREADINFO struct.

HACCEL

Handle to an accelerator table.

HBITMAP

Handle to a bitmap.

HBRUSH

Handle to a brush.

HCURSOR

Handle to a cursor.

HDC

Handle to a device context.

HDROP

Handle to an internal drop structure.

HDWP

Handle to a deferred window position.

HELPINFO

HELPINFO struct.

HEVENT

Handle to an event. Originally just a HANDLE.

HFILE

Handle to a file. Originally just a HANDLE.

HFILEMAP

Handle to a file mapping. Originally just a HANDLE.

HFILEMAPVIEW

Address of a mapped view. Originally just an LPVOID.

HFINDFILE

Handle to a file search. Originally just a HANDLE.

HFONT

Handle to a font.

HGLOBAL

Handle to a global memory block. Originally just a HANDLE.

HHOOK

Handle to a hook.

HICON

Handle to an icon.

HIMAGELIST

Handle to an image list.

HINSTANCE

Handle to an instance, same as HMODULE.

HKEY

Handle to a registry key.

HLOCAL

Handle to a local memory block.

HMENU

Handle to a menu.

HMONITOR

Handle to a display monitor.

HPEN

Handle to a pen GDI object.

HPIPE

Handle to an anonymous pipe. Originally just a HANDLE.

HPROCESS

Handle to a process. Originally just a HANDLE.

HPROCESSLIST

Handle to a process list snapshot. Originally just a HANDLE.

HRGN

Handle to a region GDI object.

HRSRC

Handle to a resource. Originally just a HANDLE.

HRSRCMEM

Handle to a resource memory block. Originally just an HGLOBAL.

HTHEME

Handle to a theme.

HTHREAD

Handle to a thread. Originally just a HANDLE.

HTREEITEM

Handle to an tree view item.

HUPDATERSRC

Handle to an updateable resource. Originally just a HANDLE.

HWND

Handle to a window.

IDispatch

IDispatch COM interface over IDispatchVT. Inherits from IUnknown.

IDispatchVT

IDispatch virtual table.

IID

COM interface ID. Just a safe abstraction over a GUID.

IMAGELISTDRAWPARAMS

IMAGELISTDRAWPARAMS struct.

IPersist

IPersist COM interface over IPersistVT. Inherits from IUnknown.

IPersistVT

IPersist virtual table.

ITypeInfo

ITypeInfo COM interface over ITypeInfoVT. Inherits from IUnknown.

ITypeInfoVT

ITypeInfo virtual table.

IUnknown

IUnknown COM interface over IUnknownVT. It’s the base to all COM interfaces.

IUnknownVT

IUnknown virtual table, base to all COM virtual tables.

Ini

Keeps sections and key/value pairs of a .ini file, also doing parsing and serialization of the data.

IniEntry

A single key/value pair of an IniSection of an Ini.

IniSection

A single section of an Ini.

LANGID

LANGID language identifier.

LCID

LCID locale identifier.

LITEM

LITEM struct.

LOGBRUSH

LOGBRUSH struct.

LOGFONT

LOGFONT struct.

LOGPEN

LOGPEN struct.

LVCOLUMN

LVCOLUMN struct.

LVFINDINFO

LVFINDINFO struct.

LVHITTESTINFO

LVHITTESTINFO struct.

LVITEM

LVITEM struct.

LVITEMINDEX

LVITEMINDEX struct.

MEMORYSTATUSEX

MEMORYSTATUSEX struct.

MENUBARINFO

MENUBARINFO struct.

MENUINFO

MENUINFO struct.

MENUITEMINFO

MENUITEMINFO struct.

MINMAXINFO

MINMAXINFO struct.

MONITORINFO

MONITORINFO struct.

MSG

MSG struct.

MSLLHOOKSTRUCT

MSLLHOOKSTRUCT struct.

MappedFile

Manages a memory-mapped file, which can be read/written through slices.

NCCALCSIZE_PARAMS

NCCALCSIZE_PARAMS struct.

NMBCDROPDOWN

NMBCDROPDOWN struct.

NMBCHOTITEM

NMBCHOTITEM struct.

NMCHAR

NMCHAR struct.

NMCUSTOMDRAW

NMCUSTOMDRAW struct.

NMDATETIMECHANGE

NMDATETIMECHANGE struct.

NMDATETIMEFORMAT

NMDATETIMEFORMAT struct.

NMDATETIMEFORMATQUERY

NMDATETIMEFORMATQUERY struct.

NMDATETIMESTRING

NMDATETIMESTRING struct.

NMDATETIMEWMKEYDOWN

NMDATETIMEWMKEYDOWN struct.

NMDAYSTATE

NMDAYSTATE struct.

NMHDR

NMHDR struct.

NMIPADDRESS

NMIPADDRESS struct.

NMITEMACTIVATE

NMITEMACTIVATE struct.

NMLINK

NMLINK struct.

NMLISTVIEW

NMLISTVIEW struct.

NMLVCACHEHINT

NMLVCACHEHINT struct.

NMLVCUSTOMDRAW

NMLVCUSTOMDRAW struct.

NMLVDISPINFO

NMLVDISPINFO struct.

NMLVEMPTYMARKUP

NMLVEMPTYMARKUP struct.

NMLVFINDITEM

NMLVFINDITEM struct.

NMLVGETINFOTIP

NMLVGETINFOTIP struct.

NMLVKEYDOWN

NMLVKEYDOWN struct.

NMLVLINK

NMLVLINK struct.

NMLVODSTATECHANGE

NMLVODSTATECHANGE struct.

NMLVSCROLL

NMLVSCROLL struct.

NMMOUSE

NMMOUSE struct.

NMSELCHANGE

NMSELCHANGE struct.

NMTRBTHUMBPOSCHANGING

NMTRBTHUMBPOSCHANGING struct.

NMTREEVIEW

NMTREEVIEW struct.

NMTVASYNCDRAW

NMTVASYNCDRAW struct.

NMTVCUSTOMDRAW

NMTVCUSTOMDRAW stuct.

NMTVITEMCHANGE

NMTVITEMCHANGE struct.

NMVIEWCHANGE

NMVIEWCHANGE struct.

NONCLIENTMETRICS

NONCLIENTMETRICS struct.

NOTIFYICONDATA

NOTIFYICONDATA struct.

OSVERSIONINFOEX

OSVERSIONINFOEX struct.

OVERLAPPED

OVERLAPPED struct.

PAINTSTRUCT

PAINTSTRUCT struct.

PBRANGE

PBRANGE struct.

POINT

POINT struct.

PROCESSENTRY32

PROCESSENTRY32 struct.

PROCESS_INFORMATION

PROCESS_INFORMATION struct.

RECT

RECT struct.

RGBQUAD

RGBQUAD struct.

ResourceInfo

Retrieves data from an embedded resource, which can be read from an executable file or a DLL.

SCROLLINFO

SCROLLINFO struct.

SECURITY_ATTRIBUTES

SECURITY_ATTRIBUTES struct.

SECURITY_DESCRIPTOR

SECURITY_DESCRIPTOR struct.

SHFILEINFO

SHFILEINFO struct.

SHFILEOPSTRUCT

SHFILEOPSTRUCT struct.

SIZE

SIZE struct.

STARTUPINFO

STARTUPINFO struct.

STYLESTRUCT

STYLESTRUCT struct.

SYSTEMTIME

SYSTEMTIME struct.

SYSTEM_INFO

SYSTEM_INFO struct.

TASKDIALOGCONFIG

TASKDIALOGCONFIG struct.

TASKDIALOG_BUTTON

TASKDIALOG_BUTTON struct.

TBADDBITMAP

TBADDBITMAP struct.

TBBUTTON

TBBUTTON struct.

TEXTMETRIC

TEXTMETRIC struct.

TIME_ZONE_INFORMATION

TIME_ZONE_INFORMATION struct.

TITLEBARINFOEX

TITLEBARINFOEX struct.

TRACKMOUSEEVENT

TRACKMOUSEEVENT struct.

TVINSERTSTRUCT

TVINSERTSTRUCT struct.

TVITEM

TVITEM struct.

TVITEMEX

TVITEMEX struct.

VS_FIXEDFILEINFO

VS_FIXEDFILEINFO struct.

WIN32_FIND_DATA

WIN32_FIND_DATA struct.

WINDOWINFO

WINDOWINFO struct.

WINDOWPLACEMENT

WINDOWPLACEMENT struct.

WINDOWPOS

WINDOWPOS struct.

WNDCLASSEX

WNDCLASSEX struct.

WString

Stores a Vec<u16> buffer for a null-terminated Unicode UTF-16 wide string natively used by Windows.

Enums

AccelMenuCtrl

Variant parameters of a wm::Command message.

AtomStr

Variant parameter used in window class functions:

BmpIcon

Variant parameter for:

BmpIdbRes

Variant parameter for:

BmpPtrStr

Variant parameter for:

DispfNup

Variant parameter for:

Encoding

String encodings that can be guessed by WString::guess_encoding.

HwndFocus

Variant parameter for:

HwndHmenu

Variant parameter for:

HwndPlace

Variant parameter for:

HwndPointId

Variant parameter for:

IconId

Variant parameter for:

IconIdTdicon

Variant parameter for:

IdIdcStr

Variant parameter for:

IdIdiStr

Variant parameter for:

IdMenu

Variant parameter used in menu methods:

IdPos

Variant parameter for:

IdStr

Variant parameter for:

IdTdiconStr

Variant parameter for:

IndexStr

Variant parameter for:

MappedFileAccess

Access type for MappedFile::open.

MenuEnum

Variant parameter for:

NccspRect

Variant parameter for:

RegistryValue

Registry value types.

ResStrs

Variant parameter for:

RtStr

Variant parameter for:

TreeitemTvi

Variant parameter for:

Traits

ComInterface

Trait to any COM object, which encapsulates a COM interface pointer.

Functions

AdjustWindowRectEx

AdjustWindowRectEx function.

AnyPopup

AnyPopup function.

AttachThreadInput

AttachThreadInput function.

ChangeDisplaySettings

ChangeDisplaySettings function.

ChooseColor

ChooseColor function.

ClipCursor

ClipCursor function.

CloseClipboard

CloseClipboard function.

CoCreateInstance

CoCreateInstance function.

CoInitializeEx

CoInitializeEx function. Returns some error codes as success status.

CoTaskMemFree

CoTaskMemFree function.

CoUninitialize

CoUninitialize function.

CommDlgExtendedError

CommDlgExtendedError function.

CommandLineToArgv

CommandLineToArgv function.

CopyFile

CopyFile function.

DecryptFile

DecryptFile function.

DeleteFile

DeleteFile function.

DispatchMessage

DispatchMessage function.

EmptyClipboard

EmptyClipboard function.

EncryptFile

EncryptFile function.

EncryptionDisable

EncryptionDisable function.

EndMenu

EndMenu function.

EnumDisplaySettingsEx

EnumDisplaySettingsEx function

EnumWindows

EnumWindows function.

ExpandEnvironmentStrings

ExpandEnvironmentStrings function.

FileTimeToSystemTime

FileTimeToSystemTime function.

GetAsyncKeyState

GetAsyncKeyState function.

GetBinaryType

GetBinaryType function.

GetClipCursor

GetClipCursor function.

GetCommandLine

GetCommandLine function.

GetComputerName

GetComputerName function.

GetCurrentDirectory

GetCurrentDirectory function.

GetCurrentProcessId

GetCurrentProcessId function.

GetCurrentThreadId

GetCurrentThreadId function.

GetCursorPos

GetCursorPos function.

GetDialogBaseUnits

GetDialogBaseUnits function.

GetDoubleClickTime

GetDoubleClickTime function.

GetEnvironmentStrings

GetEnvironmentStrings function.

GetFileAttributes

GetFileAttributes function.

GetFileVersionInfo

GetFileVersionInfo function.

GetFileVersionInfoSize

GetFileVersionInfoSize function.

GetFirmwareType

GetFirmwareType function.

GetGUIThreadInfo

GetGUIThreadInfo function.

GetLargePageMinimum

GetLargePageMinimum function.

GetLastError

GetLastError function.

GetLogicalDriveStrings

GetLogicalDriveStrings function.

GetMenuCheckMarkDimensions

GetMenuCheckMarkDimensions function.

GetMessage

GetMessage function.

GetMessagePos

GetMessagePos function.

GetNativeSystemInfo

GetNativeSystemInfo function.

GetQueueStatus

GetQueueStatus function.

GetStartupInfo

GetStartupInfo function.

GetSysColor

GetSysColor function.

GetSystemDirectory

GetSystemDirectory function.

GetSystemInfo

GetSystemInfo function.

GetSystemMetrics

GetSystemMetrics function.

GetSystemTime

GetSystemTime function.

GetSystemTimeAsFileTime

GetSystemTimeAsFileTime function.

GetSystemTimePreciseAsFileTime

GetSystemTimePreciseAsFileTime function.

GetSystemTimes

GetSystemTimes function.

GetTempPath

GetTempPath function.

GetTickCount64

GetTickCount64 function.

GetUserName

GetUserName function.

GlobalMemoryStatusEx

GlobalMemoryStatusEx function.

HIBYTE

HIBYTE function. Originally a macro.

HIDWORD

Returns the high-order u32 of an u64.

HIWORD

HIWORD function. Originally a macro.

HRESULT_FROM_WIN32

HRESULT_FROM_WIN32 function. Originally a macro.

InitCommonControls

InitCommonControls function.

IsGUIThread

IsGUIThread function.

IsNativeVhdBoot

IsNativeVhdBoot function.

IsWindows7OrGreater

IsWindows7OrGreater function.

IsWindows8OrGreater

IsWindows8OrGreater function.

IsWindows8Point1OrGreater

IsWindows8Point1OrGreater function.

IsWindows10OrGreater

IsWindows10OrGreater function.

IsWindowsServer

IsWindowsServer function.

IsWindowsVersionOrGreater

IsWindowsVersionOrGreater function.

IsWindowsVistaOrGreater

IsWindowsVistaOrGreater function.

IsWow64Message

IsWow64Message function.

LOBYTE

LOBYTE function. Originally a macro.

LODWORD

Returns the low-order u32 of an u64.

LOWORD

LOWORD function. Originally a macro.

LockSetForegroundWindow

LockSetForegroundWindow function.

MAKEDWORD

Function that implements MAKELONG, MAKEWPARAM, and MAKELPARAM macros.

MAKEQWORD

Similar to MAKEDWORD, but for u64.

MAKEWORD

MAKEWORD function. Originally a macro.

MoveFile

MoveFile function.

MulDiv

MulDiv function.

MultiByteToWideChar

MultiByteToWideChar function.

OutputDebugString

OutputDebugString function.

PeekMessage

PeekMessage function.

PostQuitMessage

PostQuitMessage function.

QueryPerformanceCounter

QueryPerformanceCounter function.

QueryPerformanceFrequency

QueryPerformanceFrequency function.

RegisterClassEx

RegisterClassEx function.

ReleaseCapture

ReleaseCapture function.

ReplaceFile

ReplaceFile function.

SHAddToRecentDocs⚠

SHAddToRecentDocs function.

SHFileOperation

SHFileOperation function.

SHGetFileInfo

SHGetFileInfo function.

SetCaretBlinkTime

SetCaretBlinkTime function.

SetCaretPos

SetCaretPos function.

SetClipboardData

SetClipboardData function.

SetCurrentDirectory

SetCurrentDirectory function.

SetCursorPos

SetCursorPos function.

SetLastError

SetLastError function.

SetProcessDPIAware

SetProcessDPIAware function.

Shell_NotifyIcon

Shell_NotifyIcon function.

ShowCursor

ShowCursor function.

Sleep

Sleep function.

SoundSentry

SoundSentry function.

SystemParametersInfo⚠

SystemParametersInfo function.

SystemTimeToFileTime

SystemTimeToFileTime function.

SystemTimeToTzSpecificLocalTime

SystemTimeToTzSpecificLocalTime function.

TaskDialogIndirect

TaskDialogIndirect function.

TrackMouseEvent

TrackMouseEvent function.

TranslateMessage

TranslateMessage function.

UnregisterClass

UnregisterClass function.

VarQueryValue⚠

VarQueryValue function.

VerSetConditionMask

VerSetConditionMask function.

VerifyVersionInfo

VerifyVersionInfo function.

WaitMessage

WaitMessage function.

WideCharToMultiByte

WideCharToMultiByte function.

Type Definitions

BoxResult

A specialized Result which returns a Box<dyn Error> on failure.

CCHOOKPROC

Type alias to CCHOOKPROC callback function.

DLGPROC

Type alias to DLGPROC callback function.

HOOKPROC

Type alias to HOOKPROC callback function.

PFTASKDIALOGCALLBACK

Type alias to PFTASKDIALOGCALLBACK calback function.

SUBCLASSPROC

Type alias to SUBCLASSPROC callback function.

TIMERPROC

Type alias to TIMERPROC callback function.

WNDPROC

Type alias to WNDPROC callback function.

WinResult

A specialized Result for Win32 operations, which returns an ERROR on failure.