Crate winsafe

Win32 GUI and related APIs in safe, idiomatic Rust.

WinSafe exposes native Win32 constants, structs and functions related to GUI work. This means you'll find only a selected subset of the Win32 API – if you're looking for a comprehensive Win32 coverage, take a look at winapi or windows crates, which are unsafe, but have everything.

Modules overview

The Win32 bindings are divided into a few modules:

root – the root of the crate has Win32 free functions and structs;

co – types and values of Win32 constants;

msg – window messages;

gui – high-level GUI wrappers.

And additionally:

shell – Win32 shell COM interfaces.

Other modules are planned to be added in the future, featuring more COM interfaces.

Native function calls

The best way to understand the idea behind WinSafe bindings is comparing them to the correspondent C code.

For example, take the following C code:

HWND hwnd = GetDesktopWindow();
SetFocus(hwnd);

This is equivalent to:

use winsafe::HWND;

let hwnd = HWND::GetDesktopWindow();
hwnd.SetFocus();

Note how GetDesktopWindow is a static method of HWND, and SetFocus is an instance method called directly upon hwnd. All native handles (HWND, HDC, HINSTANCE, etc.) are structs, thus:

native Win32 functions that return a handle are static methods in WinSafe;

native Win32 functions whose first parameter is a handle are instance methods.

Now this C code:

PostQuitMessage(0);

Is equivalent to:

use winsafe::PostQuitMessage;

PostQuitMessage(0);

Since PostQuitMessage is a free function, it's simply at the root of the crate.

Native constants

All native Win32 constants can be found in the co module. They're all typed, what means that different constant types cannot be mixed (unless you explicitly say so).

The name of the constant type is often its prefix. For example, constants of MessageBox function, like MB_OKCANCEL, belong to a type called MB.

For example, take the following C code:

let hwnd = GetDesktopWindow();
MessageBox(hwnd, "Hello, world", "My hello", MB_OKCANCEL | MB_ICONINFORMATION);

This is equivalent to:

use winsafe::{co::MB, HWND};

let hwnd = HWND::GetDesktopWindow();
hwnd.MessageBox("Hello, world", "Title", MB::OKCANCEL | MB::ICONINFORMATION)
    .unwrap();

The method MessageBox, like all native functions that can return errors, will return WinResult, which can contain an ERROR constant.

Native structs

WinSafe implements native Win32 structs in a very restricted way. First off, fields which control the size of the struct – often named cbSize – are private and automatically set when the struct is instantiated.

Pointer fields are also private, and they can be set and retrieved only through getter and setter methods. In particular, when setting a string pointer field, you need to pass a reference to a WString buffer, which will keep the actual string contents.

For example, the following C code:

WNDCLASSEX wcx = {0};
wcx.cbSize = sizeof(WNDCLASSEX);
wcx.lpszClassName = "MY_WINDOW";

if (RegisterClassEx(&wcx) == 0) {
    DWORD err = GetLastError();
    // handle error...
}

Is equivalent to:

use winsafe::{RegisterClassEx, WNDCLASSEX, WString};

let buf = WString::from_str("MY_WINDOW");

let mut wcx = WNDCLASSEX::default();
wcx.set_lpszClassName(&buf);

if let Err(err) = RegisterClassEx(&wcx) {
    // handle error...
}

Note how you don't need to call GetLastError to retrieve the error code: it's returned by the method itself.

Text encoding

Windows natively uses Unicode UTF-16.

WinSafe uses Unicode UTF-16 internally but exposes idiomatic UTF-8, performing conversions automatically when needed, so you don't have to worry about OsString or any low-level conversion.

However, there are cases where a string conversion is still needed, like when dealing with native Win32 structs. In such cases, you can use the WString struct, which is also capable of working as a buffer to receive text from Win32 calls.

GUI

On top of the Win32 bindings, WinSafe features a set of high-level GUI structs, which scaffolds the boilerplate needed to build native Win32 GUI applications. Unless you're doing something really specific, these high-level wrappers are highly recommended.

GUI structs can be found in module gui.

Modules

co
Win32 constants and types of constants.

gui
High-level GUI abstractions.

msg
Parameters of window messages.

shell
Shell COM interfaces.

Structs

ACCEL
ACCEL struct.

ATOM
ATOM returned by RegisterClassEx.

BITMAPINFOHEADER
BITMAPINFOHEADER struct.

BUTTON_IMAGELIST
BUTTON_IMAGELIST struct.

BUTTON_SPLITINFO
BUTTON_SPLITINFO struct.

CLSID
COM class ID. Just a safe abstraction over a GUID.

COLORREF
COLORREF struct.

CREATESTRUCT
CREATESTRUCT struct.

DATETIMEPICKERINFO
DATETIMEPICKERINFO struct.

GUID
GUID struct.

HACCEL
Handle to an accelerator table. Exposes methods.

HBITMAP
Handle to a bitmap GDI object. Exposes methods.

HBRUSH
Handle to a brush GDI object. Exposes methods.

HCURSOR
Handle to a cursor. Exposes methods.

HDC
Handle to a device context. Exposes methods.

HDROP
Handle to an internal drop structure. Exposes methods.

HDWP
Handle to a deferred window position structure. Exposes methods.

HFONT
Handle to a font GDI object. Exposes methods.

HHOOK
Handle to a hook. Exposes methods.

HICON
Handle to an icon. Exposes methods.

HIMAGELIST
Handle to an image list. Exposes methods.

HINSTANCE
Handle to an instance, same as HMODULE. Exposes methods.

HKEY
Handle to a registry key. Exposes methods.

HLOCAL
Handle to a local memory block. Exposes methods.

HMENU
Handle to a menu. Exposes methods.

HPEN
Handle to a pen GDI object. Exposes methods.

HRGN
Handle to a region GDI object. Exposes methods.

HTHEME
Handle to a theme. Exposes methods.

HTREEITEM
Handle to an tree view item. Exposes methods.

HWND
Handle to a window. Exposes methods.

IID
COM interface ID. Just a safe abstraction over a GUID.

IMAGELISTDRAWPARAMS
IMAGELISTDRAWPARAMS struct.

IUnknown
IUnknown interface is the base to all COM interfaces.

IUnknownVtbl
IUnknownVtbl is the base to all COM interface virtual tables.

LITEM
LITEM struct.

LOGFONT
LOGFONT struct.

LVCOLUMN
LVCOLUMN struct.

LVFINDINFO
LVFINDINFO struct.

LVITEM
LVITEM struct.

MENUINFO
MENUINFO struct.

MENUITEMINFO
MENUITEMINFO struct.

MINMAXINFO
MINMAXINFO struct.

MSG
MSG struct.

NCCALCSIZE_PARAMS
NCCALCSIZE_PARAMS struct.

NMBCDROPDOWN
NMBCDROPDOWN struct.

NMBCHOTITEM
NMBCHOTITEM struct.

NMCHAR
NMCHAR struct.

NMCUSTOMDRAW
NMCUSTOMDRAW struct.

NMDATETIMECHANGE
NMDATETIMECHANGE struct.

NMDATETIMEFORMAT
NMDATETIMEFORMAT struct.

NMDATETIMEFORMATQUERY
NMDATETIMEFORMATQUERY struct.

NMDATETIMESTRING
NMDATETIMESTRING struct.

NMDATETIMEWMKEYDOWN
NMDATETIMEWMKEYDOWN struct.

NMHDR
NMHDR struct.

NMIPADDRESS
NMIPADDRESS struct.

NMITEMACTIVATE
NMITEMACTIVATE struct.

NMLINK
NMLINK struct.

NMLISTVIEW
NMLISTVIEW struct.

NMLVCACHEHINT
NMLVCACHEHINT struct.

NMLVDISPINFO
NMLVDISPINFO struct.

NMLVEMPTYMARKUP
NMLVEMPTYMARKUP struct.

NMLVFINDITEM
NMLVFINDITEM struct.

NMLVGETINFOTIP
NMLVGETINFOTIP struct.

NMLVKEYDOWN
NMLVKEYDOWN struct.

NMLVLINK
NMLVLINK struct.

NMLVODSTATECHANGE
NMLVODSTATECHANGE struct.

NMLVSCROLL
NMLVSCROLL struct.

NMMOUSE
NMMOUSE struct.

NMTVASYNCDRAW
NMTVASYNCDRAW struct.

NONCLIENTMETRICS
NONCLIENTMETRICS struct.

OSVERSIONINFOEX
OSVERSIONINFOEX struct.

PAINTSTRUCT
PAINTSTRUCT struct.

POINT
POINT struct.

RECT
RECT struct.

SECURITY_ATTRIBUTES
SECURITY_ATTRIBUTES struct.

SIZE
SIZE struct.

STYLESTRUCT_WS
STYLESTRUCT struct for WS.

STYLESTRUCT_WS_EX
STYLESTRUCT struct for WS_EX.

SYSTEMTIME
SYSTEMTIME struct.

TRACKMOUSEEVENT
TRACKMOUSEEVENT struct.

WINDOWINFO
WINDOWINFO struct.

WINDOWPLACEMENT
WINDOWPLACEMENT struct.

WINDOWPOS
WINDOWPOS struct.

WNDCLASSEX
WNDCLASSEX struct.

WString
Stores a Vec<u16> buffer for an Unicode UTF-16 wide string natively used by Windows.

Enums

AtomStr
Variant parameter used in window class functions:

BitmapIcon
Variant parameter for:

BitmapPtrStr
Variant parameter used in menu methods:

BroadNull
Variant parameter for:

HwndHmenu
Variant parameter for:

HwndPlace
Variant parameter for:

IdIdcStr
Variant parameter for:

IdIdiStr
Variant parameter for:

IdMenu
Variant parameter used in menu methods:

IdPos
Variant parameter used in menu methods:

IdStr
Variant parameter for:

NccspRect
Variant parameter for:

RegistryValue
Variant value returned by RegQueryValueEx.

WsWsex
Variant parameter for:

Traits

Vtbl
Trait for any COM virtual table.

Functions

AdjustWindowRectEx
AdjustWindowRectEx function.

CoCreateInstance
CoCreateInstance function.

CoInitializeEx
CoInitializeEx function. Returns some error codes as success status.

CoUninitialize
CoUninitialize function.

DispatchMessage
DispatchMessage function.

GetAsyncKeyState
GetAsyncKeyState function.

GetDialogBaseUnits
GetDialogBaseUnits function.

GetDoubleClickTime
GetDoubleClickTime function.

GetEnvironmentStrings
GetEnvironmentStrings function.

GetLastError
GetLastError function.

GetMessage
GetMessage function.

GetQueueStatus
GetQueueStatus function.

GetSysColor
GetSysColor function.

GetSystemMetrics
GetSystemMetrics function.

HIBYTE
HIBYTE function. Originally a macro.

HIDWORD
Returns the high-order u32 of an u64.

HIWORD
HIWORD function. Originally a macro.

InitCommonControls
InitCommonControls function.

IsGUIThread
IsGUIThread function.

IsWindows7OrGreater
IsWindows7OrGreater

IsWindows8OrGreater
IsWindows8OrGreater function.

IsWindows8Point1OrGreater
IsWindows8Point1OrGreater function.

IsWindows10OrGreater
IsWindows10OrGreater

IsWindowsServer
IsWindowsServer

IsWindowsVersionOrGreater
IsWindowsVersionOrGreater function.

IsWindowsVistaOrGreater
IsWindowsVistaOrGreater function.

LOBYTE
LOBYTE function. Originally a macro.

LODWORD
Returns the low-order u32 of an u64.

LOWORD
LOWORD function. Originally a macro.

LockSetForegroundWindow
LockSetForegroundWindow function.

MAKEDWORD
Function that implements MAKELONG, MAKEWPARAM, and MAKELPARAM macros.

MAKEWORD
MAKEWORD function. Originally a macro.

MulDiv
MulDiv function.

OutputDebugString
OutputDebugString function.

PeekMessage
PeekMessage function.

PostMessage
PostMessage function. Note that this function is asychronous.

PostQuitMessage
PostQuitMessage function.

RegisterClassEx
RegisterClassEx function.

SetLastError
SetLastError function.

SetProcessDPIAware
SetProcessDPIAware function.

Sleep
Sleep function.

SystemParametersInfo⚠
SystemParametersInfo function.

TrackMouseEvent
TrackMouseEvent function.

TranslateMessage
TranslateMessage function.

UnregisterClass
UnregisterClass function.

VerSetConditionMask
VerSetConditionMask function.

VerifyVersionInfo
VerifyVersionInfo function.

Type Definitions

DLGPROC
Type alias to DLGPROC callback function.

HOOKPROC
Type alias to HOOKPROC callback function.

PPVtbl
Type alias to pointer to pointer to any COM virtual table.

SUBCLASSPROC
Type alias to SUBCLASSPROC callback function.

TIMERPROC
Type alias to TIMERPROC callback function.

WNDENUMPROC
Type alias to WNDENUMPROC callbak function.

WNDPROC
Type alias to WNDPROC callback function.

WinResult
A specialized Result for Win32 operations, which return an ERROR on failure.