Crate winsafe

Windows API and GUI in safe, idiomatic Rust.

Crate • GitHub • Docs (stable) • Docs (master branch) • Examples

WinSafe has:

low-level Win32 API constants, functions and structs;

high-level structs to build native Win32 GUI applications.

Usage

Add the dependency in your Cargo.toml:

[dependencies]
winsafe = { version = "0.0.19", features = [] }

Then you must enable the Cargo features you want to be included – these modules are named after native Windows DLL and library names, mostly.

The following Cargo features are available so far:

Feature	Description

comctl	ComCtl32.dll, for Common Controls

dshow	DirectShow

dwm	Desktop Window Manager

dxgi	DirectX Graphics Infrastructure

gdi	Gdi32.dll, the Windows GDI

gui	The WinSafe high-level GUI abstractions

kernel	Kernel32.dll, Advapi32.dll and Ktmw32.dll – all others will include it

mf	Media Foundation

ole	OLE and basic COM support

oleaut	OLE Automation

shell	Shell32.dll and Shlwapi.dll, the COM-based Windows Shell

taskschd	Task Scheduler

user	User32.dll and ComDlg32.dll, the basic Windows GUI support

uxtheme	UxTheme.dll, extended window theming

version	Version.dll, to manipulate *.exe version info

If you’re looking for a comprehensive Win32 coverage, take a look at winapi or windows crates, which are unsafe, but have everything.

The GUI API

WinSafe features idiomatic bindings for the Win32 API, but on top of that, it features a set of high-level GUI structs, which scaffolds the boilerplate needed to build native Win32 GUI applications, event-oriented. Unless you’re doing something really specific, these high-level wrappers are highly recommended – you’ll usually start with the WindowMain.

One of the greatest strenghts of the GUI API is supporting the use of resource files, which can be created with a WYSIWYG resource editor.

GUI structs can be found in module gui.

Native function calls

The best way to understand the idea behind WinSafe bindings is comparing them to the correspondent C code.

For example, take the following C code:

HWND hwnd = GetDesktopWindow();
SetFocus(hwnd);

This is equivalent to:

use winsafe::{prelude::*, HWND};

let hwnd = HWND::GetDesktopWindow();
hwnd.SetFocus();

Note how GetDesktopWindow is a static method of HWND, and SetFocus is an instance method called directly upon hwnd. All native handles (HWND, HDC, HINSTANCE, etc.) are structs, thus:

native Win32 functions that return a handle are static methods in WinSafe;

native Win32 functions whose first parameter is a handle are instance methods.

Now this C code:

PostQuitMessage(0);

Is equivalent to:

use winsafe::PostQuitMessage;

PostQuitMessage(0);

Since PostQuitMessage is a free function, it’s simply at the root of the crate.

Also note that some functions which require a cleanup routine – like BeginPaint, for example – will return the resource wrapped in a guard, which will perform the cleanup automatically. You’ll never have to manually call EndPaint.

Sending messages are a special case, see the msg module.

Native constants

All native Win32 constants can be found in the co module. They’re all typed, what means that different constant types cannot be mixed (unless you explicitly say so).

Technically, each constant type is simply a newtype with a couple implementations, including those allowing bitflag operations. Also, all constant values can be converted to its underlying integer type.

The name of the constant type is often its prefix. For example, constants of MessageBox function, like MB_OKCANCEL, belong to a type called MB.

For example, take the following C code:

let hwnd = GetDesktopWindow();
MessageBox(hwnd, "Hello, world", "My hello", MB_OKCANCEL | MB_ICONINFORMATION);

This is equivalent to:

use winsafe::{prelude::*, co::MB, HWND};

let hwnd = HWND::GetDesktopWindow();
hwnd.MessageBox("Hello, world", "Title", MB::OKCANCEL | MB::ICONINFORMATION)?;

The method MessageBox, like most functions that can return errors, will return SysResult, which can contain an ERROR constant.

Native structs

WinSafe implements native Win32 structs in a very restricted way. First off, fields which control the size of the struct – often named cbSize – are private and automatically set when the struct is instantiated.

Pointer fields are also private, and they can be set and retrieved only through getter and setter methods. In particular, when setting a string pointer field, you need to pass a reference to a WString buffer, which will keep the actual string contents.

For example, the following C code:

WNDCLASSEX wcx = {0};
wcx.cbSize = sizeof(WNDCLASSEX);
wcx.lpszClassName = "MY_WINDOW";

if (RegisterClassEx(&wcx) == 0) {
    DWORD err = GetLastError();
    // handle error...
}

Is equivalent to:

use winsafe::{RegisterClassEx, WNDCLASSEX, WString};

let mut wcx = WNDCLASSEX::default();

let mut buf = WString::from_str("MY_WINDOW");
wcx.set_lpszClassName(Some(&mut buf));

if let Err(err) = RegisterClassEx(&wcx) {
    // handle error...
}

Note how you don’t need to call GetLastError to retrieve the error code: it’s returned by the method itself in the SysResult.

Text encoding

Windows natively uses Unicode UTF-16.

WinSafe uses Unicode UTF-16 internally but exposes idiomatic UTF-8, performing conversions automatically when needed, so you don’t have to worry about OsString or any low-level conversion.

However, there are cases where a string conversion is still needed, like when dealing with native Win32 structs. In such cases, you can use the WString struct, which is also capable of working as a buffer to receive text from Win32 calls.

Errors and result aliases

WinSafe declares a few Result aliases which are returned by its functions and methods:

Alias	Error	Used for

SysResult	ERROR	Standard system errors.

HrResult	HRESULT	COM errors.

AnyResult	Box<dyn Error + Send + Sync>	Holding different error types. All other Result aliases can be converted into it.

Utilities

Beyond the GUI API, WinSafe features a few high-level abstractions to deal with some particularly complex Win32 topics. Unless you need something specific, prefer using these over the raw, native calls:

Utility	Used for

Encoding	String encodings.

File	File read/write and other operations.

FileMapped	Memory-mapped file operations.

path	File path operations.

WString	Managing native wide strings.

Modules

co

Native constants.

guard

RAII implementation for various resources, which automatically perform cleanup routines when the object goes out of scope.

path

File path utilities.

prelude

The WinSafe prelude.

Macros

seq_ids

Generates sequential u16 constants starting from the given value.

Structs

ACL

ACL struct.

BY_HANDLE_FILE_INFORMATION

BY_HANDLE_FILE_INFORMATION struct.

CLAIM_SECURITY_ATTRIBUTES_INFORMATION

CLAIM_SECURITY_ATTRIBUTES_INFORMATION struct.

CLAIM_SECURITY_ATTRIBUTE_FQBN_VALUE

CLAIM_SECURITY_ATTRIBUTE_FQBN_VALUE struct.

CLAIM_SECURITY_ATTRIBUTE_OCTET_STRING_VALUE

CLAIM_SECURITY_ATTRIBUTE_OCTET_STRING_VALUE struct.

CLAIM_SECURITY_ATTRIBUTE_V1

CLAIM_SECURITY_ATTRIBUTE_V1 struct.

CONSOLE_READCONSOLE_CONTROL

CONSOLE_READCONSOLE_CONTROL struct.

DEV_BROADCAST_DEVICEINTERFACE

DEV_BROADCAST_DEVICEINTERFACE struct.

DEV_BROADCAST_HANDLE

DEV_BROADCAST_HANDLE struct.

DEV_BROADCAST_HDR

DEV_BROADCAST_HDR struct.

DEV_BROADCAST_OEM

DEV_BROADCAST_OEM struct.

DEV_BROADCAST_PORT

DEV_BROADCAST_PORT struct.

DEV_BROADCAST_VOLUME

DEV_BROADCAST_VOLUME struct.

DISK_SPACE_INFORMATION

DISK_SPACE_INFORMATION struct.

FILETIME

FILETIME struct.

File

Manages an HFILE handle, which provides file read/write and other operations. It is closed automatically when the object goes out of scope.

FileMapped

Manages an HFILEMAP handle, which provides memory-mapped file operations, including read/write through slices. It is closed automatically when the object goes out of scope.

GUID

GUID struct.

HACCESSTOKEN

Handle to an access token. Originally just a HANDLE.

HEAPLIST32

HEAPLIST32 struct.

HEVENT

Handle to a named or unnamed event object. Originally just a HANDLE.

HEVENTLOG

Handle to an event log. Originally just a HANDLE.

HFILE

Handle to a file. Originally just a HANDLE.

HFILEMAP

Handle to a file mapping. Originally just a HANDLE.

HFILEMAPVIEW

Address of a mapped view. Originally just an LPVOID.

HFINDFILE

Handle to a file search. Originally just a HANDLE.

HGLOBAL

Handle to a global memory block. Originally just a HANDLE.

HHEAP

Handle to a heap object. Originally just a HANDLE.

HINSTANCE

Handle to an instance, same as HMODULE.

HKEY

Handle to a registry key.

HLOCAL

Handle to a local memory block.

HPIPE

Handle to an anonymous pipe. Originally just a HANDLE.

HPROCESS

Handle to a process. Originally just a HANDLE.

HPROCESSLIST

Handle to a process list snapshot. Originally just a HANDLE.

HRSRC

Handle to a resource. Originally just a HANDLE.

HRSRCMEM

Handle to a resource memory block. Originally just an HGLOBAL.

HSC

Handle to a Service Control Manager. Originally SC_HANDLE.

HSERVICE

Handle to a service. Originally SC_HANDLE.

HSERVICESTATUS

Handle to a service status. Originally SERVICE_STATUS_HANDLE.

HSTD

Handle to a standard device. Originally just a HANDLE.

HTHREAD

Handle to a thread. Originally just a HANDLE.

HTRANSACTION

Handle to a transaction. Originally just a HANDLE.

HUPDATERSRC

Handle to an updateable resource. Originally just a HANDLE.

LANGID

LANGID language identifier.

LCID

LCID locale identifier.

LUID

LUID identifier.

LUID_AND_ATTRIBUTES

LUID_AND_ATTRIBUTES struct.

MEMORYSTATUSEX

MEMORYSTATUSEX struct.

MODULEENTRY32

MODULEENTRY32 struct.

OSVERSIONINFOEX

OSVERSIONINFOEX struct.

OVERLAPPED

OVERLAPPED struct.

POWERBROADCAST_SETTING

POWERBROADCAST_SETTING struct.

PROCESSENTRY32

PROCESSENTRY32 struct.

PROCESSOR_NUMBER

PROCESSOR_NUMBER struct.

PROCESS_HEAP_ENTRY

PROCESS_HEAP_ENTRY struct.

PROCESS_HEAP_ENTRY_Block

PROCESS_HEAP_ENTRY Block.

PROCESS_HEAP_ENTRY_Region

PROCESS_HEAP_ENTRY Region.

PROCESS_INFORMATION

PROCESS_INFORMATION struct.

SECURITY_ATTRIBUTES

SECURITY_ATTRIBUTES struct.

SECURITY_DESCRIPTOR

SECURITY_DESCRIPTOR struct.

SERVICE_STATUS

SERVICE_STATUS struct.

SERVICE_TIMECHANGE_INFO

SERVICE_TIMECHANGE_INFO struct.

SID

SID struct.

SID_AND_ATTRIBUTES

SID_AND_ATTRIBUTES struct.

SID_AND_ATTRIBUTES_HASH

SID_AND_ATTRIBUTES_HASH struct.

SID_IDENTIFIER_AUTHORITY

SID_IDENTIFIER_AUTHORITY struct.

STARTUPINFO

STARTUPINFO struct.

SYSTEMTIME

SYSTEMTIME struct.

SYSTEM_INFO

SYSTEM_INFO struct.

THREADENTRY32

THREADENTRY32 struct.

TIME_ZONE_INFORMATION

TIME_ZONE_INFORMATION struct.

TOKEN_ACCESS_INFORMATION

TOKEN_ACCESS_INFORMATION struct.

TOKEN_APPCONTAINER_INFORMATION

TOKEN_APPCONTAINER_INFORMATION struct.

TOKEN_DEFAULT_DACL

TOKEN_DEFAULT_DACL struct.

TOKEN_ELEVATION

TOKEN_ELEVATION struct.

TOKEN_GROUPS

TOKEN_GROUPS struct.

TOKEN_GROUPS_AND_PRIVILEGES

TOKEN_GROUPS_AND_PRIVILEGES struct.

TOKEN_LINKED_TOKEN

TOKEN_LINKED_TOKEN struct.

TOKEN_MANDATORY_LABEL

TOKEN_MANDATORY_LABEL struct.

TOKEN_MANDATORY_POLICY

TOKEN_MANDATORY_POLICY struct.

TOKEN_ORIGIN

TOKEN_ORIGIN struct.

TOKEN_OWNER

TOKEN_OWNER struct.

TOKEN_PRIMARY_GROUP

TOKEN_PRIMARY_GROUP struct.

TOKEN_PRIVILEGES

TOKEN_PRIVILEGES struct.

TOKEN_SOURCE

TOKEN_SOURCE struct.

TOKEN_STATISTICS

TOKEN_STATISTICS struct.

TOKEN_USER

TOKEN_USER struct.

VALENT

VALENT struct.

WIN32_FIND_DATA

WIN32_FIND_DATA struct.

WString

Stores a [u16] buffer for a null-terminated Unicode UTF-16 wide string natively used by Windows.

WTSSESSION_NOTIFICATION

WTSSESSION_NOTIFICATION struct.

Enums

ClaimSecurityAttr

Variable parameter for:

DisabPriv

Variable parameter for:

Encoding

String encodings.

FileAccess

Access types for File::open and FileMapped::open.

IdStr

A resource identifier.

PowerSetting

Variant parameter for:

PowerSettingAwayMode

Variant parameter for:

PowerSettingLid

Variant parameter for:

RegistryValue

Registry value types.

RtStr

A predefined resource identifier.

SvcCtl

Notification content for HSERVICESTATUS::RegisterServiceCtrlHandlerEx callback, describing co::SERVICE_CONTROL.

SvcCtlDeviceEvent

Notification content for SvcCtl.

SvcCtlPowerEvent

Notification content for SvcCtl.

TokenInfo

Variant parameter for:

Functions

AllocateAndInitializeSid

AllocateAndInitializeSid function.

ConvertSidToStringSid

ConvertSidToStringSid function.

ConvertStringSidToSid

ConvertStringSidToSid function.

CopyFile

CopyFile function.

CopySid

CopySid function.

CreateDirectory

CreateDirectory function.

CreateWellKnownSid

CreateWellKnownSid function.

DecryptFile

DecryptFile function.

DeleteFile

DeleteFile function.

EncryptFile

EncryptFile function.

EncryptionDisable

EncryptionDisable function.

EqualDomainSid

EqualDomainSid function.

EqualPrefixSid

EqualPrefixSid function.

EqualSid

EqualSid function.

ExitProcess

ExitProcess function.

ExitThread

ExitThread function.

ExpandEnvironmentStrings

ExpandEnvironmentStrings function.

FileTimeToSystemTime

FileTimeToSystemTime function.

FlushProcessWriteBuffers

FlushProcessWriteBuffers function.

FormatMessage⚠

FormatMessage function.

GetBinaryType

GetBinaryType function.

GetCommandLine

GetCommandLine function.

GetComputerName

GetComputerName function.

GetCurrentDirectory

GetCurrentDirectory function.

GetCurrentProcessId

GetCurrentProcessId function.

GetCurrentThreadId

GetCurrentThreadId function.

GetDiskFreeSpaceEx

GetDiskFreeSpaceEx function.

GetDiskSpaceInformation

GetDiskSpaceInformation function.

GetDriveType

GetDriveType function.

GetEnvironmentStrings

GetEnvironmentStrings function.

GetFileAttributes

GetFileAttributes function.

GetFirmwareType

GetFirmwareType function.

GetLargePageMinimum

GetLargePageMinimum function.

GetLastError

GetLastError function.

GetLengthSid

GetLengthSid function.

GetLocalTime

GetLocalTime function.

GetLogicalDriveStrings

GetLogicalDriveStrings function.

GetLogicalDrives

GetLogicalDrives function.

GetNativeSystemInfo

GetNativeSystemInfo function.

GetPrivateProfileSection

GetPrivateProfileSection function.

GetPrivateProfileSectionNames

GetPrivateProfileSectionNames function.

GetPrivateProfileString

GetPrivateProfileString function.

GetSidLengthRequired

GetSidLengthRequired function.

GetStartupInfo

GetStartupInfo function.

GetSystemDirectory

GetSystemDirectory function.

GetSystemFileCacheSize

GetSystemFileCacheSize function.

GetSystemInfo

GetSystemInfo function.

GetSystemTime

GetSystemTime function.

GetSystemTimeAsFileTime

GetSystemTimeAsFileTime function.

GetSystemTimePreciseAsFileTime

GetSystemTimePreciseAsFileTime function.

GetSystemTimes

GetSystemTimes function.

GetTempFileName

GetTempFileName function.

GetTempPath

GetTempPath function.

GetTickCount64

GetTickCount64 function.

GetUserName

GetUserName function.

GetVolumeInformation

GetVolumeInformation function.

GetVolumePathName

GetVolumePathName function.

GetWindowsAccountDomainSid

GetWindowsAccountDomainSid function.

GlobalMemoryStatusEx

GlobalMemoryStatusEx function.

HIBYTE

HIBYTE macro.

HIDWORD

Returns the high-order u32 of an u64.

HIWORD

HIWORD macro.

InitializeSecurityDescriptor

InitializeSecurityDescriptor function.

InitiateSystemShutdown

InitiateSystemShutdown function.

InitiateSystemShutdownEx

InitiateSystemShutdownEx function.

IsDebuggerPresent

IsDebuggerPresent function.

IsNativeVhdBoot

IsNativeVhdBoot function.

IsValidSecurityDescriptor

IsValidSecurityDescriptor function.

IsValidSid

IsValidSid function.

IsWellKnownSid

IsWellKnownSid function.

IsWindows7OrGreater

IsWindows7OrGreater function.

IsWindows8OrGreater

IsWindows8OrGreater function.

IsWindows8Point1OrGreater

IsWindows8Point1OrGreater function.

IsWindows10OrGreater

IsWindows10OrGreater function.

IsWindowsServer

IsWindowsServer function.

IsWindowsVersionOrGreater

IsWindowsVersionOrGreater function.

IsWindowsVistaOrGreater

IsWindowsVistaOrGreater function.

LOBYTE

LOBYTE macro.

LODWORD

Returns the low-order u32 of an u64.

LOWORD

LOWORD macro.

LookupAccountName

LookupAccountName function.

LookupAccountSid

LookupAccountSid function.

LookupPrivilegeName

LookupPrivilegeName function.

LookupPrivilegeValue

LookupPrivilegeValue function.

MAKEDWORD

Function that implements MAKELONG, MAKEWPARAM, and MAKELPARAM macros.

MAKEQWORD

Similar to MAKEDWORD, but for u64.

MAKEWORD

MAKEWORD macro.

MoveFile

MoveFile function.

MulDiv

MulDiv function.

MultiByteToWideChar

MultiByteToWideChar function.

OutputDebugString

OutputDebugString function.

QueryPerformanceCounter

QueryPerformanceCounter function.

QueryPerformanceFrequency

QueryPerformanceFrequency function.

ReplaceFile

ReplaceFile function.

SetCurrentDirectory

SetCurrentDirectory function.

SetFileAttributes

SetFileAttributes function.

SetLastError

SetLastError function.

SetThreadStackGuarantee

SetThreadStackGuarantee function.

Sleep

Sleep function.

SwitchToThread

SwitchToThread function.

SystemTimeToFileTime

SystemTimeToFileTime function.

SystemTimeToTzSpecificLocalTime

SystemTimeToTzSpecificLocalTime function.

VerSetConditionMask

VerSetConditionMask function.

VerifyVersionInfo

VerifyVersionInfo function.

WideCharToMultiByte

WideCharToMultiByte function.

WritePrivateProfileString

WritePrivateProfileString function.

Type Aliases

AnyResult

A Result alias which returns a Box<dyn Error + Send + Sync> on failure.

SysResult

A Result alias for native system error codes, which returns an ERROR on failure.