Module winsafe::co

Native constants.

All types can be converted from/to their underlying integer type. Some types implement the NativeBitflag trait and bitflag operations.

Among these constant types, three are error types: CDERR, ERROR and HRESULT.

Structs

ACCELFuser

ACCELL fVirt (u8).

ACCESS_RIGHTSkernel

Standard access rights flags (u32).

ADgdi

HDC::SetArcDirection dir (i32).

ADRFcomctl

NMTVASYNCDRAW dwRetFlags (u32).

APPCOMMANDuser

wm::AppCommand commands (u16).

BCMcomctl

Button control messages (u32) from comctl.

BCNcomctl

Button control WM_NOTIFY notifications (i32).

BCSIFcomctl

BUTTON_SPLITINFO mask (u32).

BCSScomctl

BUTTON_SPLITINFO uSplitStyle (u32).

BHIDshell

IShellItem::BindToHandler bhid (GUID).

BIgdi

BITMAPINFOHEADER biCompression (u32).

BIAcomctl

BUTTON_IMAGELIST uAlign (u32).

BKMODEgdi

HDC::SetBkMode mode (i32).

BMuser

Button control messages (u32).

BNuser

Button control WM_COMMAND notifications (u16).

BSuser

Button control styles (u32).

BSSgdi

LOGBRUSH lbStyle (u32).

BSTuser

bm::GetCheck return value (u32).

BTNScomctl

Toolbar control styles (u8).

CBuser

Combo box control messages (u32).

CBEMcomctl

ComboBoxEx control messages (u32).

CBES_EXcomctl

Extended combo box styles (u32).

CBNuser

Combo box control WM_COMMAND notifications (u16).

CBSuser

Combo box control styles (u32).

CCcomdlg

CHOOSECOLOR Flags (u32).

CCMcomctl

Generic common controls messages (u32).

CDDScomctl

NMCUSTOMDRAW dwDrawStage (u32).

CDERRcomdlg

Common dialog box error codes.

CDIScomctl

NMCUSTOMDRAW uItemState (u32).

CDRFcomctl

NM_CUSTOMDRAW return value (u32).

CDSuser

ChangeDisplaySettings flags (u32).

CFuser

Standard clipboard formats (u32).

CHARSETgdi

HFONT::CreateFont char_set and LOGFONT lfCharset (u8).

CLIPgdi

HFONT::CreateFont clip_precision and LOGFONT lfClipPrecision (u8).

CLRuser

IMAGELISTDRAWPARAMS rgbFg (u32).

CLSCTXole

CLSCTX enumeration (u32).

CLSIDole

A COM class ID, from which the interfaces are created (GUID).

CMDuser

wm::Command notification codes (u16).

COINITole

COINIT enumeration (u32).

COLORuser

System colors (i32).

CPkernel

WideCharToMultiByte and MultiByteToWideChar code_page identifiers (u16).

CREATEkernel

Process creation flags (u32).

CSuser

Window class styles (u32).

DDLuser

cb::Dir and lb::Dir attributes (u16).

DIBgdi

LOGBRUSH lbColor (u32).

DISPOSITIONkernel

HFILE::CreateFile creation_disposition (u32).

DISP_CHANGEuser

ChangeDisplaySettings return value (u32).

DLGCuser

wm::GetDlgCode return value (u16).

DLGIDuser

Dialog built-in IDs (u16). These are also returned from HWND::MessageBox and HWND::TaskDialog.

DMuser

DEVMODE dmFields (u32).

DMBINuser

DEVMODE dmDefaultSource (i16).

DMCOLORuser

DEVMODE dmColor (i16).

DMDFOuser

DEVMODE dmDisplayFixedOutput (u32).

DMDISPLAYFLAGSuser

DEVMODE dmDisplayFlags (u32).

DMDITHERuser

DEVMODE dmDitherType (u32).

DMDOuser

DEVMODE dmDisplayOrientation (u32).

DMDUPuser

DEVMODE dmDuplex (i16).

DMICMuser

DEVMODE dmICMIntent (u32).

DMICMMETHODuser

DEVMODE dmICMMethod (u32).

DMMEDIAuser

DEVMODE dmMediaType (u32).

DMNUPuser

DEVMODE dmNup (u32).

DMORIENTuser

DEVMODE dmOrientation (i16).

DMPAPERuser

DEVMODE dmPaperSize (i16).

DMRESuser

DEVMODE dmPrintQuality (i16).

DMTTuser

DEVMODE dmTTOption (i16).

DROPEFFECTshell

DROPEFFECT constants (u32).

DSHOW_SERVICEdshow

IMFGetService::GetService service_guid (GUID).

DTMcomctl

Date and time picker control messages (u32).

DTNcomctl

Date and time picker control WM_NOTIFY notifications (i32).

DTScomctl

Date and time picker control styles (u32).

ECuser

em::GetImeStatus margins to set (u16).

EDSuser

EnumDisplaySettingsEx flags (u32).

EIMESuser

em::GetImeStatus and em::SetImeStatus status (u16).

EMuser

Edit control messages (u32).

EMFcomctl

NMLVEMPTYMARKUP dwFlags (u32).

ENuser

Edit control WM_COMMAND notifications (u16).

ENDSESSIONuser

wm::EndSession event (u32).

ENUM_SETTINGSuser

EnumDisplaySettingsEx mode_num (u32).

ERRORkernel

A native system error code (u32) retrieved by GetLastError function.

ESuser

Edit control styles (u32).

FACILITYole

HRESULT facility (u32).

FAPPCOMMANDuser

wm::AppCommand input event (u16).

FDAPshell

FDAP enumeration (u32).

FFgdi

LOGFONT lfPitchAndFamily (u8) used with PITCH.

FILE_ATTRIBUTEkernel

File attribute flags, also other flags from HFILE::CreateFile flags_and_attrs (u32).

FILE_MAPkernel

HFILEMAP::MapViewOfFile desired_access (u32).

FILE_SHAREkernel

HFILE::CreateFile share_mode (u32).

FILE_STARTING_POINTkernel

HFILE::SetFilePointerEx move_method (u32).

FILE_TYPEkernel

HFILE::GetFileType return value (u32).

FILTER_STATEdshow

FILTER_STATE enumeration (u32).

FIRMWARE_TYPEkernel

FIRMWARE_TYPE enumeration (u32).

FOshell

SHFILEOPSTRUCT wFunc (u32).

FOFshell

SHFILEOPSTRUCT fFlags (u16).

FORMAT_MESSAGEkernel

FormatMessage dwFlags (u32).

FOSshell

_FILEOPENDIALOGOPTIONS enumeration (u32).

FWgdi

HFONT::CreateFont weight and LOGFONT lfWeight (u32).

GAuser

HWND::GetAncestor flags (u32).

GCLPuser

HWND::GetClassLongPtr index (i32).

GDCgdi

HDC::GetDeviceCaps index (i32).

GDTcomctl

NMDATETIMECHANGE and NMDATETIMESTRING dwFlags (u32).

GDTRcomctl

dtm::GetRange return value (u32).

GENERICkernel

Generic access rights flags (u32).

GMgdi

HDC::SetGraphicsMode mode (i32).

GMDIuser

HMENU::GetMenuDefaultItem flags (u32).

GMEMkernel

HGLOBAL::GlobalAlloc and HGLOBAL::GlobalReAlloc flags (u32).

GRkernel

HPROCESS::GetGuiResources flags (u32).

GUIuser

GUITHREADINFO flags (u32).

GWuser

HWND::GetWindow cmd (u32).

GWLPuser

HWND::GetWindowLongPtr and HWND::SetWindowLongPtr index (i32).

GWL_Cuser

wm::StyleChanged and wm::StyleChanging change (i8).

HDFcomctl

HDITEM fmt (i32).

HDFTcomctl

HDITEM typeFilter (i32).

HDIcomctl

HDITEM mask (i32).

HDIScomctl

HDITEM state (i32).

HDMcomctl

Header control messages (u32).

HDNcomctl

Header control WM_NOTIFY notifications (i32).

HDScomctl

Header control styles (u32).

HDSILcomctl

hdm::GetImageList and hdm::SetImageList which (u8).

HELPINFOuser

HELPINFO iContextType (i32).

HELPWuser

HWND::WinHelp uCommand (u32).

HHTcomctl

HDHITTESTINFO flags (u32).

HICFcomctl

NMBCHOTITEM dwFlags (u32).

HRESULTole

A native COM error code (u32).

HSgdi

HBRUSH::CreateHatchBrush hatch (i32).

HTuser

wm::NcHitTest, wm::SetCursor hit_test (u16).

HWND_PLACEuser

HWND::SetWindowPos hWndInsertAfter (isize).

ICCcomctl

INITCOMMONCONTROLSEX icc (u32).

ICON_SZuser

wm::SetIcon icon size (u8).

IDBcomctl

TBADDBITMAP nID (usize).

IDCuser

HINSTANCE::LoadCursor lpCursorName (isize).

IDIuser

HINSTANCE::LoadIcon lpIconName (isize).

IIDole

A COM interface ID, which uniquely identifies the interface (GUID).

ILCcomctl

HIMAGELIST::Create flags (u32).

ILDcomctl

IMAGELISTDRAWFLAGS enumeration (u32).

ILScomctl

IMAGELISTSTATEFLAGS enumeration (u32).

IMAGE_TYPEuser

bm::GetImage img_type (u8).

IPMcomctl

IP address control messages (u32).

IPNcomctl

IP address control WM_NOTIFY notifications (i32).

ISMEXuser

InSendMessageEx return value (u32).

KEYadvapi

HKEY::OpenKeyEx access_rights (u32).

KFshell

KNOWN_FOLDER_FLAG enumeration (u32).

KNOWNFOLDERIDshell

Known system folders (GUID).

LANGkernel

Language identifier (u16).

LBuser

List box control messages (u32).

LBNuser

List box control WM_COMMAND notifications (u16).

LBSuser

List box control styles (u32).

LIFcomctl

LITEM mask (u32).

LIScomctl

LITEM state (u32).

LMcomctl

SysLink control messages (u32).

LOCKTYPEshlwapi

LOCKTYPE enumeration (u32).

LRuser

HINSTANCE::LoadImageBitmap, HINSTANCE::LoadImageCursor and HINSTANCE::LoadImageIcon load.

LSFWuser

LockSetForegroundWindow lock_code (u32).

LVAcomctl

lvm::Arrange arrangement (u16).

LVBKIFcomctl

LVBKIMAGE uFlags (u32).

LVCDIcomctl

NMLVCUSTOMDRAW dwItemType (u32).

LVCFcomctl

LVCOLUMN mask (u32).

LVCFMT_Ccomctl

LVCOLUMN mask (i32).

LVCFMT_Icomctl

LVITEM piColFmt (i32).

LVFFcomctl

LVFOOTERINFO mask (u32).

LVFIcomctl

LVFINDINFO flags (u32).

LVFIFcomctl

LVFOOTERITEM mask (u32).

LVFIScomctl

LVFOOTERITEM state (u32).

LVGA_FHcomctl

LVGROUP uAlign (u32).

LVGA_HEADERcomctl

NMLVCUSTOMDRAW uAlign (u32).

LVGFcomctl

LVGROUP mask (u32).

LVGGRcomctl

lvm::GetGroupRect flags (i32).

LVGITcomctl

NMLVGETINFOTIP dwFlags (u32).

LVGMFcomctl

LVGROUPMETRICS mask (u32).

LVGScomctl

LVGROUP state (u32).

LVHTcomctl

LVHITTESTINFO flags (u32).

LVIFcomctl

LVITEM mask (u32).

LVIMcomctl

LVINSERTMARK dwFlags (u32).

LVIRcomctl

lvm::GetItemRect portion (u8).

LVIScomctl

ListView item states (u32).

LVI_GROUPIDcomctl

LVITEM iGroupId (i32).

LVKFcomctl

NMITEMACTIVATE uKeyFlags (u32).

LVMcomctl

List view control messages (u32).

LVNcomctl

List view control WM_NOTIFY notifications (i32).

LVNIcomctl

lvm::GetNextItem relationship (u32).

LVScomctl

List view control styles (u32).

LVSICFcomctl

lvm::SetItemCount behavior (u32).

LVSILcomctl

lvm::GetImageList kind (u8).

LVS_EXcomctl

Extended list view control styles (u32).

LVTVIFcomctl

LVTILEVIEWINFO dwFlags (u32).

LVTVIMcomctl

LVTILEVIEWINFO dwMask (u32).

LV_VIEWcomctl

ListView views (u32).

LWScomctl

SysLink control styles (u32).

MBuser

HWND::MessageBox flags (u32).

MBCkernel

MultiByteToWideChar flags (u32).

MCMcomctl

Month calendar control messages (u32).

MCMVcomctl

NMVIEWCHANGE dwOldView and dwNewView (u32).

MCNcomctl

Month calendar control WM_NOTIFY notifications (i32).

MCScomctl

Month calendar control styles (u32).

MCSCcomctl

dtm::GetMcColor color (u8).

MEDIATYPEdshow

AM_MEDIA_TYPE majortype (GUID).

MEDIA_FORMATdshow

AM_MEDIA_TYPE formattype, originally with FORMAT prefix (GUID).

MFuser

HMENU::AppendMenu flags, HMENU::GetMenuState return value, HWND::HiliteMenuItem hilite (u32).

MFSuser

MENUITEMINFO fState (u32).

MFTuser

MENUITEMINFO fType (u32).

MFVideoARModedshow

MFVideoAspectRatioMode enumeration (u32).

MIIMuser

MENUITEMINFO fMask (u32).

MIMuser

MENUINFO fMask (u32).

MKuser

wm::LButtonDown (and similar) virtual keys (u16).

MNDuser

wm::MenuDrag return value (u8).

MNSuser

MENUINFO dwStyle (u32).

MONITORuser

HMONITOR::MonitorFromPoint, HMONITOR::MonitorFromRect, HWND::MonitorFromWindow flags (u32).

MONITORINFOFuser

MONITORINFO dwFlags (u32).

MSGFuser

wm::EnterIdle reason (u8).

NIFshell

NOTIFYICONDATA uFlags (u32).

NIIFshell

NOTIFYICONDATA dwInfoFlags (u32).

NIMshell

Shell_NotifyIcon message (u32).

NISshell

NOTIFYICONDATA dwState and dwStateFlags (u32).

NMcomctl

wm::Notify notification codes (i32).

OBJIDuser

HWND::GetMenuBarInfo idObject (i32).

OCRuser

HCURSOR::SetSystemCursor id (u32).

ODAuser

DRAWITEMSTRUCT itemAction (u32).

ODSuser

DRAWITEMSTRUCT itemState (u32).

ODTuser

DRAWITEMSTRUCT CtlType (u32).

ODT_Cuser

COMPAREITEMSTRUCT and DELETEITEMSTRUCT CtlType (u32).

OUT_PRECISgdi

HFONT::CreateFont out_precision and LOGFONT lfOutPrecision (u8).

PAGEkernel

HFILE::CreateFileMapping protect (u32).

PBMcomctl

Progress bar control messages (u32).

PBScomctl

Progress bar control styles (u32).

PBSTcomctl

Progress bar states (u32).

PICTYPEshlwapi

PICTYPE constants (i16).

PITCHgdi

HFONT::CreateFont pitch_and_family and LOGFONT lfPitchAndFamily (u8) used with FF.

PMuser

PeekMessage remove_msg (u32).

PROCESSkernel

Process security and access rights (u32).

PROCESSORkernel

SYSTEM_INFO dwProcessorType (u32).

PROCESSOR_ARCHITECTUREkernel

SYSTEM_INFO wProcessorArchitecture (u16).

PROCESS_NAMEkernel

QueryFullProcessImageName flags (u32).

PSgdi

HPEN::CreatePen style (i32).

QSuser

GetQueueStatus flags (u32).

QUALITYgdi

HFONT::CreateFont quality and LOGFONT lfQuality (u8).

RBcomctl

Rebar control messages (u32).

RBNcomctl

Rebar control WM_NOTIFY notifications (i32).

RBScomctl

Rebar control styles (u32).

RDWuser

HWND::RedrawWindow flags (u32).

REGadvapi

Registry value types (u32).

REGIONuser

HWND::GetUpdateRgn, HWND::GetWindowRgn and HDC::SelectObjectRgn return value (i32).

REG_OPTIONadvapi

HKEY::OpenKeyEx options (u32).

REPLACEFILEkernel

ReplaceFile flags (u32).

ROPgdi

Raster operation code (u32). HDC::BitBlt rop, HDC::PatBlt rop and IMAGELISTDRAWPARAMS dwRop.

RRFadvapi

HKEY::GetValue dwFlags (u32).

RTkernel

Predefined resource types (u16).

SBcomctl

Status bar control messages (u32).

SBARScomctl

Status bar control styles (u32).

SBBuser

HWND::GetScrollInfo, HWND::SetScrollInfo and HWND::SetScrollRange bar (i32). Originally has SB prefix.

SBNcomctl

Status bar control WM_NOTIFY notifications (i32).

SBTcomctl

sb::GetText, sb::GetTextLength and sb::SetText drawing operation (u16).

SB_EMuser

em::Scroll action.

SB_REQuser

wm::HScroll and wm::VScroll request (u16).

SCuser

wm::SysCommand type of system command requested (u32).

SCSkernel

GetBinaryType return value (u32).

SECTIONkernel

Composes FILE_MAP (u32).

SECURITY_IMPERSONATIONkernel

SECURITY_IMPERSONATION_LEVEL enumeration (u32).

SEEKING_FLAGSdshow

IMediaSeeking::SetPositions flags (u32).

SE_ERRshell

HWND::ShellExecute return value (u32).

SFGAOshell

SFGAO constants (u32).

SHARDshell

SHARD enumeration (u32).

SHGFIshell

SHGetFileInfo flags (u32).

SIFuser

SCROLLINFO fMask (u32).

SIGDNshell

SIGDN enumeration (u32).

SIZE_Ruser

wm::Size request (u8).

SMuser

GetSystemMetrics index (i32).

SORTkernel

Sort order identifiers (u16).

SPIuser

SystemParametersInfo action (u32).

SPIFuser

SystemParametersInfo win_ini (u32).

SSuser

Label control styles (u32).

STANDARD_RIGHTSkernel

Standard access rights flags (u32).

STAPuxtheme

HTHEME::GetThemeAppProperties return value (u32).

STARTFkernel

HPROCESS::CreateProcess dwFlags (u32).

STATE_SYSTEMuser

DATETIMEPICKERINFO stateCheck and stateButton, TITLEBARINFOEX rgstate, COMBOBOXINFO stateButton (u32).

STGCshlwapi

STGC enumeration (u32).

STMcomctl

Static control messages (u32).

STNuser

Static control WM_COMMAND notifications (u16).

STOCK_BRUSHgdi

HBRUSH::GetStockObject i (i32).

STOCK_FONTgdi

HFONT::GetStockObject i (i32).

STOCK_PENgdi

HPEN::GetStockObject i (i32).

STPFLAGshell

STPFLAG enumeration (u32).

STREAM_SEEKshlwapi

STREAM_SEEK enumeration (u32).

STRETCH_MODEgdi

HDC::SetStretchBltMode mode (i32).

SUBLANGkernel

Sublanguage identifier (u16).

SWkernel

HWND::ShowWindow show_cmd (i32).

SWPuser

HWND::SetWindowPos flags (u32).

SW_Suser

wm::ShowWindow status (u8).

TAgdi

HDC::SetTextAlign align (u32).

TBcomctl

NMTRBTHUMBPOSCHANGING nReason (i32).

TBMcomctl

Toolbar control messages (u32).

TBNcomctl

Toolbar control WM_NOTIFY notifications (i32).

TBPFshell

ITaskbarList3::SetProgressState tbpFlags (u32).

TBScomctl

Trackbar control styles (u32).

TBSTATEcomctl

Toolbar button states (u8).

TCIScomctl

Tab control item states (u32).

TCMcomctl

Tab control messages (u32).

TCNcomctl

Tab control WM_NOTIFY notifications (i32).

TCScomctl

Tab control styles (u32).

TCS_EXcomctl

Extended tab control styles (u32).

TDCBFcomctl

HWND::TaskDialog and TASKDIALOGCONFIG dwCommonButtons (i32).

TDFcomctl

TASKDIALOGCONFIG dwFlags (i32).

TD_ICONcomctl

HWND::TaskDialog pszIcon and TASKDIALOGCONFIG pszMainIcon (u16). Originally has TD prefix and ICON suffix.

TH32CSkernel

HPROCESSLIST::CreateToolhelp32Snapshot flags (u32).

THREAD_CREATEkernel

HTHREAD::CreateThread flags (u32). Originally has no prefix.

TIME_FORMATdshow

IMediaSeeking::SetTimeFormat format (GUID).

TMEuser

TrackMouseEvent dwFlags (u32).

TMTuxtheme

Theme property identifiers (i32).

TOKENkernel

Token access rights.

TPMuser

TrackPopupMenu flags (u32).

TRBMcomctl

Trackbar control messages (u32).

TRBNcomctl

Trackbar control WM_NOTIFY notifications (i32).

TTIcomctl

EDITBALLOONTIP ttiIcon (i32).

TVEcomctl

tvm::Expand action (u32).

TVGNcomctl

tvm::GetNextItem which (u32).

TVIcomctl

TVINSERTSTRUCT hInsertAfter (isize).

TVIFcomctl

TVITEM mask (u32).

TVIScomctl

Tree view item states (u32)

TVIS_EXcomctl

TVITEMEX uStateEx (u32).

TVMcomctl

Tree view control messages (u32).

TVNcomctl

Tree view control WM_NOTIFY notifications (i32).

TVScomctl

Tree view control styles (u32).

TVSILcomctl

tvm::GetImageList and tvm::SetImageList kind.

TVS_EXcomctl

Extended tree view control styles (u32).

UDMcomctl

Up-down control messages (u32).

UDNcomctl

Up-down control WM_NOTIFY notifications (i32).

UDScomctl

Up-down control styles (u32).

UOIuser

HPROCESS::SetUserObjectInformation index (i32).

VER_CONDkernel

VerSetConditionMask condition (u8).

VER_MASKkernel

VerifyVersionInfo and VerSetConditionMask type_mask (u32).

VER_NTkernel

OSVERSIONINFOEX wProductType (u8).

VER_PLATFORMkernel

OSVERSIONINFOEX dwPlatformId (u32).

VER_SUITEkernel

OSVERSIONINFOEX wSuiteMask (u16).

VFTversion

VS_FIXEDFILEINFO dwFileType (u32).

VFT2version

VS_FIXEDFILEINFO dwFileSubType (u32).

VKuser

Virtual key codes (u16).

VK_DIRcomctl

LVFINDINFO vkDirection (u16).

VOSversion

VS_FIXEDFILEINFO dwFileOS (u32).

VSuxtheme

Visual style parts and states. These are originally two separated constant types, but since they form a hierarchy, one single constant type can represent both part and state.

VS_FFversion

VS_FIXEDFILEINFO dwFileFlags (u32).

WAuser

wm::Activate activation state (u16).

WAITkernel

HPROCESS::WaitForSingleObject return value (u32).

WCkernel

WideCharToMultiByte flags (u32).

WDAuser

HWND::GetWindowDisplayAffinity and HWND::SetWindowDisplayAffinity dwAffinity (u32).

WHuser

HHOOK::CallNextHookEx code and HHOOK::SetWindowsHookEx hook_id (i32).

WIN32kernel

_WIN32 version definitions (u16).

WMuser

Window message codes (u32).

WMPNuser

wm::ParentNotify event (u16).

WMSZuser

wm::Sizing window edge (u8).

WPFuser

WINDOWPLACEMENT flags (u32).

WSuser

Window styles (u32).

WS_EXuser

Extended window styles (u32).

WVRuser

wm::NcCalcSize return flags (u32).