winsafe

Module co

Native constants.

Among these constant types, three are error types: CDERR, ERROR and HRESULT.

Structs

ACCELFuser

ACCELL fVirt (u8).

ACCESS_RIGHTSkernel

Standard access rights flags (u32).

ACTCTX_FLAGkernel

ACTCTX dwFlags (u32).

AC_SRCuser

BLENDFUNCTION BlendOp and AlphaFormat (u8).

ADgdi

HDC::SetArcDirection dir (i32).

ADRFcomctl

NMTVASYNCDRAW dwRetFlags (u32).

ADVFole

ADVF enumeration (u32).

APPCOMMANDuser

wm::AppCommand commands (u16).

BCMcomctl

Button control messages (u32) from comctl.

BCNcomctl

Button control WM_NOTIFY notifications (i32).

BCSIFcomctl

BUTTON_SPLITINFO mask (u32).

BCSScomctl

BUTTON_SPLITINFO uSplitStyle (u32).

BHIDshell

IShellItem::BindToHandler bhid (GUID).

BIgdi

BITMAPINFOHEADER biCompression (u32).

BIAcomctl

BUTTON_IMAGELIST uAlign (u32).

BINDole

BIND_FLAGS enumeration (u32).

BKMODEgdi

HDC::SetBkMode mode (i32).

BMuser

Button control messages (u32).

BNuser

Button control WM_COMMAND notifications (u16).

BSuser

Button control styles (u32).

BSFuser

BroadcastSystemMessage flags (u32).

BSMuser

BroadcastSystemMessage info and return value (u32).

BSSgdi

LOGBRUSH lbStyle (u32).

BSTuser

bm::GetCheck return value (u32).

BTNScomctl

Toolbar control styles (u8).

CALcomctl

MCM_GETCALID return type calendar IDs (u32).

CBuser

Combo box control messages (u32).

CBEMcomctl

ComboBoxEx control messages (u32).

CBES_EXcomctl

Extended combo box styles (u32).

CBNuser

Combo box control WM_COMMAND notifications (u16).

CBSuser

Combo box control styles (u32).

CCuser

CHOOSECOLOR Flags (u32).

CCMcomctl

Generic common controls messages (u32).

CDDScomctl

NMCUSTOMDRAW dwDrawStage (u32).

CDERRuser

Common dialog box error codes.

CDIScomctl

NMCUSTOMDRAW uItemState (u32).

CDRFcomctl

NM_CUSTOMDRAW return value (u32).

CDSuser

ChangeDisplaySettings flags (u32).

CFuser

Standard clipboard formats (u16).

CHARSETgdi

HFONT::CreateFont char_set and LOGFONT lfCharset (u8).

CLAIM_SECURITY_ATTRIBUTEkernel

CLAIM_SECURITY_ATTRIBUTE_V1 Flags (u16);

CLAIM_SECURITY_ATTRIBUTE_TYPEkernel

CLAIM_SECURITY_ATTRIBUTE_V1 ValueType (u16);

CLIPgdi

HFONT::CreateFont clip_precision and LOGFONT lfClipPrecision (u8).

CLRuser

IMAGELISTDRAWPARAMS rgbFg (u32).

CLSCTXole

CLSCTX enumeration (u32).

CLSIDole

A COM class ID, from which the interfaces are created (GUID).

CMDuser

wm::Command notification codes (u16).

COINITole

COINIT enumeration (u32).

COLORuser

System colors (i32).

CONSOLEkernel

HSTD::SetConsoleMode mode (u32).

CPkernel

WideCharToMultiByte and MultiByteToWideChar code_page identifiers (u16).

CREATEkernel

Process creation flags (u32).

CREATE_EVENTkernel

HEVENT::CreateEventEx flags (u32).

CSuser

Window class styles (u32).

CURSORuser

CURSORINFO flags (u32).

CUR_OBJgdi

HDC::GetCurrentObject kind (u32).

DBTkernel

wm::DeviceChange event (u16).

DBTFkernel

DEV_BROADCAST_VOLUME dbcv_flags (u16).

DBT_DEVTYPkernel

DEV_BROADCAST_HDR dbch_devicetype (u32).

DCuser

HWND::DrawCaption flags (u32).

DCXuser

GetDCEx flags (u32).

DDCuser

DIALOG_DPI_CHANGE_BEHAVIORS enumeration (u32).

DDLuser

cb::Dir and lb::Dir attributes (u16).

DEACTIVATE_ACTCTXkernel

HACTCTX::DeactivateActCtx flags (u32).

DESKTOP_RIGHTSuser

Desktop security and access rights flags (u32).

DFuser

HDESK::OpenDesktop flags (u32).

DIBgdi

LOGBRUSH lbColor (u32).

DISPATCHoleaut

IDispatch::Invoke flags (u16).

DISPIDoleaut

DISPID constants (i32) for DISPPARAMS.

DISPLAY_DEVICEuser

DISPLAY_DEVICE StateFlags (u32).

DISPOSITIONkernel

HFILE::CreateFile creation_disposition (u32).

DISP_CHANGEuser

ChangeDisplaySettings return value (u32).

DLGCuser

wm::GetDlgCode return value (u16).

DLGIDuser

Dialog built-in IDs (u16). These are also returned from HWND::MessageBox and HWND::TaskDialog.

DMuser

DEVMODE dmFields (u32).

DMBINuser

DEVMODE dmDefaultSource (i16).

DMCOLLATEuser

DEVMODE dmCollate (i16).

DMCOLORuser

DEVMODE dmColor (i16).

DMDFOuser

DEVMODE dmDisplayFixedOutput (u32).

DMDISPLAYFLAGSuser

DEVMODE dmDisplayFlags (u32).

DMDITHERuser

DEVMODE dmDitherType (u32).

DMDOuser

DEVMODE dmDisplayOrientation (u32).

DMDUPuser

DEVMODE dmDuplex (i16).

DMICMuser

DEVMODE dmICMIntent (u32).

DMICMMETHODuser

DEVMODE dmICMMethod (u32).

DMMEDIAuser

DEVMODE dmMediaType (u32).

DMNUPuser

DEVMODE dmNup (u32).

DMORIENTuser

DEVMODE dmOrientation (i16).

DMPAPERuser

DEVMODE dmPaperSize (i16).

DMRESuser

DEVMODE dmPrintQuality (i16).

DMTTuser

DEVMODE dmTTOption (i16).

DPI_HOSTING_BEHAVIORuser

DPI_HOSTING_BEHAVIOR enumeration (u32).

DRIVEkernel

GetDriveType return type (u32).

DROPEFFECTole

DROPEFFECT constants (u32).

DTuser

HDC::DrawText format (u32).

DTMcomctl

Date and time picker control messages (u32).

DTNcomctl

Date and time picker control WM_NOTIFY notifications (i32).

DTScomctl

Date and time picker control styles (u32).

DVASPECTole

DVASPECT enumeration (u32).

DWMFLIP3DWINDOWPOLICYdwm

DWMFLIP3DWINDOWPOLICY enumeration (u32).

DWMNCRENDERINGPOLICYdwm

DWMNCRENDERINGPOLICY enumeration (u32).

DWMSBTdwm

DWMWA system backdrop type (u32).

DWMSCdwm

DwmShowContact show_contact (u32).

DWMWAdwm

DWMWINDOWATTRIBUTE enumeration (u32).

DWMWCPdwm

DWMWA window corner preference (u32).

DWM_CLOAKEDdwm

DWMWA cloaked flags (u32).

DWM_SITdwm

DwmSetIconicLivePreviewBitmap sit_flags (u32).

DXGI_ADAPTER_FLAGdxgi

DXGI_ADAPTER_FLAG enumeration (u32).

DXGI_COMPUTE_PREEMPTIONdxgi

DXGI_COMPUTE_PREEMPTION_GRANULARITY enumeration.

DXGI_CREATE_FACTORYdxgi

CreateDXGIFactory2 Flags (u32).

DXGI_ENUM_MODESdxgi

DXGI_ENUM_MODES enumeration (u32).

DXGI_FORMATdxgi

DXGI_FORMAT enumeration (u32).

DXGI_GRAPHICS_PREEMPTIONdxgi

DXGI_GRAPHICS_PREEMPTION_GRANULARITY enumeration (u32).

DXGI_MAPdxgi

IDXGISurface::Map map_flags (u32).

DXGI_MODE_ROTATIONdxgi

DXGI_MODE_ROTATION enumeration (u32).

DXGI_MODE_SCALINGdxgi

DXGI_MODE_SCALING enumeration (u32).

DXGI_MODE_SCANLINE_ORDERdxgi

DXGI_MODE_SCANLINE_ORDER enumeration (u32).

DXGI_MWAdxgi

IDXGIFactory::GetWindowAssociation flags (u32).

DXGI_PRESENTdxgi

DXGI_PRESENT enumeration (u32).

DXGI_RESIDENCYdxgi

DXGI_RESIDENCY enumeration (u32).

DXGI_RESOURCE_PRIORITYdxgi

IDXGIResource::GetEvictionPriority and IDXGIResource::SetEvictionPriority eviction_priority (u32).

DXGI_SWAP_CHAIN_FLAGdxgi

DXGI_SWAP_CHAIN_FLAG enumeration (u32).

DXGI_SWAP_EFFECTdxgi

DXGI_SWAP_EFFECT enumeration (u32).

DXGI_USAGEdxgi

DXGI_USAGE flags (u32).

ECuser

em::GetImeStatus margins to set (u16).

EDDuser

EnumDisplayDevices flags (u32).

EDSuser

EnumDisplaySettingsEx flags (u32).

EIMESuser

em::GetImeStatus and em::SetImeStatus status (u16).

EMuser

Edit control messages (u32).

EMFcomctl

NMLVEMPTYMARKUP dwFlags (u32).

ENuser

Edit control WM_COMMAND notifications (u16).

ENDSESSIONuser

wm::EndSession event (u32).

ENUM_SETTINGSuser

EnumDisplaySettingsEx mode_num (u32).

ERRORkernel

A native system error code (u32), commonly retrieved by GetLastError function.

ESuser

Edit control styles (u32).

ESBuser

HWND::EnableScrollBar arrows (u32).

EVENTLOGadvapi

HEVENTLOG::ReportEvent event_type u16. undefined

EVENT_RIGHTSkernel

Event access rights (u32).

EWXuser

ExitWindowsEx flags (u32).

FACILITYole

HRESULT facility (u32).

FAPPCOMMANDuser

wm::AppCommand input event (u16).

FDAPshell

FDAP enumeration (u32).

FDEORshell

FDE_OVERWRITE_RESPONSE enumeration (u32).

FDESVRshell

FDE_SHAREVIOLATION_RESPONSE enumeration (u32).

FFgdi

LOGFONT lfPitchAndFamily (u8) used with PITCH.

FILE_ATTRIBUTEkernel

File attributes (u32).

FILE_CACHEkernel

GetSystemFileCacheSize returned flags (u32).

FILE_FLAGkernel

HFILE::CreateFile flags (u32).

FILE_MAPkernel

HFILEMAP::MapViewOfFile desired_access (u32).

FILE_SECURITYkernel

HFILE::CreateFile security (u32).

FILE_SHAREkernel

HFILE::CreateFile share_mode (u32).

FILE_STARTING_POINTkernel

HFILE::SetFilePointerEx move_method (u32).

FILE_TYPEkernel

HFILE::GetFileType return value (u32).

FILE_VOLkernel

GetVolumeInformation flags (u32).

FILTER_STATEdshow

FILTER_STATE enumeration (u32).

FIRMWARE_TYPEkernel

FIRMWARE_TYPE enumeration (u32).

FLASHWuser

FLASHWINFO dwFlags (u32).

FOshell

SHFILEOPSTRUCT wFunc (u32).

FOFshell

SHFILEOPSTRUCT fFlags (u16).

FORMwinspool

FORM_INFO_1 and FORM_INFO_2 Flags (u32).

FORMAT_MESSAGEkernel

FormatMessage dwFlags (u32).

FOSshell

_FILEOPENDIALOGOPTIONS enumeration (u32).

FWgdi

HFONT::CreateFont weight and LOGFONT lfWeight (u32).

GAuser

HWND::GetAncestor flags (u32).

GCLPuser

HWND::GetClassLongPtr index (i32).

GDCgdi

HDC::GetDeviceCaps index (i32).

GDTcomctl

NMDATETIMECHANGE and NMDATETIMESTRING dwFlags (u32).

GDTRcomctl

dtm::GetRange return value (u32).

GENERICkernel

Generic access rights flags (u32).

GET_MODULE_HANDLE_EX_FLAGkernel

HINSTANCE::GetModuleHandleEx flags (u32).

GMgdi

HDC::SetGraphicsMode mode (i32).

GMDIuser

HMENU::GetMenuDefaultItem flags (u32).

GMEMkernel

HGLOBAL::GlobalAlloc and HGLOBAL::GlobalReAlloc flags (u32).

GMRcomctl

GPSshell

GETPROPERTYSTOREFLAGS enumeration (u32).

GRkernel

HPROCESS::GetGuiResources flags (u32).

GUIuser

GUITHREADINFO flags (u32).

GWuser

HWND::GetWindow cmd (u32).

GWLPuser

HWND::GetWindowLongPtr and HWND::SetWindowLongPtr index (i32).

GWL_Cuser

wm::StyleChanged and wm::StyleChanging change (i8).

HDFcomctl

HDITEM fmt (i32).

HDFTcomctl

HDITEM typeFilter (i32).

HDIcomctl

HDITEM and NMHDDISPINFO mask (i32).

HDIScomctl

HDITEM state (i32).

HDMcomctl

Header control messages (u32).

HDNcomctl

Header control WM_NOTIFY notifications (i32).

HDScomctl

Header control styles (u32).

HDSILcomctl

hdm::GetImageList and hdm::SetImageList which (u8).

HEAP_ALLOCkernel

HHEAP::HeapAlloc flags (u32).

HEAP_CREATEkernel

HHEAP::HeapCreate options (u32).

HEAP_INFORMATIONkernel

HHEAP::HeapSetInformation information class (u32).

HEAP_REALLOCkernel

HHEAP::HeapReAlloc options (u32).

HEAP_SIZEkernel

HHEAP::HeapSize and HHEAP::HeapCompact options (u32).

HELPINFOuser

HELPINFO iContextType (i32).

HELPWuser

HWND::WinHelp uCommand (u32).

HF32kernel

HEAPLIST32 dwFlags (u32).

HHTcomctl

HDHITTESTINFO flags (u32).

HICFcomctl

NMBCHOTITEM dwFlags (u32).

HOTKEYFuser

wm::GetHotkey and wm::SetHotKey modifiers.

HRESULTole

A native COM error code (u32), which can be considered a superset of ERROR.

HSgdi

HBRUSH::CreateHatchBrush hatch (i32).

HTuser

wm::NcHitTest, wm::SetCursor hit_test (u16).

HTTP_ADDREQwininet

HINTERNETREQUEST::HttpAddRequestHeaders modifiers (u32).

HTTP_QUERYwininet

HINTERNETREQUEST::HttpQueryInfo info (u32).

HTTP_QUERY_FLAGwininet

HINTERNETREQUEST::HttpQueryInfo flags (u32).

HWND_PLACEuser

HWND::SetWindowPos hWndInsertAfter (isize).

ICCcomctl

INITCOMMONCONTROLSEX icc (u32).

ICON_SZuser

wm::SetIcon icon size (u8).

ICUwininet

InternetCanonicalizeUrl, InternetCombineUrl and InternetCrackUrl flags (u32).

IDBcomctl

TBADDBITMAP nID (usize).

IDCuser

HINSTANCE::LoadCursor lpCursorName (u32).

IDIuser

HINSTANCE::LoadIcon lpIconName (u32).

ID_PBcomctl

psm::GetResult return type (i32).

IIDole

A COM interface ID, which uniquely identifies the interface (GUID).

ILCcomctl

HIMAGELIST::Create flags (u32).

ILDcomctl

IMAGELISTDRAWFLAGS enumeration (u32).

ILScomctl

IMAGELISTSTATEFLAGS enumeration (u32).

IMAGE_TYPEuser

bm::GetImage img_type; stm::GetImage img_type (u8).

INPUTuser

INPUT type (u32).

INTERNET_DEFAULT_PORTwininet

HINTERNET::InternetOpen port (u16).

INTERNET_FLAGwininet

Internet API flags (u32).

INTERNET_OPEN_TYPEwininet

HINTERNET::InternetOpen access_type (u32).

INTERNET_SCHEMEwininet

INTERNET_SCHEME enumeration (i32).

INTERNET_SERVICEwininet

HINTERNET::InternetOpen service (u32).

INVOKEKINDoleaut

INVOKEKIND enumeration (u32).

IPMcomctl

IP address control messages (u32).

IPNcomctl

IP address control WM_NOTIFY notifications (i32).

ISMEXuser

InSendMessageEx return value (u32).

KEYadvapi

Registry access rights (u32).

KEYEVENTFuser

KEYBDINPUT dwFlags (u32).

KFshell

KNOWN_FOLDER_FLAG enumeration (u32).

KNOWNFOLDERIDshell

Known system folders (GUID).

LANGkernel

Language identifier (u16).

LAYOUTuser

GetProcessDefaultLayout and SetProcessDefaultLayout value (u32).

LBuser

List box control messages (u32).

LBNuser

List box control WM_COMMAND notifications (u16).

LBSuser

List box control styles (u32).

LIFcomctl

LITEM mask (u32).

LIScomctl

LITEM state (u32).

LLKHFuser

KBDLLHOOKSTRUCT flags (u32).

LMcomctl

SysLink control messages (u32).

LMEMkernel

HLOCAL::LocalAlloc and HLOCAL::LocalReAlloc flags (u32).

LOCKTYPEole

LOCKTYPE enumeration (u32).

LRgdi

HINSTANCE::LoadImageBitmap, HINSTANCE::LoadImageCursor and HINSTANCE::LoadImageIcon load (u32).

LSFWuser

LockSetForegroundWindow lock_code (u32).

LVAcomctl

lvm::Arrange arrangement (u16).

LVBKIFcomctl

LVBKIMAGE uFlags (u32).

LVCDIcomctl

NMLVCUSTOMDRAW dwItemType (u32).

LVCFcomctl

LVCOLUMN mask (u32).

LVCFMT_Ccomctl

LVCOLUMN mask (i32).

LVCFMT_Icomctl

LVITEM piColFmt (i32).

LVFFcomctl

LVFOOTERINFO mask (u32).

LVFIcomctl

LVFINDINFO flags (u32).

LVFIFcomctl

LVFOOTERITEM mask (u32).

LVFIScomctl

LVFOOTERITEM state (u32).

LVGA_FHcomctl

LVGROUP uAlign (u32).

LVGA_HEADERcomctl

NMLVCUSTOMDRAW uAlign (u32).

LVGFcomctl

LVGROUP mask (u32).

LVGGRcomctl

lvm::GetGroupRect flags (i32).

LVGITcomctl

NMLVGETINFOTIP dwFlags (u32).

LVGMFcomctl

LVGROUPMETRICS mask (u32).

LVGScomctl

LVGROUP state (u32).

LVHTcomctl

LVHITTESTINFO flags (u32).

LVIFcomctl

LVITEM mask (u32).

LVIMcomctl

LVINSERTMARK dwFlags (u32).

LVIRcomctl

lvm::GetItemRect portion (u8).

LVIScomctl

ListView item states (u32).

LVI_GROUPIDcomctl

LVITEM iGroupId (i32).

LVKFcomctl

NMITEMACTIVATE uKeyFlags (u32).

LVMcomctl

List view control messages (u32).

LVNcomctl

List view control WM_NOTIFY notifications (i32).

LVNIcomctl

lvm::GetNextItem relationship (u32).

LVScomctl

List view control styles (u32).

LVSICFcomctl

lvm::SetItemCount behavior (u32).

LVSILcomctl

lvm::GetImageList kind (u8).

LVS_EXcomctl

Extended list view control styles (u32).

LVTVIFcomctl

LVTILEVIEWINFO dwFlags (u32).

LVTVIMcomctl

LVTILEVIEWINFO dwMask (u32).

LV_VIEWcomctl

ListView views (u32).

LWAuser

HWND::SetLayeredWindowAttributes flags (u32).

LWScomctl

SysLink control styles (u32).

MAPVKuser

MapVirtualKey uMapType (u32).

MBuser

HWND::MessageBox flags (u32).

MBCkernel

MultiByteToWideChar flags (u32).

MBPuser

MessageBeep uType (u32).

MCGIFcomctl

This is a bitflag constant.

MCGIPcomctl

MCGRIDINFO dwPart (u32).

MCHTcomctl

MCHITTESTINFO uHit (u32).

MCMcomctl

Month calendar control messages (u32).

MCMVcomctl

NMVIEWCHANGE dwOldView and dwNewView (u32).

MCNcomctl

Month calendar control WM_NOTIFY notifications (i32).

MCScomctl

Month calendar control styles (u32).

MCSCcomctl

dtm::GetMcColor color (u8).

MDITILEuser

HWND::TileWindows how (u32).

MEmf

IMFMediaEvent::GetType return value (u32).

MEDIATYPEdshow

AM_MEDIA_TYPE majortype (GUID).

MEDIA_FORMATdshow

AM_MEDIA_TYPE formattype, originally with FORMAT prefix (GUID).

MEM_STATEkernel

MEMORY_BASIC_INFORMATION State (u32).

MEM_TYPEkernel

MEMORY_BASIC_INFORMATION Type (u32).

MFuser

HMENU::AppendMenu flags, HMENU::GetMenuState return value, HWND::HiliteMenuItem hilite (u32).

MFASYNCmf

IMFAsyncCallback::GetParameters flags (u32).

MFBYTESTREAMmf

IMFByteStream::GetCapabilities flags (u32).

MFBYTESTREAM_SEEK_FLAGmf

IMFByteStream::Seek flags (u32).

MFBYTESTREAM_SEEK_ORIGINmf

MFBYTESTREAM_SEEK_ORIGIN enumeration (u32).

MFCLOCK_CHARACTERISTICS_FLAGmf

MFCLOCK_CHARACTERISTICS_FLAGS enumeration (u32).

MFCLOCK_RELATIONAL_FLAGmf

MFCLOCK_RELATIONAL_FLAGS enumeration (u32).

MFCLOCK_STATEmf

MFCLOCK_STATE enumeration (u32).

MFMEDIASOURCEmf

MFMEDIASOURCE_CHARACTERISTICS enumeration (u32).

MFSuser

MENUITEMINFO fState (u32).

MFSESSIONCAPmf

IMFMediaSession::GetSessionCapabilities caps (u32).

MFSESSION_GETFULLTOPOLOGYmf

MFSESSION_GETFULLTOPOLOGY_FLAGS enumeration (u32).

MFSESSION_SETTOPOLOGYmf

MFSESSION_SETTOPOLOGY_FLAGS enumeration (u32).

MFSTARTUPmf

MFStartup flags (u32).

MFTuser

MENUITEMINFO fType (u32).

MFVideoARModemf

MFVideoAspectRatioMode enumeration (u32).

MFVideoRenderPrefsmf

MFVideoRenderPrefs enumeration (u32).

MF_ATTRIBUTEmf

MF_ATTRIBUTE_TYPE enumeration (u32).

MF_ATTRIBUTES_MATCHmf

MF_ATTRIBUTES_MATCH_TYPE enumeration (u32).

MF_EVENT_FLAGmf

IMFMediaEvent::GetType return type (u32).

MF_OBJECTmf

MF_OBJECT_TYPE enumeration (u32).

MF_RESOLUTIONmf

Source resolver flags (u32).

MF_SERVICEmf

IMFGetService::GetService service_guid (GUID).

MF_TIME_FORMATmf

IMFMediaSession::Start time_format (GUID).

MF_TOPOLOGYmf

MF_TOPOLOGY_TYPE enumeration (u32).

MIIMuser

MENUITEMINFO fMask (u32).

MIMuser

MENUINFO fMask (u32).

MKuser

wm::LButtonDown (and similar) virtual keys (u16).

MKRREDUCEole

How far a moniker should be reduced (u32).

MKSYSole

Moniker classes (u32).

MNDuser

wm::MenuDrag return value (u8).

MNSuser

MENUINFO dwStyle (u32).

MODuser

RegisterHotKey modifiers (u16).

MONITORuser

HMONITOR::MonitorFromPoint, HMONITOR::MonitorFromRect, HWND::MonitorFromWindow flags (u32).

MONITORINFOFuser

MONITORINFOEX dwFlags (u32).

MONITOR_DISPLAY_STATEkernel

MONITOR_DISPLAY_STATE enumeration (u32).

MOUSEEVENTFuser

MOUSEINPUT dwFlags (u32).

MOVEFILEkernel

MoveFileEx flags (u32).

MSGFuser

wm::EnterIdle reason (u8).

NIFshell

NOTIFYICONDATA uFlags (u32).

NIIFshell

NOTIFYICONDATA dwInfoFlags (u32).

NIMshell

Shell_NotifyIcon message (u32).

NISshell

NOTIFYICONDATA dwState and dwStateFlags (u32).

NMcomctl

Notification codes shared among controls (i32).

OBJIDuser

HWND::GetMenuBarInfo idObject (i32).

OBMuser

HINSTANCE::LoadImageBitmap OEM image identifier (u32).

OCRuser

HINSTANCE::LoadImageCursor and HCURSOR::SetSystemCursor OEM cursor identifier (u32).

ODAuser

DRAWITEMSTRUCT itemAction (u32).

ODSuser

DRAWITEMSTRUCT itemState (u32).

ODTuser

DRAWITEMSTRUCT CtlType (u32).

ODT_Cuser

COMPAREITEMSTRUCT and DELETEITEMSTRUCT CtlType (u32).

OICuser

HINSTANCE::LoadImageIcon OEM icon identifier (u32).

OUT_PRECISgdi

HFONT::CreateFont out_precision and LOGFONT lfOutPrecision (u8).

PAGEkernel

Memory protection constants (u32).

PBMcomctl

Progress bar control messages (u32).

PBScomctl

Progress bar control styles (u32).

PBSTcomctl

Progress bar states (u32).

PBTkernel

wm::PowerBroadcast power-management event (u32).

PCgdi

PALETTEENTRY peFlags (u8).

PDMshell

IOperationsProgressDialog::SetMode mode (u32).

PDOPSshell

PDOPSTATUS enumeration (u32).

PFDgdi

PIXELFORMATDESCRIPTOR dwFlags (u32).

PFD_TYPEgdi

PIXELFORMATDESCRIPTOR iPixelType (u8).

PICTUREole

PICTUREATTRIBUTES enumeration (u32).

PICTYPEole

PICTYPE constants (i16).

PIN_DIRECTIONdshow

PIN_DIRECTION enumeration (u32).

PITCHgdi

HFONT::CreateFont pitch_and_family and LOGFONT lfPitchAndFamily (u8) used with FF.

PKEYoleaut

PROPERTYKEY struct.

PMuser

PeekMessage remove_msg (u32).

POWER_SAVINGSkernel

Power scheme identifiers (GUID).

POWER_SETTINGkernel

Power setting identifiers (GUID).

PRINTER_ACCESSwinspool

PRINTER_DEFAULTS DesiredAccess (u32).

PRINTER_ATTRIBUTE_2winspool

PRINTER_INFO_2 Attributes (u32).

PRINTER_ATTRIBUTE_4winspool

PRINTER_INFO_4 Attributes (u32).

PRINTER_CONNECTIONwinspool

PRINTER_CONNECTION_INFO_1 dwFlags (u32).

PRINTER_ENUMwinspool

EnumPrinters2 and EnumPrinters4 flags (u32).

PRINTER_OPTIONwinspool

PRINTER_OPTION enumeration (u32).

PRINTER_STATUSwinspool

PRINTER_INFO_2 status (u32).

PRIORITY_CLASSkernel

GetPriorityClass and SetPriorityClass priority_class (u32).

PROCESSkernel

Process security and access rights (u32).

PROCESSORkernel

SYSTEM_INFO dwProcessorType (u32).

PROCESSOR_ARCHITECTUREkernel

SYSTEM_INFO wProcessorArchitecture (u16).

PROCESS_AFFINITYkernel

QueryProcessAffinityUpdateMode and SetProcessAffinityUpdateMode flags (u32).

PROCESS_HEAPkernel

PROCESS_HEAP_ENTRY wFlags (u16).

PROCESS_NAMEkernel

QueryFullProcessImageName flags (u32).

PROGDLGshell

IOperationsProgressDialog::StartProgressDialog flags (u32).

PROPERTYORIGINuxtheme

PROPERTYORIGIN enumeration (u32).

PSgdi

HPEN::CreatePen style (i32).

PSBTNcomctl

psm::PressButton index (u32).

PSCBcomctl

PFNPROPSHEETCALLBACK message codes (u32).

PSHcomctl

PROPSHEETHEADER dwFlags (u32).

PSMcomctl

Property sheet messages (u32).

PSNcomctl

Property sheet WM_NOTIFY notifications (i32).

PSNRETcomctl

PSN_APPLY return type (u32).

PSPcomctl

PROPSHEETPAGE dwFlags (u32).

PSWIZBcomctl

psm::EnableWizButtons, psm::SetButtonText, psm::SetWizButtons and psm::ShowWizButtons buttons (u32).

PSWIZBFcomctl

psm::SetWizButtons elev_icon (u32).

PTgdi

HDC::PolyDraw aj (u8).

QSuser

GetQueueStatus flags (u32).

QUALITYgdi

HFONT::CreateFont quality and LOGFONT lfQuality (u8).

RBcomctl

Rebar control messages (u32).

RBNcomctl

Rebar control WM_NOTIFY notifications (i32).

RBScomctl

Rebar control styles (u32).

RDWuser

HWND::RedrawWindow flags (u32).

REGadvapi

Registry value types (u32).

REGIONuser

HWND::GetUpdateRgn, HWND::GetWindowRgn, HDC::ExcludeUpdateRgn and HDC::SelectObject return value (i32).

REG_DISPOSITIONadvapi

HKEY::RegCreateKeyEx creation disposition (u32).

REG_OPTIONadvapi

HKEY::RegOpenKeyEx options (u32).

REG_RESTOREadvapi

Registry restore flags (u32).

REG_SAVEadvapi

Registry save flags (u32).

REPLACEFILEkernel

ReplaceFile flags (u32).

RGNgdi

HDC::SelectClipPath mode (i32).

RIDadvapi

The portion of a SID that identifies a user or group in relation to the authority that issued the SID.

ROPgdi

Raster operation code for HDC::BitBlt, HDC::PatBlt and IMAGELISTDRAWPARAMS (u32).

RPC_C_AUTHNole

Authentication service constants (u32).

RPC_C_AUTHN_LEVELole

Authentication level constants (u32).

RPC_C_AUTHZole

Authorization constants (u32).

RPC_C_IMP_LEVELole

Impersonation level constants (u32).

RPC_C_QOS_CAPABILITIESole

Quality of service capabilities (u32).

RRFadvapi

HKEY::GetValue dwFlags (u32).

RTkernel

Predefined resource types (u16).

SBcomctl

Status bar control messages (u32).

SBARScomctl

Status bar control styles (u32).

SBBuser

HWND::EnableScrollBar, HWND::GetScrollInfo, HWND::SetScrollInfo, HWND::SetScrollRange and HWND::ShowScrollBar bar (i32).

SBNcomctl

Status bar control WM_NOTIFY notifications (i32).

SBTcomctl

sb::GetText, sb::GetTextLength and sb::SetText drawing operation (u16).

SB_EMuser

em::Scroll action.

SB_REQuser

wm::HScroll and wm::VScroll request (u16).

SCuser

wm::SysCommand type of system command requested (u32).

SCROLLWuser

ScrollWindowEx flags (u32).

SCSkernel

GetBinaryType return value (u32).

SC_MANAGERadvapi

Service Control Manager access rights flags (u32).

SEkernel

SECURITY_DESCRIPTOR_CONTROL bit flags (u16).

SECkernel

HFILE::CreateFileMapping sec (u32).

SECTIONkernel

Composes FILE_MAP (u32).

SECURITY_IMPERSONATIONadvapi

SECURITY_IMPERSONATION_LEVEL enumeration (u32).

SECURITY_INFORMATIONkernel

SECURITY_INFORMATION flags (u32).

SEC_WINNT_AUTH_IDENTITYole

COAUTHIDENTITY Flags (u32).

SEEKING_FLAGSdshow

IMediaSeeking::SetPositions flags (u32).

SEE_MASKadvapi and shell

SHELLEXECUTEINFO fMask (u32).

SERVICEkernel

Service access rights flags (u32).

SERVICE_ACCEPTadvapi

SERVICE_STATUS dwControlsAccepted (u32).

SERVICE_CONTROLadvapi

LPHANDLER_FUNCTION_EX dwControl (u32).

SERVICE_ERRORadvapi

HSC::CreateService error_control (u32).

SERVICE_STARTadvapi

HSC::CreateService start_type (u32).

SERVICE_STATEadvapi

HSERVICESTATUS::SetServiceStatus current_state (u32)

SERVICE_TYPEadvapi

HSC::CreateService service_type (u32).

SEVERITYole

HRESULT severity (u8).

SE_PRIVadvapi

Privilege constants (&'static str).

SE_PRIV_ATTRadvapi

Privilege attributes (u32).

SFGAOshell

SFGAO constants (u32).

SHARDshell

SHARD enumeration (u32).

SHCIDSshell

IShellFolder::CompareIDs sorting_flags (u32).

SHCONTFshell

SHCONTF enumeration.

SHGFIshell

SHGetFileInfo flags (u32).

SHGSIshell

SHGetStockIconInfo flags (u32).

SHTDN_REASONkernel

Shutdown reason codes (u32).

SICHINTFshell

SICHINTF enumeration (u32).

SID_NAME_USEadvapi

SID_NAME_USE enumeration (u32).

SIFuser

SCROLLINFO fMask (u32).

SIGDNshell

SIGDN enumeration (u32).

SIIDshell

SHSTOCKICONID enumeration, SHGetStockIconInfo siid (u32).

SIZE_Ruser

wm::Size request (u8).

SLGPshell

IShellLink::GetPath flags (u32).

SLRole

BIND_OPTS3 and IShellLink::Resolve flags (u32).

SMuser

GetSystemMetrics index (i32).

SMTOuser

SendMessageTimeout flags (u32).

SORTkernel

Sort order identifiers (u16).

SPACTIONshell

SPACTION enumeration (u32).

SPIuser

SystemParametersInfo action (u32).

SPIFuser

SystemParametersInfo win_ini (u32).

SSuser

Label control styles (u32).

STANDARD_RIGHTSkernel

Standard access rights flags (u32).

STAPuxtheme

HTHEME::GetThemeAppProperties return value (u32).

STARTFkernel

CreateProcess dwFlags (u32).

STATE_SYSTEMuser

DATETIMEPICKERINFO stateCheck and stateButton, TITLEBARINFOEX rgstate, COMBOBOXINFO stateButton (u32).

STD_HANDLEkernel

GetStdHandle std_handle (u32).

STGCole

STGC enumeration (u32).

STGMole

STGM enumeration (u32).

STGMOVEole

STGMOVE enumeration (u32).

STMcomctl

Static control messages (u32).

STNuser

Static control WM_COMMAND notifications (u16).

STOCK_BRUSHgdi

HBRUSH::GetStockObject i (i32).

STOCK_FONTgdi

HFONT::GetStockObject i (i32).

STOCK_PENgdi

HPEN::GetStockObject i (i32).

STPFLAGshell

STPFLAG enumeration (u32).

STREAM_SEEKole

STREAM_SEEK enumeration (u32).

STRETCH_MODEgdi

HDC::SetStretchBltMode mode (i32).

STRING_FORMwinspool

FORM_INFO_2 StringType (u32).

SUBLANGkernel

Sublanguage identifier (u16).

SWkernel

HWND::ShowWindow show_cmd (i32).

SWPuser

HWND::SetWindowPos flags (u32).

SW_Suser

wm::ShowWindow status (u8).

SYSTEM_POWER_CONDITIONkernel

SYSTEM_POWER_CONDITION enumeration (u32).

TAgdi

HDC::SetTextAlign align (u32).

TASK_ACTION_TYPEtaskschd

TASK_ACTION_TYPE enumeration (u32);

TASK_COMPATIBILITYtaskschd

TASK_COMPATIBILITY enumeration (u32).

TASK_CREATIONtaskschd

TASK_CREATION enumeration (u32).

TASK_INSTANCEStaskschd

TASK_INSTANCES_POLICY enumeration (u32).

TASK_LOGONtaskschd

TASK_LOGON_TYPE enumeration (u32).

TASK_RUNLEVELtaskschd

TASK_RUNLEVEL_TYPE enumeration (u32).

TASK_STATEtaskschd

TASK_STATE enumeration (u32).

TASK_TRIGGER_TYPE2taskschd

TASK_TRIGGER_TYPE2 enumeration (u32).

TBcomctl

NMTRBTHUMBPOSCHANGING nReason (i32).

TBBFcomctl

tbm::GetBitmapFlags return value (u32).

TBIFcomctl

TBBUTTONINFO dwFlags (u32).

TBIMHTcomctl

TBINSERTMARK dwFlags (u32).

TBMcomctl

Toolbar control messages (u32).

TBMFcomctl

TBMETRICS dwMask (u32).

TBNcomctl

Toolbar control WM_NOTIFY notifications (i32).

TBPFshell

ITaskbarList3::SetProgressState tbpFlags (u32).

TBScomctl

Trackbar control styles (u32).

TBSTATEcomctl

Toolbar button states (u8).

TBSTYLE_EXcomctl

Extended toolbar control styles (u32).

TBTScomctl

trbm::SetTipSide location (u8).

TCHTcomctl

TCHITTESTINFO flags (u32).

TCIFcomctl

TCITEM mask (u32).

TCIScomctl

Tab control item states (u32).

TCMcomctl

Tab control messages (u32).

TCNcomctl

Tab control WM_NOTIFY notifications (i32).

TCScomctl

Tab control styles (u32).

TCS_EXcomctl

Extended tab control styles (u32).

TDCBFcomctl

HWND::TaskDialog and TASKDIALOGCONFIG dwCommonButtons (i32).

TDFcomctl

TASKDIALOGCONFIG dwFlags (i32).

TDNcomctl

PFTASKDIALOGCALLBACK msg (u32).

TD_ICONcomctl

HWND::TaskDialog pszIcon and IconIdTd Td (u16).

TH32CSkernel

HPROCESSLIST::CreateToolhelp32Snapshot flags (u32).

THEMESIZEuxtheme

THEMESIZE enumeration (u32).

THREAD_CREATEkernel

HTHREAD::CreateThread flags (u32).

TIME_FORMATdshow

IMediaSeeking::SetTimeFormat format (GUID).

TMEuser

TrackMouseEvent dwFlags (u32).

TMTuxtheme

Theme property identifiers (i32).

TOKENkernel

Token access rights.

TOKEN_ELEVATION_TYPEadvapi

TOKEN_ELEVATION_TYPE enumeration (u32).

TOKEN_INFORMATION_CLASSadvapi

TOKEN_INFORMATION_CLASS enumeration (u32).

TOKEN_MANDATORY_POLICYkernel

TOKEN_MANDATORY_POLICY Policy (u32).

TOKEN_TYPEadvapi

TOKEN_TYPE enumeration (u32).

TPMuser

TrackPopupMenu flags (u32).

TRANSACTIONadvapi

Transaction access masks (u32).

TRANSACTION_OPTadvapi

CrateTransaction options (u32).

TRBMcomctl

Trackbar control messages (u32).

TRBNcomctl

Trackbar control WM_NOTIFY notifications (i32).

TSFshell

_TRANSFER_SOURCE_FLAGS enumeration (u32).

TTIcomctl

EDITBALLOONTIP ttiIcon (i32).

TVCcomctl

NMTREEVIEW action (u32).

TVEcomctl

tvm::Expand action (u32).

TVGNcomctl

tvm::GetNextItem which (u32).

TVHTcomctl

TVHITTESTINFO flags (u32).

TVIcomctl

TVINSERTSTRUCT hInsertAfter (isize).

TVIFcomctl

TVITEM mask (u32).

TVIScomctl

Tree view item states (u32)

TVIS_EXcomctl

TVITEMEX uStateEx (u32).

TVMcomctl

Tree view control messages (u32).

TVNcomctl

Tree view control WM_NOTIFY notifications (i32).

TVScomctl

Tree view control styles (u32).

TVSBFcomctl

tvm::SetBorder action (u32).

TVSILcomctl

tvm::GetImageList and tvm::SetImageList kind.

TVS_EXcomctl

Extended tree view control styles (u32).

TYMEDole

TYMED enumeration (u32).

UDMcomctl

Up-down control messages (u32).

UDNcomctl

Up-down control WM_NOTIFY notifications (i32).

UDScomctl

Up-down control styles (u32).

ULWuser

UpdateLayeredWindow flags (u32).

UOIuser

HPROCESS::SetUserObjectInformation index (i32).

USER_ACTIVITY_PRESENCEkernel

PowerSetting::GlobalUserPresence enumeration (u32).

VER_CONDkernel

VerSetConditionMask condition (u8).

VER_MASKkernel

VerifyVersionInfo and VerSetConditionMask type_mask (u32).

VER_NTkernel

OSVERSIONINFOEX wProductType (u8).

VER_PLATFORMkernel

OSVERSIONINFOEX dwPlatformId (u32).

VER_SUITEkernel

OSVERSIONINFOEX wSuiteMask (u16).

VFTversion

VS_FIXEDFILEINFO dwFileType (u32).

VFT2version

VS_FIXEDFILEINFO dwFileSubType (u32).

VKuser

Virtual key codes (u16).

VK_DIRcomctl

LVFINDINFO vkDirection (u16).

VOSversion

VS_FIXEDFILEINFO dwFileOS (u32).

VSuxtheme

Visual style parts and states. These are originally two separated constant types, but since they form a hierarchy, one single constant type can represent both part and state.

VS_FFversion

VS_FIXEDFILEINFO dwFileFlags (u32).

VToleaut

VARENUM enumeration (u16).

WAuser

wm::Activate activation state (u16).

WAITkernel

HPROCESS::WaitForSingleObject return value (u32).

WCkernel

WideCharToMultiByte flags (u32).

WDAuser

HWND::GetWindowDisplayAffinity and HWND::SetWindowDisplayAffinity dwAffinity (u32).

WELL_KNOWN_SID_TYPEadvapi

WELL_KNOWN_SID_TYPE enumeration (u32).

WHuser

HHOOK::CallNextHookEx code and HHOOK::SetWindowsHookEx hook_id (i32).

WIN32kernel

_WIN32 version definitions (u16).

WMuser

Window message codes (u32).

WMPNuser

wm::ParentNotify event (u16).

WMSZuser

wm::Sizing window edge (u8).

WPFuser

WINDOWPLACEMENT flags (u32).

WSuser

Window styles (u32).

WS_EXuser

Extended window styles (u32).

WTSkernel

wm::WtsSessionChange state (u32).

WVRuser

wm::NcCalcSize return flags (u32).