Module winsafe::co

Native constants.

All types can be converted from/to their underlying integer type. They all implement the NativeConst trait; those who can be combined as bitflags also implement NativeBitflag.

Among these constant types, three are error types: [CDERR], ERROR and [HRESULT].

Structs

ACCESS_RIGHTS

Standard access rights flags (u32).

CLAIM_SECURITY_ATTRIBUTE

CLAIM_SECURITY_ATTRIBUTE_V1 Flags (u16);

CLAIM_SECURITY_ATTRIBUTE_TYPE

CLAIM_SECURITY_ATTRIBUTE_V1 ValueType (u16);

CONSOLE

HSTD::SetConsoleMode mode (u32).

CP

WideCharToMultiByte and MultiByteToWideChar code_page identifiers (u16).

CREATE

Process creation flags (u32).

CREATE_EVENT

HEVENT::CreateEventEx flags (u32).

DBT

wm::DeviceChange event (u16).

DBTF

DEV_BROADCAST_VOLUME dbcv_flags (u16).

DBT_DEVTYP

DEV_BROADCAST_HDR dbch_devicetype (u32).

DISPOSITION

HFILE::CreateFile creation_disposition (u32).

DRIVE

GetDriveType return type (u32).

ERROR

A native system error code (u32), commonly retrieved by GetLastError function.

EVENTLOG

HEVENTLOG::ReportEvent event_type u16.

EVENT_RIGHTS

Event access rights (u32).

FILE_ATTRIBUTE

File attributes (u32).

FILE_CACHE

GetSystemFileCacheSize returned flags (u32).

FILE_FLAG

HFILE::CreateFile flags (u32).

FILE_MAP

HFILEMAP::MapViewOfFile desired_access (u32).

FILE_SECURITY

HFILE::CreateFile security (u32).

FILE_SHARE

HFILE::CreateFile share_mode (u32).

FILE_STARTING_POINT

HFILE::SetFilePointerEx move_method (u32).

FILE_TYPE

HFILE::GetFileType return value (u32).

FILE_VOL

GetVolumeInformation flags (u32).

FIRMWARE_TYPE

FIRMWARE_TYPE enumeration (u32).

FORMAT_MESSAGE

FormatMessage dwFlags (u32).

GENERIC

Generic access rights flags (u32).

GMEM

HGLOBAL::GlobalAlloc and HGLOBAL::GlobalReAlloc flags (u32).

GR

HPROCESS::GetGuiResources flags (u32).

HEAP_ALLOC

HHEAP::HeapAlloc flags (u32).

HEAP_CREATE

HHEAP::HeapCreate options (u32).

HEAP_INFORMATION

HHEAP::HeapSetInformation information class (u32).

HEAP_REALLOC

HHEAP::HeapReAlloc options (u32).

HEAP_SIZE

HHEAP::HeapSize and HHEAP::HeapCompact options (u32).

HF32

HEAPLIST32 dwFlags (u32).

KEY

Registry access rights (u32).

LANG

Language identifier (u16).

LMEM

HLOCAL::LocalAlloc and HLOCAL::LocalReAlloc flags (u32).

MBC

MultiByteToWideChar flags (u32).

MONITOR_DISPLAY_STATE

MONITOR_DISPLAY_STATE enumeration (u32).

PAGE

HFILE::CreateFileMapping protect (u32).

PBT

wm::PowerBroadcast power-management event (u32).

POWER_SAVINGS

Power scheme identifiers (GUID).

POWER_SETTING

Power setting identifiers (GUID).

PRIORITY_CLASS

GetPriorityClass and SetPriorityClass priority_class (u32).

PROCESS

Process security and access rights (u32).

PROCESSOR

SYSTEM_INFO dwProcessorType (u32).

PROCESSOR_ARCHITECTURE

SYSTEM_INFO wProcessorArchitecture (u16).

PROCESS_AFFINITY

QueryProcessAffinityUpdateMode and SetProcessAffinityUpdateMode flags (u32).

PROCESS_HEAP

PROCESS_HEAP_ENTRY wFlags (u16).

PROCESS_NAME

QueryFullProcessImageName flags (u32).

REG

Registry value types (u32).

REG_DISPOSITION

HKEY::RegCreateKeyEx creation disposition (u32).

REG_OPTION

HKEY::RegOpenKeyEx options (u32).

REG_RESTORE

Registry restore flags (u32).

REG_SAVE

Registry save flags (u32).

REPLACEFILE

ReplaceFile flags (u32).

RID

The portion of a SID that identifies a user or group in relation to the authority that issued the SID.

RRF

HKEY::GetValue dwFlags (u32).

RT

Predefined resource types (u16).

SCS

GetBinaryType return value (u32).

SC_MANAGER

Service Control Manager access rights flags (u32).

SE

SECURITY_DESCRIPTOR_CONTROL bit flags (u16).

SECTION

Composes FILE_MAP (u32).

SECURITY_IMPERSONATION

SECURITY_IMPERSONATION_LEVEL enumeration (u32).

SECURITY_INFORMATION

SECURITY_INFORMATION flags (u32).

SERVICE

Service access rights flags (u32).

SERVICE_ACCEPT

SERVICE_STATUS dwControlsAccepted (u32).

SERVICE_CONTROL

LPHANDLER_FUNCTION_EX dwControl (u32).

SERVICE_ERROR

HSC::CreateService error_control (u32).

SERVICE_START

HSC::CreateService start_type (u32).

SERVICE_STATE

HSERVICESTATUS::SetServiceStatus current_state (u32)

SERVICE_TYPE

HSC::CreateService service_type (u32).

SE_PRIV

Privilege constants (&'static str).

SE_PRIV_ATTR

Privilege attributes (u32).

SHTDN_REASON

Shutdown reason codes (u32).

SID_NAME_USE

SID_NAME_USE enumeration (u32).

SORT

Sort order identifiers (u16).

STANDARD_RIGHTS

Standard access rights flags (u32).

STARTF

HPROCESS::CreateProcess dwFlags (u32).

STD_HANDLE

GetStdHandle std_handle (u32).

SUBLANG

Sublanguage identifier (u16).

SW

HWND::ShowWindow show_cmd (i32).

SYSTEM_POWER_CONDITION

SYSTEM_POWER_CONDITION enumeration (u32).

TH32CS

HPROCESSLIST::CreateToolhelp32Snapshot flags (u32).

THREAD_CREATE

HTHREAD::CreateThread flags (u32).

TOKEN

Token access rights.

TOKEN_ELEVATION_TYPE

TOKEN_ELEVATION_TYPE enumeration (u32).

TOKEN_INFORMATION_CLASS

TOKEN_INFORMATION_CLASS enumeration (u32).

TOKEN_MANDATORY_POLICY

TOKEN_MANDATORY_POLICY Policy (u32).

TOKEN_TYPE

TOKEN_TYPE enumeration (u32).

TRANSACTION

Transaction access masks (u32).

TRANSACTION_OPT

CrateTransaction options (u32).

USER_ACTIVITY_PRESENCE

PowerSetting::GlobalUserPresence enumeration (u32).

VER_COND

VerSetConditionMask condition (u8).

VER_MASK

VerifyVersionInfo and VerSetConditionMask type_mask (u32).

VER_NT

OSVERSIONINFOEX wProductType (u8).

VER_PLATFORM

OSVERSIONINFOEX dwPlatformId (u32).

VER_SUITE

OSVERSIONINFOEX wSuiteMask (u16).

WAIT

HPROCESS::WaitForSingleObject return value (u32).

WC

WideCharToMultiByte flags (u32).

WELL_KNOWN_SID_TYPE

WELL_KNOWN_SID_TYPE enumeration (u32).

WIN32

_WIN32 version definitions (u16).

WTS

wm::WtsSessionChange state (u32).