Struct winsafe::co::CLSCTX

#[repr(transparent)]
pub struct CLSCTX(_);

CLSCTX enumeration (u32).

Implementations

impl CLSCTX

pub fn has(&self, other: CLSCTX) -> bool

Tells whether other bitflag style is present. Equivalent to (val & other) != 0.

This method is common to all constant types.

impl CLSCTX

pub const INPROC_SERVER: Self

The code that creates and manages objects of this class is a DLL that runs in the same process as the caller of the function specifying the class context.

pub const INPROC_HANDLER: Self

The code that manages objects of this class is an in-process handler. This is a DLL that runs in the client process and implements client-side structures of this class when instances of the class are accessed remotely.

pub const LOCAL_SERVER: Self

The EXE code that creates and manages objects of this class runs on same machine but is loaded in a separate process space.

pub const REMOTE_SERVER: Self

A remote context. The LocalServer32 or LocalService code that creates and manages objects of this class is run on a different computer.

pub const NO_CODE_DOWNLOAD: Self

Disables the downloading of code from the directory service or the Internet. This flag cannot be set at the same time as CLSCTX::ENABLE_CODE_DOWNLOAD.

pub const NO_CUSTOM_MARSHAL: Self

Specify if you want the activation to fail if it uses custom marshalling.

pub const ENABLE_CODE_DOWNLOAD: Self

Enables the downloading of code from the directory service or the Internet. This flag cannot be set at the same time as CLSCTX::NO_CODE_DOWNLOAD.

pub const NO_FAILURE_LOG: Self

The CLSCTX::NO_FAILURE_LOG can be used to override the logging of failures in CoCreateInstanceEx.

pub const DISABLE_AAA: Self

Disables activate-as-activator (AAA) activations for this activation only.

pub const ENABLE_AAA: Self

Enables activate-as-activator (AAA) activations for this activation only.

pub const FROM_DEFAULT_CONTEXT: Self

Begin this activation from the default context of the current apartment.

pub const ACTIVATE_X86_SERVER: Self

Activate or connect to a 32-bit version of the server; fail if one is not registered.

pub const ACTIVATE_32_BIT_SERVER: Self

Activate or connect to a 32-bit version of the server; fail if one is not registered.

pub const ACTIVATE_64_BIT_SERVER: Self

Activate or connect to a 64 bit version of the server; fail if one is not registered.

pub const ACTIVATE_AAA_AS_IU: Self

Specify this flag for Interactive User activation behavior for As-Activator servers.

pub const ACTIVATE_ARM32_SERVER: Self

(No official docs for this entry.)