winsafe::co

Struct FACILITY

pub struct FACILITY(/* private fields */);

Available on crate feature ole only.

HRESULT facility (u32).

Implementations

impl FACILITY

pub const unsafe fn as_mut(&mut self) -> &mut u32

Returns a mutable reference to the underlying raw value.

Safety

Be sure the integer being set is meaningful for the actual type.

pub const unsafe fn from_raw(v: u32) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u32

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl FACILITY

pub const NULL: Self

pub const RPC: Self

pub const DISPATCH: Self

pub const STORAGE: Self

pub const ITF: Self

pub const WIN32: Self

pub const WINDOWS: Self

pub const SSPI: Self

pub const SECURITY: Self

pub const CONTROL: Self

pub const CERT: Self

pub const INTERNET: Self

pub const MEDIASERVER: Self

pub const MSMQ: Self

pub const SETUPAPI: Self

pub const SCARD: Self

pub const COMPLUS: Self

pub const AAF: Self

pub const URT: Self

pub const ACS: Self

pub const DPLAY: Self

pub const UMI: Self

pub const SXS: Self

pub const WINDOWS_CE: Self

pub const HTTP: Self

pub const USERMODE_COMMONLOG: Self

pub const WER: Self

pub const USERMODE_FILTER_MANAGER: Self

pub const BACKGROUNDCOPY: Self

pub const CONFIGURATION: Self

pub const WIA: Self

pub const STATE_MANAGEMENT: Self

pub const METADIRECTORY: Self

pub const WINDOWSUPDATE: Self

pub const DIRECTORYSERVICE: Self

pub const GRAPHICS: Self

pub const SHELL: Self

pub const NAP: Self

pub const TPM_SERVICES: Self

pub const TPM_SOFTWARE: Self

pub const UI: Self

pub const XAML: Self

pub const ACTION_QUEUE: Self

pub const PLA: Self

pub const WINDOWS_SETUP: Self

pub const FVE: Self

pub const FWP: Self

pub const WINRM: Self

pub const NDIS: Self

pub const USERMODE_HYPERVISOR: Self

pub const CMI: Self

pub const USERMODE_VIRTUALIZATION: Self

pub const USERMODE_VOLMGR: Self

pub const BCD: Self

pub const USERMODE_VHD: Self

pub const USERMODE_HNS: Self

pub const SDIAG: Self

pub const WEBSERVICES: Self

pub const WINPE: Self

pub const WPN: Self

pub const WINDOWS_STORE: Self

pub const INPUT: Self

pub const EAP: Self

pub const WINDOWS_DEFENDER: Self

pub const OPC: Self

pub const XPS: Self

pub const MBN: Self

pub const POWERSHELL: Self

pub const RAS: Self

pub const P2P_INT: Self

pub const P2P: Self

pub const DAF: Self

pub const BLUETOOTH_ATT: Self

pub const AUDIO: Self

pub const STATEREPOSITORY: Self

pub const VISUALCPP: Self

pub const SCRIPT: Self

pub const PARSE: Self

pub const BLB: Self

pub const BLB_CLI: Self

pub const WSBAPP: Self

pub const BLBUI: Self

pub const USN: Self

pub const USERMODE_VOLSNAP: Self

pub const TIERING: Self

pub const WSB_ONLINE: Self

pub const ONLINE_ID: Self

pub const DEVICE_UPDATE_AGENT: Self

pub const DRVSERVICING: Self

pub const DLS: Self

pub const DELIVERY_OPTIMIZATION: Self

pub const USERMODE_SPACES: Self

pub const USER_MODE_SECURITY_CORE: Self

pub const USERMODE_LICENSING: Self

pub const SOS: Self

pub const DEBUGGERS: Self

pub const SPP: Self

pub const RESTORE: Self

pub const DMSERVER: Self

pub const DEPLOYMENT_SERVICES_SERVER: Self

pub const DEPLOYMENT_SERVICES_IMAGING: Self

pub const DEPLOYMENT_SERVICES_MANAGEMENT: Self

pub const DEPLOYMENT_SERVICES_UTIL: Self

pub const DEPLOYMENT_SERVICES_BINLSVC: Self

pub const DEPLOYMENT_SERVICES_PXE: Self

pub const DEPLOYMENT_SERVICES_TFTP: Self

pub const DEPLOYMENT_SERVICES_TRANSPORT_MANAGEMENT: Self

pub const DEPLOYMENT_SERVICES_DRIVER_PROVISIONING: Self

pub const DEPLOYMENT_SERVICES_MULTICAST_SERVER: Self

pub const DEPLOYMENT_SERVICES_MULTICAST_CLIENT: Self

pub const DEPLOYMENT_SERVICES_CONTENT_PROVIDER: Self

pub const LINGUISTIC_SERVICES: Self

pub const AUDIOSTREAMING: Self

pub const ACCELERATOR: Self

pub const WMAAECMA: Self

pub const DIRECTMUSIC: Self

pub const DIRECT3D10: Self

pub const DXGI: Self

pub const DXGI_DDI: Self

pub const DIRECT3D11: Self

pub const DIRECT3D11_DEBUG: Self

pub const DIRECT3D12: Self

pub const DIRECT3D12_DEBUG: Self

pub const LEAP: Self

pub const AUDCLNT: Self

pub const WINCODEC_DWRITE_DWM: Self

pub const WINML: Self

pub const DIRECT2D: Self

pub const DEFRAG: Self

pub const USERMODE_SDBUS: Self

pub const JSCRIPT: Self

pub const PIDGENX: Self

pub const EAS: Self

pub const WEB: Self

pub const WEB_SOCKET: Self

pub const MOBILE: Self

pub const SQLITE: Self

pub const UTC: Self

pub const WEP: Self

pub const SYNCENGINE: Self

pub const XBOX: Self

pub const GAME: Self

pub const PIX: Self