Struct winsafe::co::HRESULT

#[repr(transparent)]
pub struct HRESULT(_);

A native COM error code (u32).

Implements the standard Error trait.

Implements the Debug and Display traits to show the error code along with the error description, taken from FormattedError trait. For example, the code below:

use winsafe::co::HRESULT;

println!("{}", HRESULT::E_INVALIDARG);

Will print:

[0x80070057 2147942487] The parameter is incorrect.

Implementations

impl HRESULT

pub const S_OK: Self

Operation successful.

pub const S_FALSE: Self

pub const E_UNEXPECTED: Self

pub const E_NOTIMPL: Self

pub const E_OUTOFMEMORY: Self

pub const E_INVALIDARG: Self

pub const E_NOINTERFACE: Self

pub const E_POINTER: Self

pub const E_HANDLE: Self

pub const E_ABORT: Self

pub const E_FAIL: Self

pub const E_ACCESSDENIED: Self

pub const E_PENDING: Self

pub const E_BOUNDS: Self

pub const E_CHANGED_STATE: Self

pub const E_ILLEGAL_STATE_CHANGE: Self

pub const E_ILLEGAL_METHOD_CALL: Self

pub const RPC_E_CALL_REJECTED: Self

pub const RPC_E_CALL_CANCELED: Self

pub const RPC_E_CANTPOST_INSENDCALL: Self

pub const RPC_E_CANTCALLOUT_INASYNCCALL: Self

pub const RPC_E_CANTCALLOUT_INEXTERNALCALL: Self

pub const RPC_E_CONNECTION_TERMINATED: Self

pub const RPC_E_SERVER_DIED: Self

pub const RPC_E_CLIENT_DIED: Self

pub const RPC_E_INVALID_DATAPACKET: Self

pub const RPC_E_CANTTRANSMIT_CALL: Self

pub const RPC_E_CLIENT_CANTMARSHAL_DATA: Self

pub const RPC_E_CLIENT_CANTUNMARSHAL_DATA: Self

pub const RPC_E_SERVER_CANTMARSHAL_DATA: Self

pub const RPC_E_SERVER_CANTUNMARSHAL_DATA: Self

pub const RPC_E_INVALID_DATA: Self

pub const RPC_E_INVALID_PARAMETER: Self

pub const RPC_E_CANTCALLOUT_AGAIN: Self

pub const RPC_E_SERVER_DIED_DNE: Self

pub const RPC_E_SYS_CALL_FAILED: Self

pub const RPC_E_OUT_OF_RESOURCES: Self

pub const RPC_E_ATTEMPTED_MULTITHREAD: Self

pub const RPC_E_NOT_REGISTERED: Self

pub const RPC_E_FAULT: Self

pub const RPC_E_SERVERFAULT: Self

pub const RPC_E_CHANGED_MODE: Self

pub const RPC_E_INVALIDMETHOD: Self

pub const RPC_E_DISCONNECTED: Self

pub const RPC_E_RETRY: Self

pub const RPC_E_SERVERCALL_RETRYLATER: Self

pub const RPC_E_SERVERCALL_REJECTED: Self

pub const RPC_E_INVALID_CALLDATA: Self

pub const RPC_E_CANTCALLOUT_ININPUTSYNCCALL: Self

pub const RPC_E_WRONG_THREAD: Self

pub const RPC_E_THREAD_NOT_INIT: Self

pub const RPC_E_VERSION_MISMATCH: Self

pub const RPC_E_INVALID_HEADER: Self

pub const RPC_E_INVALID_EXTENSION: Self

pub const RPC_E_INVALID_IPID: Self

pub const RPC_E_INVALID_OBJECT: Self

pub const RPC_E_CALL_COMPLETE: Self

pub const RPC_E_UNSECURE_CALL: Self

pub const RPC_E_TOO_LATE: Self

pub const RPC_E_NO_GOOD_SECURITY_PACKAGES: Self

pub const RPC_E_ACCESS_DENIED: Self

pub const RPC_E_REMOTE_DISABLED: Self

pub const RPC_E_INVALID_OBJREF: Self

pub const RPC_E_NO_CONTEXT: Self

pub const RPC_E_TIMEOUT: Self

pub const RPC_E_NO_SYNC: Self

pub const RPC_E_FULLSIC_REQUIRED: Self

pub const RPC_E_INVALID_STD_NAME: Self

pub const RPC_E_UNEXPECTED: Self

pub const CO_E_NOTINITIALIZED: Self

pub const CO_E_ALREADYINITIALIZED: Self

pub const CO_E_CANTDETERMINECLASS: Self

pub const CO_E_CLASSSTRING: Self

pub const CO_E_IIDSTRING: Self

pub const CO_E_APPNOTFOUND: Self

pub const CO_E_APPSINGLEUSE: Self

pub const CO_E_ERRORINAPP: Self

pub const CO_E_DLLNOTFOUND: Self

pub const CO_E_ERRORINDLL: Self

pub const CO_E_WRONGOSFORAPP: Self

pub const CO_E_OBJNOTREG: Self

pub const CO_E_OBJISREG: Self

pub const CO_E_OBJNOTCONNECTED: Self

pub const CO_E_APPDIDNTREG: Self

pub const CO_E_RELEASED: Self

pub const MF_E_PLATFORM_NOT_INITIALIZED: Self

pub const MF_E_BUFFERTOOSMALL: Self

pub const MF_E_INVALIDREQUEST: Self

pub const MF_E_INVALIDSTREAMNUMBER: Self

pub const MF_E_INVALIDMEDIATYPE: Self

pub const MF_E_NOTACCEPTING: Self

pub const MF_E_NOT_INITIALIZED: Self

pub const MF_E_UNSUPPORTED_REPRESENTATION: Self

pub const MF_E_NO_MORE_TYPES: Self

pub const MF_E_UNSUPPORTED_SERVICE: Self

pub const MF_E_UNEXPECTED: Self

pub const MF_E_INVALIDNAME: Self

pub const MF_E_INVALIDTYPE: Self

pub const MF_E_INVALID_FILE_FORMAT: Self

pub const MF_E_INVALIDINDEX: Self

pub const MF_E_INVALID_TIMESTAMP: Self

pub const MF_E_UNSUPPORTED_SCHEME: Self

pub const MF_E_UNSUPPORTED_BYTESTREAM_TYPE: Self

pub const MF_E_UNSUPPORTED_TIME_FORMAT: Self

pub const MF_E_NO_SAMPLE_TIMESTAMP: Self

pub const MF_E_NO_SAMPLE_DURATION: Self

pub const MF_E_INVALID_STREAM_DATA: Self

pub const MF_E_RT_UNAVAILABLE: Self

pub const MF_E_UNSUPPORTED_RATE: Self

pub const MF_E_THINNING_UNSUPPORTED: Self

pub const MF_E_REVERSE_UNSUPPORTED: Self

pub const MF_E_UNSUPPORTED_RATE_TRANSITION: Self

pub const MF_E_RATE_CHANGE_PREEMPTED: Self

pub const MF_E_NOT_FOUND: Self

pub const MF_E_NOT_AVAILABLE: Self

pub const MF_E_NO_CLOCK: Self

pub const MF_S_MULTIPLE_BEGIN: Self

pub const MF_E_MULTIPLE_BEGIN: Self

pub const MF_E_MULTIPLE_SUBSCRIBERS: Self

pub const MF_E_TIMER_ORPHANED: Self

pub const MF_E_STATE_TRANSITION_PENDING: Self

pub const MF_E_UNSUPPORTED_STATE_TRANSITION: Self

pub const MF_E_UNRECOVERABLE_ERROR_OCCURRED: Self

pub const MF_E_SAMPLE_HAS_TOO_MANY_BUFFERS: Self

pub const MF_E_SAMPLE_NOT_WRITABLE: Self

pub const MF_E_INVALID_KEY: Self

pub const MF_E_BAD_STARTUP_VERSION: Self

pub const MF_E_UNSUPPORTED_CAPTION: Self

pub const MF_E_INVALID_POSITION: Self

pub const MF_E_ATTRIBUTENOTFOUND: Self

pub const MF_E_PROPERTY_TYPE_NOT_ALLOWED: Self

pub const MF_E_TOPO_INVALID_OPTIONAL_NODE: Self

pub const MF_E_TOPO_CANNOT_FIND_DECRYPTOR: Self

pub const MF_E_TOPO_CODEC_NOT_FOUND: Self

pub const MF_E_TOPO_CANNOT_CONNECT: Self

pub const MF_E_TOPO_UNSUPPORTED: Self

pub const MF_E_TOPO_INVALID_TIME_ATTRIBUTES: Self

pub const MF_E_TOPO_LOOPS_IN_TOPOLOGY: Self

pub const MF_E_TOPO_MISSING_PRESENTATION_DESCRIPTOR: Self

pub const MF_E_TOPO_MISSING_STREAM_DESCRIPTOR: Self

pub const MF_E_TOPO_STREAM_DESCRIPTOR_NOT_SELECTED: Self

pub const MF_E_TOPO_MISSING_SOURCE: Self

pub const MF_E_TOPO_SINK_ACTIVATES_UNSUPPORTED: Self

pub const MF_E_TRANSFORM_TYPE_NOT_SET: Self

impl HRESULT

pub fn code(self) -> u16

HRESULT_CODE method. Originally a macro.

pub fn facility(self) -> FACILITY

HRESULT_FACILITY method. Originally a macro.

pub fn severity(self) -> u8

HRESULT_SEVERITY method. Originally a macro.