winsafe::co

Struct LVM

pub struct LVM(/* private fields */);

Available on crate feature comctl only.

List view control messages (u32).

This is a window message, convertible to WM.

Implementations

impl LVM

pub const unsafe fn as_mut(&mut self) -> &mut u32

Returns a mutable reference to the underlying raw value.

Safety

Be sure the integer being set is meaningful for the actual type.

pub const unsafe fn from_raw(v: u32) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u32

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl LVM

pub const fn has(&self, other: Self) -> bool

Returns true if the other bitflag value is present.

Equivalent to (val & other) != 0.

impl LVM

pub const SETUNICODEFORMAT: Self

pub const GETUNICODEFORMAT: Self

pub const GETBKCOLOR: Self

pub const SETBKCOLOR: Self

pub const GETIMAGELIST: Self

pub const SETIMAGELIST: Self

pub const GETITEMCOUNT: Self

pub const DELETEITEM: Self

pub const DELETEALLITEMS: Self

pub const GETCALLBACKMASK: Self

pub const SETCALLBACKMASK: Self

pub const GETNEXTITEM: Self

pub const GETITEMRECT: Self

pub const SETITEMPOSITION: Self

pub const GETITEMPOSITION: Self

pub const HITTEST: Self

pub const ENSUREVISIBLE: Self

pub const SCROLL: Self

pub const REDRAWITEMS: Self

pub const ARRANGE: Self

pub const GETEDITCONTROL: Self

pub const DELETECOLUMN: Self

pub const GETCOLUMNWIDTH: Self

pub const SETCOLUMNWIDTH: Self

pub const GETHEADER: Self

pub const CREATEDRAGIMAGE: Self

pub const GETVIEWRECT: Self

pub const GETTEXTCOLOR: Self

pub const SETTEXTCOLOR: Self

pub const GETTEXTBKCOLOR: Self

pub const SETTEXTBKCOLOR: Self

pub const GETTOPINDEX: Self

pub const GETCOUNTPERPAGE: Self

pub const GETORIGIN: Self

pub const UPDATE: Self

pub const SETITEMSTATE: Self

pub const GETITEMSTATE: Self

pub const SETITEMCOUNT: Self

pub const SORTITEMS: Self

pub const SETITEMPOSITION32: Self

pub const GETSELECTEDCOUNT: Self

pub const GETITEMSPACING: Self

pub const SETICONSPACING: Self

pub const SETEXTENDEDLISTVIEWSTYLE: Self

pub const GETEXTENDEDLISTVIEWSTYLE: Self

pub const GETSUBITEMRECT: Self

pub const SUBITEMHITTEST: Self

pub const SETCOLUMNORDERARRAY: Self

pub const GETCOLUMNORDERARRAY: Self

pub const SETHOTITEM: Self

pub const GETHOTITEM: Self

pub const SETHOTCURSOR: Self

pub const GETHOTCURSOR: Self

pub const APPROXIMATEVIEWRECT: Self

pub const SETWORKAREAS: Self

pub const GETSELECTIONMARK: Self

pub const SETSELECTIONMARK: Self

pub const GETWORKAREAS: Self

pub const SETHOVERTIME: Self

pub const GETHOVERTIME: Self

pub const GETNUMBEROFWORKAREAS: Self

pub const SETTOOLTIPS: Self

pub const GETITEM: Self

pub const SETITEM: Self

pub const INSERTITEM: Self

pub const GETTOOLTIPS: Self

pub const SORTITEMSEX: Self

pub const FINDITEM: Self

pub const GETSTRINGWIDTH: Self

pub const GETGROUPSTATE: Self

pub const GETFOCUSEDGROUP: Self

pub const GETCOLUMN: Self

pub const SETCOLUMN: Self

pub const INSERTCOLUMN: Self

pub const GETGROUPRECT: Self

pub const GETITEMTEXT: Self

pub const SETITEMTEXT: Self

pub const GETISEARCHSTRING: Self

pub const EDITLABEL: Self

pub const SETBKIMAGE: Self

pub const GETBKIMAGE: Self

pub const SETSELECTEDCOLUMN: Self

pub const SETVIEW: Self

pub const GETVIEW: Self

pub const INSERTGROUP: Self

pub const SETGROUPINFO: Self

pub const GETGROUPINFO: Self

pub const REMOVEGROUP: Self

pub const MOVEGROUP: Self

pub const GETGROUPCOUNT: Self

pub const GETGROUPINFOBYINDEX: Self

pub const MOVEITEMTOGROUP: Self

pub const SETGROUPMETRICS: Self

pub const GETGROUPMETRICS: Self

pub const ENABLEGROUPVIEW: Self

pub const SORTGROUPS: Self

pub const INSERTGROUPSORTED: Self

pub const REMOVEALLGROUPS: Self

pub const HASGROUP: Self

pub const SETTILEVIEWINFO: Self

pub const GETTILEVIEWINFO: Self

pub const SETTILEINFO: Self

pub const GETTILEINFO: Self

pub const SETINSERTMARK: Self

pub const GETINSERTMARK: Self

pub const INSERTMARKHITTEST: Self

pub const GETINSERTMARKRECT: Self

pub const SETINSERTMARKCOLOR: Self

pub const GETINSERTMARKCOLOR: Self

pub const SETINFOTIP: Self

pub const GETSELECTEDCOLUMN: Self

pub const ISGROUPVIEWENABLED: Self

pub const GETOUTLINECOLOR: Self

pub const SETOUTLINECOLOR: Self

pub const CANCELEDITLABEL: Self

pub const MAPINDEXTOID: Self

pub const MAPIDTOINDEX: Self

pub const ISITEMVISIBLE: Self

pub const GETEMPTYTEXT: Self

pub const GETFOOTERRECT: Self

pub const GETFOOTERINFO: Self

pub const GETFOOTERITEMRECT: Self

pub const GETFOOTERITEM: Self

pub const GETITEMINDEXRECT: Self

pub const SETITEMINDEXSTATE: Self

pub const GETNEXTITEMINDEX: Self