winsafe::co

Struct ME

pub struct ME(/* private fields */);

Available on crate feature mf only.

IMFMediaEvent::GetType return value (u32).

Implementations

impl ME

pub const unsafe fn as_mut(&mut self) -> &mut u32

Returns a mutable reference to the underlying raw value.

Safety

Be sure the integer being set is meaningful for the actual type.

pub const unsafe fn from_raw(v: u32) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u32

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl ME

pub const Unknown: Self

pub const Error: Self

pub const ExtendedType: Self

pub const NonFatalError: Self

pub const GenericV1Anchor: Self

pub const SessionUnknown: Self

pub const SessionTopologySet: Self

pub const SessionTopologiesCleared: Self

pub const SessionStarted: Self

pub const SessionPaused: Self

pub const SessionStopped: Self

pub const SessionClosed: Self

pub const SessionEnded: Self

pub const SessionRateChanged: Self

pub const SessionScrubSampleComplete: Self

pub const SessionCapabilitiesChanged: Self

pub const SessionTopologyStatus: Self

pub const SessionNotifyPresentationTime: Self

pub const NewPresentation: Self

pub const LicenseAcquisitionStart: Self

pub const LicenseAcquisitionCompleted: Self

pub const IndividualizationStart: Self

pub const IndividualizationCompleted: Self

pub const EnablerProgress: Self

pub const EnablerCompleted: Self

pub const PolicyError: Self

pub const PolicyReport: Self

pub const BufferingStarted: Self

pub const BufferingStopped: Self

pub const ConnectStart: Self

pub const ConnectEnd: Self

pub const ReconnectStart: Self

pub const ReconnectEnd: Self

pub const RendererEvent: Self

pub const SessionStreamSinkFormatChanged: Self

pub const SessionV1Anchor: Self

pub const SourceUnknown: Self

pub const SourceStarted: Self

pub const StreamStarted: Self

pub const SourceSeeked: Self

pub const StreamSeeked: Self

pub const NewStream: Self

pub const UpdatedStream: Self

pub const SourceStopped: Self

pub const StreamStopped: Self

pub const SourcePaused: Self

pub const StreamPaused: Self

pub const EndOfPresentation: Self

pub const EndOfStream: Self

pub const MediaSample: Self

pub const StreamTick: Self

pub const StreamThinMode: Self

pub const StreamFormatChanged: Self

pub const SourceRateChanged: Self

pub const EndOfPresentationSegment: Self

pub const SourceCharacteristicsChanged: Self

pub const SourceRateChangeRequested: Self

pub const SourceMetadataChanged: Self

pub const SequencerSourceTopologyUpdated: Self

pub const SourceV1Anchor: Self

pub const SinkUnknown: Self

pub const StreamSinkStarted: Self

pub const StreamSinkStopped: Self

pub const StreamSinkPaused: Self

pub const StreamSinkRateChanged: Self

pub const StreamSinkRequestSample: Self

pub const StreamSinkMarker: Self

pub const StreamSinkPrerolled: Self

pub const StreamSinkScrubSampleComplete: Self

pub const StreamSinkFormatChanged: Self

pub const StreamSinkDeviceChanged: Self

pub const QualityNotify: Self

pub const SinkInvalidated: Self

pub const AudioSessionNameChanged: Self

pub const AudioSessionVolumeChanged: Self

pub const AudioSessionDeviceRemoved: Self

pub const AudioSessionServerShutdown: Self

pub const AudioSessionGroupingParamChanged: Self

pub const AudioSessionIconChanged: Self

pub const AudioSessionFormatChanged: Self

pub const AudioSessionDisconnected: Self

pub const AudioSessionExclusiveModeOverride: Self

pub const SinkV1Anchor: Self

pub const CaptureAudioSessionVolumeChanged: Self

pub const CaptureAudioSessionDeviceRemoved: Self

pub const CaptureAudioSessionFormatChanged: Self

pub const CaptureAudioSessionDisconnected: Self

pub const CaptureAudioSessionExclusiveModeOverride: Self

pub const CaptureAudioSessionServerShutdown: Self

pub const SinkV2Anchor: Self

pub const TrustUnknown: Self

pub const PolicyChanged: Self

pub const ContentProtectionMessage: Self

pub const PolicySet: Self

pub const TrustV1Anchor: Self

pub const WMDRMLicenseBackupCompleted: Self

pub const WMDRMLicenseBackupProgress: Self

pub const WMDRMLicenseRestoreCompleted: Self

pub const WMDRMLicenseRestoreProgress: Self

pub const WMDRMLicenseAcquisitionCompleted: Self

pub const WMDRMIndividualizationCompleted: Self

pub const WMDRMIndividualizationProgress: Self

pub const WMDRMProximityCompleted: Self

pub const WMDRMLicenseStoreCleaned: Self

pub const WMDRMRevocationDownloadCompleted: Self

pub const WMDRMV1Anchor: Self

pub const TransformUnknown: Self

pub const TransformNeedInput: Self

pub const TransformHaveOutput: Self

pub const TransformDrainComplete: Self

pub const TransformMarker: Self

pub const TransformInputStreamStateChanged: Self

pub const ByteStreamCharacteristicsChanged: Self

pub const VideoCaptureDeviceRemoved: Self

pub const VideoCaptureDevicePreempted: Self

pub const StreamSinkFormatInvalidated: Self

pub const EncodingParameters: Self

pub const ContentProtectionMetadata: Self

pub const DeviceThermalStateChanged: Self