winsafe::co

Struct SIID

pub struct SIID(/* private fields */);

Available on crate feature shell only.

SHSTOCKICONID enumeration, SHGetStockIconInfo siid (u32).

Implementations

impl SIID

pub const unsafe fn as_mut(&mut self) -> &mut u32

Returns a mutable reference to the underlying raw value.

Safety

Be sure the integer being set is meaningful for the actual type.

pub const unsafe fn from_raw(v: u32) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u32

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl SIID

pub const DOCNOASSOC: Self

pub const DOCASSOC: Self

pub const APPLICATION: Self

pub const FOLDER: Self

pub const FOLDEROPEN: Self

pub const DRIVE525: Self

pub const DRIVE35: Self

pub const DRIVEREMOVE: Self

pub const DRIVEFIXED: Self

pub const DRIVENET: Self

pub const DRIVENETDISABLED: Self

pub const DRIVECD: Self

pub const DRIVERAM: Self

pub const WORLD: Self

pub const SERVER: Self

pub const PRINTER: Self

pub const MYNETWORK: Self

pub const FIND: Self

pub const HELP: Self

pub const SHARE: Self

pub const LINK: Self

pub const SLOWFILE: Self

pub const RECYCLER: Self

pub const RECYCLERFULL: Self

pub const MEDIACDAUDIO: Self

pub const LOCK: Self

pub const AUTOLIST: Self

pub const PRINTERNET: Self

pub const SERVERSHARE: Self

pub const PRINTERFAX: Self

pub const PRINTERFAXNET: Self

pub const PRINTERFILE: Self

pub const STACK: Self

pub const MEDIASVCD: Self

pub const STUFFEDFOLDER: Self

pub const DRIVEUNKNOWN: Self

pub const DRIVEDVD: Self

pub const MEDIADVD: Self

pub const MEDIADVDRAM: Self

pub const MEDIADVDRW: Self

pub const MEDIADVDR: Self

pub const MEDIADVDROM: Self

pub const MEDIACDAUDIOPLUS: Self

pub const MEDIACDRW: Self

pub const MEDIACDR: Self

pub const MEDIACDBURN: Self

pub const MEDIABLANKCD: Self

pub const MEDIACDROM: Self

pub const AUDIOFILES: Self

pub const IMAGEFILES: Self

pub const VIDEOFILES: Self

pub const MIXEDFILES: Self

pub const FOLDERBACK: Self

pub const FOLDERFRONT: Self

pub const SHIELD: Self

pub const WARNING: Self

pub const INFO: Self

pub const ERROR: Self

pub const KEY: Self

pub const SOFTWARE: Self

pub const RENAME: Self

pub const DELETE: Self

pub const MEDIAAUDIODVD: Self

pub const MEDIAMOVIEDVD: Self

pub const MEDIAENHANCEDCD: Self

pub const MEDIAENHANCEDDVD: Self

pub const MEDIAHDDVD: Self

pub const MEDIABLURAY: Self

pub const MEDIAVCD: Self

pub const MEDIADVDPLUSR: Self

pub const MEDIADVDPLUSRW: Self

pub const DESKTOPPC: Self

pub const MOBILEPC: Self

pub const USERS: Self

pub const MEDIASMARTMEDIA: Self

pub const MEDIACOMPACTFLASH: Self

pub const DEVICECELLPHONE: Self

pub const DEVICECAMERA: Self

pub const DEVICEVIDEOCAMERA: Self

pub const DEVICEAUDIOPLAYER: Self

pub const NETWORKCONNECT: Self

pub const INTERNET: Self

pub const ZIPFILE: Self

pub const SETTINGS: Self

pub const DRIVEHDDVD: Self

pub const DRIVEBD: Self

pub const MEDIAHDDVDROM: Self

pub const MEDIAHDDVDR: Self

pub const MEDIAHDDVDRAM: Self

pub const MEDIABDROM: Self

pub const MEDIABDR: Self

pub const MEDIABDRE: Self

pub const CLUSTEREDDRIVE: Self

pub const MAX_ICONS: Self