winsafe::co

Struct SPI

pub struct SPI(/* private fields */);

Available on crate feature user only.

SystemParametersInfo action (u32).

Implementations

impl SPI

pub const unsafe fn as_mut(&mut self) -> &mut u32

Returns a mutable reference to the underlying raw value.

Safety

Be sure the integer being set is meaningful for the actual type.

pub const unsafe fn from_raw(v: u32) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u32

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl SPI

pub const GETBEEP: Self

pub const SETBEEP: Self

pub const GETMOUSE: Self

pub const SETMOUSE: Self

pub const GETBORDER: Self

pub const SETBORDER: Self

pub const GETKEYBOARDSPEED: Self

pub const SETKEYBOARDSPEED: Self

pub const LANGDRIVER: Self

pub const ICONHORIZONTALSPACING: Self

pub const GETSCREENSAVETIMEOUT: Self

pub const SETSCREENSAVETIMEOUT: Self

pub const GETSCREENSAVEACTIVE: Self

pub const SETSCREENSAVEACTIVE: Self

pub const GETGRIDGRANULARITY: Self

pub const SETGRIDGRANULARITY: Self

pub const SETDESKWALLPAPER: Self

pub const SETDESKPATTERN: Self

pub const GETKEYBOARDDELAY: Self

pub const SETKEYBOARDDELAY: Self

pub const ICONVERTICALSPACING: Self

pub const GETICONTITLEWRAP: Self

pub const SETICONTITLEWRAP: Self

pub const GETMENUDROPALIGNMENT: Self

pub const SETMENUDROPALIGNMENT: Self

pub const SETDOUBLECLKWIDTH: Self

pub const SETDOUBLECLKHEIGHT: Self

pub const GETICONTITLELOGFONT: Self

pub const SETDOUBLECLICKTIME: Self

pub const SETMOUSEBUTTONSWAP: Self

pub const SETICONTITLELOGFONT: Self

pub const GETFASTTASKSWITCH: Self

pub const SETFASTTASKSWITCH: Self

pub const SETDRAGFULLWINDOWS: Self

pub const GETDRAGFULLWINDOWS: Self

pub const GETNONCLIENTMETRICS: Self

pub const SETNONCLIENTMETRICS: Self

pub const GETMINIMIZEDMETRICS: Self

pub const SETMINIMIZEDMETRICS: Self

pub const GETICONMETRICS: Self

pub const SETICONMETRICS: Self

pub const SETWORKAREA: Self

pub const GETWORKAREA: Self

pub const SETPENWINDOWS: Self

pub const GETHIGHCONTRAST: Self

pub const SETHIGHCONTRAST: Self

pub const GETKEYBOARDPREF: Self

pub const SETKEYBOARDPREF: Self

pub const GETSCREENREADER: Self

pub const SETSCREENREADER: Self

pub const GETANIMATION: Self

pub const SETANIMATION: Self

pub const GETFONTSMOOTHING: Self

pub const SETFONTSMOOTHING: Self

pub const SETDRAGWIDTH: Self

pub const SETDRAGHEIGHT: Self

pub const SETHANDHELD: Self

pub const GETLOWPOWERTIMEOUT: Self

pub const GETPOWEROFFTIMEOUT: Self

pub const SETLOWPOWERTIMEOUT: Self

pub const SETPOWEROFFTIMEOUT: Self

pub const GETLOWPOWERACTIVE: Self

pub const GETPOWEROFFACTIVE: Self

pub const SETLOWPOWERACTIVE: Self

pub const SETPOWEROFFACTIVE: Self

pub const SETCURSORS: Self

pub const SETICONS: Self

pub const GETDEFAULTINPUTLANG: Self

pub const SETDEFAULTINPUTLANG: Self

pub const SETLANGTOGGLE: Self

pub const GETWINDOWSEXTENSION: Self

pub const SETMOUSETRAILS: Self

pub const GETMOUSETRAILS: Self

pub const SETSCREENSAVERRUNNING: Self

pub const SCREENSAVERRUNNING: Self

pub const GETFILTERKEYS: Self

pub const SETFILTERKEYS: Self

pub const GETTOGGLEKEYS: Self

pub const SETTOGGLEKEYS: Self

pub const GETMOUSEKEYS: Self

pub const SETMOUSEKEYS: Self

pub const GETSHOWSOUNDS: Self

pub const SETSHOWSOUNDS: Self

pub const GETSTICKYKEYS: Self

pub const SETSTICKYKEYS: Self

pub const GETACCESSTIMEOUT: Self

pub const SETACCESSTIMEOUT: Self

pub const GETSERIALKEYS: Self

pub const SETSERIALKEYS: Self

pub const GETSOUNDSENTRY: Self

pub const SETSOUNDSENTRY: Self

pub const GETSNAPTODEFBUTTON: Self

pub const SETSNAPTODEFBUTTON: Self

pub const GETMOUSEHOVERWIDTH: Self

pub const SETMOUSEHOVERWIDTH: Self

pub const GETMOUSEHOVERHEIGHT: Self

pub const SETMOUSEHOVERHEIGHT: Self

pub const GETMOUSEHOVERTIME: Self

pub const SETMOUSEHOVERTIME: Self

pub const GETWHEELSCROLLLINES: Self

pub const SETWHEELSCROLLLINES: Self

pub const GETMENUSHOWDELAY: Self

pub const SETMENUSHOWDELAY: Self

pub const GETWHEELSCROLLCHARS: Self

pub const SETWHEELSCROLLCHARS: Self

pub const GETSHOWIMEUI: Self

pub const SETSHOWIMEUI: Self

pub const GETMOUSESPEED: Self

pub const SETMOUSESPEED: Self

pub const GETSCREENSAVERRUNNING: Self

pub const GETDESKWALLPAPER: Self

pub const GETAUDIODESCRIPTION: Self

pub const SETAUDIODESCRIPTION: Self

pub const GETSCREENSAVESECURE: Self

pub const SETSCREENSAVESECURE: Self

pub const GETHUNGAPPTIMEOUT: Self

pub const SETHUNGAPPTIMEOUT: Self

pub const GETWAITTOKILLTIMEOUT: Self

pub const SETWAITTOKILLTIMEOUT: Self

pub const GETWAITTOKILLSERVICETIMEOUT: Self

pub const SETWAITTOKILLSERVICETIMEOUT: Self

pub const GETMOUSEDOCKTHRESHOLD: Self

pub const SETMOUSEDOCKTHRESHOLD: Self

pub const GETPENDOCKTHRESHOLD: Self

pub const SETPENDOCKTHRESHOLD: Self

pub const GETWINARRANGING: Self

pub const SETWINARRANGING: Self

pub const GETMOUSEDRAGOUTTHRESHOLD: Self

pub const SETMOUSEDRAGOUTTHRESHOLD: Self

pub const GETPENDRAGOUTTHRESHOLD: Self

pub const SETPENDRAGOUTTHRESHOLD: Self

pub const GETMOUSESIDEMOVETHRESHOLD: Self

pub const SETMOUSESIDEMOVETHRESHOLD: Self

pub const GETPENSIDEMOVETHRESHOLD: Self

pub const SETPENSIDEMOVETHRESHOLD: Self

pub const GETDRAGFROMMAXIMIZE: Self

pub const SETDRAGFROMMAXIMIZE: Self

pub const GETSNAPSIZING: Self

pub const SETSNAPSIZING: Self

pub const GETDOCKMOVING: Self

pub const SETDOCKMOVING: Self