winsafe::co

Struct SUBLANG

pub struct SUBLANG(/* private fields */);

Available on crate feature kernel only.

Sublanguage identifier (u16).

Implementations

impl SUBLANG

pub const unsafe fn as_mut(&mut self) -> &mut u16

Returns a mutable reference to the underlying raw value.

Safety

Be sure the integer being set is meaningful for the actual type.

pub const unsafe fn from_raw(v: u16) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u16

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl SUBLANG

pub const NEUTRAL: Self

pub const DEFAULT: Self

pub const SYS_DEFAULT: Self

pub const CUSTOM_DEFAULT: Self

pub const CUSTOM_UNSPECIFIED: Self

pub const UI_CUSTOM_DEFAULT: Self

pub const AFRIKAANS_SOUTH_AFRICA: Self

pub const ALBANIAN_ALBANIA: Self

pub const ALSATIAN_FRANCE: Self

pub const AMHARIC_ETHIOPIA: Self

pub const ARABIC_SAUDI_ARABIA: Self

pub const ARABIC_IRAQ: Self

pub const ARABIC_EGYPT: Self

pub const ARABIC_LIBYA: Self

pub const ARABIC_ALGERIA: Self

pub const ARABIC_MOROCCO: Self

pub const ARABIC_TUNISIA: Self

pub const ARABIC_OMAN: Self

pub const ARABIC_YEMEN: Self

pub const ARABIC_SYRIA: Self

pub const ARABIC_JORDAN: Self

pub const ARABIC_LEBANON: Self

pub const ARABIC_KUWAIT: Self

pub const ARABIC_UAE: Self

pub const ARABIC_BAHRAIN: Self

pub const ARABIC_QATAR: Self

pub const ARMENIAN_ARMENIA: Self

pub const ASSAMESE_INDIA: Self

pub const AZERI_LATIN: Self

pub const AZERI_CYRILLIC: Self

pub const AZERBAIJANI_AZERBAIJAN_LATIN: Self

pub const AZERBAIJANI_AZERBAIJAN_CYRILLIC: Self

pub const BANGLA_INDIA: Self

pub const BANGLA_BANGLADESH: Self

pub const BASHKIR_RUSSIA: Self

pub const BASQUE_BASQUE: Self

pub const BELARUSIAN_BELARUS: Self

pub const BENGALI_INDIA: Self

pub const BENGALI_BANGLADESH: Self

pub const BOSNIAN_BOSNIA_HERZEGOVINA_LATIN: Self

pub const BOSNIAN_BOSNIA_HERZEGOVINA_CYRILLIC: Self

pub const BRETON_FRANCE: Self

pub const BULGARIAN_BULGARIA: Self

pub const CATALAN_CATALAN: Self

pub const CENTRAL_KURDISH_IRAQ: Self

pub const CHEROKEE_CHEROKEE: Self

pub const CHINESE_TRADITIONAL: Self

pub const CHINESE_SIMPLIFIED: Self

pub const CHINESE_HONGKONG: Self

pub const CHINESE_SINGAPORE: Self

pub const CHINESE_MACAU: Self

pub const CORSICAN_FRANCE: Self

pub const CZECH_CZECH_REPUBLIC: Self

pub const CROATIAN_CROATIA: Self

pub const CROATIAN_BOSNIA_HERZEGOVINA_LATIN: Self

pub const DANISH_DENMARK: Self

pub const DARI_AFGHANISTAN: Self

pub const DIVEHI_MALDIVES: Self

pub const DUTCH: Self

pub const DUTCH_BELGIAN: Self

pub const ENGLISH_US: Self

pub const ENGLISH_UK: Self

pub const ENGLISH_AUS: Self

pub const ENGLISH_CAN: Self

pub const ENGLISH_NZ: Self

pub const ENGLISH_EIRE: Self

pub const ENGLISH_SOUTH_AFRICA: Self

pub const ENGLISH_JAMAICA: Self

pub const ENGLISH_CARIBBEAN: Self

pub const ENGLISH_BELIZE: Self

pub const ENGLISH_TRINIDAD: Self

pub const ENGLISH_ZIMBABWE: Self

pub const ENGLISH_PHILIPPINES: Self

pub const ENGLISH_INDIA: Self

pub const ENGLISH_MALAYSIA: Self

pub const ENGLISH_SINGAPORE: Self

pub const ESTONIAN_ESTONIA: Self

pub const FAEROESE_FAROE_ISLANDS: Self

pub const FILIPINO_PHILIPPINES: Self

pub const FINNISH_FINLAND: Self

pub const FRENCH: Self

pub const FRENCH_BELGIAN: Self

pub const FRENCH_CANADIAN: Self

pub const FRENCH_SWISS: Self

pub const FRENCH_LUXEMBOURG: Self

pub const FRENCH_MONACO: Self

pub const FRISIAN_NETHERLANDS: Self

pub const FULAH_SENEGAL: Self

pub const GALICIAN_GALICIAN: Self

pub const GEORGIAN_GEORGIA: Self

pub const GERMAN: Self

pub const GERMAN_SWISS: Self

pub const GERMAN_AUSTRIAN: Self

pub const GERMAN_LUXEMBOURG: Self

pub const GERMAN_LIECHTENSTEIN: Self

pub const GREEK_GREECE: Self

pub const GREENLANDIC_GREENLAND: Self

pub const GUJARATI_INDIA: Self

pub const HAUSA_NIGERIA_LATIN: Self

pub const HAWAIIAN_US: Self

pub const HEBREW_ISRAEL: Self

pub const HINDI_INDIA: Self

pub const HUNGARIAN_HUNGARY: Self

pub const ICELANDIC_ICELAND: Self

pub const IGBO_NIGERIA: Self

pub const INDONESIAN_INDONESIA: Self

pub const INUKTITUT_CANADA: Self

pub const INUKTITUT_CANADA_LATIN: Self

pub const IRISH_IRELAND: Self

pub const ITALIAN: Self

pub const ITALIAN_SWISS: Self

pub const JAPANESE_JAPAN: Self

pub const KANNADA_INDIA: Self

pub const KASHMIRI_SASIA: Self

pub const KASHMIRI_INDIA: Self

pub const KAZAK_KAZAKHSTAN: Self

pub const KHMER_CAMBODIA: Self

pub const KICHE_GUATEMALA: Self

pub const KINYARWANDA_RWANDA: Self

pub const KONKANI_INDIA: Self

pub const KOREAN: Self

pub const KYRGYZ_KYRGYZSTAN: Self

pub const LAO_LAO: Self

pub const LATVIAN_LATVIA: Self

pub const LITHUANIAN: Self

pub const LOWER_SORBIAN_GERMANY: Self

pub const LUXEMBOURGISH_LUXEMBOURG: Self

pub const MACEDONIAN_MACEDONIA: Self

pub const MALAY_MALAYSIA: Self

pub const MALAY_BRUNEI_DARUSSALAM: Self

pub const MALAYALAM_INDIA: Self

pub const MALTESE_MALTA: Self

pub const MAORI_NEW_ZEALAND: Self

pub const MAPUDUNGUN_CHILE: Self

pub const MARATHI_INDIA: Self

pub const MOHAWK_MOHAWK: Self

pub const MONGOLIAN_CYRILLIC_MONGOLIA: Self

pub const MONGOLIAN_PRC: Self

pub const NEPALI_INDIA: Self

pub const NEPALI_NEPAL: Self

pub const NORWEGIAN_BOKMAL: Self

pub const NORWEGIAN_NYNORSK: Self

pub const OCCITAN_FRANCE: Self

pub const ODIA_INDIA: Self

pub const ORIYA_INDIA: Self

pub const PASHTO_AFGHANISTAN: Self

pub const PERSIAN_IRAN: Self

pub const POLISH_POLAND: Self

pub const PORTUGUESE: Self

pub const PORTUGUESE_BRAZILIAN: Self

pub const PULAR_SENEGAL: Self

pub const PUNJABI_INDIA: Self

pub const PUNJABI_PAKISTAN: Self

pub const QUECHUA_BOLIVIA: Self

pub const QUECHUA_ECUADOR: Self

pub const QUECHUA_PERU: Self

pub const ROMANIAN_ROMANIA: Self

pub const ROMANSH_SWITZERLAND: Self

pub const RUSSIAN_RUSSIA: Self

pub const SAKHA_RUSSIA: Self

pub const SAMI_NORTHERN_NORWAY: Self

pub const SAMI_NORTHERN_SWEDEN: Self

pub const SAMI_NORTHERN_FINLAND: Self

pub const SAMI_LULE_NORWAY: Self

pub const SAMI_LULE_SWEDEN: Self

pub const SAMI_SOUTHERN_NORWAY: Self

pub const SAMI_SOUTHERN_SWEDEN: Self

pub const SAMI_SKOLT_FINLAND: Self

pub const SAMI_INARI_FINLAND: Self

pub const SANSKRIT_INDIA: Self

pub const SCOTTISH_GAELIC: Self

pub const SERBIAN_BOSNIA_HERZEGOVINA_LATIN: Self

pub const SERBIAN_BOSNIA_HERZEGOVINA_CYRILLIC: Self

pub const SERBIAN_MONTENEGRO_LATIN: Self

pub const SERBIAN_MONTENEGRO_CYRILLIC: Self

pub const SERBIAN_SERBIA_LATIN: Self

pub const SERBIAN_SERBIA_CYRILLIC: Self

pub const SERBIAN_CROATIA: Self

pub const SERBIAN_LATIN: Self

pub const SERBIAN_CYRILLIC: Self

pub const SINDHI_INDIA: Self

pub const SINDHI_PAKISTAN: Self

pub const SINDHI_AFGHANISTAN: Self

pub const SINHALESE_SRI_LANKA: Self

pub const SOTHO_NORTHERN_SOUTH_AFRICA: Self

pub const SLOVAK_SLOVAKIA: Self

pub const SLOVENIAN_SLOVENIA: Self

pub const SPANISH: Self

pub const SPANISH_MEXICAN: Self

pub const SPANISH_MODERN: Self

pub const SPANISH_GUATEMALA: Self

pub const SPANISH_COSTA_RICA: Self

pub const SPANISH_PANAMA: Self

pub const SPANISH_DOMINICAN_REPUBLIC: Self

pub const SPANISH_VENEZUELA: Self

pub const SPANISH_COLOMBIA: Self

pub const SPANISH_PERU: Self

pub const SPANISH_ARGENTINA: Self

pub const SPANISH_ECUADOR: Self

pub const SPANISH_CHILE: Self

pub const SPANISH_URUGUAY: Self

pub const SPANISH_PARAGUAY: Self

pub const SPANISH_BOLIVIA: Self

pub const SPANISH_EL_SALVADOR: Self

pub const SPANISH_HONDURAS: Self

pub const SPANISH_NICARAGUA: Self

pub const SPANISH_PUERTO_RICO: Self

pub const SPANISH_US: Self

pub const SWAHILI_KENYA: Self

pub const SWEDISH: Self

pub const SWEDISH_FINLAND: Self

pub const SYRIAC_SYRIA: Self

pub const TAJIK_TAJIKISTAN: Self

pub const TAMAZIGHT_ALGERIA_LATIN: Self

pub const TAMAZIGHT_MOROCCO_TIFINAGH: Self

pub const TAMIL_INDIA: Self

pub const TAMIL_SRI_LANKA: Self

pub const TATAR_RUSSIA: Self

pub const TELUGU_INDIA: Self

pub const THAI_THAILAND: Self

pub const TIBETAN_PRC: Self

pub const TIGRIGNA_ERITREA: Self

pub const TIGRINYA_ERITREA: Self

pub const TIGRINYA_ETHIOPIA: Self

pub const TSWANA_BOTSWANA: Self

pub const TSWANA_SOUTH_AFRICA: Self

pub const TURKISH_TURKEY: Self

pub const TURKMEN_TURKMENISTAN: Self

pub const UIGHUR_PRC: Self

pub const UKRAINIAN_UKRAINE: Self

pub const UPPER_SORBIAN_GERMANY: Self

pub const URDU_PAKISTAN: Self

pub const URDU_INDIA: Self

pub const UZBEK_LATIN: Self

pub const UZBEK_CYRILLIC: Self

pub const VALENCIAN_VALENCIA: Self

pub const VIETNAMESE_VIETNAM: Self

pub const WELSH_UNITED_KINGDOM: Self

pub const WOLOF_SENEGAL: Self

pub const XHOSA_SOUTH_AFRICA: Self

pub const YAKUT_RUSSIA: Self

pub const YI_PRC: Self

pub const YORUBA_NIGERIA: Self

pub const ZULU_SOUTH_AFRICA: Self