winsafe::co

Struct TMT

pub struct TMT(/* private fields */);

Available on crate feature uxtheme only.

Theme property identifiers (i32).

Implementations

impl TMT

pub const unsafe fn as_mut(&mut self) -> &mut i32

Returns a mutable reference to the underlying raw value.

Safety

Be sure the integer being set is meaningful for the actual type.

pub const unsafe fn from_raw(v: i32) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> i32

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl TMT

pub const DIBDATA: Self

pub const GLYPHDIBDATA: Self

pub const ENUM: Self

pub const STRING: Self

pub const INT: Self

pub const BOOL: Self

pub const COLOR: Self

pub const MARGINS: Self

pub const FILENAME: Self

pub const SIZE: Self

pub const POSITION: Self

pub const RECT: Self

pub const FONT: Self

pub const INTLIST: Self

pub const HBITMAP: Self

pub const DISKSTREAM: Self

pub const STREAM: Self

pub const BITMAPREF: Self

pub const FLOAT: Self

pub const FLOATLIST: Self

pub const COLORSCHEMES: Self

pub const SIZES: Self

pub const CHARSET: Self

pub const NAME: Self

pub const DISPLAYNAME: Self

pub const TOOLTIP: Self

pub const COMPANY: Self

pub const AUTHOR: Self

pub const COPYRIGHT: Self

pub const URL: Self

pub const VERSION: Self

pub const DESCRIPTION: Self

pub const FIRST_RCSTRING_NAME: Self

pub const LAST_RCSTRING_NAME: Self

pub const CAPTIONFONT: Self

pub const SMALLCAPTIONFONT: Self

pub const MENUFONT: Self

pub const STATUSFONT: Self

pub const MSGBOXFONT: Self

pub const ICONTITLEFONT: Self

pub const HEADING1FONT: Self

pub const HEADING2FONT: Self

pub const BODYFONT: Self

pub const FIRSTFONT: Self

pub const LASTFONT: Self

pub const FLATMENUS: Self

pub const FIRSTBOOL: Self

pub const LASTBOOL: Self

pub const SIZINGBORDERWIDTH: Self

pub const SCROLLBARWIDTH: Self

pub const SCROLLBARHEIGHT: Self

pub const CAPTIONBARWIDTH: Self

pub const CAPTIONBARHEIGHT: Self

pub const SMCAPTIONBARWIDTH: Self

pub const SMCAPTIONBARHEIGHT: Self

pub const MENUBARWIDTH: Self

pub const MENUBARHEIGHT: Self

pub const PADDEDBORDERWIDTH: Self

pub const FIRSTSIZE: Self

pub const LASTSIZE: Self

pub const MINCOLORDEPTH: Self

pub const FIRSTINT: Self

pub const LASTINT: Self

pub const CSSNAME: Self

pub const XMLNAME: Self

pub const LASTUPDATED: Self

pub const ALIAS: Self

pub const FIRSTSTRING: Self

pub const LASTSTRING: Self

pub const SCROLLBAR: Self

pub const BACKGROUND: Self

pub const ACTIVECAPTION: Self

pub const INACTIVECAPTION: Self

pub const MENU: Self

pub const WINDOW: Self

pub const WINDOWFRAME: Self

pub const MENUTEXT: Self

pub const WINDOWTEXT: Self

pub const CAPTIONTEXT: Self

pub const ACTIVEBORDER: Self

pub const INACTIVEBORDER: Self

pub const APPWORKSPACE: Self

pub const HIGHLIGHT: Self

pub const HIGHLIGHTTEXT: Self

pub const BTNFACE: Self

pub const BTNSHADOW: Self

pub const GRAYTEXT: Self

pub const BTNTEXT: Self

pub const INACTIVECAPTIONTEXT: Self

pub const BTNHIGHLIGHT: Self

pub const DKSHADOW3D: Self

pub const LIGHT3D: Self

pub const INFOTEXT: Self

pub const INFOBK: Self

pub const BUTTONALTERNATEFACE: Self

pub const HOTTRACKING: Self

pub const GRADIENTACTIVECAPTION: Self

pub const GRADIENTINACTIVECAPTION: Self

pub const MENUHILIGHT: Self

pub const MENUBAR: Self

pub const FIRSTCOLOR: Self

pub const LASTCOLOR: Self

pub const FROMHUE1: Self

pub const FROMHUE2: Self

pub const FROMHUE3: Self

pub const FROMHUE4: Self

pub const FROMHUE5: Self

pub const TOHUE1: Self

pub const TOHUE2: Self

pub const TOHUE3: Self

pub const TOHUE4: Self

pub const TOHUE5: Self

pub const FROMCOLOR1: Self

pub const FROMCOLOR2: Self

pub const FROMCOLOR3: Self

pub const FROMCOLOR4: Self

pub const FROMCOLOR5: Self

pub const TOCOLOR1: Self

pub const TOCOLOR2: Self

pub const TOCOLOR3: Self

pub const TOCOLOR4: Self

pub const TOCOLOR5: Self

pub const TRANSPARENT: Self

pub const AUTOSIZE: Self

pub const BORDERONLY: Self

pub const COMPOSITED: Self

pub const BGFILL: Self

pub const GLYPHTRANSPARENT: Self

pub const GLYPHONLY: Self

pub const ALWAYSSHOWSIZINGBAR: Self

pub const MIRRORIMAGE: Self

pub const UNIFORMSIZING: Self

pub const INTEGRALSIZING: Self

pub const SOURCEGROW: Self

pub const SOURCESHRINK: Self

pub const DRAWBORDERS: Self

pub const NOETCHEDEFFECT: Self

pub const TEXTAPPLYOVERLAY: Self

pub const TEXTGLOW: Self

pub const TEXTITALIC: Self

pub const COMPOSITEDOPAQUE: Self

pub const LOCALIZEDMIRRORIMAGE: Self

pub const IMAGECOUNT: Self

pub const ALPHALEVEL: Self

pub const BORDERSIZE: Self

pub const ROUNDCORNERWIDTH: Self

pub const ROUNDCORNERHEIGHT: Self

pub const GRADIENTRATIO1: Self

pub const GRADIENTRATIO2: Self

pub const GRADIENTRATIO3: Self

pub const GRADIENTRATIO4: Self

pub const GRADIENTRATIO5: Self

pub const PROGRESSCHUNKSIZE: Self

pub const PROGRESSSPACESIZE: Self

pub const SATURATION: Self

pub const TEXTBORDERSIZE: Self

pub const ALPHATHRESHOLD: Self

pub const WIDTH: Self

pub const HEIGHT: Self

pub const GLYPHINDEX: Self

pub const TRUESIZESTRETCHMARK: Self

pub const MINDPI1: Self

pub const MINDPI2: Self

pub const MINDPI3: Self

pub const MINDPI4: Self

pub const MINDPI5: Self

pub const TEXTGLOWSIZE: Self

pub const FRAMESPERSECOND: Self

pub const PIXELSPERFRAME: Self

pub const ANIMATIONDELAY: Self

pub const GLOWINTENSITY: Self

pub const OPACITY: Self

pub const COLORIZATIONCOLOR: Self

pub const COLORIZATIONOPACITY: Self

pub const MINDPI6: Self

pub const MINDPI7: Self

pub const GLYPHFONT: Self

pub const IMAGEFILE: Self

pub const IMAGEFILE1: Self

pub const IMAGEFILE2: Self

pub const IMAGEFILE3: Self

pub const IMAGEFILE4: Self

pub const IMAGEFILE5: Self

pub const GLYPHIMAGEFILE: Self

pub const IMAGEFILE6: Self

pub const IMAGEFILE7: Self

pub const TEXT: Self

pub const CLASSICVALUE: Self

pub const OFFSET: Self

pub const TEXTSHADOWOFFSET: Self

pub const MINSIZE: Self

pub const MINSIZE1: Self

pub const MINSIZE2: Self

pub const MINSIZE3: Self

pub const MINSIZE4: Self

pub const MINSIZE5: Self

pub const NORMALSIZE: Self

pub const MINSIZE6: Self

pub const MINSIZE7: Self

pub const SIZINGMARGINS: Self

pub const CONTENTMARGINS: Self

pub const CAPTIONMARGINS: Self

pub const BORDERCOLOR: Self

pub const FILLCOLOR: Self

pub const TEXTCOLOR: Self

pub const EDGELIGHTCOLOR: Self

pub const EDGEHIGHLIGHTCOLOR: Self

pub const EDGESHADOWCOLOR: Self

pub const EDGEDKSHADOWCOLOR: Self

pub const EDGEFILLCOLOR: Self

pub const TRANSPARENTCOLOR: Self

pub const GRADIENTCOLOR1: Self

pub const GRADIENTCOLOR2: Self

pub const GRADIENTCOLOR3: Self

pub const GRADIENTCOLOR4: Self

pub const GRADIENTCOLOR5: Self

pub const SHADOWCOLOR: Self

pub const GLOWCOLOR: Self

pub const TEXTBORDERCOLOR: Self

pub const TEXTSHADOWCOLOR: Self

pub const GLYPHTEXTCOLOR: Self

pub const GLYPHTRANSPARENTCOLOR: Self

pub const FILLCOLORHINT: Self

pub const BORDERCOLORHINT: Self

pub const ACCENTCOLORHINT: Self

pub const TEXTCOLORHINT: Self

pub const HEADING1TEXTCOLOR: Self

pub const HEADING2TEXTCOLOR: Self

pub const BODYTEXTCOLOR: Self

pub const BGTYPE: Self

pub const BORDERTYPE: Self

pub const FILLTYPE: Self

pub const SIZINGTYPE: Self

pub const HALIGN: Self

pub const CONTENTALIGNMENT: Self

pub const VALIGN: Self

pub const OFFSETTYPE: Self

pub const ICONEFFECT: Self

pub const TEXTSHADOWTYPE: Self

pub const IMAGELAYOUT: Self

pub const GLYPHTYPE: Self

pub const IMAGESELECTTYPE: Self

pub const GLYPHFONTSIZINGTYPE: Self

pub const TRUESIZESCALINGTYPE: Self

pub const USERPICTURE: Self

pub const DEFAULTPANESIZE: Self

pub const BLENDCOLOR: Self

pub const CUSTOMSPLITRECT: Self

pub const ANIMATIONBUTTONRECT: Self

pub const ANIMATIONDURATION: Self

pub const TRANSITIONDURATIONS: Self

pub const SCALEDBACKGROUND: Self

pub const ATLASIMAGE: Self

pub const ATLASINPUTIMAGE: Self

pub const ATLASRECT: Self