Struct winsafe::co::VS

pub struct VS { /* private fields */ }

This is supported on crate feature uxtheme only.

Visual style parts and states. These are originally two separated constant types, but since they form a hierarchy, one single constant type can represent both part and state.

Implementations

impl VS

pub const fn new(part: i32, state: i32) -> VS

pub const fn part(&self) -> i32

pub const fn state(&self) -> i32

impl VS

pub const AEROWIZARD_TITLEBAR_S_TITLEBAR_ACTIVE: Self

pub const AEROWIZARD_TITLEBAR_S_TITLEBAR_INACTIVE: Self

pub const AEROWIZARD_HEADERAREA_S_HEADERAREA_NOMARGIN: Self

pub const AEROWIZARD_CONTENTAREA_S_CONTENTAREA_NOMARGIN: Self

pub const AEROWIZARD_COMMANDAREA: Self

pub const AEROWIZARD_BUTTON: Self

impl VS

pub const BUTTON_PUSHBUTTON_NORMAL: Self

pub const BUTTON_PUSHBUTTON_HOT: Self

pub const BUTTON_PUSHBUTTON_PRESSED: Self

pub const BUTTON_PUSHBUTTON_DISABLED: Self

pub const BUTTON_PUSHBUTTON_DEFAULTED: Self

pub const BUTTON_PUSHBUTTON_DEFAULTED_ANIMATING: Self

pub const BUTTON_RADIOBUTTON_UNCHECKEDNORMAL: Self

pub const BUTTON_RADIOBUTTON_UNCHECKEDHOT: Self

pub const BUTTON_RADIOBUTTON_UNCHECKEDPRESSED: Self

pub const BUTTON_RADIOBUTTON_UNCHECKEDDISABLED: Self

pub const BUTTON_RADIOBUTTON_CHECKEDNORMAL: Self

pub const BUTTON_RADIOBUTTON_CHECKEDHOT: Self

pub const BUTTON_RADIOBUTTON_CHECKEDPRESSED: Self

pub const BUTTON_RADIOBUTTON_CHECKEDDISABLED: Self

pub const BUTTON_CHECKBOX_UNCHECKEDNORMAL: Self

pub const BUTTON_CHECKBOX_UNCHECKEDHOT: Self

pub const BUTTON_CHECKBOX_UNCHECKEDPRESSED: Self

pub const BUTTON_CHECKBOX_UNCHECKEDDISABLED: Self

pub const BUTTON_CHECKBOX_CHECKEDNORMAL: Self

pub const BUTTON_CHECKBOX_CHECKEDHOT: Self

pub const BUTTON_CHECKBOX_CHECKEDPRESSED: Self

pub const BUTTON_CHECKBOX_CHECKEDDISABLED: Self

pub const BUTTON_CHECKBOX_MIXEDNORMAL: Self

pub const BUTTON_CHECKBOX_MIXEDHOT: Self

pub const BUTTON_CHECKBOX_MIXEDPRESSED: Self

pub const BUTTON_CHECKBOX_MIXEDDISABLED: Self

pub const BUTTON_CHECKBOX_IMPLICITNORMAL: Self

pub const BUTTON_CHECKBOX_IMPLICITHOT: Self

pub const BUTTON_CHECKBOX_IMPLICITPRESSED: Self

pub const BUTTON_CHECKBOX_IMPLICITDISABLED: Self

pub const BUTTON_CHECKBOX_EXCLUDEDNORMAL: Self

pub const BUTTON_CHECKBOX_EXCLUDEDHOT: Self

pub const BUTTON_CHECKBOX_EXCLUDEDPRESSED: Self

pub const BUTTON_CHECKBOX_EXCLUDEDDISABLED: Self

pub const BUTTON_GROUPBOX_NORMAL: Self

pub const BUTTON_GROUPBOX_DISABLED: Self

pub const BUTTON_USERBUTTON: Self

pub const BUTTON_COMMANDLINK_NORMAL: Self

pub const BUTTON_COMMANDLINK_HOT: Self

pub const BUTTON_COMMANDLINK_PRESSED: Self

pub const BUTTON_COMMANDLINK_DISABLED: Self

pub const BUTTON_COMMANDLINK_DEFAULTED: Self

pub const BUTTON_COMMANDLINK_DEFAULTED_ANIMATING: Self

pub const BUTTON_COMMANDLINKGLYPH_NORMAL: Self

pub const BUTTON_COMMANDLINKGLYPH_HOT: Self

pub const BUTTON_COMMANDLINKGLYPH_PRESSED: Self

pub const BUTTON_COMMANDLINKGLYPH_DISABLED: Self

pub const BUTTON_COMMANDLINKGLYPH_DEFAULTED: Self

pub const BUTTON_RADIOBUTTON_HCDISABLED: Self

pub const BUTTON_CHECKBOX_HCDISABLED: Self

pub const BUTTON_GROUPBOX_HCDISABLED: Self

pub const BUTTON_PUSHBUTTONDROPDOWN_NORMAL: Self

pub const BUTTON_PUSHBUTTONDROPDOWN_DISABLED: Self

impl VS

pub const COMBOBOX_DROPDOWNBUTTON_NORMAL: Self

pub const COMBOBOX_DROPDOWNBUTTON_HOT: Self

pub const COMBOBOX_DROPDOWNBUTTON_PRESSED: Self

pub const COMBOBOX_DROPDOWNBUTTON_DISABLED: Self

pub const COMBOBOX_BACKGROUND: Self

pub const COMBOBOX_TRANSPARENTBACKGROUND_NORMAL: Self

pub const COMBOBOX_TRANSPARENTBACKGROUND_HOT: Self

pub const COMBOBOX_TRANSPARENTBACKGROUND_DISABLED: Self

pub const COMBOBOX_TRANSPARENTBACKGROUND_FOCUSED: Self

pub const COMBOBOX_BORDER_NORMAL: Self

pub const COMBOBOX_BORDER_HOT: Self

pub const COMBOBOX_BORDER_FOCUSED: Self

pub const COMBOBOX_BORDER_DISABLED: Self

pub const COMBOBOX_READONLY_NORMAL: Self

pub const COMBOBOX_READONLY_HOT: Self

pub const COMBOBOX_READONLY_PRESSED: Self

pub const COMBOBOX_READONLY_DISABLED: Self

pub const COMBOBOX_DROPDOWNBUTTONRIGHT_NORMAL: Self

pub const COMBOBOX_DROPDOWNBUTTONRIGHT_HOT: Self

pub const COMBOBOX_DROPDOWNBUTTONRIGHT_PRESSED: Self

pub const COMBOBOX_DROPDOWNBUTTONRIGHT_DISABLED: Self

pub const COMBOBOX_DROPDOWNBUTTONLEFT_NORMAL: Self

pub const COMBOBOX_DROPDOWNBUTTONLEFT_HOT: Self

pub const COMBOBOX_DROPDOWNBUTTONLEFT_PRESSED: Self

pub const COMBOBOX_DROPDOWNBUTTONLEFT_DISABLED: Self

pub const COMBOBOX_CUEBANNER_NORMAL: Self

pub const COMBOBOX_CUEBANNER_HOT: Self

pub const COMBOBOX_CUEBANNER_PRESSED: Self

pub const COMBOBOX_CUEBANNER_DISABLED: Self

pub const COMBOBOX_DROPDOWNITEM_NORMAL: Self

pub const COMBOBOX_DROPDOWNITEM_HIGHLIGHTED: Self

impl VS

pub const COMMUNICATIONS_TAB_NORMAL: Self

pub const COMMUNICATIONS_TAB_HOT: Self

pub const COMMUNICATIONS_TAB_SELECTED: Self

impl VS

pub const CONTROLPANEL_NAVIGATIONPANE: Self

pub const CONTROLPANEL_CONTENTPANE: Self

pub const CONTROLPANEL_NAVIGATIONPANELABEL: Self

pub const CONTROLPANEL_CONTENTPANELABEL: Self

pub const CONTROLPANEL_TITLE: Self

pub const CONTROLPANEL_BODYTEXT: Self

pub const CONTROLPANEL_HELPLINK_NORMAL: Self

pub const CONTROLPANEL_HELPLINK_HOT: Self

pub const CONTROLPANEL_HELPLINK_PRESSED: Self

pub const CONTROLPANEL_HELPLINK_DISABLED: Self

pub const CONTROLPANEL_TASKLINK_NORMAL: Self

pub const CONTROLPANEL_TASKLINK_HOT: Self

pub const CONTROLPANEL_TASKLINK_PRESSED: Self

pub const CONTROLPANEL_TASKLINK_DISABLED: Self

pub const CONTROLPANEL_TASKLINK_PAGE: Self

pub const CONTROLPANEL_GROUPTEXT: Self

pub const CONTROLPANEL_CONTENTLINK_NORMAL: Self

pub const CONTROLPANEL_CONTENTLINK_HOT: Self

pub const CONTROLPANEL_CONTENTLINK_PRESSED: Self

pub const CONTROLPANEL_CONTENTLINK_DISABLED: Self

pub const CONTROLPANEL_SECTIONTITLELINK_NORMAL: Self

pub const CONTROLPANEL_SECTIONTITLELINK_HOT: Self

pub const CONTROLPANEL_LARGECOMMANDAREA: Self

pub const CONTROLPANEL_SMALLCOMMANDAREA: Self

pub const CONTROLPANEL_BUTTON: Self

pub const CONTROLPANEL_MESSAGETEXT: Self

pub const CONTROLPANEL_NAVIGATIONPANELINE: Self

pub const CONTROLPANEL_CONTENTPANELINE: Self

pub const CONTROLPANEL_BANNERAREA: Self

pub const CONTROLPANEL_BODYTITLE: Self

impl VS

pub const DATEPICKER_DATETEXT_NORMAL: Self

pub const DATEPICKER_DATETEXT_DISABLED: Self

pub const DATEPICKER_DATETEXT_SELECTED: Self

pub const DATEPICKER_DATEBORDER_NORMAL: Self

pub const DATEPICKER_DATEBORDER_HOT: Self

pub const DATEPICKER_DATEBORDER_FOCUSED: Self

pub const DATEPICKER_DATEBORDER_DISABLED: Self

pub const DATEPICKER_SHOWCALENDARBUTTONRIGHT_NORMAL: Self

pub const DATEPICKER_SHOWCALENDARBUTTONRIGHT_HOT: Self

pub const DATEPICKER_SHOWCALENDARBUTTONRIGHT_PRESSED: Self

pub const DATEPICKER_SHOWCALENDARBUTTONRIGHT_DISABLED: Self

impl VS

pub const DRAGDROP_COPY_HIGHLIGHT: Self

pub const DRAGDROP_COPY_NOHIGHLIGHT: Self

pub const DRAGDROP_MOVE_HIGHLIGHT: Self

pub const DRAGDROP_MOVE_NOHIGHLIGHT: Self

pub const DRAGDROP_UPDATEMETADATA_HIGHLIGHT: Self

pub const DRAGDROP_UPDATEMETADATA_NOHIGHLIGHT: Self

pub const DRAGDROP_CREATELINK_HIGHLIGHT: Self

pub const DRAGDROP_CREATELINK_NOHIGHLIGHT: Self

pub const DRAGDROP_WARNING_HIGHLIGHT: Self

pub const DRAGDROP_WARNING_NOHIGHLIGHT: Self

pub const DRAGDROP_NONE_HIGHLIGHT: Self

pub const DRAGDROP_NONE_NOHIGHLIGHT: Self

pub const DRAGDROP_IMAGEBG: Self

pub const DRAGDROP_TEXTBG: Self

impl VS

pub const EDIT_EDITTEXT_NORMAL: Self

pub const EDIT_EDITTEXT_HOT: Self

pub const EDIT_EDITTEXT_SELECTED: Self

pub const EDIT_EDITTEXT_DISABLED: Self

pub const EDIT_EDITTEXT_FOCUSED: Self

pub const EDIT_EDITTEXT_READONLY: Self

pub const EDIT_EDITTEXT_ASSIST: Self

pub const EDIT_EDITTEXT_CUEBANNER: Self

pub const EDIT_CARET: Self

pub const EDIT_BACKGROUND_NORMAL: Self

pub const EDIT_BACKGROUND_HOT: Self

pub const EDIT_BACKGROUND_DISABLED: Self

pub const EDIT_BACKGROUND_FOCUSED: Self

pub const EDIT_BACKGROUND_READONLY: Self

pub const EDIT_BACKGROUND_ASSIST: Self

pub const EDIT_PASSWORD: Self

pub const EDIT_BACKGROUNDWITHBORDER_NORMAL: Self

pub const EDIT_BACKGROUNDWITHBORDER_HOT: Self

pub const EDIT_BACKGROUNDWITHBORDER_DISABLED: Self

pub const EDIT_BACKGROUNDWITHBORDER_FOCUSED: Self

pub const EDIT_EDITBORDER_NOSCROLL_NORMAL: Self

pub const EDIT_EDITBORDER_NOSCROLL_HOT: Self

pub const EDIT_EDITBORDER_NOSCROLL_FOCUSED: Self

pub const EDIT_EDITBORDER_NOSCROLL_DISABLED: Self

pub const EDIT_EDITBORDER_HSCROLL_NORMAL: Self

pub const EDIT_EDITBORDER_HSCROLL_HOT: Self

pub const EDIT_EDITBORDER_HSCROLL_FOCUSED: Self

pub const EDIT_EDITBORDER_HSCROLL_DISABLED: Self

pub const EDIT_EDITBORDER_VSCROLL_NORMAL: Self

pub const EDIT_EDITBORDER_VSCROLL_HOT: Self

pub const EDIT_EDITBORDER_VSCROLL_FOCUSED: Self

pub const EDIT_EDITBORDER_VSCROLL_DISABLED: Self

pub const EDIT_EDITBORDER_HVSCROLL_NORMAL: Self

pub const EDIT_EDITBORDER_HVSCROLL_HOT: Self

pub const EDIT_EDITBORDER_HVSCROLL_FOCUSED: Self

pub const EDIT_EDITBORDER_HVSCROLL_DISABLED: Self

impl VS

pub const EXPLORERBAR_HEADERBACKGROUND: Self

pub const EXPLORERBAR_HEADERCLOSE_NORMAL: Self

pub const EXPLORERBAR_HEADERCLOSE_HOT: Self

pub const EXPLORERBAR_HEADERCLOSE_PRESSED: Self

pub const EXPLORERBAR_HEADERPIN_NORMAL: Self

pub const EXPLORERBAR_HEADERPIN_HOT: Self

pub const EXPLORERBAR_HEADERPIN_PRESSED: Self

pub const EXPLORERBAR_HEADERPIN_SELECTEDNORMAL: Self

pub const EXPLORERBAR_HEADERPIN_SELECTEDHOT: Self

pub const EXPLORERBAR_HEADERPIN_SELECTEDPRESSED: Self

pub const EXPLORERBAR_IEBARMENU_NORMAL: Self

pub const EXPLORERBAR_IEBARMENU_HOT: Self

pub const EXPLORERBAR_IEBARMENU_PRESSED: Self

pub const EXPLORERBAR_NORMALGROUPBACKGROUND: Self

pub const EXPLORERBAR_NORMALGROUPCOLLAPSE_NORMAL: Self

pub const EXPLORERBAR_NORMALGROUPCOLLAPSE_HOT: Self

pub const EXPLORERBAR_NORMALGROUPCOLLAPSE_PRESSED: Self

pub const EXPLORERBAR_NORMALGROUPEXPAND_NORMAL: Self

pub const EXPLORERBAR_NORMALGROUPEXPAND_HOT: Self

pub const EXPLORERBAR_NORMALGROUPEXPAND_PRESSED: Self

pub const EXPLORERBAR_NORMALGROUPHEAD: Self

pub const EXPLORERBAR_SPECIALGROUPBACKGROUND: Self

pub const EXPLORERBAR_SPECIALGROUPCOLLAPSE_NORMAL: Self

pub const EXPLORERBAR_SPECIALGROUPCOLLAPSE_HOT: Self

pub const EXPLORERBAR_SPECIALGROUPCOLLAPSE_PRESSED: Self

pub const EXPLORERBAR_SPECIALGROUPEXPAND_NORMAL: Self

pub const EXPLORERBAR_SPECIALGROUPEXPAND_HOT: Self

pub const EXPLORERBAR_SPECIALGROUPEXPAND_PRESSED: Self

pub const EXPLORERBAR_SPECIALGROUPHEAD: Self

impl VS

pub const FLYOUT_HEADER: Self

pub const FLYOUT_BODY_NORMAL: Self

pub const FLYOUT_BODY_EMPHASIZED: Self

pub const FLYOUT_LABEL_NORMAL: Self

pub const FLYOUT_LABEL_SELECTED: Self

pub const FLYOUT_LABEL_EMPHASIZED: Self

pub const FLYOUT_LABEL_DISABLED: Self

pub const FLYOUT_LINK_NORMAL: Self

pub const FLYOUT_LINK_HOVER: Self

pub const FLYOUT_DIVIDER: Self

pub const FLYOUT_WINDOW: Self

pub const FLYOUT_LINKAREA: Self

pub const FLYOUT_LINKHEADER_NORMAL: Self

pub const FLYOUT_LINKHEADER_HOVER: Self

impl VS

pub const HEADER_HEADERITEM_NORMAL: Self

pub const HEADER_HEADERITEM_HOT: Self

pub const HEADER_HEADERITEM_PRESSED: Self

pub const HEADER_HEADERITEM_SORTEDNORMAL: Self

pub const HEADER_HEADERITEM_SORTEDHOT: Self

pub const HEADER_HEADERITEM_SORTEDPRESSED: Self

pub const HEADER_HEADERITEM_ICONNORMAL: Self

pub const HEADER_HEADERITEM_ICONHOT: Self

pub const HEADER_HEADERITEM_ICONPRESSED: Self

pub const HEADER_HEADERITEM_ICONSORTEDNORMAL: Self

pub const HEADER_HEADERITEM_ICONSORTEDHOT: Self

pub const HEADER_HEADERITEM_ICONSORTEDPRESSED: Self

pub const HEADER_HEADERITEMLEFT_NORMAL: Self

pub const HEADER_HEADERITEMLEFT_HOT: Self

pub const HEADER_HEADERITEMLEFT_PRESSED: Self

pub const HEADER_HEADERITEMRIGHT_NORMAL: Self

pub const HEADER_HEADERITEMRIGHT_HOT: Self

pub const HEADER_HEADERITEMRIGHT_PRESSED: Self

pub const HEADER_HEADERSORTARROW_SORTEDUP: Self

pub const HEADER_HEADERSORTARROW_SORTEDDOWN: Self

pub const HEADER_HEADERDROPDOWN_NORMAL: Self

pub const HEADER_HEADERDROPDOWN_SOFTHOT: Self

pub const HEADER_HEADERDROPDOWN_HOT: Self

pub const HEADER_HEADERDROPDOWNFILTER_NORMAL: Self

pub const HEADER_HEADERDROPDOWNFILTER_SOFTHOT: Self

pub const HEADER_HEADERDROPDOWNFILTER_HOT: Self

pub const HEADER_HEADEROVERFLOW_NORMAL: Self

pub const HEADER_HEADEROVERFLOW_HOT: Self

impl VS

pub const LISTBOX_BORDER_HSCROLL_NORMAL: Self

pub const LISTBOX_BORDER_HSCROLL_FOCUSED: Self

pub const LISTBOX_BORDER_HSCROLL_HOT: Self

pub const LISTBOX_BORDER_HSCROLL_DISABLED: Self

pub const LISTBOX_BORDER_HVSCROLL_NORMAL: Self

pub const LISTBOX_BORDER_HVSCROLL_FOCUSED: Self

pub const LISTBOX_BORDER_HVSCROLL_HOT: Self

pub const LISTBOX_BORDER_HVSCROLL_DISABLED: Self

pub const LISTBOX_BORDER_NOSCROLL_NORMAL: Self

pub const LISTBOX_BORDER_NOSCROLL_FOCUSED: Self

pub const LISTBOX_BORDER_NOSCROLL_HOT: Self

pub const LISTBOX_BORDER_NOSCROLL_DISABLED: Self

pub const LISTBOX_BORDER_VSCROLL_NORMAL: Self

pub const LISTBOX_BORDER_VSCROLL_FOCUSED: Self

pub const LISTBOX_BORDER_VSCROLL_HOT: Self

pub const LISTBOX_BORDER_VSCROLL_DISABLED: Self

pub const LISTBOX_ITEM_HOT: Self

pub const LISTBOX_ITEM_HOTSELECTED: Self

pub const LISTBOX_ITEM_SELECTED: Self

pub const LISTBOX_ITEM_SELECTEDNOTFOCUS: Self

impl VS

pub const LISTVIEW_LISTITEM_NORMAL: Self

pub const LISTVIEW_LISTITEM_HOT: Self

pub const LISTVIEW_LISTITEM_SELECTED: Self

pub const LISTVIEW_LISTITEM_DISABLED: Self

pub const LISTVIEW_LISTITEM_SELECTEDNOTFOCUS: Self

pub const LISTVIEW_LISTITEM_HOTSELECTED: Self

pub const LISTVIEW_LISTGROUP: Self

pub const LISTVIEW_LISTDETAIL: Self

pub const LISTVIEW_LISTSORTEDDETAIL: Self

pub const LISTVIEW_EMPTYTEXT: Self

pub const LISTVIEW_GROUPHEADER_OPEN: Self

pub const LISTVIEW_GROUPHEADER_OPENHOT: Self

pub const LISTVIEW_GROUPHEADER_OPENSELECTED: Self

pub const LISTVIEW_GROUPHEADER_OPENSELECTEDHOT: Self

pub const LISTVIEW_GROUPHEADER_OPENSELECTEDNOTFOCUSED: Self

pub const LISTVIEW_GROUPHEADER_OPENSELECTEDNOTFOCUSEDHOT: Self

pub const LISTVIEW_GROUPHEADER_OPENMIXEDSELECTION: Self

pub const LISTVIEW_GROUPHEADER_OPENMIXEDSELECTIONHOT: Self

pub const LISTVIEW_GROUPHEADER_CLOSE: Self

pub const LISTVIEW_GROUPHEADER_CLOSEHOT: Self

pub const LISTVIEW_GROUPHEADER_CLOSESELECTED: Self

pub const LISTVIEW_GROUPHEADER_CLOSESELECTEDHOT: Self

pub const LISTVIEW_GROUPHEADER_CLOSESELECTEDNOTFOCUSED: Self

pub const LISTVIEW_GROUPHEADER_CLOSESELECTEDNOTFOCUSEDHOT: Self

pub const LISTVIEW_GROUPHEADER_CLOSEMIXEDSELECTION: Self

pub const LISTVIEW_GROUPHEADER_CLOSEMIXEDSELECTIONHOT: Self

pub const LISTVIEW_GROUPHEADERLINE_OPEN: Self

pub const LISTVIEW_GROUPHEADERLINE_OPENHOT: Self

pub const LISTVIEW_GROUPHEADERLINE_OPENSELECTED: Self

pub const LISTVIEW_GROUPHEADERLINE_OPENSELECTEDHOT: Self

pub const LISTVIEW_GROUPHEADERLINE_OPENSELECTEDNOTFOCUSED: Self

pub const LISTVIEW_GROUPHEADERLINE_OPENSELECTEDNOTFOCUSEDHOT: Self

pub const LISTVIEW_GROUPHEADERLINE_OPENMIXEDSELECTION: Self

pub const LISTVIEW_GROUPHEADERLINE_OPENMIXEDSELECTIONHOT: Self

pub const LISTVIEW_GROUPHEADERLINE_CLOSE: Self

pub const LISTVIEW_GROUPHEADERLINE_CLOSEHOT: Self

pub const LISTVIEW_GROUPHEADERLINE_CLOSESELECTED: Self

pub const LISTVIEW_GROUPHEADERLINE_CLOSESELECTEDHOT: Self

pub const LISTVIEW_GROUPHEADERLINE_CLOSESELECTEDNOTFOCUSED: Self

pub const LISTVIEW_GROUPHEADERLINE_CLOSESELECTEDNOTFOCUSEDHOT: Self

pub const LISTVIEW_GROUPHEADERLINE_CLOSEMIXEDSELECTION: Self

pub const LISTVIEW_GROUPHEADERLINE_CLOSEMIXEDSELECTIONHOT: Self

pub const LISTVIEW_EXPANDBUTTON_NORMAL: Self

pub const LISTVIEW_EXPANDBUTTON_HOVER: Self

pub const LISTVIEW_EXPANDBUTTON_PUSHED: Self

pub const LISTVIEW_COLLAPSEBUTTON_NORMAL: Self

pub const LISTVIEW_COLLAPSEBUTTON_HOVER: Self

pub const LISTVIEW_COLLAPSEBUTTON_PUSHED: Self

pub const LISTVIEW_COLUMNDETAIL: Self

impl VS

pub const MENU_MENUITEM_TMSCHEMA: Self

pub const MENU_MENUDROPDOWN_TMSCHEMA: Self

pub const MENU_MENUBARITEM_TMSCHEMA: Self

pub const MENU_MENUBARDROPDOWN_TMSCHEMA: Self

pub const MENU_CHEVRON_TMSCHEMA: Self

pub const MENU_SEPARATOR_TMSCHEMA: Self

pub const MENU_BARBACKGROUND_ACTIVE: Self

pub const MENU_BARBACKGROUND_INACTIVE: Self

pub const MENU_BARITEM_NORMAL: Self

pub const MENU_BARITEM_HOT: Self

pub const MENU_BARITEM_PUSHED: Self

pub const MENU_BARITEM_DISABLED: Self

pub const MENU_BARITEM_DISABLEDHOT: Self

pub const MENU_BARITEM_DISABLEDPUSHED: Self

pub const MENU_POPUPBACKGROUND: Self

pub const MENU_POPUPBORDERS: Self

pub const MENU_POPUPCHECK_CHECKMARKNORMAL: Self

pub const MENU_POPUPCHECK_CHECKMARKDISABLED: Self

pub const MENU_POPUPCHECK_BULLETNORMAL: Self

pub const MENU_POPUPCHECK_BULLETDISABLED: Self

pub const MENU_POPUPCHECKBACKGROUND_DISABLED: Self

pub const MENU_POPUPCHECKBACKGROUND_NORMAL: Self

pub const MENU_POPUPCHECKBACKGROUND_BITMAP: Self

pub const MENU_POPUPGUTTER: Self

pub const MENU_POPUPITEM_NORMAL: Self

pub const MENU_POPUPITEM_HOT: Self

pub const MENU_POPUPITEM_DISABLED: Self

pub const MENU_POPUPITEM_DISABLEDHOT: Self

pub const MENU_POPUPSEPARATOR: Self

pub const MENU_POPUPSUBMENU_NORMAL: Self

pub const MENU_POPUPSUBMENU_DISABLED: Self

pub const MENU_SYSTEMCLOSE_NORMAL: Self

pub const MENU_SYSTEMCLOSE_DISABLED: Self

pub const MENU_SYSTEMMAXIMIZE_NORMAL: Self

pub const MENU_SYSTEMMAXIMIZE_DISABLED: Self

pub const MENU_SYSTEMMINIMIZE_NORMAL: Self

pub const MENU_SYSTEMMINIMIZE_DISABLED: Self

pub const MENU_SYSTEMRESTORE_NORMAL: Self

pub const MENU_SYSTEMRESTORE_DISABLED: Self

impl VS

pub const NAVIGATION_BACKBUTTON_NORMAL: Self

pub const NAVIGATION_BACKBUTTON_HOT: Self

pub const NAVIGATION_BACKBUTTON_PRESSED: Self

pub const NAVIGATION_BACKBUTTON_DISABLED: Self

pub const NAVIGATION_FORWARDBUTTON_NORMAL: Self

pub const NAVIGATION_FORWARDBUTTON_HOT: Self

pub const NAVIGATION_FORWARDBUTTON_PRESSED: Self

pub const NAVIGATION_FORWARDBUTTON_DISABLED: Self

pub const NAVIGATION_MENUBUTTON_NORMAL: Self

pub const NAVIGATION_MENUBUTTON_HOT: Self

pub const NAVIGATION_MENUBUTTON_PRESSED: Self

pub const NAVIGATION_MENUBUTTON_DISABLED: Self

impl VS

pub const PROGRESS_BAR: Self

pub const PROGRESS_BARVERT: Self

pub const PROGRESS_CHUNK: Self

pub const PROGRESS_CHUNKVERT: Self

pub const PROGRESS_FILL_NORMAL: Self

pub const PROGRESS_FILL_ERROR: Self

pub const PROGRESS_FILL_PAUSED: Self

pub const PROGRESS_FILL_PARTIAL: Self

pub const PROGRESS_FILLVERT_NORMAL: Self

pub const PROGRESS_FILLVERT_ERROR: Self

pub const PROGRESS_FILLVERT_PAUSED: Self

pub const PROGRESS_FILLVERT_PARTIAL: Self

pub const PROGRESS_PULSEOVERLAY: Self

pub const PROGRESS_MOVEOVERLAY: Self

pub const PROGRESS_PULSEOVERLAYVERT: Self

pub const PROGRESS_MOVEOVERLAYVERT: Self

pub const PROGRESS_TRANSPARENTBAR_NORMAL: Self

pub const PROGRESS_TRANSPARENTBAR_PARTIAL: Self

pub const PROGRESS_TRANSPARENTBARVERT_NORMAL: Self

pub const PROGRESS_TRANSPARENTBARVERT_PARTIAL: Self

impl VS

pub const REBAR_GRIPPER: Self

pub const REBAR_GRIPPERVERT: Self

pub const REBAR_BAND: Self

pub const REBAR_CHEVRON_NORMAL: Self

pub const REBAR_CHEVRON_HOT: Self

pub const REBAR_CHEVRON_PRESSED: Self

pub const REBAR_CHEVRONVERT_NORMAL: Self

pub const REBAR_CHEVRONVERT_HOT: Self

pub const REBAR_CHEVRONVERT_PRESSED: Self

pub const REBAR_BACKGROUND: Self

pub const REBAR_SPLITTER_NORMAL: Self

pub const REBAR_SPLITTER_HOT: Self

pub const REBAR_SPLITTER_PRESSED: Self

pub const REBAR_SPLITTERVERT_NORMAL: Self

pub const REBAR_SPLITTERVERT_HOT: Self

pub const REBAR_SPLITTERVERT_PRESSED: Self

impl VS

pub const SCROLLBAR_ARROWBTN_UPNORMAL: Self

pub const SCROLLBAR_ARROWBTN_UPHOT: Self

pub const SCROLLBAR_ARROWBTN_UPPRESSED: Self

pub const SCROLLBAR_ARROWBTN_UPDISABLED: Self

pub const SCROLLBAR_ARROWBTN_DOWNNORMAL: Self

pub const SCROLLBAR_ARROWBTN_DOWNHOT: Self

pub const SCROLLBAR_ARROWBTN_DOWNPRESSED: Self

pub const SCROLLBAR_ARROWBTN_DOWNDISABLED: Self

pub const SCROLLBAR_ARROWBTN_LEFTNORMAL: Self

pub const SCROLLBAR_ARROWBTN_LEFTHOT: Self

pub const SCROLLBAR_ARROWBTN_LEFTPRESSED: Self

pub const SCROLLBAR_ARROWBTN_LEFTDISABLED: Self

pub const SCROLLBAR_ARROWBTN_RIGHTNORMAL: Self

pub const SCROLLBAR_ARROWBTN_RIGHTHOT: Self

pub const SCROLLBAR_ARROWBTN_RIGHTPRESSED: Self

pub const SCROLLBAR_ARROWBTN_RIGHTDISABLED: Self

pub const SCROLLBAR_ARROWBTN_UPHOVER: Self

pub const SCROLLBAR_ARROWBTN_DOWNHOVER: Self

pub const SCROLLBAR_ARROWBTN_LEFTHOVER: Self

pub const SCROLLBAR_ARROWBTN_RIGHTHOVER: Self

pub const SCROLLBAR_THUMBBTNHORZ_NORMAL: Self

pub const SCROLLBAR_THUMBBTNHORZ_HOT: Self

pub const SCROLLBAR_THUMBBTNHORZ_PRESSED: Self

pub const SCROLLBAR_THUMBBTNHORZ_DISABLED: Self

pub const SCROLLBAR_THUMBBTNHORZ_HOVER: Self

pub const SCROLLBAR_THUMBBTNVERT_NORMAL: Self

pub const SCROLLBAR_THUMBBTNVERT_HOT: Self

pub const SCROLLBAR_THUMBBTNVERT_PRESSED: Self

pub const SCROLLBAR_THUMBBTNVERT_DISABLED: Self

pub const SCROLLBAR_THUMBBTNVERT_HOVER: Self

pub const SCROLLBAR_LOWERTRACKHORZ_NORMAL: Self

pub const SCROLLBAR_LOWERTRACKHORZ_HOT: Self

pub const SCROLLBAR_LOWERTRACKHORZ_PRESSED: Self

pub const SCROLLBAR_LOWERTRACKHORZ_DISABLED: Self

pub const SCROLLBAR_LOWERTRACKHORZ_HOVER: Self

pub const SCROLLBAR_UPPERTRACKHORZ_NORMAL: Self

pub const SCROLLBAR_UPPERTRACKHORZ_HOT: Self

pub const SCROLLBAR_UPPERTRACKHORZ_PRESSED: Self

pub const SCROLLBAR_UPPERTRACKHORZ_DISABLED: Self

pub const SCROLLBAR_UPPERTRACKHORZ_HOVER: Self

pub const SCROLLBAR_LOWERTRACKVERT_NORMAL: Self

pub const SCROLLBAR_LOWERTRACKVERT_HOT: Self

pub const SCROLLBAR_LOWERTRACKVERT_PRESSED: Self

pub const SCROLLBAR_LOWERTRACKVERT_DISABLED: Self

pub const SCROLLBAR_LOWERTRACKVERT_HOVER: Self

pub const SCROLLBAR_UPPERTRACKVERT_NORMAL: Self

pub const SCROLLBAR_UPPERTRACKVERT_HOT: Self

pub const SCROLLBAR_UPPERTRACKVERT_PRESSED: Self

pub const SCROLLBAR_UPPERTRACKVERT_DISABLED: Self

pub const SCROLLBAR_UPPERTRACKVERT_HOVER: Self

pub const SCROLLBAR_GRIPPERHORZ_NORMAL: Self

pub const SCROLLBAR_GRIPPERHORZ_HOT: Self

pub const SCROLLBAR_GRIPPERHORZ_PRESSED: Self

pub const SCROLLBAR_GRIPPERHORZ_DISABLED: Self

pub const SCROLLBAR_GRIPPERHORZ_HOVER: Self

pub const SCROLLBAR_GRIPPERVERT_NORMAL: Self

pub const SCROLLBAR_GRIPPERVERT_HOT: Self

pub const SCROLLBAR_GRIPPERVERT_PRESSED: Self

pub const SCROLLBAR_GRIPPERVERT_DISABLED: Self

pub const SCROLLBAR_GRIPPERVERT_HOVER: Self

pub const SCROLLBAR_SIZEBOX_RIGHTALIGN: Self

pub const SCROLLBAR_SIZEBOX_LEFTALIGN: Self

pub const SCROLLBAR_SIZEBOX_TOPRIGHTALIGN: Self

pub const SCROLLBAR_SIZEBOX_TOPLEFTALIGN: Self

pub const SCROLLBAR_SIZEBOX_HALFBOTTOMRIGHTALIGN: Self

pub const SCROLLBAR_SIZEBOX_HALFBOTTOMLEFTALIGN: Self

pub const SCROLLBAR_SIZEBOX_HALFTOPRIGHTALIGN: Self

pub const SCROLLBAR_SIZEBOX_HALFTOPLEFTALIGN: Self

pub const SCROLLBAR_SIZEBOXBKGND_NORMAL: Self

pub const SCROLLBAR_SIZEBOXBKGND_HOT: Self

pub const SCROLLBAR_SIZEBOXBKGND_PRESSED: Self

pub const SCROLLBAR_SIZEBOXBKGND_DISABLED: Self

pub const SCROLLBAR_SIZEBOXBKGND_HOVER: Self

impl VS

pub const SPIN_UP_NORMAL: Self

pub const SPIN_UP_HOT: Self

pub const SPIN_UP_PRESSED: Self

pub const SPIN_UP_DISABLED: Self

pub const SPIN_DOWN_NORMAL: Self

pub const SPIN_DOWN_HOT: Self

pub const SPIN_DOWN_PRESSED: Self

pub const SPIN_DOWN_DISABLED: Self

pub const SPIN_UPHORZ_NORMAL: Self

pub const SPIN_UPHORZ_HOT: Self

pub const SPIN_UPHORZ_PRESSED: Self

pub const SPIN_UPHORZ_DISABLED: Self

pub const SPIN_DOWNHORZ_NORMAL: Self

pub const SPIN_DOWNHORZ_HOT: Self

pub const SPIN_DOWNHORZ_PRESSED: Self

pub const SPIN_DOWNHORZ_DISABLED: Self

impl VS

pub const STATUS_PANE: Self

pub const STATUS_GRIPPERPANE: Self

pub const STATUS_GRIPPER: Self

impl VS

pub const TAB_TABITEM_NORMAL: Self

pub const TAB_TABITEM_HOT: Self

pub const TAB_TABITEM_SELECTED: Self

pub const TAB_TABITEM_DISABLED: Self

pub const TAB_TABITEM_FOCUSED: Self

pub const TAB_TABITEMLEFTEDGE_NORMAL: Self

pub const TAB_TABITEMLEFTEDGE_HOT: Self

pub const TAB_TABITEMLEFTEDGE_SELECTED: Self

pub const TAB_TABITEMLEFTEDGE_DISABLED: Self

pub const TAB_TABITEMLEFTEDGE_FOCUSED: Self

pub const TAB_TABITEMRIGHTEDGE_NORMAL: Self

pub const TAB_TABITEMRIGHTEDGE_HOT: Self

pub const TAB_TABITEMRIGHTEDGE_SELECTED: Self

pub const TAB_TABITEMRIGHTEDGE_DISABLED: Self

pub const TAB_TABITEMRIGHTEDGE_FOCUSED: Self

pub const TAB_TABITEMBOTHEDGE_NORMAL: Self

pub const TAB_TABITEMBOTHEDGE_HOT: Self

pub const TAB_TABITEMBOTHEDGE_SELECTED: Self

pub const TAB_TABITEMBOTHEDGE_DISABLED: Self

pub const TAB_TABITEMBOTHEDGE_FOCUSED: Self

pub const TAB_TOPTABITEM_NORMAL: Self

pub const TAB_TOPTABITEM_HOT: Self

pub const TAB_TOPTABITEM_SELECTED: Self

pub const TAB_TOPTABITEM_DISABLED: Self

pub const TAB_TOPTABITEM_FOCUSED: Self

pub const TAB_TOPTABITEMLEFTEDGE_NORMAL: Self

pub const TAB_TOPTABITEMLEFTEDGE_HOT: Self

pub const TAB_TOPTABITEMLEFTEDGE_SELECTED: Self

pub const TAB_TOPTABITEMLEFTEDGE_DISABLED: Self

pub const TAB_TOPTABITEMLEFTEDGE_FOCUSED: Self

pub const TAB_TOPTABITEMRIGHTEDGE_NORMAL: Self

pub const TAB_TOPTABITEMRIGHTEDGE_HOT: Self

pub const TAB_TOPTABITEMRIGHTEDGE_SELECTED: Self

pub const TAB_TOPTABITEMRIGHTEDGE_DISABLED: Self

pub const TAB_TOPTABITEMRIGHTEDGE_FOCUSED: Self

pub const TAB_TOPTABITEMBOTHEDGE_NORMAL: Self

pub const TAB_TOPTABITEMBOTHEDGE_HOT: Self

pub const TAB_TOPTABITEMBOTHEDGE_SELECTED: Self

pub const TAB_TOPTABITEMBOTHEDGE_DISABLED: Self

pub const TAB_TOPTABITEMBOTHEDGE_FOCUSED: Self

pub const TAB_PANE: Self

pub const TAB_BODY: Self

pub const TAB_AEROWIZARDBODY: Self

impl VS

pub const TASKDIALOG_PRIMARYPANEL: Self

pub const TASKDIALOG_MAININSTRUCTIONPANE: Self

pub const TASKDIALOG_MAINICON: Self

pub const TASKDIALOG_CONTENTPANE_STANDALONE: Self

pub const TASKDIALOG_CONTENTICON: Self

pub const TASKDIALOG_EXPANDEDCONTENT: Self

pub const TASKDIALOG_COMMANDLINKPANE: Self

pub const TASKDIALOG_SECONDARYPANEL: Self

pub const TASKDIALOG_CONTROLPANE: Self

pub const TASKDIALOG_BUTTONSECTION: Self

pub const TASKDIALOG_BUTTONWRAPPER: Self

pub const TASKDIALOG_EXPANDOTEXT: Self

pub const TASKDIALOG_EXPANDOBUTTON_NORMAL: Self

pub const TASKDIALOG_EXPANDOBUTTON_HOVER: Self

pub const TASKDIALOG_EXPANDOBUTTON_PRESSED: Self

pub const TASKDIALOG_EXPANDOBUTTON_EXPANDEDNORMAL: Self

pub const TASKDIALOG_EXPANDOBUTTON_EXPANDEDHOVER: Self

pub const TASKDIALOG_EXPANDOBUTTON_EXPANDEDPRESSED: Self

pub const TASKDIALOG_EXPANDOBUTTON_NORMALDISABLED: Self

pub const TASKDIALOG_EXPANDOBUTTON_EXPANDEDDISABLED: Self

pub const TASKDIALOG_VERIFICATIONTEXT: Self

pub const TASKDIALOG_FOOTNOTEPANE: Self

pub const TASKDIALOG_FOOTNOTEAREA: Self

pub const TASKDIALOG_FOOTNOTESEPARATOR: Self

pub const TASKDIALOG_EXPANDEDFOOTERAREA: Self

pub const TASKDIALOG_PROGRESSBAR: Self

pub const TASKDIALOG_IMAGEALIGNMENT: Self

pub const TASKDIALOG_RADIOBUTTONPANE: Self

impl VS

pub const TEXTSTYLE_MAININSTRUCTION: Self

pub const TEXTSTYLE_INSTRUCTION: Self

pub const TEXTSTYLE_BODYTITLE: Self

pub const TEXTSTYLE_BODYTEXT: Self

pub const TEXTSTYLE_SECONDARYTEXT: Self

pub const TEXTSTYLE_HYPERLINKTEXT_HYPERLINK_NORMAL: Self

pub const TEXTSTYLE_HYPERLINKTEXT_HYPERLINK_HOT: Self

pub const TEXTSTYLE_HYPERLINKTEXT_HYPERLINK_PRESSED: Self

pub const TEXTSTYLE_HYPERLINKTEXT_HYPERLINK_DISABLED: Self

pub const TEXTSTYLE_EXPANDED: Self

pub const TEXTSTYLE_LABEL: Self

pub const TEXTSTYLE_CONTROLLABEL_CONTROLLABEL_NORMAL: Self

pub const TEXTSTYLE_CONTROLLABEL_CONTROLLABEL_DISABLED: Self

impl VS

pub const TOOLBAR_BUTTON_NORMAL: Self

pub const TOOLBAR_BUTTON_HOT: Self

pub const TOOLBAR_BUTTON_PRESSED: Self

pub const TOOLBAR_BUTTON_DISABLED: Self

pub const TOOLBAR_BUTTON_CHECKED: Self

pub const TOOLBAR_BUTTON_HOTCHECKED: Self

pub const TOOLBAR_BUTTON_NEARHOT: Self

pub const TOOLBAR_BUTTON_OTHERSIDEHOT: Self

pub const TOOLBAR_DROPDOWNBUTTON_NORMAL: Self

pub const TOOLBAR_DROPDOWNBUTTON_HOT: Self

pub const TOOLBAR_DROPDOWNBUTTON_PRESSED: Self

pub const TOOLBAR_DROPDOWNBUTTON_DISABLED: Self

pub const TOOLBAR_DROPDOWNBUTTON_CHECKED: Self

pub const TOOLBAR_DROPDOWNBUTTON_HOTCHECKED: Self

pub const TOOLBAR_DROPDOWNBUTTON_NEARHOT: Self

pub const TOOLBAR_DROPDOWNBUTTON_OTHERSIDEHOT: Self

pub const TOOLBAR_SPLITBUTTON_NORMAL: Self

pub const TOOLBAR_SPLITBUTTON_HOT: Self

pub const TOOLBAR_SPLITBUTTON_PRESSED: Self

pub const TOOLBAR_SPLITBUTTON_DISABLED: Self

pub const TOOLBAR_SPLITBUTTON_CHECKED: Self

pub const TOOLBAR_SPLITBUTTON_HOTCHECKED: Self

pub const TOOLBAR_SPLITBUTTON_NEARHOT: Self

pub const TOOLBAR_SPLITBUTTON_OTHERSIDEHOT: Self

pub const TOOLBAR_SPLITBUTTONDROPDOWN_NORMAL: Self

pub const TOOLBAR_SPLITBUTTONDROPDOWN_HOT: Self

pub const TOOLBAR_SPLITBUTTONDROPDOWN_PRESSED: Self

pub const TOOLBAR_SPLITBUTTONDROPDOWN_DISABLED: Self

pub const TOOLBAR_SPLITBUTTONDROPDOWN_CHECKED: Self

pub const TOOLBAR_SPLITBUTTONDROPDOWN_HOTCHECKED: Self

pub const TOOLBAR_SPLITBUTTONDROPDOWN_NEARHOT: Self

pub const TOOLBAR_SPLITBUTTONDROPDOWN_OTHERSIDEHOT: Self

pub const TOOLBAR_SEPARATOR_NORMAL: Self

pub const TOOLBAR_SEPARATOR_HOT: Self

pub const TOOLBAR_SEPARATOR_PRESSED: Self

pub const TOOLBAR_SEPARATOR_DISABLED: Self

pub const TOOLBAR_SEPARATOR_CHECKED: Self

pub const TOOLBAR_SEPARATOR_HOTCHECKED: Self

pub const TOOLBAR_SEPARATOR_NEARHOT: Self

pub const TOOLBAR_SEPARATOR_OTHERSIDEHOT: Self

pub const TOOLBAR_SEPARATORVERT_NORMAL: Self

pub const TOOLBAR_SEPARATORVERT_HOT: Self

pub const TOOLBAR_SEPARATORVERT_PRESSED: Self

pub const TOOLBAR_SEPARATORVERT_DISABLED: Self

pub const TOOLBAR_SEPARATORVERT_CHECKED: Self

pub const TOOLBAR_SEPARATORVERT_HOTCHECKED: Self

pub const TOOLBAR_SEPARATORVERT_NEARHOT: Self

pub const TOOLBAR_SEPARATORVERT_OTHERSIDEHOT: Self

pub const TOOLBAR_DROPDOWNBUTTONGLYPH_NORMAL: Self

pub const TOOLBAR_DROPDOWNBUTTONGLYPH_HOT: Self

pub const TOOLBAR_DROPDOWNBUTTONGLYPH_PRESSED: Self

pub const TOOLBAR_DROPDOWNBUTTONGLYPH_DISABLED: Self

pub const TOOLBAR_DROPDOWNBUTTONGLYPH_CHECKED: Self

pub const TOOLBAR_DROPDOWNBUTTONGLYPH_HOTCHECKED: Self

pub const TOOLBAR_DROPDOWNBUTTONGLYPH_NEARHOT: Self

pub const TOOLBAR_DROPDOWNBUTTONGLYPH_OTHERSIDEHOT: Self

impl VS

pub const TOOLTIP_STANDARD_NORMAL: Self

pub const TOOLTIP_STANDARD_LINK: Self

pub const TOOLTIP_STANDARDTITLE: Self

pub const TOOLTIP_BALLOON_NORMAL: Self

pub const TOOLTIP_BALLOON_LINK: Self

pub const TOOLTIP_BALLOONTITLE: Self

pub const TOOLTIP_CLOSE_NORMAL: Self

pub const TOOLTIP_CLOSE_HOT: Self

pub const TOOLTIP_CLOSE_PRESSED: Self

pub const TOOLTIP_BALLOONSTEM_POINTINGUPLEFTWALL: Self

pub const TOOLTIP_BALLOONSTEM_POINTINGUPCENTERED: Self

pub const TOOLTIP_BALLOONSTEM_POINTINGUPRIGHTWALL: Self

pub const TOOLTIP_BALLOONSTEM_POINTINGDOWNRIGHTWALL: Self

pub const TOOLTIP_BALLOONSTEM_POINTINGDOWNCENTERED: Self

pub const TOOLTIP_BALLOONSTEM_POINTINGDOWNLEFTWALL: Self

pub const TOOLTIP_WRENCH_NORMAL: Self

pub const TOOLTIP_WRENCH_HOT: Self

pub const TOOLTIP_WRENCH_PRESSED: Self

impl VS

pub const TRACKBAR_TRACK_NORMAL: Self

pub const TRACKBAR_TRACKVERT_NORMAL: Self

pub const TRACKBAR_THUMB_NORMAL: Self

pub const TRACKBAR_THUMB_HOT: Self

pub const TRACKBAR_THUMB_PRESSED: Self

pub const TRACKBAR_THUMB_FOCUSED: Self

pub const TRACKBAR_THUMB_DISABLED: Self

pub const TRACKBAR_THUMBBOTTOM_NORMAL: Self

pub const TRACKBAR_THUMBBOTTOM_HOT: Self

pub const TRACKBAR_THUMBBOTTOM_PRESSED: Self

pub const TRACKBAR_THUMBBOTTOM_FOCUSED: Self

pub const TRACKBAR_THUMBBOTTOM_DISABLED: Self

pub const TRACKBAR_THUMBTOP_NORMAL: Self

pub const TRACKBAR_THUMBTOP_HOT: Self

pub const TRACKBAR_THUMBTOP_PRESSED: Self

pub const TRACKBAR_THUMBTOP_FOCUSED: Self

pub const TRACKBAR_THUMBTOP_DISABLED: Self

pub const TRACKBAR_THUMBVERT_NORMAL: Self

pub const TRACKBAR_THUMBVERT_HOT: Self

pub const TRACKBAR_THUMBVERT_PRESSED: Self

pub const TRACKBAR_THUMBVERT_FOCUSED: Self

pub const TRACKBAR_THUMBVERT_DISABLED: Self

pub const TRACKBAR_THUMBLEFT_NORMAL: Self

pub const TRACKBAR_THUMBLEFT_HOT: Self

pub const TRACKBAR_THUMBLEFT_PRESSED: Self

pub const TRACKBAR_THUMBLEFT_FOCUSED: Self

pub const TRACKBAR_THUMBLEFT_DISABLED: Self

pub const TRACKBAR_THUMBRIGHT_NORMAL: Self

pub const TRACKBAR_THUMBRIGHT_HOT: Self

pub const TRACKBAR_THUMBRIGHT_PRESSED: Self

pub const TRACKBAR_THUMBRIGHT_FOCUSED: Self

pub const TRACKBAR_THUMBRIGHT_DISABLED: Self

pub const TRACKBAR_TICS_NORMAL: Self

pub const TRACKBAR_TICSVERT_NORMAL: Self

impl VS

pub const TREEVIEW_TREEITEM_NORMAL: Self

pub const TREEVIEW_TREEITEM_HOT: Self

pub const TREEVIEW_TREEITEM_SELECTED: Self

pub const TREEVIEW_TREEITEM_DISABLED: Self

pub const TREEVIEW_TREEITEM_SELECTEDNOTFOCUS: Self

pub const TREEVIEW_TREEITEM_HOTSELECTED: Self

pub const TREEVIEW_GLYPH_CLOSED: Self

pub const TREEVIEW_GLYPH_OPENED: Self

pub const TREEVIEW_BRANCH: Self

pub const TREEVIEW_HOTGLYPH_CLOSED: Self

pub const TREEVIEW_HOTGLYPH_OPENED: Self

impl VS

pub const USERTILE_STROKEBACKGROUND: Self

pub const USERTILE_HOVERBACKGROUND_NORMAL: Self

pub const USERTILE_HOVERBACKGROUND_HOT: Self

pub const USERTILE_HOVERBACKGROUND_PRESSED: Self

impl VS

pub const TEXTSELECTIONGRIPPER_GRIPPER_NORMAL: Self

pub const TEXTSELECTIONGRIPPER_GRIPPER_CENTERED: Self

impl VS

pub const WINDOW_CAPTION_ACTIVE: Self

pub const WINDOW_CAPTION_INACTIVE: Self

pub const WINDOW_CAPTION_DISABLED: Self

pub const WINDOW_SMALLCAPTION_ACTIVE: Self

pub const WINDOW_SMALLCAPTION_INACTIVE: Self

pub const WINDOW_SMALLCAPTION_DISABLED: Self

pub const WINDOW_MINCAPTION_ACTIVE: Self

pub const WINDOW_MINCAPTION_INACTIVE: Self

pub const WINDOW_MINCAPTION_DISABLED: Self

pub const WINDOW_SMALLMINCAPTION: Self

pub const WINDOW_MAXCAPTION_ACTIVE: Self

pub const WINDOW_MAXCAPTION_INACTIVE: Self

pub const WINDOW_MAXCAPTION_DISABLED: Self

pub const WINDOW_SMALLMAXCAPTION: Self

pub const WINDOW_FRAMELEFT_ACTIVE: Self

pub const WINDOW_FRAMELEFT_INACTIVE: Self

pub const WINDOW_FRAMERIGHT_ACTIVE: Self

pub const WINDOW_FRAMERIGHT_INACTIVE: Self

pub const WINDOW_FRAMEBOTTOM_ACTIVE: Self

pub const WINDOW_FRAMEBOTTOM_INACTIVE: Self

pub const WINDOW_SMALLFRAMELEFT_ACTIVE: Self

pub const WINDOW_SMALLFRAMELEFT_INACTIVE: Self

pub const WINDOW_SMALLFRAMERIGHT_ACTIVE: Self

pub const WINDOW_SMALLFRAMERIGHT_INACTIVE: Self

pub const WINDOW_SMALLFRAMEBOTTOM_ACTIVE: Self

pub const WINDOW_SMALLFRAMEBOTTOM_INACTIVE: Self

pub const WINDOW_SYSBUTTON_NORMAL: Self

pub const WINDOW_SYSBUTTON_HOT: Self

pub const WINDOW_SYSBUTTON_PUSHED: Self

pub const WINDOW_SYSBUTTON_DISABLED: Self

pub const WINDOW_MDISYSBUTTON: Self

pub const WINDOW_MINBUTTON_NORMAL: Self

pub const WINDOW_MINBUTTON_HOT: Self

pub const WINDOW_MINBUTTON_PUSHED: Self

pub const WINDOW_MINBUTTON_DISABLED: Self

pub const WINDOW_MDIMINBUTTON_NORMAL: Self

pub const WINDOW_MDIMINBUTTON_HOT: Self

pub const WINDOW_MDIMINBUTTON_PUSHED: Self

pub const WINDOW_MDIMINBUTTON_DISABLED: Self

pub const WINDOW_MAXBUTTON_NORMAL: Self

pub const WINDOW_MAXBUTTON_HOT: Self

pub const WINDOW_MAXBUTTON_PUSHED: Self

pub const WINDOW_MAXBUTTON_DISABLED: Self

pub const WINDOW_CLOSEBUTTON_NORMAL: Self

pub const WINDOW_CLOSEBUTTON_HOT: Self

pub const WINDOW_CLOSEBUTTON_PUSHED: Self

pub const WINDOW_CLOSEBUTTON_DISABLED: Self

pub const WINDOW_SMALLCLOSEBUTTON_NORMAL: Self

pub const WINDOW_SMALLCLOSEBUTTON_HOT: Self

pub const WINDOW_SMALLCLOSEBUTTON_PUSHED: Self

pub const WINDOW_SMALLCLOSEBUTTON_DISABLED: Self

pub const WINDOW_MDICLOSEBUTTON_NORMAL: Self

pub const WINDOW_MDICLOSEBUTTON_HOT: Self

pub const WINDOW_MDICLOSEBUTTON_PUSHED: Self

pub const WINDOW_MDICLOSEBUTTON_DISABLED: Self

pub const WINDOW_RESTOREBUTTON_NORMAL: Self

pub const WINDOW_RESTOREBUTTON_HOT: Self

pub const WINDOW_RESTOREBUTTON_PUSHED: Self

pub const WINDOW_RESTOREBUTTON_DISABLED: Self

pub const WINDOW_MDIRESTOREBUTTON_NORMAL: Self

pub const WINDOW_MDIRESTOREBUTTON_HOT: Self

pub const WINDOW_MDIRESTOREBUTTON_PUSHED: Self

pub const WINDOW_MDIRESTOREBUTTON_DISABLED: Self

pub const WINDOW_HELPBUTTON_NORMAL: Self

pub const WINDOW_HELPBUTTON_HOT: Self

pub const WINDOW_HELPBUTTON_PUSHED: Self

pub const WINDOW_HELPBUTTON_DISABLED: Self

pub const WINDOW_MDIHELPBUTTON: Self

pub const WINDOW_HORZSCROLL_NORMAL: Self

pub const WINDOW_HORZSCROLL_HOT: Self

pub const WINDOW_HORZSCROLL_PUSHED: Self

pub const WINDOW_HORZSCROLL_DISABLED: Self

pub const WINDOW_HORZTHUMB_NORMAL: Self

pub const WINDOW_HORZTHUMB_HOT: Self

pub const WINDOW_HORZTHUMB_PUSHED: Self

pub const WINDOW_HORZTHUMB_DISABLED: Self

pub const WINDOW_VERTSCROLL_NORMAL: Self

pub const WINDOW_VERTSCROLL_HOT: Self

pub const WINDOW_VERTSCROLL_PUSHED: Self

pub const WINDOW_VERTSCROLL_DISABLED: Self

pub const WINDOW_VERTTHUMB_NORMAL: Self

pub const WINDOW_VERTTHUMB_HOT: Self

pub const WINDOW_VERTTHUMB_PUSHED: Self

pub const WINDOW_VERTTHUMB_DISABLED: Self

pub const WINDOW_DIALOG: Self

pub const WINDOW_CAPTIONSIZINGTEMPLATE: Self

pub const WINDOW_SMALLCAPTIONSIZINGTEMPLATE: Self

pub const WINDOW_FRAMELEFTSIZINGTEMPLATE: Self

pub const WINDOW_SMALLFRAMELEFTSIZINGTEMPLATE: Self

pub const WINDOW_FRAMERIGHTSIZINGTEMPLATE: Self

pub const WINDOW_SMALLFRAMERIGHTSIZINGTEMPLATE: Self

pub const WINDOW_FRAMEBOTTOMSIZINGTEMPLATE: Self

pub const WINDOW_SMALLFRAMEBOTTOMSIZINGTEMPLATE: Self

pub const WINDOW_FRAME_ACTIVE: Self

pub const WINDOW_FRAME_INACTIVE: Self

pub const WINDOW_BORDER: Self

impl VS

pub const LINK_HYPERLINK_NORMALTEXT: Self

pub const LINK_HYPERLINK_LINKTEXT: Self

impl VS

pub const EMPTYMARKUP_MARKUPTEXT_NORMALTEXT: Self

pub const EMPTYMARKUP_MARKUPTEXT_LINKTEXT: Self

impl VS

pub const STATIC_TEXT: Self

impl VS

pub const PAGE_UP: Self

pub const PAGE_DOWN: Self

pub const PAGE_UPHORZ: Self

pub const PAGE_DOWNHORZ: Self

impl VS

pub const MONTHCAL_BACKGROUND: Self

pub const MONTHCAL_BORDERS: Self

pub const MONTHCAL_GRIDBACKGROUND: Self

pub const MONTHCAL_COLHEADERSPLITTER: Self

pub const MONTHCAL_GRIDCELLBACKGROUND_SELECTED: Self

pub const MONTHCAL_GRIDCELLBACKGROUND_HOT: Self

pub const MONTHCAL_GRIDCELLBACKGROUND_SELECTEDHOT: Self

pub const MONTHCAL_GRIDCELLBACKGROUND_SELECTEDNOTFOCUSED: Self

pub const MONTHCAL_GRIDCELLBACKGROUND_TODAY: Self

pub const MONTHCAL_GRIDCELLBACKGROUND_TODAYSELECTED: Self

pub const MONTHCAL_GRIDCELL_HOT: Self

pub const MONTHCAL_GRIDCELL_HASSTATE: Self

pub const MONTHCAL_GRIDCELL_HASSTATEHOT: Self

pub const MONTHCAL_GRIDCELL_TODAY: Self

pub const MONTHCAL_GRIDCELL_TODAYSELECTED: Self

pub const MONTHCAL_GRIDCELL_SELECTED: Self

pub const MONTHCAL_GRIDCELL_SELECTEDHOT: Self

pub const MONTHCAL_GRIDCELLUPPER_HOT: Self

pub const MONTHCAL_GRIDCELLUPPER_HASSTATE: Self

pub const MONTHCAL_GRIDCELLUPPER_HASSTATEHOT: Self

pub const MONTHCAL_GRIDCELLUPPER_SELECTED: Self

pub const MONTHCAL_GRIDCELLUPPER_SELECTEDHOT: Self

pub const MONTHCAL_TRAILINGGRIDCELL_HOT: Self

pub const MONTHCAL_TRAILINGGRIDCELL_HASSTATE: Self

pub const MONTHCAL_TRAILINGGRIDCELL_HASSTATEHOT: Self

pub const MONTHCAL_TRAILINGGRIDCELL_TODAY: Self

pub const MONTHCAL_TRAILINGGRIDCELL_TODAYSELECTED: Self

pub const MONTHCAL_TRAILINGGRIDCELL_SELECTED: Self

pub const MONTHCAL_TRAILINGGRIDCELL_SELECTEDHOT: Self

pub const MONTHCAL_TRAILINGGRIDCELLUPPER_HOT: Self

pub const MONTHCAL_TRAILINGGRIDCELLUPPER_HASSTATE: Self

pub const MONTHCAL_TRAILINGGRIDCELLUPPER_HASSTATEHOT: Self

pub const MONTHCAL_TRAILINGGRIDCELLUPPER_SELECTED: Self

pub const MONTHCAL_TRAILINGGRIDCELLUPPER_SELECTEDHOT: Self

pub const MONTHCAL_NAVNEXT_NORMAL: Self

pub const MONTHCAL_NAVNEXT_HOT: Self

pub const MONTHCAL_NAVNEXT_PRESSED: Self

pub const MONTHCAL_NAVNEXT_DISABLED: Self

pub const MONTHCAL_NAVPREV_NORMAL: Self

pub const MONTHCAL_NAVPREV_HOT: Self

pub const MONTHCAL_NAVPREV_PRESSED: Self

pub const MONTHCAL_NAVPREV_DISABLED: Self

impl VS

pub const CLOCK_TIME_NORMAL: Self

pub const CLOCK_TIME_HOT: Self

pub const CLOCK_TIME_PRESSED: Self

impl VS

pub const TRAYNOTIFY_BACKGROUND: Self

pub const TRAYNOTIFY_ANIMBACKGROUND: Self

impl VS

pub const TASKBAR_BACKGROUNDBOTTOM: Self

pub const TASKBAR_BACKGROUNDRIGHT: Self

pub const TASKBAR_BACKGROUNDTOP: Self

pub const TASKBAR_BACKGROUNDLEFT: Self

pub const TASKBAR_SIZINGBARBOTTOM: Self

pub const TASKBAR_SIZINGBARRIGHT: Self

pub const TASKBAR_SIZINGBARTOP: Self

pub const TASKBAR_SIZINGBARLEFT: Self

impl VS

pub const TASKBAND_GROUPCOUNT: Self

pub const TASKBAND_FLASHBUTTON: Self

pub const TASKBAND_FLASHBUTTONGROUPMENU: Self

impl VS

pub const STARTPANEL_USERPANE: Self

pub const STARTPANEL_MOREPROGRAMS: Self

pub const STARTPANEL_MOREPROGRAMSARROW_NORMAL: Self

pub const STARTPANEL_MOREPROGRAMSARROW_HOT: Self

pub const STARTPANEL_MOREPROGRAMSARROW_PRESSED: Self

pub const STARTPANEL_PROGLIST: Self

pub const STARTPANEL_PROGLISTSEPARATOR: Self

pub const STARTPANEL_PLACESLIST: Self

pub const STARTPANEL_PLACESLISTSEPARATOR: Self

pub const STARTPANEL_LOGOFF: Self

pub const STARTPANEL_LOGOFFBUTTONS_NORMAL: Self

pub const STARTPANEL_LOGOFFBUTTONS_HOT: Self

pub const STARTPANEL_LOGOFFBUTTONS_PRESSED: Self

pub const STARTPANEL_USERPICTURE: Self

pub const STARTPANEL_PREVIEW: Self

pub const STARTPANEL_MOREPROGRAMSTAB_NORMAL: Self

pub const STARTPANEL_MOREPROGRAMSTAB_HOT: Self

pub const STARTPANEL_MOREPROGRAMSTAB_SELECTED: Self

pub const STARTPANEL_MOREPROGRAMSTAB_DISABLED: Self

pub const STARTPANEL_MOREPROGRAMSTAB_FOCUSED: Self

pub const STARTPANEL_NSCHOST: Self

pub const STARTPANEL_SOFTWAREEXPLORER_NORMAL: Self

pub const STARTPANEL_SOFTWAREEXPLORER_HOT: Self

pub const STARTPANEL_SOFTWAREEXPLORER_SELECTED: Self

pub const STARTPANEL_SOFTWAREEXPLORER_DISABLED: Self

pub const STARTPANEL_SOFTWAREEXPLORER_FOCUSED: Self

pub const STARTPANEL_OPENBOX_NORMAL: Self

pub const STARTPANEL_OPENBOX_HOT: Self

pub const STARTPANEL_OPENBOX_SELECTED: Self

pub const STARTPANEL_OPENBOX_DISABLED: Self

pub const STARTPANEL_OPENBOX_FOCUSED: Self

pub const STARTPANEL_SEARCHVIEW: Self

pub const STARTPANEL_MOREPROGRAMSARROWBACK_NORMAL: Self

pub const STARTPANEL_MOREPROGRAMSARROWBACK_HOT: Self

pub const STARTPANEL_MOREPROGRAMSARROWBACK_PRESSED: Self

pub const STARTPANEL_TOPMATCH: Self

pub const STARTPANEL_LOGOFFSPLITBUTTONDROPDOWN: Self

impl VS

pub const MENUBAND_NEWAPPBUTTON_NORMAL: Self

pub const MENUBAND_NEWAPPBUTTON_HOT: Self

pub const MENUBAND_NEWAPPBUTTON_PRESSED: Self

pub const MENUBAND_NEWAPPBUTTON_DISABLED: Self

pub const MENUBAND_NEWAPPBUTTON_CHECKED: Self

pub const MENUBAND_NEWAPPBUTTON_HOTCHECKED: Self

pub const MENUBAND_SEPERATOR: Self