winsafe::co

Struct WM

pub struct WM(/* private fields */);

Available on crate feature user only.

Window message codes (u32).

Note: Control-specific messages have their own types, which are convertible to WM.

Implementations

impl WM

pub const unsafe fn as_mut(&mut self) -> &mut u32

Returns a mutable reference to the underlying raw value.

Safety

Be sure the integer being set is meaningful for the actual type.

pub const unsafe fn from_raw(v: u32) -> Self

Constructs a new object by wrapping the given integer value.

Safety

Be sure the given value is meaningful for the actual type.

pub const fn raw(&self) -> u32

Returns the primitive integer underlying value.

This method is similar to Into, but it is const, therefore it can be used in const contexts.

impl WM

pub const NULL: Self

pub const CREATE: Self

pub const DESTROY: Self

pub const MOVE: Self

pub const SIZE: Self

pub const ACTIVATE: Self

pub const SETFOCUS: Self

pub const KILLFOCUS: Self

pub const ENABLE: Self

pub const SETREDRAW: Self

pub const SETTEXT: Self

pub const GETTEXT: Self

pub const GETTEXTLENGTH: Self

pub const PAINT: Self

pub const CLOSE: Self

pub const QUERYENDSESSION: Self

pub const QUERYOPEN: Self

pub const ENDSESSION: Self

pub const QUIT: Self

pub const ERASEBKGND: Self

pub const SYSCOLORCHANGE: Self

pub const SHOWWINDOW: Self

pub const WININICHANGE: Self

pub const DEVMODECHANGE: Self

pub const ACTIVATEAPP: Self

pub const FONTCHANGE: Self

pub const TIMECHANGE: Self

pub const CANCELMODE: Self

pub const SETCURSOR: Self

pub const MOUSEACTIVATE: Self

pub const CHILDACTIVATE: Self

pub const QUEUESYNC: Self

pub const GETMINMAXINFO: Self

pub const PAINTICON: Self

pub const ICONERASEBKGND: Self

pub const NEXTDLGCTL: Self

pub const SPOOLERSTATUS: Self

pub const DRAWITEM: Self

pub const MEASUREITEM: Self

pub const DELETEITEM: Self

pub const VKEYTOITEM: Self

pub const CHARTOITEM: Self

pub const SETFONT: Self

pub const GETFONT: Self

pub const SETHOTKEY: Self

pub const GETHOTKEY: Self

pub const QUERYDRAGICON: Self

pub const COMPAREITEM: Self

pub const GETOBJECT: Self

pub const COPYDATA: Self

pub const COMPACTING: Self

pub const COMMNOTIFY: Self

pub const WINDOWPOSCHANGING: Self

pub const WINDOWPOSCHANGED: Self

pub const POWER: Self

pub const NOTIFY: Self

pub const INPUTLANGCHANGEREQUEST: Self

pub const INPUTLANGCHANGE: Self

pub const TCARD: Self

pub const HELP: Self

pub const USERCHANGED: Self

pub const NOTIFYFORMAT: Self

pub const CONTEXTMENU: Self

pub const STYLECHANGING: Self

pub const STYLECHANGED: Self

pub const DISPLAYCHANGE: Self

pub const GETICON: Self

pub const SETICON: Self

pub const NCCREATE: Self

pub const NCDESTROY: Self

pub const NCCALCSIZE: Self

pub const NCHITTEST: Self

pub const NCPAINT: Self

pub const NCACTIVATE: Self

pub const GETDLGCODE: Self

pub const SYNCPAINT: Self

pub const NCMOUSEMOVE: Self

pub const NCLBUTTONDOWN: Self

pub const NCLBUTTONUP: Self

pub const NCLBUTTONDBLCLK: Self

pub const NCRBUTTONDOWN: Self

pub const NCRBUTTONUP: Self

pub const NCRBUTTONDBLCLK: Self

pub const NCMBUTTONDOWN: Self

pub const NCMBUTTONUP: Self

pub const NCMBUTTONDBLCLK: Self

pub const NCXBUTTONDOWN: Self

pub const NCXBUTTONUP: Self

pub const NCXBUTTONDBLCLK: Self

pub const INPUT_DEVICE_CHANGE: Self

pub const INPUT: Self

pub const KEYFIRST: Self

pub const KEYDOWN: Self

pub const KEYUP: Self

pub const CHAR: Self

pub const DEADCHAR: Self

pub const SYSKEYDOWN: Self

pub const SYSKEYUP: Self

pub const SYSCHAR: Self

pub const SYSDEADCHAR: Self

pub const UNICHAR: Self

pub const KEYLAST: Self

pub const IME_STARTCOMPOSITION: Self

pub const IME_ENDCOMPOSITION: Self

pub const IME_COMPOSITION: Self

pub const IME_KEYLAST: Self

pub const INITDIALOG: Self

pub const COMMAND: Self

pub const SYSCOMMAND: Self

pub const TIMER: Self

pub const HSCROLL: Self

pub const VSCROLL: Self

pub const INITMENU: Self

pub const INITMENUPOPUP: Self

pub const GESTURE: Self

pub const GESTURENOTIFY: Self

pub const MENUSELECT: Self

pub const MENUCHAR: Self

pub const ENTERIDLE: Self

pub const MENURBUTTONUP: Self

pub const MENUDRAG: Self

pub const MENUGETOBJECT: Self

pub const UNINITMENUPOPUP: Self

pub const MENUCOMMAND: Self

pub const CHANGEUISTATE: Self

pub const UPDATEUISTATE: Self

pub const QUERYUISTATE: Self

pub const CTLCOLORMSGBOX: Self

pub const CTLCOLOREDIT: Self

pub const CTLCOLORLISTBOX: Self

pub const CTLCOLORBTN: Self

pub const CTLCOLORDLG: Self

pub const CTLCOLORSCROLLBAR: Self

pub const CTLCOLORSTATIC: Self

pub const MN_GETHMENU: Self

Originally has no WM prefix.

pub const MOUSEFIRST: Self

pub const MOUSEMOVE: Self

pub const LBUTTONDOWN: Self

pub const LBUTTONUP: Self

pub const LBUTTONDBLCLK: Self

pub const RBUTTONDOWN: Self

pub const RBUTTONUP: Self

pub const RBUTTONDBLCLK: Self

pub const MBUTTONDOWN: Self

pub const MBUTTONUP: Self

pub const MBUTTONDBLCLK: Self

pub const MOUSEWHEEL: Self

pub const MOUSEHWHEEL: Self

pub const XBUTTONDOWN: Self

pub const XBUTTONUP: Self

pub const XBUTTONDBLCLK: Self

pub const MOUSELAST: Self

pub const PARENTNOTIFY: Self

pub const ENTERMENULOOP: Self

pub const EXITMENULOOP: Self

pub const NEXTMENU: Self

pub const SIZING: Self

pub const CAPTURECHANGED: Self

pub const MOVING: Self

pub const POWERBROADCAST: Self

pub const DEVICECHANGE: Self

pub const MDICREATE: Self

pub const MDIDESTROY: Self

pub const MDIACTIVATE: Self

pub const MDIRESTORE: Self

pub const MDINEXT: Self

pub const MDIMAXIMIZE: Self

pub const MDITILE: Self

pub const MDICASCADE: Self

pub const MDIICONARRANGE: Self

pub const MDIGETACTIVE: Self

pub const MDISETMENU: Self

pub const ENTERSIZEMOVE: Self

pub const EXITSIZEMOVE: Self

pub const DROPFILES: Self

pub const MDIREFRESHMENU: Self

pub const POINTERDEVICECHANGE: Self

pub const POINTERDEVICEINRANGE: Self

pub const POINTERDEVICEOUTOFRANGE: Self

pub const TOUCH: Self

pub const NCPOINTERUPDATE: Self

pub const NCPOINTERDOWN: Self

pub const NCPOINTERUP: Self

pub const POINTERUPDATE: Self

pub const POINTERDOWN: Self

pub const POINTERUP: Self

pub const POINTERENTER: Self

pub const POINTERLEAVE: Self

pub const POINTERACTIVATE: Self

pub const POINTERCAPTURECHANGED: Self

pub const TOUCHHITTESTING: Self

pub const POINTERWHEEL: Self

pub const POINTERHWHEEL: Self

pub const DM_POINTERHITTEST: Self

Originally has no WM prefix.

pub const POINTERROUTEDTO: Self

pub const POINTERROUTEDAWAY: Self

pub const POINTERROUTEDRELEASED: Self

pub const IME_SETCONTEXT: Self

pub const IME_NOTIFY: Self

pub const IME_CONTROL: Self

pub const IME_COMPOSITIONFULL: Self

pub const IME_SELECT: Self

pub const IME_CHAR: Self

pub const IME_REQUEST: Self

pub const IME_KEYDOWN: Self

pub const IME_KEYUP: Self

pub const MOUSEHOVER: Self

pub const MOUSELEAVE: Self

pub const NCMOUSEHOVER: Self

pub const NCMOUSELEAVE: Self

pub const WTSSESSION_CHANGE: Self

pub const TABLET_FIRST: Self

pub const TABLET_LAST: Self

pub const DPICHANGED: Self

pub const DPICHANGED_BEFOREPARENT: Self

pub const DPICHANGED_AFTERPARENT: Self

pub const GETDPISCALEDSIZE: Self

pub const CUT: Self

pub const COPY: Self

pub const PASTE: Self

pub const CLEAR: Self

pub const UNDO: Self

pub const RENDERFORMAT: Self

pub const RENDERALLFORMATS: Self

pub const DESTROYCLIPBOARD: Self

pub const DRAWCLIPBOARD: Self

pub const PAINTCLIPBOARD: Self

pub const VSCROLLCLIPBOARD: Self

pub const SIZECLIPBOARD: Self

pub const ASKCBFORMATNAME: Self

pub const CHANGECBCHAIN: Self

pub const HSCROLLCLIPBOARD: Self

pub const QUERYNEWPALETTE: Self

pub const PALETTEISCHANGING: Self

pub const PALETTECHANGED: Self

pub const HOTKEY: Self

pub const PRINT: Self

pub const PRINTCLIENT: Self

pub const APPCOMMAND: Self

pub const THEMECHANGED: Self

pub const CLIPBOARDUPDATE: Self

pub const DWMCOMPOSITIONCHANGED: Self

pub const DWMNCRENDERINGCHANGED: Self

pub const DWMCOLORIZATIONCOLORCHANGED: Self

pub const DWMWINDOWMAXIMIZEDCHANGE: Self

pub const DWMSENDICONICTHUMBNAIL: Self

pub const DWMSENDICONICLIVEPREVIEWBITMAP: Self

pub const GETTITLEBARINFOEX: Self

pub const HANDHELDFIRST: Self

pub const HANDHELDLAST: Self

pub const AFXFIRST: Self

pub const AFXLAST: Self

pub const PENWINFIRST: Self

pub const PENWINLAST: Self

pub const APP: Self

pub const USER: Self