Struct winsafe::co::WS

#[repr(transparent)]
pub struct WS(_);

Available on crate feature user only.

Window styles (u32).

Note: Control-specific styles have their own types, which are convertible to WS.

This is a bitflag constant, which implements the NativeBitflag trait.

Implementations

impl WS

pub const OVERLAPPED: Self = Self(0)

The window is an overlapped window. An overlapped window has a title bar and a border. Same as the WS::TILED style.

pub const POPUP: Self = Self(2147483648)

The window is a pop-up window. This style cannot be used with the WS_CHILD style.

pub const CHILD: Self = Self(1073741824)

The window is a child window. A window with this style cannot have a menu bar. This style cannot be used with the WS::POPUP style.

pub const MINIMIZE: Self = Self(536870912)

The window is initially minimized. Same as the WS::ICONIC style.

pub const VISIBLE: Self = Self(268435456)

The window is initially visible. This style can be turned on and off by using the HWND::ShowWindow or HWND::SetWindowPos function.

pub const DISABLED: Self = Self(134217728)

The window is initially disabled. A disabled window cannot receive input from the user. To change this after a window has been created use the HWND::EnableWindow function.

pub const CLIPSIBLINGS: Self = Self(67108864)

Clips child windows relative to each other; that is when a particular child window receives a wm::Paint message, the WS::CLIPSIBLINGS style clips all other overlapping child windows out of the region of the child window to be updated. If WS::CLIPSIBLINGS is not specified and child windows overlap it is possible when drawing within the client area of a child window to draw within the client area of a neighboring child window.

pub const CLIPCHILDREN: Self = Self(33554432)

Excludes the area occupied by child windows when drawing occurs within the parent window. This style is used when creating the parent window.

pub const MAXIMIZE: Self = Self(16777216)

The window is initially maximized.

pub const CAPTION: Self = Self(12582912)

The window has a title bar (includes the WS::BORDER style).

pub const BORDER: Self = Self(8388608)

The window has a thin-line border.

pub const DLGFRAME: Self = Self(4194304)

The window has a border of a style typically used with dialog boxes. A window with this style cannot have a title bar.

pub const VSCROLL: Self = Self(2097152)

The window has a vertical scroll bar.

pub const HSCROLL: Self = Self(1048576)

The window has a horizontal scroll bar.

pub const SYSMENU: Self = Self(524288)

The window has a window menu on its title bar. The WS::CAPTION style must also be specified.

pub const THICKFRAME: Self = Self(262144)

The window has a sizing border. Same as the WS::SIZEBOX style.

pub const GROUP: Self = Self(131072)

The window is the first control of a group of controls. The group consists of this first control and all controls defined after it up to the next control with the WS::GROUP style. The first control in each group usually has the WS::TABSTOP style so that the user can move from group to group. The user can subsequently change the keyboard focus from one control in the group to the next control in the group by using the direction keys.

You can turn this style on and off to change dialog box navigation. To change this style after a window has been created use the HWND::SetWindowLongPtr function.

pub const TABSTOP: Self = Self(65536)

The window is a control that can receive the keyboard focus when the user presses the TAB key. Pressing the TAB key changes the keyboard focus to the next control with the WS::TABSTOP style.

You can turn this style on and off to change dialog box navigation. To change this style after a window has been created use the HWND::SetWindowLongPtr function. For user-created windows and modeless dialogs to work with tab stops alter the message loop to call the HWND::IsDialogMessage function.

pub const MINIMIZEBOX: Self = Self(131072)

The window has a minimize button. Cannot be combined with the WS_EX::CONTEXTHELP style. The WS::SYSMENU style must also be specified.

pub const MAXIMIZEBOX: Self = Self(65536)

The window has a maximize button. Cannot be combined with the WS_EX::CONTEXTHELP style. The WS::SYSMENU style must also be specified.

pub const TILED: Self = Self(Self::OVERLAPPED.0)

The window is an overlapped window. An overlapped window has a title bar and a border. Same as the WS::OVERLAPPED style.

pub const ICONIC: Self = Self(Self::MINIMIZE.0)

The window is initially minimized. Same as the WS::MINIMIZE style.

pub const SIZEBOX: Self = Self(Self::THICKFRAME.0)

The window has a sizing border. Same as the WS::THICKFRAME style.

pub const TILEDWINDOW: Self = Self(Self::OVERLAPPEDWINDOW.0)

The window is an overlapped window. Same as the WS::OVERLAPPEDWINDOW style.

pub const OVERLAPPEDWINDOW: Self = Self(Self::OVERLAPPED.0 | Self::CAPTION.0 | Self::SYSMENU.0 | Self::THICKFRAME.0 | Self::MINIMIZEBOX.0 | Self::MAXIMIZEBOX.0)

The window is an overlapped window. Same as the WS::TILEDWINDOW style.

pub const POPUPWINDOW: Self = Self(Self::POPUP.0 | Self::BORDER.0 | Self::SYSMENU.0)

The window is a pop-up window. This style cannot be used with the WS::CHILD style.

pub const CHILDWINDOW: Self = Self(Self::CHILD.0)

Same as the WS::CHILD style.