Module winsafe::guard

RAII implementation for handles.

Most handles require cleanup routines, but some handles require cleanup routines only under certain circumstances, what makes it impossible to universally implement these routines for the handle types.

So, when a handle needs a cleanup routine, it is instead returned wrapped into a guard.

Structs

ClipboardGuarduser

RAII implementation for clipboard which automatically calls CloseClipboard when the object goes out of scope.

HandleGuardkernel

RAII implementation for a Handle which automatically calls CloseHandle when the object goes out of scope.

HdcDeleteGuardgdi

RAII implementation for HDC which automatically calls DeleteDC when the object goes out of scope.

HdcPaintGuarduser

RAII implementation for HDC which automatically calls EndPaint when the object goes out of scope.

HdcReleaseGuarduser

RAII implementation for HDC which automatically calls ReleaseDC when the object goes out of scope.

HdwpGuarduser

RAII implementation for HDWP which automatically calls EndDeferWindowPos when the object goes out of scope.

HfileLockGuardkernel

RAII implementation for the HFILE lock which automatically calls UnlockFile when the object goes out of scope.

HfilemapviewGuardkernel

RAII implementation for HFILEMAPVIEW which automatically calls UnmapViewOfFile when the object goes out of scope.

HfindfileGuardkernel

RAII implementation for HFINDFILE which automatically calls FindClose when the object goes out of scope.

HglobalGuardkernel

RAII implementation for HGLOBAL which automatically calls GlobalFree when the object goes out of scope.

HimagelistDragGuardcomctl

RAII implementation for imagelist drag which automatically calls ImageList_EndDrag when the object goes out of scope.

HinstanceGuardkernel

RAII implementation for HINSTANCE which automatically calls FreeLibrary when the object goes out of scope.

HkeyGuardadvapi

RAII implementation for HKEY which automatically calls RegCloseKey when the object goes out of scope.

HthemeGuarduxtheme

RAII implementation for HTHEME which automatically calls CloseThemeData when the object goes out of scope.

HupdatersrcGuardkernel

RAII implementation HUPDATERSRC which automatically calls EndUpdateResource when the object goes out of scope.

ProcessInformationGuardkernel

RAII implementation for PROCESS_INFORMATION which automatically calls CloseHandle on hProcess and hThread fields when the object goes out of scope.