Module winsafe::guard

RAII implementation for various resources, which automatically perform cleanup routines when the object goes out of scope.

The guards are named after the functions they call.

Structs

CloseClipboardGuarduser

RAII implementation for clipboard which automatically calls CloseClipboard when the object goes out of scope.

CloseDesktopGuarduser

RAII implementation for HDESK which automatically calls CloseDesktop when the object goes out of scope.

CloseHandleGuardkernel

RAII implementation for a Handle which automatically calls CloseHandle when the object goes out of scope.

CloseHandlePiGuardkernel

RAII implementation for PROCESS_INFORMATION which automatically calls CloseHandle on hProcess and hThread fields when the object goes out of scope.

CloseThemeDataGuarduxtheme

RAII implementation for HTHEME which automatically calls CloseThemeData when the object goes out of scope.

CoUninitializeGuardole

RAII implementation which automatically calls CoUninitialize when the object goes out of scope.

DeleteDCGuardgdi

RAII implementation for HDC which automatically calls DeleteDC when the object goes out of scope.

DeleteObjectGuardgdi

RAII implementation for a GdiObject which automatically calls DeleteObject when the object goes out of scope.

DestroyAcceleratorTableGuarduser

RAII implementation for HACCEL which automatically calls DestroyAcceleratorTable when the object goes out of scope.

DestroyCursorGuarduser

RAII implementation for HCURSOR which automatically calls DestroyCursor when the object goes out of scope.

DestroyIconGuarduser

RAII implementation for HICON which automatically calls DestroyIcon when the object goes out of scope.

DestroyIconShfiGuardshell

RAII implementation for SHFILEINFO which automatically calls DestroyIcon on hIcon field when the object goes out of scope.

DestroyIconSiiGuardshell

RAII implementation for SHSTOCKICONINFO which automatically calls DestroyIcon on hIcon field when the object goes out of scope.

EndDeferWindowPosGuarduser

RAII implementation for HDWP which automatically calls EndDeferWindowPos when the object goes out of scope.

EndPaintGuarduser

RAII implementation for HDC which automatically calls EndPaint when the object goes out of scope.

EndUpdateResourceGuardkernel

RAII implementation HUPDATERSRC which automatically calls EndUpdateResource when the object goes out of scope.

FindCloseGuardkernel

RAII implementation for HFINDFILE which automatically calls FindClose when the object goes out of scope.

FreeLibraryGuardkernel

RAII implementation for HINSTANCE which automatically calls FreeLibrary when the object goes out of scope.

GlobalFreeGuardkernel

RAII implementation for HGLOBAL which automatically calls GlobalFree when the object goes out of scope.

GlobalUnlockGuardkernel

RAII implementation for HGLOBAL lock which automatically calls GlobalUnlock when the object goes out of scope.

ImageListDestroyGuardcomctl

RAII implementation for HIMAGELIST which automatically calls ImageList_Destroy when the object goes out of scope.

ImageListEndDragGuardcomctl

RAII implementation for image list drag which automatically calls ImageList_EndDrag when the object goes out of scope.

RegCloseKeyGuardadvapi

RAII implementation for HKEY which automatically calls RegCloseKey when the object goes out of scope.

ReleaseCaptureGuarduser

RAII implementation for HWND which automatically calls ReleaseCapture when the object goes out of scope.

ReleaseDCGuarduser

RAII implementation for HDC which automatically calls ReleaseDC when the object goes out of scope.

SelectObjectGuardgdi

RAII implementation for HDC::SelectObject calls, which automatically selects the previous GDI object at the end of the scope.

UnlockFileGuardkernel

RAII implementation for the HFILE lock which automatically calls UnlockFile when the object goes out of scope.

UnmapViewOfFileGuardkernel

RAII implementation for HFILEMAPVIEW which automatically calls UnmapViewOfFile when the object goes out of scope.