Module winsafe::guard

RAII implementation for various resources, which automatically perform cleanup routines when the object goes out of scope.

The guards are named after the functions they call.

Structs

CloseHandleGuard

RAII implementation for a Handle which automatically calls CloseHandle when the object goes out of scope.

CloseHandlePiGuard

RAII implementation for PROCESS_INFORMATION which automatically calls CloseHandle on hProcess and hThread fields when the object goes out of scope.

CloseServiceHandleGuard

RAII implementation for HSC which automatically calls CloseServiceHandle when the object goes out of scope.

CloseServiceHandleSvcGuard

RAII implementation for HSERVICE which automatically calls CloseServiceHandle when the object goes out of scope.

DeregisterEventSourceGuard

RAII implementation for HEVENTLOG which automatically calls DeregisterEventSource when the object goes out of scope.

EndUpdateResourceGuard

RAII implementation HUPDATERSRC which automatically calls EndUpdateResource when the object goes out of scope.

FindCloseGuard

RAII implementation for HFINDFILE which automatically calls FindClose when the object goes out of scope.

FreeLibraryGuard

RAII implementation for HINSTANCE which automatically calls FreeLibrary when the object goes out of scope.

FreeSidGuard

RAII implementation for SID, returned by AllocateAndInitializeSid, which automatically calls FreeSid when the object goes out of scope.

GlobalFreeGuard

RAII implementation for HGLOBAL which automatically calls GlobalFree when the object goes out of scope.

GlobalUnlockGuard

RAII implementation for HGLOBAL lock which automatically calls GlobalUnlock when the object goes out of scope.

HeapDestroyGuard

RAII implementation for HHEAP which automatically calls HeapDestroy when the object goes out of scope.

HeapFreeGuard

RAII implementation for the memory allocated by HHEAP::HeapAlloc which automatically calls HeapFree when the object goes out of scope.

HeapUnlockGuard

RAII implementation for HHEAP which automatically calls HeapUnlock when the object goes out of scope.

LocalFreeGuard

RAII implementation for HLOCAL which automatically calls LocalFree when the object goes out of scope.

LocalFreeSidGuard

RAII implementation for SID, returned by ConvertStringSidToSid, which automatically calls LocalFree when the object goes out of scope.

LocalUnlockGuard

RAII implementation for HLOCAL lock which automatically calls LocalUnlock when the object goes out of scope.

RegCloseKeyGuard

RAII implementation for HKEY which automatically calls RegCloseKey when the object goes out of scope.

SidGuard

RAII implementation for SID, returned by CopySid, CreateWellKnownSid, GetWindowsAccountDomainSid and LookupAccountName, which automatically frees the underlying memory block when the object goes out of scope.

TokenGroupsGuard

RAII implementation for TOKEN_GROUPS which manages the allocated memory.

TokenPrivilegesGuard

RAII implementation for TOKEN_PRIVILEGES which manages the allocated memory.

UnlockFileGuard

RAII implementation for the HFILE lock which automatically calls UnlockFile when the object goes out of scope.

UnmapViewOfFileGuard

RAII implementation for HFILEMAPVIEW which automatically calls UnmapViewOfFile when the object goes out of scope.