Module winsafe::gui

High-level GUI abstractions for custom windows and native controls.

You'll probably want to start your GUI application using the CustomMain.

Modules

events
Structs which expose the event methods of windows and controls.

Structs

Button
Native button control.

ButtonOpts
Options to create a Button programatically with Button::new.

CheckBox
Native check box control.

CheckBoxOpts
Options to create a CheckBox programatically with CheckBox::new.

ComboBox
Native combo box control.

ComboBoxOpts
Options to create a ComboBox programatically with ComboBox::new.

CustomControl
Custom child control.

CustomControlOpts
Options to create a CustomControl programatically with CustomControl::new.

CustomMain
Custom main application window. Usually, this is the first window of your application, launched directly from the main function.

CustomMainOpts
Options to create a CustomMain programatically with CustomMain::new.

CustomModal
Custom modal window.

CustomModalOpts
Options to create a CustomModal programatically with CustomModal::new.

DateTimePicker
Native date and time picker control.

DateTimePickerOpts
Options to create a DateTimePicker programatically with DateTimePicker::new.

Edit
Native edit control.

EditOpts
Options to create an Edit programatically with Edit::new.

Label
Native label control.

LabelOpts
Options to create a Label programatically with label::new.

ListBox
Native list box control.

ListBoxOpts
Options to create a ListBox programatically with ListBox::new.

ListView
Native list view control.

ListViewOpts
Options to create a ListView programatically with ListView::new.

MonthCalendar
Native month calendar control.

MonthCalendarOpts
Options to create a MonthCalendarBox programatically with MonthCalendarBox::new.

RadioButton
Native radio button control.

RadioButtonOpts
Options to create a RadioButton programatically with RadioButton::new.

RadioGroup
A group of native RadioButton controls.

Resizer
When the parent window is resized, automatically adjusts position and size of child controls.

StatusBar
Native status bar control, which has one or more parts.

Enums

Resz
In Resizer::add, determines how the child controls will be adjusted automatically when the parent window is resized.

StatusBarPart
Used when adding the parts in StatusBar::new.

Traits

Child
Trait to any child control.

Parent
Trait to any window which can host child controls.