Module winsafe::gui

High-level GUI abstractions for user windows and native controls. They can be created programmatically or by loading resources from a .res file. These files can be created with a WYSIWYG resource editor.

You’ll probably want to start your GUI application using the WindowMain.

Modules

events

Events exposed by windows and controls, which allow the handling of native Windows messages.

spec

Structs which expose specialized methods of controls.

Structs

Buttongui

Native button control.

ButtonOptsgui

Options to create a Button programmatically with Button::new.

CheckBoxgui

Native check box control, actually a variation of the ordinary Button: just a button with a specific style.

CheckBoxOptsgui

Options to create a CheckBox programmatically with CheckBox::new.

ComboBoxgui

Native combo box control.

ComboBoxOptsgui

Options to create a ComboBox programmatically with ComboBox::new.

DateTimePickergui

Native date and time picker control.

DateTimePickerOptsgui

Options to create a DateTimePicker programmatically with DateTimePicker::new.

Editgui

Native edit control.

EditOptsgui

Options to create an Edit programmatically with Edit::new.

Labelgui

Native label control.

LabelOptsgui

Options to create a Label programmatically with label::new.

ListBoxgui

Native list box control. Not to be confused with the more complex list view control.

ListBoxOptsgui

Options to create a ListBox programmatically with ListBox::new.

ListViewgui

Native list view control. Not to be confused with the simpler list box control.

ListViewOptsgui

Options to create a ListView programmatically with ListView::new.

MonthCalendargui

Native month calendar control.

MonthCalendarOptsgui

Options to create a MonthCalendar programmatically with MonthCalendar::new.

ProgressBargui

Native progress bar control.

ProgressBarOptsgui

Options to create a ProgressBar programmatically with ProgressBar::new.

RadioButtongui

Native radio button control, actually a variation of the ordinary Button: just a button with a specific style.

RadioButtonOptsgui

Options to create a RadioButton programmatically with RadioGroup::new.

RadioGroupgui

A group of native RadioButton controls.

StatusBargui

Native status bar control, which has one or more parts.

Trackbargui

Native trackbar control.

TrackbarOptsgui

Options to create a Trackbar programmatically with Trackbar::new.

TreeViewgui

Native tree view control.

TreeViewOptsgui

Options to create a TreeView programmatically with TreeView::new.

WindowControlgui

An user child window, which can handle events. Can be programmatically created or load a dialog resource from a .res file.

WindowControlOptsgui

Options to create a WindowControl programmatically with WindowControl::new.

WindowMaingui

An user main window, which can handle events. Usually, this is the first window of your application, launched directly from the main function. Can be programmatically created or load a dialog resource from a .res file.

WindowMainOptsgui

Options to create a WindowMain programmatically with WindowMain::new.

WindowModalgui

An user modal window, which can handle events. Can be programmatically created or load a dialog resource from a .res file.

WindowModalOptsgui

Options to create a WindowModal programmatically with WindowModal::new.

Enums

CheckStategui

Possible states of a CheckBox control.

Horzgui

Specifies the horizontal behavior of the control when the parent window is resized.

StatusBarPartgui

Used when adding the parts in StatusBar::new.

Vertgui

Specifies the vertical behavior of the control when the parent window is resized.