Module winsafe::gui

High-level GUI abstractions for user windows and native controls.

You’ll probably want to start your GUI application using the WindowMain.

Modules

events
Structs which expose the event methods of windows and controls.

Structs

Button
Native button control.

ButtonOpts
Options to create a Button programmatically with Button::new.

CheckBox
Native check box control, actually a variation of the ordinary Button: just a button with a specific style.

CheckBoxOpts
Options to create a CheckBox programmatically with CheckBox::new.

ComboBox
Native combo box control.

ComboBoxItems
Exposes item methods of a ComboBox control.

ComboBoxOpts
Options to create a ComboBox programmatically with ComboBox::new.

DateTimePicker
Native date and time picker control.

DateTimePickerOpts
Options to create a DateTimePicker programmatically with DateTimePicker::new.

Edit
Native edit control.

EditOpts
Options to create an Edit programmatically with Edit::new.

Label
Native label control.

LabelOpts
Options to create a Label programmatically with label::new.

ListBox
Native list box control. Not to be confused with the more complex list view control.

ListBoxItems
Exposes item methods of a ListBox control.

ListBoxOpts
Options to create a ListBox programmatically with ListBox::new.

ListView
Native list view control. Not to be confused with the simpler list box control.

ListViewColumns
Exposes column methods of a ListView control.

ListViewItems
Exposes item methods of a ListView control.

ListViewOpts
Options to create a ListView programmatically with ListView::new.

MonthCalendar
Native month calendar control.

MonthCalendarOpts
Options to create a MonthCalendar programmatically with MonthCalendar::new.

ProgressBar
Native progress bar control.

ProgressBarOpts
Options to create a ProgressBar programmatically with ProgressBar::new.

RadioButton
Native radio button control, actually a variation of the ordinary Button: just a button with a specific style.

RadioButtonOpts
Options to create a RadioButton programmatically with RadioGroup::new.

RadioGroup
A group of native RadioButton controls.

Resizer
When the parent window is resized, automatically adjusts position and size of child controls.

StatusBar
Native status bar control, which has one or more parts.

Trackbar
Native trackbar control.

TrackbarOpts
Options to create a Trackbar programmatically with Trackbar::new.

WindowControl
An user child window, which can handle events. Can be programmatically created or load a dialog resource from a .rc script.

WindowControlOpts
Options to create a WindowControl programmatically with WindowControl::new.

WindowMain
An user main window, which can handle events. Usually, this is the first window of your application, launched directly from the main function. Can be programmatically created or load a dialog resource from a .rc script.

WindowMainOpts
Options to create a WindowMain programmatically with WindowMain::new.

WindowModal
An user modal window, which can handle events. Can be programmatically created or load a dialog resource from a .rc script.

WindowModalOpts
Options to create a WindowModal programmatically with WindowModal::new.

Enums

Resz
In Resizer::new, determines how the child controls will be adjusted automatically when the parent window is resized.

StatusBarPart
Used when adding the parts in StatusBar::new.

Traits

Child
Trait to any child control.

Parent
Trait to any window which can host child controls.