Module winsafe::gui::events

Structs which expose the event methods of windows and controls.

You cannot directly instantiate these objects, they are created internally by the controls.

Structs

ButtonEvents

Exposes button control notifications.

ComboBoxEvents

Exposes combo box control notifications.

DateTimePickerEvents

Exposes date and time picker control notifications.

EditEvents

Exposes edit control notifications.

LabelEvents

Exposes label control notifications.

ListBoxEvents

Exposes list box control notifications.

ListViewEvents

Exposes list view control notifications.

MonthCalendarEvents

Exposes month calendar control notifications.

RadioGroupEvents

Exposes button control notifications for a RadioGroup.

StatusBarEvents

Exposes status bar control notifications.

TrackbarEvents

Exposes trackbar control notifications.

TreeViewEvents

Exposes tree view control notifications.

WindowEvents

Exposes window messages.