Struct winsafe::gui::events::ButtonEvents

pub struct ButtonEvents(_);

Available on crate feature gui only.

Exposes button control notifications.

These event methods are just proxies to the WindowEvents of the parent window, who is the real responsible for the child event handling.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl ButtonEvents

pub fn bcn_drop_down<F>(&self, func: F)where
F: Fn(&NMBCDROPDOWN) -> AnyResult<()> + 'static,

BCN_DROPDOWN notification.

pub fn bcn_hot_item_change<F>(&self, func: F)where
F: Fn(&NMBCHOTITEM) -> AnyResult<()> + 'static,

BCN_HOTITEMCHANGE notification.

pub fn bn_clicked<F>(&self, func: F)where
F: Fn() -> AnyResult<()> + 'static,

BN_CLICKED command notification.

Examples

use winsafe::prelude::*;
use winsafe::{gui, AnyResult};

let btn: gui::Button; // initialized somewhere

btn.on().bn_clicked({
    let btn = btn.clone(); // to pass into the closure
    move || -> AnyResult<()> {
        println!("HWND: {}", btn.hwnd());
        Ok(())
    }
});

pub fn bn_dbl_clk<F>(&self, func: F)where
F: Fn() -> AnyResult<()> + 'static,

BN_DBLCLK command notification.

pub fn bn_kill_focus<F>(&self, func: F)where
F: Fn() -> AnyResult<()> + 'static,

BN_KILLFOCUS command notification.

pub fn bn_set_focus<F>(&self, func: F)where
F: Fn() -> AnyResult<()> + 'static,

BN_SETFOCUS command notification.

pub fn nm_custom_draw<F>(&self, func: F)where
F: Fn(&NMCUSTOMDRAW) -> AnyResult<CDRF> + 'static,

NM_CUSTOMDRAW notification.