Struct winsafe::gui::events::ComboBoxEvents

pub struct ComboBoxEvents { /* fields omitted */ }

Exposes combo box control notifications.

These event methods are just proxies to the MsgEvents of the parent window, who is the real responsible for the child event handling.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl ComboBoxEvents

pub fn cbn_close_up<F>(&self, func: F) where
F: FnMut() + 'static,

CBN_CLOSEUP notification.

Sent when the list box of a combo box has been closed.

pub fn cbn_dbl_clk<F>(&self, func: F) where
F: FnMut() + 'static,

CBN_DBLCLK notification.

Sent when the user double-clicks a string in the list box of a combo box.

pub fn cbn_drop_down<F>(&self, func: F) where
F: FnMut() + 'static,

CBN_DROPDOWN notification.

Sent when the list box of a combo box is about to be made visible.

pub fn cbn_edit_change<F>(&self, func: F) where
F: FnMut() + 'static,

CBN_EDITCHANGE notification.

Sent after the user has taken an action that may have altered the text in the edit control portion of a combo box. Unlike the CBN_EDITUPDATE notification code, this notification code is sent after the system updates the screen.

pub fn cbn_edit_update<F>(&self, func: F) where
F: FnMut() + 'static,

CBN_EDITUPDATE notification.

Sent when the edit control portion of a combo box is about to display altered text. This notification code is sent after the control has formatted the text, but before it displays the text.

pub fn cbn_err_space<F>(&self, func: F) where
F: FnMut() + 'static,

CBN_ERRSPACE notification.

Sent when a combo box cannot allocate enough memory to meet a specific request.

pub fn cbn_kill_focus<F>(&self, func: F) where
F: FnMut() + 'static,

CBN_KILLFOCUS notification.

Sent when a combo box loses the keyboard focus.

pub fn cbn_sel_change<F>(&self, func: F) where
F: FnMut() + 'static,

CBN_SELCHANGE notification.

Sent when the user changes the current selection in the list box of a combo box. The user can change the selection by clicking in the list box or by using the arrow keys.

pub fn cbn_sel_end_cancel<F>(&self, func: F) where
F: FnMut() + 'static,

CBN_SELENDCANCEL notification.

Sent when the user selects an item, but then selects another control or closes the dialog box. It indicates the user's initial selection is to be ignored.

pub fn cbn_sel_end_ok<F>(&self, func: F) where
F: FnMut() + 'static,

CBN_SELENDOK notification.

Sent when the user selects a list item, or selects an item and then closes the list. It indicates that the user's selection is to be processed.

pub fn cbn_set_focus<F>(&self, func: F) where
F: FnMut() + 'static,

CBN_SETFOCUS notification.

Sent when a combo box receives the keyboard focus.