Struct winsafe::gui::events::EditEvents

pub struct EditEvents(_);

Exposes edit control notifications.

These event methods are just proxies to the WindowEvents of the parent window, who is the real responsible for the child event handling.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl EditEvents

pub fn en_align_ltr_ec<F>(&self, func: F) where
F: Fn() -> ErrResult<()> + 'static,

EN_ALIGN_LTR_EC command notification.

Sent when the user has changed the edit control direction to left-to-right.

pub fn en_align_rtl_ec<F>(&self, func: F) where
F: Fn() -> ErrResult<()> + 'static,

EN_ALIGN_LTR_EC command notification.

Sent when the user has changed the edit control direction to right-to-left.

pub fn en_change<F>(&self, func: F) where
F: Fn() -> ErrResult<()> + 'static,

EN_CHANGE command notification.

Sent when the user has taken an action that may have altered text in an edit control. Unlike the EN_UPDATE notification code, this notification code is sent after the system updates the screen.

Examples

use winsafe::prelude::*;
use winsafe::{gui, ErrResult};

let txt: gui::Edit; // initialized somewhere

txt.on().en_change({
    let txt = txt.clone(); // pass into the closure
    move || -> ErrResult<()> {
        println!("Text: {}", btn.hwnd().GetWindowText()?);
        Ok(())
    }
});

pub fn en_err_space<F>(&self, func: F) where
F: Fn() -> ErrResult<()> + 'static,

EN_ERRSPACE command notification.

Sent when an edit control cannot allocate enough memory to meet a specific request.

pub fn en_h_scroll<F>(&self, func: F) where
F: Fn() -> ErrResult<()> + 'static,

EN_HSCROLL command notification.

Sent when the user clicks an edit control’s horizontal scroll bar. Notified before the screen is updated.

pub fn en_kill_focus<F>(&self, func: F) where
F: Fn() -> ErrResult<()> + 'static,

EN_KILLFOCUS command notification.

Sent when an edit control loses the keyboard focus.

pub fn en_max_text<F>(&self, func: F) where
F: Fn() -> ErrResult<()> + 'static,

EN_MAXTEXT command notification.

Sent when the current text insertion has exceeded the specified number of characters for the edit control. The text insertion has been truncated.

This notification code is also sent when an edit control does not have the ES_AUTOHSCROLL style and the number of characters to be inserted would exceed the width of the edit control.

This notification code is also sent when an edit control does not have the ES_AUTOVSCROLL style and the total number of lines resulting from a text insertion would exceed the height of the edit control.

pub fn en_set_focus<F>(&self, func: F) where
F: Fn() -> ErrResult<()> + 'static,

EN_SETFOCUS command notification.

Sent when an edit control receives the keyboard focus.

pub fn en_update<F>(&self, func: F) where
F: Fn() -> ErrResult<()> + 'static,

EN_UPDATE command notification.

Sent when an edit control is about to redraw itself. This notification code is sent after the control has formatted the text, but before it displays the text. This makes it possible to resize the edit control window, if necessary.

pub fn en_v_scroll<F>(&self, func: F) where
F: Fn() -> ErrResult<()> + 'static,

EN_VSCROLL command notification.

Sent when the user clicks an edit control’s vertical scroll bar or when the user scrolls the mouse wheel over the edit control. Notified before the screen is updated.