Struct winsafe::gui::events::ListViewEvents

pub struct ListViewEvents(_);

Available on crate feature gui only.

Exposes list view control notifications.

These event methods are just proxies to the WindowEvents of the parent window, who is the real responsible for the child event handling.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl ListViewEvents

pub fn lvn_begin_drag<F>(&self, func: F)where
F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static,

LVN_BEGINDRAG notification.

pub fn lvn_begin_label_edit<F>(&self, func: F)where
F: Fn(&NMLVDISPINFO<'_>) -> AnyResult<bool> + 'static,

LVN_BEGINLABELEDIT notification.

pub fn lvn_begin_r_drag<F>(&self, func: F)where
F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static,

LVN_BEGINRDRAG notification.

pub fn lvn_begin_scroll<F>(&self, func: F)where
F: Fn(&NMLVSCROLL) -> AnyResult<()> + 'static,

LVN_BEGINSCROLL notification.

pub fn lvn_column_click<F>(&self, func: F)where
F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static,

LVN_COLUMNCLICK notification.

pub fn lvn_column_drop_down<F>(&self, func: F)where
F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static,

LVN_COLUMNDROPDOWN notification.

pub fn lvn_column_overflow_click<F>(&self, func: F)where
F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static,

LVN_COLUMNOVERFLOWCLICK notification.

pub fn lvn_delete_all_items<F>(&self, func: F)where
F: Fn(&NMLISTVIEW) -> AnyResult<bool> + 'static,

LVN_DELETEALLITEMS notification.

pub fn lvn_delete_item<F>(&self, func: F)where
F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static,

LVN_DELETEITEM notification.

Examples

use winsafe::prelude::*;
use winsafe::{gui, AnyResult, NMLISTVIEW};

let list: gui::ListView; // initialized somewhere

list.on().lvn_delete_item(|p: &NMLISTVIEW| -> AnyResult<()> {
    println!("Item: {}", p.iItem);
    Ok(())
});

pub fn lvn_end_label_edit<F>(&self, func: F)where
F: Fn(&NMLVDISPINFO<'_>) -> AnyResult<bool> + 'static,

LVN_ENDLABELEDIT notification.

pub fn lvn_end_scroll<F>(&self, func: F)where
F: Fn(&NMLVSCROLL) -> AnyResult<()> + 'static,

LVN_ENDSCROLL notification.

pub fn lvn_get_disp_info<F>(&self, func: F)where
F: Fn(&NMLVDISPINFO<'_>) -> AnyResult<()> + 'static,

LVN_GETDISPINFO notification.

pub fn lvn_get_empty_markup<F>(&self, func: F)where
F: Fn(&mut NMLVEMPTYMARKUP) -> AnyResult<bool> + 'static,

LVN_GETEMPTYMARKUP notification.

pub fn lvn_get_info_tip<F>(&self, func: F)where
F: Fn(&NMLVGETINFOTIP<'_>) -> AnyResult<()> + 'static,

LVN_GETINFOTIP notification.

pub fn lvn_hot_track<F>(&self, func: F)where
F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static,

LVN_HOTTRACK notification.

pub fn lvn_incremental_search<F>(&self, func: F)where
F: Fn(&NMLVFINDITEM<'_>) -> AnyResult<()> + 'static,

LVN_INCREMENTALSEARCH notification.

pub fn lvn_insert_item<F>(&self, func: F)where
F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static,

LVN_INSERTITEM notification.

pub fn lvn_item_activate<F>(&self, func: F)where
F: Fn(&NMITEMACTIVATE) -> AnyResult<()> + 'static,

LVN_ITEMACTIVATE notification.

pub fn lvn_item_changed<F>(&self, func: F)where
F: Fn(&NMLISTVIEW) -> AnyResult<()> + 'static,

LVN_ITEMCHANGED notification.

Examples

use winsafe::prelude::*;
use winsafe::{gui, AnyResult, NMLISTVIEW};

let list: gui::ListView; // initialized somewhere

list.on().lvn_item_changed(|p: &NMLISTVIEW| -> AnyResult<()> {
    println!("Item: {}", p.iItem);
    Ok(())
});

pub fn lvn_item_changing<F>(&self, func: F)where
F: Fn(&NMLISTVIEW) -> AnyResult<bool> + 'static,

LVN_ITEMCHANGING notification.

pub fn lvn_key_down<F>(&self, func: F)where
F: Fn(&NMLVKEYDOWN) -> AnyResult<()> + 'static,

LVN_KEYDOWN notification.

Examples

use winsafe::prelude::*;
use winsafe::{co, gui, AnyResult, NMLVKEYDOWN};

let list: gui::ListView; // initialized somewhere

list.on().lvn_key_down(|p: &NMLVKEYDOWN| -> AnyResult<()> {
    if p.wVKey == co::VK::DELETE {
        println!("DEL key was pressed.");
    }
    Ok(())
});

pub fn lvn_link_click<F>(&self, func: F)where
F: Fn(&NMLVLINK) -> AnyResult<()> + 'static,

LVN_LINKCLICK notification.

pub fn lvn_marquee_begin<F>(&self, func: F)where
F: Fn() -> AnyResult<()> + 'static,

LVN_MARQUEEBEGIN notification.

pub fn lvn_od_cache_hint<F>(&self, func: F)where
F: Fn(&NMLVCACHEHINT) -> AnyResult<()> + 'static,

LVN_ODCACHEHINT notification.

pub fn lvn_od_find_item<F>(&self, func: F)where
F: Fn(&mut NMLVFINDITEM<'_>) -> AnyResult<Option<u32>> + 'static,

LVN_ODFINDITEM notification.

pub fn lvn_od_state_changed<F>(&self, func: F)where
F: Fn(&NMLVODSTATECHANGE) -> AnyResult<()> + 'static,

LVN_ODSTATECHANGED notification.

pub fn lvn_set_disp_info<F>(&self, func: F)where
F: Fn(&NMLVDISPINFO<'_>) -> AnyResult<()> + 'static,

LVN_SETDISPINFO notification.

pub fn nm_click<F>(&self, func: F)where
F: Fn(&NMITEMACTIVATE) -> AnyResult<()> + 'static,

NM_CLICK notification.

pub fn nm_custom_draw<F>(&self, func: F)where
F: Fn(&NMLVCUSTOMDRAW) -> AnyResult<CDRF> + 'static,

NM_CUSTOMDRAW notification.

pub fn nm_dbl_clk<F>(&self, func: F)where
F: Fn(&NMITEMACTIVATE) -> AnyResult<()> + 'static,

NM_DBLCLK notification.

pub fn nm_hover<F>(&self, func: F)where
F: Fn() -> AnyResult<i32> + 'static,

NM_HOVER notification.

pub fn nm_kill_focus<F>(&self, func: F)where
F: Fn() -> AnyResult<()> + 'static,

NM_KILLFOCUS notification.

pub fn nm_r_click<F>(&self, func: F)where
F: Fn(&NMITEMACTIVATE) -> AnyResult<i32> + 'static,

NM_RCLICK notification.

pub fn nm_r_dbl_clk<F>(&self, func: F)where
F: Fn(&NMITEMACTIVATE) -> AnyResult<()> + 'static,

NM_RDBLCLK notification.

pub fn nm_released_capture<F>(&self, func: F)where
F: Fn() -> AnyResult<()> + 'static,

NM_RELEASEDCAPTURE notification.

pub fn nm_return<F>(&self, func: F)where
F: Fn() -> AnyResult<()> + 'static,

NM_RETURN notification.

pub fn nm_set_focus<F>(&self, func: F)where
F: Fn() -> AnyResult<()> + 'static,

NM_SETFOCUS notification.