Struct winsafe::gui::events::ListViewEvents

pub struct ListViewEvents(_);

Exposes list view control notifications.

These event methods are just proxies to the WindowEvents of the parent window, who is the real responsible for the child event handling.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl ListViewEvents

pub fn lvn_begin_drag<F>(&self, func: F) where
F: Fn(&NMLISTVIEW) -> ErrResult<()> + 'static,

LVN_BEGINDRAG notification.

Notifies that a drag-and-drop operation involving the left mouse button is being initiated.

pub fn lvn_begin_label_edit<F>(&self, func: F) where
F: Fn(&NMLVDISPINFO<'_>) -> ErrResult<bool> + 'static,

LVN_BEGINLABELEDIT notification.

Notifies about the start of label editing for an item.

pub fn lvn_begin_r_drag<F>(&self, func: F) where
F: Fn(&NMLISTVIEW) -> ErrResult<()> + 'static,

LVN_BEGINRDRAG notification.

Notifies that a drag-and-drop operation involving the right mouse button is being initiated.

pub fn lvn_begin_scroll<F>(&self, func: F) where
F: Fn(&NMLVSCROLL) -> ErrResult<()> + 'static,

LVN_BEGINSCROLL notification.

Notifies when a scrolling operation starts.

pub fn lvn_column_click<F>(&self, func: F) where
F: Fn(&NMLISTVIEW) -> ErrResult<()> + 'static,

LVN_COLUMNCLICK notification.

Notifies that a column header was clicked while the list-view control was in report mode.

pub fn lvn_column_drop_down<F>(&self, func: F) where
F: Fn(&NMLISTVIEW) -> ErrResult<()> + 'static,

LVN_COLUMNDROPDOWN notification.

Sent by a list-view control when the list-view’s drop-down button is pressed.

pub fn lvn_column_overflow_click<F>(&self, func: F) where
F: Fn(&NMLISTVIEW) -> ErrResult<()> + 'static,

LVN_COLUMNOVERFLOWCLICK notification.

Sent by a list-view control when its overflow button is clicked.

pub fn lvn_delete_all_items<F>(&self, func: F) where
F: Fn(&NMLISTVIEW) -> ErrResult<bool> + 'static,

LVN_DELETEALLITEMS notification.

Notifies that all items in the control are about to be deleted.

pub fn lvn_delete_item<F>(&self, func: F) where
F: Fn(&NMLISTVIEW) -> ErrResult<()> + 'static,

LVN_DELETEITEM notification.

Notifies that an item is about to be deleted.

Examples

use winsafe::prelude::*;
use winsafe::{gui, ErrResult, NMLISTVIEW};

let list: gui::ListView; // initialized somewhere

list.on().lvn_delete_item(|p: &NMLISTVIEW| -> ErrResult<()> {
    println!("Item: {}", p.iItem);
    Ok(())
});

pub fn lvn_end_label_edit<F>(&self, func: F) where
F: Fn(&NMLVDISPINFO<'_>) -> ErrResult<bool> + 'static,

LVN_ENDLABELEDIT notification.

Notifies about the end of label editing for an item.

pub fn lvn_end_scroll<F>(&self, func: F) where
F: Fn(&NMLVSCROLL) -> ErrResult<()> + 'static,

LVN_ENDSCROLL notification.

Notifies when a scrolling operation ends.

pub fn lvn_get_disp_info<F>(&self, func: F) where
F: Fn(&NMLVDISPINFO<'_>) -> ErrResult<()> + 'static,

LVN_GETDISPINFO notification.

It is a request to provide information needed to display or sort a list-view item.

pub fn lvn_get_empty_markup<F>(&self, func: F) where
F: Fn(&mut NMLVEMPTYMARKUP) -> ErrResult<bool> + 'static,

LVN_GETEMPTYMARKUP notification.

Sent by list-view control when the control has no items. The notification code is a request for the parent window to provide markup text.

pub fn lvn_get_info_tip<F>(&self, func: F) where
F: Fn(&NMLVGETINFOTIP<'_>) -> ErrResult<()> + 'static,

LVN_GETINFOTIP notification.

Sent by a large icon view list-view control that has the LVS_EX_INFOTIP extended style. This notification code is sent when the list-view control is requesting additional text information to be displayed in a tooltip.

pub fn lvn_hot_track<F>(&self, func: F) where
F: Fn(&NMLISTVIEW) -> ErrResult<()> + 'static,

LVN_HOTTRACK notification.

Sent by a list-view control when the user moves the mouse over an item. This notification code is only sent by list-view controls that have the LVS_EX_TRACKSELECT extended list-view style.

pub fn lvn_incremental_search<F>(&self, func: F) where
F: Fn(&NMLVFINDITEM<'_>) -> ErrResult<()> + 'static,

LVN_INCREMENTALSEARCH notification.

Notifies that an incremental search has started.

pub fn lvn_insert_item<F>(&self, func: F) where
F: Fn(&NMLISTVIEW) -> ErrResult<()> + 'static,

LVN_INSERTITEM notification.

Notifies that a new item was inserted.

pub fn lvn_item_activate<F>(&self, func: F) where
F: Fn(&NMITEMACTIVATE) -> ErrResult<()> + 'static,

LVN_ITEMACTIVATE notification.

Sent by a list-view control when the user activates an item.

pub fn lvn_item_changed<F>(&self, func: F) where
F: Fn(&NMLISTVIEW) -> ErrResult<()> + 'static,

LVN_ITEMCHANGED notification.

Notifies that an item has changed.

Examples

use winsafe::prelude::*;
use winsafe::{gui, ErrResult, NMLISTVIEW};

let list: gui::ListView; // initialized somewhere

list.on().lvn_item_changed(|p: &NMLISTVIEW| -> ErrResult<()> {
    println!("Item: {}", p.iItem);
    Ok(())
});

pub fn lvn_item_changing<F>(&self, func: F) where
F: Fn(&NMLISTVIEW) -> ErrResult<bool> + 'static,

LVN_ITEMCHANGING notification.

Notifies that an item is changing.

pub fn lvn_key_down<F>(&self, func: F) where
F: Fn(&NMLVKEYDOWN) -> ErrResult<()> + 'static,

LVN_KEYDOWN notification.

Notifies that a key has been pressed.

Examples

use winsafe::prelude::*;
use winsafe::{gui, ErrResult, NMLVKEYDOWN};

let list: gui::ListView; // initialized somewhere

list.on().lvn_delete_item(|p: &NMLVKEYDOWN| -> ErrResult<()> {
    if p.wVKey == co::VK::DELETE {
        println!("DEL key was pressed.");
    }
    Ok(())
});

pub fn lvn_link_click<F>(&self, func: F) where
F: Fn(&NMLVLINK) -> ErrResult<()> + 'static,

LVN_LINKCLICK notification.

Notifies that a link has been clicked on.

pub fn lvn_marquee_begin<F>(&self, func: F) where
F: Fn() -> ErrResult<()> + 'static,

LVN_MARQUEEBEGIN notification.

Notifies that a bounding box (marquee) selection has begun.

pub fn lvn_od_cache_hint<F>(&self, func: F) where
F: Fn(&NMLVCACHEHINT) -> ErrResult<()> + 'static,

LVN_ODCACHEHINT notification.

Sent by a virtual list-view control when the contents of its display area have changed. For example, a list-view control sends this notification code when the user scrolls the control’s display.

pub fn lvn_od_find_item<F>(&self, func: F) where
F: Fn(&mut NMLVFINDITEM<'_>) -> ErrResult<Option<u32>> + 'static,

LVN_ODFINDITEM notification.

Sent by a virtual list-view control when it needs the owner to find a particular callback item. For example, the control will send this notification code when it receives shortcut keyboard input or when it receives an lvm::FindItem message.

pub fn lvn_od_state_changed<F>(&self, func: F) where
F: Fn(&NMLVODSTATECHANGE) -> ErrResult<()> + 'static,

LVN_ODSTATECHANGED notification.

Sent by a list-view control when the state of an item or range of items has changed.

pub fn lvn_set_disp_info<F>(&self, func: F) where
F: Fn(&NMLVDISPINFO<'_>) -> ErrResult<()> + 'static,

LVN_SETDISPINFO notification.

Notifies that you must update the information it maintains for an item.

pub fn nm_click<F>(&self, func: F) where
F: Fn(&NMITEMACTIVATE) -> ErrResult<()> + 'static,

NM_CLICK notification.

Sent by a list-view control when the user clicks an item with the left mouse button.

pub fn nm_custom_draw<F>(&self, func: F) where
F: Fn(&NMLVCUSTOMDRAW) -> ErrResult<CDRF> + 'static,

NM_CUSTOMDRAW notification.

Sent by a list-view control to notify about drawing operations.

pub fn nm_dbl_clk<F>(&self, func: F) where
F: Fn(&NMITEMACTIVATE) -> ErrResult<()> + 'static,

NM_DBLCLK notification.

Sent by a list-view control when the user double-clicks an item with the left mouse button.

pub fn nm_hover<F>(&self, func: F) where
F: Fn() -> ErrResult<i32> + 'static,

NM_HOVER notification.

Sent by a list-view control when the mouse hovers over an item.

pub fn nm_kill_focus<F>(&self, func: F) where
F: Fn() -> ErrResult<()> + 'static,

NM_KILLFOCUS notification.

Notifies that the control has lost the input focus.

pub fn nm_r_click<F>(&self, func: F) where
F: Fn(&NMITEMACTIVATE) -> ErrResult<i32> + 'static,

NM_RCLICK notification.

Sent by a list-view control when the user clicks an item with the right mouse button.

pub fn nm_r_dbl_clk<F>(&self, func: F) where
F: Fn(&NMITEMACTIVATE) -> ErrResult<()> + 'static,

NM_RDBLCLK notification.

Sent by a list-view control when the user double-clicks an item with the right mouse button.

pub fn nm_released_capture<F>(&self, func: F) where
F: Fn() -> ErrResult<()> + 'static,

NM_RELEASEDCAPTURE notification.

Notifies that the control is releasing mouse capture.

pub fn nm_return<F>(&self, func: F) where
F: Fn() -> ErrResult<()> + 'static,

NM_RETURN notification.

Notifies that the control has the input focus and that the user has pressed the ENTER key.

pub fn nm_set_focus<F>(&self, func: F) where
F: Fn() -> ErrResult<()> + 'static,

NM_SETFOCUS notification.

Notifies that the control has received the input focus.