Struct winsafe::gui::events::MsgEvents

pub struct MsgEvents(_);

Exposes window messages.

You cannot directly instantiate this object, it is created internally by the window.

Implementations

impl MsgEvents

pub fn wm<F>(&self, ident: WM, func: F) where
F: FnMut(Wm) -> isize + 'static,

Event to any window message.

Note: Instead of using this event, you should always prefer the specific events, which will give you the correct message parameters. This generic method should be used when you have a custom, non-standard window message.

Examples

Handling a custom, user-defined message:

use winsafe::co::WM;
use winsafe::gui::WindowMain;

let wnd: WindowMain; // initialize it somewhere...

let CUSTOM_MSG = WM::from(0x1234);

wnd.on().wm(CUSTOM_MSG, {
  let wnd = wnd.clone(); // pass into the closure
  move |parms| {
    println!("HWND: {}, msg ID: {}", wnd.hwnd(), parms.msg_id);
    0
  }
});

pub fn wm_timer<F>(&self, timer_id: u32, func: F) where
F: FnMut() + 'static,

WM_TIMER message, narrowed to a specific timer ID.

pub fn wm_command<F>(&self, code: CMD, ctrl_id: u16, func: F) where
F: FnMut() + 'static,

WM_COMMAND message, for specific code and control ID.

A command notification must be narrowed by the command code and the control ID, so the closure will be fired for that specific control at that specific event.

Note: Instead of using this event, you should always prefer the specific command notifications, which will give you the correct message parameters. This generic method should be used when you have a custom, non-standard window notification.

pub fn wm_command_accel_menu<F>(&self, ctrl_id: u16, func: F) where
F: FnMut() + 'static,

WM_COMMAND message, handling both co::CMD::Accelerator and co::CMD::Menu, for a specific command ID.

Ideal to be used with menu commands whose IDs are shared with accelerators.

Examples

Closing the window on ESC key:

use winsafe::co;
use winsafe::gui::WindowMain;
use winsafe::msg::WmClose;

let wnd: WindowMain; // initialize it somewhere...

wnd.on().wm_command_accel_menu(co::DLGID::CANCEL.into(), {
  let wnd = wnd.clone(); // pass into the closure
  move || {
    wnd.hwnd().PostMessage(WmClose {}).unwrap();
  }
});

pub fn wm_notify<F>(&self, id_from: u16, code: NM, func: F) where
F: FnMut(WmNotify<'_>) -> isize + 'static,

WM_NOTIFY message, for specific ID and notification code.

A notification must be narrowed by the notification code and the control ID, so the closure will be fired for that specific control at the specific event.

Note: Instead of using this event, you should always prefer the specific notifications, which will give you the correct notification struct. This generic method should be used when you have a custom, non-standard window notification.

pub fn wm_activate<F>(&self, func: F) where
F: FnMut(WmActivate) + 'static,

WM_ACTIVATEAPP message.

Warning: default handled in;

non-dialog CustomMain.

pub fn wm_activate_app<F>(&self, func: F) where
F: FnMut(WmActivateApp) + 'static,

WM_ACTIVATEAPP message.

pub fn wm_app_command<F>(&self, func: F) where
F: FnMut(WmAppCommand) + 'static,

WM_APPCOMMAND message.

pub fn wm_cancel_mode<F>(&self, func: F) where
F: FnMut() + 'static,

WM_CANCELMODE message.

pub fn wm_child_activate<F>(&self, func: F) where
F: FnMut() + 'static,

WM_CHILDACTIVATE message.

pub fn wm_close<F>(&self, func: F) where
F: FnMut() + 'static,

WM_CLOSE message.

Warning: default handled in:

dialog CustomMain;

dialog CustomModal;

non-dialog CustomModal.

pub fn wm_context_menu<F>(&self, func: F) where
F: FnMut() + 'static,

WM_CONTEXTMENU message.

pub fn wm_create<F>(&self, func: F) where
F: FnMut(WmCreate<'_, '_, '_>) -> i32 + 'static,

WM_CREATE message, sent only to non-dialog windows. Dialog windows receive WM_INITDIALOG instead.

This message is sent right after the window is created, being ideal to perform initial setups.

Examples

use winsafe::gui::WindowMain;

let wnd: WindowMain; // initialize it somewhere...

wnd.on().wm_create({
  let wnd = wnd.clone(); // pass into the closure
  move |parms| {
    println!("HWND: {}, client area: {}x{}",
      wnd.hwnd(),
      parms.createstruct.cx, parms.createstruct.cy);
    0
  }
});

pub fn wm_ctl_color_btn<F>(&self, func: F) where
F: FnMut(WmCtlColorBtn) -> HDC + 'static,

WM_CTLCOLORBTN message.

pub fn wm_ctl_color_dlg<F>(&self, func: F) where
F: FnMut(WmCtlColorDlg) -> HDC + 'static,

WM_CTLCOLORDLG message.

pub fn wm_ctl_color_edit<F>(&self, func: F) where
F: FnMut(WmCtlColorEdit) -> HDC + 'static,

WM_CTLCOLOREDIT message.

pub fn wm_ctl_color_list_box<F>(&self, func: F) where
F: FnMut(WmCtlColorListBox) -> HDC + 'static,

WM_CTLCOLORLISTBOX message.

pub fn wm_ctl_color_scroll_bar<F>(&self, func: F) where
F: FnMut(WmCtlColorScrollBar) -> HDC + 'static,

WM_CTLCOLORSCROLLBAR message.

pub fn wm_ctl_color_static<F>(&self, func: F) where
F: FnMut(WmCtlColorStatic) -> HDC + 'static,

WM_CTLCOLORSTATIC message.

pub fn wm_destroy<F>(&self, func: F) where
F: FnMut() + 'static,

WM_DESTROY message.

pub fn wm_drop_files<F>(&self, func: F) where
F: FnMut(WmDropFiles) + 'static,

WM_DROPFILES message.

pub fn wm_enable<F>(&self, func: F) where
F: FnMut(WmEnable) + 'static,

WM_ENABLE message.

pub fn wm_end_session<F>(&self, func: F) where
F: FnMut(WmEndSession) + 'static,

WM_ENDSESSION message.

pub fn wm_enter_idle<F>(&self, func: F) where
F: FnMut(WmEnterIdle) + 'static,

WM_ENTERIDLE message.

pub fn wm_enter_size_move<F>(&self, func: F) where
F: FnMut(WmEnterSizeMove) + 'static,

WM_ENTERSIZEMOVE message.

pub fn wm_erase_bkgnd<F>(&self, func: F) where
F: FnMut(WmEraseBkgnd) -> i32 + 'static,

WM_ERASEBKGND message.

pub fn wm_exit_size_move<F>(&self, func: F) where
F: FnMut(WmExitSizeMove) + 'static,

WM_EXITSIZEMOVE message.

pub fn wm_get_min_max_info<F>(&self, func: F) where
F: FnMut(WmGetMinMaxInfo<'_>) + 'static,

WM_GETMINMAXINFO message.

pub fn wm_init_dialog<F>(&self, func: F) where
F: FnMut(WmInitDialog) -> bool + 'static,

WM_INITDIALOG message, sent only to dialog windows. Non-dialog windows receive WM_CREATE instead.

This message is sent right after the dialog is created, being ideal to perform initial setups.

pub fn wm_init_menu_popup<F>(&self, func: F) where
F: FnMut(WmInitMenuPopup) + 'static,

WM_INITMENUPOPUP message.

pub fn wm_l_button_dbl_clk<F>(&self, func: F) where
F: FnMut(WmLButtonDblClk) + 'static,

WM_LBUTTONDBLCLK message.

pub fn wm_l_button_down<F>(&self, func: F) where
F: FnMut(WmLButtonDown) + 'static,

WM_LBUTTONDOWN message.

pub fn wm_l_button_up<F>(&self, func: F) where
F: FnMut(WmLButtonUp) + 'static,

WM_LBUTTONUP message.

pub fn wm_m_button_dbl_clk<F>(&self, func: F) where
F: FnMut(WmMButtonDblClk) + 'static,

WM_MBUTTONDBLCLK message.

pub fn wm_m_button_down<F>(&self, func: F) where
F: FnMut(WmMButtonDown) + 'static,

WM_MBUTTONDOWN message.

pub fn wm_m_button_up<F>(&self, func: F) where
F: FnMut(WmMButtonUp) + 'static,

WM_MBUTTONUP message.

pub fn wm_mouse_hover<F>(&self, func: F) where
F: FnMut(WmMouseHover) + 'static,

WM_MOUSEHOVER message.

pub fn wm_mouse_move<F>(&self, func: F) where
F: FnMut(WmMouseMove) + 'static,

WM_MOUSEMOVE message.

pub fn wm_move<F>(&self, func: F) where
F: FnMut(WmMove) + 'static,

WM_MOVE message.

pub fn wm_moving<F>(&self, func: F) where
F: FnMut(WmMoving<'_>) + 'static,

WM_MOVING message.

pub fn wm_nc_calc_size<F>(&self, func: F) where
F: FnMut(WmNcCalcSize<'_, '_>) -> WVR + 'static,

WM_NCCALCSIZE message.

pub fn wm_nc_create<F>(&self, func: F) where
F: FnMut(WmNcCreate<'_, '_, '_>) -> bool + 'static,

WM_NCCREATE message.

pub fn wm_nc_destroy<F>(&self, func: F) where
F: FnMut() + 'static,

WM_NCDESTROY message.

Warning: default handled in;

non-dialog CustomMain;

dialog CustomMain.

pub fn wm_nc_paint<F>(&self, func: F) where
F: FnMut(WmNcPaint) + 'static,

WM_NCPAINT message.

Warning: default handled in:

non-dialog CustomControl;

dialog CustomControl.

pub fn wm_null<F>(&self, func: F) where
F: FnMut() + 'static,

WM_NULL message.

Usually this message is not handled.

pub fn wm_paint<F>(&self, func: F) where
F: FnMut() + 'static,

WM_PAINT message.

pub fn wm_query_open<F>(&self, func: F) where
F: FnMut(WmQueryOpen) -> bool + 'static,

WM_QUERYOPEN message.

pub fn wm_r_button_dbl_clk<F>(&self, func: F) where
F: FnMut(WmRButtonDblClk) + 'static,

WM_RBUTTONDBLCLK message.

pub fn wm_r_button_down<F>(&self, func: F) where
F: FnMut(WmRButtonDown) + 'static,

WM_RBUTTONDOWN message.

pub fn wm_r_button_up<F>(&self, func: F) where
F: FnMut(WmRButtonUp) + 'static,

WM_RBUTTONUP message.

pub fn wm_set_focus<F>(&self, func: F) where
F: FnMut(WmSetFocus) + 'static,

WM_SETFOCUS message.

Warning: default handled in:

non-dialog CustomMain;

non-dialog CustomModal.

pub fn wm_set_font<F>(&self, func: F) where
F: FnMut(WmSetFont) + 'static,

WM_SETFONT message.

pub fn wm_set_icon<F>(&self, func: F) where
F: FnMut(WmSetIcon) -> Option<HICON> + 'static,

WM_SETICON message.

pub fn wm_show_window<F>(&self, func: F) where
F: FnMut(WmShowWindow) + 'static,

WM_SHOWWINDOW message.

pub fn wm_size<F>(&self, func: F) where
F: FnMut(WmSize) + 'static,

WM_SIZE message.

Examples

use winsafe::gui::WindowMain;

let wnd: WindowMain; // initialize it somewhere...

wnd.on().wm_size({
  let wnd = wnd.clone(); // pass into the closure
  move |parms| {
    println!("HWND: {}, client area: {}x{}",
      wnd.hwnd(),
      parms.width, parms.height);
  }
});

pub fn wm_sizing<F>(&self, func: F) where
F: FnMut(WmSizing<'_>) + 'static,

WM_SIZING message.

pub fn wm_style_changed<F>(&self, func: F) where
F: FnMut(WmStyleChanged<'_>) + 'static,

WM_STYLECHANGED message.

pub fn wm_style_changing<F>(&self, func: F) where
F: FnMut(WmStyleChanging<'_>) + 'static,

WM_STYLECHANGING message.

pub fn wm_theme_changed<F>(&self, func: F) where
F: FnMut(WmThemeChanged) + 'static,

WM_THEMECHANGED message.

pub fn wm_window_pos_changed<F>(&self, func: F) where
F: FnMut(WmWindowPosChanged<'_>) + 'static,

WM_WINDOWPOSCHANGED message.

pub fn wm_window_pos_changing<F>(&self, func: F) where
F: FnMut(WmWindowPosChanging<'_>) + 'static,

WM_WINDOWPOSCHANGING message.

pub fn wm_x_button_dbl_clk<F>(&self, func: F) where
F: FnMut(WmXButtonDblClk) + 'static,

WM_XBUTTONDBLCLK message.

pub fn wm_x_button_down<F>(&self, func: F) where
F: FnMut(WmXButtonDown) + 'static,

WM_XBUTTONDOWN message.

pub fn wm_x_button_up<F>(&self, func: F) where
F: FnMut(WmXButtonUp) + 'static,

WM_XBUTTONUP message.