Struct winsafe::gui::events::WindowEvents

pub struct WindowEvents(_);

Exposes window messages.

You cannot directly instantiate this object, it is created internally by the window.

Implementations

impl WindowEvents

pub fn wm<F>(&self, ident: WM, func: F) where
F: FnMut(WndMsg) -> isize + 'static,

Event to any window message.

Note: Instead of using this event, you should always prefer the specific events, which will give you the correct message parameters. This generic method should be used when you have a custom, non-standard window message.

Examples

Handling a custom, user-defined message:

use winsafe::{co, gui::WindowMain};

let wnd: WindowMain; // initialize it somewhere...

let CUSTOM_MSG = co::WM::from(0x1234);

wnd.on().wm(CUSTOM_MSG, {
    let wnd = wnd.clone(); // pass into the closure
    move |parms| {
        println!("HWND: {}, msg ID: {}", wnd.hwnd(), parms.msg_id);
        0
    }
});

pub fn wm_timer<F>(&self, timer_id: u32, func: F) where
F: FnMut() + 'static,

WM_TIMER message, narrowed to a specific timer ID.

Posted to the installing thread's message queue when a timer expires.

pub fn wm_command<F>(&self, code: CMD, ctrl_id: u16, func: F) where
F: FnMut() + 'static,

WM_COMMAND message, for specific code and control ID.

A command notification must be narrowed by the command code and the control ID, so the closure will be fired for that specific control at that specific event.

Note: Instead of using this event, you should always prefer the specific command notifications, which will give you the correct message parameters. This generic method should be used when you have a custom, non-standard window notification.

pub fn wm_command_accel_menu<F>(&self, ctrl_id: u16, func: F) where
F: FnMut() + 'static,

WM_COMMAND message, handling both co::CMD::Accelerator and co::CMD::Menu, for a specific command ID.

Ideal to be used with menu commands whose IDs are shared with accelerators.

Examples

Closing the window on ESC key:

use winsafe::{co, gui::WindowMain, msg::wm};

let wnd: WindowMain; // initialize it somewhere...

wnd.on().wm_command_accel_menu(co::DLGID::CANCEL.into(), {
    let wnd = wnd.clone(); // pass into the closure
    move || {
        wnd.hwnd().PostMessage(wm::Close {}).unwrap();
    }
});

pub fn wm_notify<F>(&self, id_from: u16, code: NM, func: F) where
F: FnMut(Notify<'_>) -> isize + 'static,

WM_NOTIFY message, for specific ID and notification code.

A notification must be narrowed by the notification code and the control ID, so the closure will be fired for that specific control at the specific event.

Note: Instead of using this event, you should always prefer the specific notifications, which will give you the correct notification struct. This generic method should be used when you have a custom, non-standard window notification.

pub fn wm_activate<F>(&self, func: F) where
F: FnMut(Activate) + 'static,

WM_ACTIVATE message.

Sent to both the window being activated and the window being deactivated. If the windows use the same input queue, the message is sent synchronously, first to the window procedure of the top-level window being deactivated, then to the window procedure of the top-level window being activated. If the windows use different input queues, the message is sent asynchronously, so the window is activated immediately.

Default handling

If you handle this event, you'll overwrite the default handling in:

non-dialog CustomMain.

pub fn wm_activate_app<F>(&self, func: F) where
F: FnMut(ActivateApp) + 'static,

WM_ACTIVATEAPP message.

Sent when a window belonging to a different application than the active window is about to be activated. The message is sent to the application whose window is being activated and to the application whose window is being deactivated.

pub fn wm_app_command<F>(&self, func: F) where
F: FnMut(AppCommand) + 'static,

WM_APPCOMMAND message.

Notifies a window that the user generated an application command event, for example, by clicking an application command button using the mouse or typing an application command key on the keyboard.

pub fn wm_cancel_mode<F>(&self, func: F) where
F: FnMut() + 'static,

WM_CANCELMODE message.

Sent to cancel certain modes, such as mouse capture. For example, the system sends this message to the active window when a dialog box or message box is displayed. Certain functions also send this message explicitly to the specified window regardless of whether it is the active window. For example, the EnableWindow function sends this message when disabling the specified window.

pub fn wm_char<F>(&self, func: F) where
F: FnMut(Char) + 'static,

WM_CHAR message.

Posted to the window with the keyboard focus when a WM_KEYDOWN message is translated by the TranslateMessage function. The WM_CHAR message contains the character code of the key that was pressed.

pub fn wm_child_activate<F>(&self, func: F) where
F: FnMut() + 'static,

WM_CHILDACTIVATE message.

Sent to a child window when the user clicks the window's title bar or when the window is activated, moved, or sized.

pub fn wm_close<F>(&self, func: F) where
F: FnMut() + 'static,

WM_CLOSE message.

Sent as a signal that a window or an application should terminate.

Default handling

If you handle this event, you'll overwrite the default handling in:

dialog CustomMain;

dialog CustomModal;

non-dialog CustomModal.

pub fn wm_context_menu<F>(&self, func: F) where
F: FnMut() + 'static,

WM_CONTEXTMENU message.

Notifies a window that the user desires a context menu to appear. The user may have clicked the right mouse button (right-clicked) in the window, pressed Shift+F10 or pressed the applications key (context menu key) available on some keyboards.

pub fn wm_create<F>(&self, func: F) where
F: FnMut(Create<'_, '_, '_>) -> i32 + 'static,

WM_CREATE message, sent only to non-dialog windows. Dialog windows receive WM_INITDIALOG instead.

Sent when an application requests that a window be created by calling the CreateWindowEx function. The message is sent before the function returns. The window procedure of the new window receives this message after the window is created, but before the window becomes visible.

Examples

use winsafe::gui::WindowMain;

let wnd: WindowMain; // initialize it somewhere...

wnd.on().wm_create({
    let wnd = wnd.clone(); // pass into the closure
    move |parms| {
        println!("HWND: {}, client area: {}x{}",
            wnd.hwnd(),
            parms.createstruct.cx,
            parms.createstruct.cy,
        );
        0
    }
});

pub fn wm_ctl_color_btn<F>(&self, func: F) where
F: FnMut(CtlColorBtn) -> HDC + 'static,

WM_CTLCOLORBTN message.

Sent to the parent window of a button before drawing the button. The parent window can change the button's text and background colors. However, only owner-drawn buttons respond to the parent window processing this message.

pub fn wm_ctl_color_dlg<F>(&self, func: F) where
F: FnMut(CtlColorDlg) -> HDC + 'static,

WM_CTLCOLORDLG message.

Sent to a dialog box before the system draws the dialog box. By responding to this message, the dialog box can set its text and background colors using the specified display device context handle.

pub fn wm_ctl_color_edit<F>(&self, func: F) where
F: FnMut(CtlColorEdit) -> HDC + 'static,

WM_CTLCOLOREDIT message.

An edit control that is not read-only or disabled sends the message to its parent window when the control is about to be drawn. By responding to this message, the parent window can use the specified device context handle to set the text and background colors of the edit control.

pub fn wm_ctl_color_list_box<F>(&self, func: F) where
F: FnMut(CtlColorListBox) -> HDC + 'static,

WM_CTLCOLORLISTBOX message.

Sent to the parent window of a list box before the system draws the list box. By responding to this message, the parent window can set the text and background colors of the list box by using the specified display device context handle.

pub fn wm_ctl_color_scroll_bar<F>(&self, func: F) where
F: FnMut(CtlColorScrollBar) -> HDC + 'static,

WM_CTLCOLORSCROLLBAR message.

Sent to the parent window of a scroll bar control when the control is about to be drawn. By responding to this message, the parent window can use the display context handle to set the background color of the scroll bar control.

pub fn wm_ctl_color_static<F>(&self, func: F) where
F: FnMut(CtlColorStatic) -> HDC + 'static,

WM_CTLCOLORSTATIC message.

A static control, or an edit control that is read-only or disabled, sends the message to its parent window when the control is about to be drawn. By responding to this message, the parent window can use the specified device context handle to set the text foreground and background colors of the static control.

pub fn wm_dead_char<F>(&self, func: F) where
F: FnMut(DeadChar) + 'static,

WM_DEADCHAR message.

Posted to the window with the keyboard focus when a WM_KEYUP message is translated by the TranslateMessage function. WM_DEADCHAR specifies a character code generated by a dead key. A dead key is a key that generates a character, such as the umlaut (double-dot), that is combined with another character to form a composite character. For example, the umlaut-O character (Ö) is generated by typing the dead key for the umlaut character, and then typing the O key.

pub fn wm_destroy<F>(&self, func: F) where
F: FnMut() + 'static,

WM_DESTROY message.

Sent when a window is being destroyed. It is sent to the window procedure of the window being destroyed after the window is removed from the screen.

This message is sent first to the window being destroyed and then to the child windows (if any) as they are destroyed. During the processing of the message, it can be assumed that all child windows still exist.

pub fn wm_drop_files<F>(&self, func: F) where
F: FnMut(DropFiles) + 'static,

WM_DROPFILES message.

Sent when the user drops a file on the window of an application that has registered itself as a recipient of dropped files.

pub fn wm_enable<F>(&self, func: F) where
F: FnMut(Enable) + 'static,

WM_ENABLE message.

Sent when an application changes the enabled state of a window. It is sent to the window whose enabled state is changing. This message is sent before the EnableWindow function returns, but after the enabled state (WS_DISABLED style bit) of the window has changed.

pub fn wm_end_session<F>(&self, func: F) where
F: FnMut(EndSession) + 'static,

WM_ENDSESSION message.

Sent to an application after the system processes the results of the WM_QUERYENDSESSION message. The WM_ENDSESSION message informs the application whether the session is ending.

pub fn wm_enter_idle<F>(&self, func: F) where
F: FnMut(EnterIdle) + 'static,

WM_ENTERIDLE message.

Sent to the owner window of a modal dialog box or menu that is entering an idle state. A modal dialog box or menu enters an idle state when no messages are waiting in its queue after it has processed one or more previous messages.

pub fn wm_enter_size_move<F>(&self, func: F) where
F: FnMut(EnterSizeMove) + 'static,

WM_ENTERSIZEMOVE message.

Sent one time to a window after it enters the moving or sizing modal loop. The window enters the moving or sizing modal loop when the user clicks the window's title bar or sizing border, or when the window passes the WM_SYSCOMMAND message to the DefWindowProc function and the wParam parameter of the message specifies the SC_MOVE or SC_SIZE value. The operation is complete when DefWindowProc returns.

The system sends the message regardless of whether the dragging of full windows is enabled.

pub fn wm_erase_bkgnd<F>(&self, func: F) where
F: FnMut(EraseBkgnd) -> i32 + 'static,

WM_ERASEBKGND message.

Sent when the window background must be erased (for example, when a window is resized). The message is sent to prepare an invalidated portion of a window for painting.

pub fn wm_exit_size_move<F>(&self, func: F) where
F: FnMut(ExitSizeMove) + 'static,

WM_EXITSIZEMOVE message.

Sent one time to a window, after it has exited the moving or sizing modal loop. The window enters the moving or sizing modal loop when the user clicks the window's title bar or sizing border, or when the window passes the WM_SYSCOMMAND message to the DefWindowProc function and the wParam parameter of the message specifies the SC_MOVE or SC_SIZE value. The operation is complete when DefWindowProc returns.

pub fn wm_get_min_max_info<F>(&self, func: F) where
F: FnMut(GetMinMaxInfo<'_>) + 'static,

WM_GETMINMAXINFO message.

Sent to a window when the size or position of the window is about to change. An application can use this message to override the window's default maximized size and position, or its default minimum or maximum tracking size.

pub fn wm_help<F>(&self, func: F) where
F: FnMut(Help<'_>) + 'static,

WM_HELP message.

Indicates that the user pressed the F1 key.

pub fn wm_init_dialog<F>(&self, func: F) where
F: FnMut(InitDialog) -> bool + 'static,

WM_INITDIALOG message, sent only to dialog windows. Non-dialog windows receive WM_CREATE instead.

Sent to the dialog box procedure immediately before a dialog box is displayed. Dialog box procedures typically use this message to initialize controls and carry out any other initialization tasks that affect the appearance of the dialog box.

pub fn wm_init_menu_popup<F>(&self, func: F) where
F: FnMut(InitMenuPopup) + 'static,

WM_INITMENUPOPUP message.

Sent when a drop-down menu or submenu is about to become active. This allows an application to modify the menu before it is displayed, without changing the entire menu.

pub fn wm_key_down<F>(&self, func: F) where
F: FnMut(KeyDown) + 'static,

WM_KEYDOWN message.

Posted to the window with the keyboard focus when a nonsystem key is pressed. A nonsystem key is a key that is pressed when the ALT key is not pressed.

pub fn wm_key_up<F>(&self, func: F) where
F: FnMut(KeyUp) + 'static,

WM_KEYUP message.

Posted to the window with the keyboard focus when a nonsystem key is released. A nonsystem key is a key that is pressed when the ALT key is not pressed, or a keyboard key that is pressed when a window has the keyboard focus.

pub fn wm_kill_focus<F>(&self, func: F) where
F: FnMut(KillFocus) + 'static,

WM_KILLFOCUS message.

Sent to a window immediately before it loses the keyboard focus.

pub fn wm_l_button_dbl_clk<F>(&self, func: F) where
F: FnMut(LButtonDblClk) + 'static,

WM_LBUTTONDBLCLK message.

Posted when the user double-clicks the left mouse button while the cursor is in the client area of a window. If the mouse is not captured, the message is posted to the window beneath the cursor. Otherwise, the message is posted to the window that has captured the mouse.

pub fn wm_l_button_down<F>(&self, func: F) where
F: FnMut(LButtonDown) + 'static,

WM_LBUTTONDOWN message.

Posted when the user presses the left mouse button while the cursor is in the client area of a window. If the mouse is not captured, the message is posted to the window beneath the cursor. Otherwise, the message is posted to the window that has captured the mouse.

pub fn wm_l_button_up<F>(&self, func: F) where
F: FnMut(LButtonUp) + 'static,

WM_LBUTTONUP message.

Posted when the user releases the left mouse button while the cursor is in the client area of a window. If the mouse is not captured, the message is posted to the window beneath the cursor. Otherwise, the message is posted to the window that has captured the mouse.

pub fn wm_m_button_dbl_clk<F>(&self, func: F) where
F: FnMut(MButtonDblClk) + 'static,

WM_MBUTTONDBLCLK message.

Posted when the user double-clicks the middle mouse button while the cursor is in the client area of a window. If the mouse is not captured, the message is posted to the window beneath the cursor. Otherwise, the message is posted to the window that has captured the mouse.

pub fn wm_m_button_down<F>(&self, func: F) where
F: FnMut(MButtonDown) + 'static,

WM_MBUTTONDOWN message.

Posted when the user presses the middle mouse button while the cursor is in the client area of a window. If the mouse is not captured, the message is posted to the window beneath the cursor. Otherwise, the message is posted to the window that has captured the mouse.

pub fn wm_m_button_up<F>(&self, func: F) where
F: FnMut(MButtonUp) + 'static,

WM_MBUTTONUP message.

Posted when the user releases the middle mouse button while the cursor is in the client area of a window. If the mouse is not captured, the message is posted to the window beneath the cursor. Otherwise, the message is posted to the window that has captured the mouse.

pub fn wm_mouse_hover<F>(&self, func: F) where
F: FnMut(MouseHover) + 'static,

WM_MOUSEHOVER message.

Posted to a window when the cursor hovers over the client area of the window for the period of time specified in a prior call to TrackMouseEvent.

pub fn wm_mouse_move<F>(&self, func: F) where
F: FnMut(MouseMove) + 'static,

WM_MOUSEMOVE message.

Posted to a window when the cursor moves. If the mouse is not captured, the message is posted to the window that contains the cursor. Otherwise, the message is posted to the window that has captured the mouse.

pub fn wm_move<F>(&self, func: F) where
F: FnMut(Move) + 'static,

WM_MOVE message.

Sent after a window has been moved.

pub fn wm_moving<F>(&self, func: F) where
F: FnMut(Moving<'_>) + 'static,

WM_MOVING message.

Sent to a window that the user is moving. By processing this message, an application can monitor the position of the drag rectangle and, if needed, change its position.

pub fn wm_nc_calc_size<F>(&self, func: F) where
F: FnMut(NcCalcSize<'_, '_>) -> WVR + 'static,

WM_NCCALCSIZE message.

Sent when the size and position of a window's client area must be calculated. By processing this message, an application can control the content of the window's client area when the size or position of the window changes.

pub fn wm_nc_create<F>(&self, func: F) where
F: FnMut(NcCreate<'_, '_, '_>) -> bool + 'static,

WM_NCCREATE message.

Sent prior to the WM_CREATE message when a window is first created.

pub fn wm_nc_destroy<F>(&self, func: F) where
F: FnMut() + 'static,

WM_NCDESTROY message.

Notifies a window that its nonclient area is being destroyed. The DestroyWindow function sends the message to the window following the WM_DESTROY message. WM_DESTROY is used to free the allocated memory object associated with the window.

The WM_NCDESTROY message is sent after the child windows have been destroyed. In contrast, WM_DESTROY is sent before the child windows are destroyed.

Default handling

If you handle this event, you'll overwrite the default handling in:

non-dialog CustomMain;

dialog CustomMain.

pub fn wm_nc_paint<F>(&self, func: F) where
F: FnMut(NcPaint) + 'static,

WM_NCPAINT message.

Sent to a window when its frame must be painted.

Default handling

If you handle this event, you'll overwrite the default handling in:

non-dialog CustomControl;

dialog CustomControl.

pub fn wm_null<F>(&self, func: F) where
F: FnMut() + 'static,

WM_NULL message.

Performs no operation. An application sends the message if it wants to post a message that the recipient window will ignore.

pub fn wm_paint<F>(&self, func: F) where
F: FnMut() + 'static,

WM_PAINT message.

Sent when the system or another application makes a request to paint a portion of an application's window. The message is sent when the UpdateWindow or RedrawWindow function is called, or by the DispatchMessage function when the application obtains a WM_PAINT message by using the GetMessage or PeekMessage function.

pub fn wm_parent_notify<F>(&self, func: F) where
F: FnMut(ParentNotify) + 'static,

WM_PARENTNOTIFY message.

Sent to a window when a significant action occurs on a descendant window.

pub fn wm_query_open<F>(&self, func: F) where
F: FnMut(QueryOpen) -> bool + 'static,

WM_QUERYOPEN message.

Sent to an icon when the user requests that the window be restored to its previous size and position.

pub fn wm_r_button_dbl_clk<F>(&self, func: F) where
F: FnMut(RButtonDblClk) + 'static,

WM_RBUTTONDBLCLK message.

Posted when the user double-clicks the right mouse button while the cursor is in the client area of a window. If the mouse is not captured, the message is posted to the window beneath the cursor. Otherwise, the message is posted to the window that has captured the mouse.

pub fn wm_r_button_down<F>(&self, func: F) where
F: FnMut(RButtonDown) + 'static,

WM_RBUTTONDOWN message.

Posted when the user presses the right mouse button while the cursor is in the client area of a window. If the mouse is not captured, the message is posted to the window beneath the cursor. Otherwise, the message is posted to the window that has captured the mouse.

pub fn wm_r_button_up<F>(&self, func: F) where
F: FnMut(RButtonUp) + 'static,

WM_RBUTTONUP message.

Posted when the user releases the right mouse button while the cursor is in the client area of a window. If the mouse is not captured, the message is posted to the window beneath the cursor. Otherwise, the message is posted to the window that has captured the mouse.

pub fn wm_set_focus<F>(&self, func: F) where
F: FnMut(SetFocus) + 'static,

WM_SETFOCUS message.

Sent to a window after it has gained the keyboard focus.

Default handling

If you handle this event, you'll overwrite the default handling in:

non-dialog CustomMain;

non-dialog CustomModal.

pub fn wm_set_font<F>(&self, func: F) where
F: FnMut(SetFont) + 'static,

WM_SETFONT message.

Sets the font that a control is to use when drawing text.

pub fn wm_set_icon<F>(&self, func: F) where
F: FnMut(SetIcon) -> Option<HICON> + 'static,

WM_SETICON message.

Associates a new large or small icon with a window. The system displays the large icon in the Alt+TAB dialog box, and the small icon in the window caption.

pub fn wm_show_window<F>(&self, func: F) where
F: FnMut(ShowWindow) + 'static,

WM_SHOWWINDOW message.

Sent to a window when the window is about to be hidden or shown.

pub fn wm_size<F>(&self, func: F) where
F: FnMut(Size) + 'static,

WM_SIZE message.

Sent to a window after its size has changed.

Examples

use winsafe::gui::WindowMain;

let wnd: WindowMain; // initialize it somewhere...

wnd.on().wm_size({
    let wnd = wnd.clone(); // pass into the closure
    move |parms| {
        println!("HWND: {}, client area: {}x{}",
            wnd.hwnd(),
            parms.width,
            parms.height,
        );
    }
});

pub fn wm_sizing<F>(&self, func: F) where
F: FnMut(Sizing<'_>) + 'static,

WM_SIZING message.

Sent to a window that the user is resizing. By processing this message, an application can monitor the size and position of the drag rectangle and, if needed, change its size or position.

pub fn wm_style_changed<F>(&self, func: F) where
F: FnMut(StyleChanged<'_>) + 'static,

WM_STYLECHANGED message.

Sent to a window after the SetWindowLongPtr function has changed one or more of the window's styles.

pub fn wm_style_changing<F>(&self, func: F) where
F: FnMut(StyleChanging<'_>) + 'static,

WM_STYLECHANGING message.

Sent to a window when the SetWindowLongPtr function is about to change one or more of the window's styles.

pub fn wm_sys_char<F>(&self, func: F) where
F: FnMut(SysChar) + 'static,

WM_SYSCHAR message.

Posted to the window with the keyboard focus when a WM_SYSKEYDOWN message is translated by the TranslateMessage function. It specifies the character code of a system character key that is, a character key that is pressed while the ALT key is down.

pub fn wm_sys_command<F>(&self, func: F) where
F: FnMut(SysCommand) + 'static,

WM_SYSCOMMAND message.

A window receives this message when the user chooses a command from the Window menu (formerly known as the system or control menu) or when the user chooses the maximize button, minimize button, restore button, or close button.

pub fn wm_sys_dead_char<F>(&self, func: F) where
F: FnMut(SysDeadChar) + 'static,

WM_SYSDEADCHAR message.

Sent to the window with the keyboard focus when a WM_SYSKEYDOWN message is translated by the TranslateMessage function. WM_SYSDEADCHAR specifies the character code of a system dead key that is, a dead key that is pressed while holding down the ALT key.

pub fn wm_sys_key_down<F>(&self, func: F) where
F: FnMut(SysKeyDown) + 'static,

WM_SYSKEYDOWN message.

Posted to the window with the keyboard focus when the user presses the F10 key (which activates the menu bar) or holds down the ALT key and then presses another key. It also occurs when no window currently has the keyboard focus; in this case, the WM_SYSKEYDOWN message is sent to the active window. The window that receives the message can distinguish between these two contexts by checking the context code in the lParam parameter.

pub fn wm_sys_key_up<F>(&self, func: F) where
F: FnMut(SysKeyUp) + 'static,

WM_SYSKEYUP message.

Posted to the window with the keyboard focus when the user releases a key that was pressed while the ALT key was held down. It also occurs when no window currently has the keyboard focus; in this case, the WM_SYSKEYUP message is sent to the active window. The window that receives the message can distinguish between these two contexts by checking the context code in the lParam parameter.

pub fn wm_theme_changed<F>(&self, func: F) where
F: FnMut(ThemeChanged) + 'static,

WM_THEMECHANGED message.

Broadcast to every window following a theme change event. Examples of theme change events are the activation of a theme, the deactivation of a theme, or a transition from one theme to another.

pub fn wm_window_pos_changed<F>(&self, func: F) where
F: FnMut(WindowPosChanged<'_>) + 'static,

WM_WINDOWPOSCHANGED message.

Sent to a window whose size, position, or place in the Z order has changed as a result of a call to the SetWindowPos function or another window-management function.

pub fn wm_window_pos_changing<F>(&self, func: F) where
F: FnMut(WindowPosChanging<'_>) + 'static,

WM_WINDOWPOSCHANGING message.

Sent to a window whose size, position, or place in the Z order is about to change as a result of a call to the SetWindowPos function or another window-management function.

pub fn wm_x_button_dbl_clk<F>(&self, func: F) where
F: FnMut(XButtonDblClk) + 'static,

WM_XBUTTONDBLCLK message.

Posted when the user double-clicks the first or second X button while the cursor is in the client area of a window. If the mouse is not captured, the message is posted to the window beneath the cursor. Otherwise, the message is posted to the window that has captured the mouse.

pub fn wm_x_button_down<F>(&self, func: F) where
F: FnMut(XButtonDown) + 'static,

WM_XBUTTONDOWN message.

Posted when the user presses the first or second X button while the cursor is in the client area of a window. If the mouse is not captured, the message is posted to the window beneath the cursor. Otherwise, the message is posted to the window that has captured the mouse.

pub fn wm_x_button_up<F>(&self, func: F) where
F: FnMut(XButtonUp) + 'static,

WM_XBUTTONUP message.

Posted when the user releases the first or second X button while the cursor is in the client area of a window. If the mouse is not captured, the message is posted to the window beneath the cursor. Otherwise, the message is posted to the window that has captured the mouse.