Struct winsafe::gui::spec::ComboBoxItems

pub struct ComboBoxItems<'a> { /* fields omitted */ }

Exposes item methods of a ComboBox control.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl<'a> ComboBoxItems<'a>

pub fn add<S: AsRef<str>>(&self, items: &[S]) -> WinResult<()>

Adds new texts by sending cb::AddString messages.

Examples

use winsafe::prelude::*;
use winsafe::ComboBox;

let cmb_names: ComboBox; // initialized somewhere

cmb_names.items().add(&["John", "Mary"]);

pub fn count(&self) -> WinResult<u32>

Retrieves the number of items by sending a cb::GetCount message.

pub fn delete(&self, index: u32) -> WinResult<()>

Deletes the item at the given index by sending a cb::DeleteString message.

pub fn delete_all(&self)

Deletes all items by sending a cb::ResetContent message.

pub fn iter(&self) -> impl Iterator<Item = String>

Returns an iterator over the texts.

Examples

use winsafe::prelude::*;
use winsafe::gui;

let my_combo: gui::ComboBox; // initialized somewhere

for text in my_combo.items().iter() {
    println!("Text {}", text);
}

pub fn select(&self, index: Option<u32>)

Sets the currently selected index, or clears it, by sending a cb::SetCurSel message.

pub fn selected_index(&self) -> Option<u32>

Retrieves the index of the currently selected item, if any, by sending a cb::GetCurSel message.

pub fn selected_text(&self) -> Option<String>

Retrieves the currently selected text, if any, by calling selected_index and text methods.

pub fn text(&self, index: u32) -> Option<String>

Retrieves the text at the given position, if any, by sending a cb::GetLbText message.