Struct winsafe::gui::spec::ListViewColumns

pub struct ListViewColumns<'a> { /* fields omitted */ }

Exposes column methods of a ListView control.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl<'a> ListViewColumns<'a>

pub fn add(&self, texts_and_widths: &[(impl AsRef<str>, u32)]) -> WinResult<()>

Adds many columns at once by sending an lvm::InsertColumn message.

Widths will be adjusted to match current system DPI.

Examples

use winsafe::prelude::*;
use winsafe::gui;

let my_list: gui::ListView; // initialized somewhere

my_list.columns().add(&[
    ("Name", 300),
    ("Address", 500),
])?;

pub fn count(&self) -> WinResult<u32>

Retrieves the number of columns by sending an hdm::GetItemCount message to the handle returned by lvm::GetHeader.

pub fn info(&self, column_index: u32, lvc: &mut LVCOLUMN<'_>) -> WinResult<()>

Retrieves information about the column by sending an lvm::GetColumn message.

pub fn set_info(&self, column_index: u32, lvc: &LVCOLUMN<'_>) -> WinResult<()>

Sets information of the column by sending an lvm::SetColumn message.

pub fn set_title(&self, column_index: u32, text: &str) -> WinResult<()>

Sets the title of the column by calling set_info.

pub fn set_width(&self, column_index: u32, width: u32) -> WinResult<()>

Sets the width of the column by sending an lvm::SetColumnWidth message.

Width will be adjusted to match current system DPI.

pub fn set_width_to_fill(&self, column_index: u32) -> WinResult<()>

Sets the width of the column by sending an lvm::SetColumnWidth message. The width will be calculated to fill the remaining space.

pub fn title(&self, column_index: u32) -> WinResult<String>

Retrieves the title of the column by calling info.

pub fn width(&self, column_index: u32) -> WinResult<u32>

Retrieves the width of the column by sending an lvm::GetColumnWidth message.