Struct winsafe::gui::spec::TreeViewItems

pub struct TreeViewItems<'a> { /* private fields */ }

Available on crate feature gui only.

Exposes item methods of a TreeView control.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl<'a> TreeViewItems<'a>

pub fn add_root(&self, text: &str, icon_index: Option<u32>) -> TreeViewItem<'a>

Adds a new root item by sending a tvm::InsertItem message, and returns the newly added item.

pub fn delete_all(&self)

Deletes all items by sending a tvm::DeleteItem message.

pub fn count(&self) -> u32

Retrieves the total number of items by sending a tvm::GetCount message.

pub fn count_visible(&self) -> u32

Retrieves the number of visible items by sending a tvm::GetVisibleCount message.

pub fn end_edit_label_now(&self, save: bool)

Ends the editing of the item’s text by sending a tvm::EndEditLabelNow message.

pub const fn get(&self, hitem: HTREEITEM) -> TreeViewItem<'a>

Retrieves the item of the given handle.

Note: This method is cheap – even if htreeitem is invalid, an object will still be returned. However, operations upon this object will fail.

pub fn iter_selected(&self) -> impl Iterator<Item = TreeViewItem<'a>> + 'a

Returns an iterator over the selected items.

pub fn iter_root(&self) -> impl Iterator<Item = TreeViewItem<'a>> + 'a

Returns an iterator over the root items.