Struct winsafe::gui::spec::TreeViewItems

pub struct TreeViewItems { /* fields omitted */ }

Exposes item methods of a TreeView control.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl TreeViewItems

pub fn add_child(
&self,
hparent: HTREEITEM,
text: &str,
icon_index: Option<u32>
) -> WinResult<HTREEITEM>

Adds a new child item by sending a tvm::InsertItem message.

pub fn add_root(
&self,
text: &str,
icon_index: Option<u32>
) -> WinResult<HTREEITEM>

Adds a new root item by sending a tvm::InsertItem message.

pub fn children(&self, hparent: Option<HTREEITEM>) -> Vec<HTREEITEM>

Retrieves the children of the given item by sending tvm::GetNextItem messages.

To retrieve the root nodes, pass None as hparent.

pub fn delete(&self, hitem: HTREEITEM) -> WinResult<()>

Deletes an item by sending a tvm::DeleteItem message.

pub fn delete_all(&self) -> WinResult<()>

Deletes all items by sending a tvm::DeleteItem message.

pub fn count(&self) -> u32

Retrieves the total number of items by sending a tvm::GetCount message.

pub fn edit_label(&self, hitem: HTREEITEM) -> WinResult<HWND>

Begins in-place editing of the item’s text by sending a tvm::EditLabel message.

Returns a handle to the edit control.

pub fn end_edit_label_now(&self, save: bool) -> WinResult<()>

Ends the editing of the item’s text by sending a tvm::EndEditLabelNow message.

pub fn ensure_visible(&self, hitem: HTREEITEM) -> bool

Ensures that a tree-view item is visible, expanding the parent item or scrolling the tree-view control, if necessary, by sending a tvm::EnsureVisible message.

Returns whether a scroll occurred and no items were expanded.

pub fn expand(&self, hitem: HTREEITEM, expand: bool) -> WinResult<()>

Expands or collapse an item by sending a tvm::Expand message.

pub fn is_expanded(&self, hitem: HTREEITEM) -> bool

Tells if the item is expanded by sending a tvm::GetItemState message.

pub fn is_root(&self, hitem: HTREEITEM) -> bool

Tells if the item is a root by sending a tvm::GetNextItem message.

pub fn next_sibling(&self, hitem: HTREEITEM) -> Option<HTREEITEM>

Retrieves the next sibling of the item by sending a tvm::GetNextItem message.

pub fn parent(&self, hitem: HTREEITEM) -> Option<HTREEITEM>

Retrieves the parent of the item by sending a tvm::GetNextItem message.

pub fn prev_sibling(&self, hitem: HTREEITEM) -> Option<HTREEITEM>

Retrieves the previous sibling of the item by sending a tvm::GetNextItem message.

pub fn selected(&self, hitem: HTREEITEM) -> Option<HTREEITEM>

Retrieves the selected item by sending a tvm::GetNextItem message.

pub fn set_text(&self, hitem: HTREEITEM, text: &str) -> WinResult<()>

Sets the text of the item by sending a tvm::SetItem message.

pub fn text(&self, hitem: HTREEITEM) -> WinResult<String>

Retrieves the text of the item by sending a tvm::GetItem message.

Examples

use winsafe::{gui, HTREEITEM};

let my_tree: gui::TreeView; // initialized somewhere
let my_item: HTREEITEM;

println!("Text: {}", my_tree.items().text(my_item)?);