Struct winsafe::gui::CheckBox

pub struct CheckBox(_);

Native check box control, actually a variation of the ordinary Button: just a button with a specific style.

Implementations

impl CheckBox

pub fn new(parent: &impl Parent, opts: CheckBoxOpts) -> CheckBox

Instantiates a new CheckBox object, to be created on the parent window with HWND::CreateWindowEx.

pub fn new_dlg(
parent: &impl Parent,
ctrl_id: u16,
horz_resize: Horz,
vert_resize: Vert
) -> CheckBox

Instantiates a new CheckBox object, to be loaded from a dialog resource with HWND::GetDlgItem.

pub fn check_state(&self) -> CheckState

Retrieves the current check state by sending a bm::GetCheck message.

pub fn emulate_click(&self)

Emulates the click event for the check box by sending a bm::Click message.

pub fn is_checked(&self) -> bool

Calls check_state and compares the result with CheckState::Checked.

pub fn resize_to_text(&self) -> WinResult<()>

Resizes the control to exactly fit current text.

pub fn set_check_state(&self, state: CheckState)

Sets the current check state by sending a bm::SetCheck message.

pub fn set_check_state_and_trigger(&self, state: CheckState) -> WinResult<()>

Sets the current check state by sending a bm::SetCheck message, then sends a wm::Command message to the parent, so it can handle the event.

pub fn set_text(&self, text: &str) -> WinResult<()>

Sets the text by calling SetWindowText.

pub fn set_text_and_resize(&self, text: &str) -> WinResult<()>

Calls set_text and resizes the control to exactly fit the new text.

pub fn text(&self) -> WinResult<String>

Retrieves the text by calling HWND::GetWindowText.