Struct winsafe::gui::DateTimePicker

pub struct DateTimePicker(_);

Available on crate feature gui only.

Native date and time picker control.

Implementations

impl DateTimePicker

pub fn new(parent: &impl GuiParent, opts: DateTimePickerOpts) -> DateTimePicker

Instantiates a new DateTimePicker object, to be created on the parent window with HWND::CreateWindowEx.

pub fn new_dlg(
parent: &impl GuiParent,
ctrl_id: u16,
resize_behavior: (Horz, Vert)
) -> DateTimePicker

Instantiates a new DateTimePicker object, to be loaded from a dialog resource with HWND::GetDlgItem.

pub fn date(&self, st: &mut SYSTEMTIME)

Retrieves the currently selected date by sending a dtm::GetSystemTime message.

pub fn set_date(&self, st: &SYSTEMTIME)

Sets the currently selected date by sending a dtm::SetSystemTime message.