Struct winsafe::gui::Edit

pub struct Edit(_);

Available on crate feature gui only.

Native edit control.

Implementations

impl Edit

pub fn new(parent: &impl GuiParent, opts: EditOpts) -> Edit

Instantiates a new Edit object, to be created on the parent window with HWND::CreateWindowEx.

pub fn new_dlg(
parent: &impl GuiParent,
ctrl_id: u16,
resize_behavior: (Horz, Vert)
) -> Edit

Instantiates a new Edit object, to be loaded from a dialog resource with HWND::GetDlgItem.

pub fn hide_balloon_tip(&self)

Hides any balloon tip by sending an em::HideBalloonTip message.

pub fn iter_lines<'a>(&'a self) -> impl Iterator<Item = String> + 'a

Returns an iterator over the lines in the Edit.

Examples

use winsafe::prelude::*;
use winsafe::gui;

let my_edit: gui::Edit; // initialized somewhere

for line in my_edit.iter_lines() {
    println!("{}", line);
}

pub fn limit_text(&self, max_chars: Option<u32>)

Limits the number of characters that can be type by sending an /// em::SetLimitText message.

pub fn set_selection(&self, start: Option<u32>, end: Option<u32>)

Sets the selection range of the text by sending an em::SetSel message.

Examples

Selecting all text in the control:

use winsafe::prelude::*;
use winsafe::gui;

let my_edit: gui::Edit; // initialized somewhere

my_edit.set_selection(Some(0), None);

Clearing the selection:

use winsafe::gui;

let my_edit: gui::Edit; // initialized somewhere

my_edit.set_selection(None, None);

pub fn show_ballon_tip(&self, title: &str, text: &str, icon: TTI)

Displays a balloon tip by sending an em::ShowBalloonTip message.