winsafe::gui

Struct Header

pub struct Header(/* private fields */);

Available on crate feature gui only.

Native header control.

Implementations

impl Header

pub fn on_subclass(&self) -> &impl GuiEventsWindow

Exposes the subclass events. If at least one event exists, the control will be subclassed.

Note: Subclassing may impact performance, use with care.

Panics

Panics if the control or the parent window are already created. Events must be set before control and parent window creation.

pub fn on(&self) -> &impl GuiEventsHeader

Exposes the specific control events.

Panics

Panics if the control is already created. Events must be set before control creation.

impl Header

pub fn new(parent: &(impl GuiParent + 'static), opts: HeaderOpts<'_>) -> Self

Instantiates a new Header object, to be created on the parent window with HWND::CreateWindowEx.

Panics

Panics if the parent window was already created โ that is, you cannot dynamically create a Header in an event closure.

pub fn new_dlg( parent: &(impl GuiParent + 'static), ctrl_id: u16, resize_behavior: (Horz, Vert), ) -> Self

Instantiates a new Header object, to be loaded from a dialog resource with HWND::GetDlgItem.

Panics

Panics if the parent dialog was already created โ that is, you cannot dynamically create a Header in an event closure.

pub fn image_list(&self, kind: HDSIL) -> HrResult<HIMAGELIST>

Retrieves one of the associated image lists by sending an hdm::GetImageList message.

Image lists are lazy-initialized: the first time you call this method for a given image list, it will be created and assigned with hdm::SetImageList.

The image list is owned by the control.

pub const fn items(&self) -> HeaderItems<'_>

Item methods.