Struct winsafe::gui::ListView

pub struct ListView(_);

Native list view control. Not to be confused with the simpler list box control.

Implements Child trait.

Implementations

impl ListView

pub fn new(parent: &dyn Parent, opts: ListViewOpts) -> ListView

Instantiates a new ListView object, to be created on the parent window with HWND::CreateWindowEx.

pub fn new_dlg(
parent: &dyn Parent,
ctrl_id: u16,
context_menu: Option<HMENU>
) -> ListView

Instantiates a new ListView object, to be loaded from a dialog resource with HWND::GetDlgItem.

Note: The optional context_menu is shared: it must be destroyed manually after the control is destroyed. But note that menus loaded from resources don’t need to be destroyed.

pub fn hwnd(&self) -> HWND

Returns the underlying handle for this control.

Note: the handle is initially null, receiving an actual value only after the control is physically created, what usually happens right before WM_CREATE or WM_INITDIALOG events.

pub fn ctrl_id(&self) -> u16

Returns the control ID.

pub fn focus(&self) -> WinResult<()>

Focuses the control by sending a wm::NextDlgCtl message.

pub fn on_subclass(&self) -> &WindowEvents

Exposes the subclass events. If at least one event exists, the control will be subclassed.

Note: Subclassing may impact performance, use with care.

Panics

Panics if the control or the parent window are already created. Events must be set before control and parent window creation.

pub fn on(&self) -> &ListViewEvents

Exposes the control events.

These event methods are just proxies to the WindowEvents of the parent window, who is the real responsible for the child event handling.

Panics

Panics if the control or the parent window are already created. Events must be set before control and parent window creation.

pub fn columns(&self) -> &ListViewColumns

Exposes the column methods.

pub fn context_menu(&self) -> Option<HMENU>

Returns the context menu attached to this list view, if any.

The context menu is attached when the list view is created, either by calling ListView::new or ListView::new_dlg.

pub fn image_list(&self, kind: LVSIL) -> Option<HIMAGELIST>

Retrieves one of the associated image lists by sending an lvm::GetImageList message.

pub fn items(&self) -> &ListViewItems

Exposes the item methods.

pub fn current_view(&self) -> LV_VIEW

Retrieves the current view by sending an lvm::GetView message.

pub fn set_current_view(&self, view: LV_VIEW) -> WinResult<()>

Sets the current view by sending an lvm::SetView message.

pub fn set_image_list(
&self,
kind: LVSIL,
himagelist: HIMAGELIST
) -> Option<HIMAGELIST>

Sets the one of the associated image lists by sending an lvm::SetImageList message.

Returns the previous image list, if any.

pub fn toggle_extended_style(&self, set: bool, ex_style: LVS_EX)

Toggles the given extended list view styles by sending an lvm::SetExtendedListViewStyle message.