winsafe::gui

Struct ListView

pub struct ListView<T: 'static = ()>(/* private fields */);

Available on crate feature gui only.

Native list view control. Not to be confused with the simpler list box control.

The generic parameter specifies the type of the object that will be embedded on each item โ if you donโt want to store anything, just use () as the type. Internally, this storage is implemented with pointers in the itemโs LPARAM fields.

You also have access to the embedded Header control, which renders the columns.

Examples

The example below creates a main window with a list view, which has two columns.

use winsafe::{self as w, co, gui, prelude::*};

fn main() {
    if let Err(e) = MainWindow::create_and_run() {
        eprintln!("{}", e);
    }
}

#[derive(Clone)]
struct MainWindow {
    wnd: gui::WindowMain,
    lst: gui::ListView,
}

impl MainWindow {
    fn create_and_run() -> w::AnyResult<i32> {
        let wnd = gui::WindowMain::new(gui::WindowMainOpts {
            title: "Main window",
            size: gui::dpi(300, 150),
            ..Default::default()
        });
        let lst = gui::ListView::new(
            &wnd,
            gui::ListViewOpts {
                size: gui::dpi(260, 120),
                position: gui::dpi(10, 10),
                columns: &[("First", 100), ("Second", 100)],
                ..Default::default()
            },
        );

        let new_self = Self { wnd, lst };
        new_self.events();

        new_self.wnd.run_main(None)
    }

    fn events(&self) {
        let self2 = self.clone();
        self.wnd.on().wm_create(move |_| {
            self2.lst.items().add(&["Hello", "World"], None, ())?;
            self2.lst.items().add(&["Another", "Line"], None, ())?;
            self2.lst.items().add(&["Third", "Row"], None, ())?;
            Ok(0)
        });

        let self2 = self.clone();
        self.lst.on().lvn_item_changed(move |_| {
            let num_selec = self2.lst.items().selected_count();
            let text = format!("{num_selec} selected");
            self2.wnd.hwnd().SetWindowText(&text)?;
            Ok(())
        });
    }
}

Implementations

impl<T> ListView<T>

pub fn on_subclass(&self) -> &impl GuiEventsWindow

Exposes the subclass events. If at least one event exists, the control will be subclassed.

Note: Subclassing may impact performance, use with care.

Panics

Panics if the control or the parent window are already created. Events must be set before control and parent window creation.

pub fn on(&self) -> &impl GuiEventsListView

Exposes the specific control events.

Panics

Panics if the control is already created. Events must be set before control creation.

impl<T> ListView<T>

pub fn new(parent: &(impl GuiParent + 'static), opts: ListViewOpts<'_>) -> Self

Instantiates a new ListView object, to be created on the parent window with HWND::CreateWindowEx.

Panics

Panics if the parent window was already created โ that is, you cannot dynamically create a ListView in an event closure.

pub fn new_dlg( parent: &(impl GuiParent + 'static), ctrl_id: u16, resize_behavior: (Horz, Vert), context_menu_id: Option<u16>, ) -> Self

Instantiates a new ListView object, to be loaded from a dialog resource with HWND::GetDlgItem.

The context_menu_id must point to the root menu whose first submenu will be effectively displayed as the context menu.

Since the image list is always managed by the control itself, LVS::SHAREIMAGELISTS style will always be added.

Panics

Panics if the parent dialog was already created โ that is, you cannot dynamically create a ListView in an event closure.

Panics if the context menu, when specified, does not exist.

pub const fn cols(&self) -> ListViewCols<'_, T>

Column methods.

pub fn context_menu(&self) -> Option<HMENU>

The context submenu is optionally set in ListView::new or ListView::new_dlg. If set, this method returns a handle to its first submenu, because the first submenu is the one to be effectively displayed by the control.

pub fn current_view(&self) -> LV_VIEW

Retrieves the current view by sending an lvm::GetView message.

pub fn header(&self) -> Option<&Header>

Returns the embedded Header of the list view, if any.

The Header is tried to be initialized during the internal wm::Create and wm::InitDialog processing. If the list view doesnโt happen to have a header โ only report view lists have headers โ, the internal Header is removed, and this method will forever return None.

This means you may call header() to add events to the internal Header, but if the list view has no header, these events will simply be discarded.

Examples

use winsafe::{self as w, prelude::*, gui};

let my_list: gui::ListView; // initialized somewhere

let my_list2 = my_list.clone();
my_list.header().unwrap().on().hdn_item_click(move |p| {
    let h = my_list2.header().unwrap().items().get(p.iItem as _);
    println!("Col: {}", h.text());
    Ok(())
});

pub fn image_list(&self, kind: LVSIL) -> HrResult<HIMAGELIST>

Retrieves one of the associated image lists by sending an lvm::GetImageList message.

Image lists are lazy-initialized: the first time you call this method for a given image list, it will be created and assigned with lvm::SetImageList.

The image list is owned by the control.

pub const fn items(&self) -> ListViewItems<'_, T>

Item methods.

pub fn set_current_view(&self, view: LV_VIEW) -> SysResult<()>

Sets the current view by sending an lvm::SetView message.

pub fn set_extended_style(&self, set: bool, ex_style: LVS_EX)

Sets or unsets the given extended list view styles by sending an lvm::SetExtendedListViewStyle message.

pub fn set_redraw(&self, can_redraw: bool)

Allows or disallows the redrawing of the control by sending a wm::SetRedraw message.