Struct winsafe::gui::ListViewItems

pub struct ListViewItems { /* fields omitted */ }

Exposes item methods of a ListView control.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl ListViewItems

pub fn add(&self, text: &str, icon_index: Option<u32>) -> WinResult<u32>

Appends a new item by sending an LVM_INSERTITEM message, and returns its index.

pub fn count(&self) -> u32

Retrieves the total number of items by sending an LVM_GETITEMCOUNT message.

pub fn delete(&self, item_indexes: &[u32]) -> WinResult<()>

Deletes the items at the given indexes by sending an LVM_DELETEITEM message

pub fn delete_all(&self) -> WinResult<()>

Deletes all items by sending an LVM_DELETEALLITEMS message.

pub fn ensure_visible(&self, item_index: u32) -> WinResult<()>

Scrolls the list by sending an LVM_ENSUREVISIBLE message so that an item is visible in the list.

pub fn focused(&self) -> Option<u32>

Retrieves the index of the focused item by sending an LVM_GETNEXTITEM message

pub fn is_focused(&self, item_index: u32) -> bool

Tells if the item is the focused one by sending an LVM_GETITEMSTATE message.

pub fn is_selected(&self, item_index: u32) -> bool

Tells if the item is selected by sending an LVM_GETITEMSTATE message.

pub fn is_visible(&self, item_index: u32) -> bool

Tells if the item is currently visible by sending an LVM_ISITEMVISIBLE message.

pub fn selected(&self) -> Vec<u32>

Retrieves the indexes of the selected items by sending LVM_GETNEXTITEM messages.

pub fn selected_count(&self) -> u32

Retrieves the number of selected items by sending an LVM_GETSELECTEDCOUNT message.

pub fn set_focused(&self, item_index: u32) -> WinResult<()>

Sets the focused item by sending an LVM_SETITEMSTATE message.

pub fn set_selected(&self, set: bool, item_indexes: &[u32]) -> WinResult<()>

Sets or remove the selection from the given item indexes by sending LVM_SETITEMSTATE messages.

pub fn set_selected_all(&self, set: bool) -> WinResult<()>

Sets or remove the selection for all items by sending an LVM_SETITEMSTATE message.

pub fn set_text(
&self,
item_index: u32,
column_index: u32,
text: &str
) -> WinResult<()>

Sets the text of an item under a column by sending an LVM_SETITEMTEXT message.

pub fn text(&self, item_index: u32, column_index: u32) -> String

Retrieves the text of an item under a column by sending an LVM_GETITEMTEXT message.