Struct winsafe::gui::RadioGroup

pub struct RadioGroup(_);

A group of native RadioButton controls.

Implementations

impl RadioGroup

pub fn new(parent: &dyn Parent, opts: &[RadioButtonOpts]) -> RadioGroup

Instantiates a new RadioGroup object, each RadioButton to be created on the parent window with CreateWindowEx.

Panics

Panics if no options are passed.

pub fn new_dlg(parent: &dyn Parent, ctrl_ids: &[u16]) -> RadioGroup

Instantiates a new RadioGroup object, to be loaded from a dialog resource with GetDlgItem.

Panics

Panics if no control IDs are passed.

pub fn on(&self) -> &RadioGroupEvents

Exposes the radio group events.

These event methods are just proxies to the WindowEvents of the parent window, who is the real responsible for the child event handling.

Panics

Panics if the control or the parent window are already created. Events must be set before control and parent window creation.

pub fn iter(&self) -> Iter<'_, RadioButton>

Returns an iterator over the internal RadioButton slice.

Example

Changing the text of all radio buttons to "One":

use winsafe::gui::RadioGroup;

let radio_group: RadioGroup; // initialized somewhere

for single_radio in me.rads.iter() {
    single_radio.hwnd().SetWindowText("One").unwrap();
}

pub fn checked(&self) -> Option<&RadioButton>

Returns the currently checked RadioButton of this group, if any.

pub fn count(&self) -> usize

Returns the number of RadioButton controls in this group.