winsafe::gui

Struct TreeViewItem

pub struct TreeViewItem<'a, T: 'static = ()> { /* private fields */ }

Available on crate feature gui only.

A single item of a TreeView control.

Each object keeps an unique HTREEITEM handle.

You cannot directly instantiate this object, it is created internally by the control.

Implementations

impl<'a, T> TreeViewItem<'a, T>

pub fn add_child( &self, text: &str, icon_index: Option<u32>, data: T, ) -> SysResult<Self>

Adds a new child item by sending a tvm::InsertItem message, and returns the newly added item.

pub fn data(&self) -> Rc<RefCell<T>>

Returns a Rc/RefCell with the stored data by sending an lvm::GetItem message.

Panics

Panics if the TreeView doesnโt have an actual type, that is, if it was declared as TreeView<()>.

Panics if the item index is invalid.

pub fn delete(&self) -> SysResult<()>

Deletes the item by sending a tvm::DeleteItem message.

pub fn edit_label(&self) -> SysResult<HWND>

Begins in-place editing of the itemโs text by sending a tvm::EditLabel message.

Returns a handle to the edit control.

pub fn ensure_visible(&self) -> bool

Ensures that a tree-view item is visible, expanding the parent item or scrolling the tree-view control, if necessary, by sending a tvm::EnsureVisible message.

Returns whether a scroll occurred and no items were expanded.

pub fn expand(&self, expand: bool) -> SysResult<()>

Expands or collapse the item by sending a tvm::Expand message.

pub const fn htreeitem(&self) -> &HTREEITEM

Returns the underlying handle of the item.

pub fn is_expanded(&self) -> bool

Tells if the item is expanded by sending a tvm::GetItemState message.

pub fn is_root(&self) -> bool

Tells if the item is a root by sending a tvm::GetNextItem message.

pub fn iter_children(&self) -> impl Iterator<Item = TreeViewItem<'a, T>> + 'a

Returns an iterator over the child items.

pub fn iter_next_siblings( &self, ) -> impl Iterator<Item = TreeViewItem<'a, T>> + 'a

Returns an iterator over the next sibling items.

pub fn iter_prev_siblings( &self, ) -> impl Iterator<Item = TreeViewItem<'a, T>> + 'a

Returns an iterator over the previous sibling items.

pub fn parent(&self) -> Option<Self>

Retrieves the parent of the item by sending a tvm::GetNextItem message.

pub fn set_text(&self, text: &str) -> SysResult<()>

Sets the text of the item by sending a tvm::SetItem message.

pub fn text(&self) -> SysResult<String>

Retrieves the text of the item by sending a tvm::GetItem message.