Trait winsafe::gui::Parent

pub trait Parent {
    fn hwnd_ref(&self) -> &HWND;

    fn user_events_ref(&self) -> &WindowEvents;

    fn privileged_events_ref(&self) -> &WindowEvents;
}

Trait to any window which can host child controls.

Required methods

fn hwnd_ref(&self) -> &HWND

Returns a reference to the window handle.

fn user_events_ref(&self) -> &WindowEvents

Returns a reference to the user events object.

When an user event is added, it will overwrite the previous one.

fn privileged_events_ref(&self) -> &WindowEvents

Returns a reference to the privileged events object, which is used internally to automate some tasks.

All privileged events are executed, and their result is ignored.

Loading content...