Module winsafe::msg

Parameters of window messages.

Structs

BcmGetIdealSize
BCM_GETIDEALSIZE message parameters.

BcmGetImageList
BCM_GETIMAGELIST message parameters.

BcmGetNote
BCM_GETNOTE message parameters.

BcmGetNoteLength
BCM_GETNOTELENGTH message, which has no parameters.

BcmGetSplitInfo
BCM_GETSPLITINFO message parameters.

BcmGetTextMargin
BCM_GETTEXTMARGIN message parameters.

BcmSetDropDownState
BCM_SETDROPDOWNSTATE message parameters.

BcmSetImageList
BCM_SETIMAGELIST message parameters.

BcmSetNote
BCM_SETNOTE message parameters.

BcmSetShield
BCM_SETSHIELD message parameters.

BcmSetSplitInfo
BCM_SETSPLITINFO message parameters.

BcmSetTextMargin
BCM_SETTEXTMARGIN message parameters.

BmClick
BM_CLICK message, which has no parameters.

BmGetCheck
BM_GETCHECK message parameters.

BmGetImage
BM_GETIMAGE message parameters.

BmGetState
BM_GETSTATE message, which has no parameters.

BmSetCheck
BM_SETCHECK message parameters.

BmSetDontClick
BM_SETDONTCLICK message parameters.

BmSetImage
BM_SETIMAGE message parameters.

BmSetState
BM_SETSTATE message parameters.

BmSetStyle
BM_SETSTYLE message parameters.

CbAddString
CB_ADDSTRING message parameters.

CbDeleteString
CB_DELETESTRING message parameters.

CbGetCount
CB_GETCOUNT message, which has no parameters.

CbGetCurSel
CB_GETCURSEL message, which has no parameters.

CbGetLbText
CB_GETLBTEXT message parameters.

CbGetLbTextLen
CB_GETLBTEXTLEN message parameters.

CbGetMinVisible
CB_GETMINVISIBLE message, which has no parameters.

CbGetTopIndex
CB_GETTOPINDEX message parameters.

CbInitStorage
CB_INITSTORAGE message parameters.

CbResetContent
CB_RESETCONTENT message, which has no parameters.

CbSetCurSel
CB_SETCURSEL message parameters.

CbSetMinVisible
CB_SETMINVISIBLE message parameters.

CbSetTopIndex
CB_SETTOPINDEX message parameters.

CbShowDropDown
CB_SHOWDROPDOWN message parameters.

HdmGetItemCount
HDM_GETITEMCOUNT message, which has no parameters.

LvmDeleteAllItems
LVM_DELETEALLITEMS message, which has no parameters.

LvmDeleteItem
LVM_DELETEITEM message parameters.

LvmEnsureVisible
LVM_ENSUREVISIBLE message parameters.

LvmGetColumn
LVM_GETCOLUMN message parameters.

LvmGetColumnWidth
LVM_GETCOLUMNWIDTH message parameters.

LvmGetExtendedListViewStyle
LVM_GETEXTENDEDLISTVIEWSTYLE message, which has no parameters.

LvmGetHeader
LVM_GETHEADER message, which has no parameters.

LvmGetItemCount
LVM_GETITEMCOUNT message, which has no parameters.

LvmGetItemState
LVM_GETITEMSTATE message parameters.

LvmGetItemText
LVM_GETITEMTEXT message parameters.

LvmGetNextItem
LVM_GETNEXTITEM message parameters.

LvmGetSelectedCount
LVM_GETSELECTEDCOUNT message, which has no parameters.

LvmGetView
LVM_GETVIEW message, which has no parameters.

LvmInsertColumn
LVM_INSERTCOLUMN message parameters.

LvmInsertItem
LVM_INSERTITEM message parameters.

LvmIsGroupViewEnabled
LVM_ISGROUPVIEWENABLED message, which has no parameters.

LvmIsItemVisible
LVM_ISITEMVISIBLE message parameters.

LvmRedrawItems
LVM_REDRAWITEMS message parameters.

LvmScroll
LVM_SCROLL message parameters.

LvmSetColumn
LVM_SETCOLUMN message parameters.

LvmSetExtendedListViewStyle
LVM_SETEXTENDEDLISTVIEWSTYLE message parameters.

LvmSetItem
LVM_SETITEM message parameters.

LvmSetItemState
LVM_SETITEMSTATE message parameters.

LvmSetItemText
LVM_SETITEMTEXT message parameters.

LvmSetSelectedColumn
LVM_SETSELECTEDCOLUMN message parameters.

LvmSetView
LVM_SETVIEW message parameters.

LvmUpdate
LVM_UPDATE message parameters.

SbGetIcon
SB_GETICON message parameters.

SbGetParts
SB_GETPARTS message parameters.

SbGetText
SB_GETTEXT message parameters.

SbGetTextLength
SB_GETTEXTLENGTH message parameters.

SbGetTipText
SB_GETTIPTEXT message parameters.

SbSetIcon
SB_SETICON

SbSetParts
SB_SETPARTS message parameters.

SbSetText
SB_SETTEXT message parameters.

SbSetTipText
SB_SETTIPTEXT message parameters.

StmGetIcon
STM_GETICON message, which has no parameters.

StmSetIcon
STM_SETICON message parameters.

Wm
Generic window message parameters.

WmActivate
WM_ACTIVATE message parameters.

WmActivateApp
WM_ACTIVATEAPP message parameters.

WmAppCommand
WM_APPCOMMAND message parameters.

WmCancelMode
WM_CANCELMODE message, which has no parameters.

WmChildActivate
WM_CHILDACTIVATE message, which has no parameters.

WmClose
WM_CLOSE message, which has no parameters.

WmCommand
WM_COMMAND message parameters.

WmContextMenu
WM_CONTEXTMENU message parameters.

WmCreate
WM_CREATE message parameters.

WmCtlColorBtn
WM_CTLCOLORBTN message parameters.

WmCtlColorDlg
WM_CTLCOLORDLG message parameters.

WmCtlColorEdit
WM_CTLCOLOREDIT message parameters.

WmCtlColorListBox
WM_CTLCOLORLISTBOX message parameters.

WmCtlColorScrollBar
WM_CTLCOLORSCROLLBAR message parameters.

WmCtlColorStatic
WM_CTLCOLORSTATIC message parameters.

WmDestroy
WM_DESTROY message, which has no parameters.

WmDropFiles
WM_DROPFILES message parameters.

WmEnable
WM_ENABLE message parameters.

WmEndSession
WM_ENDSESSION message parameters.

WmEnterIdle
WM_ENTERIDLE message parameters.

WmEnterSizeMove
WM_ENTERSIZEMOVE message, which has no parameters.

WmEraseBkgnd
WM_ERASEBKGND message parameters.

WmExitSizeMove
WM_EXITSIZEMOVE message, which has no parameters.

WmGetMinMaxInfo
WM_GETMINMAXINFO message parameters.

WmInitDialog
WM_INITDIALOG message parameters.

WmInitMenuPopup
WM_INITMENUPOPUP message parameters.

WmLButtonDblClk
WM_LBUTTONDBLCLK message parameters.

WmLButtonDown
WM_LBUTTONDOWN message parameters.

WmLButtonUp
WM_LBUTTONUP message parameters.

WmMButtonDblClk
WM_MBUTTONDBLCLK message parameters.

WmMButtonDown
WM_MBUTTONDOWN message parameters.

WmMButtonUp
WM_MBUTTONUP message parameters.

WmMouseHover
WM_MOUSEHOVER message parameters.

WmMouseMove
WM_MOUSEMOVE message parameters.

WmMove
WM_MOVE message parameters.

WmMoving
WM_MOVING message parameters.

WmNcCalcSize
WM_NCCALCSIZE message parameters.

WmNcCreate
WM_NCCREATE message parameters.

WmNcDestroy
WM_NCDESTROY message, which has no parameters.

WmNcPaint
WM_NCPAINT message parameters.

WmNotify
WM_NOTIFY message parameters.

WmNull
WM_NULL message, which has no parameters.

WmPaint
WM_PAINT message, which has no parameters.

WmQueryOpen
WM_QUERYOPEN message, which has no parameters.

WmRButtonDblClk
WM_RBUTTONDBLCLK message parameters.

WmRButtonDown
WM_RBUTTONDOWN message parameters.

WmRButtonUp
WM_RBUTTONUP message parameters.

WmSetFocus
WM_SETFOCUS message parameters.

WmSetFont
WM_SETFONT message parameters.

WmSetIcon
WM_SETICON message parameters.

WmShowWindow
WM_SHOWWINDOW message parameters.

WmSize
WM_SIZE message parameters.

WmSizing
WM_SIZING message parameters.

WmStyleChanged
WM_STYLECHANGED message parameters.

WmStyleChanging
WM_STYLECHANGING message parameters.

WmThemeChanged
WM_THEMECHANGED message, which has no parameters.

WmTimer
WM_TIMER message parameters.

WmWindowPosChanged
WM_WINDOWPOSCHANGED message parameters.

WmWindowPosChanging
WM_WINDOWPOSCHANGING message parameters.

WmXButtonDblClk
WM_XBUTTONDBLCLK message parameters.

WmXButtonDown
WM_XBUTTONDOWN message parameters.

WmXButtonUp
WM_XBUTTONUP message parameters.

Traits

Message
Trait to the parameters of a message that can be sent.

MessageHandleable
Trait to the parameters of a message that can be sent and handled.