Module winsafe::msg

Parameters of window messages.

WndMsg is the generic message, with WPARAM and LPARAM fields. Other messages belong to a module according to its prefix, for example, BM_CLICK can be found in bm module.

Examples

We want to delete the 3rd element of a ListView control. This can be done by sending it an LVM_DELETEITEM message via SendMessage. The message itself is a struct, which is initialized with the specific message parameters.

The message struct also defines the data type returned by SendMessage. In the example below, LVM_DELETEITEM returns WinResult<()>.

use winsafe::{HWND, msg::lvm};

let hlistview: HWND; // initialize it somewhere...

hlistview.SendMessage({
    lvm::DeleteItem {
        index: 2,
    },
}).expect("Failed to delete item 2.");

Messages are organized into modules according to their prefixes: wm (window messages), lvm (list view messages), and so on.

Modules

bm
Button control messages, whose constants have BM and BCM prefixes.

cb
Combo box control messages, whose constants have CB prefix.

dtm
Date and time picker control messages, whose constants have DTM prefix.

em
Edit control messages, whose constants have EM prefix.

hdm
Header control messages, whose constants have HDM prefix.

lb
List box control messages, whose constants have LB prefix.

lvm
List view control messages, whose constants have LVM prefix.

mcm
Month calendar control messages, whose constants have MCM prefix.

sb
Status bar control messages, whose constants have SB prefix.

stm
Static control messages, whose constants have STM prefix.

wm
Generic window messages, whose constants have WM prefix.

Structs

WndMsg
Generic window message parameters: WPARAM and LPARAM.

Traits

MsgSend
Trait to the parameters of a message that can be sent. Implemented by all messages.

MsgSendRecv
Trait to the parameters of a message that can be sent and handled. Implemented by WndMsg and all msg::wm messages.