Trait winsafe::msg::Message

pub trait Message {
    type RetType;
    fn convert_ret(&self, v: isize) -> Self::RetType;

    fn as_generic_wm(&self) -> Wm;
}

Trait to the parameters of a message that can be sent.

Allows the conversion to the generic Wm parameters, and also defines the return type of the message.

Used in functions like SendMessage and DefWindowProc.

Associated Types

type RetType

The specific type of the value returned by the message.

Loading content...
Required methods

fn convert_ret(&self, v: isize) -> Self::RetType

Converts the generic isize return value to the specific type returned by the message.

fn as_generic_wm(&self) -> Wm

Converts the message parameters to the generic Wm parameters.

Loading content...