Trait winsafe::msg::MessageHandleable

pub trait MessageHandleable: Message {
    fn from_generic_wm(parm: Wm) -> Self;
}

Trait to the parameters of a message that can be sent and handled.

Allows the conversion from and to the generic Wm parameters, and also defines the return type of the message.

Used in functions like SendMessage and DefWindowProc.

Required methods

fn from_generic_wm(parm: Wm) -> Self

Converts the generic Wm parameters into the message parameters.

Loading content...