Trait winsafe::msg::MsgSend

pub trait MsgSend {
    type RetType;
    fn convert_ret(&self, v: isize) -> Self::RetType;

    fn as_generic_wm(&mut self) -> WndMsg;
}

Trait to the parameters of a message that can be sent. Implemented by all messages.

Allows the conversion to the generic WndMsg parameters, and also defines the return type of the message.

Used in functions like SendMessage and DefWindowProc.

Associated Types

type RetType

The specific type of the value returned by the message.

Required methods

fn convert_ret(&self, v: isize) -> Self::RetType

Converts the generic isize return value to the specific type returned by the message.

fn as_generic_wm(&mut self) -> WndMsg

Converts the specific message parameters struct into the generic WndMsg message struct.