Trait winsafe::msg::MsgSendRecv

pub trait MsgSendRecv: MsgSend {
    fn from_generic_wm(parm: WndMsg) -> Self;
}

Trait to the parameters of a message that can be sent and handled. Implemented by WndMsg and all msg::wm messages.

Allows the conversion from and to the generic WndMsg parameters, and also defines the return type of the message.

Required methods

fn from_generic_wm(parm: WndMsg) -> Self

Converts the generic WndMsg parameters struct into the specific message struct.

Loading content...