Module winsafe::path

File path utilities.

Some of the functions are similar to std::path::Path ones, but here they work directly upon &str instead of &OsStr.

Functions

exe_path

Returns the path of the current EXE file, without the EXE filename, and without a trailing backslash.

get_file_name

Extracts the file name from a full path, if any.

get_path

Extracts the full path, but the last part.

has_extension

Tells whether the full path ends in one of the given extensions, case-insensitive.

iter

Returns an iterator over each part of the path.

replace_extension

Replaces the extension by the given one.

replace_file_name

Replaces the file name by the given one.

replace_path

Keeps the file name and replaces the path by the given one.

rtrim_backslash

Removes a trailing backslash, if any.