Module winsafe::prelude

The WinSafe prelude.

The purpose of this module is to alleviate imports of many common traits. To use it, add a glob import to the top of all your modules that use the library:

use winsafe::prelude::*;

Traits

FormattedErrorkernel

A system error which can be formatted with FormatMessage.

GdiObjectgdi

This trait is enabled with the gdi feature, and implements methods for any HGDIOBJ handle, which is the base handle for GDI objects.

GuiChildgui

Any child window.

GuiChildFocusgui

Any child window which can be focused.

GuiEventsgui

Exposes the basic window message methods of WindowEvents.

GuiNativeControlgui

Any native control, which can be subclassed.

GuiNativeControlEventsgui

Events of a native control.

GuiParentgui

Any window which can host child controls.

GuiTabgui

Any window which can be rendered inside a native Tab control.

GuiThreadgui

Allows a window to spawn new threads which can return errors, and run closures in the original UI thread.

GuiWindowgui

Any window. Exposes the underlying window handle.

GuiWindowTextgui

Any window which can get/set text.

Handlekernel

A native handle, implemented by all handle types.

MsgSenduser

Parameters of a message which can be sent. Implemented by all undefined messages.

MsgSendRecvuser

Parameters of a message which can be sent and handled (received). Implemented by WndMsg and all msg::wm messages.

NativeBitflagkernel

A native typed bitflag constant.

NativeConstkernel

A native typed constant.

advapi_Hkeyadvapi

This trait is enabled with the advapi feature, and provides methods for HKEY.

advapi_ktm_Hkeyadvapi and ktm

This trait is enabled with advapi and ktm features, and provides methods for HKEY.

comctl_Himagelistcomctl

This trait is enabled with the comctl feature, and provides methods for HIMAGELIST.

comctl_Hwndcomctl

This trait is enabled with the comctl feature, and provides methods for HWND.

comctl_ole_Hwndcomctl and ole

This trait is enabled with comctl and ole features, and provides methods for HWND.

comctl_shell_Himagelistcomctl and shell

This trait is enabled with comctl and shell features, and provides methods for HIMAGELIST.

dshow_IBaseFilterdshow

This trait is enabled with the dshow feature, and provides methods for IBaseFilter.

dshow_IEnumFiltersdshow

This trait is enabled with the dshow feature, and provides methods for IEnumFilters.

dshow_IEnumMediaTypesdshow

This trait is enabled with the dshow feature, and provides methods for IEnumMediaTypes.

dshow_IFileSinkFilterdshow

This trait is enabled with the dshow feature, and provides methods for IFileSinkFilter.

dshow_IFilterGraphdshow

This trait is enabled with the dshow feature, and provides methods for IFilterGraph.

dshow_IGraphBuilderdshow

This trait is enabled with the dshow feature, and provides methods for IGraphBuilder.

dshow_IMFGetServicedshow

This trait is enabled with the dshow feature, and provides methods for IMFGetService.

dshow_IMFVideoDisplayControldshow

This trait is enabled with the dshow feature, and provides methods for IMFVideoDisplayControl.

dshow_IMediaControldshow

This trait is enabled with the dshow feature, and provides methods for IMediaControl.

dshow_IMediaFilterdshow

This trait is enabled with the dshow feature, and provides methods for IMediaFilter.

dshow_IMediaSeekingdshow

This trait is enabled with the dshow feature, and provides methods for IMediaSeeking.

dshow_IPindshow

This trait is enabled with the dshow feature, and provides methods for IPin.

dshow_gdi_IMFVideoDisplayControldshow and gdi

This trait is enabled with dshow and gdi features, and provides methods for IMFVideoDisplayControl.

gdi_Hbitmapgdi

This trait is enabled with the gdi feature, and provides methods for HBITMAP.

gdi_Hbrushgdi

This trait is enabled with the gdi feature, and provides methods for HBRUSH.

gdi_Hdcgdi

This trait is enabled with the gdi feature, and provides methods for HDC.

gdi_Hfontgdi

This trait is enabled with the gdi feature, and provides methods for HFONT.

gdi_Hinstancegdi

This trait is enabled with the gdi feature, and provides methods for HINSTANCE.

gdi_Hpengdi

This trait is enabled with the gdi feature, and provides methods for HPEN.

gdi_Hrgngdi

This trait is enabled with the gdi feature, and provides methods for HRGN.

gdi_ole_Hdcgdi and ole

This trait is enabled with gdi and ole features, and provides methods for HDC.

gdi_ole_IPicturegdi and ole

This trait is enabled with gdi and ole features, and provides methods for IPicture.

kernel_Haccesstokenkernel

This trait is enabled with the kernel feature, and provides methods for HACCESSTOKEN.

kernel_Hfilekernel

This trait is enabled with the kernel feature, and provides methods for HFILE.

kernel_Hfilemapkernel

This trait is enabled with the kernel feature, and provides methods for HFILEMAP.

kernel_Hfilemapviewkernel

This trait is enabled with the kernel feature, and provides methods for HFILEMAPVIEW.

kernel_Hfindfilekernel

This trait is enabled with the kernel feature, and provides methods for HFINDFILE.

kernel_Hglobalkernel

This trait is enabled with the kernel feature, and provides methods for HGLOBAL.

kernel_Hinstancekernel

This trait is enabled with the kernel feature, and provides methods for HINSTANCE.

kernel_Hlocalkernel

This trait is enabled with the kernel feature, and provides methods for HLOCAL.

kernel_Hpipekernel

This trait is enabled with the kernel feature, and provides methods for HPIPE.

kernel_Hprocesskernel

This trait is enabled with the kernel feature, and provides methods for HPROCESS.

kernel_Hprocesslistkernel

This trait is enabled with the kernel feature, and provides methods for HPROCESSLIST.

kernel_Hthreadkernel

This trait is enabled with the kernel feature, and provides methods for HTHREAD.

kernel_Hupdatersrckernel

This trait is enabled with the kernel feature, and provides methods for HUPDATERSRC.

ktm_Htransactionktm

This trait is enabled with the ktm feature, and provides methods for HTRANSACTION.

msimg_Hdcmsimg

This trait is enabled with the msimg feature, and provides methods for HDC.

ole_Hwndole

This trait is enabled with the ole feature, and provides methods for HWND.

ole_IBindCtxole

This trait is enabled with the ole feature, and provides methods for IBindCtx.

ole_IDataObjectole

This trait is enabled with the ole feature, and provides methods for IDataObject.

ole_IDropTargetole

This trait is enabled with the ole feature, and provides methods for IDropTarget.

ole_IPersistole

This trait is enabled with the ole feature, and provides methods for IPersist.

ole_IPictureole

This trait is enabled with the ole feature, and provides methods for IPicture.

ole_ISequentialStreamole

This trait is enabled with the ole feature, and provides methods for ISequentialStream.

ole_IStreamole

IStream methods from ole feature.

ole_IUnknownole

This trait is enabled with the ole feature, and provides methods for IUnknown. It is the base trait for all COM traits.

oleaut_IDispatcholeaut

This trait is enabled with the oleaut feature, and provides methods for IDispatch.

oleaut_IPictureoleaut

This trait is enabled with the oleaut feature, and provides methods for IDispatch.

oleaut_IPropertyStoreoleaut

This trait is enabled with the oleaut feature, and provides methods for IPropertyStore.

oleaut_ITypeInfooleaut

This trait is enabled with the oleaut feature, and provides methods for ITypeInfo.

oleaut_Variantoleaut

Methods common to VARIANT and PROPVARIANT structs.

shell_Hdropshell

This trait is enabled with the shell feature, and provides methods for HDROP.

shell_Hwndshell

This trait is enabled with the shell feature, and provides methods for HWND.

shell_IEnumShellItemsshell

This trait is enabled with the shell feature, and provides methods for IEnumShellItems.

shell_IFileDialogshell

This trait is enabled with the shell feature, and provides methods for IFileDialog.

shell_IFileOpenDialogshell

This trait is enabled with the shell feature, and provides methods for IFileOpenDialog.

shell_IFileSaveDialogshell

This trait is enabled with the shell feature, and provides methods for IFileSaveDialog.

shell_IModalWindowshell

This trait is enabled with the shell feature, and provides methods for IModalWindow.

shell_IShellItemshell

This trait is enabled with the shell feature, and provides methods for IShellItem.

shell_IShellItem2shell

This trait is enabled with the shell feature, and provides methods for IShellItem2.

shell_IShellItemArrayshell

This trait is enabled with the shell feature, and provides methods for IShellItemArray.

shell_IShellLinkshell

This trait is enabled with the shell feature, and provides methods for IShellLink.

shell_IStreamshell

This trait is enabled with the shell feature, and provides methods for IStream.

shell_ITaskbarListshell

This trait is enabled with the shell feature, and provides methods for ITaskbarList.

shell_ITaskbarList2shell

This trait is enabled with the shell feature, and provides methods for ITaskbarList2.

shell_ITaskbarList3shell

This trait is enabled with the shell feature, and provides methods for ITaskbarList3.

shell_ITaskbarList4shell

This trait is enabled with the shell feature, and provides methods for ITaskbarList4.

user_Hacceluser

This trait is enabled with the user feature, and provides methods for HACCEL.

user_Hcursoruser

This trait is enabled with the user feature, and provides methods for HCURSOR.

user_Hdcuser

This trait is enabled with the user feature, and provides methods for HDC.

user_Hdwpuser

This trait is enabled with the user feature, and provides methods for HDWP.

user_Hhookuser

This trait is enabled with the user feature, and provides methods for HHOOK.

user_Hiconuser

This trait is enabled with the user feature, and provides methods for HICON.

user_Hinstanceuser

This trait is enabled with the user feature, and provides methods for HINSTANCE.

user_Hmenuuser

This trait is enabled with the user feature, and provides methods for HMENU.

user_Hmonitoruser

This trait is enabled with the user feature, and provides methods for HMONITOR.

user_Hprocessuser

This trait is enabled with the user feature, and provides methods for HPROCESS.

user_Hwnduser

This trait is enabled with the user feature, and provides methods for HWND.

uxtheme_Hthemeuxtheme

This trait is enabled with the uxtheme feature, and provides methods for HTHEME.

uxtheme_Hwnduxtheme

This trait is enabled with the uxtheme feature, and provides methods for HWND.