Module winsafe::prelude

Commonly imported traits.

Intended to be used as:

use winsafe::prelude::*;

Re-exports

pub use super::events::prelude::*;

Traits

Child

Any child window.

Focus

Focus a child control with wm::NextDlgCtl message.

NativeControl

Any native child control.

NativeControlEvents

Exposes the native control events.

Parent

Trait to any window which can host child controls.

ParentEvents

Exposes all parent window events and control notifications through WindowEventsAll.

UiThread

Allows running code in the original UI thread.

Window

Any window. Exposes the underlying window handle.