Module winsafe::prelude

The WinSafe prelude.

The purpose of this module is to alleviate imports of many common traits by adding a glob import to the top of your module:

use winsafe::prelude::*;

Traits

AsAny

Used to convert a reference to the Any trait.

ComInterface

Any COM object, which encapsulates a COM interface pointer.

ComctlHimagelist

HIMAGELIST methods from comctl feature.

ComctlHwnd

HWND methods from comctl feature.

ComctlOleHwnd

HWND methods from comctl+ole features.

ComctlShellHimagelist

HIMAGELIST methods from comctl+shell features.

DshowGdiIMFVideoDisplayControl

IMFVideoDisplayControl methods from dshow+gdi feature.

DshowIBaseFilter

IBaseFilter methods from dshow feature.

DshowIEnumFilters

IEnumFilters methods from dshow feature.

DshowIFileSinkFilter

IFileSinkFilter methods from dshow feature.

DshowIFilterGraph

IFilterGraph methods methods from dshow feature.

DshowIGraphBuilder

IGraphBuilder methods methods from dshow feature.

DshowIMFGetService

IMFGetService methods methods from dshow feature.

DshowIMFVideoDisplayControl

IMFVideoDisplayControl methods from dshow feature.

DshowIMediaControl

IMediaControl methods from dshow feature.

DshowIMediaFilter

IMediaFilter methods methods from dshow feature.

DshowIMediaSeeking

IMediaSeeking methods methods from dshow feature.

DshowIPin

IPin methods methods from dshow feature.

FormattedError

A system error which can be formatted with FormatMessage.

GdiHbitmap

HBITMAP methods from gdi feature.

GdiHbrush

HBRUSH methods from gdi feature.

GdiHdc

HDC methods from gdi feature.

GdiHfont

HFONT methods from gdi feature.

GdiHpen

HPEN methods from gdi feature.

GdiHrgn

HRGN methods from gdi feature.

GdiOleautHdc

HDC methods from gdi+oleaut feature.

GdiOleautIPicture

IPicture methods from gdi+oleaut feature.

GuiChild

Any child window.

GuiEventsView

Exposes the methods of WindowEvents.

GuiEventsViewAll

Exposes the methods of WindowEventsAll.

GuiFocusControl

Any child window which can be focused.

GuiNativeControl

Any native control, which can be subclassed.

GuiNativeControlEvents

Events of a native control.

GuiParent

Any window which can host child controls.

GuiTextControl

Any child window which can get/set text.

GuiThread

Allows a window to spawn new threads which can return errors, and run closures in the original UI thread.

GuiWindow

Any window. Exposes the underlying window handle.

Handle

A native handle.

HandleClose

A Handle which can be closed.

HandleGdi

Any HGDIOBJ handle, which is a specific handle for GDI objects.

KernelHfile

HFILE methods from kernel feature.

KernelHfilemap

HFILEMAP methods from kernel feature.

KernelHfilemapview

HFILEMAPVIEW methods from kernel feature.

KernelHfindfile

HFILEMAPVIEW methods from kernel feature.

KernelHglobal

HGLOBAL methods from kernel feature.

KernelHinstance

HINSTANCE methods from kernel feature.

KernelHlocal

HLOCAL methods from kernel feature.

KernelHpipe

HPIPE methods from kernel feature.

KernelHprocess

HPROCESS methods from kernel feature.

KernelHprocesslist

HPROCESSLIST methods from kernel feature.

KernelHthread

HTHREAD methods from kernel feature.

KernelHupdatersrc

HUPDATERSRC methods from kernel feature.

MsgSend

Trait to the parameters of a message that can be sent. Implemented by all messages.

MsgSendRecv

Trait to the parameters of a message that can be sent and handled. Implemented by WndMsg and all msg::wm messages.

MsimgHdc

HDC methods from msimg feature.

NativeBitflag

A native bitflag constant.

OleIUnknown

IUnknown methods from ole feature.

OleautIDispatch

IDispatch methods from oleaut feature.

OleautIPicture

IPicture methods from oleaut feature.

OleautITypeInfo

ITypeInfo methods from oleaut feature.

ShellHdrop

HDROP methods from shell feature.

ShellHwnd

HWND methods from shell feature.

ShellIFileDialog

IFileDialog methods from shell feature.

ShellIFileOpenDialog

IFileOpenDialog methods from shell feature.

ShellIFileSaveDialog

IFileSaveDialog methods from shell feature.

ShellIModalWindow

IModalWindow methods from shell feature.

ShellIShellItem

IShellItem methods from shell feature.

ShellIShellItemArray

IShellItemArray methods from shell feature.

ShellITaskbarList

ITaskbarList methods from shell feature.

ShellITaskbarList2

ITaskbarList2 methods from shell feature.

ShellITaskbarList3

ITaskbarList3 methods from shell feature.

ShellITaskbarList4

ITaskbarList4 methods from shell feature.

ShlwapiIPersist

IPersist methods from shlwapi feature.

ShlwapiISequentialStream

ISequentialStream methods from shlwapi feature.

ShlwapiIStream

IStream methods from shlwapi feature.

UserHaccel

HACCEL methods from user feature.

UserHcursor

HCURSOR methods from user feature.

UserHdc

HDC methods from user feature.

UserHdwp

HDWP methods from user feature.

UserHhook

HHOOK methods from user feature.

UserHicon

HICON methods from user feature.

UserHinstance

HINSTANCE methods from user feature.

UserHmenu

HMENU methods from user feature.

UserHmonitor

HMONITOR methods from user feature.

UserHprocess

HPROCESS methods from user feature.

UserHwnd

HWND methods from user feature.

UxthemeHtheme

HTHEME methods from uxtheme feature.

UxthemeHwnd

HWND methods from uxtheme feature.