Module winsafe::prelude

The WinSafe prelude.

The purpose of this module is to alleviate imports of many common traits by adding a glob import to the top of your module:

use winsafe::prelude::*;

Traits

AdvapiHkeyadvapi

HKEY methods from advapi feature.

ComInterfaceole

Any COM object, which encapsulates a COM interface pointer.

ComctlHimagelistcomctl

HIMAGELIST methods from comctl feature.

ComctlHwndcomctl

HWND methods from comctl feature.

ComctlOleHwndcomctl and ole

HWND methods from comctl+ole features.

ComctlShellHimagelistcomctl and shell

HIMAGELIST methods from comctl+shell features.

DshowGdiIMFVideoDisplayControldshow and gdi

IMFVideoDisplayControl methods from dshow+gdi feature.

DshowIBaseFilterdshow

IBaseFilter methods from dshow feature.

DshowIEnumFiltersdshow

IEnumFilters methods from dshow feature.

DshowIFileSinkFilterdshow

IFileSinkFilter methods from dshow feature.

DshowIFilterGraphdshow

IFilterGraph methods methods from dshow feature.

DshowIGraphBuilderdshow

IGraphBuilder methods methods from dshow feature.

DshowIMFGetServicedshow

IMFGetService methods methods from dshow feature.

DshowIMFVideoDisplayControldshow

IMFVideoDisplayControl methods from dshow feature.

DshowIMediaControldshow

IMediaControl methods from dshow feature.

DshowIMediaFilterdshow

IMediaFilter methods methods from dshow feature.

DshowIMediaSeekingdshow

IMediaSeeking methods methods from dshow feature.

DshowIPindshow

IPin methods methods from dshow feature.

FormattedErrorkernel

A system error which can be formatted with FormatMessage.

GdiHbitmapgdi

HBITMAP methods from gdi feature.

GdiHbrushgdi

HBRUSH methods from gdi feature.

GdiHdcgdi

HDC methods from gdi feature.

GdiHfontgdi

HFONT methods from gdi feature.

GdiHpengdi

HPEN methods from gdi feature.

GdiHrgngdi

HRGN methods from gdi feature.

GdiOleautHdc

HDC methods from gdi+oleaut feature.

GdiOleautIPicturegdi and oleaut

IPicture methods from gdi+oleaut feature.

GuiChildgui

Any child window.

GuiChildFocusgui

Any child window which can be focused.

GuiEventsgui

Exposes the basic window message methods of WindowEvents.

GuiNativeControlgui

Any native control, which can be subclassed.

GuiNativeControlEventsgui

Events of a native control.

GuiParentgui

Any window which can host child controls.

GuiThreadgui

Allows a window to spawn new threads which can return errors, and run closures in the original UI thread.

GuiWindowgui

Any window. Exposes the underlying window handle.

GuiWindowTextgui

Any window which can get/set text.

Handlekernel

A native handle.

HandleClosekernel

A Handle which can be closed.

HandleGdigdi

Any HGDIOBJ handle, which is a specific handle for GDI objects.

KernelHaccesstokenkernel

HACCESSTOKEN methods from kernel feature.

KernelHfilekernel

HFILE methods from kernel feature.

KernelHfilemapkernel

HFILEMAP methods from kernel feature.

KernelHfilemapviewkernel

HFILEMAPVIEW methods from kernel feature.

KernelHfindfilekernel

HFILEMAPVIEW methods from kernel feature.

KernelHglobalkernel

HGLOBAL methods from kernel feature.

KernelHinstancekernel

HINSTANCE methods from kernel feature.

KernelHlocalkernel

HLOCAL methods from kernel feature.

KernelHpipekernel

HPIPE methods from kernel feature.

KernelHprocesskernel

HPROCESS methods from kernel feature.

KernelHprocesslistkernel

HPROCESSLIST methods from kernel feature.

KernelHthreadkernel

HTHREAD methods from kernel feature.

KernelHupdatersrckernel

HUPDATERSRC methods from kernel feature.

MsgSenduser

Trait to the parameters of a message that can be sent. Implemented by all messages.

MsgSendRecvuser

Trait to the parameters of a message that can be sent and handled. Implemented by WndMsg and all msg::wm messages.

MsimgHdcmsimg

HDC methods from msimg feature.

NativeBitflagkernel

A native bitflag constant.

OleIUnknownole

IUnknown methods from ole feature.

OleautIDispatcholeaut

IDispatch methods from oleaut feature.

OleautIPictureoleaut

IPicture methods from oleaut feature.

OleautITypeInfooleaut

ITypeInfo methods from oleaut feature.

ShellHdropshell

HDROP methods from shell feature.

ShellHwndshell

HWND methods from shell feature.

ShellIFileDialogshell

IFileDialog methods from shell feature.

ShellIFileOpenDialogshell

IFileOpenDialog methods from shell feature.

ShellIFileSaveDialogshell

IFileSaveDialog methods from shell feature.

ShellIModalWindowshell

IModalWindow methods from shell feature.

ShellIShellItemshell

IShellItem methods from shell feature.

ShellIShellItemArrayshell

IShellItemArray methods from shell feature.

ShellIShellLinkshell

IShellLink methods from shell feature.

ShellITaskbarListshell

ITaskbarList methods from shell feature.

ShellITaskbarList2shell

ITaskbarList2 methods from shell feature.

ShellITaskbarList3shell

ITaskbarList3 methods from shell feature.

ShellITaskbarList4shell

ITaskbarList4 methods from shell feature.

ShlwapiIBindCtxshlwapi

IBindCtx methods from shlwapi feature.

ShlwapiIPersistshlwapi

IPersist methods from shlwapi feature.

ShlwapiISequentialStreamshlwapi

ISequentialStream methods from shlwapi feature.

ShlwapiIStreamshlwapi

IStream methods from shlwapi feature.

UserHacceluser

HACCEL methods from user feature.

UserHcursoruser

HCURSOR methods from user feature.

UserHdcuser

HDC methods from user feature.

UserHdwpuser

HDWP methods from user feature.

UserHhookuser

HHOOK methods from user feature.

UserHiconuser

HICON methods from user feature.

UserHinstanceuser

HINSTANCE methods from user feature.

UserHmenuuser

HMENU methods from user feature.

UserHmonitoruser

HMONITOR methods from user feature.

UserHprocessuser

HPROCESS methods from user feature.

UserHwnduser

HWND methods from user feature.

UxthemeHthemeuxtheme

HTHEME methods from uxtheme feature.

UxthemeHwnduxtheme

HWND methods from uxtheme feature.