Module winsafe::prelude

The WinSafe prelude.

The purpose of this module is to alleviate imports of many common traits by adding a glob import to the top of your module:

use winsafe::prelude::*;

Traits

AsAny

Used to convert a reference to the Any trait.

Child

Any child window.

ComInterface

Any COM object, which encapsulates a COM interface pointer.

EventsView

Exposes the methods of WindowEvents.

EventsViewAll

Exposes the methods of WindowEventsAll.

FocusControl

Any child window which can be focused.

FormattedError

A system error which can be formatted with FormatMessage.

Handle

Any Windows handle.

HandleClose

Any Handle which can be closed.

HandleGdi

Any HGDIOBJ handle, which is a specific handle for GDI objects.

IBaseFilterT

Exposes the IBaseFilter methods.

IDispatchT

Exposes the IDispatch methods.

IEnumFiltersT

Exposes the IEnumFilters methods.

IFileDialogT

Exposes the IFileDialog methods.

IFileOpenDialogT

Exposes the IFileOpenDialog methods.

IFileSaveDialogT

Exposes the IFileSaveDialog methods.

IFileSinkFilterT

Exposes the IFileSinkFilter methods.

IFilterGraphT

Exposes the IFilterGraph methods.

IGraphBuilderT

Exposes the IGraphBuilder methods.

IMFGetServiceT

Exposes the IMFGetService methods.

IMFVideoDisplayControlT

Exposes the IMFVideoDisplayControl methods.

IMediaControlT

Exposes the IMediaControl methods.

IMediaFilterT

Exposes the IMediaFilter methods.

IMediaSeekingT

Exposes the IMediaSeeking methods.

IModalWindowT

Exposes the IModalWindow methods.

IPersistT

Exposes the IPersist methods.

IPictureT

Exposes the IPicture methods.

IPinT

Exposes the IPin methods.

IShellItemArrayT

Exposes the IShellItemArray methods.

IShellItemT

Exposes the IShellItem methods.

ITaskbarList2T

Exposes the ITaskbarList2 methods.

ITaskbarList3T

Exposes the ITaskbarList3 methods.

ITaskbarList4T

Exposes the ITaskbarList4 methods.

ITaskbarListT

Exposes the ITaskbarList methods.

ITypeInfoT

Exposes the ITypeInfo methods.

IUnknownT

Exposes the IUnknown methods.

NativeConstant

Any native Windows constant.

NativeControl

Any native control, which can be subclassed.

NativeControlEvents

Events of a native control.

Parent

Any window which can host child controls.

TextControl

Any child window which can get/set text.

UiThread

Allows a window to spawn new threads which can return errors, and run closures in the original UI thread.

Window

Any window. Exposes the underlying window handle.