winsafe

Module prelude

Available on crate feature kernel only.

The WinSafe prelude.

The purpose of this module is to alleviate imports of many common traits. To use it, add a glob import to the top of all your modules that use the library:

use winsafe::prelude::*;

Traits

GdiObjectgdi

This trait is enabled with the gdi feature, and implements methods for any HGDIOBJ handle, which is the base handle for GDI objects.

GuiControlgui

Any child control.

GuiParentgui

Any window which can host child controls.

GuiWindowgui

Any window. Exposes the underlying window handle.

Handle

A native handle, implemented by all handle types.

MsgSenduser

Parameters of a message which can be sent. Implemented by all defined messages.

MsgSendRecvuser

Parameters of a message which can be sent and handled (received). Implemented by WndMsg and all msg::wm messages.

SystemError

A system error which can be formatted with FormatMessage, exhibiting a description string provided by the OS.

advapi_Haccesstokenadvapi

This trait is enabled with the advapi feature, and provides methods for HACCESSTOKEN.

advapi_Heventlogadvapi

This trait is enabled with the advapi feature, and provides methods for HEVENTLOG.

advapi_Hkeyadvapi

This trait is enabled with the advapi feature, and provides methods for HKEY.

advapi_Hprocessadvapi

This trait is enabled with the advapi feature, and provides methods for HPROCESS.

advapi_Hscadvapi

This trait is enabled with the advapi feature, and provides methods for HSC.

advapi_Hserviceadvapi

This trait is enabled with the advapi feature, and provides methods for HSERVICE.

advapi_Hservicestatusadvapi

This trait is enabled with the advapi feature, and provides methods for HSERVICESTATUS.

advapi_Hthreadadvapi

This trait is enabled with the advapi feature, and provides methods for HTHREAD.

advapi_Htransactionadvapi

This trait is enabled with the advapi feature, and provides methods for HTRANSACTION.

comctl_Himagelistcomctl

This trait is enabled with the comctl feature, and provides methods for HIMAGELIST.

comctl_Hpropsheetpagecomctl

This trait is enabled with the comctl feature, and provides methods for HPROPSHEETPAGE.

comctl_Hwndcomctl

This trait is enabled with the comctl feature, and provides methods for HWND.

comctl_gdi_Himagelistcomctl and gdi

This trait is enabled with comctl and gdi features, and provides methods for HIMAGELIST.

comctl_shell_Himagelistcomctl and shell

This trait is enabled with comctl and shell features, and provides methods for HIMAGELIST.

dshow_IBaseFilterdshow

This trait is enabled with the dshow feature, and provides methods for IBaseFilter.

dshow_IEnumFiltersdshow

This trait is enabled with the dshow feature, and provides methods for IEnumFilters.

dshow_IEnumMediaTypesdshow

This trait is enabled with the dshow feature, and provides methods for IEnumMediaTypes.

dshow_IEnumPinsdshow

This trait is enabled with the dshow feature, and provides methods for IEnumPins.

dshow_IFileSinkFilterdshow

This trait is enabled with the dshow feature, and provides methods for IFileSinkFilter.

dshow_IFilterGraphdshow

This trait is enabled with the dshow feature, and provides methods for IFilterGraph.

dshow_IFilterGraph2dshow

This trait is enabled with the dshow feature, and provides methods for IFilterGraph2.

dshow_IGraphBuilderdshow

This trait is enabled with the dshow feature, and provides methods for IGraphBuilder.

dshow_IMediaControldshow

This trait is enabled with the dshow feature, and provides methods for IMediaControl.

dshow_IMediaFilterdshow

This trait is enabled with the dshow feature, and provides methods for IMediaFilter.

dshow_IMediaSeekingdshow

This trait is enabled with the dshow feature, and provides methods for IMediaSeeking.

dshow_IPindshow

This trait is enabled with the dshow feature, and provides methods for IPin.

dwm_Hwnddwm

This trait is enabled with the dwm feature, and provides methods for HWND.

dxgi_IDXGIAdapterdxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGIAdapter.

dxgi_IDXGIAdapter1dxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGIAdapter1.

dxgi_IDXGIAdapter2dxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGIAdapter2.

dxgi_IDXGIDevicedxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGIDevice.

dxgi_IDXGIDeviceSubObjectdxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGIDeviceSubObject.

dxgi_IDXGIFactorydxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGIFactory.

dxgi_IDXGIFactory1dxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGIFactory1.

dxgi_IDXGIKeyedMutexdxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGIKeyedMutex.

dxgi_IDXGIObjectdxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGIObject.

dxgi_IDXGIOutputdxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGIOutput.

dxgi_IDXGIResourcedxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGIResource.

dxgi_IDXGISurfacedxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGISurface.

dxgi_IDXGISwapChaindxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGISwapChain.

gdi_Hbitmapgdi

This trait is enabled with the gdi feature, and provides methods for HBITMAP.

gdi_Hbrushgdi

This trait is enabled with the gdi feature, and provides methods for HBRUSH.

gdi_Hdcgdi

This trait is enabled with the gdi feature, and provides methods for HDC.

gdi_Hfontgdi

This trait is enabled with the gdi feature, and provides methods for HFONT.

gdi_Hinstancegdi

This trait is enabled with the gdi feature, and provides methods for HINSTANCE.

gdi_Hpalettegdi

This trait is enabled with the gdi feature, and provides methods for HFONT.

gdi_Hpengdi

This trait is enabled with the gdi feature, and provides methods for HPEN.

gdi_Hrgngdi

This trait is enabled with the gdi feature, and provides methods for HRGN.

gdi_mf_IMFVideoDisplayControlgdi and mf

This trait is enabled with gdi and mf features, and provides methods for IMFVideoDisplayControl.

kernel_Hevent

This trait is enabled with the kernel feature, and provides methods for HEVENT.

kernel_Hfile

This trait is enabled with the kernel feature, and provides methods for HFILE.

kernel_Hfilemap

This trait is enabled with the kernel feature, and provides methods for HFILEMAP.

kernel_Hfilemapview

This trait is enabled with the kernel feature, and provides methods for HFILEMAPVIEW.

kernel_Hfindfile

This trait is enabled with the kernel feature, and provides methods for HFINDFILE.

kernel_Hglobal

This trait is enabled with the kernel feature, and provides methods for HGLOBAL.

kernel_Hheap

This trait is enabled with the kernel feature, and provides methods for HHEAP.

kernel_Hinstance

This trait is enabled with the kernel feature, and provides methods for HINSTANCE.

kernel_Hlocal

This trait is enabled with the kernel feature, and provides methods for HLOCAL.

kernel_Hpipe

This trait is enabled with the kernel feature, and provides methods for HPIPE.

kernel_Hprocess

This trait is enabled with the kernel feature, and provides methods for HPROCESS.

kernel_Hprocesslist

This trait is enabled with the kernel feature, and provides methods for HPROCESSLIST.

kernel_Hstd

This trait is enabled with the kernel feature, and provides methods for HSTD.

kernel_Hthread

This trait is enabled with the kernel feature, and provides methods for HTHREAD.

kernel_Hupdatersrc

This trait is enabled with the kernel feature, and provides methods for HUPDATERSRC.

mf_IMFAsyncResultmf

This trait is enabled with the mf feature, and provides methods for IMFAsyncResult.

mf_IMFAttributesmf

This trait is enabled with the mf feature, and provides methods for IMFAttributes.

mf_IMFByteStreammf

This trait is enabled with the mf feature, and provides methods for IMFByteStream.

mf_IMFClockmf

This trait is enabled with the mf feature, and provides methods for IMFClock.

mf_IMFCollectionmf

This trait is enabled with the mf feature, and provides methods for IMFCollection.

mf_IMFGetServicemf

This trait is enabled with the mf feature, and provides methods for IMFGetService.

mf_IMFMediaEventmf

This trait is enabled with the mf feature, and provides methods for IMFMediaEvent.

mf_IMFMediaEventGeneratormf

This trait is enabled with the mf feature, and provides methods for IMFMediaEventGenerator.

mf_IMFMediaSessionmf

This trait is enabled with the mf feature, and provides methods for IMFMediaSession.

mf_IMFMediaSourcemf

This trait is enabled with the mf feature, and provides methods for IMFMediaSource.

mf_IMFMediaTypeHandlermf

This trait is enabled with the mf feature, and provides methods for IMFMediaTypeHandler.

mf_IMFPresentationDescriptormf

This trait is enabled with the mf feature, and provides methods for IMFPresentationDescriptor.

mf_IMFSourceResolvermf

This trait is enabled with the mf feature, and provides methods for IMFSourceResolver.

mf_IMFStreamDescriptormf

This trait is enabled with the mf feature, and provides methods for IMFStreamDescriptor.

mf_IMFTopologymf

This trait is enabled with the mf feature, and provides methods for IMFTopology.

mf_IMFTopologyNodemf

This trait is enabled with the mf feature, and provides methods for IMFTopologyNode.

mf_IMFVideoDisplayControlmf

This trait is enabled with the mf feature, and provides methods for IMFVideoDisplayControl.

ole_Hwndole

This trait is enabled with the ole feature, and provides methods for HWND.

ole_IAdviseSinkole

This trait is enabled with the ole feature, and provides methods for IAdviseSink.

ole_IBindCtxole

This trait is enabled with the ole feature, and provides methods for IBindCtx.

ole_IDataObjectole

This trait is enabled with the ole feature, and provides methods for IDataObject.

ole_IMonikerole

This trait is enabled with the ole feature, and provides methods for IMoniker.

ole_IPersistole

This trait is enabled with the ole feature, and provides methods for IPersist.

ole_IPersistFileole

This trait is enabled with the ole feature, and provides methods for IPersistFile.

ole_IPersistStreamole

This trait is enabled with the ole feature, and provides methods for IPersistStream.

ole_IPictureole

This trait is enabled with the ole feature, and provides methods for IPicture.

ole_ISequentialStreamole

This trait is enabled with the ole feature, and provides methods for ISequentialStream.

ole_IStorageole

This trait is enabled with the ole feature, and provides methods for IStorage.

ole_IStreamole

IStream methods from ole feature.

ole_IUnknownole

This trait is enabled with the ole feature, and provides methods for IUnknown. It is the base trait for all COM traits.

oleaut_IDispatcholeaut

This trait is enabled with the oleaut feature, and provides methods for IDispatch.

oleaut_IPropertyStoreoleaut

This trait is enabled with the oleaut feature, and provides methods for IPropertyStore.

oleaut_ITypeInfooleaut

This trait is enabled with the oleaut feature, and provides methods for ITypeInfo.

psapi_Hprocesspsapi

This trait is enabled with the psapi feature, and provides methods for HPROCESS.

shell_Hdropshell

This trait is enabled with the shell feature, and provides methods for HDROP.

shell_Hwndshell

This trait is enabled with the shell feature, and provides methods for HWND.

shell_IEnumShellItemsshell

This trait is enabled with the shell feature, and provides methods for IEnumShellItems.

shell_IFileDialogshell

This trait is enabled with the shell feature, and provides methods for IFileDialog.

shell_IFileOpenDialogshell

This trait is enabled with the shell feature, and provides methods for IFileOpenDialog.

shell_IFileOperationshell

This trait is enabled with the shell feature, and provides methods for IFileOperation.

shell_IFileSaveDialogshell

This trait is enabled with the shell feature, and provides methods for IFileSaveDialog.

shell_IModalWindowshell

This trait is enabled with the shell feature, and provides methods for IModalWindow.

shell_IShellItemshell

This trait is enabled with the shell feature, and provides methods for IShellItem.

shell_IShellItem2shell

This trait is enabled with the shell feature, and provides methods for IShellItem2.

shell_IShellItemArrayshell

This trait is enabled with the shell feature, and provides methods for IShellItemArray.

shell_IShellLinkshell

This trait is enabled with the shell feature, and provides methods for IShellLink.

shell_ITaskbarListshell

This trait is enabled with the shell feature, and provides methods for ITaskbarList.

shell_ITaskbarList2shell

This trait is enabled with the shell feature, and provides methods for ITaskbarList2.

shell_ITaskbarList3shell

This trait is enabled with the shell feature, and provides methods for ITaskbarList3.

shell_ITaskbarList4shell

This trait is enabled with the shell feature, and provides methods for ITaskbarList4.

taskschd_IActiontaskschd

This trait is enabled with the taskschd feature, and provides methods for IAction.

taskschd_IActionCollectiontaskschd

This trait is enabled with the taskschd feature, and provides methods for IActionCollection.

taskschd_IBootTriggertaskschd

This trait is enabled with the taskschd feature, and provides methods for IBootTrigger.

taskschd_IComHandlerActiontaskschd

This trait is enabled with the taskschd feature, and provides methods for IComHandlerAction.

taskschd_IDailyTriggertaskschd

This trait is enabled with the taskschd feature, and provides methods for IDailyTrigger.

taskschd_IEmailActiontaskschd

This trait is enabled with the taskschd feature, and provides methods for IEmailAction.

taskschd_IEventTriggertaskschd

This trait is enabled with the taskschd feature, and provides methods for IEventTrigger.

taskschd_IExecActiontaskschd

This trait is enabled with the taskschd feature, and provides methods for IExecAction.

taskschd_IIdleTriggertaskschd

This trait is enabled with the taskschd feature, and provides methods for IIdleTrigger.

taskschd_ILogonTriggertaskschd

This trait is enabled with the taskschd feature, and provides methods for ILogonTrigger.

taskschd_IRegisteredTasktaskschd

This trait is enabled with the taskschd feature, and provides methods for IRegisteredTask.

taskschd_IRegistrationInfotaskschd

This trait is enabled with the taskschd feature, and provides methods for IRegistrationInfo.

taskschd_ITaskDefinitiontaskschd

This trait is enabled with the taskschd feature, and provides methods for ITaskDefinition.

taskschd_ITaskFoldertaskschd

This trait is enabled with the taskschd feature, and provides methods for ITaskFolder.

taskschd_ITaskServicetaskschd

This trait is enabled with the taskschd feature, and provides methods for ITaskService.

taskschd_ITriggertaskschd

This trait is enabled with the taskschd feature, and provides methods for ITriggerCollection.

taskschd_ITriggerCollectiontaskschd

This trait is enabled with the taskschd feature, and provides methods for ITriggerCollection.

user_Hacceluser

This trait is enabled with the user feature, and provides methods for HACCEL.

user_Hclipboarduser

This trait is enabled with the user feature, and provides methods for HCLIPBOARD.

user_Hcursoruser

This trait is enabled with the user feature, and provides methods for HCURSOR.

user_Hdcuser

This trait is enabled with the user feature, and provides methods for HDC.

user_Hdeskuser

This trait is enabled with the user feature, and provides methods for HDESK.

user_Hdwpuser

This trait is enabled with the user feature, and provides methods for HDWP.

user_Hhookuser

This trait is enabled with the user feature, and provides methods for HHOOK.

user_Hiconuser

This trait is enabled with the user feature, and provides methods for HICON.

user_Hinstanceuser

This trait is enabled with the user feature, and provides methods for HINSTANCE.

user_Hmenuuser

This trait is enabled with the user feature, and provides methods for HMENU.

user_Hmonitoruser

This trait is enabled with the user feature, and provides methods for HMONITOR.

user_Hprocessuser

This trait is enabled with the user feature, and provides methods for HPROCESS.

user_Hwnduser

This trait is enabled with the user feature, and provides methods for HWND.

uxtheme_Hthemeuxtheme

This trait is enabled with the uxtheme feature, and provides methods for HTHEME.

uxtheme_Hwnduxtheme

This trait is enabled with the uxtheme feature, and provides methods for HWND.

version_Hversioninfoversion

This trait is enabled with the version feature, and provides methods for HVERSIONINFO.

winspool_Hprinterwinspool

This trait is enabled with the winspool feature, and provides methods for HPRINTER.

winspool_Hwndwinspool

This trait is enabled with the winspool feature, and provides methods for HWND.