winsafe

Module prelude

The WinSafe prelude.

The purpose of this module is to alleviate imports of many common traits. To use it, add a glob import to the top of all your modules that use the library:

use winsafe::prelude::*;

Traits

GdiObjectgdi

This trait is enabled with the gdi feature, and implements methods for any HGDIOBJ handle, which is the base handle for GDI objects.

GuiControlgui

This trait is enabled with the gui feature, and is implemented by all child controls.

GuiEventsButtongui

This trait is enabled with the gui feature, and exposes button control notifications.

GuiEventsComboBoxgui

This trait is enabled with the gui feature, and exposes combo box control notifications.

GuiEventsDateTimePickergui

This trait is enabled with the gui feature, and exposes date and time picker control notifications.

GuiEventsEditgui

This trait is enabled with the gui feature, and exposes edit control notifications.

GuiEventsHeadergui

This trait is enabled with the gui feature, and exposes header control notifications.

GuiEventsLabelgui

This trait is enabled with the gui feature, and exposes label control notifications.

GuiEventsListBoxgui

This trait is enabled with the gui feature, and exposes list box control notifications.

GuiEventsListViewgui

This trait is enabled with the gui feature, and exposes list view control notifications.

GuiEventsMonthCalendargui

This trait is enabled with the gui feature, and exposes month calendar control notifications.

GuiEventsParentgui

This trait is enabled with the gui feature, and provides events for window notifications.

GuiEventsPropSheetPagegui

This trait is enabled with the gui feature, and exposes property sheet page control notifications.

GuiEventsRadioGroupgui

This trait is enabled with the gui feature, and exposes button control notifications for a RadioGroup.

GuiEventsStatusBargui

This trait is enabled with the gui feature, and exposes status bar control notifications.

GuiEventsTabgui

This trait is enabled with the gui feature, and exposes tab control notifications.

GuiEventsTrackbargui

This trait is enabled with the gui feature, and exposes trackbar control notifications.

GuiEventsTreeViewgui

This trait is enabled with the gui feature, and exposes tree view control notifications.

GuiEventsUpDowngui

This trait is enabled with the gui feature, and exposes up-down control notifications.

GuiEventsWindowgui

This trait is enabled with the gui feature, and provides events for subclassed native controls.

GuiParentgui

This trait is enabled with the gui feature, and is implemented by all windows which can host child controls.

GuiWindowgui

This trait is enabled with the gui feature, and is implemented by all windows. Exposes the underlying window handle.

Handlekernel

A native handle, implemented by all handle types.

MsgSenduser

This trait is enabled with the gui feature, and is implemented by the parameters of messages which can be sent. Implemented by all defined messages.

MsgSendRecvuser

This trait is enabled with the gui feature, and is implemented by the parameters of messages which can be sent and handled (received). Implemented by WndMsg and all msg::wm messages.

SystemErrorkernel

This trait is enabled with the kernel feature, and is implemented by all system error types which can be formatted with FormatMessage, exhibiting a description string provided by the OS.

dshow_IBaseFilterdshow

This trait is enabled with the dshow feature, and provides methods for IBaseFilter.

dshow_IEnumFiltersdshow

This trait is enabled with the dshow feature, and provides methods for IEnumFilters.

dshow_IEnumMediaTypesdshow

This trait is enabled with the dshow feature, and provides methods for IEnumMediaTypes.

dshow_IEnumPinsdshow

This trait is enabled with the dshow feature, and provides methods for IEnumPins.

dshow_IFileSinkFilterdshow

This trait is enabled with the dshow feature, and provides methods for IFileSinkFilter.

dshow_IFilterGraphdshow

This trait is enabled with the dshow feature, and provides methods for IFilterGraph.

dshow_IFilterGraph2dshow

This trait is enabled with the dshow feature, and provides methods for IFilterGraph2.

dshow_IGraphBuilderdshow

This trait is enabled with the dshow feature, and provides methods for IGraphBuilder.

dshow_IMediaControldshow

This trait is enabled with the dshow feature, and provides methods for IMediaControl.

dshow_IMediaFilterdshow

This trait is enabled with the dshow feature, and provides methods for IMediaFilter.

dshow_IMediaSeekingdshow

This trait is enabled with the dshow feature, and provides methods for IMediaSeeking.

dshow_IPindshow

This trait is enabled with the dshow feature, and provides methods for IPin.

dxgi_IDXGIAdapterdxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGIAdapter.

dxgi_IDXGIAdapter1dxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGIAdapter1.

dxgi_IDXGIAdapter2dxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGIAdapter2.

dxgi_IDXGIDevicedxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGIDevice.

dxgi_IDXGIDeviceSubObjectdxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGIDeviceSubObject.

dxgi_IDXGIFactorydxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGIFactory.

dxgi_IDXGIFactory1dxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGIFactory1.

dxgi_IDXGIFactory2dxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGIFactory2.

dxgi_IDXGIKeyedMutexdxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGIKeyedMutex.

dxgi_IDXGIObjectdxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGIObject.

dxgi_IDXGIOutputdxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGIOutput.

dxgi_IDXGIResourcedxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGIResource.

dxgi_IDXGISurfacedxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGISurface.

dxgi_IDXGISwapChaindxgi

This trait is enabled with the dxgi feature, and provides methods for IDXGISwapChain.

gdi_mf_IMFVideoDisplayControlgdi and mf

This trait is enabled with gdi and mf features, and provides methods for IMFVideoDisplayControl.

mf_IMFAsyncResultmf

This trait is enabled with the mf feature, and provides methods for IMFAsyncResult.

mf_IMFAttributesmf

This trait is enabled with the mf feature, and provides methods for IMFAttributes.

mf_IMFByteStreammf

This trait is enabled with the mf feature, and provides methods for IMFByteStream.

mf_IMFClockmf

This trait is enabled with the mf feature, and provides methods for IMFClock.

mf_IMFCollectionmf

This trait is enabled with the mf feature, and provides methods for IMFCollection.

mf_IMFGetServicemf

This trait is enabled with the mf feature, and provides methods for IMFGetService.

mf_IMFMediaEventmf

This trait is enabled with the mf feature, and provides methods for IMFMediaEvent.

mf_IMFMediaEventGeneratormf

This trait is enabled with the mf feature, and provides methods for IMFMediaEventGenerator.

mf_IMFMediaSessionmf

This trait is enabled with the mf feature, and provides methods for IMFMediaSession.

mf_IMFMediaSourcemf

This trait is enabled with the mf feature, and provides methods for IMFMediaSource.

mf_IMFMediaTypeHandlermf

This trait is enabled with the mf feature, and provides methods for IMFMediaTypeHandler.

mf_IMFPresentationDescriptormf

This trait is enabled with the mf feature, and provides methods for IMFPresentationDescriptor.

mf_IMFSourceResolvermf

This trait is enabled with the mf feature, and provides methods for IMFSourceResolver.

mf_IMFStreamDescriptormf

This trait is enabled with the mf feature, and provides methods for IMFStreamDescriptor.

mf_IMFTopologymf

This trait is enabled with the mf feature, and provides methods for IMFTopology.

mf_IMFTopologyNodemf

This trait is enabled with the mf feature, and provides methods for IMFTopologyNode.

mf_IMFVideoDisplayControlmf

This trait is enabled with the mf feature, and provides methods for IMFVideoDisplayControl.

ole_IAdviseSinkole

This trait is enabled with the ole feature, and provides methods for IAdviseSink.

ole_IBindCtxole

This trait is enabled with the ole feature, and provides methods for IBindCtx.

ole_IDataObjectole

This trait is enabled with the ole feature, and provides methods for IDataObject.

ole_IMonikerole

This trait is enabled with the ole feature, and provides methods for IMoniker.

ole_IPersistole

This trait is enabled with the ole feature, and provides methods for IPersist.

ole_IPersistFileole

This trait is enabled with the ole feature, and provides methods for IPersistFile.

ole_IPersistStreamole

This trait is enabled with the ole feature, and provides methods for IPersistStream.

ole_IPictureole

This trait is enabled with the ole feature, and provides methods for IPicture.

ole_ISequentialStreamole

This trait is enabled with the ole feature, and provides methods for ISequentialStream.

ole_IStorageole

This trait is enabled with the ole feature, and provides methods for IStorage.

ole_IStreamole

IStream methods from ole feature.

ole_IUnknownole

This trait is enabled with the ole feature, and provides methods for IUnknown. It is the base trait for all COM traits.

oleaut_IDispatcholeaut

This trait is enabled with the oleaut feature, and provides methods for IDispatch.

oleaut_IPropertyStoreoleaut

This trait is enabled with the oleaut feature, and provides methods for IPropertyStore.

oleaut_ITypeInfooleaut

This trait is enabled with the oleaut feature, and provides methods for ITypeInfo.

shell_IEnumShellItemsshell

This trait is enabled with the shell feature, and provides methods for IEnumShellItems.

shell_IFileDialogshell

This trait is enabled with the shell feature, and provides methods for IFileDialog.

shell_IFileOpenDialogshell

This trait is enabled with the shell feature, and provides methods for IFileOpenDialog.

shell_IFileOperationshell

This trait is enabled with the shell feature, and provides methods for IFileOperation.

shell_IFileSaveDialogshell

This trait is enabled with the shell feature, and provides methods for IFileSaveDialog.

shell_IModalWindowshell

This trait is enabled with the shell feature, and provides methods for IModalWindow.

shell_IShellFoldershell

This trait is enabled with the shell feature, and provides methods for IShellFolder.

shell_IShellItemshell

This trait is enabled with the shell feature, and provides methods for IShellItem.

shell_IShellItem2shell

This trait is enabled with the shell feature, and provides methods for IShellItem2.

shell_IShellItemArrayshell

This trait is enabled with the shell feature, and provides methods for IShellItemArray.

shell_IShellLinkshell

This trait is enabled with the shell feature, and provides methods for IShellLink.

shell_ITaskbarListshell

This trait is enabled with the shell feature, and provides methods for ITaskbarList.

shell_ITaskbarList2shell

This trait is enabled with the shell feature, and provides methods for ITaskbarList2.

shell_ITaskbarList3shell

This trait is enabled with the shell feature, and provides methods for ITaskbarList3.

shell_ITaskbarList4shell

This trait is enabled with the shell feature, and provides methods for ITaskbarList4.

taskschd_IActiontaskschd

This trait is enabled with the taskschd feature, and provides methods for IAction.

taskschd_IActionCollectiontaskschd

This trait is enabled with the taskschd feature, and provides methods for IActionCollection.

taskschd_IBootTriggertaskschd

This trait is enabled with the taskschd feature, and provides methods for IBootTrigger.

taskschd_IComHandlerActiontaskschd

This trait is enabled with the taskschd feature, and provides methods for IComHandlerAction.

taskschd_IDailyTriggertaskschd

This trait is enabled with the taskschd feature, and provides methods for IDailyTrigger.

taskschd_IEmailActiontaskschd

This trait is enabled with the taskschd feature, and provides methods for IEmailAction.

taskschd_IEventTriggertaskschd

This trait is enabled with the taskschd feature, and provides methods for IEventTrigger.

taskschd_IExecActiontaskschd

This trait is enabled with the taskschd feature, and provides methods for IExecAction.

taskschd_IIdleTriggertaskschd

This trait is enabled with the taskschd feature, and provides methods for IIdleTrigger.

taskschd_ILogonTriggertaskschd

This trait is enabled with the taskschd feature, and provides methods for ILogonTrigger.

taskschd_IPrincipaltaskschd

This trait is enabled with the taskschd feature, and provides methods for IPrincipal.

taskschd_IRegisteredTasktaskschd

This trait is enabled with the taskschd feature, and provides methods for IRegisteredTask.

taskschd_IRegistrationInfotaskschd

This trait is enabled with the taskschd feature, and provides methods for IRegistrationInfo.

taskschd_ITaskDefinitiontaskschd

This trait is enabled with the taskschd feature, and provides methods for ITaskDefinition.

taskschd_ITaskFoldertaskschd

This trait is enabled with the taskschd feature, and provides methods for ITaskFolder.

taskschd_ITaskServicetaskschd

This trait is enabled with the taskschd feature, and provides methods for ITaskService.

taskschd_ITaskSettingstaskschd

This trait is enabled with the taskschd feature, and provides methods for ITaskSettings.

taskschd_ITriggertaskschd

This trait is enabled with the taskschd feature, and provides methods for ITrigger.

taskschd_ITriggerCollectiontaskschd

This trait is enabled with the taskschd feature, and provides methods for ITriggerCollection.