Module winsafe::prelude

The WinSafe prelude.

The purpose of this module is to alleviate imports of many common traits. To use it, add a glob import to the top of all your modules that use the library:

use winsafe::prelude::*;

Traits

FormattedError

A system error which can be formatted with FormatMessage.

Handle

A native handle, implemented by all handle types.

IntUnderlying

A type which has a primitive integer as its underlying type.

NativeBitflag

A native typed bitflag constant.

NativeConst

A native typed constant.

NativeStrConst

A native typed constant mapped to a string.

VariableSized

A struct natively defined with its last element being an 1-element array. The struct is dynamically allocated to accomodate a variable number of elements, whose amount is stored in a specific field. The elements are accessed by reaching the past-struct memory room.

kernel_Haccesstoken

This trait is enabled with the kernel feature, and provides methods for HACCESSTOKEN.

kernel_Hevent

This trait is enabled with the kernel feature, and provides methods for HEVENT.

kernel_Heventlog

This trait is enabled with the kernel feature, and provides methods for HEVENTLOG.

kernel_Hfile

This trait is enabled with the kernel feature, and provides methods for HFILE.

kernel_Hfilemap

This trait is enabled with the kernel feature, and provides methods for HFILEMAP.

kernel_Hfilemapview

This trait is enabled with the kernel feature, and provides methods for HFILEMAPVIEW.

kernel_Hfindfile

This trait is enabled with the kernel feature, and provides methods for HFINDFILE.

kernel_Hglobal

This trait is enabled with the kernel feature, and provides methods for HGLOBAL.

kernel_Hheap

This trait is enabled with the kernel feature, and provides methods for HHEAP.

kernel_Hinstance

This trait is enabled with the kernel feature, and provides methods for HINSTANCE.

kernel_Hkey

This trait is enabled with the kernel feature, and provides methods for HKEY.

kernel_Hlocal

This trait is enabled with the kernel feature, and provides methods for HLOCAL.

kernel_Hpipe

This trait is enabled with the kernel feature, and provides methods for HPIPE.

kernel_Hprocess

This trait is enabled with the kernel feature, and provides methods for HPROCESS.

kernel_Hprocesslist

This trait is enabled with the kernel feature, and provides methods for HPROCESSLIST.

kernel_Hsc

This trait is enabled with the kernel feature, and provides methods for HSC.

kernel_Hservice

This trait is enabled with the kernel feature, and provides methods for HSERVICE.

kernel_Hservicestatus

This trait is enabled with the kernel feature, and provides methods for HSERVICESTATUS.

kernel_Hstd

This trait is enabled with the kernel feature, and provides methods for HSTD.

kernel_Hthread

This trait is enabled with the kernel feature, and provides methods for HTHREAD.

kernel_Htransaction

This trait is enabled with the kernel feature, and provides methods for HTRANSACTION.

kernel_Hupdatersrc

This trait is enabled with the kernel feature, and provides methods for HUPDATERSRC.