Trait winsafe::prelude::GdiObject

pub trait GdiObject: Handle + Any { }

Available on crate features kernel and gdi only.

This trait is enabled with the gdi feature, and implements methods for any HGDIOBJ handle, which is the base handle for GDI objects.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Object Safety

This trait is not object safe.