winsafe::prelude

Trait GdiObject

pub trait GdiObject: Handle + Any { }

Available on crate features kernel and gdi only.

This trait is enabled with the gdi feature, and implements methods for any HGDIOBJ handle, which is the base handle for GDI objects.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.