Trait winsafe::prelude::GuiNativeControl

pub trait GuiNativeControl: GuiChild {
    fn on_subclass(&self) -> &WindowEvents;
}

Available on crate feature gui only.

Any native control, which can be subclassed.

Required Methods

fn on_subclass(&self) -> &WindowEvents

Exposes the subclass events. If at least one event exists, the control will be subclassed.

Note: Subclassing may impact performance, use with care.

Panics

Panics if the control or the parent window are already created. Events must be set before control and parent window creation.