Trait winsafe::prelude::GuiParent

pub trait GuiParent: GuiWindow {
    // Required methods
    fn on(&self) -> &WindowEventsAll;
    unsafe fn as_base(&self) -> *mut c_void;
}

Available on crate feature gui only.

Any window which can host child controls.

Required Methods

fn on(&self) -> &WindowEventsAll

Exposes methods to handle the basic window messages, plus timer and native control notifications.

Panics

Panics if the window is already created. Events must be set before window creation.

unsafe fn as_base(&self) -> *mut c_void

Returns a pointer to the inner base window structure.

Safety

The returned pointer must be cast to the specific underlying window struct.

This method is used internally by the library, and not intended to be used externally.