Trait winsafe::prelude::GuiWindow

pub trait GuiWindow: AsAny {
    fn hwnd(&self) -> HWND;
}

This is supported on crate feature gui only.

Any window. Exposes the underlying window handle.

Required methods

fn hwnd(&self) -> HWND

Returns the underlying handle for this control.

Note that the handle is initially null, receiving an actual value only after the control is physically created, what usually happens right before WM_CREATE or WM_INITDIALOG events.