Trait winsafe::prelude::Handle

pub trait Handle: Debug + Display + LowerHex + UpperHex + Copy + Clone + PartialEq + Eq + Send {
    const NULL: Self;

    unsafe fn from_ptr<T>(p: *mut T) -> Self;

    unsafe fn as_ptr(self) -> *mut c_void;

    fn is_null(self) -> bool { ... }

    fn as_opt(self) -> Option<Self> { ... }
}

Any Windows handle.

Associated Constants

const NULL: Self

The null, invalid handle.

Required methods

unsafe fn from_ptr<T>(p: *mut T) -> Self

Creates a new handle instance by wrapping a pointer.

unsafe fn as_ptr(self) -> *mut c_void

Returning the underlying raw pointer.

Provided methods

fn is_null(self) -> bool

Tells if the handle is invalid (null).

fn as_opt(self) -> Option<Self>

Returns None if the handle is null, otherwise returns Some(&Self).