Trait winsafe::prelude::Handle

pub trait Handle: Sized + PartialEq + Eq + Send + Hash + Debug + Display + LowerHex + UpperHex {
    const NULL: Self;
    const INVALID: Self;

    // Required methods
    unsafe fn from_ptr(p: *mut c_void) -> Self;
    unsafe fn as_mut(&mut self) -> &mut *mut c_void;
    fn ptr(&self) -> *mut c_void;

    // Provided methods
    unsafe fn raw_copy(&self) -> Self { ... }
    fn as_opt(&self) -> Option<&Self> { ... }
}

A native handle, implemented by all handle types.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Required Associated Constants

const NULL: Self

The null, uninitialized handle; equals to 0.

const INVALID: Self

The invalid handle; equals to -1.

Operations upon this handle will fail with ERROR::INVALID_HANDLE error code.

Required Methods

unsafe fn from_ptr(p: *mut c_void) -> Self

Creates a new handle object by wrapping a pointer.

This method can be used as an escape hatch to interoperate with other libraries.

Safety

Be sure the pointer has the correct type and isn’t owned by anyone else, otherwise you may cause memory access violations.

unsafe fn as_mut(&mut self) -> &mut *mut c_void

Returns a mutable reference to the underlying raw pointer.

This method can be used as an escape hatch to interoperate with other libraries.

Safety

This method exposes the raw pointer used by raw Windows calls. It’s an opaque pointer to an internal Windows structure, and no dereferencings should be attempted.

fn ptr(&self) -> *mut c_void

Returns the underlying raw pointer.

This method exposes the raw pointer used by raw Windows calls. It’s an opaque pointer to an internal Windows structure, and no dereferencings should be attempted.

This method can be used as an escape hatch to interoperate with other libraries.

Provided Methods

unsafe fn raw_copy(&self) -> Self

Returns a raw copy of the underlying handle pointer.

Safety

As the name implies, raw_copy returns a raw copy of the handle, so closing one of the copies won’t close the others. This means a handle can be used after it has been closed, what can lead to errors and undefined behavior. Even worse: sometimes Windows reuses handle values, so you can call a method on a completely different handle type, what can be catastrophic.

However, in some cases the Windows API demands a copy of the handle – raw_copy is an escape hatch to fill this gap.

fn as_opt(&self) -> Option<&Self>

Returns None if the handle is null or invalid, otherwise returns Some(&self).

Examples

use winsafe::{self as w, prelude::*};

let hfile = w::HFILE::NULL;

match hfile.as_opt() {
    Some(hfile) => println!("Never prints"),
    None => println!("The handle is null"),
}

Object Safety

This trait is not object safe.