winsafe::prelude

Trait IntUnderlying

pub trait IntUnderlying:
    Sized
    + Clone
    + Copy
    + PartialEq
    + Eq
    + PartialOrd
    + Ord
    + Hash
    + Send
    + Into<Self::Raw>
    + AsRef<Self::Raw> {
    type Raw;

    // Required method
    unsafe fn as_mut(&mut self) -> &mut Self::Raw;
}

Available on crate feature kernel only.

A type which has a primitive integer as its underlying type.

Required Associated Types

type Raw

The underlying raw integer type.

Required Methods

unsafe fn as_mut(&mut self) -> &mut Self::Raw

Returns a mutable reference to the underlying raw value.

Safety

Be sure the integer being set is meaningful for the actual type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.