Trait winsafe::prelude::KernelHprocess

pub trait KernelHprocess: Handle {

    fn CreateProcess(
        application_name: Option<&str>, 
        command_line: Option<&str>, 
        process_attrs: Option<&mut SECURITY_ATTRIBUTES<'_>>, 
        thread_attrs: Option<&mut SECURITY_ATTRIBUTES<'_>>, 
        inherit_handles: bool, 
        creation_flags: CREATE, 
        environment: Option<Vec<String>>, 
        current_dir: Option<&str>, 
        si: &mut STARTUPINFO<'_, '_>
    ) -> WinResult<PROCESS_INFORMATION> { ... }

    fn ExitProcess(exit_code: u32) { ... }

    fn FlushInstructionCache(
        self, 
        base_address: *mut c_void, 
        size: u64
    ) -> WinResult<()> { ... }

    fn FlushProcessWriteBuffers() { ... }

    fn GetCurrentProcess() -> HPROCESS { ... }

    fn GetExitCodeProcess(self) -> WinResult<u32> { ... }

    fn GetGuiResources(self, flags: GR) -> WinResult<u32> { ... }

    fn GetProcessId(self) -> WinResult<u32> { ... }

    fn GetProcessTimes(
        self, 
        creation: &mut FILETIME, 
        exit: &mut FILETIME, 
        kernel: &mut FILETIME, 
        user: &mut FILETIME
    ) -> WinResult<()> { ... }

    fn IsWow64Process(self) -> WinResult<bool> { ... }

    fn OpenProcess(
        desired_access: PROCESS, 
        inherit_handle: bool, 
        process_id: u32
    ) -> WinResult<HPROCESS> { ... }

    fn QueryFullProcessImageName(self, flags: PROCESS_NAME) -> WinResult<String> { ... }

    fn WaitForSingleObject(self, milliseconds: Option<u32>) -> WinResult<WAIT> { ... }

}

HPROCESS methods from kernel feature.

Provided methods

fn CreateProcess(
application_name: Option<&str>,
command_line: Option<&str>,
process_attrs: Option<&mut SECURITY_ATTRIBUTES<'_>>,
thread_attrs: Option<&mut SECURITY_ATTRIBUTES<'_>>,
inherit_handles: bool,
creation_flags: CREATE,
environment: Option<Vec<String>>,
current_dir: Option<&str>,
si: &mut STARTUPINFO<'_, '_>
) -> WinResult<PROCESS_INFORMATION>

CreateProcess static method.

Note: Process and thread handles are returned in the PROCESS_INFORMATION struct, and they must be paired with their respective HPROCESS::CloseHandle and HTHREAD::CloseHandle calls.

fn ExitProcess(exit_code: u32)

ExitProcess static method.

fn FlushInstructionCache(
self,
base_address: *mut c_void,
size: u64
) -> WinResult<()>

FlushInstructionCache method.

fn FlushProcessWriteBuffers()

FlushProcessWriteBuffers static method.

fn GetCurrentProcess() -> HPROCESS

GetCurrentProcess static method.

fn GetExitCodeProcess(self) -> WinResult<u32>

GetExitCodeProcess method.

fn GetGuiResources(self, flags: GR) -> WinResult<u32>

GetGuiResources method.

fn GetProcessId(self) -> WinResult<u32>

GetProcessId method.

fn GetProcessTimes(
self,
creation: &mut FILETIME,
exit: &mut FILETIME,
kernel: &mut FILETIME,
user: &mut FILETIME
) -> WinResult<()>

GetProcessTimes method.

fn IsWow64Process(self) -> WinResult<bool>

IsWow64Process method.

fn OpenProcess(
desired_access: PROCESS,
inherit_handle: bool,
process_id: u32
) -> WinResult<HPROCESS>

OpenProcess static method.

Note: Must be paired with an HPROCESS::CloseHandle call.

fn QueryFullProcessImageName(self, flags: PROCESS_NAME) -> WinResult<String>

QueryFullProcessImageName method.

fn WaitForSingleObject(self, milliseconds: Option<u32>) -> WinResult<WAIT>

WaitForSingleObject method.