Trait winsafe::prelude::MsgSend

pub unsafe trait MsgSend {
    type RetType;

    // Required methods
    fn convert_ret(&self, v: isize) -> Self::RetType;
    fn as_generic_wm(&mut self) -> WndMsg;
}

Available on crate features kernel and user only.

Parameters of a message which can be sent. Implemented by all defined messages.

Allows the conversion to the generic WndMsg parameters, and also defines the return type of the message.

Used in functions like SendMessage and DefWindowProc.

Safety

Messages manipulate pointers, copies and window states. Improper use may lead to undefined behavior.

Required Associated Types

type RetType

The specific type of the value returned by the message.

Required Methods

fn convert_ret(&self, v: isize) -> Self::RetType

Converts the generic isize return value to the specific type returned by the message.

fn as_generic_wm(&mut self) -> WndMsg

Converts the specific message parameters struct into the generic WndMsg message struct.