winsafe::prelude

Trait MsgSend

pub trait MsgSend {
    type RetType;

    // Required methods
    unsafe fn isize_to_ret(&self, v: isize) -> Self::RetType;
    fn as_generic_wm(&mut self) -> WndMsg;
}

Available on crate feature user only.

This trait is enabled with the gui feature, and is implemented by the parameters of messages which can be sent. Implemented by all defined messages.

Allows the conversion to the generic WndMsg parameters, and also defines the return type of the message.

Used in functions like SendMessage and DefWindowProc.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Required Associated Types

type RetType

The specific type of the value returned by the message.

Required Methods

unsafe fn isize_to_ret(&self, v: isize) -> Self::RetType

Unmarshaling method which converts the generic isize return value to the specific type returned by the message.

Safety

Return values often involve pointers and require casts, make sure the conversions are correct.

fn as_generic_wm(&mut self) -> WndMsg

Marshaling method which converts the specific message parameters struct into the generic WndMsg message struct.