Trait winsafe::prelude::MsgSendRecv

pub unsafe trait MsgSendRecv: MsgSend {
    fn from_generic_wm(parm: WndMsg) -> Self;
}

Available on crate feature user only.

Parameters of a message which can be sent and handled (received). Implemented by WndMsg and all msg::wm messages.

Allows the conversion from and to the generic WndMsg parameters, and also defines the return type of the message.

Required Methods

fn from_generic_wm(parm: WndMsg) -> Self

Converts the generic WndMsg parameters struct into the specific message struct.