Trait winsafe::prelude::MsgSendRecv

pub unsafe trait MsgSendRecv: MsgSend {
    // Required method
    fn from_generic_wm(parm: WndMsg) -> Self;
}

Available on crate features kernel and user only.

Parameters of a message which can be sent and handled (received). Implemented by WndMsg and all msg::wm messages.

Allows the conversion from and to the generic WndMsg parameters, and also defines the return type of the message.

Safety

Messages manipulate pointers, copies and window states. Improper use may lead to undefined behavior.

Required Methods

fn from_generic_wm(parm: WndMsg) -> Self

Converts the generic WndMsg parameters struct into the specific message struct.

Object Safety

This trait is not object safe.