winsafe::prelude

Trait MsgSendRecv

pub trait MsgSendRecv: MsgSend {
    // Required method
    unsafe fn from_generic_wm(parm: WndMsg) -> Self;
}

Available on crate features kernel and user only.

Parameters of a message which can be sent and handled (received). Implemented by WndMsg and all msg::wm messages.

Allows the conversion from and to the generic WndMsg parameters, and also defines the return type of the message.

Required Methods

unsafe fn from_generic_wm(parm: WndMsg) -> Self

Unmarshaling method which converts the generic WndMsg parameters struct into the specific message struct.

Safety

Message parameters often involve pointers and require casts, make sure the conversions are correct.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.