Trait winsafe::prelude::NativeConst

pub trait NativeConst: Sized + Default + Clone + Copy + PartialEq + Eq + Send + Hash + From<Self::Raw> + Into<Self::Raw> + Debug + Display + LowerHex + UpperHex + Binary + Octal {
    type Raw;
}

Available on crate feature kernel only.

A native typed constant.

If the values of this constant type can be combined as bitflags, it will also implement the NativeBitflag trait.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Required Associated Types

type Raw

The underlying type of this constant.