Trait winsafe::prelude::NativeConst

pub trait NativeConst: IntUnderlying + Default + Debug + Display + LowerHex + UpperHex + Binary + Octal { }

A native typed constant.

If the values of this constant type can be combined as bitflags, it will also implement the NativeBitflag trait.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Object Safety

This trait is not object safe.