Trait winsafe::prelude::UserHwnd

pub trait UserHwnd: Handle {
    const BROADCAST: HWND;

    fn hinstance(self) -> HINSTANCE { ... }

    fn ArrangeIconicWindows(self) -> WinResult<u32> { ... }

    fn BeginPaint(self, ps: &mut PAINTSTRUCT) -> WinResult<HDC> { ... }

    fn BringWindowToTop(self) -> WinResult<()> { ... }

    fn ChildWindowFromPoint(self, pt: POINT) -> Option<HWND> { ... }

    fn ClientToScreen(self, pt: &mut POINT) -> WinResult<()> { ... }

    fn ClientToScreenRc(self, rc: &mut RECT) -> WinResult<()> { ... }

    fn CloseWindow(self) -> WinResult<()> { ... }

    fn CreateWindowEx(
        ex_style: WS_EX, 
        class_name: AtomStr, 
        title: Option<&str>, 
        style: WS, 
        pos: POINT, 
        size: SIZE, 
        hwnd_parent: Option<HWND>, 
        hmenu: IdMenu, 
        hinstance: HINSTANCE, 
        lparam: Option<isize>
    ) -> WinResult<HWND> { ... }

    fn DefWindowProc<M>(self, msg: M) -> M::RetType
    where
        M: MsgSend,
    { ... }

    fn DestroyWindow(self) -> WinResult<()> { ... }

    fn DrawMenuBar(self) -> WinResult<()> { ... }

    fn EnableWindow(self, enable: bool) -> bool { ... }

    fn EndDialog(self, result: isize) -> WinResult<()> { ... }

    fn EndPaint(self, ps: &PAINTSTRUCT) { ... }

    fn EnumChildWindows<F>(self, func: F)
    where
        F: Fn(HWND) -> bool,
    { ... }

    fn FindWindow(
        class_name: Option<AtomStr>, 
        title: Option<&str>
    ) -> WinResult<HWND> { ... }

    fn FindWindowEx(
        self, 
        hwnd_child_after: Option<HWND>, 
        class_name: AtomStr, 
        title: Option<&str>
    ) -> WinResult<HWND> { ... }

    fn GetActiveWindow() -> Option<HWND> { ... }

    fn GetAltTabInfo(
        self, 
        item: Option<u32>, 
        ati: &mut ALTTABINFO, 
        sz_item_text: Option<u32>
    ) -> WinResult<String> { ... }

    fn GetAncestor(self, flags: GA) -> Option<HWND> { ... }

    fn GetCapture() -> Option<HWND> { ... }

    fn GetClassLongPtr(self, index: GCLP) -> usize { ... }

    fn GetClassName(self) -> WinResult<String> { ... }

    fn GetClientRect(self) -> WinResult<RECT> { ... }

    fn GetDC(self) -> WinResult<HDC> { ... }

    fn GetDesktopWindow() -> HWND { ... }

    fn GetDlgCtrlID(self) -> WinResult<u16> { ... }

    fn GetDlgItem(self, ctrl_id: u16) -> WinResult<HWND> { ... }

    fn GetFocus() -> Option<HWND> { ... }

    fn GetForegroundWindow() -> Option<HWND> { ... }

    fn GetLastActivePopup(self) -> Option<HWND> { ... }

    fn GetMenu(self) -> Option<HMENU> { ... }

    fn GetMenuBarInfo(
        self, 
        obj_id: OBJID, 
        item_id: u32, 
        mbi: &mut MENUBARINFO
    ) -> WinResult<()> { ... }

    fn GetMenuItemRect(
        self, 
        hmenu: HMENU, 
        item_pos: u32, 
        rc_item: &mut RECT
    ) -> WinResult<()> { ... }

    fn GetNextDlgGroupItem(
        self, 
        hwnd_ctrl: HWND, 
        previous: bool
    ) -> WinResult<HWND> { ... }

    fn GetNextDlgTabItem(
        self, 
        hwnd_ctrl: HWND, 
        previous: bool
    ) -> WinResult<HWND> { ... }

    fn GetParent(self) -> WinResult<HWND> { ... }

    fn GetScrollInfo(self, bar: SBB, si: &mut SCROLLINFO) -> WinResult<()> { ... }

    fn GetScrollPos(self, bar: SBB) -> WinResult<i32> { ... }

    fn GetShellWindow() -> Option<HWND> { ... }

    fn GetSystemMenu(self, revert: bool) -> Option<HMENU> { ... }

    fn GetTopWindow(self) -> WinResult<Option<HWND>> { ... }

    fn GetUpdateRgn(self, hrgn: HRGN, erase: bool) -> WinResult<REGION> { ... }

    fn GetWindow(self, cmd: GW) -> WinResult<HWND> { ... }

    fn GetWindowDC(self) -> WinResult<HDC> { ... }

    fn GetWindowDisplayAffinity(self) -> WinResult<WDA> { ... }

    fn GetWindowInfo(self, wi: &mut WINDOWINFO) -> WinResult<()> { ... }

    fn GetWindowLongPtr(self, index: GWLP) -> isize { ... }

    fn GetWindowPlacement(self, wp: &mut WINDOWPLACEMENT) -> WinResult<()> { ... }

    fn GetWindowRect(self) -> WinResult<RECT> { ... }

    fn GetWindowRgn(self, hrgn: HRGN) -> WinResult<REGION> { ... }

    fn GetWindowRgnBox(self, rc: &mut RECT) -> WinResult<REGION> { ... }

    fn GetWindowText(self) -> WinResult<String> { ... }

    fn GetWindowTextLength(self) -> WinResult<i32> { ... }

    fn GetWindowThreadProcessId(self) -> (u32, u32) { ... }

    fn HiliteMenuItem(self, hmenu: HMENU, id_or_pos: IdPos, hilite: MF) -> bool { ... }

    fn InvalidateRect(self, rc: Option<&RECT>, erase: bool) -> WinResult<()> { ... }

    fn InvalidateRgn(self, hrgn: HRGN, erase: bool) { ... }

    fn IsChild(self, hwnd_possible_child: HWND) -> bool { ... }

    fn IsDialogMessage(self, msg: &mut MSG) -> bool { ... }

    fn IsIconic(self) -> bool { ... }

    fn IsWindow(self) -> bool { ... }

    fn IsWindowEnabled(self) -> bool { ... }

    fn IsWindowUnicode(self) -> bool { ... }

    fn IsWindowVisible(self) -> bool { ... }

    fn IsZoomed(self) -> bool { ... }

    fn KillTimer(self, event_id: usize) -> WinResult<()> { ... }

    fn LogicalToPhysicalPoint(self, pt: *mut POINT) -> WinResult<()> { ... }

    fn MapDialogRect(self, rc: &mut RECT) -> WinResult<()> { ... }

    fn MessageBox(
        self, 
        text: &str, 
        caption: &str, 
        flags: MB
    ) -> WinResult<DLGID> { ... }

    fn MonitorFromWindow(self, flags: MONITOR) -> HMONITOR { ... }

    fn MoveWindow(self, pos: POINT, size: SIZE, repaint: bool) -> WinResult<()> { ... }

    fn OpenClipboard(self) -> WinResult<()> { ... }

    fn PostMessage<M>(self, msg: M) -> WinResult<()>
    where
        M: MsgSend,
    { ... }

    fn RealChildWindowFromPoint(
        self, 
        pt_parent_client_coords: POINT
    ) -> Option<HWND> { ... }

    fn RealGetWindowClass(self) -> WinResult<String> { ... }

    fn RedrawWindow(
        self, 
        rc_update: &RECT, 
        hrgn_update: HRGN, 
        flags: RDW
    ) -> WinResult<()> { ... }

    fn ReleaseDC(self, hdc: HDC) -> WinResult<()> { ... }

    fn ScreenToClient(self, pt: &mut POINT) -> WinResult<()> { ... }

    fn ScreenToClientRc(self, rc: &mut RECT) -> WinResult<()> { ... }

    fn SendMessage<M>(self, msg: M) -> M::RetType
    where
        M: MsgSend,
    { ... }

    fn SetCapture(&self) -> Option<HWND> { ... }

    fn SetFocus(self) -> Option<HWND> { ... }

    fn SetForegroundWindow(self) -> bool { ... }

    fn SetMenu(self, hmenu: HMENU) -> WinResult<()> { ... }

    fn SetParent(self, hwnd_new_parent: HWND) -> WinResult<Option<HWND>> { ... }

    fn SetScrollInfo(self, bar: SBB, si: &SCROLLINFO, redraw: bool) -> i32 { ... }

    fn SetScrollPos(self, b: SBB, pos: i32, redraw: bool) -> WinResult<i32> { ... }

    fn SetScrollRange(
        self, 
        bar: SBB, 
        min_pos: i32, 
        max_pos: i32, 
        redraw: bool
    ) -> WinResult<()> { ... }

    fn SetTimer(
        self, 
        event_id: usize, 
        elapse_ms: u32, 
        timer_func: Option<TIMERPROC>
    ) -> WinResult<usize> { ... }

    fn SetWindowDisplayAffinity(self, affinity: WDA) -> WinResult<()> { ... }

    fn SetWindowLongPtr(self, index: GWLP, new_long: isize) -> isize { ... }

    fn SetWindowPlacement(self, wp: &WINDOWPLACEMENT) -> WinResult<()> { ... }

    fn SetWindowPos(
        self, 
        hwnd_insert_after: HwndPlace, 
        pos: POINT, 
        size: SIZE, 
        flags: SWP
    ) -> WinResult<()> { ... }

    fn SetWindowRgn(self, hrgn: HRGN, redraw: bool) -> WinResult<()> { ... }

    fn SetWindowText(self, text: &str) -> WinResult<()> { ... }

    fn ShowCaret(self) -> WinResult<()> { ... }

    fn ShowWindow(self, show_cmd: SW) -> bool { ... }

    fn ShowWindowAsync(self, show_cmd: SW) -> WinResult<()> { ... }

    fn TranslateAccelerator(
        self, 
        haccel_table: HACCEL, 
        msg: &mut MSG
    ) -> WinResult<()> { ... }

    fn UpdateWindow(self) -> WinResult<()> { ... }

    fn ValidateRect(self, rc: &RECT) -> WinResult<()> { ... }

    fn ValidateRgn(self, hrgn: HRGN) -> WinResult<()> { ... }

    fn WindowFromPhysicalPoint(pt: POINT) -> Option<HWND> { ... }

    fn WindowFromPoint(pt: POINT) -> Option<HWND> { ... }

    fn WinHelp(self, help_file: &str, cmd: HELPW, data: usize) -> WinResult<()> { ... }

}

This is supported on crate feature user only.

HWND methods from user feature.

Associated Constants

const BROADCAST: HWND

Represents all top-level windows in HWND::PostMessage and HWND::SendMessage.

Provided methods

fn hinstance(self) -> HINSTANCE

GetWindowLongPtr wrapper to retrieve the window HINSTANCE.

fn ArrangeIconicWindows(self) -> WinResult<u32>

ArrangeIconicWindows method.

fn BeginPaint(self, ps: &mut PAINTSTRUCT) -> WinResult<HDC>

BeginPaint method.

Note: Must be paired with an HWND::EndPaint call.

Examples

use winsafe::prelude::*;
use winsafe::{HWND, PAINTSTRUCT};

let hwnd: HWND; // initialized somewhere

let mut ps = PAINTSTRUCT::default();
let hdc = hwnd.BeginPaint(&mut ps)?;

// hdc painting...

hwnd.EndPaint(&ps);

fn BringWindowToTop(self) -> WinResult<()>

BringWindowToTop method.

fn ChildWindowFromPoint(self, pt: POINT) -> Option<HWND>

ChildWindowFromPoint method.

fn ClientToScreen(self, pt: &mut POINT) -> WinResult<()>

ClientToScreen method.

fn ClientToScreenRc(self, rc: &mut RECT) -> WinResult<()>

ClientToScreen method for a RECT.

fn CloseWindow(self) -> WinResult<()>

CloseWindow method.

Note that this method will actually minimize the window, not destroy it.

fn CreateWindowEx(
ex_style: WS_EX,
class_name: AtomStr,
title: Option<&str>,
style: WS,
pos: POINT,
size: SIZE,
hwnd_parent: Option<HWND>,
hmenu: IdMenu,
hinstance: HINSTANCE,
lparam: Option<isize>
) -> WinResult<HWND>

CreateWindowEx static method.

fn DefWindowProc<M>(self, msg: M) -> M::RetType where
M: MsgSend,

DefWindowProc method.

The return type is variable, being defined by the RetType associated type of the MsgSend trait. That means each message can define its own return type.

fn DestroyWindow(self) -> WinResult<()>

DestroyWindow method.

fn DrawMenuBar(self) -> WinResult<()>

DrawMenuBar method.

fn EnableWindow(self, enable: bool) -> bool

EnableWindow method.

fn EndDialog(self, result: isize) -> WinResult<()>

EndDialog method.

fn EndPaint(self, ps: &PAINTSTRUCT)

EndPaint method.

fn EnumChildWindows<F>(self, func: F) where
F: Fn(HWND) -> bool,

EnumChildWindows method.

Examples

use winsafe::prelude::*;
use winsafe::HWND;

let hwnd: HWND; // initialized somewhere

hwnd.EnumChildWindows(|hchild: HWND| -> bool {
    println!("Child HWND: {}", hchild);
    true
});

fn FindWindow(
class_name: Option<AtomStr>,
title: Option<&str>
) -> WinResult<HWND>

FindWindow static method.

fn FindWindowEx(
self,
hwnd_child_after: Option<HWND>,
class_name: AtomStr,
title: Option<&str>
) -> WinResult<HWND>

FindWindowEx method.

fn GetActiveWindow() -> Option<HWND>

GetActiveWindow static method.

fn GetAltTabInfo(
self,
item: Option<u32>,
ati: &mut ALTTABINFO,
sz_item_text: Option<u32>
) -> WinResult<String>

GetAltTabInfo method.

If item is None, the item text is not retrieved.

The sz_item_text is the maximum number of expected chars for the item text. If None, defaults to 100.

fn GetAncestor(self, flags: GA) -> Option<HWND>

GetAncestor method.

fn GetCapture() -> Option<HWND>

GetCapture static method.

fn GetClassLongPtr(self, index: GCLP) -> usize

GetClassLongPtr method.

fn GetClassName(self) -> WinResult<String>

GetClassName method.

fn GetClientRect(self) -> WinResult<RECT>

GetClientRect method.

fn GetDC(self) -> WinResult<HDC>

GetDC method.

Note: Must be paired with an HWND::ReleaseDC call.

fn GetDesktopWindow() -> HWND

GetDesktopWindow static method.

fn GetDlgCtrlID(self) -> WinResult<u16>

GetDlgCtrlID method.

fn GetDlgItem(self, ctrl_id: u16) -> WinResult<HWND>

GetDlgItem method.

fn GetFocus() -> Option<HWND>

GetFocus static method.

fn GetForegroundWindow() -> Option<HWND>

GetForegroundWindow static method.

fn GetLastActivePopup(self) -> Option<HWND>

GetLastActivePopup method.

fn GetMenu(self) -> Option<HMENU>

GetMenu method.

fn GetMenuBarInfo(
self,
obj_id: OBJID,
item_id: u32,
mbi: &mut MENUBARINFO
) -> WinResult<()>

GetMenuBarInfo method.

fn GetMenuItemRect(
self,
hmenu: HMENU,
item_pos: u32,
rc_item: &mut RECT
) -> WinResult<()>

GetMenuItemRect method.

fn GetNextDlgGroupItem(self, hwnd_ctrl: HWND, previous: bool) -> WinResult<HWND>

GetNextDlgGroupItem method.

fn GetNextDlgTabItem(self, hwnd_ctrl: HWND, previous: bool) -> WinResult<HWND>

GetNextDlgTabItem method.

fn GetParent(self) -> WinResult<HWND>

GetParent method.

fn GetScrollInfo(self, bar: SBB, si: &mut SCROLLINFO) -> WinResult<()>

GetScrollInfo method.

fn GetScrollPos(self, bar: SBB) -> WinResult<i32>

GetScrollPos method.

fn GetShellWindow() -> Option<HWND>

GetShellWindow static method.

fn GetSystemMenu(self, revert: bool) -> Option<HMENU>

GetSystemMenu method.

fn GetTopWindow(self) -> WinResult<Option<HWND>>

GetTopWindow method.

fn GetUpdateRgn(self, hrgn: HRGN, erase: bool) -> WinResult<REGION>

GetUpdateRgn method.

fn GetWindow(self, cmd: GW) -> WinResult<HWND>

GetWindow method.

fn GetWindowDC(self) -> WinResult<HDC>

GetWindowDC method.

Note: Must be paired with an HWND::ReleaseDC call.

fn GetWindowDisplayAffinity(self) -> WinResult<WDA>

GetWindowDisplayAffinity method.

fn GetWindowInfo(self, wi: &mut WINDOWINFO) -> WinResult<()>

GetWindowInfo method.

fn GetWindowLongPtr(self, index: GWLP) -> isize

GetWindowLongPtr method.

fn GetWindowPlacement(self, wp: &mut WINDOWPLACEMENT) -> WinResult<()>

GetWindowPlacement method.

fn GetWindowRect(self) -> WinResult<RECT>

GetWindowRect method.

fn GetWindowRgn(self, hrgn: HRGN) -> WinResult<REGION>

GetWindowRgn method.

fn GetWindowRgnBox(self, rc: &mut RECT) -> WinResult<REGION>

GetWindowRgnBox method.

fn GetWindowText(self) -> WinResult<String>

GetWindowText method.

use winsafe::prelude::*;
use winsafe::HWND;

let hwnd: HWND; // initialized somewhere

let text = hwnd.GetWindowText()?;
println!("Text: {}", text);

fn GetWindowTextLength(self) -> WinResult<i32>

GetWindowTextLength method. Does not count the terminating null.

fn GetWindowThreadProcessId(self) -> (u32, u32)

GetWindowThreadProcessId method.

Returns thread ID and process ID, respectively.

fn HiliteMenuItem(self, hmenu: HMENU, id_or_pos: IdPos, hilite: MF) -> bool

HiliteMenuItem method.

fn InvalidateRect(self, rc: Option<&RECT>, erase: bool) -> WinResult<()>

InvalidateRect method.

Examples

Most of the time you’ll just want update the entire client area:

use winsafe::prelude::*;
use winsafe::HWND;

let hwnd: HWND; // initialized somewhere

hwnd.InvalidateRect(None, true)?;

fn InvalidateRgn(self, hrgn: HRGN, erase: bool)

InvalidateRgn method.

fn IsChild(self, hwnd_possible_child: HWND) -> bool

IsChild method.

fn IsDialogMessage(self, msg: &mut MSG) -> bool

IsDialogMessage method.

fn IsIconic(self) -> bool

IsIconic method.

fn IsWindow(self) -> bool

IsWindow method.

fn IsWindowEnabled(self) -> bool

IsWindowEnabled method.

fn IsWindowUnicode(self) -> bool

IsWindowUnicode method.

fn IsWindowVisible(self) -> bool

IsWindowVisible method.

fn IsZoomed(self) -> bool

IsZoomed method.

fn KillTimer(self, event_id: usize) -> WinResult<()>

KillTimer method.

fn LogicalToPhysicalPoint(self, pt: *mut POINT) -> WinResult<()>

LogicalToPhysicalPoint method.

fn MapDialogRect(self, rc: &mut RECT) -> WinResult<()>

MapDialogRect method.

fn MessageBox(self, text: &str, caption: &str, flags: MB) -> WinResult<DLGID>

MessageBox method.

Consider using the more modern TaskDialog method.

Examples

A modal message box, which blocks its parent:

use winsafe::prelude::*;
use winsafe::{co, HWND};

let my_hwnd: HWND; // initialized somewhere

hwnd.MessageBox("Hello, world", "title",
    co::MB::OKCANCEL | co::MB::ICONINFORMATION)?;

Usually the message box has a valid parent window, however, if for some reason you don’t have a window to serve as parent, you still can show a non-modal, parent-less message box by retrieving the desktop handle:

use winsafe::prelude::*;
use winsafe::{co, HWND};

HWND::GetDesktopWindow()
    .MessageBox("Hello, world", "Title", co::MB::ICONEXCLAMATION)?;

fn MonitorFromWindow(self, flags: MONITOR) -> HMONITOR

MonitorFromWindow method.

fn MoveWindow(self, pos: POINT, size: SIZE, repaint: bool) -> WinResult<()>

MoveWindow method.

fn OpenClipboard(self) -> WinResult<()>

OpenClipboard method.

Note: Must be paired with a CloseClipboard call.

fn PostMessage<M>(self, msg: M) -> WinResult<()> where
M: MsgSend,

PostMessage method. Note that this method is asychronous.

Examples

Programatically closing a window:

use winsafe::prelude::*;
use winsafe::{HWND, msg::wm};

let hwnd: HWND; // initialized somewhere

hwnd.PostMessage(wm::Close {})?;

Sending a message to all top-level windows:

use winsafe::prelude::*;
use winsafe::{HWND, msg::wm};

HWND::BROADCAST.PostMessage(
    wm::ExitMenuLoop { is_shortcut: false },
)?;

fn RealChildWindowFromPoint(
self,
pt_parent_client_coords: POINT
) -> Option<HWND>

RealChildWindowFromPoint method.

fn RealGetWindowClass(self) -> WinResult<String>

RealGetWindowClassW method.

fn RedrawWindow(
self,
rc_update: &RECT,
hrgn_update: HRGN,
flags: RDW
) -> WinResult<()>

RedrawWindow method.

fn ReleaseDC(self, hdc: HDC) -> WinResult<()>

ReleaseDC method.

fn ScreenToClient(self, pt: &mut POINT) -> WinResult<()>

ScreenToClient method.

fn ScreenToClientRc(self, rc: &mut RECT) -> WinResult<()>

ScreenToClient method for a RECT.

fn SendMessage<M>(self, msg: M) -> M::RetType where
M: MsgSend,

SendMessage method.

The return type is variable, being defined by the RetType associated type of the MsgSend trait. That means each message can define its own return type.

Examples

Sending a lvm::SetItem list view message, which demands a reference to an LVITEM object:

use winsafe::prelude::*;
use winsafe::{co, HWND, LVITEM, msg::lvm};

let hwnd: HWND; // initialized somewhere

let mut lvi = LVITEM::default(); // object to be sent
lvi.mask = co::LVIF::IMAGE;
lvi.iImage = 3;

hwnd.SendMessage(lvm::SetItem {
    lvitem: &lvi,
});

fn SetCapture(&self) -> Option<HWND>

SetCapture method.

Note: Must be paired with a ReleaseCapture call.

fn SetFocus(self) -> Option<HWND>

SetFocus method.

fn SetForegroundWindow(self) -> bool

SetForegroundWindow method.

fn SetMenu(self, hmenu: HMENU) -> WinResult<()>

SetMenu method.

fn SetParent(self, hwnd_new_parent: HWND) -> WinResult<Option<HWND>>

SetParent method.

fn SetScrollInfo(self, bar: SBB, si: &SCROLLINFO, redraw: bool) -> i32

SetScrollInfo method.

fn SetScrollPos(self, b: SBB, pos: i32, redraw: bool) -> WinResult<i32>

SetScrollPos method.

fn SetScrollRange(
self,
bar: SBB,
min_pos: i32,
max_pos: i32,
redraw: bool
) -> WinResult<()>

SetScrollRange method.

fn SetTimer(
self,
event_id: usize,
elapse_ms: u32,
timer_func: Option<TIMERPROC>
) -> WinResult<usize>

SetTimer method.

fn SetWindowDisplayAffinity(self, affinity: WDA) -> WinResult<()>

SetWindowDisplayAffinity method.

fn SetWindowLongPtr(self, index: GWLP, new_long: isize) -> isize

SetWindowLongPtr method.

fn SetWindowPlacement(self, wp: &WINDOWPLACEMENT) -> WinResult<()>

SetWindowPlacement method.

fn SetWindowPos(
self,
hwnd_insert_after: HwndPlace,
pos: POINT,
size: SIZE,
flags: SWP
) -> WinResult<()>

SetWindowPos method.

fn SetWindowRgn(self, hrgn: HRGN, redraw: bool) -> WinResult<()>

SetWindowRgn method.

fn SetWindowText(self, text: &str) -> WinResult<()>

SetWindowText method.

fn ShowCaret(self) -> WinResult<()>

ShowCaret method.

fn ShowWindow(self, show_cmd: SW) -> bool

ShowWindow method.

fn ShowWindowAsync(self, show_cmd: SW) -> WinResult<()>

ShowWindowAsync method.

fn TranslateAccelerator(
self,
haccel_table: HACCEL,
msg: &mut MSG
) -> WinResult<()>

TranslateAccelerator method.

fn UpdateWindow(self) -> WinResult<()>

UpdateWindow method.

fn ValidateRect(self, rc: &RECT) -> WinResult<()>

ValidateRect method.

fn ValidateRgn(self, hrgn: HRGN) -> WinResult<()>

ValidateRgn method.

fn WindowFromPhysicalPoint(pt: POINT) -> Option<HWND>

WindowFromPhysicalPoint static method.

fn WindowFromPoint(pt: POINT) -> Option<HWND>

WindowFromPoint static method.

fn WinHelp(self, help_file: &str, cmd: HELPW, data: usize) -> WinResult<()>

WinHelp method.