Trait winsafe::prelude::advapi_Hkey

pub trait advapi_Hkey: Handle {
    const CLASSES_ROOT: HKEY = _;
    const CURRENT_USER: HKEY = _;
    const LOCAL_MACHINE: HKEY = _;
    const USERS: HKEY = _;
    const PERFORMANCE_DATA: HKEY = _;
    const CURRENT_CONFIG: HKEY = _;
    const DYN_DATA: HKEY = _;
    const CURRENT_USER_LOCAL_SETTINGS: HKEY = _;
    const PERFORMANCE_TEXT: HKEY = _;
    const PERFORMANCE_NLSTEXT: HKEY = _;

    fn RegEnumKeyEx(
        &self
    ) -> SysResult<Box<dyn Iterator<Item = SysResult<String>> + '_>> { ... }
    fn RegEnumValue(
        &self
    ) -> SysResult<Box<dyn Iterator<Item = SysResult<(String, REG)>> + '_>> { ... }
    fn RegFlushKey(&self) -> SysResult<()> { ... }
    fn RegGetValue(
        &self,
        sub_key: Option<&str>,
        value: Option<&str>
    ) -> SysResult<RegistryValue> { ... }
    fn RegOpenKeyEx(
        &self,
        sub_key: &str,
        options: REG_OPTION,
        access_rights: KEY
    ) -> SysResult<HkeyGuard> { ... }
    fn RegQueryInfoKey(
        &self,
        class: Option<&mut WString>,
        num_sub_keys: Option<&mut u32>,
        max_sub_key_name_len: Option<&mut u32>,
        max_class_len: Option<&mut u32>,
        num_values: Option<&mut u32>,
        max_value_name_len: Option<&mut u32>,
        max_value_len: Option<&mut u32>,
        security_descr_len: Option<&mut u32>,
        last_write_time: Option<&mut FILETIME>
    ) -> SysResult<()> { ... }
    fn RegQueryValueEx(&self, value: &str) -> SysResult<RegistryValue> { ... }
    fn RegSetKeyValue(
        &self,
        sub_key: &str,
        value: &str,
        data: RegistryValue
    ) -> SysResult<()> { ... }
    fn RegSetValueEx(&self, value: &str, data: RegistryValue) -> SysResult<()> { ... }
}

Available on crate feature advapi only.

This trait is enabled with the advapi feature, and provides methods for HKEY.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Associated Constants

const CLASSES_ROOT: HKEY = _

Predefined registry key, always open.

const CURRENT_USER: HKEY = _

Predefined registry key, always open.

const LOCAL_MACHINE: HKEY = _

Predefined registry key, always open.

const USERS: HKEY = _

Predefined registry key, always open.

const PERFORMANCE_DATA: HKEY = _

Predefined registry key, always open.

const CURRENT_CONFIG: HKEY = _

Predefined registry key, always open.

const DYN_DATA: HKEY = _

Predefined registry key, always open.

const CURRENT_USER_LOCAL_SETTINGS: HKEY = _

Predefined registry key, always open.

const PERFORMANCE_TEXT: HKEY = _

Predefined registry key, always open.

const PERFORMANCE_NLSTEXT: HKEY = _

Predefined registry key, always open.

Provided Methods

fn RegEnumKeyEx(
&self
) -> SysResult<Box<dyn Iterator<Item = SysResult<String>> + '_>>

Returns an iterator over the names of the keys, which calls RegEnumKeyEx repeatedly.

Examples

use winsafe::prelude::*;
use winsafe::{co, HKEY};

let hkey = HKEY::CURRENT_USER.RegOpenKeyEx(
    "Control Panel",
    co::REG_OPTION::default(),
    co::KEY::READ,
)?;

for key_name in hkey.RegEnumKeyEx()? {
    let key_name = key_name?;
    println!("{}", key_name);
}

fn RegEnumValue(
&self
) -> SysResult<Box<dyn Iterator<Item = SysResult<(String, REG)>> + '_>>

Returns an iterator of the names and types of the values, which calls RegEnumValue repeatedly.

Examples

use winsafe::prelude::*;
use winsafe::{co, HKEY, SysResult};

let hkey = HKEY::CURRENT_USER.RegOpenKeyEx(
    "Control Panel\\Appearance",
    co::REG_OPTION::default(),
    co::KEY::READ,
)?;

for value_and_type in hkey.RegEnumValue()? {
    let (value, reg_type) = value_and_type?;
    println!("{}, {}", value, reg_type);
}

// Collecting into a Vec
let values_and_types: Vec<(String, co::REG)> =
    hkey.RegEnumValue()?
        .collect::<SysResult<Vec<_>>>()?;

fn RegFlushKey(&self) -> SysResult<()>

RegFlushKey method.

fn RegGetValue(
&self,
sub_key: Option<&str>,
value: Option<&str>
) -> SysResult<RegistryValue>

RegGetValue method.

The data type will be automatically queried with a first call to RegGetValue.

Note that this method validates some race conditions, returning co::ERROR::TRANSACTION_REQUEST_NOT_VALID and co::ERROR::INVALID_DATA.

Examples

use winsafe::prelude::*;
use winsafe::{HKEY, RegistryValue};

let val = HKEY::CURRENT_USER.RegGetValue(
    Some("Control Panel\\Mouse"),
    Some("Beep"),
)?;

match val {
    RegistryValue::Dword(n) => println!("Number u32: {}", n),
    RegistryValue::Qword(n) => println!("Number u64: {}", n),
    RegistryValue::Sz(str) => println!("String: {}", str),
    RegistryValue::MultiSz(strs) => {
       println!("Multi string:");
       for s in strs.iter() {
           print!("[{}] ", s);
       }
       println!("");
    },
    RegistryValue::Binary(bin) => {
        println!("Binary:");
        for b in bin.iter() {
            print!("{:02x} ", b);
        }
        println!("");
    },
    RegistryValue::None => println!("No value"),
}

fn RegOpenKeyEx(
&self,
sub_key: &str,
options: REG_OPTION,
access_rights: KEY
) -> SysResult<HkeyGuard>

RegOpenKeyEx method.

Examples

use winsafe::prelude::*;
use winsafe::{co, HKEY};

let hkey = HKEY::CURRENT_USER.RegOpenKeyEx(
    "Control Panel\\Mouse",
    co::REG_OPTION::default(),
    co::KEY::READ,
)?;

fn RegQueryInfoKey(
&self,
class: Option<&mut WString>,
num_sub_keys: Option<&mut u32>,
max_sub_key_name_len: Option<&mut u32>,
max_class_len: Option<&mut u32>,
num_values: Option<&mut u32>,
max_value_name_len: Option<&mut u32>,
max_value_len: Option<&mut u32>,
security_descr_len: Option<&mut u32>,
last_write_time: Option<&mut FILETIME>
) -> SysResult<()>

RegQueryInfoKey method.

fn RegQueryValueEx(&self, value: &str) -> SysResult<RegistryValue>

RegQueryValueEx method.

The data type will be automatically queried with a first call to RegQueryValueEx.

Note that this method validates some race conditions, returning co::ERROR::TRANSACTION_REQUEST_NOT_VALID and co::ERROR::INVALID_DATA.

Examples

use winsafe::prelude::*;
use winsafe::{co, HKEY, RegistryValue};

let hkey = HKEY::CURRENT_USER.RegOpenKeyEx(
    "Control Panel\\Mouse",
    co::REG_OPTION::default(),
    co::KEY::READ,
)?;

let val = hkey.RegQueryValueEx("Beep")?;

match val {
    RegistryValue::Dword(n) => println!("Number u32: {}", n),
    RegistryValue::Qword(n) => println!("Number u64: {}", n),
    RegistryValue::Sz(str) => println!("String: {}", str),
    RegistryValue::MultiSz(strs) => {
       println!("Multi string:");
       for s in strs.iter() {
           print!("[{}] ", s);
       }
       println!("");
    },
    RegistryValue::Binary(bin) => {
        println!("Binary:");
        for b in bin.iter() {
            print!("{:02x} ", b);
        }
        println!("");
    },
    RegistryValue::None => println!("No value"),
}

fn RegSetKeyValue(
&self,
sub_key: &str,
value: &str,
data: RegistryValue
) -> SysResult<()>

RegSetKeyValue method.

If the value doesn’t exist, if will be created. If new type is different from current type, new type will take over.

Examples

use winsafe::prelude::*;
use winsafe::{HKEY, RegistryValue, WString};

HKEY::CURRENT_USER.RegSetKeyValue(
    "Software\\My Company",
    "Color",
    RegistryValue::Sz("blue".to_owned()),
)?;

fn RegSetValueEx(&self, value: &str, data: RegistryValue) -> SysResult<()>

RegSetValueEx method.

If the value doesn’t exist, if will be created. If new type is different from current type, new type will prevail.

Examples

use winsafe::prelude::*;
use winsafe::{co, HKEY, RegistryValue, WString};

let hkey = HKEY::CURRENT_USER.RegOpenKeyEx(
    "Console\\Git Bash",
    co::REG_OPTION::default(),
    co::KEY::ALL_ACCESS,
)?;

hkey.RegSetValueEx(
    "Color",
    RegistryValue::Sz("blue".to_owned()),
)?;