Trait winsafe::prelude::advapi_Hkey

pub trait advapi_Hkey: Handle {
    const CLASSES_ROOT: HKEY = _;
    const CURRENT_USER: HKEY = _;
    const LOCAL_MACHINE: HKEY = _;
    const USERS: HKEY = _;
    const PERFORMANCE_DATA: HKEY = _;
    const PERFORMANCE_TEXT: HKEY = _;
    const PERFORMANCE_NLSTEXT: HKEY = _;
    const CURRENT_CONFIG: HKEY = _;
    const DYN_DATA: HKEY = _;
    const CURRENT_USER_LOCAL_SETTINGS: HKEY = _;

    fn CloseKey(self) -> SysResult<()> { ... }
    fn EnumKeyEx<'a>(
        self
    ) -> SysResult<Box<dyn Iterator<Item = SysResult<String>> + 'a>> { ... }
    fn EnumValue<'a>(
        self
    ) -> SysResult<Box<dyn Iterator<Item = SysResult<(String, REG)>> + 'a>> { ... }
    fn GetValue(self, sub_key: &str, value: &str) -> SysResult<RegistryValue> { ... }
    fn OpenKeyEx(
        self,
        sub_key: &str,
        options: REG_OPTION,
        access_rights: KEY
    ) -> SysResult<HKEY> { ... }
    fn QueryInfoKey(
        self,
        class: Option<&mut WString>,
        num_sub_keys: Option<&mut u32>,
        max_sub_key_name_len: Option<&mut u32>,
        max_class_len: Option<&mut u32>,
        num_values: Option<&mut u32>,
        max_value_name_len: Option<&mut u32>,
        max_value_len: Option<&mut u32>,
        security_descr_len: Option<&mut u32>,
        last_write_time: Option<&mut FILETIME>
    ) -> SysResult<()> { ... }
    fn QueryValueEx(self, value: &str) -> SysResult<RegistryValue> { ... }
    fn SetKeyValue(
        self,
        sub_key: &str,
        value: &str,
        data: RegistryValue
    ) -> SysResult<()> { ... }
    fn SetValueEx(self, value: &str, data: RegistryValue) -> SysResult<()> { ... }
}

Available on crate feature advapi only.

This trait is enabled with the advapi feature, and provides methods for HKEY.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Associated Constants

const CLASSES_ROOT: HKEY = _

Predefined registry key, always open.

const CURRENT_USER: HKEY = _

Predefined registry key, always open.

const LOCAL_MACHINE: HKEY = _

Predefined registry key, always open.

const USERS: HKEY = _

Predefined registry key, always open.

const PERFORMANCE_DATA: HKEY = _

Predefined registry key, always open.

const PERFORMANCE_TEXT: HKEY = _

Predefined registry key, always open.

const PERFORMANCE_NLSTEXT: HKEY = _

Predefined registry key, always open.

const CURRENT_CONFIG: HKEY = _

Predefined registry key, always open.

const DYN_DATA: HKEY = _

Predefined registry key, always open.

const CURRENT_USER_LOCAL_SETTINGS: HKEY = _

Predefined registry key, always open.

Provided Methods

fn CloseKey(self) -> SysResult<()>

RegCloseKey method.

fn EnumKeyEx<'a>(
self
) -> SysResult<Box<dyn Iterator<Item = SysResult<String>> + 'a>>

Returns an iterator over the names of the keys, which calls RegEnumKeyEx repeatedly.

Examples

use winsafe::prelude::*;
use winsafe::{co, HKEY};

let hkey = HKEY::CURRENT_USER.OpenKeyEx(
    "Control Panel",
    co::REG_OPTION::default(),
    co::KEY::READ,
)?;

for key_name in hkey.EnumKeyEx()? {
    let key_name = key_name?;
    println!("{}", key_name);
}

hkey.CloseKey()?;

fn EnumValue<'a>(
self
) -> SysResult<Box<dyn Iterator<Item = SysResult<(String, REG)>> + 'a>>

Returns an iterator of the names and types of the values, which calls RegEnumValue repeatedly.

Examples

use winsafe::prelude::*;
use winsafe::{co, HKEY};

let hkey = HKEY::CURRENT_USER.OpenKeyEx(
    "Control Panel\\Appearance",
    co::REG_OPTION::default(),
    co::KEY::READ,
)?;

for value_and_type in hkey.EnumValue()? {
    let (value, reg_type) = value_and_type?;
    println!("{}, {}", value, reg_type);
}

hkey.CloseKey()?;

fn GetValue(self, sub_key: &str, value: &str) -> SysResult<RegistryValue>

RegGetValue method.

The data type will be automatically queried with a first call to RegGetValue.

Examples

use winsafe::prelude::*;
use winsafe::{HKEY, RegistryValue};

let val = HKEY::CURRENT_USER.GetValue(
    "Control Panel\\Mouse",
    "Beep",
)?;

match val {
    RegistryValue::Dword(n) => println!("Number u32: {}", n),
    RegistryValue::Qword(n) => println!("Number u64: {}", n),
    RegistryValue::Sz(str) => println!("String: {}", str),
    RegistryValue::Binary(bin) => {
        println!("Binary:");
        for b in bin.iter() {
            print!("{:02x} ", b);
        }
        println!("");
    },
    RegistryValue::None => println!("No value"),
}

fn OpenKeyEx(
self,
sub_key: &str,
options: REG_OPTION,
access_rights: KEY
) -> SysResult<HKEY>

RegOpenKeyEx method.

Note: Must be paired with an HKEY::CloseKey call.

Examples

use winsafe::prelude::*;
use winsafe::{co, HKEY};

let hkey = HKEY::CURRENT_USER.OpenKeyEx(
    "Control Panel\\Mouse",
    co::REG_OPTION::default(),
    co::KEY::READ,
)?;

hkey.CloseKey()?;

fn QueryInfoKey(
self,
class: Option<&mut WString>,
num_sub_keys: Option<&mut u32>,
max_sub_key_name_len: Option<&mut u32>,
max_class_len: Option<&mut u32>,
num_values: Option<&mut u32>,
max_value_name_len: Option<&mut u32>,
max_value_len: Option<&mut u32>,
security_descr_len: Option<&mut u32>,
last_write_time: Option<&mut FILETIME>
) -> SysResult<()>

RegQueryInfoKey method.

fn QueryValueEx(self, value: &str) -> SysResult<RegistryValue>

RegQueryValueEx method.

The data type will be automatically queried with a first call to RegQueryValueEx.

Examples

use winsafe::prelude::*;
use winsafe::{co, HKEY, RegistryValue};

let hkey = HKEY::CURRENT_USER.OpenKeyEx(
    "Control Panel\\Mouse",
    co::REG_OPTION::default(),
    co::KEY::READ,
)?;

let val = hkey.QueryValueEx("Beep")?;

match val {
    RegistryValue::Dword(n) => println!("Number u32: {}", n),
    RegistryValue::Qword(n) => println!("Number u64: {}", n),
    RegistryValue::Sz(str) => println!("String: {}", str),
    RegistryValue::Binary(bin) => {
        println!("Binary:");
        for b in bin.iter() {
            print!("{:02x} ", b);
        }
        println!("");
    },
    RegistryValue::None => println!("No value"),
}

hkey.CloseKey()?;

fn SetKeyValue(
self,
sub_key: &str,
value: &str,
data: RegistryValue
) -> SysResult<()>

RegSetKeyValue method.

If the value doesn’t exist, if will be created. If new type is different from current type, new type will take over.

Examples

use winsafe::prelude::*;
use winsafe::{HKEY, RegistryValue, WString};

HKEY::CURRENT_USER.SetKeyValue(
    "Software\\My Company",
    "Color",
    RegistryValue::Sz(WString::from_str("blue")),
)?;

fn SetValueEx(self, value: &str, data: RegistryValue) -> SysResult<()>

RegSetValueEx method.

If the value doesn’t exist, if will be created. If new type is different from current type, new type will prevail.

Examples

use winsafe::prelude::*;
use winsafe::{co, HKEY, RegistryValue, WString};

let hkey = HKEY::CURRENT_USER.OpenKeyEx(
    "Console\\Git Bash",
    co::REG_OPTION::default(),
    co::KEY::ALL_ACCESS,
)?;

hkey.SetValueEx(
    "Color",
    RegistryValue::Sz(WString::from_str("blue")),
)?;

hkey.CloseKey()?;