Trait winsafe::prelude::comctl_Himagelist

pub trait comctl_Himagelist: Handle {

    fn Add(
        &self,
        hbmp_image: &HBITMAP,
        hbmp_mask: Option<&HBITMAP>
    ) -> SysResult<u32> { ... }
    fn AddIcon(&self, hicon: &HICON) -> SysResult<u32> { ... }
    fn AddMasked(
        &self,
        hbmp_image: &HBITMAP,
        color_mask: COLORREF
    ) -> SysResult<u32> { ... }
    fn BeginDrag(
        &self,
        itrack: u32,
        hotspot: POINT
    ) -> SysResult<HimagelistDragGuard<'_>> { ... }
    fn Create(
        image_sz: SIZE,
        flags: ILC,
        initial_size: i32,
        grow_size: i32
    ) -> SysResult<HIMAGELIST> { ... }
    fn Destroy(&self) -> SysResult<()> { ... }
    fn DragMove(&self, x: i32, y: i32) -> SysResult<()> { ... }
    fn DragShowNolock(show: bool) -> SysResult<()> { ... }
    fn GetIconSize(&self) -> SysResult<SIZE> { ... }
    fn GetImageCount(&self) -> u32 { ... }
    fn Remove(&self, index: Option<u32>) -> SysResult<()> { ... }
    fn ReplaceIcon(&self, index: Option<u32>, hicon_new: &HICON) -> SysResult<u32> { ... }
    fn SetImageCount(&self, new_count: u32) -> SysResult<()> { ... }

}

Available on crate feature comctl only.

This trait is enabled with the comctl feature, and provides methods for HIMAGELIST.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn Add(&self, hbmp_image: &HBITMAP, hbmp_mask: Option<&HBITMAP>) -> SysResult<u32>

ImageList_Add method.

A copy of the bitmap is made and stored in the image list, so you’re free to release the original bitmap.

fn AddIcon(&self, hicon: &HICON) -> SysResult<u32>

ImageList_AddIcon method.

A copy of the icon is made and stored in the image list, so you’re free to release the original icon.

fn AddMasked(&self, hbmp_image: &HBITMAP, color_mask: COLORREF) -> SysResult<u32>

ImageList_AddMasked method.

A copy of the bitmap is made and stored in the image list, so you’re free to release the original bitmap.

fn BeginDrag(
&self,
itrack: u32,
hotspot: POINT
) -> SysResult<HimagelistDragGuard<'_>>

ImageList_BeginDrag method.

In the original C implementation, you must call ImageList_EndDrag as a cleanup operation.

Here, the cleanup is performed automatically, because BeginDrag returns an HimagelistDragGuard, which automatically calls ImageList_EndDrag when the guard goes out of scope. You must, however, keep the guard alive, otherwise the cleanup will be performed right away.

use winsafe::prelude::*;
use winsafe::{HIMAGELIST, POINT};

let himgl: HIMAGELIST; // initialized somewhere

let _drag = himgl.BeginDrag(0, POINT::new(0, 0))?; // keep guard alive

fn Create(
image_sz: SIZE,
flags: ILC,
initial_size: i32,
grow_size: i32
) -> SysResult<HIMAGELIST>

ImageList_Create static method.

Note: Must be paired with an HIMAGELIST::Destroy call.

Examples

use winsafe::prelude::*;
use winsafe::{co, HIMAGELIST, SIZE};

let himgl = HIMAGELIST::Create(
    SIZE::new(16, 16), co::ILC::COLOR32, 1, 1)?;

himgl.Destroy()?;

fn Destroy(&self) -> SysResult<()>

ImageList_Destroy method.

After calling this method, the handle will be invalidated and further operations will fail with ERROR::INVALID_HANDLE error code.

fn DragMove(&self, x: i32, y: i32) -> SysResult<()>

ImageList_DragMove method.

fn DragShowNolock(show: bool) -> SysResult<()>

ImageList_DragShowNolock static method.

fn GetIconSize(&self) -> SysResult<SIZE>

ImageList_GetIconSize method.

fn GetImageCount(&self) -> u32

ImageList_GetImageCount method.

fn Remove(&self, index: Option<u32>) -> SysResult<()>

ImageList_Remove method.

fn ReplaceIcon(&self, index: Option<u32>, hicon_new: &HICON) -> SysResult<u32>

ImageList_ReplaceIcon method.

Note: A copy of the bitmap is made, and this copy is then stored. You’re still responsible for freeing the original bitmap.

fn SetImageCount(&self, new_count: u32) -> SysResult<()>

ImageList_SetImageCount methods.