Trait winsafe::prelude::comctl_Himagelist

pub trait comctl_Himagelist: Handle {

    fn Add(
        self,
        hbmp_image: HBITMAP,
        hbmp_mask: Option<HBITMAP>
    ) -> SysResult<u32> { ... }
    fn AddIcon(self, hicon: HICON) -> SysResult<u32> { ... }
    fn AddMasked(
        self,
        hbmp_image: HBITMAP,
        color_mask: COLORREF
    ) -> SysResult<u32> { ... }
    fn BeginDrag(self, track: u32, hotspot: POINT) -> SysResult<()> { ... }
    fn Create(
        image_sz: SIZE,
        flags: ILC,
        initial_size: i32,
        grow_size: i32
    ) -> SysResult<HIMAGELIST> { ... }
    fn Destroy(self) -> SysResult<()> { ... }
    fn DragMove(self, x: i32, y: i32) -> SysResult<()> { ... }
    fn DragShowNolock(show: bool) -> SysResult<()> { ... }
    fn EndDrag() { ... }
    fn GetIconSize(self) -> SysResult<SIZE> { ... }
    fn GetImageCount(self) -> u32 { ... }
    fn Remove(self, index: Option<u32>) -> SysResult<()> { ... }
    fn ReplaceIcon(self, index: Option<u32>, hicon_new: HICON) -> SysResult<u32> { ... }
    fn SetImageCount(self, new_count: u32) -> SysResult<()> { ... }

}

Available on crate feature comctl only.

This trait is enabled with the comctl feature, and provides methods for HIMAGELIST.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn Add(self, hbmp_image: HBITMAP, hbmp_mask: Option<HBITMAP>) -> SysResult<u32>

ImageList_Add method.

Note: A copy of the bitmap is made, and this copy is then stored. You’re still responsible for freeing the original bitmap.

fn AddIcon(self, hicon: HICON) -> SysResult<u32>

ImageList_AddIcon method.

The icon can be destroyed right after this call, its handle its duplicated inside the image list.

fn AddMasked(self, hbmp_image: HBITMAP, color_mask: COLORREF) -> SysResult<u32>

ImageList_AddMasked method.

fn BeginDrag(self, track: u32, hotspot: POINT) -> SysResult<()>

ImageList_BeginDrag method.

Note: Must be paired with an HIMAGELIST::EndDrag call.

fn Create(
image_sz: SIZE,
flags: ILC,
initial_size: i32,
grow_size: i32
) -> SysResult<HIMAGELIST>

ImageList_Create static method.

Note: Must be paired with an HIMAGELIST::Destroy call.

Examples

use winsafe::prelude::*;
use winsafe::{co, HIMAGELIST, SIZE};

let himgl = HIMAGELIST::Create(
    SIZE::new(16, 16), co::ILC::COLOR32, 1, 1)?;

himgl.Destroy()?;

fn Destroy(self) -> SysResult<()>

ImageList_Destroy method.

fn DragMove(self, x: i32, y: i32) -> SysResult<()>

ImageList_DragMove method.

fn DragShowNolock(show: bool) -> SysResult<()>

ImageList_DragShowNolock static method.

fn EndDrag()

ImageList_EndDrag static method.

fn GetIconSize(self) -> SysResult<SIZE>

ImageList_GetIconSize method.

fn GetImageCount(self) -> u32

ImageList_GetImageCount method.

fn Remove(self, index: Option<u32>) -> SysResult<()>

ImageList_Remove method.

fn ReplaceIcon(self, index: Option<u32>, hicon_new: HICON) -> SysResult<u32>

ImageList_ReplaceIcon method.

The icon can be destroyed right after this call, its handle its duplicated inside the image list.

fn SetImageCount(self, new_count: u32) -> SysResult<()>

ImageList_SetImageCount methods.