winsafe::prelude

Trait comctl_Himagelist

pub trait comctl_Himagelist: Handle {

    // Provided methods
    fn iter(&self) -> impl Iterator<Item = HrResult<DestroyIconGuard>> + '_ { ... }
    fn Add(
        &self,
        hbmp_image: &HBITMAP,
        hbmp_mask: Option<&HBITMAP>,
    ) -> HrResult<u32> { ... }
    fn AddIcon(&self, hicon: &HICON) -> HrResult<u32> { ... }
    fn AddMasked(
        &self,
        hbmp_image: &HBITMAP,
        color_mask: COLORREF,
    ) -> HrResult<u32> { ... }
    fn BeginDrag(
        &self,
        itrack: u32,
        hotspot: POINT,
    ) -> HrResult<ImageListEndDragGuard<'_>> { ... }
    fn Create(
        image_sz: SIZE,
        flags: ILC,
        initial_size: i32,
        grow_size: i32,
    ) -> HrResult<ImageListDestroyGuard> { ... }
    fn DragMove(&self, x: i32, y: i32) -> HrResult<()> { ... }
    fn DragShowNolock(show: bool) -> HrResult<()> { ... }
    fn Draw(
        &self,
        index: u32,
        hdc_dest: &HDC,
        dest: POINT,
        style: ILD,
    ) -> HrResult<()> { ... }
    fn DrawEx(
        &self,
        index: u32,
        hdc_dest: &HDC,
        dest: POINT,
        img_portion: Option<SIZE>,
        background_color: ClrDefNone,
        foreground_color: ClrDefNone,
        style: ILD,
    ) -> HrResult<()> { ... }
    fn Duplicate(&self) -> HrResult<ImageListDestroyGuard> { ... }
    fn ExtractIcon(&self, index: u32) -> HrResult<DestroyIconGuard> { ... }
    fn GetBkColor(&self) -> COLORREF { ... }
    fn GetIcon(&self, index: u32, flags: ILD) -> HrResult<DestroyIconGuard> { ... }
    fn GetIconSize(&self) -> HrResult<SIZE> { ... }
    fn GetImageCount(&self) -> u32 { ... }
    fn Remove(&self, index: Option<u32>) -> HrResult<()> { ... }
    fn ReplaceIcon(&self, index: u32, hicon_new: &HICON) -> HrResult<u32> { ... }
    fn SetBkColor(&self, bk_color: Option<COLORREF>) -> Option<COLORREF> { ... }
    unsafe fn SetImageCount(&self, new_count: u32) -> HrResult<()> { ... }
    fn Write(&self, stream: &impl ole_IStream) -> HrResult<()> { ... }

}

Available on crate features kernel and comctl only.

This trait is enabled with the comctl feature, and provides methods for HIMAGELIST.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn iter(&self) -> impl Iterator<Item = HrResult<DestroyIconGuard>> + '_

Returns an iterator over all icons in the image list, by calling HIMAGELIST::ExtractIcon for each one.

Examples

Collecting the icons into a Vec:

use winsafe::{self as w, prelude::*};

let himgl: w::HIMAGELIST; // initialized somewhere

let icons = himgl.iter()
    .collect::<w::HrResult<Vec<_>>>()?;

fn Add( &self, hbmp_image: &HBITMAP, hbmp_mask: Option<&HBITMAP>, ) -> HrResult<u32>

ImageList_Add function.

A copy of the bitmap is made and stored in the image list, so you’re free to release the original bitmap.

fn AddIcon(&self, hicon: &HICON) -> HrResult<u32>

ImageList_AddIcon macro.

A copy of the icon is made and stored in the image list, so you’re free to release the original icon.

fn AddMasked(&self, hbmp_image: &HBITMAP, color_mask: COLORREF) -> HrResult<u32>

ImageList_AddMasked function.

A copy of the bitmap is made and stored in the image list, so you’re free to release the original bitmap.

fn BeginDrag( &self, itrack: u32, hotspot: POINT, ) -> HrResult<ImageListEndDragGuard<'_>>

ImageList_BeginDrag function.

In the original C implementation, you must call ImageList_EndDrag as a cleanup operation.

Here, the cleanup is performed automatically, because BeginDrag returns an ImageListEndDragGuard, which automatically calls ImageList_EndDrag when the guard goes out of scope. You must, however, keep the guard alive, otherwise the cleanup will be performed right away.

Examples

use winsafe::{self as w, prelude::*};

let himgl: w::HIMAGELIST; // initialized somewhere

let _drag = himgl.BeginDrag(0, w::POINT::new(0, 0))?; // keep guard alive

fn Create( image_sz: SIZE, flags: ILC, initial_size: i32, grow_size: i32, ) -> HrResult<ImageListDestroyGuard>

ImageList_Create function.

Examples

use winsafe::{self as w, prelude::*, co};

let himgl = w::HIMAGELIST::Create(
    w::SIZE::new(16, 16),
    co::ILC::COLOR32,
    1,
    1,
)?;

// ImageList_Destroy() automatically called

fn DragMove(&self, x: i32, y: i32) -> HrResult<()>

ImageList_DragMove function.

fn DragShowNolock(show: bool) -> HrResult<()>

ImageList_DragShowNolock function.

fn Draw( &self, index: u32, hdc_dest: &HDC, dest: POINT, style: ILD, ) -> HrResult<()>

ImageList_Draw function.

fn DrawEx( &self, index: u32, hdc_dest: &HDC, dest: POINT, img_portion: Option<SIZE>, background_color: ClrDefNone, foreground_color: ClrDefNone, style: ILD, ) -> HrResult<()>

ImageList_DrawEx function.

fn Duplicate(&self) -> HrResult<ImageListDestroyGuard>

ImageList_Duplicate function.

fn ExtractIcon(&self, index: u32) -> HrResult<DestroyIconGuard>

ImageList_ExtractIcon macro.

A copy of the stored icon is returned.

fn GetBkColor(&self) -> COLORREF

ImageList_GetBkColor function.

fn GetIcon(&self, index: u32, flags: ILD) -> HrResult<DestroyIconGuard>

ImageList_GetIcon function.

A copy of the stored icon is returned.

fn GetIconSize(&self) -> HrResult<SIZE>

ImageList_GetIconSize function.

fn GetImageCount(&self) -> u32

ImageList_GetImageCount function.

fn Remove(&self, index: Option<u32>) -> HrResult<()>

ImageList_Remove function.

fn ReplaceIcon(&self, index: u32, hicon_new: &HICON) -> HrResult<u32>

ImageList_ReplaceIcon function.

A copy of the icon is made and stored in the image list, so you’re free to release the original icon.

fn SetBkColor(&self, bk_color: Option<COLORREF>) -> Option<COLORREF>

ImageList_SetBkColor function.

unsafe fn SetImageCount(&self, new_count: u32) -> HrResult<()>

ImageList_SetImageCount function.

Safety

If the size is increased, you must call HIMAGELIST::ReplaceIcon to fill the new indexes, otherwise draw operations will be unpredictable.

fn Write(&self, stream: &impl ole_IStream) -> HrResult<()>

ImageList_Write function.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.