Trait winsafe::prelude::comctl_Hwnd

pub trait comctl_Hwnd: Handle {
    fn DefSubclassProc<M>(self, msg: M) -> M::RetType
   where
        M: MsgSend,
    { ... }
    fn RemoveWindowSubclass(
        self,
        subclass_func: SUBCLASSPROC,
        subclass_id: usize
    ) -> SysResult<()> { ... }
    fn SetWindowSubclass(
        self,
        subclass_proc: SUBCLASSPROC,
        subclass_id: usize,
        ref_data: usize
    ) -> SysResult<()> { ... }
}

Available on crate feature comctl only.

This trait is enabled with the comctl feature, and provides methods for HWND.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn DefSubclassProc<M>(self, msg: M) -> M::RetTypewhere
M: MsgSend,

DefSubclassProc method.

The return type is variable, being defined by the RetType associated type of the MsgSend trait. That means each message can define its own return type.

fn RemoveWindowSubclass(
self,
subclass_func: SUBCLASSPROC,
subclass_id: usize
) -> SysResult<()>

RemoveWindowSubclass method.

fn SetWindowSubclass(
self,
subclass_proc: SUBCLASSPROC,
subclass_id: usize,
ref_data: usize
) -> SysResult<()>

SetWindowSubclass method.