Trait winsafe::prelude::comctl_shell_Himagelist

pub trait comctl_shell_Himagelist: comctl_Himagelist {
    fn add_icon_from_shell(self, file_extensions: &[&str]) -> SysResult<()> { ... }
}

Available on crate features comctl and shell only.

This trait is enabled with comctl and shell features, and provides methods for HIMAGELIST.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn add_icon_from_shell(self, file_extensions: &[&str]) -> SysResult<()>

Calls SHGetFileInfo to retrieve one or more shell file icons, then passes them to AddIcon.

Examples

use winsafe::prelude::*;
use winsafe::{co, HIMAGELIST, SIZE};

let himgl = HIMAGELIST::Create(
    SIZE::new(16, 16), co::ILC::COLOR32, 1, 1)?;

himgl.add_icon_from_shell(&["mp3", "wav"])?;

himgl.Destroy()?;