winsafe::prelude

Trait comctl_shell_Himagelist

pub trait comctl_shell_Himagelist: comctl_Himagelist {
    // Provided method
    fn add_icons_from_shell(
        &self,
        file_extensions: &[impl AsRef<str>],
    ) -> HrResult<()> { ... }
}

Available on crate features comctl and shell and kernel only.

This trait is enabled with comctl and shell features, and provides methods for HIMAGELIST.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn add_icons_from_shell( &self, file_extensions: &[impl AsRef<str>], ) -> HrResult<()>

Calls SHGetFileInfo to retrieve one or more shell file icons, then passes them to AddIcon.

Examples

use winsafe::{self as w, prelude::*, co};

let mut himgl = w::HIMAGELIST::Create(
    w::SIZE::new(16, 16),
    co::ILC::COLOR32,
    1,
    1,
)?;

himgl.add_icons_from_shell(&["mp3", "wav"])?;

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.