Trait winsafe::prelude::gdi_Hbrush

pub trait gdi_Hbrush: gdi_Hgdiobj {
    fn from_sys_color(color: COLOR) -> HBRUSH { ... }
    fn CreateBrushIndirect(lb: &LOGBRUSH) -> WinResult<HBRUSH> { ... }
    fn CreateHatchBrush(hatch: HS, color: COLORREF) -> WinResult<HBRUSH> { ... }
    fn CreatePatternBrush(hbmp: HBITMAP) -> WinResult<HBRUSH> { ... }
    fn CreateSolidBrush(color: COLORREF) -> WinResult<HBRUSH> { ... }
    fn GetObject(self, pv: &mut LOGBRUSH) -> WinResult<()> { ... }
    fn GetStockObject(sb: STOCK_BRUSH) -> WinResult<HBRUSH> { ... }
    fn GetSysColorBrush(index: COLOR) -> WinResult<HBRUSH> { ... }
    fn UnrealizeObject(self) -> WinResult<()> { ... }
}

Available on crate feature gdi only.

This trait is enabled with the gdi feature, and provides methods for HBRUSH.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn from_sys_color(color: COLOR) -> HBRUSH

Creates a brush with the given system color.

Note: This should be used only to initialize the WNDCLASSEX’s hbrBackground field. Any other use will yield an invalid handle.

fn CreateBrushIndirect(lb: &LOGBRUSH) -> WinResult<HBRUSH>

CreateBrushIndirect static method.

Note: Must be paired with an HBRUSH::DeleteObject call.

fn CreateHatchBrush(hatch: HS, color: COLORREF) -> WinResult<HBRUSH>

CreateHatchBrush static method.

Note: Must be paired with an HBRUSH::DeleteObject call.

fn CreatePatternBrush(hbmp: HBITMAP) -> WinResult<HBRUSH>

CreatePatternBrush static method.

Note: Must be paired with an HBRUSH::DeleteObject call.

fn CreateSolidBrush(color: COLORREF) -> WinResult<HBRUSH>

CreateSolidBrush static method.

Note: Must be paired with an HBRUSH::DeleteObject call.

fn GetObject(self, pv: &mut LOGBRUSH) -> WinResult<()>

GetObject method.

fn GetStockObject(sb: STOCK_BRUSH) -> WinResult<HBRUSH>

GetStockObject static method.

fn GetSysColorBrush(index: COLOR) -> WinResult<HBRUSH>

GetSysColorBrush static method.

fn UnrealizeObject(self) -> WinResult<()>

UnrealizeObject method.