Trait winsafe::prelude::gdi_Hbrush

pub trait gdi_Hbrush: GdiObject {
    fn from_sys_color(color: COLOR) -> HBRUSH { ... }
    fn CreateBrushIndirect(lb: &LOGBRUSH) -> SysResult<GdiObjectGuard<HBRUSH>> { ... }
    fn CreateHatchBrush(
        hatch: HS,
        color: COLORREF
    ) -> SysResult<GdiObjectGuard<HBRUSH>> { ... }
    fn CreatePatternBrush(hbmp: &HBITMAP) -> SysResult<GdiObjectGuard<HBRUSH>> { ... }
    fn CreateSolidBrush(color: COLORREF) -> SysResult<GdiObjectGuard<HBRUSH>> { ... }
    fn GetObject(&self, pv: &mut LOGBRUSH) -> SysResult<()> { ... }
    fn GetStockObject(sb: STOCK_BRUSH) -> SysResult<HBRUSH> { ... }
    fn GetSysColorBrush(index: COLOR) -> SysResult<HBRUSH> { ... }
    fn UnrealizeObject(&self) -> SysResult<()> { ... }
}

Available on crate feature gdi only.

This trait is enabled with the gdi feature, and provides methods for HBRUSH.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn from_sys_color(color: COLOR) -> HBRUSH

Creates a brush with the given system color.

Note: This should be used only to initialize the WNDCLASSEX’s hbrBackground field. Any other use will yield an invalid handle.

fn CreateBrushIndirect(lb: &LOGBRUSH) -> SysResult<GdiObjectGuard<HBRUSH>>

CreateBrushIndirect static method.

fn CreateHatchBrush(
hatch: HS,
color: COLORREF
) -> SysResult<GdiObjectGuard<HBRUSH>>

CreateHatchBrush static method.

fn CreatePatternBrush(hbmp: &HBITMAP) -> SysResult<GdiObjectGuard<HBRUSH>>

CreatePatternBrush static method.

fn CreateSolidBrush(color: COLORREF) -> SysResult<GdiObjectGuard<HBRUSH>>

CreateSolidBrush static method.

fn GetObject(&self, pv: &mut LOGBRUSH) -> SysResult<()>

GetObject method.

fn GetStockObject(sb: STOCK_BRUSH) -> SysResult<HBRUSH>

GetStockObject static method.

fn GetSysColorBrush(index: COLOR) -> SysResult<HBRUSH>

GetSysColorBrush static method.

fn UnrealizeObject(&self) -> SysResult<()>

UnrealizeObject method.