Trait winsafe::prelude::kernel_Hfile

pub trait kernel_Hfile: Handle {
    fn CreateFile(
        file_name: &str,
        desired_access: GENERIC,
        share_mode: FILE_SHARE,
        security_attrs: Option<&mut SECURITY_ATTRIBUTES<'_>>,
        creation_disposition: DISPOSITION,
        flags_and_attrs: FILE_ATTRIBUTE,
        hfile_template: Option<&HFILE>
    ) -> SysResult<(HandleGuard<HFILE>, ERROR)> { ... }
    fn CreateFileMapping(
        &self,
        mapping_attrs: Option<&mut SECURITY_ATTRIBUTES<'_>>,
        protect: PAGE,
        max_size: Option<u64>,
        mapping_name: Option<&str>
    ) -> SysResult<HandleGuard<HFILEMAP>> { ... }
    fn GetFileInformationByHandle(
        &self,
        fi: &mut BY_HANDLE_FILE_INFORMATION
    ) -> SysResult<()> { ... }
    fn GetFileSizeEx(&self) -> SysResult<usize> { ... }
    fn GetFileType(&self) -> SysResult<FILE_TYPE> { ... }
    fn LockFile(
        &self,
        offset: u64,
        num_bytes_to_lock: u64
    ) -> SysResult<HfileLockGuard<'_, Self>> { ... }
    fn ReadFile(
        &self,
        buffer: &mut [u8],
        overlapped: Option<&mut OVERLAPPED>
    ) -> SysResult<u32> { ... }
    fn SetEndOfFile(&self) -> SysResult<()> { ... }
    fn SetFilePointerEx(
        &self,
        distance_to_move: i64,
        move_method: FILE_STARTING_POINT
    ) -> SysResult<i64> { ... }
    fn WriteFile(
        &self,
        data: &[u8],
        overlapped: Option<&mut OVERLAPPED>
    ) -> SysResult<u32> { ... }
}

Available on crate feature kernel only.

This trait is enabled with the kernel feature, and provides methods for HFILE.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn CreateFile(
file_name: &str,
desired_access: GENERIC,
share_mode: FILE_SHARE,
security_attrs: Option<&mut SECURITY_ATTRIBUTES<'_>>,
creation_disposition: DISPOSITION,
flags_and_attrs: FILE_ATTRIBUTE,
hfile_template: Option<&HFILE>
) -> SysResult<(HandleGuard<HFILE>, ERROR)>

CreateFile static method.

The error code is also returned because it can carry information even if the file is successfully open.

Examples

Opening an existing file as read-only:

use winsafe::prelude::*;
use winsafe::{co, HFILE};

let (hfile, status) = HFILE::CreateFile(
    "C:\\Temp\\test.txt",
    co::GENERIC::READ,
    co::FILE_SHARE::READ,
    None,
    co::DISPOSITION::OPEN_EXISTING,
    co::FILE_ATTRIBUTE::NORMAL,
    None,
)?;

Opening a file for read and write. If the file doesn’t exist, create it:

use winsafe::prelude::*;
use winsafe::{co, HFILE};

let (hfile, status) = HFILE::CreateFile(
    "C:\\Temp\\test.txt",
    co::GENERIC::READ | co::GENERIC::WRITE,
    co::FILE_SHARE::NoValue,
    None,
    co::DISPOSITION::OPEN_ALWAYS,
    co::FILE_ATTRIBUTE::NORMAL,
    None,
)?;

fn CreateFileMapping(
&self,
mapping_attrs: Option<&mut SECURITY_ATTRIBUTES<'_>>,
protect: PAGE,
max_size: Option<u64>,
mapping_name: Option<&str>
) -> SysResult<HandleGuard<HFILEMAP>>

CreateFileMapping method.

fn GetFileInformationByHandle(
&self,
fi: &mut BY_HANDLE_FILE_INFORMATION
) -> SysResult<()>

GetFileInformationByHandle method.

fn GetFileSizeEx(&self) -> SysResult<usize>

GetFileSizeEx method.

fn GetFileType(&self) -> SysResult<FILE_TYPE>

GetFileType method.

fn LockFile(
&self,
offset: u64,
num_bytes_to_lock: u64
) -> SysResult<HfileLockGuard<'_, Self>>

LockFile method.

Note that the guard returned by this method must be kept alive while the lock is used. This is necessary because, when the guard goes out of scope, it will automatically call UnlockFile.

In the example below, the guard is kept alive:

use winsafe::prelude::*;
use winsafe::HFILE;

let hfile: HFILE; // initialized somewhere

let total_size = hfile.GetFileSizeEx()?;

let _lock = hfile.LockFile(0, total_size as _)?; // keep the returned guard alive

// file read/write operations...

fn ReadFile(
&self,
buffer: &mut [u8],
overlapped: Option<&mut OVERLAPPED>
) -> SysResult<u32>

ReadFile method.

Returns the number of bytes read.

fn SetEndOfFile(&self) -> SysResult<()>

SetEndOfFile method.

fn SetFilePointerEx(
&self,
distance_to_move: i64,
move_method: FILE_STARTING_POINT
) -> SysResult<i64>

SetFilePointerEx method.

fn WriteFile(
&self,
data: &[u8],
overlapped: Option<&mut OVERLAPPED>
) -> SysResult<u32>

WriteFile method.

Returns the number of bytes written.