Trait winsafe::prelude::kernel_Hfilemapview

pub trait kernel_Hfilemapview: Handle {
    fn as_mut_slice<'a>(self, len: usize) -> &'a mut [u8] { ... }
    fn as_slice<'a>(self, len: usize) -> &'a [u8] { ... }
    fn UnmapViewOfFile(self) -> SysResult<()> { ... }
}

Available on crate feature kernel only.

This trait is enabled with the kernel feature, and provides methods for HFILEMAPVIEW.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn as_mut_slice<'a>(self, len: usize) -> &'a mut [u8]

Returns a slice representing the mapped memory. You can modify the contents. You should call this method only if the file has write access.

Note: If the file is resized to a smaller size, the slice will still map the bytes beyond the file. This may cause serious errors. So, if the file is resized, re-generate the slice by calling as_slice again.

fn as_slice<'a>(self, len: usize) -> &'a [u8]

Returns a slice representing the mapped memory.

Note: If the file is resized to a smaller size, the slice will still map the bytes beyond the file. This may cause serious errors. So, if the file is resized, re-generate the slice by calling as_slice again.

Examples

Reading the contents of a file into a string:

use winsafe::prelude::*;
use winsafe::{co, HFILE};

let (hfile, _) = HFILE::CreateFile(
    "C:\\Temp\\test.txt",
    co::GENERIC::READ,
    co::FILE_SHARE::READ,
    None,
    co::DISPOSITION::OPEN_EXISTING,
    co::FILE_ATTRIBUTE::NORMAL,
    None,
)?;

let hmap = hfile.CreateFileMapping(
    None,
    co::PAGE::READONLY,
    None,
    None,
)?;

let view = hmap.MapViewOfFile(co::FILE_MAP::READ, 0, None)?;

let slice = view.as_slice(hfile.GetFileSizeEx()?);
let text = std::str::from_utf8(slice)?;

view.UnmapViewOfFile()?;
hmap.CloseHandle()?;
hfile.CloseHandle()?;

println!("{}", text);

fn UnmapViewOfFile(self) -> SysResult<()>

UnmapViewOfFile method.