Trait winsafe::prelude::kernel_Hfindfile

pub trait kernel_Hfindfile: Handle {
    fn iter<'a>(
        path_and_pattern: &'a str
    ) -> Box<dyn Iterator<Item = SysResult<String>> + 'a> { ... }
    fn FindFirstFile(
        file_name: &str,
        wfd: &mut WIN32_FIND_DATA
    ) -> SysResult<(HfindfileGuard, bool)> { ... }
    fn FindNextFile(&self, wfd: &mut WIN32_FIND_DATA) -> SysResult<bool> { ... }
}

Available on crate feature kernel only.

This trait is enabled with the kernel feature, and provides methods for HFINDFILE.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn iter<'a>(
path_and_pattern: &'a str
) -> Box<dyn Iterator<Item = SysResult<String>> + 'a>

Returns an iterator over the found items, by calling HFINDFILE::FindFirstFile and HFINDFILE::FindNextFile. The handle is automatically closed.

Examples

Enumerating all TXT files in a directory:

use winsafe::prelude::*;
use winsafe::{AnyResult, HFINDFILE};

for file_path in HFINDFILE::iter("C:\\Temp\\*.txt") {
    let file_path = file_path?;
    println!("File: {}", file_path);
}

HFINDFILE::iter("C:\\Temp\\*.txt")
    .try_for_each(|file_path| -> AnyResult<()> {
        let file_path = file_path?;
        println!("File: {}", file_path);
        Ok(())
    })?;

Collecting the TXTs into a Vec:

use winsafe::prelude::*;
use winsafe::{HFINDFILE, SysResult};

let file_paths = HFINDFILE::iter("C:\\Temp\\*.txt")
    .collect::<SysResult<Vec<_>>>()?;

fn FindFirstFile(
file_name: &str,
wfd: &mut WIN32_FIND_DATA
) -> SysResult<(HfindfileGuard, bool)>

FindFirstFile static method.

This method is rather tricky, consider using HFINDFILE::iter.

fn FindNextFile(&self, wfd: &mut WIN32_FIND_DATA) -> SysResult<bool>

FindNextFile method.

This method is rather tricky, consider using HFINDFILE::iter.