Trait winsafe::prelude::kernel_Hglobal

pub trait kernel_Hglobal: Handle {
    fn GlobalAlloc(flags: GMEM, num_bytes: usize) -> SysResult<HGLOBAL> { ... }
    fn GlobalFlags(self) -> SysResult<GMEM> { ... }
    fn GlobalFree(self) -> SysResult<()> { ... }
    fn GlobalLock<'a>(self) -> SysResult<&'a mut [u8]> { ... }
    fn GlobalReAlloc(self, num_bytes: usize, flags: GMEM) -> SysResult<HGLOBAL> { ... }
    fn GlobalSize(self) -> SysResult<usize> { ... }
    fn GlobalUnlock(self) -> SysResult<()> { ... }
}

Available on crate feature kernel only.

This trait is enabled with the kernel feature, and provides methods for HGLOBAL.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn GlobalAlloc(flags: GMEM, num_bytes: usize) -> SysResult<HGLOBAL>

GlobalAlloc static method.

Note: Must be paired with an HGLOBAL::GlobalFree call.

fn GlobalFlags(self) -> SysResult<GMEM>

GlobalFlags method.

fn GlobalFree(self) -> SysResult<()>

GlobalFree method.

fn GlobalLock<'a>(self) -> SysResult<&'a mut [u8]>

GlobalLock method.

Calls HGLOBAL::GlobalSize to retrieve the size of the memory block.

Note: Must be paired with an HGLOBAL::GlobalUnlock call.

fn GlobalReAlloc(self, num_bytes: usize, flags: GMEM) -> SysResult<HGLOBAL>

GlobalReAlloc method.

Note: Must be paired with an HGLOBAL::GlobalFree call.

fn GlobalSize(self) -> SysResult<usize>

GlobalSize method.

fn GlobalUnlock(self) -> SysResult<()>

GlobalUnlock method.