Trait winsafe::prelude::kernel_Hglobal

pub trait kernel_Hglobal: Handle {
    fn GlobalAlloc(flags: GMEM, num_bytes: usize) -> SysResult<HglobalGuard> { ... }
    fn GlobalFlags(&self) -> SysResult<GMEM> { ... }
    fn GlobalLock(&self) -> SysResult<&mut [u8]> { ... }
    fn GlobalReAlloc(&self, num_bytes: usize, flags: GMEM) -> SysResult<()> { ... }
    fn GlobalSize(&self) -> SysResult<usize> { ... }
    fn GlobalUnlock(&self) -> SysResult<()> { ... }
}

Available on crate feature kernel only.

This trait is enabled with the kernel feature, and provides methods for HGLOBAL.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn GlobalAlloc(flags: GMEM, num_bytes: usize) -> SysResult<HglobalGuard>

GlobalAlloc static method.

fn GlobalFlags(&self) -> SysResult<GMEM>

GlobalFlags method.

fn GlobalLock(&self) -> SysResult<&mut [u8]>

GlobalLock method.

Calls HGLOBAL::GlobalSize to retrieve the size of the memory block.

Note: Must be paired with an HGLOBAL::GlobalUnlock call.

fn GlobalReAlloc(&self, num_bytes: usize, flags: GMEM) -> SysResult<()>

GlobalReAlloc method.

Originally this method returns the handle to the reallocated memory object; here the original handle is automatically updated.

fn GlobalSize(&self) -> SysResult<usize>

GlobalSize method.

fn GlobalUnlock(&self) -> SysResult<()>

GlobalUnlock method.