Trait winsafe::prelude::kernel_Hinstance

pub trait kernel_Hinstance: Handle {
    fn EnumResourceLanguages<F>(
        &self,
        resource_type: RtStr,
        resource_id: IdStr,
        func: F
    ) -> SysResult<()>
    where
        F: Fn(LANGID) -> bool,
    { ... }
    fn EnumResourceNames<F>(&self, resource_type: RtStr, func: F) -> SysResult<()>
    where
        F: Fn(IdStr) -> bool,
    { ... }
    fn EnumResourceTypes<F>(&self, func: F) -> SysResult<()>
    where
        F: Fn(RtStr) -> bool,
    { ... }
    fn FindResource(
        &self,
        resource_id: IdStr,
        resource_type: RtStr
    ) -> SysResult<HRSRC> { ... }
    fn FindResourceEx(
        &self,
        resource_id: IdStr,
        resource_type: RtStr,
        language: Option<LANGID>
    ) -> SysResult<HRSRC> { ... }
    fn GetModuleFileName(&self) -> SysResult<String> { ... }
    fn GetModuleHandle(module_name: Option<&str>) -> SysResult<HINSTANCE> { ... }
    fn GetProcAddress(&self, proc_name: &str) -> SysResult<*const c_void> { ... }
    fn LoadLibrary(lib_file_name: &str) -> SysResult<HinstanceGuard> { ... }
    fn LoadResource(&self, res_info: &HRSRC) -> SysResult<HRSRCMEM> { ... }
    fn LockResource(
        &self,
        res_info: &HRSRC,
        hres_loaded: &HRSRCMEM
    ) -> SysResult<&[u8]> { ... }
    fn SizeofResource(&self, res_info: &HRSRC) -> SysResult<u32> { ... }
}

Available on crate feature kernel only.

This trait is enabled with the kernel feature, and provides methods for HINSTANCE.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn EnumResourceLanguages<F>(
&self,
resource_type: RtStr,
resource_id: IdStr,
func: F
) -> SysResult<()>where
F: Fn(LANGID) -> bool,

EnumResourceLanguages method.

fn EnumResourceNames<F>(&self, resource_type: RtStr, func: F) -> SysResult<()>where
F: Fn(IdStr) -> bool,

EnumResourceNames method.

fn EnumResourceTypes<F>(&self, func: F) -> SysResult<()>where
F: Fn(RtStr) -> bool,

EnumResourceTypes method.

fn FindResource(
&self,
resource_id: IdStr,
resource_type: RtStr
) -> SysResult<HRSRC>

FindResource method.

For an example, see HINSTANCE::LockResource.

fn FindResourceEx(
&self,
resource_id: IdStr,
resource_type: RtStr,
language: Option<LANGID>
) -> SysResult<HRSRC>

FindResourceEx method.

For an example, see HINSTANCE::LockResource.

fn GetModuleFileName(&self) -> SysResult<String>

GetModuleFileName method.

Examples

Retrieving the full path of currently running .exe file:

use winsafe::prelude::*;
use winsafe::HINSTANCE;

let exe_name = HINSTANCE::NULL.GetModuleFileName()?;

println!("EXE: {}", exe_name);

fn GetModuleHandle(module_name: Option<&str>) -> SysResult<HINSTANCE>

GetModuleHandle static method.

Examples

Retrieving current module instance:

use winsafe::prelude::*;
use winsafe::HINSTANCE;

let hinstance = HINSTANCE::GetModuleHandle(None)?;

fn GetProcAddress(&self, proc_name: &str) -> SysResult<*const c_void>

GetProcAddress method.

fn LoadLibrary(lib_file_name: &str) -> SysResult<HinstanceGuard>

LoadLibrary static method.

fn LoadResource(&self, res_info: &HRSRC) -> SysResult<HRSRCMEM>

LoadResource method.

For an example, see HINSTANCE::LockResource.

fn LockResource(
&self,
res_info: &HRSRC,
hres_loaded: &HRSRCMEM
) -> SysResult<&[u8]>

LockResource method.

This method should belong to HRSRCMEM, but in order to make it safe, we automatically call HINSTANCE::SizeofResource, so it’s implemented here.

Examples

The Updating Resources example:

use winsafe::prelude::*;
use winsafe::{
    co, HINSTANCE, HUPDATERSRC, IdStr, LANGID, RtStr,
};

const IDD_HAND_ABOUTBOX: u16 = 103;
const IDD_FOOT_ABOUTBOX: u16 = 110;

let hExe = HINSTANCE::LoadLibrary("hand.exe")?;

let hRes = hExe.FindResource(
    IdStr::Id(IDD_HAND_ABOUTBOX),
    RtStr::Rt(co::RT::DIALOG),
)?;

let hResLoad = hExe.LoadResource(&hRes)?;
let lpResLock = hExe.LockResource(&hRes, &hResLoad)?;
let hUpdateRes = HUPDATERSRC::BeginUpdateResource("foot.exe", false)?;

hUpdateRes.UpdateResource(
    RtStr::Rt(co::RT::DIALOG),
    IdStr::Id(IDD_FOOT_ABOUTBOX),
    LANGID::new(co::LANG::NEUTRAL, co::SUBLANG::NEUTRAL),
    lpResLock,
)?;

fn SizeofResource(&self, res_info: &HRSRC) -> SysResult<u32>

SizeofResource method.

For an example, see HINSTANCE::LockResource.