Trait winsafe::prelude::kernel_Hlocal

pub trait kernel_Hlocal: Handle {
    // Provided methods
    fn LocalAlloc(flags: LMEM, num_bytes: usize) -> SysResult<LocalFreeGuard> { ... }
    fn LocalFlags(&self) -> SysResult<LMEM> { ... }
    fn LocalReAlloc(&mut self, num_bytes: usize, flags: LMEM) -> SysResult<()> { ... }
    fn LocalSize(&self) -> SysResult<usize> { ... }
}

Available on crate feature kernel only.

This trait is enabled with the kernel feature, and provides methods for HLOCAL.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn LocalAlloc(flags: LMEM, num_bytes: usize) -> SysResult<LocalFreeGuard>

LocalAlloc static method.

fn LocalFlags(&self) -> SysResult<LMEM>

LocalFlags method.

fn LocalReAlloc(&mut self, num_bytes: usize, flags: LMEM) -> SysResult<()>

LocalReAlloc method.

Originally this method returns the handle to the reallocated memory object; here the original handle is automatically updated.

fn LocalSize(&self) -> SysResult<usize>

LocalSize method.