Trait winsafe::prelude::kernel_Hlocal

pub trait kernel_Hlocal: Handle {
    fn LocalFree(&self) -> SysResult<()> { ... }
    fn LocalSize(&self) -> SysResult<usize> { ... }
}

Available on crate feature kernel only.

This trait is enabled with the kernel feature, and provides methods for HLOCAL.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn LocalFree(&self) -> SysResult<()>

LocalFree method.

After calling this method, the handle will be invalidated and further operations will fail with ERROR::INVALID_HANDLE error code.

fn LocalSize(&self) -> SysResult<usize>

LocalSize method.