Trait winsafe::prelude::kernel_Hprocess

pub trait kernel_Hprocess: Handle {

    fn CreateProcess(
        application_name: Option<&str>,
        command_line: Option<&str>,
        process_attrs: Option<&mut SECURITY_ATTRIBUTES<'_>>,
        thread_attrs: Option<&mut SECURITY_ATTRIBUTES<'_>>,
        inherit_handles: bool,
        creation_flags: CREATE,
        environment: Option<Vec<String>>,
        current_dir: Option<&str>,
        si: &mut STARTUPINFO<'_, '_>
    ) -> SysResult<PROCESS_INFORMATION> { ... }
    fn ExitProcess(exit_code: u32) { ... }
    fn FlushInstructionCache(
        self,
        base_address: *mut c_void,
        size: usize
    ) -> SysResult<()> { ... }
    fn FlushProcessWriteBuffers() { ... }
    fn GetCurrentProcess() -> HPROCESS { ... }
    fn GetCurrentProcessId() -> u32 { ... }
    fn GetExitCodeProcess(self) -> SysResult<u32> { ... }
    fn GetGuiResources(self, flags: GR) -> SysResult<u32> { ... }
    fn GetProcessHandleCount(self) -> SysResult<u32> { ... }
    fn GetProcessId(self) -> SysResult<u32> { ... }
    fn GetProcessTimes(
        self,
        creation: &mut FILETIME,
        exit: &mut FILETIME,
        kernel: &mut FILETIME,
        user: &mut FILETIME
    ) -> SysResult<()> { ... }
    fn IsDebuggerPresent() -> bool { ... }
    fn IsProcessCritical(self) -> SysResult<bool> { ... }
    fn IsWow64Process(self) -> SysResult<bool> { ... }
    fn OpenProcess(
        desired_access: PROCESS,
        inherit_handle: bool,
        process_id: u32
    ) -> SysResult<HPROCESS> { ... }
    fn OpenProcessToken(self, desired_access: TOKEN) -> SysResult<HACCESSTOKEN> { ... }
    fn QueryFullProcessImageName(self, flags: PROCESS_NAME) -> SysResult<String> { ... }
    fn WaitForSingleObject(self, milliseconds: Option<u32>) -> SysResult<WAIT> { ... }

}

Available on crate feature kernel only.

This trait is enabled with the kernel feature, and provides methods for HPROCESS.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn CreateProcess(
application_name: Option<&str>,
command_line: Option<&str>,
process_attrs: Option<&mut SECURITY_ATTRIBUTES<'_>>,
thread_attrs: Option<&mut SECURITY_ATTRIBUTES<'_>>,
inherit_handles: bool,
creation_flags: CREATE,
environment: Option<Vec<String>>,
current_dir: Option<&str>,
si: &mut STARTUPINFO<'_, '_>
) -> SysResult<PROCESS_INFORMATION>

CreateProcess static method.

Note: Process and thread handles are returned in the PROCESS_INFORMATION struct, and they must be paired with their respective HPROCESS::CloseHandle and HTHREAD::CloseHandle calls.

fn ExitProcess(exit_code: u32)

ExitProcess static method.

fn FlushInstructionCache(
self,
base_address: *mut c_void,
size: usize
) -> SysResult<()>

FlushInstructionCache method.

fn FlushProcessWriteBuffers()

FlushProcessWriteBuffers static method.

fn GetCurrentProcess() -> HPROCESS

GetCurrentProcess static method.

fn GetCurrentProcessId() -> u32

GetCurrentProcessId static method.

fn GetExitCodeProcess(self) -> SysResult<u32>

GetExitCodeProcess method.

fn GetGuiResources(self, flags: GR) -> SysResult<u32>

GetGuiResources method.

fn GetProcessHandleCount(self) -> SysResult<u32>

GetProcessHandleCount method.

fn GetProcessId(self) -> SysResult<u32>

GetProcessId method.

fn GetProcessTimes(
self,
creation: &mut FILETIME,
exit: &mut FILETIME,
kernel: &mut FILETIME,
user: &mut FILETIME
) -> SysResult<()>

GetProcessTimes method.

fn IsDebuggerPresent() -> bool

IsDebuggerPresent static method.

fn IsProcessCritical(self) -> SysResult<bool>

IsProcessCritical method.

fn IsWow64Process(self) -> SysResult<bool>

IsWow64Process method.

fn OpenProcess(
desired_access: PROCESS,
inherit_handle: bool,
process_id: u32
) -> SysResult<HPROCESS>

OpenProcess static method.

This method will return ERROR::INVALID_PARAMETER if you try to open a system process.

Note: Must be paired with an HPROCESS::CloseHandle call.

fn OpenProcessToken(self, desired_access: TOKEN) -> SysResult<HACCESSTOKEN>

OpenProcessToken method.

Note: Must be paired with an HACCESSTOKEN::CloseHandle call.

fn QueryFullProcessImageName(self, flags: PROCESS_NAME) -> SysResult<String>

QueryFullProcessImageName method.

fn WaitForSingleObject(self, milliseconds: Option<u32>) -> SysResult<WAIT>

WaitForSingleObject method.