winsafe::prelude
Trait kernel_Hprocess

pub trait kernel_Hprocess: Handle {

    // Provided methods
    fn CheckRemoteDebuggerPresent(&self) -> SysResult<bool> { ... }
    fn CreateProcess(
        application_name: Option<&str>,
        command_line: Option<&str>,
        process_attrs: Option<&mut SECURITY_ATTRIBUTES<'_>>,
        thread_attrs: Option<&mut SECURITY_ATTRIBUTES<'_>>,
        inherit_handles: bool,
        creation_flags: CREATE,
        environment: Option<Vec<(&str, &str)>>,
        current_dir: Option<&str>,
        si: &mut STARTUPINFO<'_, '_>,
    ) -> SysResult<CloseHandlePiGuard> { ... }
    fn FlushInstructionCache(
        &self,
        base_address: *mut c_void,
        size: usize,
    ) -> SysResult<()> { ... }
    fn GetCurrentProcess() -> HPROCESS { ... }
    fn GetExitCodeProcess(&self) -> SysResult<u32> { ... }
    fn GetGuiResources(&self, flags: GR) -> SysResult<u32> { ... }
    fn GetPriorityClass(&self) -> SysResult<PRIORITY_CLASS> { ... }
    fn GetProcessHandleCount(&self) -> SysResult<u32> { ... }
    fn GetProcessId(&self) -> SysResult<u32> { ... }
    fn GetProcessTimes(
        &self,
    ) -> SysResult<(FILETIME, FILETIME, FILETIME, FILETIME)> { ... }
    fn IsProcessCritical(&self) -> SysResult<bool> { ... }
    fn IsWow64Process(&self) -> SysResult<bool> { ... }
    fn OpenProcess(
        desired_access: PROCESS,
        inherit_handle: bool,
        process_id: u32,
    ) -> SysResult<CloseHandleGuard<HPROCESS>> { ... }
    fn QueryFullProcessImageName(
        &self,
        flags: PROCESS_NAME,
    ) -> SysResult<String> { ... }
    fn QueryProcessAffinityUpdateMode(&self) -> SysResult<PROCESS_AFFINITY> { ... }
    fn QueryProcessCycleTime(&self) -> SysResult<u64> { ... }
    fn SetPriorityClass(&self, prority_class: PRIORITY_CLASS) -> SysResult<()> { ... }
    fn SetProcessAffinityUpdateMode(
        &self,
        flags: PROCESS_AFFINITY,
    ) -> SysResult<()> { ... }
    fn SetProcessPriorityBoost(
        &self,
        disable_priority_boost: bool,
    ) -> SysResult<()> { ... }
    fn TerminateProcess(&self, exit_code: u32) -> SysResult<()> { ... }
    fn WaitForSingleObject(&self, milliseconds: Option<u32>) -> SysResult<WAIT> { ... }

}

Available on crate feature kernel only.

This trait is enabled with the kernel feature, and provides methods for HPROCESS.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn CheckRemoteDebuggerPresent(&self) -> SysResult<bool>

CheckRemoteDebuggerPresent function.

fn CreateProcess( application_name: Option<&str>, command_line: Option<&str>, process_attrs: Option<&mut SECURITY_ATTRIBUTES<'_>>, thread_attrs: Option<&mut SECURITY_ATTRIBUTES<'_>>, inherit_handles: bool, creation_flags: CREATE, environment: Option<Vec<(&str, &str)>>, current_dir: Option<&str>, si: &mut STARTUPINFO<'_, '_>, ) -> SysResult<CloseHandlePiGuard>

CreateProcess function.

fn FlushInstructionCache( &self, base_address: *mut c_void, size: usize, ) -> SysResult<()>

FlushInstructionCache function.

fn GetCurrentProcess() -> HPROCESS

GetCurrentProcess function.

fn GetExitCodeProcess(&self) -> SysResult<u32>

GetExitCodeProcess function.

fn GetGuiResources(&self, flags: GR) -> SysResult<u32>

GetGuiResources function.

fn GetPriorityClass(&self) -> SysResult<PRIORITY_CLASS>

GetPriorityClass function.

fn GetProcessHandleCount(&self) -> SysResult<u32>

GetProcessHandleCount function.

fn GetProcessId(&self) -> SysResult<u32>

GetProcessId function.

fn GetProcessTimes(&self) -> SysResult<(FILETIME, FILETIME, FILETIME, FILETIME)>

GetProcessTimes function.

Returns, respectively:

creation time;

exit time;

kernel time;

user time.

Examples

use winsafe::{self as w, prelude::*, co};

let hprocess: w::HPROCESS; // initialized somewhere

let (creation, exit, kernel, user) = hprocess.GetProcessTimes()?;

fn IsProcessCritical(&self) -> SysResult<bool>

IsProcessCritical function.

fn IsWow64Process(&self) -> SysResult<bool>

IsWow64Process function.

fn OpenProcess( desired_access: PROCESS, inherit_handle: bool, process_id: u32, ) -> SysResult<CloseHandleGuard<HPROCESS>>

OpenProcess function.

This method will return ERROR::INVALID_PARAMETER if you try to open a system process.

fn QueryFullProcessImageName(&self, flags: PROCESS_NAME) -> SysResult<String>

QueryFullProcessImageName function.

fn QueryProcessAffinityUpdateMode(&self) -> SysResult<PROCESS_AFFINITY>

QueryProcessAffinityUpdateMode function.

fn QueryProcessCycleTime(&self) -> SysResult<u64>

QueryProcessCycleTime function.

fn SetPriorityClass(&self, prority_class: PRIORITY_CLASS) -> SysResult<()>

SetPriorityClass function.

fn SetProcessAffinityUpdateMode(&self, flags: PROCESS_AFFINITY) -> SysResult<()>

SetProcessAffinityUpdateMode function.

fn SetProcessPriorityBoost(&self, disable_priority_boost: bool) -> SysResult<()>

SetProcessPriorityBoost function.

fn TerminateProcess(&self, exit_code: u32) -> SysResult<()>

TerminateProcess function.

fn WaitForSingleObject(&self, milliseconds: Option<u32>) -> SysResult<WAIT>

WaitForSingleObject function.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.