Trait winsafe::prelude::kernel_Hprocesslist

pub trait kernel_Hprocesslist: Handle {

    fn iter_heaps<'a>(
        &'a self
    ) -> Box<dyn Iterator<Item = SysResult<&'a HEAPLIST32>> + 'a> { ... }
    fn iter_modules<'a>(
        &'a self
    ) -> Box<dyn Iterator<Item = SysResult<&'a MODULEENTRY32>> + 'a> { ... }
    fn iter_processes<'a>(
        &'a self
    ) -> Box<dyn Iterator<Item = SysResult<&'a PROCESSENTRY32>> + 'a> { ... }
    fn iter_threads<'a>(
        &'a self
    ) -> Box<dyn Iterator<Item = SysResult<&'a THREADENTRY32>> + 'a> { ... }
    fn CreateToolhelp32Snapshot(
        flags: TH32CS,
        th32_process_id: Option<u32>
    ) -> SysResult<HPROCESSLIST> { ... }
    fn Heap32ListFirst(self, hl: &mut HEAPLIST32) -> SysResult<bool> { ... }
    fn Heap32ListNext(self, hl: &mut HEAPLIST32) -> SysResult<bool> { ... }
    fn Module32First(self, me: &mut MODULEENTRY32) -> SysResult<bool> { ... }
    fn Module32Next(self, me: &mut MODULEENTRY32) -> SysResult<bool> { ... }
    fn Process32First(self, pe: &mut PROCESSENTRY32) -> SysResult<bool> { ... }
    fn Process32Next(self, pe: &mut PROCESSENTRY32) -> SysResult<bool> { ... }
    fn Thread32First(self, te: &mut THREADENTRY32) -> SysResult<bool> { ... }
    fn Thread32Next(self, te: &mut THREADENTRY32) -> SysResult<bool> { ... }

}

Available on crate feature kernel only.

This trait is enabled with the kernel feature, and provides methods for HPROCESSLIST.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn iter_heaps<'a>(
&'a self
) -> Box<dyn Iterator<Item = SysResult<&'a HEAPLIST32>> + 'a>

Returns an iterator over the heaps of a process, with HEAPLIST32 structs. Calls HPROCESSLIST::Heap32ListFirst and then HPROCESSLIST::Heap32ListNext consecutively.

Examples

use winsafe::prelude::*;
use winsafe::{co, HPROCESSLIST};

let hpl = HPROCESSLIST::
    CreateToolhelp32Snapshot(co::TH32CS::SNAPHEAPLIST, None)?;

for heap_entry in hpl.iter_heaps() {
    let heap_entry = heap_entry?;
    let is_default_heap = heap_entry.dwFlags == co::HF32::DEFAULT;
    println!("{} {}",
        heap_entry.th32HeapID, heap_entry.th32ProcessID);
}

hpl.CloseHandle()?;

fn iter_modules<'a>(
&'a self
) -> Box<dyn Iterator<Item = SysResult<&'a MODULEENTRY32>> + 'a>

Returns an iterator over the modules of a process, with MODULEENTRY32 structs. Calls HPROCESSLIST::Module32First and then HPROCESSLIST::Module32Next consecutively.

Examples

use winsafe::prelude::*;
use winsafe::{co, HPROCESSLIST};

let hpl = HPROCESSLIST::
    CreateToolhelp32Snapshot(co::TH32CS::SNAPMODULE, None)?;

for mod_entry in hpl.iter_modules() {
    let mod_entry = mod_entry?;
    println!("{} {}",
        mod_entry.szModule(), mod_entry.th32ProcessID);
}

hpl.CloseHandle()?;

fn iter_processes<'a>(
&'a self
) -> Box<dyn Iterator<Item = SysResult<&'a PROCESSENTRY32>> + 'a>

Returns an iterator over the processes of a process, with PROCESSENTRY32 structs. Calls HPROCESSLIST::Process32First and then HPROCESSLIST::Process32Next consecutively.

Examples

use winsafe::prelude::*;
use winsafe::{co, HPROCESSLIST};

let hpl = HPROCESSLIST::
    CreateToolhelp32Snapshot(co::TH32CS::SNAPPROCESS, None)?;

for proc_entry in hpl.iter_processes() {
    let proc_entry = proc_entry?;
    println!("{} {} {}",
        proc_entry.szExeFile(), proc_entry.th32ProcessID, proc_entry.cntThreads);
}

hpl.CloseHandle()?;

fn iter_threads<'a>(
&'a self
) -> Box<dyn Iterator<Item = SysResult<&'a THREADENTRY32>> + 'a>

Returns an iterator over the threads of a process, with THREADENTRY32 structs. Calls HPROCESSLIST::Thread32First and then HPROCESSLIST::Thread32Next consecutively.

Examples

use winsafe::prelude::*;
use winsafe::{co, HPROCESS, HPROCESSLIST};

let hpl = HPROCESSLIST::CreateToolhelp32Snapshot(
    co::TH32CS::SNAPTHREAD,
    Some(HPROCESS::GetCurrentProcessId()),
)?;

for thread_entry in hpl.iter_threads() {
    let thread_entry = thread_entry?;
    println!("{} {}",
        thread_entry.th32ThreadID, thread_entry.th32OwnerProcessID);
}

hpl.CloseHandle()?;

fn CreateToolhelp32Snapshot(
flags: TH32CS,
th32_process_id: Option<u32>
) -> SysResult<HPROCESSLIST>

CreateToolhelp32Snapshot static method.

Note: Must be paired with an HPROCESSLIST::CloseHandle call.

fn Heap32ListFirst(self, hl: &mut HEAPLIST32) -> SysResult<bool>

HeapList32First method.

Prefer using HPROCESSLIST::iter_heaps, which is simpler.

fn Heap32ListNext(self, hl: &mut HEAPLIST32) -> SysResult<bool>

HeapList32Next method.

Prefer using HPROCESSLIST::iter_heaps, which is simpler.

fn Module32First(self, me: &mut MODULEENTRY32) -> SysResult<bool>

Module32First method.

Prefer using HPROCESSLIST::iter_modules, which is simpler.

fn Module32Next(self, me: &mut MODULEENTRY32) -> SysResult<bool>

Module32Next method.

Prefer using HPROCESSLIST::iter_modules, which is simpler.

fn Process32First(self, pe: &mut PROCESSENTRY32) -> SysResult<bool>

Process32First method.

Prefer using HPROCESSLIST::iter_processes, which is simpler.

fn Process32Next(self, pe: &mut PROCESSENTRY32) -> SysResult<bool>

Process32Next method.

Prefer using HPROCESSLIST::iter_processes, which is simpler.

fn Thread32First(self, te: &mut THREADENTRY32) -> SysResult<bool>

Thread32First method.

Prefer using HPROCESSLIST::iter_threads, which is simpler.

fn Thread32Next(self, te: &mut THREADENTRY32) -> SysResult<bool>

Thread32First method.

Prefer using HPROCESSLIST::iter_threads, which is simpler.