winsafe::prelude

Trait mf_IMFClock

pub trait mf_IMFClock: ole_IUnknown {
    // Provided methods
    fn GetClockCharacteristics(&self) -> HrResult<MFCLOCK_CHARACTERISTICS_FLAG> { ... }
    fn GetContinuityKey(&self) -> HrResult<u32> { ... }
    fn GetCorrelatedTime(&self) -> HrResult<(i64, i64)> { ... }
    fn GetProperties(&self) -> HrResult<MFCLOCK_PROPERTIES> { ... }
    fn GetState(&self) -> HrResult<MFCLOCK_STATE> { ... }
}

Available on crate feature mf only.

This trait is enabled with the mf feature, and provides methods for IMFClock.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn GetClockCharacteristics(&self) -> HrResult<MFCLOCK_CHARACTERISTICS_FLAG>

IMFClock::GetClockCharacteristics method.

fn GetContinuityKey(&self) -> HrResult<u32>

IMFClock::GetContinuityKey method.

fn GetCorrelatedTime(&self) -> HrResult<(i64, i64)>

IMFClock::GetCorrelatedTime method.

Returns the last known clock time (in units of the clock’s frequency) and the system time that corresponds to the last known clock time (in 100-nanosecond units).

fn GetProperties(&self) -> HrResult<MFCLOCK_PROPERTIES>

IMFClock::GetProperties method.

fn GetState(&self) -> HrResult<MFCLOCK_STATE>

IMFClock::GetState method.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.