Trait winsafe::prelude::mf_IMFTopologyNode

pub trait mf_IMFTopologyNode: mf_IMFAttributes {
    // Provided methods
    fn CloneFrom(&self, node: &impl mf_IMFTopologyNode) -> HrResult<()> { ... }
    fn ConnectOutput(
        &self,
        output_index: u32,
        downstream_node: &impl mf_IMFTopologyNode,
        input_index_on_downstream_node: u32
    ) -> HrResult<()> { ... }
    fn DisconnectOutput(&self, output_index: u32) -> HrResult<()> { ... }
    fn GetInput(&self, input_index: u32) -> HrResult<(IMFTopologyNode, u32)> { ... }
    fn GetInputCount(&self) -> HrResult<u32> { ... }
    fn GetNodeType(&self) -> HrResult<MF_TOPOLOGY> { ... }
    fn GetObject<T>(&self) -> HrResult<T>
       where T: ole_IUnknown { ... }
    fn GetOutput(&self, output_index: u32) -> HrResult<(IMFTopologyNode, u32)> { ... }
    fn GetOutputCount(&self) -> HrResult<u32> { ... }
    fn GetTopoNodeID(&self) -> HrResult<u64> { ... }
    fn SetObject(&self, object: &impl ole_IUnknown) -> HrResult<()> { ... }
    fn SetTopoNodeID(&self, topo_id: u64) -> HrResult<()> { ... }
}

Available on crate features kernel and mf only.

This trait is enabled with the mf feature, and provides methods for IMFTopologyNode.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn CloneFrom(&self, node: &impl mf_IMFTopologyNode) -> HrResult<()>

IMFTopologyNode::CloneFrom method.

fn ConnectOutput( &self, output_index: u32, downstream_node: &impl mf_IMFTopologyNode, input_index_on_downstream_node: u32 ) -> HrResult<()>

IMFTopologyNode::ConnectOutput method.

fn DisconnectOutput(&self, output_index: u32) -> HrResult<()>

IMFTopologyNode::DisconnectOutput method.

fn GetInput(&self, input_index: u32) -> HrResult<(IMFTopologyNode, u32)>

IMFTopologyNode::GetInput method.

Returns the node and the index of the output stream that is connected to this node’s input stream.

fn GetInputCount(&self) -> HrResult<u32>

IMFTopologyNode::GetInputCount method.

fn GetNodeType(&self) -> HrResult<MF_TOPOLOGY>

IMFTopologyNode::GetNodeType method.

fn GetObject<T>(&self) -> HrResult<T>
where T: ole_IUnknown,

IMFTopologyNode::GetObject method.

fn GetOutput(&self, output_index: u32) -> HrResult<(IMFTopologyNode, u32)>

IMFTopologyNode::GetOutput method.

Returns the node and the index of the input stream that is connected to this node’s output stream.

fn GetOutputCount(&self) -> HrResult<u32>

IMFTopologyNode::GetOutputCount method.

fn GetTopoNodeID(&self) -> HrResult<u64>

IMFTopologyNode::GetTopoNodeID method.

fn SetObject(&self, object: &impl ole_IUnknown) -> HrResult<()>

IMFTopologyNode::SetObject method

fn SetTopoNodeID(&self, topo_id: u64) -> HrResult<()>

IMFTopologyNode::SetTopoNodeID method.

Object Safety

This trait is not object safe.