winsafe::prelude

Trait ole_IDataObject

pub trait ole_IDataObject: ole_IUnknown {
    // Provided methods
    fn DAdvise(
        &self,
        formatetc: &FORMATETC<'_>,
        advf: ADVF,
        adv_sink: &impl ole_IAdviseSink,
    ) -> HrResult<u32> { ... }
    fn DUnadvise(&self, connection: u32) -> HrResult<()> { ... }
    unsafe fn GetData(
        &self,
        formatetc: &FORMATETC<'_>,
    ) -> HrResult<ReleaseStgMediumGuard> { ... }
    fn QueryGetData(&self, formatetc: &FORMATETC<'_>) -> HrResult<()> { ... }
}

Available on crate feature ole only.

This trait is enabled with the ole feature, and provides methods for IDataObject.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn DAdvise( &self, formatetc: &FORMATETC<'_>, advf: ADVF, adv_sink: &impl ole_IAdviseSink, ) -> HrResult<u32>

IDataObject::DAdvise method.

fn DUnadvise(&self, connection: u32) -> HrResult<()>

IDataObject::DUnadvise method.

unsafe fn GetData( &self, formatetc: &FORMATETC<'_>, ) -> HrResult<ReleaseStgMediumGuard>

IDataObject::GetData method.

Safety

The returned struct may contain pointers that need to be deallocated.

fn QueryGetData(&self, formatetc: &FORMATETC<'_>) -> HrResult<()>

IDataObject::QueryGetData method.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.