Trait winsafe::prelude::ole_IPicture

pub trait ole_IPicture: ole_IUnknown {
    fn get_CurDC(&self) -> HrResult<HDC> { ... }
    fn get_Height(&self) -> HrResult<i32> { ... }
    fn get_Type(&self) -> HrResult<PICTYPE> { ... }
    fn get_Width(&self) -> HrResult<i32> { ... }
    fn PictureChanged(&self) -> HrResult<()> { ... }
    fn put_KeepOriginalFormat(&self, keep: bool) -> HrResult<()> { ... }
    fn Render(
        &self,
        hdc: HDC,
        dest_pt: POINT,
        dest_sz: SIZE,
        src_offset: Option<POINT>,
        src_extent: SIZE,
        metafile_bounds: Option<&RECT>
    ) -> HrResult<()> { ... }
    fn SelectPicture(&self, hdc: HDC) -> HrResult<(HDC, HBITMAP)> { ... }
}

Available on crate feature ole only.

This trait is enabled with the ole feature, and provides methods for IPicture.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn get_CurDC(&self) -> HrResult<HDC>

IPicture::get_CurDC method.

fn get_Height(&self) -> HrResult<i32>

IPicture::get_Height method.

Note: Returns a value in HIMETRIC units. To convert it to pixels, prefer using the simpler IPicture::size_px, or use HDC::HiMetricToPixel to perform the conversion manually.

Examples

use winsafe::prelude::*;
use winsafe::{HWND, IPicture};

let pic: IPicture; // initialized somewhere

let hdc = HWND::NULL.GetDC()?;

let (_, height) = hdc.HiMetricToPixel(0, pic.get_Height()?);
println!("Height: {} px", height);

HWND::NULL.ReleaseDC(hdc)?;

fn get_Type(&self) -> HrResult<PICTYPE>

IPicture::get_Type method.

fn get_Width(&self) -> HrResult<i32>

IPicture::get_Width method.

Note: Returns a value in HIMETRIC units. To convert it to pixels, prefer using the simpler IPicture::size_px, or use HDC::HiMetricToPixel to perform the conversion manually.

Examples

use winsafe::prelude::*;
use winsafe::{HWND, IPicture};

let pic: IPicture; // initialized somewhere

let hdc = HWND::NULL.GetDC()?;

let (width, _) = hdc.HiMetricToPixel(pic.get_Width()?, 0);
println!("Width: {} px", width);

HWND::NULL.ReleaseDC(hdc)?;

fn PictureChanged(&self) -> HrResult<()>

IPicture::PictureChanged method.

fn put_KeepOriginalFormat(&self, keep: bool) -> HrResult<()>

IPicture::put_KeepOriginalFormat method.

fn Render(
&self,
hdc: HDC,
dest_pt: POINT,
dest_sz: SIZE,
src_offset: Option<POINT>,
src_extent: SIZE,
metafile_bounds: Option<&RECT>
) -> HrResult<()>

IPicture::Render method.

fn SelectPicture(&self, hdc: HDC) -> HrResult<(HDC, HBITMAP)>

IPicture::SelectPicture method.