winsafe::prelude

Trait ole_ISequentialStream

pub trait ole_ISequentialStream: ole_IUnknown {
    // Provided methods
    fn Read(&self, buffer: &mut [u8]) -> HrResult<u32> { ... }
    fn Write(&self, data: &[u8]) -> HrResult<u32> { ... }
}

Available on crate feature ole only.

This trait is enabled with the ole feature, and provides methods for ISequentialStream.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn Read(&self, buffer: &mut [u8]) -> HrResult<u32>

ISequentialStream::Read method.

Returns the number of bytes written; if this value is lower than the requested size, it means the end of stream was reached.

fn Write(&self, data: &[u8]) -> HrResult<u32>

ISequentialStream::Write method.

Returns the number of bytes written.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.