Trait winsafe::prelude::ole_IUnknown

pub trait ole_IUnknown: Clone + From<ComPtr> {
    const IID: IID;

    unsafe fn leak(&mut self) -> ComPtr;
    unsafe fn ptr(&self) -> ComPtr;
    unsafe fn vt_ref<T>(&self) -> &T;

    fn QueryInterface<T>(&self) -> HrResult<T>
    where
        T: ole_IUnknown,
    { ... }
}

Available on crate feature ole only.

This trait is enabled with the ole feature, and provides methods for IUnknown. It is the base trait for all COM traits.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Note that the IUnknownVT virtual table has two other methods: AddRef and Release. While these methods are relevant in C++, here they are abstracted away as it follows:

AddRef – called along the clone method from the Clone trait;

Release – called automatically by the Drop trait, so you don’t need to worry about it.

Required Associated Constants

const IID: IID

The COM interface ID.

Required Methods

unsafe fn leak(&mut self) -> ComPtr

Returns the pointer to the underlying COM virtual table and sets the internal pointer to null, so that Release won’t be called.

Note: Be sure to release the pointer, otherwise, as the name of this method implies, you will cause a resource leak.

unsafe fn ptr(&self) -> ComPtr

Returns the pointer to the underlying COM virtual table.

unsafe fn vt_ref<T>(&self) -> &T

Returns a reference to the underlying COM virtual table.

Note: Be sure the pointer has the correct virtual table type (or inherits from it), and it is not null.

Provided Methods

fn QueryInterface<T>(&self) -> HrResult<T>where
T: ole_IUnknown,

IUnknown::QueryInterface method.