Trait winsafe::prelude::ole_IUnknown

pub trait ole_IUnknown: Clone + From<ComPtr> {
    const IID: IID;

    // Required methods
    fn leak(&mut self) -> ComPtr;
    unsafe fn ptr(&self) -> ComPtr;
    unsafe fn vt_ref<T>(&self) -> &T;

    // Provided method
    fn QueryInterface<T>(&self) -> HrResult<T>
       where T: ole_IUnknown { ... }
}

Available on crate feature ole only.

This trait is enabled with the ole feature, and provides methods for IUnknown. It is the base trait for all COM traits.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Note that the IUnknownVT virtual table has two other methods: AddRef and Release. While these methods are relevant in C++, here they are abstracted away as it follows:

AddRef – called along the clone method from the Clone trait;

Release – called automatically by the Drop trait, so you don’t need to worry about it.

Required Associated Constants

const IID: IID

The COM interface ID.

Required Methods

fn leak(&mut self) -> ComPtr

Returns the pointer to the underlying COM virtual table and sets the internal pointer to null, so that Release won’t be called.

Be sure to release the pointer, otherwise, as the name of this method implies, you will cause a resource leak.

This method is used internally by the library, and not intended to be used externally.

unsafe fn ptr(&self) -> ComPtr

Returns the pointer to the underlying COM virtual table.

Safety

The returned pointer is allocated and alive in the source COM interface, and should not be released. Do not cast this pointer into anything else.

This method is used internally by the library, and not intended to be used externally.

unsafe fn vt_ref<T>(&self) -> &T

Returns a reference to the underlying COM virtual table.

Safety

Be sure the pointer has the correct virtual table type (or inherits from it), and is not null.

This method is used internally by the library, and not intended to be used externally.

Provided Methods

fn QueryInterface<T>(&self) -> HrResult<T>where T: ole_IUnknown,

IUnknown::QueryInterface method.