winsafe::prelude

Trait ole_IUnknown

pub trait ole_IUnknown: Clone {
    const IID: IID;

    // Required methods
    unsafe fn from_ptr(p: *mut c_void) -> Self;
    fn ptr(&self) -> *mut c_void;
    unsafe fn as_mut(&mut self) -> &mut *mut c_void;

    // Provided methods
    unsafe fn null() -> Self { ... }
    fn leak(&mut self) -> *mut c_void { ... }
    fn QueryInterface<T>(&self) -> HrResult<T>
       where T: ole_IUnknown { ... }
}

Available on crate feature ole only.

This trait is enabled with the ole feature, and provides methods for IUnknown. It is the base trait for all COM traits.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Note that the IUnknown virtual table has two other methods: AddRef and Release. While these methods are relevant in C++, here they are abstracted away as it follows:

AddRef – called along the clone method from the Clone trait;

Release – called automatically by the Drop trait, so you don’t need to worry about it.

Required Associated Constants

const IID: IID

The COM interface ID.

Required Methods

unsafe fn from_ptr(p: *mut c_void) -> Self

Creates an object from a COM virtual table pointer.

This method can be used as an escape hatch to interoperate with other libraries.

Panics

Panics if trying to create a custom COM implementation object.

Safety

Be sure the pointer points to a properly allocated COM virtual table.

fn ptr(&self) -> *mut c_void

Returns the pointer to the underlying COM virtual table.

This method can be used as an escape hatch to interoperate with other libraries.

unsafe fn as_mut(&mut self) -> &mut *mut c_void

Returns a mutable reference do the underlying COM virtual table pointer.

This method can be used as an escape hatch to interoperate with other libraries.

Panics

Panics if trying to modify a custom COM implementation object.

Safety

Be sure the pointer being set points to a properly allocated COM virtual table.

Provided Methods

unsafe fn null() -> Self

Creates an object from a null COM virtual table pointer.

Safety

The pointer must be initialized before any call. Do not call methods on a null COM pointer.

fn leak(&mut self) -> *mut c_void

Returns the pointer to the underlying COM virtual table and sets the internal pointer to null, so that Release won’t be called.

Be sure to release the pointer, otherwise, as the name of this method implies, you will cause a resource leak.

fn QueryInterface<T>(&self) -> HrResult<T>
where T: ole_IUnknown,

IUnknown::QueryInterface method.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.