Trait winsafe::prelude::shell_Hdrop

pub trait shell_Hdrop: Handle {
    fn iter<'a>(
        &'a self
    ) -> SysResult<Box<dyn Iterator<Item = SysResult<String>> + 'a>> { ... }
    fn DragFinish(self) { ... }
    fn DragQueryFile(
        self,
        ifile: Option<u32>,
        buf: Option<&mut WString>
    ) -> SysResult<u32> { ... }
    fn DragQueryPoint(self) -> (POINT, bool) { ... }
}

Available on crate feature shell only.

This trait is enabled with the shell feature, and provides methods for HDROP.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn iter<'a>(
&'a self
) -> SysResult<Box<dyn Iterator<Item = SysResult<String>> + 'a>>

Returns an iterator over the dropped files by calling HDROP::DragQueryFile consecutively, then frees the handle by calling HDROP::DragFinish.

Examples

Iterating over the strings:

use winsafe::prelude::*;
use winsafe::HDROP;

let hdrop: HDROP; // initialized somewhere

for file_path in hdrop.iter()? {
    let file_path = file_path?;
    println!("File: {}", file_path);
}

Collecting the strings into a Vec:

use winsafe::prelude::*;
use winsafe::{HDROP, SysResult};

let hdrop: HDROP; // initialized somewhere

let file_paths = hdrop.iter()?
    .collect::<SysResult<Vec<_>>>()?;

fn DragFinish(self)

DragFinish method.

Prefer using HDROP::iter, which calls DragFinish automatically.

fn DragQueryFile(
self,
ifile: Option<u32>,
buf: Option<&mut WString>
) -> SysResult<u32>

DragQueryFile method.

This method is rather tricky, consider using HDROP::iter.

fn DragQueryPoint(self) -> (POINT, bool)

DragQueryPoint method.

Returns the coordinates and whether the drop occurred in the client area of the window.