Trait winsafe::prelude::shell_IFileDialog

pub trait shell_IFileDialog: shell_IModalWindow {

    // Provided methods
    fn AddPlace(&self, si: &impl shell_IShellItem, fdap: FDAP) -> HrResult<()> { ... }
    fn Advise(&self, fde: &IFileDialogEvents) -> HrResult<u32> { ... }
    fn ClearClientData(&self) -> HrResult<()> { ... }
    fn Close(&self, hr: ERROR) -> HrResult<()> { ... }
    fn GetCurrentSelection(&self) -> HrResult<IShellItem> { ... }
    fn GetFileName(&self) -> HrResult<String> { ... }
    fn GetFileTypeIndex(&self) -> HrResult<u32> { ... }
    fn GetFolder(&self) -> HrResult<IShellItem> { ... }
    fn GetOptions(&self) -> HrResult<FOS> { ... }
    fn GetResult(&self) -> HrResult<IShellItem> { ... }
    fn SetClientGuid(&self, guid: &GUID) -> HrResult<()> { ... }
    fn SetDefaultExtension(&self, default_extension: &str) -> HrResult<()> { ... }
    fn SetDefaultFolder(&self, si: &impl shell_IShellItem) -> HrResult<()> { ... }
    fn SetFileName(&self, name: &str) -> HrResult<()> { ... }
    fn SetFileNameLabel(&self, label: &str) -> HrResult<()> { ... }
    fn SetFileTypeIndex(&self, index: u32) -> HrResult<()> { ... }
    fn SetFileTypes<S: AsRef<str>>(
        &self,
        filter_spec: &[(S, S)]
    ) -> HrResult<()> { ... }
    fn SetFolder(&self, si: &impl shell_IShellItem) -> HrResult<()> { ... }
    fn SetOkButtonLabel(&self, text: &str) -> HrResult<()> { ... }
    fn SetOptions(&self, opts: FOS) -> HrResult<()> { ... }
    fn SetTitle(&self, text: &str) -> HrResult<()> { ... }
    fn Unadvise(&self, cookie: u32) -> HrResult<()> { ... }

}

Available on crate features kernel and shell only.

This trait is enabled with the shell feature, and provides methods for IFileDialog.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn AddPlace(&self, si: &impl shell_IShellItem, fdap: FDAP) -> HrResult<()>

IFileDialog::AddPlace method.

fn Advise(&self, fde: &IFileDialogEvents) -> HrResult<u32>

IFileDialog::Advise method.

fn ClearClientData(&self) -> HrResult<()>

IFileDialog::ClearClientData method.

fn Close(&self, hr: ERROR) -> HrResult<()>

IFileDialog::Close method.

fn GetCurrentSelection(&self) -> HrResult<IShellItem>

IFileDialog::GetCurrentSelection method.

fn GetFileName(&self) -> HrResult<String>

IFileDialog::GetFileName method.

fn GetFileTypeIndex(&self) -> HrResult<u32>

IFileDialog::GetFileTypeIndex method.

fn GetFolder(&self) -> HrResult<IShellItem>

IFileDialog::GetFolder method.

fn GetOptions(&self) -> HrResult<FOS>

IFileDialog::GetOptions method.

fn GetResult(&self) -> HrResult<IShellItem>

IFileDialog::GetResult method.

If you chose a single file, this is the method to retrieve its path.

fn SetClientGuid(&self, guid: &GUID) -> HrResult<()>

IFileDialog::SetClientGuid method.

fn SetDefaultExtension(&self, default_extension: &str) -> HrResult<()>

IFileDialog::SetDefaultExtension method.

fn SetDefaultFolder(&self, si: &impl shell_IShellItem) -> HrResult<()>

IFileDialog::SetDefaultFolder method.

fn SetFileName(&self, name: &str) -> HrResult<()>

IFileDialog::SetFileName method.

fn SetFileNameLabel(&self, label: &str) -> HrResult<()>

IFileDialog::SetFileNameLabel method.

fn SetFileTypeIndex(&self, index: u32) -> HrResult<()>

IFileDialog::SetFileTypeIndex method.

Note: The index is one-based.

fn SetFileTypes<S: AsRef<str>>(&self, filter_spec: &[(S, S)]) -> HrResult<()>

IFileDialog::SetFileTypes method.

Examples

use winsafe::{self as w, prelude::*};

let file_dlg: w::IFileDialog; // initialized somewhere

file_dlg.SetFileTypes(&[
    ("Documents", "*.docx;*.txt"),
    ("Images", "*.jpg;*.png;*.bmp"),
    ("All files", "*.*"),
])?;

fn SetFolder(&self, si: &impl shell_IShellItem) -> HrResult<()>

IFileDialog::SetFolder method.

fn SetOkButtonLabel(&self, text: &str) -> HrResult<()>

IFileDialog::SetOkButtonLabel method.

fn SetOptions(&self, opts: FOS) -> HrResult<()>

IFileDialog::SetOptions method.

fn SetTitle(&self, text: &str) -> HrResult<()>

IFileDialog::SetTitle method.

fn Unadvise(&self, cookie: u32) -> HrResult<()>

IFileDialog::Unadvise method.

Object Safety

This trait is not object safe.