Trait winsafe::prelude::shell_IFileOpenDialog

pub trait shell_IFileOpenDialog: shell_IFileDialog {
    fn GetResults(&self) -> HrResult<IShellItemArray> { ... }
    fn GetSelectedItems(&self) -> HrResult<IShellItemArray> { ... }
}

Available on crate feature shell only.

This trait is enabled with the shell feature, and provides methods for IFileOpenDialog.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn GetResults(&self) -> HrResult<IShellItemArray>

IFileOpenDialog::GetResults method.

If you chose multiple files, this is the method to retrieve the paths.

Examples

Collecting the file paths into a Vec:

use winsafe::prelude::*;
use winsafe::{co, HrResult, IFileOpenDialog};

let fo: IFileOpenDialog; // initialized somewhere

let paths = fo.GetResults()?.iter()?
    .map(|shi|
        shi.and_then(|shi|
            shi.GetDisplayName(co::SIGDN::FILESYSPATH)
        ),
    )
    .collect::<HrResult<Vec<_>>>()?;

fn GetSelectedItems(&self) -> HrResult<IShellItemArray>

IFileOpenDialog::GetSelectedItems method.