Trait winsafe::prelude::shell_IShellLink

pub trait shell_IShellLink: ole_IUnknown {

    fn GetArguments(&self) -> HrResult<String> { ... }
    fn GetDescription(&self) -> HrResult<String> { ... }
    fn GetIconLocation(&self) -> HrResult<(String, i32)> { ... }
    fn GetPath(
        &self,
        fd: Option<&mut WIN32_FIND_DATA>,
        flags: SLGP
    ) -> HrResult<String> { ... }
    fn GetShowCmd(&self) -> HrResult<SW> { ... }
    fn GetWorkingDirectory(&self) -> HrResult<String> { ... }
    fn Resolve(&self, hwnd: &HWND, flags: SLR) -> HrResult<()> { ... }
    fn SetArguments(&self, args: &str) -> HrResult<()> { ... }
    fn SetDescription(&self, args: &str) -> HrResult<()> { ... }
    fn SetIconLocation(&self, path: &str, index: i32) -> HrResult<()> { ... }
    fn SetPath(&self, file: &str) -> HrResult<()> { ... }
    fn SetRelativePath(&self, file: &str) -> HrResult<()> { ... }
    fn SetShowCmd(&self, show_cmd: SW) -> HrResult<()> { ... }
    fn SetWorkingDirectory(&self, dir: &str) -> HrResult<()> { ... }

}

Available on crate feature shell only.

This trait is enabled with the shell feature, and provides methods for IShellLink.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

fn GetArguments(&self) -> HrResult<String>

IShellLinkW::GetArguments method.

fn GetDescription(&self) -> HrResult<String>

IShellLinkW::GetDescription method.

fn GetIconLocation(&self) -> HrResult<(String, i32)>

IShellLinkW::GetIconLocation method.

Returns the path of the icon and its index within the file.

fn GetPath(
&self,
fd: Option<&mut WIN32_FIND_DATA>,
flags: SLGP
) -> HrResult<String>

IShellLinkW::GetPath method.

fn GetShowCmd(&self) -> HrResult<SW>

IShellLinkW::GetShowCmd method.

fn GetWorkingDirectory(&self) -> HrResult<String>

IShellLinkW::GetWorkingDirectory method.

fn Resolve(&self, hwnd: &HWND, flags: SLR) -> HrResult<()>

IShellLinkW::Resolve method.

fn SetArguments(&self, args: &str) -> HrResult<()>

IShellLinkW::SetArguments method.

fn SetDescription(&self, args: &str) -> HrResult<()>

IShellLinkW::SetDescription method.

fn SetIconLocation(&self, path: &str, index: i32) -> HrResult<()>

IShellLinkW::SetIconLocation method.

fn SetPath(&self, file: &str) -> HrResult<()>

IShellLinkW::SetPath method.

fn SetRelativePath(&self, file: &str) -> HrResult<()>

IShellLinkW::SetRelativePath method.

fn SetShowCmd(&self, show_cmd: SW) -> HrResult<()>

IShellLinkW::SetShowCmd method.

fn SetWorkingDirectory(&self, dir: &str) -> HrResult<()>

IShellLinkW::SetWorkingDirectory method.