Trait winsafe::prelude::user_Hinstance

pub trait user_Hinstance: Handle {
    unsafe fn CreateDialogParam(
        &self,
        resource_id: IdStr,
        hwnd_parent: Option<&HWND>,
        dialog_proc: DLGPROC,
        init_param: Option<isize>
    ) -> SysResult<HWND> { ... }
    unsafe fn DialogBoxParam(
        &self,
        resource_id: IdStr,
        hwnd_parent: Option<&HWND>,
        dialog_proc: DLGPROC,
        init_param: Option<isize>
    ) -> SysResult<isize> { ... }
    fn GetClassInfoEx(
        &self,
        class_name: &str,
        wcx: &mut WNDCLASSEX<'_>
    ) -> SysResult<ATOM> { ... }
    fn LoadAccelerators(&self, table_name: IdStr) -> SysResult<HACCEL> { ... }
    fn LoadCursor(&self, resource_id: IdIdcStr) -> SysResult<HCURSOR> { ... }
    fn LoadIcon(&self, icon_id: IdIdiStr) -> SysResult<HICON> { ... }
    fn LoadImageBitmap(
        &self,
        name: IdObmStr,
        sz: SIZE,
        load: LR
    ) -> SysResult<HBITMAP> { ... }
    fn LoadImageCursor(
        &self,
        name: IdOcrStr,
        sz: SIZE,
        load: LR
    ) -> SysResult<HCURSOR> { ... }
    fn LoadImageIcon(&self, name: IdOicStr, sz: SIZE, load: LR) -> SysResult<HICON> { ... }
    fn LoadMenu(&self, resource_id: IdStr) -> SysResult<HMENU> { ... }
    fn LoadString(&self, id: u16) -> SysResult<String> { ... }
}

Available on crate feature user only.

This trait is enabled with the user feature, and provides methods for HINSTANCE.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

unsafe fn CreateDialogParam(
&self,
resource_id: IdStr,
hwnd_parent: Option<&HWND>,
dialog_proc: DLGPROC,
init_param: Option<isize>
) -> SysResult<HWND>

CreateDialogParam method.

Safety

To create a dialog, you must provide a dialog procedure.

unsafe fn DialogBoxParam(
&self,
resource_id: IdStr,
hwnd_parent: Option<&HWND>,
dialog_proc: DLGPROC,
init_param: Option<isize>
) -> SysResult<isize>

DialogBoxParam method.

Safety

To create a dialog, you must provide a dialog procedure.

fn GetClassInfoEx(
&self,
class_name: &str,
wcx: &mut WNDCLASSEX<'_>
) -> SysResult<ATOM>

GetClassInfoEx method.

Examples

Retrieving information of a window class created in our application:

use winsafe::prelude::*;
use winsafe::{HINSTANCE, WNDCLASSEX};

let mut wcx = WNDCLASSEX::default();
HINSTANCE::GetModuleHandle(None)?
    .GetClassInfoEx("SOME_CLASS_NAME", &mut wcx)?;

fn LoadAccelerators(&self, table_name: IdStr) -> SysResult<HACCEL>

LoadAccelerators method.

fn LoadCursor(&self, resource_id: IdIdcStr) -> SysResult<HCURSOR>

LoadCursor method.

Examples

Loading a system cursor:

use winsafe::prelude::*;
use winsafe::{co, HINSTANCE, IdIdcStr};

let sys_cursor = HINSTANCE::NULL
    .LoadCursor(IdIdcStr::Idc(co::IDC::ARROW))?;

fn LoadIcon(&self, icon_id: IdIdiStr) -> SysResult<HICON>

LoadIcon method.

Note: Must be paired with an HICON::DestroyIcon call.

Examples

Loading a system icon:

use winsafe::prelude::*;
use winsafe::{co, IdIdiStr, HINSTANCE};

let sys_icon = HINSTANCE::NULL
    .LoadIcon(IdIdiStr::Idi(co::IDI::INFORMATION))?;

fn LoadImageBitmap(&self, name: IdObmStr, sz: SIZE, load: LR) -> SysResult<HBITMAP>

LoadImage method for HBITMAP.

Note: Must be paired with an HBITMAP::DeleteObject call.

fn LoadImageCursor(&self, name: IdOcrStr, sz: SIZE, load: LR) -> SysResult<HCURSOR>

LoadImage method for HCURSOR.

Note: Must be paired with an HCURSOR::DestroyCursor call.

fn LoadImageIcon(&self, name: IdOicStr, sz: SIZE, load: LR) -> SysResult<HICON>

LoadImage method for HICON.

Note: Must be paired with an HICON::DestroyIcon call.

fn LoadMenu(&self, resource_id: IdStr) -> SysResult<HMENU>

LoadMenu method.

fn LoadString(&self, id: u16) -> SysResult<String>

LoadString method.