Trait winsafe::prelude::user_Hinstance

pub trait user_Hinstance: kernel_Hinstance {
    // Provided methods
    unsafe fn CreateDialogParam(
        &self,
        resource_id: IdStr,
        hwnd_parent: Option<&HWND>,
        dialog_proc: DLGPROC,
        init_param: Option<isize>
    ) -> SysResult<HWND> { ... }
    unsafe fn DialogBoxIndirectParam(
        &self,
        dialog_template: &DLGTEMPLATE,
        hwnd_parent: Option<&HWND>,
        dialog_proc: DLGPROC,
        init_param: Option<isize>
    ) -> SysResult<isize> { ... }
    unsafe fn DialogBoxParam(
        &self,
        resource_id: IdStr,
        hwnd_parent: Option<&HWND>,
        dialog_proc: DLGPROC,
        init_param: Option<isize>
    ) -> SysResult<isize> { ... }
    fn GetClassInfoEx(
        &self,
        class_name: &str,
        wcx: &mut WNDCLASSEX<'_>
    ) -> SysResult<ATOM> { ... }
    fn LoadAccelerators(&self, table_name: IdStr) -> SysResult<HACCEL> { ... }
    fn LoadCursor(&self, resource_id: IdIdcStr) -> SysResult<DestroyCursorGuard> { ... }
    fn LoadIcon(&self, icon_id: IdIdiStr) -> SysResult<DestroyIconGuard> { ... }
    fn LoadMenu(&self, resource_id: IdStr) -> SysResult<HMENU> { ... }
    fn LoadString(&self, id: u16) -> SysResult<String> { ... }
}

Available on crate features kernel and user only.

This trait is enabled with the user feature, and provides methods for HINSTANCE.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Methods

unsafe fn CreateDialogParam( &self, resource_id: IdStr, hwnd_parent: Option<&HWND>, dialog_proc: DLGPROC, init_param: Option<isize> ) -> SysResult<HWND>

CreateDialogParam function.

Safety

To create a dialog, you must provide a dialog procedure.

unsafe fn DialogBoxIndirectParam( &self, dialog_template: &DLGTEMPLATE, hwnd_parent: Option<&HWND>, dialog_proc: DLGPROC, init_param: Option<isize> ) -> SysResult<isize>

DialogBoxIndirectParam function.

Safety

To create a dialog, you must provide a dialog procedure.

unsafe fn DialogBoxParam( &self, resource_id: IdStr, hwnd_parent: Option<&HWND>, dialog_proc: DLGPROC, init_param: Option<isize> ) -> SysResult<isize>

DialogBoxParam function.

Safety

To create a dialog, you must provide a dialog procedure.

fn GetClassInfoEx( &self, class_name: &str, wcx: &mut WNDCLASSEX<'_> ) -> SysResult<ATOM>

GetClassInfoEx function.

Examples

Retrieving information of a window class created in our application:

use winsafe::{self as w, prelude::*};

let mut wcx = w::WNDCLASSEX::default();
w::HINSTANCE::GetModuleHandle(None)?
    .GetClassInfoEx("SOME_CLASS_NAME", &mut wcx)?;

fn LoadAccelerators(&self, table_name: IdStr) -> SysResult<HACCEL>

LoadAccelerators function.

fn LoadCursor(&self, resource_id: IdIdcStr) -> SysResult<DestroyCursorGuard>

LoadCursor function.

Examples

Loading a system cursor:

use winsafe::{self as w, prelude::*, co};

let sys_cursor = w::HINSTANCE::NULL
    .LoadCursor(w::IdIdcStr::Idc(co::IDC::ARROW))?;

fn LoadIcon(&self, icon_id: IdIdiStr) -> SysResult<DestroyIconGuard>

LoadIcon function.

Examples

Loading a system icon:

use winsafe::{self as w, prelude::*, co};

let sys_icon = w::HINSTANCE::NULL
    .LoadIcon(w::IdIdiStr::Idi(co::IDI::INFORMATION))?;

fn LoadMenu(&self, resource_id: IdStr) -> SysResult<HMENU>

LoadMenu function.

fn LoadString(&self, id: u16) -> SysResult<String>

LoadString function.

Object Safety

This trait is not object safe.