Trait winsafe::prelude::user_Hwnd

pub trait user_Hwnd: Handle {
    const BROADCAST: HWND = _;
    const DESKTOP: HWND = _;

    fn hinstance(&self) -> HINSTANCE { ... }
    fn ArrangeIconicWindows(&self) -> SysResult<u32> { ... }
    fn BeginPaint(&self) -> SysResult<HdcPaintGuard<'_, Self>> { ... }
    fn BringWindowToTop(&self) -> SysResult<()> { ... }
    fn ChildWindowFromPoint(&self, pt: POINT) -> Option<HWND> { ... }
    fn ClientToScreen(&self, pt: &mut POINT) -> SysResult<()> { ... }
    fn ClientToScreenRc(&self, rc: &mut RECT) -> SysResult<()> { ... }
    fn CloseWindow(&self) -> SysResult<()> { ... }
    unsafe fn CreateWindowEx(
        ex_style: WS_EX,
        class_name: AtomStr,
        title: Option<&str>,
        style: WS,
        pos: POINT,
        size: SIZE,
        hwnd_parent: Option<&HWND>,
        hmenu: IdMenu<'_>,
        hinstance: &HINSTANCE,
        lparam: Option<isize>
    ) -> SysResult<HWND> { ... }
    fn DefWindowProc<M>(&self, msg: M) -> M::RetType
    where
        M: MsgSend,
    { ... }
    fn DestroyWindow(&self) -> SysResult<()> { ... }
    fn DrawMenuBar(&self) -> SysResult<()> { ... }
    fn EnableWindow(&self, enable: bool) -> bool { ... }
    fn EndDialog(&self, result: isize) -> SysResult<()> { ... }
    fn EnumChildWindows<F>(&self, func: F)
    where
        F: Fn(HWND) -> bool,
    { ... }
    fn FindWindow(
        class_name: Option<AtomStr>,
        title: Option<&str>
    ) -> SysResult<HWND> { ... }
    fn FindWindowEx(
        &self,
        hwnd_child_after: Option<&HWND>,
        class_name: AtomStr,
        title: Option<&str>
    ) -> SysResult<HWND> { ... }
    fn GetActiveWindow() -> Option<HWND> { ... }
    fn GetAltTabInfo(
        &self,
        item: Option<u32>,
        ati: &mut ALTTABINFO,
        sz_item_text: Option<u32>
    ) -> SysResult<String> { ... }
    fn GetAncestor(&self, flags: GA) -> Option<HWND> { ... }
    fn GetCapture() -> Option<HWND> { ... }
    fn GetClassLongPtr(&self, index: GCLP) -> usize { ... }
    fn GetClassName(&self) -> SysResult<String> { ... }
    fn GetClientRect(&self) -> SysResult<RECT> { ... }
    fn GetDC(&self) -> SysResult<HdcReleaseGuard<'_, Self>> { ... }
    fn GetDesktopWindow() -> HWND { ... }
    fn GetDlgCtrlID(&self) -> SysResult<u16> { ... }
    fn GetDlgItem(&self, ctrl_id: u16) -> SysResult<HWND> { ... }
    fn GetFocus() -> Option<HWND> { ... }
    fn GetForegroundWindow() -> Option<HWND> { ... }
    fn GetLastActivePopup(&self) -> Option<HWND> { ... }
    fn GetMenu(&self) -> Option<HMENU> { ... }
    fn GetMenuBarInfo(
        &self,
        obj_id: OBJID,
        item_id: u32,
        mbi: &mut MENUBARINFO
    ) -> SysResult<()> { ... }
    fn GetMenuItemRect(
        &self,
        hmenu: &HMENU,
        item_pos: u32,
        rc_item: &mut RECT
    ) -> SysResult<()> { ... }
    fn GetNextDlgGroupItem(
        &self,
        hwnd_ctrl: &HWND,
        previous: bool
    ) -> SysResult<HWND> { ... }
    fn GetNextDlgTabItem(
        &self,
        hwnd_ctrl: &HWND,
        previous: bool
    ) -> SysResult<HWND> { ... }
    fn GetParent(&self) -> SysResult<HWND> { ... }
    fn GetScrollInfo(&self, bar: SBB, si: &mut SCROLLINFO) -> SysResult<()> { ... }
    fn GetScrollPos(&self, bar: SBB) -> SysResult<i32> { ... }
    fn GetShellWindow() -> Option<HWND> { ... }
    fn GetSystemMenu(&self, revert: bool) -> Option<HMENU> { ... }
    fn GetTopWindow(&self) -> SysResult<Option<HWND>> { ... }
    fn GetUpdateRgn(&self, hrgn: &HRGN, erase: bool) -> SysResult<REGION> { ... }
    fn GetWindow(&self, cmd: GW) -> SysResult<HWND> { ... }
    fn GetWindowDC(&self) -> SysResult<HdcReleaseGuard<'_, Self>> { ... }
    fn GetWindowDisplayAffinity(&self) -> SysResult<WDA> { ... }
    fn GetWindowInfo(&self, wi: &mut WINDOWINFO) -> SysResult<()> { ... }
    fn GetWindowLongPtr(&self, index: GWLP) -> isize { ... }
    fn GetWindowPlacement(&self, wp: &mut WINDOWPLACEMENT) -> SysResult<()> { ... }
    fn GetWindowRect(&self) -> SysResult<RECT> { ... }
    fn GetWindowRgn(&self, hrgn: &HRGN) -> SysResult<REGION> { ... }
    fn GetWindowRgnBox(&self, rc: &mut RECT) -> SysResult<REGION> { ... }
    fn GetWindowText(&self) -> SysResult<String> { ... }
    fn GetWindowTextLength(&self) -> SysResult<i32> { ... }
    fn GetWindowThreadProcessId(&self) -> (u32, u32) { ... }
    fn HiliteMenuItem(&self, hmenu: &HMENU, id_or_pos: IdPos, hilite: MF) -> bool { ... }
    fn InvalidateRect(&self, rc: Option<&RECT>, erase: bool) -> SysResult<()> { ... }
    fn InvalidateRgn(&self, hrgn: &HRGN, erase: bool) { ... }
    fn IsChild(&self, hwnd_possible_child: &HWND) -> bool { ... }
    fn IsDialogMessage(&self, msg: &mut MSG) -> bool { ... }
    fn IsIconic(&self) -> bool { ... }
    fn IsWindow(&self) -> bool { ... }
    fn IsWindowEnabled(&self) -> bool { ... }
    fn IsWindowUnicode(&self) -> bool { ... }
    fn IsWindowVisible(&self) -> bool { ... }
    fn IsZoomed(&self) -> bool { ... }
    fn KillTimer(&self, event_id: usize) -> SysResult<()> { ... }
    fn LogicalToPhysicalPoint(&self, pt: *mut POINT) -> SysResult<()> { ... }
    fn MapDialogRect(&self, rc: &mut RECT) -> SysResult<()> { ... }
    fn MessageBox(&self, text: &str, caption: &str, flags: MB) -> SysResult<DLGID> { ... }
    fn MonitorFromWindow(&self, flags: MONITOR) -> HMONITOR { ... }
    fn MoveWindow(&self, pos: POINT, size: SIZE, repaint: bool) -> SysResult<()> { ... }
    fn OpenClipboard(&self) -> SysResult<ClipboardGuard<'_>> { ... }
    fn PostMessage<M>(&self, msg: M) -> SysResult<()>
    where
        M: MsgSend + Send + Copy + 'static,
    { ... }
    fn RealChildWindowFromPoint(
        &self,
        pt_parent_client_coords: POINT
    ) -> Option<HWND> { ... }
    fn RealGetWindowClass(&self) -> SysResult<String> { ... }
    fn RedrawWindow(
        &self,
        rc_update: &RECT,
        hrgn_update: &HRGN,
        flags: RDW
    ) -> SysResult<()> { ... }
    fn ScreenToClient(&self, pt: &mut POINT) -> SysResult<()> { ... }
    fn ScreenToClientRc(&self, rc: &mut RECT) -> SysResult<()> { ... }
    fn SendMessage<M>(&self, msg: M) -> M::RetType
    where
        M: MsgSend,
    { ... }
    fn SendMessageTimeout<M>(
        &self,
        msg: M,
        flags: SMTO,
        timeout_ms: u32
    ) -> SysResult<M::RetType>
    where
        M: MsgSend,
    { ... }
    fn SetCapture(&self) -> Option<HWND> { ... }
    fn SetFocus(&self) -> Option<HWND> { ... }
    fn SetForegroundWindow(&self) -> bool { ... }
    fn SetMenu(&self, hmenu: &HMENU) -> SysResult<()> { ... }
    fn SetParent(&self, hwnd_new_parent: &HWND) -> SysResult<Option<HWND>> { ... }
    fn SetScrollInfo(&self, bar: SBB, si: &SCROLLINFO, redraw: bool) -> i32 { ... }
    fn SetScrollPos(&self, b: SBB, pos: i32, redraw: bool) -> SysResult<i32> { ... }
    fn SetScrollRange(
        &self,
        bar: SBB,
        min_pos: i32,
        max_pos: i32,
        redraw: bool
    ) -> SysResult<()> { ... }
    fn SetTimer(
        &self,
        event_id: usize,
        elapse_ms: u32,
        timer_func: Option<TIMERPROC>
    ) -> SysResult<usize> { ... }
    fn SetWindowDisplayAffinity(&self, affinity: WDA) -> SysResult<()> { ... }
    fn SetWindowLongPtr(&self, index: GWLP, new_long: isize) -> isize { ... }
    fn SetWindowPlacement(&self, wp: &WINDOWPLACEMENT) -> SysResult<()> { ... }
    fn SetWindowPos(
        &self,
        hwnd_insert_after: HwndPlace,
        pos: POINT,
        size: SIZE,
        flags: SWP
    ) -> SysResult<()> { ... }
    fn SetWindowRgn(&self, hrgn: &HRGN, redraw: bool) -> SysResult<()> { ... }
    fn SetWindowText(&self, text: &str) -> SysResult<()> { ... }
    fn ShowCaret(&self) -> SysResult<()> { ... }
    fn ShowWindow(&self, show_cmd: SW) -> bool { ... }
    fn ShowWindowAsync(&self, show_cmd: SW) -> SysResult<()> { ... }
    fn TranslateAccelerator(
        &self,
        haccel_table: &HACCEL,
        msg: &mut MSG
    ) -> SysResult<()> { ... }
    fn UpdateWindow(&self) -> SysResult<()> { ... }
    fn ValidateRect(&self, rc: &RECT) -> SysResult<()> { ... }
    fn ValidateRgn(&self, hrgn: &HRGN) -> SysResult<()> { ... }
    fn WindowFromPhysicalPoint(pt: POINT) -> Option<HWND> { ... }
    fn WindowFromPoint(pt: POINT) -> Option<HWND> { ... }
    fn WinHelp(&self, help_file: &str, cmd: HELPW, data: usize) -> SysResult<()> { ... }

}

Available on crate feature user only.

This trait is enabled with the user feature, and provides methods for HWND.

Prefer importing this trait through the prelude:

use winsafe::prelude::*;

Provided Associated Constants

const BROADCAST: HWND = _

Represents all top-level windows in HWND::PostMessage and HWND::SendMessage.

const DESKTOP: HWND = _

Represents the desktop window in HWND::GetDC.

Provided Methods

fn hinstance(&self) -> HINSTANCE

GetWindowLongPtr wrapper to retrieve the window HINSTANCE.

fn ArrangeIconicWindows(&self) -> SysResult<u32>

ArrangeIconicWindows method.

fn BeginPaint(&self) -> SysResult<HdcPaintGuard<'_, Self>>

BeginPaint method.

Examples

use winsafe::prelude::*;
use winsafe::HWND;

let hwnd: HWND; // initialized somewhere

let hdc = hwnd.BeginPaint()?;

println!("Erase background? {}", hdc.paintstruct().fErase());
println!("Painting area: {}", hdc.paintstruct().rcPaint);

// hdc painting...

Note that the HDC guard returned by this method must be kept alive while the painting is made, even if you won’t use it directly. This is necessary because, when the guard goes out of scope, it will automatically call EndPaint.

In the example below, the HDC is not used, but the returned guard is kept alive:

use winsafe::prelude::*;
use winsafe::HWND;

let hwnd: HWND; // initialized somewhere

let _ = hwnd.BeginPaint()?; // keep the returned guard alive

// hdc painting...

fn BringWindowToTop(&self) -> SysResult<()>

BringWindowToTop method.

fn ChildWindowFromPoint(&self, pt: POINT) -> Option<HWND>

ChildWindowFromPoint method.

fn ClientToScreen(&self, pt: &mut POINT) -> SysResult<()>

ClientToScreen method.

fn ClientToScreenRc(&self, rc: &mut RECT) -> SysResult<()>

ClientToScreen method for a RECT.

fn CloseWindow(&self) -> SysResult<()>

CloseWindow method.

Note that this method will actually minimize the window, not destroy it.

unsafe fn CreateWindowEx(
ex_style: WS_EX,
class_name: AtomStr,
title: Option<&str>,
style: WS,
pos: POINT,
size: SIZE,
hwnd_parent: Option<&HWND>,
hmenu: IdMenu<'_>,
hinstance: &HINSTANCE,
lparam: Option<isize>
) -> SysResult<HWND>

CreateWindowEx static method.

fn DefWindowProc<M>(&self, msg: M) -> M::RetTypewhere
M: MsgSend,

DefWindowProc method.

The return type is variable, being defined by the RetType associated type of the MsgSend trait. That means each message can define its own return type.

fn DestroyWindow(&self) -> SysResult<()>

DestroyWindow method.

Usually you don’t need to call this method directly, since it’s automatically called inside the internal message loop. The ordinary way to close a window is sending a wm::Close message.

After calling this method, the handle will be invalidated and further operations will fail with ERROR::INVALID_HANDLE error code.

fn DrawMenuBar(&self) -> SysResult<()>

DrawMenuBar method.

fn EnableWindow(&self, enable: bool) -> bool

EnableWindow method.

fn EndDialog(&self, result: isize) -> SysResult<()>

EndDialog method.

fn EnumChildWindows<F>(&self, func: F)where
F: Fn(HWND) -> bool,

EnumChildWindows method.

Examples

use winsafe::prelude::*;
use winsafe::HWND;

let hwnd: HWND; // initialized somewhere

hwnd.EnumChildWindows(|hchild: HWND| -> bool {
    println!("Child HWND: {}", hchild);
    true
});

fn FindWindow(
class_name: Option<AtomStr>,
title: Option<&str>
) -> SysResult<HWND>

FindWindow static method.

fn FindWindowEx(
&self,
hwnd_child_after: Option<&HWND>,
class_name: AtomStr,
title: Option<&str>
) -> SysResult<HWND>

FindWindowEx method.

fn GetActiveWindow() -> Option<HWND>

GetActiveWindow static method.

fn GetAltTabInfo(
&self,
item: Option<u32>,
ati: &mut ALTTABINFO,
sz_item_text: Option<u32>
) -> SysResult<String>

GetAltTabInfo method.

If item is None, the item text is not retrieved.

The sz_item_text is the maximum number of expected chars for the item text. If None, defaults to 100.

fn GetAncestor(&self, flags: GA) -> Option<HWND>

GetAncestor method.

fn GetCapture() -> Option<HWND>

GetCapture static method.

fn GetClassLongPtr(&self, index: GCLP) -> usize

GetClassLongPtr method.

fn GetClassName(&self) -> SysResult<String>

GetClassName method.

fn GetClientRect(&self) -> SysResult<RECT>

GetClientRect method.

fn GetDC(&self) -> SysResult<HdcReleaseGuard<'_, Self>>

GetDC method.

To get the device context of the desktop window, use the predefined HWND::DESKTOP.

fn GetDesktopWindow() -> HWND

GetDesktopWindow static method.

fn GetDlgCtrlID(&self) -> SysResult<u16>

GetDlgCtrlID method.

fn GetDlgItem(&self, ctrl_id: u16) -> SysResult<HWND>

GetDlgItem method.

fn GetFocus() -> Option<HWND>

GetFocus static method.

fn GetForegroundWindow() -> Option<HWND>

GetForegroundWindow static method.

fn GetLastActivePopup(&self) -> Option<HWND>

GetLastActivePopup method.

fn GetMenu(&self) -> Option<HMENU>

GetMenu method.

fn GetMenuBarInfo(
&self,
obj_id: OBJID,
item_id: u32,
mbi: &mut MENUBARINFO
) -> SysResult<()>

GetMenuBarInfo method.

fn GetMenuItemRect(
&self,
hmenu: &HMENU,
item_pos: u32,
rc_item: &mut RECT
) -> SysResult<()>

GetMenuItemRect method.

fn GetNextDlgGroupItem(&self, hwnd_ctrl: &HWND, previous: bool) -> SysResult<HWND>

GetNextDlgGroupItem method.

fn GetNextDlgTabItem(&self, hwnd_ctrl: &HWND, previous: bool) -> SysResult<HWND>

GetNextDlgTabItem method.

fn GetParent(&self) -> SysResult<HWND>

GetParent method.

fn GetScrollInfo(&self, bar: SBB, si: &mut SCROLLINFO) -> SysResult<()>

GetScrollInfo method.

fn GetScrollPos(&self, bar: SBB) -> SysResult<i32>

GetScrollPos method.

fn GetShellWindow() -> Option<HWND>

GetShellWindow static method.

fn GetSystemMenu(&self, revert: bool) -> Option<HMENU>

GetSystemMenu method.

fn GetTopWindow(&self) -> SysResult<Option<HWND>>

GetTopWindow method.

fn GetUpdateRgn(&self, hrgn: &HRGN, erase: bool) -> SysResult<REGION>

GetUpdateRgn method.

fn GetWindow(&self, cmd: GW) -> SysResult<HWND>

GetWindow method.

fn GetWindowDC(&self) -> SysResult<HdcReleaseGuard<'_, Self>>

GetWindowDC method.

fn GetWindowDisplayAffinity(&self) -> SysResult<WDA>

GetWindowDisplayAffinity method.

fn GetWindowInfo(&self, wi: &mut WINDOWINFO) -> SysResult<()>

GetWindowInfo method.

fn GetWindowLongPtr(&self, index: GWLP) -> isize

GetWindowLongPtr method.

fn GetWindowPlacement(&self, wp: &mut WINDOWPLACEMENT) -> SysResult<()>

GetWindowPlacement method.

fn GetWindowRect(&self) -> SysResult<RECT>

GetWindowRect method.

fn GetWindowRgn(&self, hrgn: &HRGN) -> SysResult<REGION>

GetWindowRgn method.

fn GetWindowRgnBox(&self, rc: &mut RECT) -> SysResult<REGION>

GetWindowRgnBox method.

fn GetWindowText(&self) -> SysResult<String>

GetWindowText method.

Returns a String, performing all necessary allocations.

use winsafe::prelude::*;
use winsafe::HWND;

let hwnd: HWND; // initialized somewhere

let text = hwnd.GetWindowText()?;
println!("Text: {}", text);

fn GetWindowTextLength(&self) -> SysResult<i32>

GetWindowTextLength method. Does not count the terminating null.

You should rarely use this method since GetWindowText returns a String, performing all necessary allocations.

fn GetWindowThreadProcessId(&self) -> (u32, u32)

GetWindowThreadProcessId method.

Returns thread ID and process ID, respectively.

fn HiliteMenuItem(&self, hmenu: &HMENU, id_or_pos: IdPos, hilite: MF) -> bool

HiliteMenuItem method.

fn InvalidateRect(&self, rc: Option<&RECT>, erase: bool) -> SysResult<()>

InvalidateRect method.

Examples

Most of the time you’ll just want update the entire client area:

use winsafe::prelude::*;
use winsafe::HWND;

let hwnd: HWND; // initialized somewhere

hwnd.InvalidateRect(None, true)?;

fn InvalidateRgn(&self, hrgn: &HRGN, erase: bool)

InvalidateRgn method.

fn IsChild(&self, hwnd_possible_child: &HWND) -> bool

IsChild method.

fn IsDialogMessage(&self, msg: &mut MSG) -> bool

IsDialogMessage method.

fn IsIconic(&self) -> bool

IsIconic method.

fn IsWindow(&self) -> bool

IsWindow method.

fn IsWindowEnabled(&self) -> bool

IsWindowEnabled method.

fn IsWindowUnicode(&self) -> bool

IsWindowUnicode method.

fn IsWindowVisible(&self) -> bool

IsWindowVisible method.

fn IsZoomed(&self) -> bool

IsZoomed method.

fn KillTimer(&self, event_id: usize) -> SysResult<()>

KillTimer method.

fn LogicalToPhysicalPoint(&self, pt: *mut POINT) -> SysResult<()>

LogicalToPhysicalPoint method.

fn MapDialogRect(&self, rc: &mut RECT) -> SysResult<()>

MapDialogRect method.

fn MessageBox(&self, text: &str, caption: &str, flags: MB) -> SysResult<DLGID>

MessageBox method.

Consider using the more modern TaskDialog method.

Examples

A modal message box, which blocks its parent:

use winsafe::prelude::*;
use winsafe::{co, HWND};

let my_hwnd: HWND; // initialized somewhere

hwnd.MessageBox("Hello, world", "title",
    co::MB::OKCANCEL | co::MB::ICONINFORMATION)?;

Usually the message box has a valid parent window, however, if for some reason you don’t have a window to serve as parent, you still can show a non-modal, parent-less message box by retrieving the desktop handle:

use winsafe::prelude::*;
use winsafe::{co, HWND};

HWND::GetDesktopWindow()
    .MessageBox("Hello, world", "Title", co::MB::ICONEXCLAMATION)?;

fn MonitorFromWindow(&self, flags: MONITOR) -> HMONITOR

MonitorFromWindow method.

fn MoveWindow(&self, pos: POINT, size: SIZE, repaint: bool) -> SysResult<()>

MoveWindow method.

fn OpenClipboard(&self) -> SysResult<ClipboardGuard<'_>>

OpenClipboard method.

Note that the guard returned by this method must be kept alive while you work upon the clipboard. This is necessary because, when the guard goes out of scope, it will automatically call CloseClipboard.

In the example below, the returned guard is kept alive:

use winsafe::prelude::*;
use winsafe::HWND;

let hwnd: HWND; // initialized somewhere

let _ = hwnd.OpenClipboard()?; // keep the returned guard alive

fn PostMessage<M>(&self, msg: M) -> SysResult<()>where
M: MsgSend + Send + Copy + 'static,

PostMessage method. Note that this method is asychronous.

fn RealChildWindowFromPoint(
&self,
pt_parent_client_coords: POINT
) -> Option<HWND>

RealChildWindowFromPoint method.

fn RealGetWindowClass(&self) -> SysResult<String>

RealGetWindowClassW method.

fn RedrawWindow(
&self,
rc_update: &RECT,
hrgn_update: &HRGN,
flags: RDW
) -> SysResult<()>

RedrawWindow method.

fn ScreenToClient(&self, pt: &mut POINT) -> SysResult<()>

ScreenToClient method.

fn ScreenToClientRc(&self, rc: &mut RECT) -> SysResult<()>

ScreenToClient method for a RECT.

fn SendMessage<M>(&self, msg: M) -> M::RetTypewhere
M: MsgSend,

SendMessage method.

The return type is variable, being defined by the RetType associated type of the MsgSend trait. That means each message can define its own return type.

Examples

Sending a bm::GetImage button message, which demands an image type parameter. Note that this specific message can also return an error, which is handled with ?:

use winsafe::prelude::*;
use winsafe::{co, HWND, msg::bm};

let hwnd: HWND; // initialized somewhere

let bmp = hwnd.SendMessage(bm::GetImage {
    img_type: co::IMAGE_TYPE::BITMAP,
})?;

Sending an em::CharFromPos edit message, which receives point coordinates and returns two values:

use winsafe::prelude::*;
use winsafe::{HWND, msg::em, POINT};

let hwnd: HWND; // initialized somewhere

let (char_pos, line_pos) = hwnd.SendMessage(
    em::CharFromPos {
        coords: POINT::new(12, 20),
    },
);

fn SendMessageTimeout<M>(
&self,
msg: M,
flags: SMTO,
timeout_ms: u32
) -> SysResult<M::RetType>where
M: MsgSend,

SendMessageTimeout method.

fn SetCapture(&self) -> Option<HWND>

SetCapture method.

Note: Must be paired with a ReleaseCapture call.

fn SetFocus(&self) -> Option<HWND>

SetFocus method.

fn SetForegroundWindow(&self) -> bool

SetForegroundWindow method.

fn SetMenu(&self, hmenu: &HMENU) -> SysResult<()>

SetMenu method.

fn SetParent(&self, hwnd_new_parent: &HWND) -> SysResult<Option<HWND>>

SetParent method.

fn SetScrollInfo(&self, bar: SBB, si: &SCROLLINFO, redraw: bool) -> i32

SetScrollInfo method.

fn SetScrollPos(&self, b: SBB, pos: i32, redraw: bool) -> SysResult<i32>

SetScrollPos method.

fn SetScrollRange(
&self,
bar: SBB,
min_pos: i32,
max_pos: i32,
redraw: bool
) -> SysResult<()>

SetScrollRange method.

fn SetTimer(
&self,
event_id: usize,
elapse_ms: u32,
timer_func: Option<TIMERPROC>
) -> SysResult<usize>

SetTimer method.

fn SetWindowDisplayAffinity(&self, affinity: WDA) -> SysResult<()>

SetWindowDisplayAffinity method.

fn SetWindowLongPtr(&self, index: GWLP, new_long: isize) -> isize

SetWindowLongPtr method.

fn SetWindowPlacement(&self, wp: &WINDOWPLACEMENT) -> SysResult<()>

SetWindowPlacement method.

fn SetWindowPos(
&self,
hwnd_insert_after: HwndPlace,
pos: POINT,
size: SIZE,
flags: SWP
) -> SysResult<()>

SetWindowPos method.

fn SetWindowRgn(&self, hrgn: &HRGN, redraw: bool) -> SysResult<()>

SetWindowRgn method.

fn SetWindowText(&self, text: &str) -> SysResult<()>

SetWindowText method.

fn ShowCaret(&self) -> SysResult<()>

ShowCaret method.

fn ShowWindow(&self, show_cmd: SW) -> bool

ShowWindow method.

fn ShowWindowAsync(&self, show_cmd: SW) -> SysResult<()>

ShowWindowAsync method.

fn TranslateAccelerator(
&self,
haccel_table: &HACCEL,
msg: &mut MSG
) -> SysResult<()>

TranslateAccelerator method.

fn UpdateWindow(&self) -> SysResult<()>

UpdateWindow method.

fn ValidateRect(&self, rc: &RECT) -> SysResult<()>

ValidateRect method.

fn ValidateRgn(&self, hrgn: &HRGN) -> SysResult<()>

ValidateRgn method.

fn WindowFromPhysicalPoint(pt: POINT) -> Option<HWND>

WindowFromPhysicalPoint static method.

fn WindowFromPoint(pt: POINT) -> Option<HWND>

WindowFromPoint static method.

fn WinHelp(&self, help_file: &str, cmd: HELPW, data: usize) -> SysResult<()>

WinHelp method.