Struct winsafe::DEVMODE

#[repr(C)]
pub struct DEVMODE {

    pub dmDriverVersion: u16,
    pub dmDriverExtra: u16,
    pub dmFields: DM,
    pub dmColor: DMCOLOR,
    pub dmDuplex: DMDUP,
    pub dmYResolution: u16,
    pub dmTTOption: DMTT,
    pub dmLogPixels: u16,
    pub dmBitsPerPel: u32,
    pub dmPelsWidth: u32,
    pub dmPelsHeight: u32,
    pub dmDisplayFrequency: u32,
    pub dmICMMethod: DMICMMETHOD,
    pub dmICMIntent: DMICM,
    pub dmMediaType: DMMEDIA,
    pub dmDitherType: DMDITHER,
    // some fields omitted

}

DEVMODE struct.

Fields
dmDriverVersion: u16dmDriverExtra: u16dmFields: DMdmColor: DMCOLORdmDuplex: DMDUPdmYResolution: u16dmTTOption: DMTTdmLogPixels: u16dmBitsPerPel: u32dmPelsWidth: u32dmPelsHeight: u32dmDisplayFrequency: u32dmICMMethod: DMICMMETHODdmICMIntent: DMICMdmMediaType: DMMEDIAdmDitherType: DMDITHER
Implementations

impl DEVMODE

pub fn dmDeviceName(&self) -> String

Returns the string field.

pub fn set_dmDeviceName(&mut self, text: &str)

Sets the string field.

pub fn dmCollate(&self) -> bool

Returns the bool field.

pub fn set_dmCollate(&mut self, val: bool)

Sets the bool field.

pub fn dmFormName(&self) -> String

Returns the string field.

pub fn set_dmFormName(&mut self, text: &str)

Sets the string field.

pub fn set_dmDisplayFlags_dmNup(&mut self, val: DispfNup)

Sets the dmDisplayFlags or the dmNup field.

pub fn dmOrientation(&self) -> DMORIENT

Returns the dmOrientation printer field, which is part of an union.

pub fn set_dmOrientation(&mut self, val: DMORIENT)

Sets the dmOrientation printer field, which is part of an union.

pub fn dmPaperSize(&self) -> DMPAPER

Returns the dmPaperSize printer field, which is part of an union.

pub fn set_dmPaperSize(&mut self, val: DMPAPER)

Sets the dmPaperSize printer field, which is part of an union.

pub fn dmPaperLength(&self) -> i16

Returns the dmPaperLength printer field, which is part of an union.

pub fn set_dmPaperLength(&mut self, val: i16)

Sets the dmPaperLength printer field, which is part of an union.

pub fn dmPaperWidth(&self) -> i16

Returns the dmPaperWidth printer field, which is part of an union.

pub fn set_dmPaperWidth(&mut self, val: i16)

Sets the dmPaperWidth printer field, which is part of an union.

pub fn dmScale(&self) -> i16

Returns the dmScale printer field, which is part of an union.

pub fn set_dmScale(&mut self, val: i16)

Sets the dmScale printer field, which is part of an union.

pub fn dmCopies(&self) -> i16

Returns the dmCopies printer field, which is part of an union.

pub fn set_dmCopies(&mut self, val: i16)

Sets the dmCopies printer field, which is part of an union.

pub fn dmDefaultSource(&self) -> DMBIN

Returns the dmDefaultSource printer field, which is part of an union.

pub fn set_dmDefaultSource(&mut self, val: DMBIN)

Sets the dmDefaultSource printer field, which is part of an union.

pub fn dmPrintQuality(&self) -> DMRES

Returns the dmPrintQuality printer field, which is part of an union.

pub fn set_dmPrintQuality(&mut self, val: DMRES)

Sets the dmPrintQuality printer field, which is part of an union.

pub fn dmPosition(&self) -> POINT

Returns the dmPosition display field, which is part of an union.

pub fn set_dmPosition(&mut self, val: POINT)

Sets the dmPosition display field, which is part of an union.

pub fn dmDisplayOrientation(&self) -> DMDO

Returns the dmDisplayOrientation display field, which is part of an union.

pub fn set_dmDisplayOrientation(&mut self, val: DMDO)

Sets the dmDisplayOrientation display field, which is part of an union.

pub fn dmDisplayFixedOutput(&self) -> DMDFO

Returns the dmDisplayFixedOutput display field, which is part of an union.

pub fn set_dmDisplayFixedOutput(&mut self, val: DMDFO)

Sets the dmDisplayFixedOutput display field, which is part of an union.