Struct winsafe::HDC

#[repr(C)]pub struct HDC { /* fields omitted */ }

Handle to a device context. Exposes methods.

Implementations

impl HDC

pub unsafe fn from_ptr<T>(p: *mut T) -> HDC

Creates a new handle instance by wrapping a pointer.

pub unsafe fn null_handle() -> Self

Creates a null, invalid handle.

pub unsafe fn as_ptr(self) -> *mut c_void

Consumes the handle returning the underlying raw pointer.

pub fn is_null(self) -> bool

Tells if the handle is invalid (null).

pub fn as_opt(self) -> Option<HDC>

Consumes the handle into an option, which is None if the handle pointer is null.

impl HDC

pub fn CreateCompatibleDC(self) -> WinResult<HDC>

CreateCompatibleDC method.

Note: Must be paired with a DeleteDC call.

pub fn DeleteDC(self) -> WinResult<()>

DeleteDC method.

pub fn GetDeviceCaps(self, index: GDC) -> i32

GetDeviceCaps method.

pub fn GetTextExtentPoint32(self, lpString: &str) -> WinResult<SIZE>

GetTextExtentPoint32 method.

pub fn LineTo(self, x: i32, y: i32) -> WinResult<()>

LineTo method.

pub fn MoveToEx(self, x: i32, y: i32, lppt: Option<&mut POINT>) -> WinResult<()>

MoveToEx method.

pub fn Pie(
self,
left: i32,
top: i32,
right: i32,
bottom: i32,
xr1: i32,
yr1: i32,
xr2: i32,
yr2: i32
) -> WinResult<()>

Pie method.

pub fn PolyBezier(self, apt: &[POINT]) -> WinResult<()>

PolyBezier method.

pub fn PolyBezierTo(self, apt: &[POINT]) -> WinResult<()>

PolyBezierTo method.

pub fn Polyline(self, apt: &[POINT]) -> WinResult<()>

Polyline method.

pub fn PolylineTo(self, apt: &[POINT]) -> WinResult<()>

PolylineTo method.

pub fn PtVisible(self, x: i32, y: i32) -> WinResult<bool>

PtVisible method.

pub fn Rectangle(
self,
left: i32,
top: i32,
right: i32,
bottom: i32
) -> WinResult<()>

Rectangle method.

pub fn RestoreDC(self, nSavedDC: i32) -> WinResult<()>

RestoreDC method.

pub fn RoundRect(
self,
left: i32,
top: i32,
right: i32,
bottom: i32,
width: i32,
height: i32
) -> WinResult<()>

RoundRect method.

pub fn SaveDC(self) -> WinResult<i32>

SaveDC method.

pub fn SelectObjectBitmap(self, h: HBITMAP) -> WinResult<HBITMAP>

SelectObject method for HBITMAP.

pub fn SelectObjectBrush(self, h: HBRUSH) -> WinResult<HBRUSH>

SelectObject method for HBRUSH.

pub fn SelectObjectFont(self, h: HFONT) -> WinResult<HFONT>

SelectObject method for HFONT.

pub fn SelectObjectPen(self, h: HPEN) -> WinResult<HPEN>

SelectObject method for HPEN.

pub fn SelectObjectRgn(self, h: HRGN) -> WinResult<REGION>

SelectObject method for HRGN.

pub fn SetBkMode(self, mode: BKMODE) -> WinResult<BKMODE>

SetBkMode method