Struct winsafe::HDC

#[repr(C)]
pub struct HDC { /* fields omitted */ }

Handle to a device context.

Implementations

impl HDC

pub const NULL: Self

The null, invalid handle.

This constant is common to all handle types.

pub unsafe fn from_ptr<T>(p: *mut T) -> HDC

Creates a new handle instance by wrapping a pointer.

This method is common to all handle types.

pub unsafe fn as_ptr(self) -> *mut c_void

Consumes the handle returning the underlying raw pointer.

This method is common to all handle types.

pub fn is_null(self) -> bool

Tells if the handle is invalid (null).

This method is common to all handle types.

pub fn as_opt(self) -> Option<HDC>

Consumes the handle into an option, which is None if the handle pointer is null.

This method is common to all handle types.

impl HDC

pub fn AbortPath(self) -> WinResult<()>

AborthPath method.

pub fn AngleArc(
self,
x: i32,
y: i32,
r: u32,
StartAngle: f32,
SweepAngle: f32
) -> WinResult<()>

AngleArc method.

pub fn BeginPath(self) -> WinResult<()>

BeginPath method.

pub fn BitBlt(
self,
x: i32,
y: i32,
cx: i32,
cy: i32,
hdcSrc: HDC,
x1: i32,
y1: i32,
rop: ROP
) -> WinResult<()>

BitBlt method.

pub fn CancelDC(self) -> WinResult<()>

CancelDC method.

pub fn Chord(
self,
x1: i32,
y1: i32,
x2: i32,
y2: i32,
x3: i32,
y3: i32,
x4: i32,
y4: i32
) -> WinResult<()>

Chord method.

pub fn CreateCompatibleBitmap(self, cx: i32, cy: i32) -> WinResult<HBITMAP>

CreateCompatibleBitmap method.

Note: Must be paired with an HBITMAP::DeleteObject call.

pub fn CreateCompatibleDC(self) -> WinResult<HDC>

CreateCompatibleDC method.

Note: Must be paired with an HDC::DeleteDC call.

pub fn DeleteDC(self) -> WinResult<()>

DeleteDC method.

pub fn EndPath(self) -> WinResult<()>

EndPath method.

pub fn FillPath(self) -> WinResult<()>

FillPath method.

pub fn FillRect(self, lprc: RECT, hbr: HBRUSH) -> WinResult<()>

FillRect method.

pub fn GetDCBrushColor(self) -> WinResult<COLORREF>

GetDCBrushColor method.

pub fn GetDCPenColor(self) -> WinResult<COLORREF>

GetDCPenColor method.

pub fn GetDeviceCaps(self, index: GDC) -> i32

GetDeviceCaps method.

pub fn GetTextColor(self) -> WinResult<COLORREF>

GetTextColor method.

pub fn GetTextExtentPoint32(self, lpString: &str) -> WinResult<SIZE>

GetTextExtentPoint32 method.

pub fn GetTextMetrics(self, lptm: &mut TEXTMETRIC) -> WinResult<()>

GetTextMetrics method.

pub fn LineTo(self, x: i32, y: i32) -> WinResult<()>

LineTo method.

pub fn MoveToEx(self, x: i32, y: i32, lppt: Option<&mut POINT>) -> WinResult<()>

MoveToEx method.

pub fn PatBlt(self, x: i32, y: i32, w: i32, h: i32, rop: ROP) -> WinResult<()>

PatBlt method.

pub fn PathToRegion(self) -> WinResult<HRGN>

PathToRegion method.

Note: Must be paired with an HRGN::DeleteObject call.

pub fn Pie(
self,
left: i32,
top: i32,
right: i32,
bottom: i32,
xr1: i32,
yr1: i32,
xr2: i32,
yr2: i32
) -> WinResult<()>

Pie method.

pub fn PolyBezier(self, apt: &[POINT]) -> WinResult<()>

PolyBezier method.

pub fn PolyBezierTo(self, apt: &[POINT]) -> WinResult<()>

PolyBezierTo method.

pub fn Polyline(self, apt: &[POINT]) -> WinResult<()>

Polyline method.

pub fn PolylineTo(self, apt: &[POINT]) -> WinResult<()>

PolylineTo method.

pub fn PtVisible(self, x: i32, y: i32) -> WinResult<bool>

PtVisible method.

pub fn Rectangle(
self,
left: i32,
top: i32,
right: i32,
bottom: i32
) -> WinResult<()>

Rectangle method.

pub fn RestoreDC(self, nSavedDC: i32) -> WinResult<()>

RestoreDC method.

pub fn RoundRect(
self,
left: i32,
top: i32,
right: i32,
bottom: i32,
width: i32,
height: i32
) -> WinResult<()>

RoundRect method.

pub fn SaveDC(self) -> WinResult<i32>

SaveDC method.

pub fn SelectObjectBitmap(self, h: HBITMAP) -> WinResult<HBITMAP>

SelectObject method for HBITMAP.

pub fn SelectObjectBrush(self, h: HBRUSH) -> WinResult<HBRUSH>

SelectObject method for HBRUSH.

pub fn SelectObjectFont(self, h: HFONT) -> WinResult<HFONT>

SelectObject method for HFONT.

pub fn SelectObjectPen(self, h: HPEN) -> WinResult<HPEN>

SelectObject method for HPEN.

pub fn SelectObjectRgn(self, h: HRGN) -> WinResult<REGION>

SelectObject method for HRGN.

pub fn SetArcDirection(self, dir: AD) -> WinResult<AD>

SetArcDirection method.

pub fn SetBkMode(self, mode: BKMODE) -> WinResult<BKMODE>

SetBkMode method.

pub fn SetDCBrushColor(self, color: COLORREF) -> WinResult<COLORREF>

SetDCBrushColor

pub fn SetDCPenColor(self, color: COLORREF) -> WinResult<COLORREF>

SetDCPenColor

pub fn SetGraphicsMode(self, iMode: GM) -> WinResult<GM>

SetGraphicsMode method.

pub fn SetTextAlign(self, align: TA) -> WinResult<TA>

SetTextAlign method.

pub fn SetTextColor(self, color: COLORREF) -> WinResult<COLORREF>

SetTextColor method.

pub fn SetTextJustification(self, extra: i32, count: i32) -> WinResult<()>

SetTextJustification method.

pub fn SetViewportExtEx(self, x: i32, y: i32) -> WinResult<SIZE>

SetViewportExtEx method.

pub fn SetViewportOrgEx(self, x: i32, y: i32) -> WinResult<POINT>

SetViewportOrgEx method.

pub fn SetWindowExtEx(self, x: i32, y: i32) -> WinResult<SIZE>

SetWindowExtEx method.

pub fn SetWindowOrgEx(self, x: i32, y: i32) -> WinResult<POINT>

SetWindowOrgEx method.

pub fn StrokeAndFillPath(self) -> WinResult<()>

StrokeAndFillPath method.

pub fn StrokePath(self) -> WinResult<()>

StrokePath method.

pub fn TextOut(self, x: i32, y: i32, lpString: &str) -> WinResult<()>

TextOut method.

pub fn TransparentBlt(
self,
xoriginDest: i32,
yoriginDest: i32,
wDest: i32,
hDest: i32,
hdcSrc: HDC,
xoriginSrc: i32,
yoriginSrc: i32,
wSrc: i32,
hSrc: i32,
crTransparent: COLORREF
) -> WinResult<()>

TransparentBlt method.

pub fn WidenPath(self) -> WinResult<()>

WidenPath method.