Struct winsafe::HFILE

#[repr(C)]pub struct HFILE { /* fields omitted */ }

Handle to a file. Originally just a HANDLE.

Implementations

impl HFILE

pub unsafe fn from_ptr<T>(p: *mut T) -> HFILE

Creates a new handle instance by wrapping a pointer.

This method is common to all handle types.

pub unsafe fn null_handle() -> Self

Creates a null, invalid handle.

This method is common to all handle types.

pub unsafe fn as_ptr(self) -> *mut c_void

Consumes the handle returning the underlying raw pointer.

This method is common to all handle types.

pub fn is_null(self) -> bool

Tells if the handle is invalid (null).

This method is common to all handle types.

pub fn as_opt(self) -> Option<HFILE>

Consumes the handle into an option, which is None if the handle pointer is null.

This method is common to all handle types.

impl HFILE

pub fn CloseHandle(self) -> WinResult<()>

CloseHandle method.

pub fn CreateFile(
lpFileName: &str,
dwDesiredAccess: GENERIC,
dwShareMode: FILE_SHARE,
lpSecurityAttributes: Option<&mut SECURITY_ATTRIBUTES>,
dwCreationDisposition: DISPOSITION,
dwFlagsAndAttributes: FILE_ATTRIBUTE,
hTemplateFile: Option<HFILE>
) -> WinResult<HFILE>

CreateFile static method.

Note: Must be paired with a CloseHandle call.

Examples

Opening an existing file as read-only:

use winsafe::{co, HFILE};

let hfile = HFILE::CreateFile(
    "C:\\Temp\\something.txt",
    co::GENERIC::READ,
    co::FILE_SHARE::READ,
    None,
    co::DISPOSITION::OPEN_EXISTING,
    co::FILE_ATTRIBUTE::NORMAL,
    None,
).unwrap();

hfile.CloseHandle().unwrap();

Opening a file for read and write. If the file doesn't exist, create it:

use winsafe::{co, HFILE};

let hfile = w::HFILE::CreateFile(
    "C:\\Temp\\something.txt",
    co::GENERIC::READ | co::GENERIC::WRITE,
    co::FILE_SHARE::NONE,
    None,
    co::DISPOSITION::OPEN_ALWAYS,
    co::FILE_ATTRIBUTE::NORMAL,
    None,
).unwrap();

hfile.CloseHandle().unwrap();

pub fn CreateFileMapping(
self,
lpFileMappingAttributes: Option<&mut SECURITY_ATTRIBUTES>,
flProtect: PAGE,
maximumSize: u64,
lpName: Option<&str>
) -> WinResult<HFILEMAP>

CreateFileMapping method.

Note: Must be paired with a CloseHandle call.

pub fn GetFileInformationByHandle(
self,
lpFileInformation: &mut BY_HANDLE_FILE_INFORMATION
) -> WinResult<()>

GetFileInformationByHandle method.

pub fn GetFileSizeEx(self) -> WinResult<i64>

GetFileSizeEx method.

pub fn GetFileType(self) -> WinResult<FILE_TYPE>

GetFileType method.

pub fn LockFile(self, offset: u64, numBytesToLock: u64) -> WinResult<()>

LockFile method.

Note: Must be paired with an UnlockFile call.

pub fn ReadFile(
self,
numBytesToRead: u32,
lpOverlapped: Option<&mut OVERLAPPED>
) -> WinResult<Vec<u8>>

ReadFile method.

pub fn SetEndOfFile(self) -> WinResult<()>

SetEndOfFile method.

pub fn SetFilePointerEx(
self,
liDistanceToMove: i64,
dwMoveMethod: FILE_STARTING_POINT
) -> WinResult<i64>

SetFilePointerEx method.

pub fn UnlockFile(self, offset: u64, numBytesToLock: u64) -> WinResult<()>

UnlockFile method.

pub fn WriteFile(
self,
buffer: &[u8],
lpOverlapped: Option<&mut OVERLAPPED>
) -> WinResult<u32>

WriteFile method.