Struct winsafe::HINSTANCE

#[repr(C)]
pub struct HINSTANCE { /* fields omitted */ }

Handle to an instance, same as HMODULE.

Implementations

impl HINSTANCE

pub const NULL: Self

The null, invalid handle.

This constant is common to all handle types.

pub unsafe fn from_ptr<T>(p: *mut T) -> HINSTANCE

Creates a new handle instance by wrapping a pointer.

This method is common to all handle types.

pub unsafe fn as_ptr(self) -> *mut c_void

Consumes the handle returning the underlying raw pointer.

This method is common to all handle types.

pub fn is_null(self) -> bool

Tells if the handle is invalid (null).

This method is common to all handle types.

pub fn as_opt(self) -> Option<HINSTANCE>

Consumes the handle into an option, which is None if the handle pointer is null.

This method is common to all handle types.

impl HINSTANCE

pub fn CreateDialogParam(
self,
lpTemplateName: u16,
hWndParent: Option<HWND>,
lpDialogFunc: DLGPROC,
dwInitParam: Option<isize>
) -> WinResult<HWND>

CreateDialogParam method.

pub fn DialogBoxParam(
self,
lpTemplateName: u16,
hWndParent: Option<HWND>,
lpDialogFunc: DLGPROC,
dwInitParam: Option<isize>
) -> WinResult<isize>

DialogBoxParam method.

pub fn FreeLibrary(self) -> WinResult<()>

FreeLibrary method.

pub fn GetClassInfoEx(
self,
lpszClass: &str,
lpwcx: &mut WNDCLASSEX<'_>
) -> WinResult<ATOM>

GetClassInfoEx method.

Examples

Retrieving information of a window class created in our application:

use winsafe::{HINSTANCE, WNDCLASSEX};

let mut wcx = WNDCLASSEX::default();
HINSTANCE::GetModuleHandle(None).unwrap()
    .GetClassInfoEx("SOME_CLASS_NAME", &mut wcx).unwrap();

pub fn GetModuleFileName(self) -> WinResult<String>

GetModuleFileName method.

Examples

Retrieving the full path of currently running .exe file:

use winsafe::HINSTANCE;

println!("EXE: {}", HINSTANCE::NULL.GetModuleFileName().unwrap());

pub fn GetModuleHandle(lpModuleName: Option<&str>) -> WinResult<HINSTANCE>

GetModuleHandle static method.

Examples

Retrieving current module instance:

use winsafe::HINSTANCE;

let hinstance = HINSTANCE::GetModuleHandle(None).unwrap();

pub fn GetProcAddress(self, lpProcName: &str) -> WinResult<*const c_void>

GetProcAddress method.

pub fn LoadAccelerators(self, lpTableName: u16) -> WinResult<HACCEL>

LoadAccelerators method.

pub fn LoadCursor(self, lpCursorName: IdIdc) -> WinResult<HCURSOR>

LoadCursor method.

Examples

Loading a system cursor:

use winsafe::{co, HINSTANCE, IdIdc};

let sys_cursor = HINSTANCE::default()
    .LoadCursor(IdIdc::Idc(co::IDC::ARROW))
    .unwrap();

pub fn LoadIcon(self, lpIconName: IdIdi) -> WinResult<HICON>

LoadIcon method.

Examples

Loading a system icon:

use winsafe::{co, IdIdi, HINSTANCE};

let sys_icon = HINSTANCE::default()
    .LoadIcon(IdIdi::Idi(co::IDI::INFORMATION))
    .unwrap();

pub fn LoadImageBitmap(
self,
name: u16,
cx: i32,
cy: i32,
fuLoad: LR
) -> WinResult<HBITMAP>

LoadImage method for HBITMAP.

pub fn LoadImageCursor(
self,
name: u16,
cx: i32,
cy: i32,
fuLoad: LR
) -> WinResult<HCURSOR>

LoadImage method for HCURSOR.

pub fn LoadImageIcon(
self,
name: u16,
cx: i32,
cy: i32,
fuLoad: LR
) -> WinResult<HICON>

LoadImage method for HICON.

pub fn LoadLibrary(lpLibFileName: &str) -> WinResult<HINSTANCE>

LoadLibrary static method.

Note: Must be paired with an HINSTANCE::FreeLibrary call.

pub fn LoadMenu(self, lpMenuName: u16) -> WinResult<HMENU>

LoadMenu method.

pub fn LoadString(self, uID: u16) -> WinResult<String>

LoadString method.