Struct winsafe::HINSTANCE

#[repr(C)]
pub struct HINSTANCE { /* fields omitted */ }

Handle to an instance, same as HMODULE.

Implementations

impl HINSTANCE

pub const NULL: Self

The null, invalid handle.

This constant is common to all handle types.

pub unsafe fn from_ptr<T>(p: *mut T) -> HINSTANCE

Creates a new handle instance by wrapping a pointer.

This method is common to all handle types.

pub unsafe fn as_ptr(self) -> *mut c_void

Consumes the handle returning the underlying raw pointer.

This method is common to all handle types.

pub fn is_null(self) -> bool

Tells if the handle is invalid (null).

This method is common to all handle types.

pub fn as_opt(self) -> Option<HINSTANCE>

Consumes the handle into an option, which is None if the handle pointer is null.

This method is common to all handle types.

impl HINSTANCE

pub fn CreateDialogParam(
self,
resource_id: IdStr,
hwnd_parent: Option<HWND>,
dialog_proc: DLGPROC,
init_param: Option<isize>
) -> WinResult<HWND>

CreateDialogParam method.

pub fn DialogBoxParam(
self,
resource_id: IdStr,
hwnd_parent: Option<HWND>,
dialog_proc: DLGPROC,
init_param: Option<isize>
) -> WinResult<isize>

DialogBoxParam method.

pub fn EnumResourceLanguages<F>(
self,
resource_type: RtStr,
resource_id: IdStr,
func: F
) -> WinResult<()> where
F: Fn(LANGID) -> bool,

EnumResourceLanguages method.

pub fn EnumResourceNames<F>(
self,
resource_type: RtStr,
func: F
) -> WinResult<()> where
F: Fn(IdStr) -> bool,

EnumResourceNames method.

pub fn EnumResourceTypes<F>(self, func: F) -> WinResult<()> where
F: Fn(RtStr) -> bool,

EnumResourceTypes method.

pub fn FindResource(
self,
resource_id: IdStr,
resource_type: RtStr
) -> WinResult<HRSRC>

FindResource method.

For an example, see HINSTANCE::LockResource.

pub fn FindResourceEx(
self,
resource_id: IdStr,
resource_type: RtStr,
language: Option<LANGID>
) -> WinResult<HRSRC>

FindResourceEx method.

For an example, see HINSTANCE::LockResource.

pub fn FreeLibrary(self) -> WinResult<()>

FreeLibrary method.

pub fn GetClassInfoEx(
self,
class_name: &str,
wcx: &mut WNDCLASSEX<'_>
) -> WinResult<ATOM>

GetClassInfoEx method.

Examples

Retrieving information of a window class created in our application:

use winsafe::{HINSTANCE, WNDCLASSEX};

let mut wcx = WNDCLASSEX::default();
HINSTANCE::GetModuleHandle(None)?
    .GetClassInfoEx("SOME_CLASS_NAME", &mut wcx)?;

pub fn GetModuleFileName(self) -> WinResult<String>

GetModuleFileName method.

Examples

Retrieving the full path of currently running .exe file:

use winsafe::HINSTANCE;

let exe_name = HINSTANCE::NULL.GetModuleFileName()?;

println!("EXE: {}", exe_name);

pub fn GetModuleHandle(module_name: Option<&str>) -> WinResult<HINSTANCE>

GetModuleHandle static method.

Examples

Retrieving current module instance:

use winsafe::HINSTANCE;

let hinstance = HINSTANCE::GetModuleHandle(None)?;

pub fn GetProcAddress(self, proc_name: &str) -> WinResult<*const c_void>

GetProcAddress method.

pub fn LoadAccelerators(self, table_name: IdStr) -> WinResult<HACCEL>

LoadAccelerators method.

pub fn LoadCursor(self, resource_id: IdIdcStr) -> WinResult<HCURSOR>

LoadCursor method.

Examples

Loading a system cursor:

use winsafe::{co, HINSTANCE, IdIdc};

let sys_cursor = HINSTANCE::default()
    .LoadCursor(IdIdc::Idc(co::IDC::ARROW))?;

pub fn LoadIcon(self, icon_id: IdIdiStr) -> WinResult<HICON>

LoadIcon method.

Examples

Loading a system icon:

use winsafe::{co, IdIdi, HINSTANCE};

let sys_icon = HINSTANCE::default()
    .LoadIcon(IdIdi::Idi(co::IDI::INFORMATION))?;

pub fn LoadImageBitmap(
self,
name: u16,
sz: SIZE,
load: LR
) -> WinResult<HBITMAP>

LoadImage method for HBITMAP.

pub fn LoadImageCursor(
self,
name: u16,
cx: i32,
cy: i32,
load: LR
) -> WinResult<HCURSOR>

LoadImage method for HCURSOR.

pub fn LoadImageIcon(
self,
name: u16,
cx: i32,
cy: i32,
load: LR
) -> WinResult<HICON>

LoadImage method for HICON.

pub fn LoadLibrary(lib_file_name: &str) -> WinResult<HINSTANCE>

LoadLibrary static method.

Note: Must be paired with an HINSTANCE::FreeLibrary call.

pub fn LoadMenu(self, resource_id: IdStr) -> WinResult<HMENU>

LoadMenu method.

pub fn LoadResource(self, res_info: HRSRC) -> WinResult<HRSRCMEM>

LoadResource method.

For an example, see HINSTANCE::LockResource.

pub fn LoadString(self, id: u16) -> WinResult<String>

LoadString method.

pub fn LockResource<'a>(
self,
res_info: HRSRC,
hres_loaded: HRSRCMEM
) -> WinResult<&'a [u8]>

LockResource method.

This method should belong to HRSRCMEM, but in order to make it safe, we automatically call HINSTANCE::SizeofResource, so it’s implemented here.

Examples

The Updating Resources example:

use winsafe::{HINSTANCE, HUPDATERSRC, LANGID};
use winsafe::{co, IdStr, RtStr};

const IDD_HAND_ABOUTBOX: u16 = 103;
const IDD_FOOT_ABOUTBOX: u16 = 110;

let hExe = HINSTANCE::LoadLibrary("hand.exe")?;

let hRes = hExe.FindResource(
    IdStr::Id(IDD_HAND_ABOUTBOX),
    RtStr::Rt(co::RT::DIALOG),
)?;

let hResLoad = hExe.LoadResource(hRes)?;
let lpResLock = hExe.LockResource(hRes, hResLoad)?;
let hUpdateRes = HUPDATERSRC::BeginUpdateResource("foot.exe", false)?;

hUpdateRes.UpdateResource(
    RtStr::Rt(co::RT::DIALOG),
    IdStr::Id(IDD_FOOT_ABOUTBOX),
    LANGID::new(co::LANG::NEUTRAL, co::SUBLANG::NEUTRAL),
    lpResLock,
)?;

hUpdateRes.EndUpdateResource(false)?;

hExe.FreeLibrary()?;

pub fn SizeofResource(self, res_info: HRSRC) -> WinResult<u32>

SizeofResource method.

For an example, see HINSTANCE::LockResource.