Struct winsafe::HKEY

#[repr(transparent)]
pub struct HKEY(_);

Available on crate feature advapi only.

Handle to a registry key.

This handle also exposes several predefined registry keys, like HKEY::CURRENT_USER, which are always open and ready to be used. Usually, they are the starting point to open a registry key.