Struct winsafe::HMENU

#[repr(C)]pub struct HMENU { /* fields omitted */ }

Handle to a menu.

Implementations

impl HMENU

pub unsafe fn from_ptr<T>(p: *mut T) -> HMENU

Creates a new handle instance by wrapping a pointer.

This method is common to all handle types.

pub unsafe fn null_handle() -> Self

Creates a null, invalid handle.

This method is common to all handle types.

pub unsafe fn as_ptr(self) -> *mut c_void

Consumes the handle returning the underlying raw pointer.

This method is common to all handle types.

pub fn is_null(self) -> bool

Tells if the handle is invalid (null).

This method is common to all handle types.

pub fn as_opt(self) -> Option<HMENU>

Consumes the handle into an option, which is None if the handle pointer is null.

This method is common to all handle types.

impl HMENU

pub fn AppendMenu(
self,
uFlags: MF,
uIDNewItem: IdMenu,
lpNewItem: BitmapPtrStr
) -> WinResult<()>

AppendMenu method.

pub fn AppendMenuItem(self, command_id: u16, text: &str) -> WinResult<()>

A more convenient AppendMenu, which appends a new item at the menu with its command ID.

pub fn AppendMenuSeparator(self) -> WinResult<()>

A more convenient AppendMenu, which appends a separator.

pub fn AppendMenuSubmenu(self, submenu: HMENU, text: &str) -> WinResult<()>

A more convenient AppendMenu, which appends a menu as a new submenu entry.

pub fn CheckMenuItem(self, uIDCheckItem: IdPos, uCheck: bool) -> WinResult<MF>

CheckMenuItem method.

pub fn CreateMenu() -> WinResult<HMENU>

CreateMenu static method.

Note: If not attached to a window, must be paired with a DestroyMenu call.

pub fn CreatePopupMenu() -> WinResult<HMENU>

CreatePopupMenu static method.

Note: If not attached to a window, must be paired with a DestroyMenu call.

pub fn DeleteMenu(self, uPosition: IdPos) -> WinResult<()>

DeleteMenu method.

pub fn DestroyMenu(self) -> WinResult<()>

DestroyMenu method.

pub fn EnableMenuItem(self, uIDEnableItem: IdPos, uEnable: MF) -> WinResult<MF>

EnableMenuItem method.

You don't need to pass MF::BYCOMMAND or MF::BYPOSITION flags, they are inferred by IdPos.

pub fn GetMenuInfo(self, lpmi: &mut MENUINFO) -> WinResult<()>

GetMenuInfo method.

pub fn GetMenuItemCount(self) -> WinResult<u32>

GetMenuItemCount method.

pub fn GetMenuItemID(self, nPos: i32) -> Option<i32>

GetMenuItemID method.

pub fn GetSubMenu(self, nPos: u32) -> Option<HMENU>

GetSubMenu method.

pub fn InsertMenu(
self,
uPosition: IdPos,
uFlags: MF,
uIDNewItem: IdMenu,
lpNewItem: BitmapPtrStr
) -> WinResult<()>

InsertMenu method.

You don't need to pass MF::BYCOMMAND or MF::BYPOSITION flags, they are inferred by IdPos.

pub fn InsertMenuItem(self, item: IdPos, lpmi: &MENUITEMINFO) -> WinResult<()>

InsertMenuItem method.

pub fn IsMenu(self) -> bool

IsMenu method.

pub fn RemoveMenu(self, uPosition: IdPos) -> WinResult<()>

RemoveMenu method.

pub fn SetMenuInfo(self, mii: &MENUINFO) -> WinResult<()>

SetMenuInfo method.

pub fn SetMenuItemInfo(self, item: IdPos, lpmii: &MENUITEMINFO) -> WinResult<()>

SetMenuItemInfo method.

pub fn TrackPopupMenu(
self,
uFlags: TPM,
x: i32,
y: i32,
hWnd: HWND
) -> WinResult<Option<i32>>

TrackPopupMenu method