Struct winsafe::HMENU

#[repr(transparent)]
pub struct HMENU(_);

Handle to a menu.

Implementations

impl HMENU

pub fn AppendMenu(
self,
flags: MF,
new_item: IdMenu,
content: BmpPtrStr
) -> WinResult<()>

AppendMenu method.

This method is rather tricky, consider using HMENU::AppendMenuEnum.

pub fn AppendMenuEnum(self, items: &[MenuEnum<'_>]) -> WinResult<()>

A more convenient HMENU::AppendMenu.

Examples

Adding multiple entries at once, with their command IDs:

use winsafe::{HMENU, MenuEnum, seq_ids};

seq_ids! {
    ID_FILE_OPEN = 2001;
    ID_FILE_SAVE
    ID_FILE_EXIT
}

let hmenu: HMENU; // initialized somewhere

hmenu.AppendMenuEnum(&[
    MenuEnum::Entry(ID_FILE_OPEN, "&Open"),
    MenuEnum::Entry(ID_FILE_OPEN, "&Save"),
    MenuEnum::Separator,
    MenuEnum::Entry(ID_FILE_EXIT, "E&xit"),
])?;

pub fn CheckMenuItem(self, id_or_pos: IdPos, check: bool) -> WinResult<MF>

CheckMenuItem method.

pub fn CheckMenuRadioItem(
self,
first: IdPos,
last: IdPos,
check: IdPos
) -> WinResult<()>

CheckMenuRadioItem method.

Panics

Panics if first, last and check don’t use the same enum field.

pub fn CreateMenu() -> WinResult<HMENU>

CreateMenu static method.

Note: If not attached to a window, must be paired with an HMENU::DestroyMenu call.

pub fn CreatePopupMenu() -> WinResult<HMENU>

CreatePopupMenu static method.

Note: If not attached to a window, must be paired with an HMENU::DestroyMenu call.

pub fn DeleteMenu(self, id_or_pos: IdPos) -> WinResult<()>

DeleteMenu method.

pub fn DestroyMenu(self) -> WinResult<()>

DestroyMenu method.

pub fn EnableMenuItem(self, id_or_pos: IdPos, enable: bool) -> WinResult<MF>

EnableMenuItem method.

Examples

Disabling a menu item:

use winsafe::{HMENU, IdPos, seq_ids};

seq_ids! {
    ID_FILE_OPEN = 101;
}

let hmenu: HMENU; // initialized somewhere

hmenu.EnableMenuItem(
    IdPos::Id(ID_FILE_OPEN),
    false,
)?;

Disabling multiple menu items at once:

use winsafe::{HMENU, IdPos, seq_ids};

seq_ids! {
    ID_FILE_OPEN = 201;
    ID_FILE_SAVE
}

let hmenu: HMENU; // initialized somewhere

[
    (ID_FILE_OPEN, "Open\tCtrl+O"),
    (ID_FILE_SAVE, "&Save"),
].iter()
    .map(|(id, txt)| hmenu.EnableMenuItem(
        IdPos::Id(*id),
        false,
    ))
    .collect::<Result<Vec<_>, _>>()?;

pub fn GetMenuDefaultItem(self, by_pos: bool, flags: GMDI) -> WinResult<IdPos>

GetMenuDefaultItem method.

pub fn GetMenuInfo(self, mi: &mut MENUINFO) -> WinResult<()>

GetMenuInfo method.

pub fn GetMenuItemCount(self) -> WinResult<u32>

GetMenuItemCount method.

pub fn GetMenuItemInfo(
self,
id_or_pos: IdPos,
mii: &mut MENUITEMINFO
) -> WinResult<()>

GetMenuItemInfo method.

pub fn GetMenuItemID(self, nPos: i32) -> Option<i32>

GetMenuItemID method.

pub fn GetMenuState(self, id_or_pos: IdPos) -> WinResult<MF>

GetMenuState method.

pub fn GetMenuString(self, id_or_pos: IdPos) -> WinResult<String>

GetMenuString method.

pub fn GetSubMenu(self, pos: u32) -> Option<HMENU>

GetSubMenu method.

pub fn InsertMenuItem(
self,
id_or_pos: IdPos,
mii: &MENUITEMINFO
) -> WinResult<()>

InsertMenuItem method.

pub fn IsMenu(self) -> bool

IsMenu method.

pub fn RemoveMenu(self, id_or_pos: IdPos) -> WinResult<()>

RemoveMenu method.

pub fn SetMenuDefaultItem(self, id_or_pos: IdPos) -> WinResult<()>

SetMenuDefaultItem method.

pub fn SetMenuInfo(self, mi: &MENUINFO) -> WinResult<()>

SetMenuInfo method.

pub fn SetMenuItemBitmaps(
self,
id_or_pos: IdPos,
hbmp_unchecked: Option<HBITMAP>,
hbmp_checked: Option<HBITMAP>
) -> WinResult<()>

SetMenuItemBitmaps method.

pub fn SetMenuItemInfo(
self,
id_or_pos: IdPos,
mii: &MENUITEMINFO
) -> WinResult<()>

SetMenuItemInfo method.

pub fn TrackPopupMenu(
self,
flags: TPM,
location: POINT,
hwnd: HWND
) -> WinResult<Option<i32>>

TrackPopupMenu method.

Note: If you just want to display a popup menu, consider the simpler HMENU::TrackPopupMenuAtPoint.

pub fn TrackPopupMenuAtPoint(
self,
pos: POINT,
hwnd_parent: HWND,
hwnd_coords_relative_to: HWND
) -> WinResult<()>

Shows the popup menu anchored at the given coordinates using TrackPopupMenu, and performs other needed operations.

This method will block until the menu disappears.