Struct winsafe::HMENU

#[repr(C)]
pub struct HMENU { /* fields omitted */ }

Handle to a menu.

Implementations

impl HMENU

pub const NULL: Self

The null, invalid handle.

This constant is common to all handle types.

pub unsafe fn from_ptr<T>(p: *mut T) -> HMENU

Creates a new handle instance by wrapping a pointer.

This method is common to all handle types.

pub unsafe fn as_ptr(self) -> *mut c_void

Consumes the handle returning the underlying raw pointer.

This method is common to all handle types.

pub fn is_null(self) -> bool

Tells if the handle is invalid (null).

This method is common to all handle types.

pub fn as_opt(self) -> Option<HMENU>

Consumes the handle into an option, which is None if the handle pointer is null.

This method is common to all handle types.

impl HMENU

pub fn AppendMenu(
self,
uFlags: MF,
uIDNewItem: IdMenu,
lpNewItem: BitmapPtrStr
) -> WinResult<()>

AppendMenu method.

This method is rather tricky, consider using AppendMenuEnum.

pub fn AppendMenuEnum(self, item: &EntrySeparatorSubmenu<'_>) -> WinResult<()>

A more convenient AppendMenu.

Examples

Adding a new menu entry, with its command ID:

use winsafe::{EntrySeparatorSubmenu, HMENU};

let my_hmenu: HMENU; // initialized somewhere

const ID_FILE_OPEN: i32 = 2001;

my_hmenu.AppendMenuEnum(
   &EntrySeparatorSubmenu::Entry(ID_FILE_OPEN, "&Open"),
).unwrap();

Adding multiple entries at once:

use winsafe::{EntrySeparatorSubmenu, HMENU};

let my_hmenu: HMENU; // initialized somewhere

const ID_FILE_OPEN: i32 = 2001;
const ID_FILE_SAVE: i32 = 2002;
const ID_FILE_EXIT: i32 = 2003;

[
    EntrySeparatorSubmenu::Entry(ID_FILE_OPEN, "&Open"),
    EntrySeparatorSubmenu::Entry(ID_FILE_OPEN, "&Save"),
    EntrySeparatorSubmenu::Separator,
    EntrySeparatorSubmenu::Entry(ID_FILE_EXIT, "E&xit"),
].iter()
    .for_each(|e| file_menu.AppendMenuEnum(e).unwrap());

pub fn CheckMenuItem(self, uIDCheckItem: IdPos, uCheck: bool) -> WinResult<MF>

CheckMenuItem method.

pub fn CreateMenu() -> WinResult<HMENU>

CreateMenu static method.

Note: If not attached to a window, must be paired with an HMENU::DestroyMenu call.

pub fn CreatePopupMenu() -> WinResult<HMENU>

CreatePopupMenu static method.

Note: If not attached to a window, must be paired with an HMENU::DestroyMenu call.

pub fn DeleteMenu(self, uPosition: IdPos) -> WinResult<()>

DeleteMenu method.

pub fn DestroyMenu(self) -> WinResult<()>

DestroyMenu method.

pub fn EnableMenuItem(
self,
uIDEnableItem: IdPos,
uEnable: bool
) -> WinResult<MF>

EnableMenuItem method.

Examples

Disabling a menu item:

use winsafe::{HMENU, IdPos};

const ID_FILE_OPEN: i32 = 101;

let hmenu: HMENU; // initialized somewhere

hmenu.EnableMenuItem(
    IdPos::Id(ID_FILE_OPEN),
    false,
).unwrap();

Disabling multiple menu items at once:

use winsafe::{HMENU, IdPos};

const ID_FILE_OPEN: i32 = 201;
const ID_FILE_SAVE: i32 = 202;

let hmenu: HMENU; // initialized somewhere

[
    (ID_FILE_OPEN, "Open\tCtrl+O"),
    (ID_FILE_SAVE, "&Save"),
].iter()
    .for_each(|(id, txt)| hmenu.EnableMenuItem(
        IdPos::Id(*id),
        false,
    ).unwrap());

pub fn GetMenuInfo(self, lpmi: &mut MENUINFO) -> WinResult<()>

GetMenuInfo method.

pub fn GetMenuItemCount(self) -> WinResult<u32>

GetMenuItemCount method.

pub fn GetMenuItemID(self, nPos: i32) -> Option<i32>

GetMenuItemID method.

pub fn GetSubMenu(self, nPos: u32) -> Option<HMENU>

GetSubMenu method.

pub fn InsertMenuItem(self, item: IdPos, lpmi: &MENUITEMINFO) -> WinResult<()>

InsertMenuItem method.

pub fn IsMenu(self) -> bool

IsMenu method.

pub fn RemoveMenu(self, uPosition: IdPos) -> WinResult<()>

RemoveMenu method.

pub fn SetMenuDefaultItem(self, uItem: IdPos) -> WinResult<()>

SetMenuDefaultItem method.

pub fn SetMenuInfo(self, mii: &MENUINFO) -> WinResult<()>

SetMenuInfo method.

pub fn SetMenuItemInfo(self, item: IdPos, lpmii: &MENUITEMINFO) -> WinResult<()>

SetMenuItemInfo method.

pub fn TrackPopupMenu(
self,
uFlags: TPM,
x: i32,
y: i32,
hWnd: HWND
) -> WinResult<Option<i32>>

TrackPopupMenu method.

Note: If you just want to display a popup menu, consider the simpler TrackPopupMenuAtPoint.

pub fn TrackPopupMenuAtPoint(
self,
pos: POINT,
hParent: HWND,
hCoordsRelativeTo: HWND
) -> WinResult<()>

Shows the popup menu anchored at the given coordinates using TrackPopupMenu, and performs other needed operations.

This method will block until the menu disappears.