Struct winsafe::HPROCESS

#[repr(C)]
pub struct HPROCESS { /* fields omitted */ }

Handle to a process. Originally just a HANDLE.

Implementations

impl HPROCESS

pub const NULL: Self

The null, invalid handle.

This constant is common to all handle types.

pub unsafe fn from_ptr<T>(p: *mut T) -> HPROCESS

Creates a new handle instance by wrapping a pointer.

This method is common to all handle types.

pub unsafe fn as_ptr(self) -> *mut c_void

Consumes the handle returning the underlying raw pointer.

This method is common to all handle types.

pub fn is_null(self) -> bool

Tells if the handle is invalid (null).

This method is common to all handle types.

pub fn as_opt(self) -> Option<HPROCESS>

Consumes the handle into an option, which is None if the handle pointer is null.

This method is common to all handle types.

impl HPROCESS

pub fn CloseHandle(self) -> WinResult<()>

CloseHandle method.

impl HPROCESS

pub fn CreateProcess(
lpApplicationName: Option<&str>,
lpCommandLine: Option<&str>,
lpProcessAttributes: Option<&mut SECURITY_ATTRIBUTES<'_>>,
lpThreadAttributes: Option<&mut SECURITY_ATTRIBUTES<'_>>,
nInheritHandles: bool,
dwCreationFlags: CREATE,
lpEnvironment: Option<Vec<String>>,
lpCurrentDirectory: Option<&str>,
lpStartupInfo: &mut STARTUPINFO<'_, '_>
) -> WinResult<PROCESS_INFORMATION>

CreateProcess static method.

Note: Process and thread handles are returned in the PROCESS_INFORMATION struct, and they must be paired with their respective HPROCESS::CloseHandle and HTHREAD::CloseHandle calls.

pub fn ExitProcess(dwExitCode: u32)

ExitProcess static method.

pub fn FlushInstructionCache(
self,
lpBaseAddress: *mut c_void,
dwSize: u64
) -> WinResult<()>

FlushInstructionCache method.

pub fn FlushProcessWriteBuffers()

FlushProcessWriteBuffers static method.

pub fn GetCurrentProcess() -> HPROCESS

GetCurrentProcess static method.

pub fn GetExitCodeProcess(self) -> WinResult<u32>

GetExitCodeProcess method.

pub fn GetGuiResources(self, uiFlags: GR) -> WinResult<u32>

GetGuiResources method.

pub fn GetProcessId(self) -> WinResult<u32>

GetProcessId method.

pub fn GetProcessTimes(
self,
lpCreationTime: &mut FILETIME,
lpExitTime: &mut FILETIME,
lpKernelTime: &mut FILETIME,
lpUserTime: &mut FILETIME
) -> WinResult<()>

GetProcessTimes method.

pub fn IsWow64Process(self) -> WinResult<bool>

IsWow64Process method.

pub fn WaitForSingleObject(self, dwMilliseconds: Option<u32>) -> WinResult<WAIT>

WaitForSingleObject method.