Struct winsafe::HWND

#[repr(C)]
pub struct HWND { /* fields omitted */ }

Handle to a window.

Implementations

impl HWND

pub const NULL: Self

The null, invalid handle.

This constant is common to all handle types.

pub unsafe fn from_ptr<T>(p: *mut T) -> HWND

Creates a new handle instance by wrapping a pointer.

This method is common to all handle types.

pub unsafe fn as_ptr(self) -> *mut c_void

Consumes the handle returning the underlying raw pointer.

This method is common to all handle types.

pub fn is_null(self) -> bool

Tells if the handle is invalid (null).

This method is common to all handle types.

pub fn as_opt(self) -> Option<HWND>

Consumes the handle into an option, which is None if the handle pointer is null.

This method is common to all handle types.

impl HWND

pub const BROADCAST: Self

Represents all top-level windows in HWND::PostMessage and HWND::SendMessage.

pub fn hinstance(self) -> HINSTANCE

GetWindowLongPtr wrapper to retrieve the window HINSTANCE.

pub fn ArrangeIconicWindows(self) -> WinResult<u32>

ArrangeIconicWindows method.

pub fn BeginPaint(self, ps: &mut PAINTSTRUCT) -> WinResult<HDC>

BeginPaint method.

Note: Must be paired with an HWND::EndPaint call.

Examples

BeginPaint is usually called inside a WM_PAINT event to paint the window client area:

use winsafe::{gui, PAINTSTRUCT};

let my_main: gui::WindowMain; // initialized somewhere

my_main.on().wm_paint({
    let my_main = my_main.clone();
    move || {
        let mut ps = PAINTSTRUCT::default();
        let hdc = my_main.hwnd().BeginPaint(&mut ps)?;

        // paint the HDC...

        my_main.hwnd().EndPaint(&ps);
        Ok(())
    }
});

pub fn BringWindowToTop(self) -> WinResult<()>

BringWindowToTop method.

pub fn ChildWindowFromPoint(self, pt: POINT) -> Option<HWND>

ChildWindowFromPoint method.

pub fn ClientToScreen(self, pt: &mut POINT) -> WinResult<()>

ClientToScreen method.

pub fn ClientToScreenRc(self, rc: &mut RECT) -> WinResult<()>

ClientToScreen method for a RECT.

pub fn CloseWindow(self) -> WinResult<()>

CloseWindow method.

Note that this method will actually minimize the window, not destroy it.

pub fn CreateWindowEx(
ex_style: WS_EX,
class_name: AtomStr,
title: Option<&str>,
style: WS,
pos: POINT,
size: SIZE,
hwnd_parent: Option<HWND>,
hmenu: IdMenu,
hinstance: HINSTANCE,
lparam: Option<isize>
) -> WinResult<HWND>

CreateWindowEx static method.

pub fn DefSubclassProc<M: MsgSend>(self, msg: M) -> M::RetType

DefSubclassProc method.

The return type is variable, being defined by the RetType associated type of the MsgSend trait. That means each message can define its own return type.

pub fn DefWindowProc<M: MsgSend>(self, msg: M) -> M::RetType

DefWindowProc method.

The return type is variable, being defined by the RetType associated type of the MsgSend trait. That means each message can define its own return type.

pub fn DestroyWindow(self) -> WinResult<()>

DestroyWindow method.

pub fn DrawMenuBar(self) -> WinResult<()>

DrawMenuBar method.

pub fn EnableWindow(self, enable: bool) -> bool

EnableWindow method.

pub fn EndDialog(self, result: isize) -> WinResult<()>

EndDialog method.

pub fn EndPaint(self, ps: &PAINTSTRUCT)

EndPaint method.

pub fn EnumChildWindows<F>(self, func: F) where
F: Fn(HWND) -> bool,

EnumChildWindows method.

Examples

use winsafe::HWND;

let my_hwnd: HWND; // initialized somewhere

my_hwnd.EnumChildWindows(|hchild: HWND| -> bool {
    println!("Child HWND: {}", hchild);
    true
});

pub fn FindWindow(
class_name: Option<AtomStr>,
title: Option<&str>
) -> WinResult<HWND>

FindWindow static method.

pub fn FindWindowEx(
self,
hwnd_child_after: Option<HWND>,
class_name: AtomStr,
title: Option<&str>
) -> WinResult<HWND>

FindWindowEx method.

pub fn GetActiveWindow() -> Option<HWND>

GetActiveWindow static method.

pub fn GetAltTabInfo(
self,
item: Option<u32>,
ati: &mut ALTTABINFO,
sz_item_text: Option<u32>
) -> WinResult<String>

GetAltTabInfo method.

If item is None, the item text is not retrieved.

The sz_item_text is the maximum number of expected chars for the item text. If None, defaults to 100.

pub fn GetAncestor(self, flags: GA) -> Option<HWND>

GetAncestor method.

pub fn GetCapture() -> Option<HWND>

GetCapture static method.

pub fn GetClassLongPtr(self, index: GCLP) -> usize

GetClassLongPtr method.

pub fn GetClassName(self) -> WinResult<String>

GetClassName method.

pub fn GetClientRect(self) -> WinResult<RECT>

GetClientRect method.

pub fn GetDC(self) -> WinResult<HDC>

GetDC method.

Note: Must be paired with an HWND::ReleaseDC call.

pub fn GetDesktopWindow() -> HWND

GetDesktopWindow static method.

pub fn GetDlgCtrlID(self) -> WinResult<u16>

GetDlgCtrlID method.

pub fn GetDlgItem(self, ctrl_id: u16) -> WinResult<HWND>

GetDlgItem method.

pub fn GetFocus() -> Option<HWND>

GetFocus static method.

pub fn GetForegroundWindow() -> Option<HWND>

GetForegroundWindow static method.

pub fn GetLastActivePopup(self) -> Option<HWND>

GetLastActivePopup method.

pub fn GetMenu(self) -> Option<HMENU>

GetMenu method.

pub fn GetMenuBarInfo(
self,
obj_id: OBJID,
item_id: u32,
mbi: &mut MENUBARINFO
) -> WinResult<()>

GetMenuBarInfo method.

pub fn GetMenuItemRect(
self,
hmenu: HMENU,
item_pos: u32,
rc_item: &mut RECT
) -> WinResult<()>

GetMenuItemRect method.

pub fn GetNextDlgGroupItem(
self,
hwnd_ctrl: HWND,
previous: bool
) -> WinResult<HWND>

GetNextDlgGroupItem method.

pub fn GetNextDlgTabItem(
self,
hwnd_ctrl: HWND,
previous: bool
) -> WinResult<HWND>

GetNextDlgTabItem method.

pub fn GetParent(self) -> WinResult<HWND>

GetParent method.

pub fn GetScrollInfo(self, bar: SBB, si: &mut SCROLLINFO) -> WinResult<()>

GetScrollInfo method.

pub fn GetScrollPos(self, bar: SBB) -> WinResult<i32>

GetScrollPos method.

pub fn GetShellWindow() -> Option<HWND>

GetShellWindow static method.

pub fn GetSystemMenu(self, revert: bool) -> Option<HMENU>

GetSystemMenu method.

pub fn GetTopWindow(self) -> WinResult<Option<HWND>>

GetTopWindow method.

pub fn GetUpdateRgn(self, hrgn: HRGN, erase: bool) -> WinResult<REGION>

GetUpdateRgn method.

pub fn GetWindow(self, cmd: GW) -> WinResult<HWND>

GetWindow method.

pub fn GetWindowDC(self) -> WinResult<HDC>

GetWindowDC method.

Note: Must be paired with an HWND::ReleaseDC call.

pub fn GetWindowDisplayAffinity(self) -> WinResult<WDA>

GetWindowDisplayAffinity method.

pub fn GetWindowInfo(self, wi: &mut WINDOWINFO) -> WinResult<()>

GetWindowInfo method.

pub fn GetWindowLongPtr(self, index: GWLP) -> isize

GetWindowLongPtr method.

pub fn GetWindowPlacement(self, wp: &mut WINDOWPLACEMENT) -> WinResult<()>

GetWindowPlacement method.

pub fn GetWindowRect(self) -> WinResult<RECT>

GetWindowRect method.

pub fn GetWindowRgn(self, hrgn: HRGN) -> WinResult<REGION>

GetWindowRgn method.

pub fn GetWindowRgnBox(self, rc: &mut RECT) -> WinResult<REGION>

GetWindowRgnBox method.

pub fn GetWindowText(self) -> WinResult<String>

GetWindowText method.

use winsafe::HWND;

let my_hwnd: HWND; // initialized somewhere

let text = my_hwnd.GetWindowText()?;
println!("Text: {}", text);

pub fn GetWindowTextLength(self) -> WinResult<i32>

GetWindowTextLength method. Does not count the terminating null.

pub fn GetWindowThreadProcessId(self) -> u32

GetWindowThreadProcessId method.

pub fn HiliteMenuItem(self, hmenu: HMENU, id_or_pos: IdPos, hilite: MF) -> bool

HiliteMenuItem method.

pub fn InvalidateRect(self, rc: Option<&RECT>, erase: bool) -> WinResult<()>

InvalidateRect method.

Examples

Most of the time you’ll just want update the entire client area:

use winsafe::HWND;

let my_hwnd: HWND; // initialized somewhere

my_hwnd.InvalidateRect(None, true)?;

pub fn InvalidateRgn(self, hrgn: HRGN, erase: bool)

InvalidateRgn method.

pub fn IsChild(self, hwnd_possible_child: HWND) -> bool

IsChild method.

pub fn IsDialogMessage(self, msg: &mut MSG) -> bool

IsDialogMessage method.

pub fn IsIconic(self) -> bool

IsIconic method.

pub fn IsWindow(self) -> bool

IsWindow method.

pub fn IsWindowEnabled(self) -> bool

IsWindowEnabled method.

pub fn IsWindowUnicode(self) -> bool

IsWindowUnicode method.

pub fn IsWindowVisible(self) -> bool

IsWindowVisible method.

pub fn IsZoomed(self) -> bool

IsZoomed method.

pub fn KillTimer(self, event_id: usize) -> WinResult<()>

KillTimer method.

pub fn LogicalToPhysicalPoint(self, pt: *mut POINT) -> WinResult<()>

LogicalToPhysicalPoint method.

pub fn MapDialogRect(self, rc: &mut RECT) -> WinResult<()>

MapDialogRect method.

pub fn MessageBox(
self,
text: &str,
caption: &str,
flags: MB
) -> WinResult<DLGID>

MessageBox method.

Consider using the more modern TaskDialog method.

Examples

A modal message box, which blocks its parent:

use winsafe::{co, HWND};

let my_hwnd: HWND; // initialized somewhere

my_hwnd.MessageBox("Hello, world", "title",
    co::MB::OKCANCEL | co::MB::ICONINFORMATION)?

Usually the message box has a valid parent window, however, if for some reason you don’t have a window to serve as parent, you still can show a non-modal, parent-less message box by retrieving the desktop handle:

HWND::GetDesktopWindow()
    .MessageBox("Hello, world", "Title", co::MB::ICONEXCLAMATION)?;

pub fn MonitorFromWindow(self, flags: MONITOR) -> HMONITOR

MonitorFromWindow method.

pub fn MoveWindow(self, pos: POINT, size: SIZE, repaint: bool) -> WinResult<()>

MoveWindow method.

pub fn OpenClipboard(self) -> WinResult<()>

OpenClipboard method.

Note: Must be paired with a CloseClipboard call.

pub fn OpenThemeData(self, class_list: &str) -> Option<HTHEME>

OpenThemeData method.

Note: Must be paired with an HTHEME::CloseThemeData call.

pub fn PostMessage<M: MsgSend>(self, msg: M) -> WinResult<()>

PostMessage method. Note that this method is asychronous.

Examples

Programatically closing a window:

use winsafe::{HWND, msg::wm};

let my_hwnd: HWND; // initialized somewhere

my_hwnd.PostMessage(wm::Close {})?;

Sending a message to all top-level windows:

use winsafe::{HWND, msg::wm};

HWND::BROADCAST.PostMessage(
    wm::ExitMenuLoop { is_shortcut: false },
)?;

pub fn RealChildWindowFromPoint(
self,
pt_parent_client_coords: POINT
) -> Option<HWND>

RealChildWindowFromPoint method.

pub fn RealGetWindowClass(self) -> WinResult<String>

RealGetWindowClassW method.

pub fn RedrawWindow(
self,
rc_update: &RECT,
hrgn_update: HRGN,
flags: RDW
) -> WinResult<()>

RedrawWindow method.

pub fn ReleaseDC(self, hdc: HDC) -> WinResult<()>

ReleaseDC method.

pub fn RemoveWindowSubclass(
self,
subclass_func: SUBCLASSPROC,
subclass_id: usize
) -> WinResult<()>

RemoveWindowSubclass method.

pub fn ScreenToClient(self, pt: &mut POINT) -> WinResult<()>

ScreenToClient method.

pub fn ScreenToClientRc(self, rc: &mut RECT) -> WinResult<()>

ScreenToClient method for a RECT.

pub fn SendMessage<M: MsgSend>(self, msg: M) -> M::RetType

SendMessage method.

The return type is variable, being defined by the RetType associated type of the MsgSend trait. That means each message can define its own return type.

Examples

Sending a lvm::SetItem list view message, which demands a reference to an LVITEM object:

use winsafe::{co, HWND, LVITEM, msg::lvm};

let my_hwnd: HWND; // initialized somewhere

let mut lvi = LVITEM::default(); // object to be sent
lvi.mask = co::LVIF::IMAGE;
lvi.iImage = 3;

my_hwnd.SendMessage(lvm::SetItem {
    lvitem: &lvi,
});

pub fn SetCapture(&self) -> Option<HWND>

SetCapture method.

Note: Must be paired with a ReleaseCapture call.

pub fn SetFocus(self) -> Option<HWND>

SetFocus method.

pub fn SetForegroundWindow(self) -> bool

SetForegroundWindow method.

pub fn SetMenu(self, hmenu: HMENU) -> WinResult<()>

SetMenu method.

pub fn SetParent(self, hwnd_new_parent: HWND) -> WinResult<Option<HWND>>

SetParent method.

pub fn SetScrollInfo(self, bar: SBB, si: &SCROLLINFO, redraw: bool) -> i32

SetScrollInfo method.

pub fn SetScrollPos(self, b: SBB, pos: i32, redraw: bool) -> WinResult<i32>

SetScrollPos method.

pub fn SetScrollRange(
self,
bar: SBB,
min_pos: i32,
max_pos: i32,
redraw: bool
) -> WinResult<()>

SetScrollRange method.

pub fn SetTimer(
self,
event_id: usize,
elapse_ms: u32,
timer_func: Option<TIMERPROC>
) -> WinResult<usize>

SetTimer method.

pub fn SetWindowDisplayAffinity(self, affinity: WDA) -> WinResult<()>

SetWindowDisplayAffinity method.

pub fn SetWindowLongPtr(self, index: GWLP, new_long: isize) -> isize

SetWindowLongPtr method.

pub fn SetWindowPlacement(self, wp: &WINDOWPLACEMENT) -> WinResult<()>

SetWindowPlacement method.

pub fn SetWindowPos(
self,
hwnd_insert_after: HwndPlace,
pos: POINT,
size: SIZE,
flags: SWP
) -> WinResult<()>

SetWindowPos method.

pub fn SetWindowRgn(self, hrgn: HRGN, redraw: bool) -> WinResult<()>

SetWindowRgn method.

pub fn SetWindowSubclass(
self,
subclass_proc: SUBCLASSPROC,
subclass_id: usize,
ref_data: usize
) -> WinResult<()>

SetWindowSubclass method.

pub fn SetWindowText(self, text: &str) -> WinResult<()>

SetWindowText method.

pub fn ShellExecute(
self,
operation: &str,
file: &str,
parameters: Option<&str>,
directory: Option<&str>,
show_cmd: SW
) -> Result<HINSTANCE, SE_ERR>

ShellExecute method.

pub fn ShowCaret(self) -> WinResult<()>

ShowCaret method.

pub fn ShowWindow(self, show_cmd: SW) -> bool

ShowWindow method.

pub fn TaskDialog(
self,
hinstance: Option<HINSTANCE>,
window_title: Option<&str>,
main_instruction: Option<&str>,
content: Option<&str>,
common_buttons: TDCBF,
icon: IdTdiconStr
) -> WinResult<DLGID>

TaskDialog method.

If you need more customization, consider the TaskDialogIndirect function.

Examples

An information message with just an OK button:

use winsafe::{co, HWND, IdTdicon};

let my_hwnd: HWND; // initialized somewhere

my_hwnd.TaskDialog(
    None,
    Some("My app name"),
    Some("Operation successful"),
    Some("The operation completed successfully."),
    co::TDCBF::OK
    IdTdicon::Tdicon(co::TD_ICON::INFORMATION),
)?;

Prompt the user to click OK or Cancel upon a question:

use winsafe::{co, HWND, IdTdicon};

let my_hwnd: HWND; // initialized somewhere

let answer = my_hwnd.TaskDialog(
    None,
    Some("My app name"),
    Some("File modified"),
    Some("The file has been modified.\nProceed closing the application?"),
    co::TDCBF::OK | co::TDCBF::CANCEL,
    IdTdicon::Tdicon(co::TD_ICON::WARNING),
)?;

if answer == co::DLGID::OK {
    println!("User clicked OK.");
}

pub fn TranslateAccelerator(
self,
haccel_table: HACCEL,
msg: &mut MSG
) -> WinResult<()>

TranslateAccelerator method.

pub fn UpdateWindow(self) -> WinResult<()>

UpdateWindow method.

pub fn ValidateRect(self, rc: &RECT) -> WinResult<()>

ValidateRect method.

pub fn ValidateRgn(self, hrgn: HRGN) -> WinResult<()>

ValidateRgn method.

pub fn WindowFromPhysicalPoint(pt: POINT) -> Option<HWND>

WindowFromPhysicalPoint static method.

pub fn WindowFromPoint(pt: POINT) -> Option<HWND>

WindowFromPoint static method.

pub fn WinHelp(self, help_file: &str, cmd: HELPW, data: usize) -> WinResult<()>

WinHelp method.