Struct winsafe::HWND

#[repr(C)]pub struct HWND { /* fields omitted */ }

Handle to a window. Exposes methods.

Implementations

impl HWND

pub unsafe fn from_ptr<T>(p: *mut T) -> HWND

Creates a new handle instance by wrapping a pointer.

pub unsafe fn null_handle() -> Self

Creates a null, invalid handle.

pub unsafe fn as_ptr(self) -> *mut c_void

Consumes the handle returning the underlying raw pointer.

pub fn is_null(self) -> bool

Tells if the handle is invalid (null).

pub fn as_opt(self) -> Option<HWND>

Consumes the handle into an option, which is None if the handle pointer is null.

impl HWND

pub fn hinstance(self) -> HINSTANCE

GetWindowLongPtr wrapper to retrieve the window HINSTANCE.

pub fn ArrangeIconicWindows(self) -> WinResult<u32>

ArrangeIconicWindows method.

pub fn BeginPaint(self, lpPaint: &mut PAINTSTRUCT) -> WinResult<HDC>

BeginPaint method.

Note: Must be paired with an EndPaint call.

pub fn BringWindowToTop(self) -> WinResult<()>

BringWindowToTop method.

pub fn ChildWindowFromPoint(self, point: POINT) -> Option<HWND>

ChildWindowFromPoint method.

pub fn ClientToScreen(self, lpPoint: &mut POINT) -> WinResult<()>

ClientToScreen method.

pub fn ClientToScreenRc(self, lpRect: &mut RECT) -> WinResult<()>

ClientToScreen method for a RECT.

pub fn CloseWindow(self) -> WinResult<()>

CloseWindow method.

Note that this method will actually minimize the window, not destroy it.

pub fn CreateWindowEx(
dwExStyle: WS_EX,
lpClassName: AtomStr,
lpWindowName: Option<&str>,
dwStyle: WS,
X: i32,
Y: i32,
nWidth: i32,
nHeight: i32,
hWndParent: Option<HWND>,
hMenu: IdMenu,
hInstance: HINSTANCE,
lpParam: Option<isize>
) -> WinResult<HWND>

CreateWindowEx static method.

pub fn DefSubclassProc<M: Message>(self, uMsg: M) -> M::RetType

DefSubclassProc method.

The return type is variable, being defined by the RetType associated type of the Message trait. That means each message can define its own return type.

pub fn DefWindowProc<M: Message>(self, uMsg: M) -> M::RetType

DefWindowProc method.

The return type is variable, being defined by the RetType associated type of the Message trait. That means each message can define its own return type.

pub fn DestroyWindow(self)

DestroyWindow method.

pub fn EnableWindow(self, bEnable: bool) -> bool

EnableWindow method.

pub fn EndDialog(self, nResult: isize) -> WinResult<()>

EndDialog method.

pub fn EndPaint(self, lpPaint: &PAINTSTRUCT)

EndPaint method.

pub fn EnumChildWindows(self, lpEnumFunc: WNDENUMPROC, lParam: isize)

EnumChildWindows method.

This method can be more performant than EnumChildWindowsVec, which passes through all children and allocates a Vec. However, it has the inconvenient of the manual function pointer.

pub fn EnumChildWindowsVec(self) -> Vec<HWND>

A more convenient EnumChildWindows, which returns a Vec with the handles of all child windows, instead of taking a function pointer.

Examples

use winsafe::HWND;

let my_hwnd: HWND; // initialize it somewhere...

for hchild in my_hwnd.EnumChildWindowsVec() {
    println!("HWND: {}", hchild);
}

pub fn FindWindow(lpClassName: &str, lpWindowName: &str) -> WinResult<HWND>

FindWindow static method.

pub fn GetActiveWindow() -> Option<HWND>

GetActiveWindow static method.

pub fn GetAncestor(self, gaFlags: GA) -> Option<HWND>

GetAncestor method.

pub fn GetCapture() -> Option<HWND>

GetCapture static method.

pub fn GetClassLongPtr(self, nIndex: GCLP) -> usize

GetClassLongPtr method.

pub fn GetClientRect(self) -> WinResult<RECT>

GetClientRect method.

pub fn GetDC(self) -> WinResult<HDC>

GetDC method.

Note: Must be paired with a ReleaseDC call.

pub fn GetDesktopWindow() -> HWND

GetDesktopWindow static method.

pub fn GetDlgCtrlID(self) -> WinResult<i32>

GetDlgCtrlID method.

pub fn GetDlgItem(self, nIDDlgItem: i32) -> WinResult<HWND>

GetDlgItem method.

pub fn GetFocus() -> Option<HWND>

GetFocus static method.

pub fn GetForegroundWindow() -> Option<HWND>

GetForegroundWindow static method.

pub fn GetNextDlgGroupItem(self, hCtl: HWND, bPrevious: bool) -> WinResult<HWND>

GetNextDlgGroupItem method.

pub fn GetNextDlgTabItem(self, hCtl: HWND, bPrevious: bool) -> WinResult<HWND>

GetNextDlgTabItem method.

pub fn GetParent(self) -> WinResult<HWND>

GetParent method.

pub fn GetUpdateRgn(self, hRgn: HRGN, bErase: bool) -> WinResult<REGION>

GetUpdateRgn method.

pub fn GetWindow(self, uCmd: GW) -> WinResult<HWND>

GetWindow method.

pub fn GetWindowDC(self) -> WinResult<HDC>

GetWindowDC method.

Note: Must be paired with a ReleaseDC call.

pub fn GetWindowDisplayAffinity(self) -> WinResult<WDA>

GetWindowDisplayAffinity method.

pub fn GetWindowInfo(self, pwi: &mut WINDOWINFO) -> WinResult<()>

GetWindowInfo method.

pub fn GetWindowLongPtr(self, nIndex: GWLP) -> isize

GetWindowLongPtr method.

pub fn GetWindowPlacement(self, lpwndpl: &mut WINDOWPLACEMENT) -> WinResult<()>

GetWindowPlacement method.

pub fn GetWindowRect(self) -> WinResult<RECT>

GetWindowRect method.

pub fn GetWindowRgn(self, hRgn: HRGN) -> WinResult<REGION>

GetWindowRgn method.

pub fn GetWindowRgnBox(self, lprc: &mut RECT) -> WinResult<REGION>

GetWindowRgnBox method.

pub fn GetWindowText(self, buf: &mut WString) -> WinResult<i32>

GetWindowText method.

The passed buffer will be automatically allocated with GetWindowTextLength.

This method can be more performant than GetWindowTextStr because the buffer can be reused, avoiding multiple allocations. However, it has the inconvenient of the manual conversion from WString to String.

Examples

use winsafe::HWND;

let my_hwnd: HWND; // initialize it somewhere...

let mut buf = WString::new();
my_hwnd.GetWindowText(&mut buf).unwrap();
println!("Text: {}", buf.to_string());

pub fn GetWindowTextLength(self) -> WinResult<i32>

GetWindowTextLength method. Does not count the terminating null.

pub fn GetWindowTextStr(self) -> WinResult<String>

A more convenient GetWindowText, which returns a String instead of requiring an external buffer.

Examples

use winsafe::HWND;

let my_hwnd: HWND; // initialize it somewhere...

let text = my_hwnd.GetWindowTextStr().unwrap();
println!("Text: {}", text);

pub fn HiliteMenuItem(
self,
hMenu: HMENU,
uIDHiliteItem: IdPos,
uHilite: MF
) -> bool

HiliteMenuItem method.

pub fn InvalidateRect(
self,
lpRect: Option<&RECT>,
bErase: bool
) -> WinResult<()>

InvalidateRect method.

Examples

Most of the time you'll just want update the entire client area:

use winsafe::HWND;

let my_hwnd: HWND; // initialize it somewhere...

my_hwnd.InvalidateRect(None, true)
    .unwrap();

pub fn InvalidateRgn(self, hRgn: HRGN, bErase: bool)

InvalidateRgn method.

pub fn IsChild(self, hWndPossibleChild: HWND) -> bool

IsChild method.

pub fn IsDialogMessage(self, lpMsg: &mut MSG) -> bool

IsDialogMessage method.

pub fn IsIconic(self) -> bool

IsIconic method.

pub fn IsWindow(self) -> bool

IsWindow method.

pub fn IsWindowEnabled(self) -> bool

IsWindowEnabled method.

pub fn IsWindowVisible(self) -> bool

IsWindowVisible method.

pub fn IsZoomed(self) -> bool

IsZoomed method.

pub fn MapDialogRect(self, lpRect: &mut RECT) -> WinResult<()>

MapDialogRect method.

pub fn MessageBox(
self,
lpText: &str,
lpCaption: &str,
uType: MB
) -> WinResult<DLGID>

MessageBox method.

Examples

A modal message box, which blocks its parent:

use winsafe::HWND;

let my_hwnd: HWND; // initialize it somewhere...

my_hwnd.MessageBox("Hello, world", "title", co::MB::OKCANCEL | co::MB::ICONINFORMATION)
    .unwrap();

Usually the message box has a valid parent window, however, if for some reason you don't have a window to serve as parent, you still can show a non-modal, parent-less message box by retrieving the desktop handle:

HWND::GetDesktopWindow()
    .MessageBox("Hello, world", "Title", co::MB::ICONEXCLAMATION)
    .unwrap();

pub fn OpenThemeData(self, pszClassList: &str) -> Option<HTHEME>

OpenThemeData method.

Note: Must be paired with a CloseThemeData call.

pub fn PostMessage<M: Message>(self, uMsg: M) -> WinResult<()>

PostMessage method. Note that this method is asychronous.

Note: To use HWND_BROADCAST or NULL as the first argument, see the PostMessage free function.

Examples

Programatically closing a window:

use winsafe::{HWND, msg::WmClose};

let my_hwnd: HWND; // initialize it somewhere...

my_hwnd.PostMessage(WmClose {}).unwrap();

pub fn RealChildWindowFromPoint(
self,
ptParentClientCoords: POINT
) -> Option<HWND>

RealChildWindowFromPoint method.

pub fn RedrawWindow(
self,
lprcUpdate: &RECT,
hrgnUpdate: HRGN,
flags: RDW
) -> WinResult<()>

RedrawWindow method.

pub fn ReleaseDC(self, hDC: HDC) -> WinResult<()>

ReleaseDC method.

pub fn RemoveWindowSubclass(
self,
pfnSubclass: SUBCLASSPROC,
uIdSubclass: usize
) -> WinResult<()>

RemoveWindowSubclass method.

pub fn ScreenToClient(self, lpPoint: &mut POINT) -> WinResult<()>

ScreenToClient method;

pub fn ScreenToClientRc(self, lpRect: &mut RECT) -> WinResult<()>

ScreenToClient method for a RECT.

pub fn SendMessage<M: Message>(self, uMsg: M) -> M::RetType

SendMessage method.

The return type is variable, being defined by the RetType associated type of the Message trait. That means each message can define its own return type.

Examples

Sending a LVM_SETITEM list view message, which demands a reference to an LVITEM object:

use winsafe::co;
use winsafe::msg::LvmSetItem;
use winsafe::{LVITEM, HWND};

let my_hwnd: HWND; // initialize it somewhere...

let mut lvi = LVITEM::default(); // object to be sent
lvi.mask = co::LVIF::IMAGE;
lvi.iImage = 3;

my_hwnd.SendMessage(LvmSetItem {
    lvitem: &lvi,
});

pub fn SetFocus(self) -> Option<HWND>

SetFocus method.

pub fn SetParent(self, hWndNewParent: HWND) -> WinResult<Option<HWND>>

SetParent method.

pub fn SetWindowDisplayAffinity(self, dwAffinity: WDA) -> WinResult<()>

SetWindowDisplayAffinity method.

pub fn SetWindowLongPtr(self, nIndex: GWLP, dwNewLong: isize) -> isize

SetWindowLongPtr method.

pub fn SetWindowPlacement(self, lpwndpl: &WINDOWPLACEMENT) -> WinResult<()>

SetWindowPlacement method.

pub fn SetWindowPos(
self,
hWndInsertAfter: HwndPlace,
X: i32,
Y: i32,
cx: u32,
cy: u32,
uFlags: SWP
) -> WinResult<()>

SetWindowPos method.

pub fn SetWindowRgn(self, hRgn: HRGN, bRedraw: bool) -> WinResult<()>

SetWindowRgn method.

pub fn SetWindowSubclass(
self,
pfnSubclass: SUBCLASSPROC,
uIdSubclass: usize,
dwRefData: usize
) -> WinResult<()>

SetWindowSubclass method.

pub fn SetWindowText(self, lpString: &str) -> WinResult<()>

SetWindowText method.

pub fn ShowCaret(self) -> WinResult<()>

ShowCaret method.

pub fn ShowWindow(self, nCmdShow: SW) -> bool

ShowWindow method.

pub fn TranslateAccelerator(
self,
hAccTable: HACCEL,
lpMsg: &mut MSG
) -> WinResult<()>

TranslateAccelerator method.

pub fn UpdateWindow(self) -> WinResult<()>

UpdateWindow method.

pub fn ValidateRect(self, lpRect: &RECT) -> WinResult<()>

ValidateRect method.

pub fn ValidateRgn(self, hRgn: HRGN) -> WinResult<()>

ValidateRgn method.

pub fn WindowFromPhysicalPoint(point: POINT) -> Option<HWND>

WindowFromPhysicalPoint static method.

pub fn WindowFromPoint(point: POINT) -> Option<HWND>

WindowFromPoint static method.