Struct winsafe::Ini

pub struct Ini {
    pub sections: Vec<IniSection>,
}

This is supported on crate feature kernel only.

Keeps sections and key/value pairs of a .ini file, also doing parsing and serialization of the data.

Examples

use winsafe::prelude::*;
use winsafe::Ini;

let fini = Ini::parse_from_file("C:\\Temp\\my_file.ini")?;

for section in fini.sections.iter() {
    println!("Section: {}", section.name);
    for entry in section.entries.iter() {
        println!("Key: {}; Value: {}", entry.key, entry.val);
    }
}

Fields
sections: Vec<IniSection>

All the sections of the file. The values can be modified at will.

Implementations

impl Ini

pub fn parse_str(contents: &str) -> Ini

Parses an Ini from a string.

pub fn parse_bytes(bytes: &[u8]) -> WinResult<Ini>

Parses an Ini from raw bytes. The Encoding will be guessed with WString::guess_encoding.

pub fn parse_from_file(ini_path: &str) -> WinResult<Ini>

Parses an Ini directly from a file.

pub fn serialize_to_str(&self) -> String

Serializes the data to a string.

pub fn serialize_to_bytes(&self) -> Vec<u8>

Serializes the data to raw bytes with String::into_bytes.

pub fn serialize_to_file(&self, ini_path: &str) -> WinResult<()>

Serializes the data directly to a file.

pub fn value(&self, section: &str, key: &str) -> Option<&str>

Returns a reference to the specified value, if any.

pub fn value_mut(&mut self, section: &str, key: &str) -> Option<&mut str>

Returns a mutable reference to the specified value, if any.