Struct winsafe::VARIANT

#[repr(C)]
pub struct VARIANT { /* private fields */ }

Available on crate feature oleaut only.

VARIANT struct.

Automatically calls VariantClear when the object goes out of scope.

Implementations

impl VARIANT

pub const fn variant_type(&self) -> VT

Returns the co::VT variant type currently being stored in the object.

pub const unsafe fn raw(&self) -> &[u8; 16]

Returns a reference to the raw data held by the VARIANT.

pub fn is_empty(&self) -> bool

Tells whether the VARIANT holds no value.

pub fn is_null(&self) -> bool

Tells whether the VARIANT holds an SQL style null.

pub fn new_bool(val: bool) -> VARIANT

Crates a new VARIANT holding a bool value.

pub fn bool(&self) -> Option<bool>

If the VARIANT holds a bool value, returns it, otherwise None.

pub fn new_bstr(val: &str) -> VARIANT

Creates a new VARIANT holding a BSTR value.

pub fn bstr(&self) -> Option<String>

If the VARIANT holds a BSTR value, returns it, otherwise None.

pub fn new_f32(val: f32) -> VARIANT

Creates a new VARIANT holding an f32 value.

pub fn f32(&self) -> Option<f32>

If the VARIANT holds an f32 value, returns it, otherwise None.

pub fn new_f64(val: f64) -> VARIANT

Creates a new VARIANT holding an f64 value.

pub fn f64(&self) -> Option<f64>

If the VARIANT holds an f64 value, returns it, otherwise None.

pub fn new_i8(val: i8) -> VARIANT

Creates a new VARIANT holding an i8 value.

pub fn i8(&self) -> Option<i8>

If the VARIANT holds an i8 value, returns it, otherwise None.

pub fn new_i16(val: i16) -> VARIANT

Creates a new VARIANT holding an i16 value.

pub fn i16(&self) -> Option<i16>

If the VARIANT holds an i16 value, returns it, otherwise None.

pub fn new_i32(val: i32) -> VARIANT

Creates a new VARIANT holding an i32 value.

pub fn i32(&self) -> Option<i32>

If the VARIANT holds an i32 value, returns it, otherwise None.

pub fn new_i64(val: i64) -> VARIANT

Creates a new VARIANT holding an i64 value.

pub fn i64(&self) -> Option<i64>

If the VARIANT holds an i64 value, returns it, otherwise None.

pub fn new_idispatch<T>(val: &T) -> VARIANT where
T: oleaut_IDispatch,

Creates a new VARIANT holding an IDispatch COM value.

Note that the IDispatch object will be cloned into the VARIANT, still being able to be used thereafter.

pub fn idispatch<T>(&self) -> Option<T> where
T: oleaut_IDispatch,

If the VARIANT holds an IDispatch COM value, returns it, otherwise None.

Note that the returned object will be a clone of the IDispatch being held.

pub fn new_iunknown<T>(val: &T) -> VARIANT where
T: ole_IUnknown,

Creates a new VARIANT holding an IUnknown COM value.

Note that the IUnknown object will be cloned into the VARIANT, still being able to be used thereafter.

pub fn iunknown<T>(&self) -> Option<T> where
T: ole_IUnknown,

If the VARIANT holds an IUnknown COM value, returns it, otherwise None.

Note that the returned object will be a clone of the IUnknown being held.

pub fn new_time(val: &SYSTEMTIME) -> WinResult<VARIANT>

Creates a new VARIANT holding a date/time value.

pub fn time(&self) -> Option<SYSTEMTIME>

If the VARIANT holds a date/time value, returns it, otherwise None.

pub fn new_u8(val: u8) -> VARIANT

Creates a new VARIANT holding an u8 value.

pub fn u8(&self) -> Option<u8>

If the VARIANT holds an u8 value, returns it, otherwise None.

pub fn new_u16(val: u16) -> VARIANT

Creates a new VARIANT holding an u16 value.

pub fn u16(&self) -> Option<u16>

If the VARIANT holds an u16 value, returns it, otherwise None.

pub fn new_u32(val: u32) -> VARIANT

Creates a new VARIANT holding an u32 value.

pub fn u32(&self) -> Option<u32>

If the VARIANT holds an u32 value, returns it, otherwise None.

pub fn new_u64(val: u64) -> VARIANT

Creates a new VARIANT holding an u64 value.

pub fn u64(&self) -> Option<u64>

If the VARIANT holds an u64 value, returns it, otherwise None.