Struct winsafe::VARIANT

#[repr(C)]
pub struct VARIANT { /* private fields */ }

Available on crate feature oleaut only.

VARIANT struct.

Automatically calls VariantClear when the object goes out of scope.

Implementations

impl VARIANT

pub fn new_idispatch(val: &impl oleaut_IDispatch) -> VARIANT

Creates a new object holding an IDispatch COM value.

Note that the IDispatch object will be cloned into the object, still being able to be used thereafter.

pub fn idispatch<T>(&self) -> Option<T>where
T: oleaut_IDispatch,

If the object holds an IDispatch COM value, returns it, otherwise None.

Note that the returned object will be a clone of the IDispatch being held.

pub fn new_iunknown<T>(val: &impl ole_IUnknown) -> VARIANT

Creates a new object holding an IUnknown COM value.

Note that the IUnknown object will be cloned into the object, still being able to be used thereafter.

pub fn iunknown<T>(&self) -> Option<T>where
T: ole_IUnknown,

If the object holds an IUnknown COM value, returns it, otherwise None.

Note that the returned object will be a clone of the IUnknown being held.