Struct winsafe::VARIANT

#[repr(C)]
pub struct VARIANT { /* private fields */ }

Available on crate feature oleaut only.

VARIANT struct.

Automatically calls VariantClear when the object goes out of scope.

The Default implementation returns a co::VT::EMPTY value.

Implementations

impl VARIANT

pub fn new_idispatch(val: &impl oleaut_IDispatch) -> Self

Creates a new object holding an IDispatch COM value.

Note that val will be cloned into the VARIANT – that is, IUnknown::AddRef will be called –, so val will remain valid to be used thereafter.

pub fn idispatch<T>(&self) -> Option<T>
where T: oleaut_IDispatch,

If the object holds an IDispatch COM value, returns it, otherwise None.

Note that the returned object will be a clone – that is, IUnknown::AddRef will be called.

pub fn new_iunknown<T>(val: &impl ole_IUnknown) -> Self

Creates a new object holding an IUnknown COM value.

Note that val will be cloned into the VARIANT – that is, IUnknown::AddRef will be called –, so val will remain valid to be used thereafter.

pub fn iunknown<T>(&self) -> Option<T>
where T: ole_IUnknown,

If the object holds an IUnknown COM value, returns it, otherwise None.

Note that the returned object will be a clone – that is, IUnknown::AddRef will be called.