Module winsafe::task_dlg

Provides high-level abstractions to TaskDialogIndirect and HWND::TaskDialog functions.

Functions

errorcomctl and ole

Displays an error modal window with an OK button.

infocomctl and ole

Displays an information modal window with an OK button.

ok_cancelcomctl and ole

Displays a question modal window with OK and Cancel buttons. The text of the OK button can be customized.

yes_nocomctl and ole

Displays a question modal window with Yes and No buttons.

yes_no_cancelcomctl and ole

Displays a question modal window with Yes, No and Cancel buttons.